Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AMQPClient.basic_cancel — Method.

Cancels a consumer.

This does not affect already delivered messages, but it does mean the server will not send any more messages for that consumer. The client may receive an arbitrary number of messages in between sending the cancel method and receiving the cancel­ok reply.

AMQPClient.basic_consume — Method.

Start a queue consumer.

queue: queue name consumertag: id of the consumer, server generates a unique tag if this is empty nolocal: do not deliver own messages no_ack: no acknowledgment needed, server automatically and silently acknowledges delivery (speed at the cost of reliability) exclusive: request exclusive access (only this consumer can access the queue) nowait: do not send a reply method

AMQPClient.basic_publish — Method.

Publish a message

This method publishes a message to a specific exchange. The message will be routed to queues as defined by the exchange configuration and distributed to any active consumers when the transaction, if any, is committed.

AMQPClient.queue_declare — Method.

Declare a queue (or query an existing queue). Returns a tuple: (boolean success/failure, queue name, message count, consumer count)

AMQPClient.queue_delete — Method.

Delete a queue. Returns a tuple: (boolean success/failure, message count)

AMQPClient.queue_purge — Method.

Purge messages from a queue. Returns a tuple: (boolean success/failure, message count)

Base.read — Method.

Read a generic frame. All frames have the following wire format:

0 1 3 7 size+7 size+8 +–––+––––-+––––-+ +––––––-+ +–––––-+ | type | channel | size | | payload | | frame-end | +–––+––––-+––––-+ +––––––-+ +–––––-+ octet short long 'size' octets octet

AMQPClient.is_method — Method.

Validate if the method frame is for the given class and method.