Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AWSCore.AWSConfig — Type.

Most AWSCore functions take a AWSConfig dictionary as the first argument. This dictionary holds AWSCredentials and AWS region configuration.

aws = AWSConfig(:creds => AWSCredentials(), :region => "us-east-1")`

AWSCore.AWSCredentials — Type.

When you interact with AWS, you specify your AWS Security Credentials to verify who you are and whether you have permission to access the resources that you are requesting. AWS uses the security credentials to authenticate and authorize your requests.

The fields access_key_id and secret_key hold the access keys used to authenticate API requests (see Creating, Modifying, and Viewing Access Keys).

Temporary Security Credentials require the extra session token field.

The user_arn and account_number fields are used to cache the result of the aws_user_arn and aws_account_number functions.

The AWSCredentials() constructor tries to load local Credentials from environment variables, ~/.aws/credentials, ~/.aws/config or EC2 instance credentials. To specify the profile to use from ~/.aws/credentials, do, for example, AWSCredentials(profile="profile-name").

AWSCore.AWSRequest — Type.

The AWSRequest dictionary describes a single API request: It contains the following keys:

:creds => AWSCredentials for authentication.

:verb => "GET", "PUT", "POST" or "DELETE"

:url => service endpoint url (returned by aws_endpoint)

:headers => HTTP headers

:content => HTTP body

:resource => HTTP request path

:region => AWS region

:service => AWS service name

AWSCore.arn — Function.

arn([::AWSConfig], service, resource, [region, [account]])

Generate an Amazon Resource Name for service and resource.

AWSCore.arn_account — Method.

arn_account(arn)

Extract account number from arn.

AWSCore.arn_iam_name — Method.

arn_iam_name

Extract IAM resource name from arn.

AWSCore.arn_iam_type — Method.

arn_iam_type

Extract IAM resource type from arn. e.g. "role", "policy"...

AWSCore.arn_region — Method.

arn_region(arn)

Extract region name from arn.

AWSCore.arn_resource — Method.

arn_resource(arn)

Extract resource name from arn.

AWSCore.arn_service — Method.

arn_service(arn)

Extract service name from arn.

AWSCore.aws_account_number — Method.

aws_account_number(::AWSConfig)

12-digit AWS Account Number.

AWSCore.aws_config — Method.

The aws_config function provides a simple way to creates an AWSConfig configuration dictionary.

>aws = aws_config()
>aws = aws_config(creds = my_credentials)
>aws = aws_config(region = "ap-southeast-2")
>aws = aws_config(profile = "profile-name")

By default, the aws_config attempts to load AWS credentials from:

AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY environemnt variables,

~/.aws/credentials or

EC2 Instance Credentials.

A ~/.aws/credentials file can be created using the AWS CLI command aws configrue. Or it can be created manually:

[default]
aws_access_key_id = AKIAXXXXXXXXXXXXXXXX
aws_secret_access_key = XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX

If your ~/.aws/credentials file contains multiple profiles you can pass the profile name as a string to the profile keyword argument (nothing by default) or select a profile by setting the AWS_PROFILE environment variable.

aws_config understands the following AWS CLI environment variables: AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_SESSION_TOKEN, AWS_DEFAULT_REGION, AWS_PROFILE and AWS_CONFIG_FILE.

An configuration dictionary can also be created directly from a key pair as follows. However, putting access credentials in source code is discouraged.

aws = aws_config(creds = AWSCredentials("AKIAXXXXXXXXXXXXXXXX",
                                        "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"))

AWSCore.aws_user_arn — Method.

aws_user_arn(::AWSConfig)

Unique Amazon Resource Name for configrued user.

e.g. "arn:aws:iam::account-ID-without-hyphens:user/Bob"

AWSCore.default_aws_config — Method.

default_aws_config returns a global shared AWSConfig object obtained by calling aws_config with no optional arguments.

AWSCore.is_arn — Method.

is_arn(arn)

Is arn in the [correct format]? (http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html)

AWSCore.localhost_is_ec2 — Method.

Is Julia running on an EC2 virtual machine?

AWSCore.localhost_is_lambda — Method.

Is Julia running in an AWS Lambda sandbox?

AWSCore.aws_args_dict — Method.

aws_args_dict(args)

Convert nested Vector{Pair} maps in args into Dict{String,Any} maps.

AWSCore.do_request — Method.

do_request(::AWSRequest)

Submit an API request, return the result.

AWSCore.dot_aws_credentials — Function.

Load Credentials from [AWS CLI ~/.aws/credentials file] or AWS CLI ~/.aws/config file.

AWSCore.dump_aws_request — Method.

Pretty-print AWSRequest dictionary.

AWSCore.ec2_instance_credentials — Method.

Load Instance Profile Credentials for EC2 virtual machine.

AWSCore.ec2_metadata — Method.

ec2_metadata(key)

Fetch EC2 meta-data for key.

AWSCore.ecs_instance_credentials — Method.

Load ECS Task Credentials

AWSCore.env_instance_credentials — Method.

Load Credentials from environment variables AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY etc. (e.g. in Lambda sandbox).

AWSCore.flatten_query — Function.

flatten_query(service, query, prefix="")

Recursivly flatten tree of Dicts and Arrays into a 1-level deep Dict.

AWSCore.mime_multipart — Method.

mime_multipart([header,] parts)

Encode parts as a MIME Multipart message.

parts is a Vector of (filename, content_type, content) Tuples.

AWSCore.rest_resource — Method.

rest_resource(request, args)

Replace {Arg} placeholders in request[:resource] with arg values.

AWSCore.service_json — Method.

service_json(aws::AWSConfig; args...)

Process request for AWS "json" service protocol.

AWSCore.service_query — Method.

service_query(aws::AWSConfig; args...)

Process request for AWS "query" service protocol.

AWSCore.service_rest_json — Method.

service_rest_json(aws::AWSConfig; args...)

Process request for AWS "rest_json" service protocol.

AWSCore.service_rest_xml — Method.

service_rest_xml(aws::AWSConfig; args...)

Process request for AWS "rest_xml" service protocol.

AWSCore.service_url — Method.

service_url(aws::AWSConfig; request)

Service endpoint URL for request.