Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AWSS3.s3_arn — Method.

s3_arn(resource)
s3_arn(bucket,path)

Amazon Resource Name for S3 resource or bucket and path.

AWSS3.s3_copy — Method.

s3_copy([::AWSConfig], bucket, path; to_bucket=bucket, to_path=path)

PUT Object - Copy

Optional Arguments

metadata::Dict=; optional x-amz-meta- headers.

AWSS3.s3_create_bucket — Method.

s3_create_bucket([:AWSConfig], bucket)

PUT Bucket

AWSS3.s3_delete — Method.

s3_delete([::AWSConfig], bucket, path; [version=]

DELETE Object

AWSS3.s3_delete_bucket — Method.

s3_delete_bucket([::AWSConfig], "bucket")

DELETE Bucket.

AWSS3.s3_delete_tags — Function.

s3_delete_tags([::AWSConfig], bucket, [path])

Delete tags from bucket or object (path).

AWSS3.s3_enable_versioning — Method.

s3_enable_versioning([::AWSConfig], bucket)

Enable versioning for bucket.

AWSS3.s3_exists — Method.

s3_exists([::AWSConfig], bucket, path [version=])

Is there an object in bucket at path?

AWSS3.s3_get — Method.

s3_get([::AWSConfig], bucket, path; <keyword arguments>)

Get Object from path in bucket.

Optional Arguments

version=: version of object to get.

retry=true: try again on "NoSuchBucket", "NoSuchKey" (common if object was recently created).

raw=false: return response as Vector{UInt8} (by default return type depends on Content-Type header).

AWSS3.s3_get_file — Method.

s3_get_file([::AWSConfig], bucket, path, filename; [version=])

Like s3_get but streams result directly to filename.

AWSS3.s3_get_meta — Method.

s3getmeta([::AWSConfig], bucket, path; [version=])

HEAD Object

Retrieves metadata from an object without returning the object itself.

AWSS3.s3_get_tags — Function.

s3_get_tags([::AWSConfig], bucket, [path])

Get tags from bucket or object (path).

AWSS3.s3_list_buckets — Function.

s3_list_buckets([::AWSConfig])

List of all buckets owned by the sender of the request.

AWSS3.s3_list_keys — Function.

s3_list_keys([::AWSConfig], bucket, [path_prefix])

Like s3_list_objects but returns object keys as Vector{String}.

AWSS3.s3_list_objects — Function.

s3_list_objects([::AWSConfig], bucket, [path_prefix])

List Objects in bucket with optional path_prefix.

Returns Vector{Dict} with keys Key, LastModified, ETag, Size, Owner, StorageClass.

AWSS3.s3_list_versions — Function.

s3_list_versions([::AWSConfig], bucket, [path_prefix])

List object versions in bucket with optional path_prefix.

AWSS3.s3_purge_versions — Function.

s3_purge_versions([::AWSConfig], bucket, [path [, pattern]])

DELETE all object versions except for the latest version.

AWSS3.s3_put — Function.

s3_put([::AWSConfig], bucket, path, data; <keyword arguments>

PUT Object data at path in bucket.

Optional Arguments

data_type=; Content-Type header.

encoding=; Content-Encoding header.

metadata::Dict=; x-amz-meta- headers.

tags::Dict=; x-amz-tagging- headers (see also s3_put_tags and s3_get_tags).

AWSS3.s3_put_cors — Method.

s3_put_cors([::AWSConfig], bucket, cors_config)

PUT Bucket cors

s3_put_cors("my_bucket", """
    <?xml version="1.0" encoding="UTF-8"?>
    <CORSConfiguration xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
        <CORSRule>
            <AllowedOrigin>http://my.domain.com</AllowedOrigin>
            <AllowedOrigin>http://my.other.domain.com</AllowedOrigin>
            <AllowedMethod>GET</AllowedMethod>
            <AllowedMethod>HEAD</AllowedMethod>
            <AllowedHeader>*</AllowedHeader>
            <ExposeHeader>Content-Range</ExposeHeader>
        </CORSRule>
    </CORSConfiguration>
"""

AWSS3.s3_put_tags — Method.

s3_put_tags([::AWSConfig], bucket, [path,] tags::Dict)

PUT tags on bucket or object (path).

See also tags= option on s3_put.

AWSS3.s3_sign_url — Function.

s3_sign_url([::AWSConfig], bucket, path, [seconds=3600];
            [verb="GET"], [content_type="application/octet-stream"])

Create a pre-signed url for bucket and path (expires after for seconds).

To create an upload URL use verb="PUT" and set content_type to match the type used in the Content-Type header of the PUT request.

url = s3_sign_url("my_bucket", "my_file.txt"; verb="PUT")
Requests.put(URI(url), "Hello!")

url = s3_sign_url("my_bucket", "my_file.txt";
                  verb="PUT", content_type="text/plain")

Requests.put(URI(url), "Hello!";
             headers=Dict("Content-Type" => "text/plain"))