Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AbstractInstances.concretetype — Method.

concretetype(T::Type)

Return a concrete mutable type with no fields which is a subtype of T. The mutability of the returned type ensures that instances of the type are unique.

AbstractInstances.oftype — Method.

oftype(T::Type)

Create a unique instance of type, concretetype(T).

AbstractInstances.singleton — Method.

singleton(T::Type)

Create an instance of type, singletontype(T).

AbstractInstances.singletontype — Method.

singletontype(T::Type)

Return a concrete immutable type with no fields which is a subtype of T. The immutability of the returned type ensures that instances of the type are singletons.