Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AhoCorasickAutomatons.ACPosition — Type.

See ?eachmatch

AhoCorasickAutomatons.AhoCorasickAutomaton — Type.

AhoCorasickAutomaton{T <: Unsigned}
        s.t.
typemax(T) >= maximum(nodes)

A 2-Array implementation of Aho–Corasick automaton(ACA)[1], Most used as an engine to mine a long text string, to get all occurences of substrings interested in the text. ACA is also adequate for unicode strings as a Set{String} or Dict{String, Unsigned} (similar as Trie, but with a far more less space usage).

The ACA acts as a key data structrue in Aho–Corasick algorithm for multiple consecutive string patterns finding[2]. This particular implementation use no-decreasing base, i.e.,

1. base[node] >= node
2. children[node] > node

This choice make it suitable for large-size key set of not-very-long keys, with faster insertion speed and moderate space usage.

Examples

using Pkg; less(joinpath(Pkg.dir("AhoCorasickAutomatons"), "test", "runtests.jl"))

Notes

Maybe of much more slower speed than a oridinary tree-based ACA, specially for random generated keys.

When inserting duplicated keys, only the last one will make sense.

[1]
Jun‐Ichi Aoe, Katsushi Morimoto and Takashi Sato 'An Efficient Implementation of Trie Structures', September 1992.

[2]
Aho, Alfred V.; Corasick, Margaret J. (June 1975). "Efficient string matching: An aid to bibliographic search". Communications of the ACM. 18 (6): 333&ndash, 340. doi:10.1145/360825.360855. MR 0371172.

Base.eachmatch — Method.

eachmatch(obj::AhoCorasickAutomaton{T}, text::AbstractString)::Vector{ACPosition{T}} where T

Search for all matches of a AhoCorasickAutomaton obj in text and return a vector of the matches. Each matches is represented as a ACPosition type, which has 3 fields:

1. s : start of match
2. t : stop of match
3. i : index of the key in *obj*, which is the original insertion order of keys to *obj*

The field i may be use as index of external property arrays, i.e., the AhoCorasickAutomaton can act as a Map{String, Any}.