Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Algencan.AlgencanSolver — Type.

Algencan solver that stores options

Algencan.AlgencanSolver — Method.

Store keyworded argments as options

Algencan.setmultwarmstart! — Method.

Set an inital value for the constaint mutipliers (warmstart)

Algencan.AlgencanMathProgModel — Type.

Algencan model, that storing solution data

Algencan.find_status — Method.

Find final status of Algencan

Algencan.julia_fc — Method.

Compute objective and constraints as required by Algencan

Algencan.julia_gjac — Method.

Compute objective gradient and constraints Jacobian as required by Algencan

Algencan.julia_hl — Method.

Compute the Hessian of the Lagrangian as required by Algencan

Algencan.julia_hlp — Method.

Compute the Hessian of the Lagrangian times p as required by Algencan

Algencan.option2vparam — Method.

Transform the option dictionary into a vparam string array

Algencan.treat_lower_bounds — Method.

Analyse the lower and upper bounds on the constraints and prepare the data structure to treat lower bounds.

Base.copy — Method.

Return a copy of the current model

MathProgBase.SolverInterface.getobjbound — Method.

Get best bound on (local) optimal value

MathProgBase.SolverInterface.getobjgap — Method.

Get gap to (local) optimality

MathProgBase.SolverInterface.getobjval — Method.

Get objective value at current primal point

MathProgBase.SolverInterface.getrawsolver — Method.

There is no inner solver, all functionality is exposed by the default interface

MathProgBase.SolverInterface.getreducedcosts — Method.

Return mutipliers associated with bound cosntraints

MathProgBase.SolverInterface.getsense — Method.

Return problem sense

MathProgBase.SolverInterface.getsolution — Method.

Return the current primal point, the solution after calling optimize

MathProgBase.SolverInterface.getsolvetime — Method.

Get the solution time

MathProgBase.SolverInterface.getvartype — Method.

Return variable types: they are all continuous

MathProgBase.SolverInterface.loadproblem! — Method.

Loads the problem with its basic data and functions in a NLPEvaluator

MathProgBase.SolverInterface.numconstr — Method.

Return the number of constraints

MathProgBase.SolverInterface.numvar — Method.

Return the number of decision variables

MathProgBase.SolverInterface.setparameters! — Method.

Set parameter for a solver, that will be default for all next models

You can use any Algencan parameter that can be set in a specification file plus :epsfeas, :epsopt, :efstain, :eostain, efacc, :eoac, :outputfnm, :specfnm.

See more details in Algencan documentation.

MathProgBase.SolverInterface.setsense! — Method.

Change optimization sense, either :Min or :Max

MathProgBase.SolverInterface.setvartype! — Method.

Algencan only deals with continuous variable, so inform if this is not the case

MathProgBase.SolverInterface.setwarmstart! — Method.

Set an initial value for the decision variables (warmstart)

MathProgBase.SolverInterface.status — Method.

Get current model status