Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AllanDeviations.allandev — Method.

allandev(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the allan deviation

#parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

#returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

AllanDeviations.hadamarddev — Method.

hadamarddev(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the hadamard deviation

#parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

#returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

AllanDeviations.mallandev — Method.

mallandev(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the modified allan deviation

#parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

#returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

AllanDeviations.mtie — Method.

mtie(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the maximal time interval error

parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

AllanDeviations.timedev — Method.

timedev(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the time deviation

#parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

#returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

AllanDeviations.totaldev — Method.

totaldev(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the total deviation

#parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

#returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.