Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AmplNLReader.write_sol — Method.

Write message msg along with primal and dual variables x and y to file.

LinearOperators.reset! — Method.

Reset evaluation counters in nlp.

NLPModels.cons! — Method.

Evaluate the constraints at x in place.

NLPModels.cons — Method.

Evaluate the constraints at x.

NLPModels.conscale — Method.

Scale the vector of constraints by the vector s.

NLPModels.ghjvprod! — Method.

Compute the vector of dot products (g, Hj*v) in place where Hj is the Hessian of the j-th constraint at x.

NLPModels.ghjvprod — Method.

Compute the vector of dot products (g, Hj*v) where Hj is the Hessian of the j-th constraint at x.

NLPModels.grad! — Method.

Evaluate the gradient of the objective function at x in place.

NLPModels.grad — Method.

Evaluate the gradient of the objective function at x.

NLPModels.hess — Method.

Evaluate the Lagrangian Hessian at (x,y) as a sparse matrix. Only the lower triangle is returned.

NLPModels.hess_coord — Method.

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format. Only the lower triangle is returned.

NLPModels.hprod! — Method.

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v in place.

NLPModels.hprod — Method.

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v.

NLPModels.jac — Method.

Evaluate the constraints Jacobian at x as a sparse matrix.

NLPModels.jac_coord — Method.

Evaluate the constraints Jacobian at x in sparse coordinate format.

NLPModels.jprod! — Method.

Evaluate the Jacobian-vector product at x in place. Warning: Currently building the Jacobian for this.

NLPModels.jprod — Method.

Evaluate the Jacobian-vector product at x. Warning: Currently building the Jacobian for this.

NLPModels.jth_con — Method.

Evaluate the j-th constraint at x.

NLPModels.jth_congrad! — Method.

Evaluate the j-th constraint gradient at x in place.

NLPModels.jth_congrad — Method.

Evaluate the j-th constraint gradient at x.

NLPModels.jth_hprod! — Method.

Evaluate the product of the j-th constraint Hessian at x with the vector v in place. The objective Hessian is used if j=0.

NLPModels.jth_hprod — Method.

Evaluate the product of the j-th constraint Hessian at x with the vector v. The objective Hessian is used if j=0.

NLPModels.jth_sparse_congrad — Method.

Evaluate the j-th constraint sparse gradient at x.

NLPModels.jtprod! — Method.

Evaluate the transposed-Jacobian-vector product at x in place. Warning: Currently building the Jacobian for this.

NLPModels.jtprod — Method.

Evaluate the transposed-Jacobian-vector product at x. Warning: Currently building the Jacobian for this.

NLPModels.lagscale — Method.

Set the scaling factor σ in the Lagrangian: L(x,y) = f(x) + σ ∑ yi ci(x).

NLPModels.obj — Method.

Evaluate the objective function of nlp at x.

NLPModels.varscale — Method.

Scale the vector of variables by the vector s.