ApproximateBayesianComputing.jl: Approximate Bayesian Computing for Julia

ApproximateBayesianComputing.jl: Approximate Bayesian Computing for Julia

This package provides for Approximate Bayesian Computing (ABC) via sequential importance sampling in Julia. Currently, it implements a single algorithm, ABC-PMC based on Beaumont et al. 2002 via abcpmcplan_type. However, it does include several extra features that will eventually be documented.

Getting Started

Add the ApproximateBayesianComputing.jl package (update once become registered)

using ApproximateBayesianComputing.jl

include("tests/runtests.jl")

See test1.jl for a demo of how to use ABC

API

ApproximateBayesianComputing.ApproximateBayesianComputing — Module.

ApproximateBayesianComputing Module providing types (abcpmcplantype, abcpopulationtype) and methods (generatetheta, initabc, updateabcpop, runabc) for using Approximate Bayesian Computing

Index

ApproximateBayesianComputing

ApproximateBayesianComputing.ApproximateBayesianComputing

ApproximateBayesianComputing.abc_population_type