-

ApproximateBayesianComputing — Module.

ApproximateBayesianComputing Module providing types (abcpmcplantype, abcpopulationtype) and methods (generatetheta, initabc, updateabcpop, runabc) for using Approximate Bayesian Computing

ApproximateBayesianComputing.abc_pmc_plan_type

abc_pmc_plan_type(gd::Function,css::Function,cd::Function,p)

ApproximateBayesianComputing.abc_population_type — Method.

abcpopulationtype(numparam::Integer, numparticles::Integer; acceptlog::abclogtype = abclogtype(), rejectlog::abclogtype = abclogtype(), repeats::Array{Int64,1} = zeros(Int64,num_particles) )

numparam: Number of model parameters for generating simulated data numparticles: Number of particles for sequential importance sampler

Optional parameters: acceptlog: Log of accepted parameters/summary statistics/distances rejectlog: Log of rejected parameters/summary statistics/distances repeats: Array indicating which particles have been repeated from previous generation