Functions

Function Reference

Function Reference

AprilTags.jl functions

Wrappers

Index

AprilTags.jl functions

AprilTags.AprilTagDetector — Type.

AprilTagDetector(tagfamily=tag36h11)

Create a default AprilTag detector with the 36h11 tag family Create an AprilTag detector with tag family in tagfamily::TagFamilies @enum TagFamilies tag36h11 tag36h10 tag25h9 tag16h5

AprilTags.freeDetector! — Function.

freeDetector!(apriltagdetector)

Free the allocated memmory

AprilTags.homography_to_pose — Function.

homography_to_pose(H, fx, fy, cx, cy, [taglength = 2.0])

Given a 3x3 homography matrix and the camera model (focal length and centre), compute the pose of the tag. The focal lengths should be given in pixels. The returned units are those of the tag size, therefore the translational components should be scaled with the tag size. Note: the tag coordinates are from (-1,-1) to (1,1), i.e. the tag size has lenght of 2 units. Optionally, the tag length (in metre) can be passed to return a scaled value.

AprilTags.homographytopose — Function.

homographytopose(H, fx, fy, cx, cy, [taglength = 2.0])

Given a 3x3 homography matrix and the camera model (focal length and centre), compute the pose of the tag. The focal lengths should be given in pixels. The returned units are those of the tag size, therefore the translational components should be scaled with the tag size. Note: the tag coordinates are from (-1,-1) to (1,1), i.e. the tag size has lenght of 2 units. Optionally, the tag length (in metre) can be passed to return a scaled value. The camara coordinate system: camera looking in positive Z axis with x to the right and y down.

AprilTags.drawTagBox! — Function.

drawTagBox!(image, tag)

Draw a box around the tag. imageCol = RGB.(image) foreach(tag->drawTagBox!(imageCol, tag), tags)

AprilTags.drawTagAxes! — Function.

drawTagAxes!(image, tag, CameraMatrix)

Draw the tag x, y, and z axes to show the orientation. imageCol = RGB.(image) foreach(tag->drawTagAxes!(imageCol, tag, K), tags)

AprilTags.getAprilTagImage — Function.

getAprilTagImage(tagIndex, tagfamily=tag36h11)

Return an image [Gray{N0f8}] for with tagIndex from tag family in tagfamily::TagFamilies @enum TagFamilies tag36h11 tag36h10 tag25h9 tag16h5

AprilTags.threadcalldetect — Function.

threadcalldetect(detector, image)

Run the april tag detector on a image

Wrappers

AprilTags.apriltag_detector_create — Function.

apriltag_detector_create()

Create a AprilTag Detector object with all fields set to default value.

AprilTags.tag36h11_create — Function.

tag36h11_create()

Create a AprilTag family object for tag36h11 with all fields set to default value.

AprilTags.tag36h11_destroy — Function.

tag36h11_destroy(tf)

Destroy the AprilTag family object.

AprilTags.apriltag_detector_add_family — Function.

apriltag_detector_add_family(tag_detector, tag_family)

Add a tag family to an AprilTag Detector object. The caller still "owns" the family and a single instance should only be provided to one apriltag detector instance.

AprilTags.apriltag_detector_detect — Function.

apriltag_detector_detect(tag_detector, image)

Detect tags from an image and return an array of apriltagdetectiont*. You can use apriltagdetectionsdestroy to free the array and the detections it contains, or call detectiondestroy and zarraydestroy yourself.

AprilTags.threadcall_apriltag_detector_detect — Function.

threadcall_apriltag_detector_detect(tag_detector, image)

Experimental call apriltagdetectordetect in a seperate thread using the experimantal @threadcall Detect tags from an image and return an array of apriltagdetectiont*. You can use apriltagdetectionsdestroy to free the array and the detections it contains, or call detectiondestroy and zarraydestroy yourself.

Index

AprilTags.AprilTagDetector

AprilTags.apriltag_detector_add_family

AprilTags.apriltag_detector_create

AprilTags.apriltag_detector_detect

AprilTags.drawTagAxes!

AprilTags.drawTagBox!

AprilTags.freeDetector!

AprilTags.getAprilTagImage

AprilTags.homography_to_pose

AprilTags.homographytopose

AprilTags.tag36h11_create

AprilTags.tag36h11_destroy

AprilTags.threadcall_apriltag_detector_detect

AprilTags.threadcalldetect