Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Arrow.Datestamp — Type.

Datestamp <: ArrowTime

Stores a date as an Int32 for Arrow formatted data.

Arrow.DictEncoding — Type.

DictEncoding{P<:ArrowVector,J} <: ArrowVector{J}

Arrow.List — Type.

List{P<:AbstractPrimitive,J} <: AbstractList{J}

An Arrow formatted array of variable length objects such as strings. The List contains "value" data as well as "offsets" which describe from which elements of the values data an element of List should be constructed. The offsets are necessarily a Primitive{Int32} while the values can be any ArrowVector type (but in most circumstances should be Primitive).

Constructors

List{J}(len::Integer, offs::Primitive{Int32}, vals::AbstractPrimitive)
List{J}(offs::Primitive{Int32}, vals::AbstractPrimitive)
List{J}(data::Vector{UInt8}, offset_idx::Integer, len::Integer, vals::AbstractPrimitive)
List{J}(data::Vector{UInt8}, offset_idx::Integer, values_idx::Integer, ::Type{C}, x::AbstractVector)
List{J}(data::Vector{UInt8}, i::Integer, ::Type{C}, x::AbstractVector)
List(data::Vector{UInt8}, i::Integer, ::Type{C}, x::AbstractVector)
List{J}(Array, ::Type{C}, x::AbstractVector)
List(Array, ::Type{C}, x::AbstractVector)
List(::Type{C}, v::AbstractVector)
List(v::AbstractVector{<:AbstractString})

Note that by default, Lists of strings will be encoded in UTF-8.

Arguments

len: the length of the List

offs: a Primitive{Int32} containing the offsets data

vals: a Primitive containing the values data

data: the data buffer in which the underlying data is stored

offset_idx: location within data where the offset data is stored

values_idx: location within data where the values data is stored

C: the encoding type (type of values), e.g. for UTF-8 strings this is UInt8. defaults to UInt8 if not given explicitly

i: the location in data where all data should be stored (offsets, then values)

x, v: array to be stored or converted

Arrow.NullableList — Type.

NullableList{P<:AbstractPrimitive,J} <: AbstractList{Union{Missing,J}}

An Arrow formatted array of variable length objects such as strings which may be null. The NullableList contains a bit mask specifying which values are null and "offsets" which describe from which elements of the values data an element of the NullableList should be constructed. The bitmask is contained in a Primitive{UInt8} while the offsets data in a Primitive{Int32}. The values can be contained in any ArrowVector type, but in most cases should be Primitive.

Constructors

NullableList{J}(len::Integer, bmask::Primitive, offs::Primitive, vals::AbstractPrimitive)
NullableList{J}(bmask::Primitive, offs::Primitive, vals::AbstractPrimitive)
NullableList{J}(data::Vector{UInt8}, bitmask_idx::Integer, offset_idx::Integer, len::Integer,
                vals::AbstractPrimitive)
NullableList{J}(data::Vector{UInt8}, bitmask_idx::Integer, offset_idx::Integer, values_idx::Integer,
                len::Integer, ::Type{C}, values_len::Integer)
NullableList{J}(data::Vector{UInt8}, bitmask_idx::Integer, offset_idx::Integer, values_idx::Integer,
                ::Type{C}, x::AbstractVector)
NullableList(data::Vector{UInt8}, i::Integer, ::Type{C}, x::AbstractVector)
NullableList(Array, ::Type{C}, x::AbstractVector)
NullableList(Array, x::AbstracVector)
NullableList(::Type{C}, v::AbstractVector)
NullableList(v::AbstractVector)

If Array is given as an argument, a contiguous array will be allocated to store the data.

Arguments

len: the length of the NullableList

bmask: the Primitive providing the bit mask

offs: the Primitive providing the offsets

vals: the AbstractPrimitive providing the values

data: a buffer for storing the data

bitmask_idx: the location in data of the bit mask

offsets_idx: the location in data of the offsets

values_idx: the location in data of the values

values_len: the total length of the values data (i.e. number of elements in the values array)

C: the data type of the values data. defaults to UInt8 when not provided

x, v: array to be stored by the NullableList

Arrow.NullablePrimitive — Type.

NullablePrimitive{J} <: AbstractPrimitive{Union{J,Missing}}

A set of values stored contiguously in data with a bit mask specifying which values should be considered null. The bit mask and data needn't necessarily coexist within the same array.

Constructors

NullablePrimitive(bmask::Primitive{UInt8}, vals::Primitive{J})
NullablePrimitive(data::Vector{UInt8}, bitmask_idx::Integer, values_idx::Integer, x::AbstractVector)
NullablePrimitive(data::Vector{UInt8}, i::Integer, x::AbstractVector)
NullablePrimitive(v::AbstractVector)
NullablePrimitive(Array, x::AbstractVector)

If Array is passed to a constructor, the bit mask and values for the NullablePrimitive will be contiguously allocated within a single array (bit mask first, then values).

Arguments

bmask: a Primitive{UInt8} containing the null bitmask for the NullablePrimitive

vals: a Primitive containing the underlying data values for the NullablePrimitive

data: a buffer in which the data for the NullablePrimitive will be stored

bitmask_idx: the location within data where the null bit mask will be stored

values_idx: the location within data where the values will be stored

x, v: values to be stored in data

Arrow.Primitive — Type.

Primitive{J} <: AbstractPrimitive{J}

A set of values which are stored contiguously in memory. All other ArrowVector objects are built from Primitives.

Constructors

Primitive{J}(data::Vector{UInt8}, i::Integer, len::Integer)
Primitive(data::Vector{UInt8}, i::Integer, x::AbstractVector)
Primitive(v::AbstractVector)
Primitive(Array, x::Abstractvector)

If Array is provided as the virst argument, the data will be allocated contiguously in a new buffer.

Arguments

data: a data buffer which the Primitive will refer to for accessing and storing values

i: the index of the location (1-based) in data where the beginning of the value data is stored

x: a vector which will be written into data[i] on construction

v: existing reference data. constructors with will reference the originalvasdata`

Arrow.TimeOfDay — Type.

TimeOfDay{P<:Dates.TimePeriod,T<:Union{Int32,Int64}} <: ArrowTime

An arrow formatted object for representing the time of day. Underlying data is Int32 for seconds and milliseconds, Int64 for microsecond and nanosecond.

Arrow.Timestamp — Type.

Timestamp{P<:Dates.TimePeriod} <: ArrowTime

Timestamp in which time is stored in units P as Int64 for Arrow formatted data.

Arrow.arrowformat — Function.

arrowformat(v::AbstractVector)
arrowformat(Array, v::AbstractVector)

Convert a vector to the appropriate arrowformat. If Array is passed, a contiguous array will be used for the data buffer.

Arrow.arrowview — Method.

arrowview(p::ArrowVector, idx::UnitRange)

Return another ArrowVector which is a view into p for the index range idx.

This can only be done with contiguous indices.

Arrow.bitmaskbytes — Method.

bitmaskbytes(A::AbstractVector)
bitmaskbytes(::Type{Union{J,Missing}}, A::AbstractVector)

Compute the number of bytes needed to store a null bitmask for the data in A. This is 0 unless J <: Union{K,Missing}. Note that this does not take into account scheme-dependent padding.

Arrow.bitpack — Method.

bitpack(A::AbstractVector{Bool})

Returns a Vector{UInt8} the bits of which are the values of A.

Arrow.bytesforbits — Method.

bytesforbits(n::Integer)

Get the number of bytes required to store n bits.

Arrow.getoffset — Method.

getoffset(l::AbstractList, i::Integer)

Retrieve offset i for list l. Note that this retrieves the Arrow formated 0-based indexed raw numbers!

Arrow.isnull — Method.

isnull(A::ArrowVector, idx)

Check whether element(s) idx of A are null.

Arrow.nullcount — Method.

nullcount(A::ArrowVector)

Return the number of nulls (missings) in A.

Arrow.offsets — Method.

offsets(v::AbstractVector)

Construct a Vector{Int32} of offsets appropriate for data appearing in v.

Arrow.padding — Method.

padding(n::Integer)

Determines the total number of bytes needed to store n bytes with padding. Note that the Arrow standard requires buffers to be aligned to 8-byte boundaries.

Arrow.rawpadded — Method.

rawpadded(A::Primitive)

Get a Vector{UInt8} describing the data of A with Arrow padding (8 bytes).

Arrow.rawvalues — Method.

rawvalues(A::ArrowVector, idx)

Gets a Vector{UInt8} of the raw values associated with the indices idx.

Arrow.rawvalues — Method.

rawvalues(A::Primitive, i)
rawvalues(A::Primitive)

Gets the raw values for elements at locations i in the form of a Vector{UInt8}.

Arrow.totalbytes — Method.

totalbytes(A::AbstractVector)
totalbytes(::Type{Union{J,Missing}}, A::AbstractVector)
totalbytes(::Type{C}, A::AbstractVector)
totalbytes(::Type{Union{J,Missing}}, ::Type{C}, A::AbstractVector)

Computes the minimum number of bytes needed to store A as an Arrow formatted primitive array or list.

To obtain the minimum bytes to store string data, one must input C the character encoding type the string will be converted to (e.g. UInt8).

Arrow.unbitpack — Method.

unbitpack(A::AbstractVector{UInt8})

Returns a Vector{Bool} the values of which are the bits of A.

Arrow.valuesbytes — Method.

valuesbytes(A::AbstractVector)
valuesbytes(::Type{C}, A::AbstractVector{<:AbstractString})

Computes the number of bytes needed to store the values of A (without converting the underlying binary type). This does not include the number of bytes needed to store metadata such as a null bitmask or offsets.

To obtain the number of values bytes needed to string data, one must input C the character encoding type the string will be converted to (e.g. UInt8).

Arrow.writepadded — Method.

writepadded(io::IO, x)
writepadded(io::IO, A::Primitive)
writepadded(io::IO, A::Arrowvector, subbuffs::Function...)

Write the data x to io with 8-byte padding. This is commonly needed in Arrow implementations since Arrow requires 8-byte boundary alignment.

If a Primitive is provided, the appropriate padded values will be written.

If an ArrowVector is provided, the ordering of the sub-buffers must be specified, and they will be written in the order given. For example writepadded(io, A, bitmask, offsets, values) will write the bit mask, offsets and then values of A.

Arrow.writepadded — Method.

writepadded(io::IO, A::Primitive)

Write the values of A to io with Arrow padding (8 bytes).

Arrow.bitmaskpointer — Method.

bitmaskpointer(A::ArrowVector)

Returns a pointer to the start of the bitmask buffer for A.

Arrow.construct — Method.

construct(::Type{T}, A::AbstractPrimitive{J}, i::Integer, len::Integer)

Construct an object of type T from len values in A starting at index i. For this to work requires the existence of a constructor of the form T(::Vector{J}).

Arrow.datapointer — Method.

datapointer(A::Primitive)

Returns a pointer to the very start of the data buffer for A (i.e. does not depend on indices).

Arrow.ellength — Method.

ellength(l::AbstractList, i::Integer)

Get the length of element i.

Arrow.encode — Method.

encode(::Type{C}, v::AbstractVector{J})

Attempt to encode the data in v as a Vector{C}. This requires that if convert(Vector{C}, x) is valid where x ∈ v. Nothing is stored for cases where x == missing.

Arrow.getbit — Method.

getbit(A::Primitive, i)

Get the ith bit of the primitive. For example, for a Primitive{UInt8] the 10th bit is the second bit of the second value.

Arrow.getvalue — Method.

getvalue(A::ArrowVector, idx)

Get the values for indices idx from A.

Arrow.offsetspointer — Method.

offsetspointer(A::AbstractList)

Returns a pointer to the start of the offsets buffer for A.

Arrow.unsafe_construct — Method.

unsafe_construct(::Type{T}, A::Primitive, i::Integer, len::Integer)

Construct an object of type T using len elements from A starting at index i (1-based indexing). This is mostly used by AbstractList objects to construct variable length objects such as strings from primitive arrays.

Users must define new methods for new types T.

Arrow.unsafe_ellength — Method.

unsafe_ellength(l::AbstractList, i::Integer)

Get the length of element i. Involves calls to unsafe_load.

Arrow.unsafe_getoffset — Method.

unsafe_getoffset(l::AbstractList, i::Integer)

Get the offset for element i. Contains a call to unsafe_load.

Arrow.unsafe_getvalue — Method.

unsafe_getvalue(A::ArrowVector, i)

Retrieve the value from memory location i using Julia 1-based indexing. i can be a single integer index, an AbstractVector of integer indices, or an AbstractVector{Bool} mask.

This typically involves a call to unsafe_load or unsafe_wrap.

Arrow.unsafe_isnull — Method.

unsafe_isnull(A::ArrowVector, idx)

Check whether element(s) idx of A are null.

Arrow.unsafe_rawbitmask — Method.

unsafe_rawbitmask(p::ArrowVector{Union{J,Missing}})

Retrieve the raw value of the null bit mask for p.

Arrow.unsafe_rawoffsets — Method.

unsafe_rawoffsets(p::AbstractList)

Retreive the raw offstets for p as a Vector{UInt8}.

Arrow.unsafe_rawpadded — Method.

unsafe_rawpadded(ptr::Ptr, len::Integer, padding::Function=identity)

Return a Vector{UInt8} padded to appropriate size specified by padding.

Arrow.unsafe_rawpaddedvalues — Method.

unsafe_rawpaddedvalues(p::ArrowVector)

Retreive raw value data for p as a Vector{UInt8}.

Arrow.unsafe_setnull! — Method.

unsafe_setnull!(A::ArrowVector{Union{J,Missing}}, x::Bool, i::Integer)

Set element i of A to be null. This involves no bounds checking and a call to unsafe_store!.

Arrow.unsafe_setnulls! — Method.

unsafe_setnulls!(A::ArrowVector, nulls::AbstractVector{Bool})

Set all the nulls for the ArrowVector. This does not check bounds and contains a call to unsafe_copy! (but does not copy directly from nulls).

Arrow.unsafe_setoffset! — Method.

unsafe_setoffset!(l::AbstractList, off::Int32, i::Integer)

Set offset i to off. Contains a call to unsafe_store!.

Arrow.unsafe_setoffsets! — Method.

unsafe_setoffsets!(l::AbstractList, off::Vector{Int32})

Set all offsets to the Vector{Int32} off. Contains a call to unsafe_copy! which copies the entirety of off.

Arrow.unsafe_setvalue! — Method.

unsafe_setvalue!(A::ArrowVector{J}, x, i)

Set the value at location i to x. If i is a single integer, x should be an element of type J. Otherwise i can be an AbstractVector{<:Integer} or AbstractVector{Bool} in which case x should be an appropriately sized AbstractVector{J}.

Arrow.valuespointer — Method.

valuespointer(A::ArrowVector)

Returns a pointer to the start of the values buffer for A.