API

API documentation

Index

AsterReader.MEDFile

AsterReader.RMEDFile

AsterReader.aster_read_data

AsterReader.aster_read_mesh

AsterReader.aster_read_nodes

AsterReader.get_element_sets

AsterReader.get_node_sets

AsterReader.parse_node_id

AsterReader.to_ascii

AsterReader.aster_read_mesh — Function.

aster_read_mesh(filename, mesh_name=nothing)

Parse code aster .med file and return mesh data in a dictionary.

Dictionary contains additional dictionaries nodes, node_sets, elements, element_sets, element_types, surface_sets and surface_types.

If mesh file contains several meshes, one must provide the mesh name as additional argument or expcetion will be thrown.

AsterReader.MEDFile — Type.

Code Aster binary file (.med).

AsterReader.RMEDFile — Type.

Code Aster result file (.rmed).

AsterReader.aster_read_data — Method.

Read nodal field from rmed file.

AsterReader.aster_read_nodes — Method.

Return nodes from result med file.

AsterReader.get_element_sets — Method.

get_element_sets(med, mesh_name)

Return element sets from med file. Return type is a dictionary, where the key is the element set id number (integer) and value is a vector of strings, containing human-readable name for element set.

Notes

One element set id can have multiple names.

AsterReader.get_node_sets — Method.

Return node sets from med file.

Notes

One node set id can have multiple names.

AsterReader.parse_node_id — Method.

parse_node_id(node_name)

Return id number from node name. Usually the node name contains the id number, i.e. N123 => 123 and so on.

AsterReader.to_ascii — Method.

Convert vector of Int8 to ASCII string.