Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Atom.handle — Method.

handle(func, typ)

Set the handler for messages of type typ to func.

Atom.handlemsg — Method.

handlemsg(typ, args...)

Tries to call the message handler corresponding to typ (which can be set with handle).

Atom.initialise — Method.

initialise(; welcome = false)

Sets up the environment for Atom.jl: Stop SIGINTs from killing Julia and send a welcome message to Atom if welcome == true.

Atom.isactive — Method.

isactive(sock)

Check whether the socked sock exists and is open.

Atom.msg — Method.

msg(typ, args...)

Prints a json string to sock that contains all function arguments in an array.

Atom.rpc — Method.

rpc(typ, args...)

Sends a message to sock via msg. Blocks until the client returns a message (which needs to send back the callback field).

Atom.serve — Method.

serve(port; kws...)

Start an asynchronous Julia server that listens on localhost:port and handle the messages sent over that socket. The kws... are passed through to initialise.