Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Authorization.AbstractClient — Type.

Required fields:

id::String;

id2permission::Dict{String, Permission}; # Resource ID => Permission

idpattern2permission::Dict{Regex, Permission}; # Resource ID pattern => Permission

type2permission::Dict{DataType, Permission}; # Resource type => Permission

Authorization.AbstractResource — Type.

Required fields:

id::String

Authorization.create! — Method.

Create resource. If successful return nothing, else return an error message as a String.

Authorization.getpermission — Method.

Returns the Permission object for client-resource combination if it exists, else returns nothing.

Authorization.haspermission — Method.

Returns true if the client has permission to act on the resource according to the given action.

Action ust be one of :create, :read, :update, :delete.

Authorization.setexpiry! — Method.

Set the expiry of every permission of the client to the given expiry.

Authorization.update! — Method.

Update resource. If successful return nothing, else return an error message as a String.

Base.delete! — Method.

Delete resource. If successful return nothing, else return an error message as a String.

Base.read — Method.

Read resource. If successful return (true, value), else return (false, errormessage::String).

Authorization.permisssions_conflict — Method.

Returns true if more than one resource ID pattern matches the resource ID.