Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BKTrees.BKTree — Type.

The B-K tree structure.

BKTrees.Node — Type.

A node in the B-K tree.

BKTrees.add! — Method.

Adds a node (i.e. Node{T}(item::T)) to the tree.

BKTrees.find — Method.

Find items in the tree whose distance is less or equal to n and returns the top k items, ordered ascending according to the distance.

BKTrees.hamming_distance — Method.

Calculates the Hamming distance (numbers of different bits) between two numbers.

BKTrees.is_empty_node — Method.

Determines whether a Node in the BKTree is empty or not.