Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BarycentricInterpolation.degree — Method.

degree(poly)

Return the degree of the polynomial specified.

BarycentricInterpolation.differentiation_matrix — Method.

differentiation_matrix(poly::AbstractPolynomial)

Return the differentiation matrix at the nodes of the polynomial specified.

P = Chebyshev2{5}()
D = differentiation_matrix(P)

Now dy/dx ≈ D*y at the nodes of the polynomial.

BarycentricInterpolation.interpolate — Method.

interpolate(poly::AbstractPolynomial, y₀, [x])

Return the value of y(x) given that y(nodes(poly)) = y₀. If the value of x is not provided, return a function y(x) that evaluates the interpolant at any x.

BarycentricInterpolation.interpolation_matrix — Method.

interpolation_matrix(poly::AbstractPolynomial, x)

Return the interpolation matrix from the nodes of poly to the point(s) x. For example :

P = Chebyshev2{5}()
x = range(-1, stop=1, length=10)
M = interpolation_matrix(P, x)

Now y(x) ≈ M*y₀ given that `y(nodes(poly)) = y₀.

BarycentricInterpolation.nodes — Function.

nodes(poly)

Return the nodes for the specified orthogonal polynomials.

BarycentricInterpolation.weights — Function.

weights(poly)

Return the Barycentric weights for the specified orthogonal polynomials.