Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BayesOpt.ConfigParameters — Method.

ConfigParameters()

Returns default parameters of BayesOpt (see initializeparametersto_default in the C API).

BayesOpt.bayes_optimization — Method.

bayes_optimization(func, lb, ub, config)

Runs continuous Bayesian optimization on func that takes vectors of length d as argument and returns a real number, within the box defined by the lowerbounds lb (a vector of length d with lowerbounds for each dimension) and upperbounds ub, using config (see ConfigParameters).

BayesOpt.bayes_optimization_categorical — Method.

bayes_optimization_categorical(func, categories, config)

Runs Bayesian optimization on func that takes vectors of length d as argument and returns a real number, with categories array of size d with the number of categories per dimension, using config (see ConfigParameters).

BayesOpt.bayes_optimization_disc — Method.

bayes_optimization_disc(func, valid_x, config)

Runs Bayesian optimization on func that takes vectors of length d as argument and returns a real number, on valid_x an array of valid points (vectors of length d) using config (see ConfigParameters).

BayesOpt.set_criteria! — Function.

set_criteria!(config, name) set criteria in config to name.

BayesOpt.set_kernel! — Function.

set_kernel!(config, name) set kernel in config to name.

BayesOpt.set_learning! — Function.

set_learning!(config, name) set learning in config to name.

BayesOpt.set_load_file! — Function.

set_load_file!(config, name) set load file in config to name.

BayesOpt.set_log_file! — Function.

set_log_file!(config, name) set log file in config to name.

BayesOpt.set_mean! — Function.

set_mean!(config, name) set mean in config to name.

BayesOpt.set_save_file! — Function.

set_save_file!(config, name) set save file in config to name.

BayesOpt.set_score! — Function.

set_score!(config, name) set score in config to name.

BayesOpt.set_surrogate! — Function.

set_surrogate!(config, name) set surrogate in config to name.