Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BayesianOptimization.BOpt — Method.

BOpt(func, model, acquisition, modeloptimizer, lowerbounds, upperbounds;
          sense = Max, maxiterations = 10^4, maxduration = Inf,
          acquisitionoptions = NamedTuple(), repetitions = 1,
          verbosity = Progress, lhs_iterations = 5*length(lowerbounds))

BayesianOptimization.BrochuBetaScaling — Type.

Scales βt of UpperConfidenceBound as

βt = √(2 * log(t^(D/2 + 2) * π^2/(3δ)))

where t is the number of observations, D is the dimensionality of the input data points and δ is a small constant (default δ = 0.1).

See Brochu E., Cora V. M., de Freitas N. (2010), "A Tutorial on Bayesian Optimization of Expensive Cost Functions, with Application to Active User Modeling and Hierarchical Reinforcement Learning", https://arxiv.org/abs/1012.2599v1 page 16.

BayesianOptimization.ExpectedImprovement — Type.

The expected improvement measures the expected improvement x - τ of a point x upon an incumbent target τ. For Gaussian distributions it is given by

(μ(x) - τ) * ϕ[(μ(x) - τ)/σ(x)] + σ(x) * Φ[(μ(x) - τ)/σ(x)]

where ϕ is the standard normal distribution function and Φ is the standard normal cumulative function, and μ(x), σ(x) are mean and standard deviation of the distribution at point x.

BayesianOptimization.MLGPOptimizer — Method.

MLGPOptimizer(; every = 10, kwargs...)

Set the GP hyperparameters to the maximum likelihood estimate every number of steps.

BayesianOptimization.MutualInformation — Type.

The mutual information measures the amount of information gained by querying at x. The parameter γ̂ gives a lower bound for the information on f from the queries {x}. For a Gaussian this is γ̂ = ∑σ²(x) and the mutual information at x is μ(x) + √(α)*(√(σ²(x)+γ̂) - √(γ̂))

where μ(x), σ(x) are mean and standard deviation of the distribution at point x.

See Contal E., Perchet V., Vayatis N. (2014), "Gaussian Process Optimization with Mutual Information" http://proceedings.mlr.press/v32/contal14.pdf

BayesianOptimization.NoBetaScaling — Type.

Applies no scaling to βt of UpperConfidenceBound.

BayesianOptimization.ProbabilityOfImprovement — Type.

The probability of improvement measures the probability that a point x leads to an improvement upon an incumbent target τ. For Gaussian distributions it is given by

Φ[(μ(x) - τ)/σ(x)]

where Φ is the standard normal cumulative distribution function and μ(x), σ(x) are mean and standard deviation of the distribution at point x.

BayesianOptimization.ThompsonSamplingSimple — Type.

The acquisition function associated with ThompsonSamplingSimple draws independent samples for each input x a function value from the model. Together with a gradient-free optimization method this leads to proposal points that might be similarly distributed as the maxima of true Thompson samples from GPs. True Thompson samples from a GP are simply functions from a GP. Maximizing these samples can be tricky, see e.g. http://hildobijl.com/Downloads/GPRT.pdf chapter 6.

BayesianOptimization.UpperConfidenceBound — Type.

For Gaussian distributions the upper confidence bound at x is given by μ(x) + βt * σ(x)

where βt is a fixed parameter in the case of NoBetaScaling or an observation size dependent parameter in the case of e.g. BrochuBetaScaling.

BayesianOptimization.UpperConfidenceBound — Method.

UpperConfidenceBound(; scaling = BrochuBetaScaling(.1), βt = 1)

BayesianOptimization.boptimize! — Method.

boptimize!(o::BOpt)

BayesianOptimization.NoModelOptimizer — Type.

Don't optimize the model ever.