Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BeliefUpdaters.DiscreteBelief — Type.

DiscreteBelief

A belief specified by a probability vector.

Normalization of b is NOT enforced at all times, but the DiscreteBeleif(pomdp, b) constructor will warn, and update(...) always returns a belief with normalized b.

BeliefUpdaters.KMarkovUpdater — Type.

Updater that stores the k most recent observations as the belief.

BeliefUpdaters.PreviousObservationUpdater — Type.

Updater that stores the most recent observation as the belief, missing if none is available.

BeliefUpdaters.uniform_belief — Method.

Return a DiscreteBelief with equal probability for each state.