Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BenchmarkProfiles.data_profile — Function.

Produce a data profile.

The 3-dimensional array H gives the performance data for each solver and each problem (smaller is better). H[k,p,s] is the k-th costly operation (e.g., function evaluation) for problem p and solver s. Failures on a given problem are represented by a negative value, an infinite value, or NaN.

The vector N gives a scaling associated to each problem. If the number of simplex gradients is being measured, N[p] should be n(p) + 1 where n(p) is the number of variables of problem p.

The threshold τ determines the tolerance in the convergence criterion

f(x) ≤ fL + τ (f0 - fL),

where for each problem f(x) is a measure recorded (e.g., the objective value), f0 is the measure at the initial point, and fL is the best measure obtained by any solver.

BenchmarkProfiles.data_ratios — Method.

Compute data ratios used to produce a data profile.

There is normally no need to call this function directly. See the documentation of data_profile() for more information.

BenchmarkProfiles.performance_profile — Method.

Produce a performance profile.

Each column of the matrix T defines the performance data for a solver (smaller is better). Failures on a given problem are represented by a negative value, an infinite value, or NaN. The optional argument logscale is used to produce a logarithmic (base 2) performance plot.

BenchmarkProfiles.performance_ratios — Method.

Compute performance ratios used to produce a performance profile.

There is normally no need to call this function directly. See the documentation of performance_profile() for more information.