Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BigArrays.BigArray — Type.

BigArray

currently, assume that the array dimension (x,y,z,...) is >= 3 all the manipulation effects in the x,y,z dimension

BigArrays.BigArray — Method.

BigArray(layerPath::AbstractString; fillMissing=true, mod::Symbol=DEFAULT_MODE)

BigArrays.BigArray — Method.

BigArray( d::AbstractBigArrayBackend, info::Info;
              mip::Int = 0,
              fillMissing::Bool=DEFAULT_FILL_MISSING,
              mode::Symbol=DEFAULT_MODE)

Parameters: d: the bigarray storage backend info: the info containing the metadata mip: mip level. 0 is the highest resolution. the mip level is like a image pyramid with difference downsampled levels. fillMissing: whether fill the missing blocks in the storage backend with zeros or not. mode: the io mode with options in {multithreading, sequential, multiprocesses, sharedarray}

BigArrays.BigArray — Method.

BigArray(info::Info; fillMissing::Bool=DEFAULT_FILL_MISSING, mode=DEFAULT_MODE)

create a new directory with random name. this function was designed for test and benchmark. we need another function the clear the whole array

Base.getindex — Method.

Base.getindex( ba::BigArray{D, T, N, C}, idxes::Union{UnitRange, Int}...) where {D,T,N,C}

get index with different modes: sharedarray, multi_processes, multithreads, sequential

Base.setindex! — Method.

Base.setindex!( ba::BigArray{D,T,N,C}, buf::Array{T,N},

setindex with different mode: sharedarray, multithreads, multiprocesses, sequential

BigArrays.adjust_volume_boundary — Method.

adjust the global and buffer range according to total volume size. shrink the range stop if the ranges passes the volume boundary.

BigArrays.get_num_chunks — Method.

get_num_chunks(ba::BigArray, idxes::Union{UnitRange,Int}...)

get number of chunks needed to do cutout from this range

BigArrays.getindex_sequential — Method.

get_index_sequential(ba::BigArray, idxes::Union{UnitRange, Int}...)

sequential implementation for debuging

BigArrays.list_missing_chunks — Method.

list_missing_chunks(ba::BigArray, idxes::Union{UnitRange, Int}...)

list the non-existing keys in the index range if the returned list is empty, then all the chunks exist in the storage backend.

BigArrays.setindex_multiprocesses! — Method.

put array in RAM to a BigArray

this version uses channel to control the number of asynchronized request

BigArrays.setindex_multithreads! — Method.

put array in RAM to a BigArray backend

this version uses channel to control the number of asynchronized request

BigArrays.setindex_sequential! — Method.

sequential function, good for debuging

BigArrays.setindex_sharedarray! — Method.

put array in RAM to a BigArray

this version uses channel to control the number of asynchronized request