Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Bio3DView.Axes — Type.

Axes(length, radius)
Axes()

Styling for visualisation of the coordinate axes. Arguments are numbers for the length and the radius of the arrows. The optional keyword argument colors is a list of 3 colors as Strings for the 3 axes.

Bio3DView.Box — Type.

Box(center, dimensions)

Data and styling for a box visualisation. Arguments are a Vector{Float64} of the center coordinates and a Vector{Float64} of the box size in each dimension. Optional keyword arguments are color and wireframe.

Bio3DView.CameraAngle — Type.

CameraAngle(posx, posy, posz, zoom, qx, qy, qz, qw)

A custom perspective to view the molecule from. Arguments are x/y/z translation, zoom, and x/y/z/w rotation quaternion. Default is CameraAngle(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0).

Bio3DView.IsoSurface — Type.

IsoSurface(voldata, isoval)

Data and styling for an isosurface visualisation. Arguments are the filepath with the volume data in "cube" format and the value to view the isosurface at. Optional keyword arguments are color, opacity, wireframe and smoothness.

Bio3DView.Style — Type.

Style(style_type)
Style(style_type, options)

A style for a molecular visualisation, with an optional Dict of options. Examples are Style("cartoon") and Style("cartoon", Dict("color"=> "spectrum", "ribbon"=> true, "thickness"=> 1.0)).

Bio3DView.Surface — Type.

Surface()
Surface(options)

A style for a molecular VDW surface visualisation, with an optional Dict of options. An example is Surface(Dict("opacity"=> 0.8, "colorscheme"=> "whiteCarbon")).

Bio3DView.viewfile — Function.

viewfile(file)
viewfile(file, format)

View a molecular structure from a file. Displays in a popup window, or in the output cell for an IJulia notebook. Arguments are the filepath and the format ("pdb", "sdf", "xyz" or "mol2"). If not provided, the format is guessed from the file extension, e.g. "myfile.xyz" is treated as being in the xyz format. Optional keyword arguments are style, surface, isosurface, box, vtkcell, axes, cameraangle, height, width, html and debug.

Bio3DView.viewpdb — Method.

viewpdb(pdbid)

View a structure from the Protein Data Bank (PDB). Displays in a popup window, or in the output cell for an IJulia notebook. Argument is the four letter PDB ID, e.g. "1AKE". Requires an internet connection to work. Optional keyword arguments are style, surface, isosurface, box, vtkcell, axes, cameraangle, height, width, html and debug.

Bio3DView.viewstring — Method.

viewstring(str, format)

View a molecular structure contained in a string. Displays in a popup window, or in the output cell for an IJulia notebook. Arguments are the molecule string and the format ("pdb", "sdf", "xyz" or "mol2"). Optional keyword arguments are style, surface, isosurface, box, vtkcell, axes, cameraangle, height, width, html and debug.