Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Blobs.Blob — Type.

A pointer to a T stored inside a Blob.

Blobs.BlobBitVector — Type.

A fixed-length bit vector whose data is stored in a Blob.

Blobs.BlobString — Type.

A string whose data is stored in a Blob.

Blobs.BlobVector — Type.

A fixed-length vector whose data is stored in a Blob.

Blobs.child_size — Method.

child_size(::Type{T}, args...) where {T}

The number of bytes needed to allocate children of T, not including self_size(T).

Defaults to 0.

Blobs.free — Method.

free(blob::Blob)

Free the underlying allocation for blob.

Blobs.init — Method.

init(blob::Blob{T}, free::Blob{Nothing}, args...)::Blob{Nothing} where T

Initialize blob, where free is the beginning of the remaining free space. Must return free + child_size(T, args...).

The default implementation where child_size(T) == 0 does nothing. Override this method to add custom initializers for your types.

Blobs.init — Method.

init(blob::Blob{T}, args...) where T

Initialize blob.

Assumes that blob it at least self_size(T) + child_size(T, args...) bytes long.

Blobs.malloc — Method.

malloc(::Type{T}, args...)::Blob{T} where T

Allocate an uninitialized Blob{T}.

Blobs.malloc_and_init — Method.

malloc_and_init(::Type{T}, args...)::Blob{T} where T

Allocate and initialize a new Blob{T}.

Blobs.self_size — Method.

self_size(::Type{T}, args...) where {T}

The number of bytes needed to allocate T itself.

Defaults to sizeof(T).