Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

BusinessDays.CompositeHolidayCalendar — Type.

Allows for combination of several Holiday Calendars.

BusinessDays.GenericHolidayCalendar — Type.

GenericHolidayCalendar

holidays: a set of holiday dates

dtmin: minimum date allowed to check for holidays in holidays set. Defaults to min(holidays...).

dtmax: maximum date allowed to check for holidays in holidays set. Defaults to max(holidays...).

cache: instance of HolidayCalendarCache.

BusinessDays.GenericHolidayCalendar — Type.

GenericHolidayCalendar(holidays, [dtmin], [dtmax], [_initcache_])

holidays: a set of holiday dates

dtmin: minimum date allowed to check for holidays in holidays set. Defaults to min(holidays...).

dtmax: maximum date allowed to check for holidays in holidays set. Defaults to max(holidays...).

_initcache_: initializes the cache for this calendar. Defaults to true.

BusinessDays.HolidayCalendar — Type.

Abstract type for Holiday Calendars.

BusinessDays.advancebdays — Method.

advancebdays(calendar, dt, bdays_count)

Increments given date dt by bdays_count. Decrements it if bdays_count is negative. bdays_count can be a Int, Vector{Int} or a UnitRange.

Computation starts by next Business Day if dt is not a Business Day.

BusinessDays.bdays — Method.

bdays(calendar, dt0, dt1) :: Dates.Day

Counts the number of Business Days between dt0 and dt1. Returns instances of Dates.Day.

Computation is always based on next Business Day if given dates are not Business Days.

BusinessDays.bdayscount — Method.

bdayscount(calendar, dt0, dt1) :: Int

Counts the number of Business Days between dt0 and dt1. Returns Int.

Computation is always based on next Business Day if given dates are not Business Days.

BusinessDays.firstbdayofmonth — Method.

firstbdayofmonth(calendar, dt)
firstbdayofmonth(calendar, yy, mm)

Returns the first business day of month.

BusinessDays.isbday — Method.

isbday(calendar, dt)

Returns false for weekends or holidays. Returns true otherwise.

BusinessDays.isholiday — Method.

isholiday(calendar, dt)

Checks if dt is a holiday based on a given calendar of holidays.

calendar can be an instance of HolidayCalendar, a Symbol or an AbstractString.

Returns boolean values.

BusinessDays.isweekday — Method.

isweekday(dt)

Returns true for Monday to Friday. Returns false otherwise.

BusinessDays.isweekend — Method.

isweekend(dt)

Returns true for Saturdays or Sundays. Returns false otherwise.

BusinessDays.lastbdayofmonth — Method.

lastbdayofmonth(calendar, dt)
lastbdayofmonth(calendar, yy, mm)

Returns the last business day of month.

BusinessDays.listbdays — Method.

listbdays(calendar, dt0::Dates.Date, dt1::Dates.Date) → Vector{Dates.Date}

Returns the list of business days between dt0 and dt1.

BusinessDays.listholidays — Method.

listholidays(calendar, dt0::Dates.Date, dt1::Dates.Date) → Vector{Dates.Date}

Returns the list of holidays between dt0 and dt1.

BusinessDays.tobday — Method.

tobday(calendar, dt; [forward=true])

Adjusts dt to next Business Day if it's not a Business Day. If isbday(dt), returns dt.

BusinessDays.Australia — Type.

Public holidays for the Australian states and territories. Although some holidays are common to all states and territories, such as Christmas Day, each state and territory also has its own additional holidays. Therefore the set of relevant holidays depends on which state/territory you are concerned with.

The Australian states and territories are:

Australian Capital Territory (ACT)

New South Wales (NSW)

Northern Territory (NT)

Queensland (QLD)

South Australia (SA)

Tasmania (TAS)

Western Australia (WA)

Victoria (VIC)

For example: cal = Australia(:QLD)

BusinessDays.AustraliaASX — Type.

Public holidays for the Australian Stock Exchange (ASX).

BusinessDays.BRSettlement — Type.

Banking holidays for Brazil (federal holidays plus Carnival).

BusinessDays.BrazilExchange — Type.

BMF&BOVESPA Exchange holidays (http://www.bmfbovespa.com.br).

BusinessDays.CanadaSettlement — Type.

Holidays for Canada

BusinessDays.CanadaTSX — Type.

Holidays for Toronto Stock Exchange

BusinessDays.HolidayCalendarCache — Type.

Data structure for calendar cache.

BusinessDays.HolidayCalendarCache — Method.

HolidayCalendarCache()

creates an empty instance of HolidayCalendarCache

BusinessDays.NullHolidayCalendar — Type.

A calendar with no holidays and no weekends. bdays returns the actual days between dates (dt1 - d10).

BusinessDays.TARGET — Type.

Holidays for TARGET Eurozone (Trans-European Automated Real-time Gross Settlement Express Transfer System)

BusinessDays.UKSettlement — Type.

Banking holidays for England and Wales.

BusinessDays.USGovernmentBond — Type.

United States Government Bond calendar.

BusinessDays.USNYSE — Type.

United States NYSE holidays.

BusinessDays.USSettlement — Type.

United States federal holidays.

BusinessDays.WeekendsOnly — Type.

A calendar with no holidays. But account for weekends.

isholiday always returns false for this calendar. Remember that isbday considers that Saturdays and Sundars are not business days.

BusinessDays.adjustweekendholidayPost — Method.

adjustweekendholidayPost(dt, [adjust_saturdays]) → Date

In the UK and Canada, if a holiday falls on Saturday or Sunday, it's observed on the next business day. This function will adjust to the next Monday.

adjust_saturdays kwarg defaults to true.

BusinessDays.adjustweekendholidayUS — Method.

adjustweekendholidayUS(dt) → Date

In the United States, if a holiday falls on Saturday, it's observed on the preceding Friday. If it falls on Sunday, it's observed on the next Monday.

BusinessDays.cleancache — Method.

cleancache([calendar])

Cleans cache for a given instance or list of HolidayCalendar, Symbol or AbstractString.

BusinessDays.easter_date — Method.

easter_date(y::Dates.Year) → Dates.Date

Returns result of easter_rata as a Dates.Date instance.

BusinessDays.easter_rata — Method.

easter_rata(y::Dates.Year) → Int

Returns Easter date as a Rata Die number.

Based on Algo R.

BusinessDays.findweekday — Method.

findweekday(weekday_target::Integer, yy::Integer, mm::Integer, occurrence::Integer, ascending::Bool) → Date

Given a year yy and month mm, finds a date where a choosen weekday occurs.

weekday_target values are declared in module Base.Dates: Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday = 1,2,3,4,5,6,7.

If ascending is true, searches from the beginning of the month. If false, searches from the end of the month.

If occurrence is 2 and weekday_target is Monday, searches the 2nd Monday of the given month, and so on.

BusinessDays.initcache — Function.

initcache(calendar, [d0], [d1])

Creates cache for a given Holiday Calendar. After calling this function, any call to isbday function, or any function that uses isbday, will be optimized to use this cache.

You can pass calendar as an instance of HolidayCalendar, Symbol or AbstractString. You can also pass calendar as an AbstractArray of those types.