Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

COSMO.assemble! — Method.

assemble!(model, P, q, constraint(s), [settings, x0, y0])

Assembles a COSMO.Model with a cost function defind by P and q, and a number of constraints.

The positive semidefinite matrix P and vector q are used to specify the cost function of the optimization problem:

min   1/2 x'Px + q'x
s.t.  Ax + b ∈ C

constraints is a COSMO.Constraint or an array of COSMO.Constraint objects that are used to describe the constraints on x.

The optinal arguments x0 and y0 can be used to provide the solver with warm starting values for the primal variable x and the dual variable y. The optinal argument settings can be used to pass custom solver settings.

COSMO.optimize! — Method.

optimize!(model)

Attempts to solve the optimization problem defined in COSMO.Model object with the user settings defined in COSMO.Settings. Returns a COSMO.Result object.

COSMO.Constraint — Type.

Constraint(A, b, convex_set, dim = 0, indices = 0:0)

Creates a COSMO constraint: Ax + b ∈ convex_set.

By default the following convex sets are supported: ZeroSet, Nonnegatives, SecondOrderCone, PsdCone.

Examples

julia> Constraint([1 0;0 1], zeros(2), COSMO.PsdCone)
Constraint
Size of A: (2, 2)
ConvexSet: COSMO.PsdCone

The optinal arguments dim and indices can be used to specify A and b for subparts of variable x. If x has dimension dim = 4, then x[2] and x[3] can be constrained to the zero cone in the following way:

Examples

julia> c = Constraint([1 0;0 1], zeros(2), COSMO.ZeroSet, 4, 2:3)
Constraint
Size of A: (2, 4)
ConvexSet: COSMO.ZeroSet

Notice that extra columns of A have been added automatically.

julia>Matrix(c.A)
2×4 Array{Float64,2}:
0.0  1.0  0.0  0.0
0.0  0.0  1.0  0.0

COSMO.Result — Type.

Result{T <: AbstractFloat}

Object returned by the COSMO solver after calling optimize!(model, settings). It has the following fields:

Fieldname	Type	Description

x	Vector{T}	Primal variable

y	Vector{T}	Dual variable

s	Vector{T}	(Primal) set variable

obj_val	T	Objective value

iter	Int64	Number of iterations

status	Symbol	Solution status

info	COSMO.ResultInfo	Struct with more information

times	COSMO.ResultTimes	Struct with several measured times

COSMO.Settings — Type.

Settings(;arg=val)

Creates a COSMO settings object that is used to pass user settings to the solver.

Argument	Description	Values (default)

rho	ADMM rho step	0.1

sigma	ADMM sigma step	1e-6.

alpha	Relaxation parameter	1.6

eps_abs	Absolute residual tolerance	1e-4

eps_rel	Relative residual tolerance	1e-4

epspriminf	Primal infeasibility tolerance	1e-4

epsdualinf	Dual infeasibility tolerance	1e-4

max_iter	Maximum number of iterations	2500

verbose	Verbose printing	false

verbose_timing	Verbose timing	false

check_termination	Check termination interval	40

check_infeasibility	Check infeasibility interval	40

scaling	Number of scaling iterations	10

adaptive_rho	Automatic adaptation of step size parameter	true

time_limit	set solver time limit in s	0

COSMO.Workspace — Type.

Model()

Initializes an empty COSMO model that can be filled with problem data using assemble!(model, P, q,constraints, [settings]).

COSMO.set! — Method.

set!(model, P, q, A, b, convex_sets, [settings])

Sets model data directly based on provided fields.

COSMO.warm_start! — Method.

warm_start!(model, [x0, y0])

Provides the COSMO.Model with warm starting values for the primal variable x and/or the dual variable y.