Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

CSTParser.parse_expression — Method.

parse_expression(ps)

Parses an expression until closer(ps) == true. Expects to enter the ParseState the token before the the beginning of the expression and ends on the last token.

Acceptable starting tokens are:

A keyword

An opening parentheses or brace.

An operator.

An instance (e.g. identifier, number, etc.)

An @.

CSTParser.check_reformat — Method.

check_reformat()

Reads and parses all files in current directory, applys formatting fixes and checks that the output AST remains the same.

CSTParser.check_span — Function.

check_span(x, neq = [])

Recursively checks whether the span of an expression equals the sum of the span of its components. Returns a vector of failing expressions.

CSTParser.compare — Method.

compare(x,y)

Recursively checks whether two Base.Expr are the same. Returns unequal sub- expressions.

CSTParser.contributes_scope — Method.

contributes_scope(x)

Checks whether the body of x is included in the toplevel namespace.

CSTParser.get_id — Method.

get_id(x)

Get the IDENTIFIER name of a variable, possibly in the presence of type declaration operators.

CSTParser.get_sig — Method.

get_sig(x)

Returns the full signature of function, macro and datatype definitions. Should only be called when has_sig(x) == true.

CSTParser.get_t — Method.

get_t(x)

Basic inference in the presence of type declarations.

CSTParser.lex_ws_comment — Method.

lex_ws_comment(l::Lexer, c)

Having hit an initial whitespace/comment/semicolon continues collecting similar Chars until they end. Returns a WS token with an indication of newlines/ semicolons. Indicating a semicolons takes precedence over line breaks as the former is equivalent to the former in most cases.

CSTParser.parse — Function.

parse(str, cont = false)

Parses the passed string. If cont is true then will continue parsing until the end of the string returning the resulting expressions in a TOPLEVEL block.

CSTParser.parse_array — Method.

parse_array(ps)

Having hit '[' return either:

A vect

A vcat

A comprehension

An array (vcat of hcats)

CSTParser.parse_call — Function.

parse_call(ps, ret)

Parses a function call. Expects to start before the opening parentheses and is passed the expression declaring the function name, ret.

CSTParser.parse_curly — Method.

parse_curly(ps, ret)

Parses the juxtaposition of ret with an opening brace. Parses a comma seperated list.

CSTParser.parse_generator — Method.

parse_generator(ps)

Having hit for not at the beginning of an expression return a generator. Comprehensions are parsed as SQUAREs containing a generator.

CSTParser.parse_if — Function.

parse_if(ps, ret, nested=false, puncs=[])

Parse an if block.

CSTParser.parse_macrocall — Method.

parse_macrocall(ps)

Parses a macro call. Expects to start on the @.

CSTParser.parse_paren — Method.

parse_paren(ps, ret)

Parses an expression starting with a (.

CSTParser.parse_ref — Method.

parse_ref(ps, ret)

Handles cases where an expression - ret - is followed by [. Parses the following bracketed expression and modifies it's .head appropriately.

CSTParser.parse_string_or_cmd — Function.

parsestringor_cmd(ps)

When trying to make an INSTANCE from a string token we must check for interpolating opoerators.

CSTParser.parse_tuple — Method.

parse_tuple(ps, ret)

ret is followed by a comma so tries to parse the rest of the tuple.

CSTParser.parse_unary — Method.

parse_unary(ps)

Having hit a unary operator at the start of an expression return a call.

CSTParser.remlineinfo! — Method.

remlineinfo!(x)

Removes line info expressions. (i.e. Expr(:line, 1))