Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

CUDAapi.find_binary — Method.

find_binary(names; locations=String[])

Similar to find_library, performs an exhaustive search for a binary in various subdirectories of locations, and finally PATH.

CUDAapi.find_libdevice — Method.

find_libdevice(targets::Vector{VersionNumber}, toolkit_dirs::Vector{String})

Look for the CUDA device library supporting targets in any of the CUDA toolkit directories toolkit_dirs. On CUDA >= 9.0, a single library unified library is discovered and returned as a string. On older toolkits, individual libraries for each of the targets are returned as a vector of strings.

CUDAapi.find_library — Method.

find_library(names; locations=String[], versions=VersionNumber[], word_size=Sys.WORD_SIZE)

Wrapper for Libdl.find_library, performing a more exhaustive search:

variants of the library name (including version numbers, platform-specific tags, etc);

various subdirectories of the locations list, and finally system library directories.

Returns the full path to the library.

CUDAapi.find_toolkit — Method.

find_toolkit()::Vector{String}

Look for directories where (parts of) the CUDA toolkit might be installed. This returns a (possibly empty) list of paths that can be used as an argument to other discovery functions.

The behavior of this function can be overridden by defining the CUDA_PATH, CUDA_HOME or CUDA_ROOT environment variables, which should point to the root of the CUDA toolkit.