API wrappers

API wrappers

This section lists the package's public functionality that directly corresponds to functionality of the CUDA driver API. In general, the abstractions stay close to those of the CUDA driver API, so for more information on certain library calls you can consult the CUDA driver API reference.

The documentation is grouped according to the modules of the driver API.

Installation properties

CUDAdrv.vendor — Function.

Returns a string identifying the vendor of your CUDA driver.

Initialization

CUDAdrv.init — Function.

Initialize the CUDA driver API.

This function is automatically called upon loading the package. You should not have to call this manually.

Error Handling

CUDAdrv.CuError — Type.

CuError(code::Integer)
CuError(code::Integer, info::String)

Create a CUDA error object with error code code. The optional info parameter indicates whether extra information, such as error logs, is known.

CUDAdrv.name — Method.

name(err::CuError)

Gets the string representation of an error code.

This name can often be used as a symbol in source code to get an instance of this error. For example:

julia> using CUDAdrv

julia> err = CuError(1)
CuError(1, ERROR_INVALID_VALUE)

julia> name(err)
"ERROR_INVALID_VALUE"

julia> CUDAdrv.ERROR_INVALID_VALUE
CuError(1, ERROR_INVALID_VALUE)

CUDAdrv.description — Method.

description(err::CuError)

Gets the string description of an error code.

Version Management

CUDAdrv.version — Method.

version()

Returns the CUDA version as reported by the driver.

Device Management

CUDAdrv.CuDevice — Type.

CuDevice(i::Integer)

Get a handle to a compute device.

CUDAdrv.devices — Function.

devices()

Get an iterator for the compute devices.

CUDAdrv.name — Method.

name(dev::CuDevice)

Returns an identifier string for the device.

CUDAdrv.totalmem — Method.

totalmem(dev::CuDevice)

Returns the total amount of memory (in bytes) on the device.

CUDAdrv.attribute — Function.

attribute(dev::CuDevice, code)

Returns information about the device.

Certain common attributes are exposed by additional convenience functions:

CUDAdrv.capability — Method.

capability(dev::CuDevice)

Returns the compute capability of the device.

CUDAdrv.warpsize — Method.

warpsize(dev::CuDevice)

Returns the warp size (in threads) of the device.

Context Management

CUDAdrv.CuContext — Type.

CuContext(dev::CuDevice, flags::CUctx_flags=SCHED_AUTO)
CuContext(f::Function, ...)

Create a CUDA context for device. A context on the GPU is analogous to a process on the CPU, with its own distinct address space and allocated resources. When a context is destroyed, the system cleans up the resources allocated to it.

Contexts are unique instances which need to be destroyed after use. For automatic management, prefer the do block syntax, which implicitly calls destroy.

CUDAdrv.destroy! — Method.

destroy!(ctx::CuContext)

Mark a context for destruction.

This does not immediately destroy the context, as there might still be dependent resources which have not been collected yet. The context will get freed as soon as all outstanding instances have been finalized.

CUDAdrv.CuCurrentContext — Function.

CuCurrentContext()

Return the current context, or nothing if there is no active context.

CUDAdrv.activate — Method.

activate(ctx::CuContext)

Binds the specified CUDA context to the calling CPU thread.

CUDAdrv.synchronize — Method.

synchronize(ctx::CuContext=CuCurrentContext())

Block for a context's tasks to complete.

The ctx parameter defaults to the current active context.

CUDAdrv.device — Method.

device()
device(ctx::Cucontext)

Returns the device for the current context. The optional ctx parameter is to make sure that the current context is really active, and hence the returned device is valid.

Primary Context Management

CUDAdrv.CuPrimaryContext — Type.

CuPrimaryContext(dev::CuDevice)

Create a primary CUDA context for a given device.

Each primary context is unique per device and is shared with CUDA runtime API. It is meant for interoperability with (applications using) the runtime API.

CUDAdrv.CuContext — Method.

CuContext(pctx::CuPrimaryContext)
CuContext(f::Function, pctx::CuPrimaryContext)

Retain the primary context on the GPU, returning a context compatible with the driver API. The primary context will be released when the returned driver context is finalized. For that reason, it is advised to use this function with do block syntax.

CUDAdrv.isactive — Method.

isactive(pctx::CuPrimaryContext)

Query whether a primary context is active.

CUDAdrv.flags — Method.

flags(pctx::CuPrimaryContext)

Query the flags of a primary context.

CUDAdrv.setflags! — Method.

setflags!(pctx::CuPrimaryContext)

Set the flags of a primary context.

CUDAdrv.unsafe_reset! — Method.

unsafe_reset!(pctx::CuPrimaryContext, [checked::Bool=true])

Explicitly destroys and cleans up all resources associated with a device's primary context in the current process. Note that this forcibly invalidates all contexts derived from this primary context, and as a result outstanding resources might become invalid.

It is normally unnecessary to call this function, as resource are automatically freed when contexts go out of scope. In the case of primary contexts, they are collected when all contexts derived from that primary context have gone out of scope.

The checked argument determines whether to verify that the primary context has become inactive after resetting the derived driver contexts. This may not be possible, eg. if the CUDA runtime API itself has retained an additional context instance.

Module Management

CUDAdrv.CuModule — Type.

CuModule(data, options::Dict{CUjit_option,Any})
CuModuleFile(path, options::Dict{CUjit_option,Any})

Create a CUDA module from a data, or a file containing data. The data may be PTX code, a CUBIN, or a FATBIN.

The options is an optional dictionary of JIT options and their respective value.

Function Management

CUDAdrv.CuFunction — Type.

CuFunction(mod::CuModule, name::String)

Acquires a function handle from a named function in a module.

Global Variable Management

CUDAdrv.CuGlobal — Type.

CuGlobal{T}(mod::CuModule, name::String)

Acquires a typed global variable handle from a named global in a module.

Base.eltype — Method.

eltype(var::CuGlobal)

Return the element type of a global variable object.

Base.get — Method.

get(var::CuGlobal)

Return the current value of a global variable.

CUDAdrv.set — Method.

set(var::CuGlobal{T}, T)

Set the value of a global variable to val

Linker

CUDAdrv.CuLink — Type.

CuLink()

Creates a pending JIT linker invocation.

CUDAdrv.add_data! — Function.

add_data!(link::CuLink, name::String, code::String)

Add PTX code to a pending link operation.

add_data!(link::CuLink, name::String, data::Vector{UInt8}, type::CUjit_input)

Add object code to a pending link operation.

CUDAdrv.add_file! — Function.

add_file!(link::CuLink, path::String, typ::CUjit_input)

Add data from a file to a link operation. The argument typ indicates the type of the contained data.

CUDAdrv.CuLinkImage — Type.

The result of a linking operation.

This object keeps its parent linker object alive, as destroying a linker destroys linked images too.

CUDAdrv.complete — Function.

complete(link::CuLink)

Complete a pending linker invocation, returning an output image.

CUDAdrv.CuModule — Method.

CuModule(img::CuLinkImage, ...)
CuModule(f::Function, img::CuLinkImage, ...)

Create a CUDA module from a completed linking operation. Options from CuModule apply.

Memory Management

CUDAdrv.Mem.alloc — Function.

Allocate linear memory on the device and return a buffer to the allocated memory. The allocated memory is suitably aligned for any kind of variable. The memory will not be freed automatically, use free(::Buffer) for that.

CUDAdrv.Mem.free — Function.

Free device memory.

CUDAdrv.Mem.set! — Function.

Initialize device memory with a repeating value.

CUDAdrv.Mem.upload — Function.

Upload memory from host to device. Executed asynchronously on stream if async is true.

CUDAdrv.Mem.upload! — Function.

Upload memory from host to device. Executed asynchronously on stream if async is true.

CUDAdrv.Mem.download — Function.

Download memory from device to host. Executed asynchronously on stream if async is true.

CUDAdrv.Mem.download! — Function.

Download memory from device to host. Executed asynchronously on stream if async is true.

CUDAdrv.Mem.transfer — Function.

Transfer memory from device to device. Executed asynchronously on stream if async is true.

CUDAdrv.Mem.transfer! — Function.

Transfer memory from device to device. Executed asynchronously on stream if async is true.

Memory info

CUDAdrv.Mem.info — Function.

info()

Returns a tuple of two integers, indicating respectively the free and total amount of memory (in bytes) available for allocation by the CUDA context.

CUDAdrv.Mem.total — Function.

total()

Returns the total amount of memory (in bytes), available for allocation by the CUDA context.

CUDAdrv.Mem.used — Function.

used()

Returns the used amount of memory (in bytes), allocated by the CUDA context.

CUDAdrv.Mem.free — Method.

free()

Returns the free amount of memory (in bytes), available for allocation by the CUDA context.

Stream Management

CUDAdrv.CuStream — Type.

CuStream(flags=0)

Create a CUDA stream.

CUDAdrv.CuDefaultStream — Function.

CuDefaultStream()

Return the default stream.

CUDAdrv.synchronize — Method.

synchronize(s::CuStream)

Wait until a stream's tasks are completed.

Event Management

CUDAdrv.CuEvent — Type.

CuEvent()

Create a new CUDA event.

CUDAdrv.record — Function.

record(e::CuEvent, stream=CuDefaultStream())

Record an event on a stream.

CUDAdrv.synchronize — Method.

synchronize(e::CuEvent)

Waits for an event to complete.

CUDAdrv.elapsed — Function.

elapsed(start::CuEvent, stop::CuEvent)

Computes the elapsed time between two events (in seconds).

CUDAdrv.@elapsed — Macro.

@elapsed stream ex
@elapsed ex

A macro to evaluate an expression, discarding the resulting value, instead returning the number of seconds it took to execute on the GPU, as a floating-point number.

Execution Control

CUDAdrv.CuDim3 — Type.

CuDim3(x)

CuDim3((x,))
CuDim3((x, y))
CuDim3((x, y, x))

A type used to specify dimensions, consisting of 3 integers for respectively the x, y and z dimension. Unspecified dimensions default to 1.

Often accepted as argument through the CuDim type alias, eg. in the case of cudacall or launch, allowing to pass dimensions as a plain integer or a tuple without having to construct an explicit CuDim3 object.

CUDAdrv.cudacall — Function.

cudacall(f::CuFunction, types, values...;
         blocks::CuDim, threads::CuDim, shmem=0, stream=CuDefaultStream())

ccall-like interface for launching a CUDA function f on a GPU.

For example:

vadd = CuFunction(md, "vadd")
a = rand(Float32, 10)
b = rand(Float32, 10)
ad = Mem.upload(a)
bd = Mem.upload(b)
c = zeros(Float32, 10)
cd = Mem.alloc(c)

cudacall(vadd, (Ptr{Cfloat},Ptr{Cfloat},Ptr{Cfloat}), ad, bd, cd;
         threads=10)
Mem.download!(c, cd)

The blocks and threads arguments control the launch configuration, and should both consist of either an integer, or a tuple of 1 to 3 integers (omitted dimensions default to 1). The types argument can contain both a tuple of types, and a tuple type, the latter being slightly faster.

CUDAdrv.launch — Function.

launch(f::CuFunction, blocks::CuDim, threads::CuDim, args...;
       shmem=0, stream=CuDefaultStream())
launch(f::CuFunction, blocks::CuDim, threads::CuDim, shmem::Int, stream::CuStream, args...)

Low-level call to launch a CUDA function f on the GPU, using blocks and threads as respectively the grid and block configuration. Dynamic shared memory is allocated according to shmem, and the kernel is launched on stream stream.

Arguments to a kernel should either be bitstype, in which case they will be copied to the internal kernel parameter buffer, or a pointer to device memory.

This is a low-level call, prefer to use cudacall instead.

Profiler Control

CUDAdrv.@profile — Macro.

@profile ex

Run expressions while activating the CUDA profiler.

Note that this API is used to programmatically control the profiling granularity by allowing profiling to be done only on selective pieces of code. It does not perform any profiling on itself, you need external tools for that.

CUDAdrv.Profile.start — Function.

start()

Enables profile collection by the active profiling tool for the current context. If profiling is already enabled, then this call has no effect.

CUDAdrv.Profile.stop — Function.

stop()

Disables profile collection by the active profiling tool for the current context. If profiling is already disabled, then this call has no effect.