Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

AbstractLattices.:∧ — Method.

J ∧ K is the largest ClosedInterval contained in both. See also *.

AbstractLattices.:∨ — Method.

J ∨ K is the smalles ClosedInterval containing both. See also +.

ClosedIntervals.EmptyInterval — Function.

EmptyInterval(T::DataType = Float64) creates an empty ClosedInterval of a given type.

ClosedIntervals.left — Method.

For a ClosedInterval I, left(I) returns its left end point.

ClosedIntervals.right — Method.

For a ClosedInterval I, right(I) returns its right end point.

Base.:* — Method.

For ClosedIntervals J and K, J*K is their intersection. Also available as J ∧ K.

Base.:+ — Method.

For ClosedIntervals J and K, J+K is the smallest ClosedInterval containing them both. This is also available as J ∨ K.

Base.:<< — Method.

For ClosedIntervals I and J, I<<J tests if I is completely to the left of J.

Base.:>> — Method.

For ClosedIntervals I and J, I>>J tests if I is completely to the right of J.

Base.in — Method.

For a number x and a ClosedInterval I (of the same type) in(x,I) tests if x is contained in the interval I.

Base.isempty — Method.

For a ClosedInterval I, isempty(I) tests if I is an empty interval.

Base.isless — Method.

Lexicographic ordering of ClosedIntervals.

Base.length — Method.

length(I) is the length of the ClosedInterval I.