Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Clustering.Hclust — Type.

Hierarchical clustering of the data returned by hclust(). The data hierarchy is defined by the merges matrix:

each row specifies which subtrees (referenced by their IDs) are merged into a higher-level subtree

negative subtree id denotes leaf node and corresponds to the datapoint at position -id

positive id denotes nontrivial subtree: the row merges[id, :] specifies its left and right subtrees, and heights[id] – its height.

This type mostly follows R's hclust class.

Clustering.MCLResult — Type.

struct MCLResult <: ClusteringResult

Result returned by mcl().

Clustering.dbscan — Method.

dbscan(points, radius ; leafsize = 20, min_neighbors = 1, min_cluster_size = 1) -> clusters

Cluster points using the DBSCAN (density-based spatial clustering of applications with noise) algorithm.

Arguments

points: matrix of points

radius::Real: query radius

Keyword Arguments

leafsize::Int: number of points binned in each leaf node in the KDTree

min_neighbors::Int: minimum number of neighbors to be a core point

min_cluster_size::Int: minimum number of points to be a valid cluster

Output

Vector{DbscanCluster}: an array of clusters with the id, size core indices and boundary indices

Example:

points = randn(3, 10000)
clusters = dbscan(points, 0.05, min_neighbors = 3, min_cluster_size = 20) # clusters with less than 20 points will be discarded

Clustering.mcl — Method.

mcl(adj::Matrix; [keyword arguments])::MCLResult

Identify clusters in the weighted graph using Markov Clustering Algorithm (MCL).

Arguments

adj::Matrix{Float64}: adjacency matrix that defines the weighted graph to cluster

add_loops::Bool: whether edges of weight 1.0 from the node to itself should be appended to the graph (enabled by default)

expansion::Number: MCL expansion constant (2)

inflation::Number: MCL inflation constant (2.0)

save_final_matrix::Bool: save final equilibrium state in the result, otherwise leave it empty; disabled by default, could be useful if MCL doesn't converge

max_iter::Integer: max number of MCL iterations

tol::Number: MCL adjacency matrix convergence threshold

prune_tol::Number: pruning threshold

display::Symbol: :none for no output or :verbose for diagnostic messages

See original MCL implementation.

Ref: Stijn van Dongen, "Graph clustering by flow simulation", 2001

Clustering.silhouettes — Method.

silhouettes(assignments::AbstractVector, [counts,] dists)
silhouettes(clustering::ClusteringResult, dists)

Compute silhouette values for individual points w.r.t. given clustering.

assignments the vector of point assignments (cluster indices)

counts the optional vector of cluster sizes (how many points assigned to each cluster; should match assignments)

clustering the output of some clustering method

dists point×point pairwise distance matrix

Returns a vector of silhouette values for each individual point.

mean(silhouettes(...)) could be used as a measure of clustering quality; higher values indicate better separation of clusters w.r.t. distances provided in dists.

References

Silhouette Wikipedia page.

Peter J. Rousseeuw (1987). "Silhouettes: a Graphical Aid to the Interpretation and Validation of Cluster Analysis". Computational and Applied Mathematics. 20: 53–65.

Clustering.vmeasure — Method.

vmeasure(assign1, assign2; β = 1.0)

V-measure between two clustering assignments.

assign1 and assign2 can be either ClusteringResult objects or assignments vectors (AbstractVector{<:Integer}).

The β parameter defines trade-off between homogeneity and completeness:

if β is greater than 1, completeness is weighted more strongly,

if β is less than 1, homogeneity is weighted more strongly.

Ref: Andrew Rosenberg and Julia Hirschberg, 2007. "V-Measure: A conditional entropy-based external cluster evaluation measure"

Clustering.AverageDistance — Type.

Average distance between a pair of points from each clusters.

Clustering.MaximumDistance — Type.

Maximum distance between a pair of point from each clusters.

Clustering.MinimalDistance — Type.

Distance between the clusters is the minimal distance between any pair of their points.

Clustering.ReducibleMetric — Type.

Base type for reducible Lance–Williams cluster metrics.

The metric d is called reducible if for any clusters A, B and C and some ρ > 0 s.t.

d(A, B) < ρ, d(A, C) > ρ, d(B, C) > ρ

it follows that

d(A∪B, C) > ρ

If the cluster metrics belongs to Lance-Williams family, there is an efficient formula that defines d(A∪B, C) using d(A, C), d(B, C) and d(A, B).

Clustering.WardDistance — Type.

Ward distance between the two clusters A and B is the amount by which merging the two clusters into a single larger cluster A∪B would increase the average squared distance of a point to its cluster centroid.

Clustering._dbscan — Method.

An implementation of DBSCAN algorithm that keeps track of both the core and boundary points

Clustering.accept_cluster! — Method.

accept_cluster!(clusters, core_selection, cluster_selection)

Accept cluster and update the clusters list

Input

clusters :: Vector{DbscanCluster}: a list of the accepted clusters

core_selection :: Vector{Bool}: selection of the core points of the cluster

cluster_selection :: Vector{Bool}: selection of all the cluster points

Clustering.update_exploration_list! — Method.

update_exploration_list!(adj_list, exploration_list, visited)

Update the queue for expanding the cluster

Input

adj_list :: Vector{Int}: indices of the neighboring points

exploration_list :: Vector{Int}: the indices that will be explored in the future

visited :: Vector{Bool}: a flag to indicate whether a point has been explored already