Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

CodeTools.codemodule — Method.

Takes Julia source code and a line number, gives back the string name of the module at that line.

CodeTools.dirsnearby — Method.

Takes a start directory and returns a set of nearby directories.

CodeTools.filemodule_ — Method.

Takes an absolute path to a file and returns a string representing the module it belongs to.

CodeTools.find_include — Method.

Takes an absolute path to a file and returns the (file, line) where that file is included or nothing.

CodeTools.includeline — Method.

Takes a given Julia source file and another (absolute) path, gives the line on which the path is included in the file or 0.