Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

CodecBzip2.Bzip2Compressor — Method.

Bzip2Compressor(;blocksize100k=9, workfactor=30, verbosity=0)

Create a bzip2 compression codec.

Arguments

blocksize100k: block size to be use for compression (1..9)

workfactor: amount of effort the standard algorithm will expend before resorting to the fallback (0..250)

verbosity: verbosity level (0..4)

CodecBzip2.Bzip2CompressorStream — Method.

Bzip2CompressorStream(stream::IO; kwargs...)

Create a bzip2 compression stream (see Bzip2Compressor for kwargs).

CodecBzip2.Bzip2Decompressor — Method.

Bzip2Decompressor(;small=false, verbosity=0)

Create a bzip2 decompression codec.

Arguments

small: flag to activate an algorithm which uses less memory

verbosity: verbosity level (0..4)

CodecBzip2.Bzip2DecompressorStream — Method.

Bzip2DecompressorStream(stream::IO; kwargs...)

Create a bzip2 decompression stream (see Bzip2Decompressor for kwargs).