Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ColorTypes.AGray32 — Type.

AGray32 uses a UInt32 representation of color, 0xAAIIIIII, where I=intensity (grayscale value) and A=alpha. Each II pair is assumed to be the same. This format is often used by external libraries such as Cairo.

You can extract the single gray value with gray(c) and the alpha as alpha(c). You can construct them directly from a UInt32, or as AGray32(i,alpha). Note that i and alpha are interpreted on a scale from 0 (black) to 1 (white).

ColorTypes.ARGB32 — Type.

ARGB32 uses a UInt32 representation of color, 0xAARRGGBB, where R=red, G=green, B=blue and A is the alpha channel. This format is often used by external libraries such as Cairo. On a little-endian machine, this type has the exact same storage format as BGRA{N0f8}.

ARGB32 colors do not have fields named alpha, r, g, b, but you can still extract the individual components with alpha(c), red(c), green(c), blue(c). You can construct them directly from a UInt32, or as ARGB32(r, g, b, alpha).

ColorTypes.AbstractRGB — Type.

AbstractRGB{T} is an abstract supertype for red/green/blue color types that can be constructed as C(r, g, b) and for which the elements can be extracted as red(c), green(c), blue(c). You should not make assumptions about internal storage order, the number of fields, or the representation. One AbstractRGB color-type, RGB24, is not parametric and does not have fields named r, g, b.

ColorTypes.AlphaColor — Type.

AlphaColor is an abstract supertype for types like ARGB, where the alpha channel comes first in the internal storage order. Note that the constructor order is still (color, alpha).

ColorTypes.BGR — Type.

BGR is a variant of RGB with the opposite storage order. Note that the constructor is still called in the order BGR(r, g, b). This storage order is noteworthy because on little-endian machines, BGRA (with transparency) can be reinterpreted to the UInt32 color format used by libraries such as Cairo and OpenGL.

ColorTypes.Color — Type.

Color{T,N} is the abstract supertype for a color (or grayscale) with no transparency.

ColorTypes.ColorAlpha — Type.

ColorAlpha is an abstract supertype for types like RGBA, where the alpha channel comes last in the internal storage order.

ColorTypes.Colorant — Type.

Colorant{T,N} is the abstract super-type of all types in ColorTypes, and refers to both (opaque) colors and colors-with-transparency (alpha channel) information. T is the element type (extractable with eltype) and N is the number of meaningful entries (extractable with length), i.e., the number of arguments you would supply to the constructor.

ColorTypes.DIN99 — Type.

DIN99 is the (L99, a99, b99) adaptation of CIELAB

ColorTypes.DIN99d — Type.

DIN99d is the (L99d, a99d, b99d) improvement on DIN99

ColorTypes.DIN99o — Type.

DIN99o is the (L99o, a99o, b99o) adaptation of CIELAB

ColorTypes.Gray — Type.

Gray is a grayscale object. You can extract its value with gray(c).

ColorTypes.Gray24 — Type.

Gray24 uses a UInt32 representation of color, 0xAAIIIIII, where I=intensity (grayscale value) and A is irrelevant. Each II pair is assumed to be the same. This format is often used by external libraries such as Cairo.

You can extract the single gray value with gray(c). You can construct them directly from a UInt32, or as Gray24(i). Note that i is interpreted on a scale from 0 (black) to 1 (white).

ColorTypes.HSB — Type.

HSB (Hue-Saturation-Brightness) is an alias for HSV.

ColorTypes.HSI — Type.

HSI is the Hue-Saturation-Intensity colorspace.

ColorTypes.HSL — Type.

HSL is the Hue-Saturation-Lightness colorspace.

ColorTypes.HSV — Type.

HSV is the Hue-Saturation-Value colorspace.

ColorTypes.LCHab — Type.

LCHab is the Luminance-Chroma-Hue, Polar-Lab colorspace

ColorTypes.LCHuv — Type.

LCHuv is the Luminance-Chroma-Hue, Polar-Luv colorspace

ColorTypes.LMS — Type.

LMS is the Long-Medium-Short colorspace based on activation of the three cone photoreceptors. Like XYZ, this is a linear color space.

ColorTypes.Lab — Type.

Lab is the CIELAB colorspace.

ColorTypes.Luv — Type.

Luv is the CIELUV colorspace

ColorTypes.RGB — Type.

RGB is the standard Red-Green-Blue (sRGB) colorspace. Values of the individual color channels range from 0 (black) to 1 (saturated). If you want "Integer" storage types (e.g., 255 for full color), use N0f8(1) instead (see FixedPointNumbers).

ColorTypes.RGB1 — Type.

RGB1 is a variant of RGB which has a padding element inserted at the beginning. In some applications it may have useful memory-alignment properties.

Like all other AbstractRGB objects, the constructor is still called RGB1(r, g, b).

ColorTypes.RGB24 — Type.

RGB24 uses a UInt32 representation of color, 0xAARRGGBB, where R=red, G=green, B=blue and A is irrelevant. This format is often used by external libraries such as Cairo.

RGB24 colors do not have fields named r, g, b, but you can still extract the individual components with red(c), green(c), blue(c). You can construct them directly from a UInt32, or as RGB(r, g, b).

ColorTypes.RGB4 — Type.

RGB4 is a variant of RGB which has a padding element inserted at the end. In some applications it may have useful memory-alignment properties.

Like all other AbstractRGB objects, the constructor is still called RGB4(r, g, b).

ColorTypes.TransparentColor — Type.

TransparentColor{C,T,N} is the abstract type for any color-with-transparency. The C parameter refers to the type of the pure color (without transparency) and can be extracted with color_type. T is the element type of both C and the alpha channel, and N has the same meaning as in Colorant (and is 1 larger than the corresponding color type).

All transparent types should support two modes of construction:

P(color, alpha)
P(component1, component2, component3, alpha) (assuming a 3-component color)

For a Colorant c, the color component can be extracted with color(c), and the alpha channel with alpha(c). Note that types such as ARGB32 do not have a field named alpha.

Most concrete types, like RGB, have both ARGB and RGBA transparent analogs. These two indicate different internal storage order (see AlphaColor and ColorAlpha, and the alphacolor and coloralpha functions).

ColorTypes.XYZ — Type.

XYZ is the CIE 1931 XYZ colorspace. It is a linear colorspace, meaning that mathematical operations such as addition, subtraction, and scaling make "colorimetric sense" in this colorspace.

ColorTypes.YCbCr — Type.

YCbCr is the Y'CbCr color encoding often used in digital photography or video

ColorTypes.YIQ — Type.

YIQ is a color encoding, for example used in NTSC transmission.

ColorTypes.xyY — Type.

xyY is the CIE 1931 xyY (chromaticity + luminance) space

ColorTypes.alpha — Method.

alpha(p) extracts the alpha component of a color. For a color without an alpha channel, it will always return 1.

ColorTypes.alphacolor — Function.

alphacolor(RGB) returns ARGB, i.e., the corresponding transparent color type with storage order (alpha, color).

ColorTypes.base_color_type — Method.

base_color_type is similar to color_type, except it "strips off" the element type. For example,

color_type(RGB{N0f8})     == RGB{N0f8}
base_color_type(RGB{N0f8}) == RGB

This can be very handy if you want to switch element types. For example:

c64 = base_color_type(c){Float64}(color(c))

converts c into a Float64 representation (potentially discarding any alpha-channel information).

ColorTypes.base_colorant_type — Method.

base_colorant_type is similar to base_color_type, but it preserves the "alpha" portion of the type.

For example,

base_color_type(ARGB{N0f8})  == RGB
base_colorant_type(ARGB{N0f8})  == ARGB

If you just want to switch element types, this is the safest default and the easiest to use:

c64 = base_colorant_type(c){Float64}(c)

ColorTypes.blue — Method.

blue(c) returns the blue component of an AbstractRGB opaque or transparent color.

ColorTypes.ccolor — Method.

ccolor ("concrete color") helps write flexible methods. The idea is that users may write convert(HSV, c) or even convert(Array{HSV}, A) without specifying the element type explicitly (e.g., convert(Array{HSV{Float32}}, A)). ccolor implements the logic "choose the user's eltype if specified, otherwise retain the eltype of the source object." However, when the source object has FixedPoint element type, and the destination only supports AbstractFloat, we choose Float32.

Usage:

ccolor(desttype, srctype) -> concrete desttype

Example:

convert{C<:Colorant}(::Type{C}, p::Colorant) = cnvt(ccolor(C,typeof(p)), p)

where cnvt is the function that performs explicit conversion.

ColorTypes.color — Method.

color(c) extracts the opaque color component from a Colorant (e.g., omits the alpha channel, if present).

ColorTypes.color_type — Method.

color_type(c) or color_type(C) (c being a color instance and C being the type) returns the type of the Color object (without alpha channel). This, and related functions like base_color_type, base_colorant_type, and ccolor are useful for manipulating types for writing generic code.

For example,

color_type(RGB)          == RGB
color_type(RGB{Float32}) == RGB{Float32}
color_type(ARGB{N0f8})     == RGB{N0f8}

ColorTypes.coloralpha — Function.

coloralpha(RGB) returns RGBA, i.e., the corresponding transparent color type with storage order (color, alpha).

ColorTypes.comp1 — Method.

comp1(c) extracts the first component you'd pass to the constructor of the corresponding object. For most color types without an alpha channel, this is just the first field, but for types like BGR that reverse the internal storage order this provides the value that you'd use to reconstruct the color.

Specifically, for any Color{T,3},

c == typeof(c)(comp1(c), comp2(c), comp3(c))

returns true.

ColorTypes.comp2 — Method.

comp2(c) extracts the second constructor argument (see comp1).

ColorTypes.comp3 — Method.

comp3(c) extracts the third constructor argument (see comp1).

ColorTypes.gray — Method.

gray(c) returns the gray component of a grayscale opaque or transparent color.

ColorTypes.green — Method.

green(c) returns the green component of an AbstractRGB opaque or transparent color.

ColorTypes.mapc — Method.

mapc(f, rgb) -> rgbf
mapc(f, rgb1, rgb2) -> rgbf

mapc applies the function f to each color channel of the input color(s), returning an output color in the same colorspace.

Examples:

julia> mapc(x->clamp(x,0,1), RGB(-0.2,0.3,1.2))
RGB{Float64}(0.0,0.3,1.0)

julia> mapc(max, RGB(0.1,0.8,0.3), RGB(0.5,0.5,0.5))
RGB{Float64}(0.5,0.8,0.5)

julia> mapc(+, RGB(0.1,0.8,0.3), RGB(0.5,0.5,0.5))
RGB{Float64}(0.6,1.3,0.8)

ColorTypes.mapreducec — Method.

mapreducec(f, op, v0, c)

Reduce across color channels of c with the binary operator op, first applying f to each channel. v0 is the neutral element used to initiate the reduction. For grayscale,

mapreducec(f, op, v0, c::Gray) = op(v0, f(comp1(c)))

whereas for RGB

mapreducec(f, op, v0, c::RGB) = op(f(comp3(c)), op(f(comp2(c)), op(v0, f(comp1(c)))))

If c has an alpha channel, it is always the last one to be folded into the reduction.

ColorTypes.red — Method.

red(c) returns the red component of an AbstractRGB opaque or transparent color.

ColorTypes.reducec — Method.

reducec(op, v0, c)

Reduce across color channels of c with the binary operator op. v0 is the neutral element used to initiate the reduction. For grayscale,

reducec(op, v0, c::Gray) = op(v0, comp1(c))

whereas for RGB

reducec(op, v0, c::RGB) = op(comp3(c), op(comp2(c), op(v0, comp1(c))))

If c has an alpha channel, it is always the last one to be folded into the reduction.