Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

CommunityDetection.community_detection_bethe — Function.

community_detection_bethe(g::AbstractGraph, k=-1; kmax=15)

Perform detection for k communities using the spectral properties of the Bethe Hessian matrix associated to g. If k is omitted or less than 1, the optimal number of communities will be automatically selected. In this case the maximum number of detectable communities is given by kmax. Return a vector containing the vertex assignments.

References

Saade et al.

CommunityDetection.community_detection_nback — Method.

community_detection_nback(g::AbstractGraph, k::Int)

Return an array, indexed by vertex, containing commmunity assignments for graph g detecting k communities. Community detection is performed using the spectral properties of the non-backtracking matrix of g.

References

Krzakala et al.

CommunityDetection.nonbacktrack_embedding — Method.

nonbacktrack_embedding(g::AbstractGraph, k::Int)

Perform spectral embedding of the non-backtracking matrix of g. Return a matrix ϕ where ϕ[:,i] are the coordinates for vertex i.

Implementation Notes

Does not explicitly construct the non_backtracking_matrix. See Nonbacktracking for details.

References

Krzakala et al..