Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ComplexPhasePortrait.phase! — Method.

ComplexPhasePortrait.phase!(args...; attributes...)

Command works on plot args 1:N and accepts keyword arguments to style the plot. Adds new plot to current_scene()

ComplexPhasePortrait.phase! — Method.

ComplexPhasePortrait.phase!(attributes::Attributes, args...; attributes...)

Like ComplexPhasePortrait.phase!(args...; attributes...) but accepts a theme as first argument. Adds new plot to current_scene()!

ComplexPhasePortrait.phase! — Method.

ComplexPhasePortrait.phase!(attributes::Attributes, args...; attributes...)

Like ComplexPhasePortrait.phase!(scene, args...; attributes...) but accepts a theme as second argument. Adds new plot to scene!

ComplexPhasePortrait.phase! — Method.

ComplexPhasePortrait.phase!(scene::SceneLike, args...; attributes...)

Command works on plot args 1:N and accepts keyword arguments to style the plot. Adds new plot to scene!

ComplexPhasePortrait.phase — Method.

ComplexPhasePortrait.phase(args...; attributes...)

Command works on plot args 1:N and accepts keyword arguments to style the plot. Creates a new scene!

ComplexPhasePortrait.phase — Method.

ComplexPhasePortrait.phase(attributes::Attributes, args...; attributes...)

Like ComplexPhasePortrait.phase(args...; attributes...) but accepts a theme as first argument. Creates a new scene!

ComplexPhasePortrait.baseColorMap — Method.

Color map for complex phase portrait with 600 elements.

ComplexPhasePortrait.sawfun — Method.

Sawtooth function over reals with period dx onto [a,b].

ComplexPhasePortrait.stepfun — Method.

Integer step function with period 1 such that [0,1]⟶[1,nmax].