Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ComputationalResources.AbstractCPU — Type.

AbstractCPU

An abstract type indicating that computation should be performed using the CPU.

ComputationalResources.AbstractResource — Type.

AbstractResource

The abstract supertype of all computational resources.

ComputationalResources.ArrayFireLibs — Type.

ArrayFireLibs()
ArrayFireLibs(settings)

Indicate that computation should be performing using the ArrayFire libraries. Optionally pass in an object specifying algorithmic settings.

Examples:

filter(ArrayFireLibs(), image, kernel)
filter(ArrayFireLibs(backend), image, kernel)

ComputationalResources.CPU1 — Type.

CPU1()
CPU1(settings)

Indicate that a computation should be performed using the CPU in single-threaded mode. Optionally pass in an object specifying algorithmic settings.

Examples:

filter(CPU1(), image, kernel)
filter(CPU1(TileSize(64,8)), image, kernel)

ComputationalResources.CPUThreads — Type.

CPUThreads()
CPUThreads(settings)

Indicate that a computation should be performed using the CPU in multi-threaded mode. Optionally pass in an object specifying algorithmic settings.

Examples:

filter(CPUThreads(), image, kernel)
filter(CPUThreads(TileSize(64,8)), image, kernel)

ComputationalResources.CUDALibs — Type.

CUDALibs()
CUDALibs(settings)

Indicate that computation should be performing using the CUDA libraries. Optionally pass in an object specifying algorithmic settings.

Examples:

filter(CUDALibs(), image, kernel)
filter(CUDALibs(backend), image, kernel)

ComputationalResources.OpenCLLibs — Type.

OpenCLLibs()
OpenCLLibs(settings)

Indicate that computation should be performing using the OpenCL libraries. Optionally pass in an object specifying algorithmic settings.

Examples:

filter(OpenCLLibs(), image, kernel)
filter(OpenCLLibs(backend), image, kernel)

ComputationalResources.TileSize — Type.

TileSize(dims)

Request that an array computation be performed using tiles (blocks) of size dims.

ComputationalResources.addresource — Method.

addresource(T)

Add T to the list of available resources. For example, addresource(OpenCLLibs) would indicate that you have a GPU and the OpenCL libraries installed.

ComputationalResources.haveresource — Method.

haveresource(T)

Returns true if T is an available resource. For example, haveresource(OpenCLLibs) tests whether the OpenCLLibs have been added as an available resource. This function is typically used inside a module's __init__ function.

Example:

# The __init__ function for MyPackage:
function __init__()
    ...  # other initialization code, possibly setting the LOAD_PATH
    if haveresource(OpenCLLibs)
        @eval using MyPackageCL  # a separate module containing OpenCL implementations
    end
    # Put additional resource checks here
    ...  # other initialization code, possibly cleaning up the LOAD_PATH
end

ComputationalResources.rmresource — Method.

rmresource(T)

Remove T from the list of available resources. For example, rmresource(OpenCLLibs) would indicate that any future package loads should avoid loading their specializations for OpenCL.

ComputationalResources.CPUProcesses — Type.

CPUProcesses()
CPUProcesses(settings)

Indicate that a computation should be performed using the CPU in multi-process mode. Processes should be added with addprocs() or julia started with julia -p N. Processes must communicate using distributed memory operations such as remote refrences. Optionally pass in an object specifying algorithmic settings.

Examples:

filter(CPUProcesses(), image, kernel)
filter(CPUProcesses(TileSize(64,8)), image, kernel)