Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Conda._install_conda — Function.

Install miniconda if it hasn't been installed yet; installconda(true) installs Conda even if it has already been installed.

Conda._installed_packages — Function.

List all installed packages as an array.

Conda._installed_packages_dict — Function.

List all installed packages as an dict of tuples with (version_number, fullname).

Conda._installer_url — Method.

Get the miniconda installer URL.

Conda._quiet — Method.

Suppress progress bar in continuous integration environments

Conda._set_conda_env — Function.

Use a cleaned up environment for the command cmd.

Any environment variable starting by CONDA or PYTHON will interact with the run.

Conda.add — Function.

Install a new package.

Conda.add_channel — Function.

Add a channel to the list of channels

Conda.bin_dir — Method.

Prefix for the executable files installed with the packages

Conda.channels — Function.

Get the list of channels used to search packages

Conda.conda_rc — Method.

Path to the condarc file

Conda.exists — Function.

Check if a given package exists.

Conda.lib_dir — Method.

Prefix for the shared libraries installed with the packages

Conda.list — Function.

List all installed packages to standard output.

Conda.parseconda — Function.

Run conda command with environment variables set and return the json output as a julia object

Conda.prefix — Method.

Prefix for installation of the environment

Conda.python_dir — Method.

Prefix where the python command lives

Conda.rm — Function.

Uninstall a package.

Conda.rm_channel — Function.

Remove a channel from the list of channels

Conda.runconda — Function.

Run conda command with environment variables set.

Conda.script_dir — Method.

Prefix for the python scripts. On UNIX, this is the same than Conda.BINDIR

Conda.search — Function.

Search a specific version of a package

Conda.search — Function.

Search packages for a string

Conda.update — Function.

Update all installed packages.

Conda.version — Function.

Get the exact version of a package as a VersionNumber.