Intervals and univariate transformations

Intervals

The interval types are different from some other interval implementations in Julia. They do not specify if the interval is open or closed at an endpoint, and also encode infiniteness and semi-infiniteness in the type, for type stable code.

ContinuousTransformations.AbstractInterval — Type.

abstract type AbstractInterval

Abstract supertype for all univariate intervals. It is not specified whether they are open or closed.

ContinuousTransformations.RealLine — Type.

RealLine()

The real line. Use the constant ℝ.

ContinuousTransformations.ℝ — Constant.

A constant for the real line.

ContinuousTransformations.PositiveRay — Type.

PositiveRay(left)

The real numbers above left. See ℝ⁺.

source

ContinuousTransformations.ℝ⁺ — Constant.

The positive real numbers.

ContinuousTransformations.NegativeRay — Type.

NegativeRay(right)

The real numbers below right. See ℝ⁻.

source

ContinuousTransformations.ℝ⁻ — Constant.

The negative real numbers.

ContinuousTransformations.Segment — Type.

Segment(left, right)

The real numbers between left and right, with $-∞ < \text{left} < \text{right} < ∞$ enforced.

source

Intervals also support the following methods in Base: minimum, maximum, in, isfinite, isinf, extrema.

Segments also support middle, linspace, and

ContinuousTransformations.width — Function.

width(s)

Width of a finite interval.

Univariate transformations

General interface

ContinuousTransformations.UnivariateTransformation — Type.

abstract type UnivariateTransformation <: ContinuousTransformation

Univariate monotone transformation, either increasing or decreasing on the whole domain (thus, a bijection).

ContinuousTransformations.isincreasing — Function.

isincreasing(transformation)

Return true (false), when the transformation is monotonically increasing (decreasing).

Specific transformations

ContinuousTransformations.Affine — Type.

Affine(α, β)

Mapping $ℝ → ℝ$ using $x ↦ α⋅x + β$.

$α > 0$ is enforced, see Negation.

source

ContinuousTransformations.IDENTITY — Function.

Identity (as an affine transformation).

ContinuousTransformations.Negation — Type.

Negation()

Mapping $ℝ → ℝ$ using $x ↦ -x$.

source

ContinuousTransformations.NEGATION — Function.

Negation()

Mapping $ℝ → ℝ$ using $x ↦ -x$.

source

ContinuousTransformations.Logistic — Type.

Logistic()

Mapping $ℝ → (0,1)$ using $x ↦ 1/(1+\exp(-x))$.

source

ContinuousTransformations.LOGISTIC — Function.

Logistic()

Mapping $ℝ → (0,1)$ using $x ↦ 1/(1+\exp(-x))$.

source

ContinuousTransformations.Exp — Type.

Exp()

Mapping $ℝ → ℝ⁺$ using $x ↦ \exp(x)$.

source

ContinuousTransformations.EXP — Function.

Exp()

Mapping $ℝ → ℝ⁺$ using $x ↦ \exp(x)$.

source

ContinuousTransformations.Logit — Type.

Logit()

Mapping $(0,1) → ℝ$ using $x ↦ \log(x/(1-x))$.

source

ContinuousTransformations.LOGIT — Function.

Logit()

Mapping $(0,1) → ℝ$ using $x ↦ \log(x/(1-x))$.

source

ContinuousTransformations.Log — Type.

Log()

Mapping $ℝ → ℝ⁺$ using $x ↦ \exp(x)$.

source

ContinuousTransformations.LOG — Function.

Log()

Mapping $ℝ → ℝ⁺$ using $x ↦ \exp(x)$.

source

ContinuousTransformations.InvRealCircle — Type.

InvRealCircle()

Mapping $(-1,1) → ℝ$ using $x ↦ x/√(1-x^2)$.

source

ContinuousTransformations.INVREALCIRCLE — Function.

InvRealCircle()

Mapping $(-1,1) → ℝ$ using $x ↦ x/√(1-x^2)$.

source

ContinuousTransformations.RealCircle — Type.

RealCircle()

Mapping $ℝ → (-1,1)$ using $x ↦ x/√(1+x^2)$.

source

ContinuousTransformations.REALCIRCLE — Function.

RealCircle()

Mapping $ℝ → (-1,1)$ using $x ↦ x/√(1+x^2)$.

source

Composing transformations

ContinuousTransformations.ComposedTransformation — Type.

ComposedTransformation(f, g)

Compose two univariate transformations, resulting in the mapping $f∘g$, or `x ↦ f(g(x)).

Use the ∘ operator for construction.