Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

CxxWrap.gcprotect — Method.

Protect a variable from garbage collection by adding it to the global array kept by CxxWrap

CxxWrap.gcunprotect — Method.

Unprotect a variable from garbage collection by removing it from the global array kept by CxxWrap

CxxWrap.ptrunion — Method.

Create a Union containing the type and a smart pointer to any type derived from it

CxxWrap.SmartPointer — Type.

Base class for smart pointers

CxxWrap.SmartPointerWithDeref — Type.

Concrete smart pointer implementation. PT is a hash for the pointer type.