Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

D3Trees.D3Tree — Method.

D3Tree(children, <keyword arguments>)

Construct a tree to be displayed using D3 in a browser or ipython notebook, specifying structure with lists of children indices.

Arguments

Required

children::Vector{Vector{Int}}: List of children for each node. E.g.
D3Tree([[2,3], [], [4], []])
creates a tree with four nodes. Nodes 2 and 3 are children of node 1, and node 4 is the only child of node 3. Nodes 2 and 4 are childless.

Keyword:

text::Vector{String} - text to appear under each node.

tooltip::Vector{String} - text to appear when hovering over each node.

style::Vector{String} - html style for each node.

link_style::Vector{String} - html style for each link.

title::String - html title.

init_expand::Integer - levels to expand initially.

init_duration::Number - duration of the initial animation in ms.

svg_height::Number - height of the svg containing the tree in px.

D3Trees.D3Tree — Method.

D3Tree(node; detect_repeat=true, kwargs...)

Construct a tree to be displayed using D3 in a browser or ipython notebook with any object, node, that implements the AbstractTrees interface.

The style may be controlled by implementing the following functions, which should return Strings for the nodes:

D3Trees.text(node)
D3Trees.tooltip(node)
D3Trees.style(node)
D3Trees.link_style(node)

Arguments

Required

node: an object that has AbstractTrees.children(node) and AbstractTrees.printnode(io::IO, node)

Keyword

detect_repeat: if true, uses a dictionary to detect whether a node has appeared previously

Also supports, the non-vector arguments of the vector-of-vectors D3Tree constructor, i.e. title, init_expand, init_duration, svg_height.

D3Trees.inbrowser — Method.

inbrowser(t::D3Tree, command::Cmd)

Open an html document with the D3Tree in a program launched with the specified command.

D3Trees.inbrowser — Method.

inbrowser(t::D3Tree, browsername::String)

Open an html document with the D3Tree in a browser with a platform-specific command.

D3Trees.link_style — Method.

D3Trees.link_style(n)

Return the html style for the link leading to the D3Trees node corresponding to AbstractTrees node n

D3Trees.style — Method.

D3Trees.style(n)

Return the html style for the D3Trees node corresponding to AbstractTrees node n

D3Trees.text — Method.

D3Trees.text(n)

Return the text to be displayed at the D3Trees node corresponding to AbstractTrees node n

D3Trees.tooltip — Method.

D3Trees.tooltip(n)

Return the text to be displayed in the tooltip for the D3Trees node corresponding to AbstractTrees node n