Defaults

Defaults

DFControl.configuredefault_pseudos — Method.

configuredefault_pseudos(server = getdefault_server(), pseudo_dirs=getdefault_pseudodirs())

Reads the specified default_pseudo_dirs on the default_server and sets up the default_pseudos variable, and also adds all the entries to the user_defaults.jl file.

DFControl.getdefault_pseudo — Function.

getdefault_pseudo(atom::Symbol, pseudo_setname=:default; pseudospecifier=nothing)

Returns the pseudo potential string linked to the atom.

DFControl.getdefault_pseudodirs — Method.

getdefault_pseudodirs()

Returns the default pseudo dirs if it's defined. If it is not defined return nothing.

DFControl.getdefault_server — Method.

getdefault_server()

Returns the default server if it's defined. If it is not defined return "".

DFControl.removedefault_input — Method.

removedefault_input(input::Symbol)

Remove input from the default_inputs variable. Also removes the stored input file.

DFControl.removedefault_pseudodir — Method.

removedefault_pseudodir(pseudo_symbol::Symbol)

Removes entry with flag pseudo_symbol from the default_pseudodirs and user_defaults.jl file.

DFControl.removedefault_pseudos — Method.

removedefault_pseudos(pseudo_symbol::Symbol)

Removes all pseudo entries with flag pseudo_symbol from the default_pseudos.

DFControl.setdefault_input — Method.

setdefault_input(input::dfinput, structure, calculation::Symbol)

Adds the input to the default_inputs variable, and writes it to a file in user_defaults folder to be read every time on load.

DFControl.setdefault_jobheader — Method.

setdefault_jobheader(lines)

Sets the header that will get added to each job.tt file, if no other header was specified.

DFControl.setdefault_pseudodir — Method.

setdefault_pseudodir(pseudo_symbol::Symbol, dir::String)

Adds an entry inside the default_pseudodirs with flag pseudo_symbol, and adds it to the user_defaults.jl file.

DFControl.setdefault_server — Method.

setdefault_server(server::String)

Sets the default server variable, and also adds it to the user_defaults.jl file.