FileIO

FileIO

DFControl.read_fermi_from_qe_output — Function.

read_fermi_from_qe_output(filename::String,T=Float64)

Reads the Fermi level from a Quantum Espresso scf calculation output file (if there is one).

DFControl.read_ks_from_qe_output — Function.

read_ks_from_qe_output(filename::String, T=Float64)

Read k-points from a Quantum Espresso bands output file in cartesian (2pi/alat in Angstrom^-1!) and crystalline coordinates. Returns (kpointscart,kpointscryst).

DFControl.read_qe_bands_file — Function.

read_qe_bands(filename::String, T=Float64)

Reads the output file of a 'bands' calculation in Quantum Espresso. Returns an array of DFBands each with the same k_points and their respective energies.

DFControl.read_qe_input — Function.

read_qe_input(filename, T=Float64; exec="pw.x",  runcommand="", run=true, structure_name="NoName")

Reads a Quantum Espresso input file. The exec get's used to find which flags are allowed in this input file, and convert the read values to the correct Types. Returns a DFInput{QE} and the Structure that is found in the input.

DFControl.read_qe_kpdos — Function.

read_qe_kpdos(filename::String,column=1;fermi=0)

Reads the kresolved partial density of states from a Quantum Espresso projwfc output file. Only use this if the flag kresolveddos=true in the projwfc input file!! The returned matrix can be readily plotted using heatmap() from Plots.jl! Optional input: column = 1 (column of the output, 1 = first column after ik and E) fermi = 0 (possible fermi offset of the read energy values) Return: Array{Float64,2}(length(kpoints),length(energies)) , (ytickvals,yticks)

DFControl.read_qe_output — Function.

read_qe_output(filename::String, T=Float64)

Reads a generic quantum espresso input, returns a dictionary with all found data in the file. Possible keys:

:fermi

:polarization

:pol_mod

:k_cryst

:k_cart

:alat

:cell_parameters

:pos_option

:atomic_positions

:total_force

:colin_mag_moments

:bands

:accuracy

DFControl.read_qe_pdos — Function.

read_qe_pdos(filename::String, column=1; fermi=0)

Reads partial dos file. One can specify the column of values to read.

DFControl.read_qe_polarization — Function.

read_qe_polarization(filename::String, T=Float64)

Returns the polarization and modulus.

DFControl.read_wannier_input — Function.

read_wannier_input(filename::String, T=Float64; runcommand= Exec(""), run=true, exec=Exec("wannier90.x"), structure_name="NoName")

Reads a DFInput{Wannier90} and the included Structure from a WANNIER90 input file.

DFControl.save — Function.

save(input::DFInput{QE}, structure, filename::String=inpath(input))

Writes a Quantum Espresso input file.

DFControl.save — Function.

save(input::DFInput{Wannier90}, structure, filename::String=inpath(input))

Writes the DFInput{Wannier90} and structure to a file, that can be interpreted by WANNIER90. The atoms in the structure must have projections defined.