Utils

Utils

DFControl.kgrid — Method.

kgrid(na, nb, nc, input)

Returns an array of k-grid points that are equally spaced, input can be either :wan or :nscf, the returned grids are appropriate as inputs for wannier90 or an nscf calculation respectively.