Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DPClustering.dpclustering — Method.

dpclustering(y::Array{Real, 1}, N::Array{Real, 1}; <keyword arguments>)

Perform dirichlet clustering on the variant allele frequency distribution of cancer sequencing data and find the number of clusters that the data supports, y is a vector of the number of reads reporting each mutant, N is the total depth at each locus.

...

Arguments

iterations = 1000: number of iterations of the gibbs samples

C = 30: Max number of clusters to consider

burninstart = round(Int64, iterations/2): Burn in of the gibbs samples

bw = 0.01: Bandwidth of density estimation

maxxaxis = 0.7:

cutoffweight = 0.05: Minimum weight to be called a cluster

verbose = true: Show progress of gibbs sampling with ProgressMeter package

A = 0.01: Hyperparameter for α, see Nik-Zainal et al

B = 0.01: Hyperparameter for α, see Nik-Zainal et al

...

DPClustering.plotresults — Method.

plotresults(dp; <keyword arguments>)

Plot results from DPClustering object. Will plot histogram of raw data with density estimates from Gibbs sampling. ...

Arguments

save = false: Set to true if you want the plot to be saved

dir = "": Directory where the plot will be saved to. Default is the current working directory.

plotname = "DPclustering": Name to call plot when saving.

...