Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Dagger.Context — Type.

A context represents a set of processors to use for a papply operation.

Dagger.OSProc — Type.

OS process - contains pid returned by addprocs

Dagger.UnitDomain — Type.

Default domain – has no information about the value

Dagger.alignfirst — Method.

alignfirst(a)

Make a subdomain a standalone domain. For example,

alignfirst(ArrayDomain(11:25, 21:100))
# => ArrayDomain((1:15), (1:80))

Dagger.cached_stage — Method.

A memoized version of stage. It is important that the tasks generated for the same DArray have the same identity, for example:

A = rand(Blocks(100,100), Float64, 1000, 1000)
compute(A+A')

must not result in computation of A twice.

Dagger.compute — Method.

A DArray object may contain a thunk in it, in which case we first turn it into a Thunk object and then compute it.

Dagger.compute — Method.

Compute a Thunk - creates the DAG, assigns ranks to nodes for tie breaking and runs the scheduler.

Dagger.domain — Function.

domain(x::T)

Returns metadata about x. This metadata will be in the domain field of a Chunk object when an object of type T is created as the result of evaluating a Thunk.

Dagger.domain — Method.

The domain of an array is a ArrayDomain

Dagger.domain — Method.

If no domain method is defined on an object, then we use the UnitDomain on it. A UnitDomain is indivisible.

Dagger.load — Method.

load(ctx, file_path)

Load an Union{Chunk, Thunk} from a file.

Dagger.load — Method.

load(ctx, ::Type{Chunk}, fpath, io)

Load a Chunk object from a file, the file path is required for creating a FileReader object

Dagger.save — Method.

save(io::IO, val)

Save a value into the IO buffer. In the case of arrays and sparse matrices, this will save it in a memory-mappable way.

load(io::IO, t::Type, domain) will load the object given its domain

Dagger.save — Method.

save(ctx, chunk::Union{Chunk, Thunk}, file_path::AbsractString)

Save a chunk to a file at file_path.

Dagger.save — Method.

special case distmem writing - write to disk on the process with the chunk.

Dagger.Chunk — Type.

A chunk with some data

Dagger.DArray — Type.

DArray{T,N,F}(domain, subdomains, chunks, concat)

An N-dimensional distributed array of element type T.

domain: the whole ArrayDomain of the array

subdomains: a DomainBlocks of the same dimensions as the array

chunks: an array of chunks of dimension N

concat: a function of type F. concat(x, y; dims=d) takes two chunks x and y and concatenates them along dimension d. cat is used by default.

Dagger.DArray — Method.

DArray(T, domain, subdomains, chunks, [concat=cat])

Creates a distributed array of element type T.

T: element type

rest of the arguments are the same as the DArray constructor.

Dagger.LocalEventLog — Type.

represents a process local events array.

A context with log_sink set to LocalEventLog() will cause events to be recorded into the

Dagger.NoOpLog — Type.

Various means of writing an event to something.

Dagger.PromotePartition — Type.

This is a way of suggesting that stage should call stage_operand with the operation and other arguments

Dagger.State — Type.

Overall state used during visualization

Dagger.Timespan — Type.

identifies

space (category, id) time (timeline, start, finish)

also tracks gc_num during this and profiling samples.

Base.view — Method.

view of a DArray chunk returns a DArray of thunks

Dagger.add_span — Method.

Add a Timespan to a given State under tl (timeline) and category.

Dagger.dependents — Function.

find the set of direct dependents for each task

Dagger.dsort_chunks — Function.

dsort_chunks(cs, [nchunks, nsamples]; options...)

Sort contents of chunks (cs) and return a new set of chunks such that the chunks when concatenated return a sorted collection. Each chunk in turn is sorted.

Args:

nchunks – the number of chunks to produce, regardless of how many chunks were given as input

nsamples – the number of elements to sample from each chunk to guess the splitters (nchunks-1 splitters) each chunk will be delimited by the splitter.

merge – a function to merge two sorted collections.

sub – a function to get a subset of the collection takes (collection, range) (defaults to getindex)

order – Base.Sort.Ordering to be used for sorting

batchsize – number of chunks to split and merge at a time (e.g. if there are 128 input chunks and 128 output chunks, and batchsize is 8, then we first sort among batches of 8 chunks – giving 16 batches. Then we sort among the first chunk of the first 8 batches (all elements less than the first splitter), then go on to the first 8 chunks of the second 8 batches, and so on...

chunks_presorted – is each chunk in the input already sorted?

sortandsample – a function to sort a chunk, then sample N elements to infer the splitters. It takes 3 arguments: (collection, N, presorted). presorted is a boolean which is true if the chunk is already sorted.

Returns a tuple of (chunk, samples) where chunk is the Dagger.Chunk object. chunk can be nothing if no change to the initial array was made (e.g. it was already sorted)

affinities – a list of processes where the output chunks should go. If the length is not equal to nchunks then affinities array is cycled through.

Dagger.get_logs! — Method.

Get the logs from each process, clear it too

Dagger.make_timespan — Method.

create a timespan given the strt and finish events

Dagger.next_state — Method.

When building state for real-time visualization, use next_state to progress gantt state.

Dagger.noffspring — Method.

recursively find the number of taks dependent on each task in the DAG. Input: dependents dict

Dagger.order — Method.

Given a root node of the DAG, calculates a total order for tie-braking

Root node gets score 1,

rest of the nodes are explored in DFS fashion but chunks of each node are explored in order of noffspring, i.e. total number of tasks depending on the result of the said node.

Args: - node: root node - ndeps: result of noffspring

Dagger.recursive_splitters — Method.

split the splitters themselves into batches.

Args:

ord – Sorting.Ordering object

splitters – the nchunks-1 splitters

batchsize – batch size

Returns: A Tuple{Vector, Vector{Vector}} – the coarse splitters which create batchsize splits, finer splitters within those batches which create a total of nchunks splits.

julia> Dagger.recursive_splitters(Dagger.default_ord,
            [10,20,30,40,50,60], 5,3)
([30], Any[[10, 20], [40, 50, 60]])

The first value [30] represents a coarse split that cuts the dataset from -Inf-30, and 30-Inf. Each part is further recursively split using the next set of splitters

Dagger.split_range — Method.

Utility function to divide the range range into n chunks

Dagger.split_range_interval — Method.

split_range_interval(range, n)

split a range into pieces each of length n or lesser

Dagger.stage_operands — Method.

an operand which should be distributed as per convenience

Dagger.thunkize — Method.

If a DArray tree has a Thunk in it, make the whole thing a big thunk

Dagger.tochunk — Method.

Create a chunk from a sequential object.

Dagger.treereduce — Method.

Tree reduce

Dagger.write_event — Method.

Write a log event