Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DataDepsGenerators.citation_text — Method.

citation_text(doi)

Uses the DOI formatted citation service to generate citation text for a given DOI. This works for DOI's issued by: CrossRef, DataCite, and mEDRA.

See https://citation.crosscite.org/docs.html#sec-4-1

DataDepsGenerators.generate — Function.

generate([repo/s], url/id, [shortname]; show_failures=false)

Generates a DataDeps code block. The only reuired parameter is the url/id.

url/id The identifier for the dataset

a URL for a landing page is normally best

Other IDs like a DOIs also work.

repo/s either a single repository/API or a list of such

this takes on of the DataRepo types exported by this package.

E.g CKAN(), or Figshare().

If not provided, this defaults to checking all of them.

If only one repo is provided, and it fails, the error will be thrown.

If multiple repos are provided, them the metadata from all of them is combined.

shortname, the name to use in the generated DataDep

if not provided will use the dataset's title, but these are often very long.

show_failures, weather or not to list all the repos that fail and why.

You generally do not want to turn this on, unless you are debugging.

It is fine and expected for most repos to fail (after all the data is probably only on one of them)

If all repos fail, then the failure list will be shown, regardless of if this is set or not.

DataDepsGenerators.aggregate — Method.

aggregate(metadatas)

Given a collection of Metadata from differenct sources, combined them to create the most complete and detailed accounting of metadata.

DataDepsGenerators.escape_multiline_string — Method.

escape_multiline_string

like Escape string, but does not escape newlines

DataDepsGenerators.filter_html — Method.

filter_html(text)

Strips any HTML tags out of the text. If that is required.

DataDepsGenerators.get_dataurls_from_webserver_index — Method.

links_from_webserver_index(url)

Extracts all the content links from a webservers directory index page. These follow a pretty standard form. This one is tested so far on Apache/2.2.15

DataDepsGenerators.getfirst — Method.

getfirst(dict, keys...)

Returns the element coresponding to the first key that is found. Returns missing if no key is found.

DataDepsGenerators.getpage — Function.

getpage(url)

downloads and parses the page from the URL

DataDepsGenerators.getpage_raw — Method.

getpage_raw(url)

Downloads the page from the URL, returning the raw (unparsed) text of the body.

DataDepsGenerators.indent — Function.

indent(str)

Indents each line in a string

DataDepsGenerators.leaf_subtypes — Method.

leaf_subtypes(T)

Returns all the nonabstract types decedent from T.

DataDepsGenerators.lift — Method.

lift(func, arg)

Calls func(arg), propagating missing values

DataDepsGenerators.match_doi — Method.

match_doi(uri::String

DataDepsGenerators.text_only — Method.

text_only(doc)

Extracts just the unformatted text (no attributes etc), from a HTML document or fragment(/s)