Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DataValues.DataValueArray — Type.

DataValueArray{T, N} is an efficient alternative to Array{DataValue{T}, N}.

DataValues.dropna! — Method.

dropna!(X::DataValueVector)

Remove missing entries of X in-place and return a Vector view of the unwrapped DataValue entries.

DataValues.dropna! — Method.

dropna!(X::AbstractVector)

Remove missing entries of X in-place and return a Vector view of the unwrapped DataValue entries. If no missing values are present, this is a no-op and X is returned.

DataValues.dropna — Method.

dropna(X::AbstractVector)

Return a vector containing only the non-missing entries of X, unwrapping DataValue entries. A copy is always returned, even when X does not contain any missing values.

DataValues.padna! — Method.

padna!(X::DataValueVector, front::Integer, back::Integer)

Insert front null entries at the beginning of X and add back null entries at the end of X. Returns X.

DataValues.padna — Method.

padna(X::DataValueVector, front::Integer, back::Integer)

return a copy of X with front null entries inserted at the beginning of the copy and back null entries inserted at the end.

Base.append! — Method.

append!(X::DataValueVector, items::AbstractVector)

Add the elements of items to the end of X.

Note that append!(X, [1, 2, 3]) is equivalent to push!(X, 1, 2, 3), where the items to be added to X are passed individually to push! and as a collection to append!.

Base.convert — Method.

convert(T, X::DataValueArray)

Convert X to an AbstractArray of type T. Note that if X contains any null entries then calling convert without supplying a replacement value for null entries will result in an error.

Currently supported return type arguments include: Array, Array{T}, Vector, Matrix.

convert(T, X::DataValueArray, replacement)

Convert X to an AbstractArray of type T and replace all null entries of X with replacement in the result.

Base.copy — Method.

copy(X::DataValueArray)

Return a shallow copy of X; the outer structure of X will be copied, but all elements will be identical to those of X.

Base.copyto! — Method.

copyto!(dest::DataValueArray, src::DataValueArray)

Copy the initialized values of a source DataValueArray into the respective indices of the destination DataValueArray. If an entry in src is null, then this method nullifies the respective entry in dest.

Base.deepcopy — Method.

Base.deepcopy(X::DataValueArray)

Return a DataValueArray object whose internal values and isna fields are deep copies of X.values and X.isna respectively.

Base.deleteat! — Method.

deleteat!(X::DataValueVector, inds)

Delete the entry at inds from X and then return X. Note that inds may be either a single scalar index or a collection of sorted, pairwise unique indices. Subsequent items after deleted entries are shifted down to fill the resulting gaps.

Base.empty! — Method.

empty!(X::DataValueVector) -> DataValueVector

Remove all elements from a DataValueVector. Returns DataValueVector{T}(), where T is the eltype of X.

Base.fill! — Method.

fill!(X::DataValueArray, x::DataValue)

Fill X with the value x and designate each entry as present. If x is an object reference, all elements will refer to the same object. Note that fill!(X, Foo()) will return X filled with the result of evaluating Foo() once.

Base.fill! — Method.

fill!(X::DataValueArray, x::DataValue)

Fill X with the value x. If x is empty, then fill!(X, x) nullifies each entry of X. Otherwise, fill!(X, x) fills X.values with the value of x and designates each entry of X as present.

Base.getindex — Method.

getindex(x::DataValue)

Attempt to access the value of x. Throw a DataValueException if the value is not present. Usually, this is written as x[].

Base.getindex — Method.

getindex{T,N}(X::DataValueArray{T,N}, I::DataValue{Int}...)

Just as above, with the additional behavior that this method throws an error if any component of the index I is missing.

Base.getindex — Method.

getindex{T,N}(X::DataValueArray{T,N}, I::Int...)

Retrieve a single entry from a DataValueArray. If the value in the entry designated by I is present, then it will be returned wrapped in a DataValue{T} container. If the value is missing, then this method returns DataValue{T}().

Base.lastindex — Method.

lastindex(X::DataValueArray)

Returns the last entry of X.

Base.length — Method.

length(X::DataValueArray)

Returns the maximum index i for which getindex(X, i) is valid.

Base.mapreduce — Method.

mapreduce(f, op::Function, X::DataValueArray; [skipna::Bool=false])

Map a function f over the elements of X and reduce the result under the operation op. One can set the behavior of this method to skip the null entries of X by setting the keyword argument skipna equal to true. If skipna behavior is enabled, f will be automatically lifted over the elements of X. Note that, in general, mapreducing over a DataValueArray will return a DataValue object regardless of whether skipna is set to true or false.

Base.ndims — Method.

ndims(X::DataValueArray)

Returns the number of dimensions of X.

Base.pop! — Method.

pop!{T}(X::DataValueVector{T})

Remove the last entry from X and return it. If the value in that entry is missing, then this method returns DataValue{T}().

Base.popfirst! — Method.

popfirst!{T}(X::DataValueVector{T})

Remove the first entry from X and return it as a DataValue object.

Base.prepend! — Method.

prepend!(X::DataValueVector, items::AbstractVector)

Add the elements of items to the beginning of X.

Note that prepend!(X, [1, 2, 3]) is equivalent to pushfirst!(X, 1, 2, 3), where the items to be added to X are passed individually to pushfirst! and as a collection to prepend!.

Base.push! — Method.

push!{T,V}(X::DataValueVector{T}, v::DataValue{V})

Insert a value at the end of X from a DataValue value v. If v is null then this method adds a null entry at the end of X. Returns X.

Base.push! — Method.

push!{T,V}(X::DataValueVector{T}, v::V)

Insert v at the end of X, which registers v as a present value.

Base.pushfirst! — Method.

pushfirst!(X::DataValueVector, v)

Insert a value v at the beginning of X and return X.

Base.pushfirst! — Method.

pushfirst!(X::DataValueVector, v::DataValue)

Insert a value at the beginning of X from a DataValue value v. If v is null then this method inserts a null entry at the beginning of X. Returns X.

Base.reduce — Method.

reduce(f, op::Function, X::DataValueArray; [skipna::Bool=false])

Reduce Xunder the operation op. One can set the behavior of this method to skip the null entries of X by setting the keyword argument skipna equal to true. If skipna behavior is enabled, f will be automatically lifted over the elements of X. Note that, in general, mapreducing over a DataValueArray will return a DataValue object regardless of whether skipna is set to true or false.

Base.resize! — Method.

resize!(X::DataValueVector, n::Int)

Resize a one-dimensional DataValueArray X to contain precisely n elements. If n is greater than the current length of X, then each new entry will be designated as null.

Base.reverse — Function.

reverse(X::DataValueVector, [s], [n])

Return a copy of X with the first n elements starting at index s (inclusive) reversed. If unspecified, s and n will default to 1 and length(X), respectively.

Base.reverse! — Function.

reverse!(X::DataValueVector, [s], [n])

Modify X by reversing the first n elements starting at index s (inclusive). If unspecified, s and n will default to 1 and length(X), respectively.

Base.setindex! — Method.

setindex!(X::DataValueArray, v::Any, I::Int...)

Set the entry of X at position I equal to v. This method always updates X.isna to indicate that the entry at index I is present and X.values to store v at I.

Base.setindex! — Method.

setindex!(X::DataValueArray, v::DataValue, I::Int...)

Set the entry of X at position I equal to a DataValue value v. If v is missing, then only X.isna is updated to indicate that the entry at index I is null. If v is not null, then X.isna is updated to indicate that the entry at index I is present and X.values is updated to store the value wrapped in v.

Base.size — Method.

size(X::DataValueArray, [d::Real])

Return a tuple containing the lengths of each dimension of X, or if d is specific, the length of X along dimension d.

Base.sizehint! — Method.

sizehint!(X::DataValueVector, newsz::Integer)

Suggest that collection X reserve capacity for at least newsz elements. This can improve performance.

Base.splice! — Function.

splice!(X::DataValueVector, i::Integer, [ins])

Remove the item at index i and return the removed item. Subsequent items are shifted down to fill the resulting gap. If specified, replacement values from an ordered collection will be spliced in place of the removed item.

Base.splice! — Method.

splice!{T<:Integer}(X::DataValueVector, rng::UnitRange{T}, [ins])

Remove items in the specified index range, and return a collection containing the removed items. Subsequent items are shifted down to fill the resulting gap. If specified, replacement values from an ordered collection will be spliced in place of the removed items.

To insert ins before an index n without removing any items, use splice!(X, n:n-1, ins).