Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DatagenCopulaBased.archcopulagen — Method.

archcopulagen(t::Int, n::Int, θ::Union{Float64, Int}, copula::String; rev::Bool = false)

Returns: t x n Matrix{Float}, t realisations of n-variate data generated from Archimedean one parameter copula.

Following copula families are supported: clayton, frank, gumbel and amh – Ali-Mikhail-Haq.

If rev == true, reverse the copula output i.e. u → 1-u (we call it reversed copula). It cor == pearson, kendall, uses correlation coeficient as a parameter

```jldoctest julia> Random.seed!(43);

julia> archcopulagen(10, 2, 1, "clayton") 10×2 Array{Float64,2}: 0.770331 0.932834 0.472847 0.0806845 0.970749 0.653029 0.622159 0.0518025 0.402461 0.228549 0.946375 0.842883 0.809076 0.129038 0.747983 0.433829 0.374341 0.437269 0.973066 0.910103

```

DatagenCopulaBased.chainfrechetcopulagen — Function.

chainfrechetcopulagen(t::Int, α::Vector{Float64}, β::Vector{Float64} = zeros(α))

Retenares data from a chain of bivariate two parameter frechet copuls with parameters vectors α and β, such that ∀ᵢ 0 α[i] + β[i] ≤1 α[i] > 0, and β[i] > 0 |α| = |β|

julia> Random.seed!(43)

julia> julia> chainfrechetcopulagen(10, [0.6, 0.4], [0.3, 0.5])
10×3 Array{Float64,2}:
 0.996764  0.996764  0.996764
 0.204033  0.795967  0.204033
 0.979901  0.979901  0.0200985
 0.120669  0.879331  0.120669
 0.453027  0.453027  0.453027
 0.800909  0.199091  0.800909
 0.54892   0.54892   0.54892
 0.933832  0.933832  0.0661679
 0.396943  0.396943  0.396943
 0.804096  0.851275  0.955881

DatagenCopulaBased.cormatgen — Function.

cormatgen(n::Int, ρ::Float64 = 0.5, ordered::Bool = false, altersing::Bool = true)

Returns symmetric correlation matrix Σ of size n x n, with reference correlation 0 < ρ < 1. If ordered = false, Σ elements varies arround ρ, i.e. σᵢⱼ ≈ ρ+δ else they drop as indices differences rise, i.e. σᵢⱼ ≳ σᵢₖ as |i-j| < |i-k|. If altersing = true, some σ are positive and some negative, else ∀ᵢⱼ σᵢⱼ ≥ 0.

julia> Random.seed!(43);

julia> julia> cormatgen(2)
2×2 Array{Float64,2}:
 1.0       0.660768
 0.660768  1.0

DatagenCopulaBased.frechetcopulagen — Method.

frechetcopulagen(t::Int, n::Int, α::Union{Int, Float64}, β::Union{Int, Float64})

Two parameters Frechet copula C = α C{max} + β C{min} + (1- α - β) C_{⟂}, supported only for n == 2

julia> Random.seed!(43);

julia> frechetcopulagen(10, 2, 0.4, 0.2)
10×2 Array{Float64,2}:
 0.180975  0.661781
 0.775377  0.775377
 0.888934  0.125437
 0.924876  0.924876
 0.408278  0.591722
 0.912603  0.740184
 0.828727  0.171273
 0.400537  0.0288987
 0.429437  0.429437
 0.955881  0.851275

DatagenCopulaBased.frechetcopulagen — Method.

frechetcopulagen(t::Int, n::Int, α::Union{Int, Float64})

Returns t realisation of n variate data generated from one parameter frechet multidimentional copula, a combination of maximal copla with weight α and independent copula with weight 1-α

julia> Random.seed!(43);

julia> frechetcopulagen(10, 2, 0.5)
10×2 Array{Float64,2}:
 0.180975  0.661781
 0.775377  0.775377
 0.888934  0.125437
 0.924876  0.924876
 0.408278  0.408278
 0.912603  0.740184
 0.828727  0.00463791
 0.400537  0.0288987
 0.429437  0.429437
 0.955881  0.851275

DatagenCopulaBased.gausscopulagen — Function.

gausscopulagen(t::Int, Σ::Matrix{Float64} = [1. 0.5; 0.5 1.])

Returns: t x n matrix of t realisations of multivariate data generated using gaussian copula with Σ - correlation matrix. If the symmetric covariance matrix is imputed, it will be converted into a correlation matrix automatically.

julia> Random.seed!(43);

julia> gausscopulagen(10)
10×2 Array{Float64,2}:
 0.589188  0.815308
 0.708285  0.924962
 0.747341  0.156994
 0.227634  0.183116
 0.227575  0.957376
 0.271558  0.364803
 0.445691  0.52792
 0.585362  0.23135
 0.498593  0.48266
 0.190283  0.594451

DatagenCopulaBased.gcop2arch — Method.

gcop2arch(x::Matrix{Float64}, inds::VP)

Takes x ∈ Rᵗˣⁿ a matrix of t realisations of data from Gaussin n-variate distribution. Return a matrix of size x, where chosen subset of marginals (inds[i][2]) has an Archimedean sub-copula (denoted by inds[i][1]), all univariate marginals are unchanged. e.g. inds = ["clayton" => [1,2]] a subset of marginals indexed by 1,2 to such with Clayton sub-copula

julia> Σ = [1. 0.5 0.5; 0.5 1. 0.5; 0.5 0.5 1];

julia> Random.seed!(42)

julia> x = rand(MvNormal(Σ), 6)'
6×3 Array{Float64,2}:
 -0.556027  -0.662861   -0.384124
 -0.299484   1.38993    -0.571326
 -0.468606  -0.0990787  -2.3464
  1.00331    1.43902     0.966819
  0.518149   1.55065     0.989712
 -0.886205   0.149748   -1.54419

julia> gcop2arch(x, ["clayton" => [1,2]]; naive::Bool = false, notnested::Bool = false)
6×3 Array{Float64,2}:
 -0.742443   0.424851  -0.384124
  0.211894   0.195774  -0.571326
 -0.989417  -0.299369  -2.3464
  0.157683   1.47768    0.966819
  0.154893   0.893253   0.989712
 -0.657297  -0.339814  -1.54419

DatagenCopulaBased.gcop2frechet — Method.

gcop2frechet(x::Matrix{Float64}, inds::VP)

Takes x ∈ Rᵗˣⁿ a matrix of t realisations of data from Gaussin n-variate distribution. Return a matrix of size x, where chosen subset of marginals (inds) has Frechet maximal sub-copula, all univariate marginals are unchanged.

julia> Σ = [1. 0.5 0.5; 0.5 1. 0.5; 0.5 0.5 1];

julia> Random.seed!(42)

julia> x = rand(MvNormal(Σ), 6)'
6×3 Array{Float64,2}:
 -0.556027  -0.662861   -0.384124
 -0.299484   1.38993    -0.571326
 -0.468606  -0.0990787  -2.3464
  1.00331    1.43902     0.966819
  0.518149   1.55065     0.989712
 -0.886205   0.149748   -1.54419

julia> gcop2frechet(x, [1,2]; naive = false)
6×3 Array{Float64,2}:
 -0.875777   -0.374723   -0.384124
  0.0960334   0.905703   -0.571326
 -0.599792   -0.0110945  -2.3464
  0.813717    1.8513      0.966819
  0.599255    1.56873     0.989712
 -0.7223     -0.172507   -1.54419

DatagenCopulaBased.gcop2marshallolkin — Function.

gcop2marshallolkin(x::Matrix{Float64}, inds::Vector{Int}, λ1::Float64 = 1., λ2::Float64 = 1.5; naive::Bool = false)

Takes x ∈ Rᵗˣⁿ a matrix of t realisations of data from Gaussin n-variate distribution. Return a matrix of size x, where chosen subset of marginals (inds) has bivariate Marshal Olkin sub-copula parametrised by free parameters λ1 and λ2. All univariate marginals are unchanged.

julia> Σ = [1. 0.5 0.5; 0.5 1. 0.5; 0.5 0.5 1];

julia> Random.seed!(42)

julia> x = rand(MvNormal(Σ), 6)'
6×3 Array{Float64,2}:
 -0.556027  -0.662861   -0.384124
 -0.299484   1.38993    -0.571326
 -0.468606  -0.0990787  -2.3464
  1.00331    1.43902     0.966819
  0.518149   1.55065     0.989712
 -0.886205   0.149748   -1.54419

julia> gcop2marshallolkin(x, [1,2], 1., 1.5; naive = false)
6×3 Array{Float64,2}:
 -0.790756   0.784371  -0.384124
 -0.28088    0.338086  -0.571326
 -0.90688   -0.509684  -2.3464
  0.738628   1.71026    0.966819
  0.353654   1.19357    0.989712
 -0.867606  -0.589929  -1.54419

DatagenCopulaBased.gcop2tstudent — Method.

gcop2tstudent(x::Matrix{Float64}, ind::Vector{Int}, ν::Int)

Takes x ∈ Rᵗˣⁿ a matrix of t realisations of data from Gaussin n-variate distribution. Return a matrix of size x, where chosen subset of marginals (ind) has a t-Student copula with ν degrees of freedom, all univariate marginals are unchanged

julia> Σ = [1. 0.5 0.5; 0.5 1. 0.5; 0.5 0.5 1];

julia> Random.seed!(42)

julia> x = rand(MvNormal(Σ), 6)'
6×3 Array{Float64,2}:
 -0.556027  -0.662861   -0.384124
 -0.299484   1.38993    -0.571326
 -0.468606  -0.0990787  -2.3464
  1.00331    1.43902     0.966819
  0.518149   1.55065     0.989712
 -0.886205   0.149748   -1.54419

julia> gcop2tstudent(x, [1,2], 6)
6×3 Array{Float64,2}:
 -0.514449  -0.49147    -0.384124
 -0.377933   1.66254    -0.571326
 -0.430426  -0.0165044  -2.3464
  0.928668   1.50472     0.966819
  0.223439   1.12372     0.989712
 -0.710786   0.239012   -1.54419

DatagenCopulaBased.marshallolkincopulagen — Function.

marshallolkincopulagen(t::Int, λ::Vector{Float64})

Returns: t x n Matrix{Float}, t realisations of n-variate data generated from Marshall-Olkin copula with parameter vector λ of non-negative elements λₛ. Number of marginals is n = ceil(Int, log(2, length(λ)-1)). Parameters are ordered as follow: λ = [λ₁, λ₂, ..., λₙ, λ₁₂, λ₁₃, ..., λ₁ₙ, λ₂₃, ..., λₙ₋₁ₙ, λ₁₂₃, ..., λ₁₂...ₙ] If reversed = true, returns data from reversed Marshall-Olkin copula.

```jldoctest

julia> Random.seed!(43)

julia> marshallolkincopulagen(10, [0.2, 1.2, 1.6]) 10×2 Array{Float64,2}: 0.99636 0.994344 0.167268 0.0619408 0.977418 0.965093 0.495167 0.0247053 0.410336 0.250159 0.778989 0.678064 0.50927 0.350059 0.925875 0.887095 0.353646 0.219006 0.782477 0.686799 ```

DatagenCopulaBased.nestedarchcopulagen — Function.

nestedarchcopulagen(t::Int, n::Vector{Int}, ϕ::Vector{Float64}, θ::Float64, copula::String, m::Int = 0)

Returns Matrix{Float} of t realisations of sum(n)+m random variables generated using nested archimedean copula, outer copula parameter is θ, inner i'th copulas parameter is ϕ[i] and size is n[i]. If m ≠ 0, last m variables are from outer copula only.

Following copula families are supported: clayton, frank, gumbel and amh – Ali-Mikhail-Haq.

Nested archimedean copula in in a form Cθ(Cϕ₁(u₁₁, ..., u₁,ₙ₁), ..., C_ϕₖ(uₖ₁, ..., uₖ,ₙₖ), u₁ , ... uₘ). Parameters must fulfill ∀ᵢ ϕ[i] ≥ θ

Basically uses Alg. 5 McNeil, A.J., 2008. 'Sampling nested Archimedean copulas'. Journal of Statistical Computation and Simulation 78, 567–581.

julia> Random.seed!(43);

julia> nestedarchcopulagen(10, [2,2], [2., 3.], 1.1, "clayton", 1)
10×5 Array{Float64,2}:
 0.414567  0.683167   0.9953    0.607738  0.793386
 0.533001  0.190563   0.17076   0.273119  0.78807
 0.572782  0.161307   0.418821  0.110356  0.661781
 0.623807  0.140974   0.295422  0.454368  0.477065
 0.386276  0.266261   0.559423  0.449874  0.294137
 0.219757  0.122586   0.371318  0.298965  0.507315
 0.322658  0.0627113  0.738565  0.919912  0.19471
 0.131938  0.0672061  0.364721  0.220329  0.662842
 0.773414  0.812113   0.639333  0.527118  0.545043
 0.958656  0.871822   0.958339  0.801866  0.862751

DatagenCopulaBased.nestedarchcopulagen — Function.

nestedarchcopulagen::Int, n::Vector{Vector{Int}}, Ψ::Vector{Vector{Float64}}, ϕ::Vector{Float64}, θ₀::Float64, copula::String = "gumbel")

Returns t realisations of ∑ᵢ ∑ⱼ nᵢⱼ variate data from double nested gumbel copula. Cθ(Cϕ₁(CΨ₁₁(u,...), ..., CCΨ₁,ₗ₁(u...)), ..., Cϕₖ(CΨₖ₁(u,...), ..., CΨₖ,ₗₖ(u,...))) where lᵢ = length(n[i])

```jldoctest julia> Random.seed!(43)

julia> nestedarchcopulagen(5, [[2,2],[2]], [[3., 2.], [4.]], [1.5, 2.1], 1.2, "gumbel") 5×6 Array{Float64,2}: 0.464403 0.711722 0.883035 0.896706 0.888614 0.826514 0.750596 0.768193 0.0659561 0.0252472 0.996014 0.989127 0.825211 0.712079 0.581356 0.507739 0.882675 0.84959 0.276326 0.0827071 0.240836 0.434629 0.0184611 0.031363 0.208422 0.504727 0.27561 0.639089 0.481855 0.573715 ```

DatagenCopulaBased.nestedarchcopulagen — Function.

nestedarchcopulagen(t::Int, θ::Vector{Float64}, copula::String = "gumbel")

Returns t realisations of length(θ)+1 variate data from hierarchically nested Gumbel copula. Cθₙ(... Cθ₂(C_θ₁(u₁, u₂), u₃)..., uₙ)

julia> Random.seed!(43)

julia> x = nestedarchcopulagen(5, [4., 3., 2.], "gumbel")
5×4 Array{Float64,2}:
 0.483466  0.621572  0.241025  0.312664
 0.827237  0.696634  0.768802  0.730543
 0.401159  0.462126  0.412573  0.72571
 0.970726  0.964746  0.940314  0.934625
 0.684486  0.614142  0.690664  0.401897

DatagenCopulaBased.tstudentcopulagen — Function.

tstudentcopulagen(t::Int, Σ::Matrix{Float64} = [1. 0.5; 0.5 1.], ν::Int=10)

Generates data using t-student Copula given Σ - correlation matrix, ν - degrees of freedom. If the symmetric covariance matrix is imputed, it will be converted into a correlation matrix automatically.

```jldoctest julia> Random.seed!(43);

julia> tstudentcopulagen(10) 10×2 Array{Float64,2}: 0.658199 0.937148 0.718244 0.92602 0.809521 0.0980325 0.263068 0.222589 0.187187 0.971109 0.245373 0.346428 0.452336 0.524498 0.57113 0.272525 0.498443 0.48082 0.113788 0.633349 ```

DatagenCopulaBased.AMHτ2θ — Method.

AMHτ2θ(τ::Float64)

Returns Ali-Mikhail-Haq copula θ parameter, givem Kendall's τ

DatagenCopulaBased.Debye — Function.

Debye(x::Float64, k::Int)

Returns float64, Debye function Dₖ(x) value

DatagenCopulaBased.Ginv — Method.

Ginv(y::Float64, α::Float64)

Returns Float64, helper for the joe/frank nested copula generator

DatagenCopulaBased.InvlaJ — Method.

InvlaJ(n::Int, α::Float64)

Returns Float64, n-th element of the inverse laplacea transform of generator of Joe nested copula

DatagenCopulaBased.are_parameters_good — Method.

areparametersgood(ϕ::Vector{Float64}, θ::Float64)

tests sufficient nesting condition given parameters, returns bool

DatagenCopulaBased.copulagen — Method.

copulagen(copula::String, r::Matrix{Float}, θ::Union{Float64, Int})

Auxiliary function used to generate data from archimedean copula (clayton, gumbel, frank or amh) parametrised by a single parameter θ given a matrix of independent [0,1] distributerd random vectors.

  julia> copulagen("clayton", [0.2 0.6 0.9; 0.4 0.5 0.8], 2.)
  2×2 Array{Float64,2}:
   0.675778  0.851993
   0.687482  0.736394

DatagenCopulaBased.elInvlaF — Method.

elInvlaF(θ₁::Float64, θ₀::Float64)

Returns Int, a single sample of the inverse laplacea transform of the generator of nested Frank copula

DatagenCopulaBased.fncopulagen — Method.

fncopulagen(α::Vector{Float64}, β::Vector{Float64}, u::Matrix{Float64})

julia> fncopulagen(2, [0.2, 0.4], [0.1, 0.1], [0.2 0.4 0.6; 0.3 0.5 0.7])
2×3 Array{Float64,2}:
 0.6  0.4  0.2
 0.7  0.5  0.3

DatagenCopulaBased.frankτ2θ — Method.

frankτ2θ(τ::Float64)

Returns a Frank copula θ parameter, givem Kendall's τ

DatagenCopulaBased.frechet — Method.

frechet(α::Float64, u::Matrix{Float64})

returns a matrix of size u with data generated using Frechet maximal copula with parameter α

DatagenCopulaBased.getV0 — Method.

getV0(θ::Union{Float64, Int}, v::Vector{Float64}, copula::String)

Returns Vector{Float} or Vector{Int} of realisations of axiliary variable V0 used to ganarate data from 1d archimedean copula with parameter Θ, given v: realisations of 1d variable uniformly distributed on [0,1]

julia> getV0(2., [0.2, 0.4, 0.6, 0.8], "clayton")
4-element Array{Float64,1}:
 0.0641848
 0.274996
 0.708326
 1.64237

DatagenCopulaBased.getcors_advanced — Method.

getcors_advanced(x::Matrix{Float64})

clusters data on a basis of a correlation

DatagenCopulaBased.levyel — Method.

levyel(θ::Union{Int, Float64})

An element from Levy stable distribution with parameters α = 1/θ, β = 1, γ = (cos(pi/(2*θ)))^θ and δ = 0. Return Float, given parameter ϴ of dostribution

DatagenCopulaBased.levygen — Method.

Return a Vector{Float64} of of pseudo cdf of Levy stable distribution with parameters α = 1/θ, β = 1, γ = (cos(pi/(2*θ)))^θ and δ = 0, given a vector of Float64 - u

DatagenCopulaBased.logseriescdf — Method.

logseriescdf(p::Float64)

Returns a vector{Float} of the discrete cdf of logarithmic distribution

DatagenCopulaBased.logseriesquantile — Method.

logseriesquantile(p::Float64, v::Vector{Float64})

Returns a vector{Float} of the v[i] th quaintlie of logaritmic distribution with parameter p

DatagenCopulaBased.mean_outer — Method.

mean_outer(Σ::Matrix{Float64}, part::Vector{Vector{Int}})

returns a mean correlation excluding internal one is subsets determined by part

DatagenCopulaBased.meanΣ — Method.

meanΣ(Σ::Matrix{Float64})

Returns Float64, a mean of the mean of lower diagal elements of a matrix

julia> s= [1. 0.2 0.3; 0.2 1. 0.4; 0.3 0.4 1.];

julia> meanΣ(s)
0.3

DatagenCopulaBased.mocopula — Method.

mocopula(u::Matrix{Float64}, n::Int, λ::Vector{Float64})

Returns: t x n Matrix{Float}, t realisations of n-variate data generated from Marshall-Olkin copula with parameter vector λ of non-negative elements λₛ, given [0,1]ᵗˣˡ ∋ u, where l = 2ⁿ-1

  julia> mocopula([0.2 0.3 0.4; 0.3 0.4 0.6; 0.4 0.5 0.7], 2, [1., 1.5, 2.])
  3×2 Array{Float64,2}:
   0.252982  0.201189
   0.464758  0.409039
   0.585662  0.5357

DatagenCopulaBased.nestedcopulag — Method.

nestedcopulag(copula::String, n::Vector{Int}, ϕ::Vector{Float64}, θ::Float64, r::Matrix{Float64})

Given [0,1]ᵗˣˡ ∋ r , returns t realisations of l-1 variate data from nested archimedean copula

```jldoctest julia> Random.seed!(43)

julia> nestedcopulag("clayton", [2, 2], [2., 3.], 1.1, [0.1 0.2 0.3 0.4 0.5; 0.2 0.3 0.4 0.5 0.6]) 2×4 Array{Float64,2}: 0.193949 0.230553 0.515404 0.557686 0.712034 0.761276 0.190189 0.208867 ```

DatagenCopulaBased.nestedfrankgen — Method.

nestedfrankgen(θ₁::Float64, θ₀::Float64, V0::Vector{Int})

Return vector of int, samples of inverse laplacea trensform of nested Frank copula given parametes and V0 - vector of samples if invlaplace of perents copula

DatagenCopulaBased.nestedstep — Method.

nestedstep(copula::String, u::Matrix{Float64}, V0::Union{Vector{Float64}, Vector{Int}}, ϕ::Float64, θ::Float64)

Given u ∈ [0,1]ᵗⁿ and V0 ∈ ℜᵗ returns u ∈ [0,1]ᵗⁿ for a given archimedean nested copula with inner copulas parameters ϕ anu auter copula parameter θ

julia> nestedstep("clayton", [0.2 0.8; 0.1 0.7], [0.2, 0.4], 2., 1.5)
2×2 Array{Float64,2}:
 0.283555  0.789899
 0.322614  0.806915

DatagenCopulaBased.norm2unifind — Function.

norm2unifind(x::Matrix{Float64}, i::Vector{Int}, cop::String)

Return uniformly distributed data from x[:,i] given a copula familly.

DatagenCopulaBased.parameters — Method.

parameters(x::Matrix{Float64}, part::Vector{Vector{Int}})

Returns parametrization by correlation for data x and partition part for nested copulas.

DatagenCopulaBased.phi — Method.

phi(u::Matrix{Float64}, θ::Union{Float64, Int}, copula::String)

Given a matrix t realisations of n variate data ℜᵗⁿ ∋ u = -log(rand(t,n))./V0 returns it transformed through an inverse generator of archimedean copula. Output is distributed uniformly on [0,1]ⁿ

julia> julia> phi([0.2 0.6; 0.4 0.8], 2., "clayton")
2×2 Array{Float64,2}:
 0.845154  0.6742
 0.745356  0.620174

DatagenCopulaBased.rand2cop — Method.

rand2cop(u1::Vector{Float64}, θ::Union{Int, Float64}, copula::String)

Returns vector of data generated using copula::String given vector of uniformly distributed u1 and copula parameter θ.

DatagenCopulaBased.sampleInvlaJ — Method.

sampleInvlaJ(α::Float64, v::Float64)

Returns Int, a sample of inverce laplacea transform of generator of Joe nested copula, given a parameter α and a random numver v ∈ [0,1]

DatagenCopulaBased.testbivθ — Method.

testbivθ(θ::Union{Float64}, copula::String)

Tests bivariate copula parameter

clayton bivariate sub-copulas with parameters (θᵢ ≥ -1) ^ ∧ (θᵢ ≠ 0). amh – Ali-Mikhail-Haq bivariate sub-copulas with parameters -1 ≥ θᵢ ≥ 1 frank bivariate sub-copulas with parameters (θᵢ ≠ 0)

DatagenCopulaBased.testind — Method.

testind(inds::Vector{Pair{String,Vector{Int64}}})

Tests if the sub copula name is supported and if their indices are disjoint.

DatagenCopulaBased.testnestedθϕ — Method.

testnestedθϕ(n::Vector{Int}, ϕ::Vector{Float64}, θ::Float64, copula::String)

Tests parameters, its hierarchy and size of parametes vector for nested archimedean copulas.

DatagenCopulaBased.testθ — Method.

function testθ(θ::Union{Float64, Int}, copula::String)

Tests the parameter θ value for archimedean copula, returns void

DatagenCopulaBased.tiltedlevygen — Method.

tiltedlevygen(V0::Vector{Float64}, α::Float64)

Returns a Vector{Float} genrated from the expotencialy tilted levy stable pdf f(x; V0, α) = exp(-V0^α) g(x; α)/exp(-V0), where g(x; α) is a stable Levy pdf with parameters α = 1/θ, β = 1, γ = (cos(pi/(2*θ)))^θ and δ = 0.

DatagenCopulaBased.usebivρ — Method.

Returns Float64, a copula parameter given a pearson or kendall correlation

For clayton or frank copula correlation fulfulling (-1 > ρᵢ > 1) ∧ (ρᵢ ≠ 0) For amh copula pearson correlation fulfilling -0.2816 > ρᵢ >= .5. while kendall -0.18 < τ < 1/3

DatagenCopulaBased.useρ — Method.

useρ(ρ::Float64, copula::String)

Tests the available pearson correlation for archimedean copula, returns Float, corresponding copula parameter θ.

julia> useρ(0.75, "gumbel")
2.294053859606698

DatagenCopulaBased.useτ — Method.

useτ(ρ::Float64, copula::String)

Tests the available kendall's correlation for archimedean copula, returns Float, corresponding copula parameter θ.

julia> useτ(0.5, "clayton")
2.0

DatagenCopulaBased.Σ_theor — Method.

Σ_theor(ϕ::Vector{Float64}, θ::Float64, part::Vector{Vector{Int}})

returns a matrix indicating a theoretical correlation according togiven parameters and partition

DatagenCopulaBased.ρ2θ — Method.

ρ2θ(ρ::Union{Float64, Int}, copula::String)

Returns a Float, an archimedean copula parameter given expected Spermann correlation ρ and a copula.

DatagenCopulaBased.τ2θ — Method.

τ2θ(τ::Float64, copula::String)

Returns Float, a single parameter of Archimedean copula, given the Kenalss τ correlation

DatagenCopulaBased.τ2λ — Method.

τ2λ(τ::Vector{Float64}, λ::Vector{Float64})

Suplement the vector of λ patrameters of Marshall-Olkin copula, given some of those parameters and a vector fo Kendall's τ correlations. Wroks fo 2,3 variate MO copula