Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DelayDiffEq.Discontinuity — Type.

Discontinuity(t, order::Int)

Object of discontinuity of order order at time t, i.e. discontinuity of orderth derivative at time t.

DelayDiffEq.MethodOfSteps — Method.

MethodOfSteps(alg; constrained::Bool=false, fixedpoint_abstol=nothing,
              fixedpoint_reltol=nothing, fixedpoint_norm=nothing,
              max_fixedpoint_iters::Int=10)

Construct an algorithm that solves delay differential equations by the method of steps, where alg is an ODE algorithm from OrdinaryDiffEq.jl without lazy interpolation upon which the calculation of steps is based.

If the algorithm is constrained only steps of size at most the minimal delay will be taken. If it is unconstrained, fixed-point iteration is applied for step sizes that exceed the minimal delay.

The absolute and relative tolerance of the fixed-point iterations can be set by fixedpoint_abstol and fixedpoint_reltol, respectively, either as scalars or vectors. Based on these tolerances error estimates are calculated during the fixed-point iterations with a norm that may be specified as fixedpoint_norm. Fixed-point iterations are stopped if the error estimate is less than 1 or after the maximal number max_fixedpoint_iters of iteration steps.

DiffEqBase.auto_dt_reset! — Method.

auto_dt_reset!(dde_int::DDEIntegrator)

Automatically determine initial time step of dde_int.

DiffEqBase.get_proposed_dt — Method.

get_proposed_dt(integrator::DDEIntegrator)

Get the time step that integrator will take after the current step.

DiffEqBase.initialize! — Method.

initialize!(integrator::DDEIntegrator)

Set initial values of integrator.

DiffEqBase.reeval_internals_due_to_modification! — Method.

reeval_internals_due_to_modification!(integrator::DDEIntegrator)

Recalculate interpolation data and update ODE integrator after changes by callbacks.

DiffEqBase.reinit! — Function.

reinit!(integrator::DDEIntegrator[, u0 = integrator.sol.prob.u0;
        t0 = integrator.sol.prob.tspan[1],
        tf = integrator.sol.prob.tspan[2],
        erase_sol = true,
        kwargs...])

Reinitialize integrator with (optionally) different initial state u0, different integration interval from t0 to tf, and erased solution if erase_sol = true.

DiffEqBase.set_proposed_dt! — Method.

set_proposed_dt!(integrator::DDEIntegrator, dt)

Set the time step that integrator will take after the current step to dt.

DiffEqBase.terminate! — Method.

terminate!(integrator::DDEIntegrator)

Stop further calculations of integrator.

DiffEqBase.u_modified! — Method.

u_modified!(integrator::DDEIntegrator, bool::Bool)

Signal integrator whether state vector u was modified by a callback.

A modified u will lead to recalculations in order to prevent discontinuities.

DelayDiffEq.DiscontinuityCallback — Method.

DiscontinuityCallback(lags, discontinuities::Vector{<:Discontinuity};
                      [interp_points::Int=10, abstol=1e-12, reltol=0])

Callback that tracks discontinuities that are propagated by dependent lags of the form (u,p,t) -> lag(u,p,t).

Hereby a number interp_points of interpolation points are used to first check for different signs of functions $f(t) = T + lag(u(t),p,t) - t$, where $T$ is time point of a previous discontinuity and $t$ is contained in the current time interval. This shows that the current time interval contains propagated discontinuities of which the exact time point is then determined by a root finding algorithm. The sign at the lower bound of the time interval, i.e. at tprev, is set to 0 with absolute tolerance abstol and relative tolerance reltol.

DelayDiffEq.HistoryFunction — Type.

HistoryFunction(h, sol, integrator)

Wrap history function h, solution sol, and integrator integrator to create a common interface for retrieving values at any time point with varying accuracy.

Before the initial time point of solution sol values are calculated by history function h, for time points in the time span of sol interpolated values of sol are returned, and after the final time point of sol an inter- or extrapolation of the current state of integrator integrator is retrieved.

DelayDiffEq.add_next_discontinuities! — Function.

add_next_discontinuities!(integrator::DDEIntegrator, order[, t=integrator.t])

Add discontinuities of next order that are propagated from discontinuity of order order at time t in integrator, but only if order is less or equal than the order of the applied method or the problem is neutral.

Discontinuities caused by constant delays are immediately calculated, and discontinuities caused by dependent delays are tracked by a callback.

DelayDiffEq.advance_ode_integrator! — Method.

advance_ode_integrator!(integrator::DDEIntegrator)

Advance ODE integrator of integrator to next time interval, values and complete interpolation data of integrator.

DelayDiffEq.agrees — Method.

agrees(h, u, p, t)

Determine whether history function evaluates to u at time point t for parameters p.

DelayDiffEq.assign_expr — Method.

assign_expr(::Val{name}, ::Type{T}, ::Type{cache})

Create expression that extracts field name of type T from cache of type cache to variable name.

Hereby u, uprev, uprev2, and function f are updated, if required.

DelayDiffEq.build_linked_cache — Method.

build_linked_cache(cache, alg, u, uprev, uprev2, f, t, dt)

Create cache for algorithm alg from existing cache cache with updated u, uprev, uprev2, f, t, and dt.

DelayDiffEq.build_solution_array — Method.

build_solution_array(integrator::DDEIntegrator)

Create a DiffEqArray of the time points and values that form the solution of integrator.

DelayDiffEq.build_solution_interpolation — Method.

build_solution_interpolation(integrator::DDEIntegrator, sol::DiffEqArray)

Create interpolation data to solution of integrator, which is formed by time points and values in sol.

DelayDiffEq.constant_extrapolant! — Method.

constant_extrapolant!(val, t, integrator::DEIntegrator, idxs, deriv)

Calculate constant extrapolation of derivative deriv at time t and indices idxs for integrator, and save result in val if val is not nothing.

DelayDiffEq.constant_extrapolant — Method.

constant_extrapolant(t, integrator::DEIntegrator, idxs, deriv)

Calculate constant extrapolation of derivative deriv at time t and indices idxs for integrator.

DelayDiffEq.determine_discontinuity_existence — Method.

determine_discontinuity_existence(prev_sign::Int, next_sign::Int, Θs, f)

Determine whether function f has a root in the interval [0, 1] by checking signs of f at 0 and 1 (prev_sign and next_sign, respectively) and at interpolation points in Θs.

This corresponds to the existence of a propagated discontinuity.

DelayDiffEq.find_discontinuity_time — Method.

find_discontinuity_time(integrator::DDEIntegrator, callback::DiscontinuityCallback,
                        prev_sign::Int, next_sign::Int, Θs, f)

Find time of propagated discontinuity for a certain dependent delay and previous discontinuity, which is root of the function f, in the current time interval of integrator with interpolation points Θs. Hereby f shows signs prev_sign and next_sign at both ends of the time interval.

DelayDiffEq.fsal_typeof — Method.

fsal_typeof(integrator::ODEIntegrator)

Return type of FSAL of integrator.

DelayDiffEq.reduce_solution! — Method.

reduce_solution!(integrator::DDEIntegrator, tmax)

Remove time points of ODE solution of integrator up to time tmax that are not required for calculation of DDE solution.

DiffEqBase.apply_callback! — Method.

apply_callback!(integrator::DDEIntegrator, callback::DiscontinuityCallback, cb_time,
                order)

Handle discontinuity of order order at time integrator.tprev + cb_time in the current time interval of integrator that was found by callback. Cause the current step to fail, and add the found discontinuity to both the heap of discontinuities and of time stops.

DiffEqBase.postamble! — Method.

postamble!(integrator::DDEIntegrator)

Clean up solution of integrator.

OrdinaryDiffEq.handle_discontinuities! — Method.

handle_discontinuities!(integrator::DDEIntegrator)

Handle discontinuities at the current time point of integrator.

OrdinaryDiffEq.perform_step! — Method.

perform_step!(integrator::DDEIntegrator)

Calculate next step of integrator.