Library

Dendriform.jl Library

Dendriform.jl Library

Index

Dendriform Types

Dendriform Operators

Dendriform Morphisms

Tools & Options

Dendriform.jl

Setup

Background

Usage

References

Index

Dendriform.BaseTree

Dendriform.Grove

Dendriform.GroveBin

Dendriform.PBTree

Base.:*

Base.:+

Base.:/

Base.:<

Base.:>

Base.:\

Base.:∪

Base.:≤

Base.:≥

Dendriform.:∨

Dendriform.:⊢

Dendriform.:⊣

Dendriform.:⊴

Dendriform.:⋖

Dendriform.:⋗

Dendriform.GroveError

Dendriform.LeftInherited

Dendriform.PrimitiveTree

Dendriform.RightInherited

Dendriform.TreeBase

Dendriform.TreeInteger

Dendriform.TreeRational

Dendriform.between

Dendriform.dashv

Dendriform.graft

Dendriform.grovebit

Dendriform.grovecheck

Dendriform.grovedisplay

Dendriform.groveindex

Dendriform.grovesort

Dendriform.left

Dendriform.over

Dendriform.posetnext

Dendriform.posetprev

Dendriform.right

Dendriform.treecheck

Dendriform.treeindex

Dendriform.treeshift

Dendriform.under

Dendriform.vdash

Dendriform.σ

Dendriform Types

Dendriform.PBTree — Type.

Planar Binary Tree with Loday's notation

Summary

mutable struct PBTree <: AbstractGrove

Fields

degr::UInt8 Y::Array{UInt8,1}

Dendriform.Grove — Type.

Grove of planar binary trees, matrix equivalence class

Summary

mutable struct Grove <: AbstractGrove

Fields

degr::UInt8 size::Int Y::Array{UInt8,2}

Dendriform.GroveBin — Type.

Compressed binary representation of grove

Summary

mutable struct GroveBin <: AbstractGrove

Fields

degr::UInt8 size::Int gbin::Integer ppos::Float16

Dendriform.BaseTree — Type.

Descending greatest integer search data for grove

Summary

mutable struct BaseTree <: Abstract Grove

Fields

μ::Array{Array{UInt8,1},1}

Dendriform Operators

Base.:∪ — Function.

∪(g...)

Returns union of AbstractGrove objects

Branching

Dendriform.graft — Function.

graft(left::AbstractPBTree, right::AbstractPBTree)

Grafts the left and right PBTree with root vertex

Dendriform.:∨ — Function.

∨(left::AbstractPBTree, right::AbstractPBTree)

Grafts the left and right AbstractPBTree objects

Dendriform.left — Function.

left(::AbstractPBTree)

Returns the left branch of an AbstractPBTree

Dendriform.right — Function.

right(::AbstractPBTree)

Returns the right branch of an AbstractPBTree

Partial Ordering

Dendriform.posetnext — Function.

posetnext(::AbstractPBTree)

Returns a Grove that covers the given tree

Dendriform.posetprev — Function.

posetprev(::AbstractPBTree)

Returns a Grove that covers the given tree

Dendriform.:⋖ — Function.

⋖(a::AbstractPBTree, b::AbstractPBTree)

Returns Bool that tells if b covers a in Tamari partial order

Dendriform.:⋗ — Function.

⋗(a::AbstractPBTree, b::AbstractPBTree)

Returns Bool that tells if a covers b in Tamari partial order

Base.:< — Function.

<(a::AbstractPBTree, b::AbstractPBTree)

Returns Bool that tells if a < b in Tamari partial order

Base.:> — Function.

>(a::AbstractPBTree, b::AbstractPBTree)

Returns Bool that tells if a > b in Tamari partial order

Base.:≤ — Function.

≤(a::AbstractPBTree, b::AbstractPBTree)

Returns Bool that tells if a ≤ b in Tamari partial order

Base.:≥ — Function.

≥(a::AbstractPBTree, b::AbstractPBTree)

Returns Bool that tells if a ≥ b in Tamari partial order

Dendriform.between — Function.

between(a::AbstractPBTree,b::AbstractPBTree)

Returns Grove of trees ordered between a and b

Dendriform.:⊴ — Function.

⊴(a::AbstractPBTree,b::AbstractPBTree)

Returns Grove of trees ordered between a and b

Dialgebra Arithmetic

Dendriform.over — Function.

over(a::AbstractPBTree, b::AbstractPBTree)

Returns PBTree obtained from a over b operation

Dendriform.under — Function.

under(a::AbstractPBTree, b::AbstractPBTree)

Returns PBTree obtained from a under b operation

Base.:/ — Function.

/(a::PBTree, b::PBTree)

Returns PBTree obtained from a over b operation

Base.:\ — Function.

\(a::PBTree, b::PBTree)

Returns PBTree obtained from a under b operation

Dendriform.dashv — Function.

dashv(a::AbstractGrove, b::AbstractGrove)

Returns Grove obtained from a ⊣ b operation

Dendriform.vdash — Function.

vdash(a::AbstractGrove, b::AbstractGrove)

Returns Grove obtained from a ⊢ b operation

Dendriform.:⊣ — Function.

⊣(a::AbstractGrove, b::AbstractGrove)

Returns Grove obtained from a ⊣ b operation

Dendriform.:⊢ — Function.

⊢(a::AbstractGrove, b::AbstractGrove)

Returns Grove obtained from a ⊢ b operation

Base.:+ — Function.

+(a::AbstractGrove, b::AbstractGrove)

Returns Grove obtained from a + b operation

Base.:* — Function.

*(a::AbstractGrove, b::AbstractGrove)

Returns Grove obtained from a × b operation

Dendriform Morphisms

Dendriform.σ — Function.

σ(g::AbstractGrove)

Applies the involution to any PBTree or Grove object

Consistency Checks

Dendriform.LeftInherited — Function.

Dendriform.LeftInherited(::AbstractPBTree)

Returns Bool that tells if PBTree is left inherited

Dendriform.RightInherited — Function.

Dendriform.RightInherited(::AbstractPBTree)

Returns Bool that tells if PBTree is right inherited

Dendriform.PrimitiveTree — Function.

Dendriform.PrimitiveTree(::AbstractPBTree)

Returns Bool that tells if PBTree is primitive

Dendriform.treecheck — Function.

treecheck(::AbstractPBTree)

Returns Bool that tells if PBTree is valid

Dendriform.grovecheck — Function.

grovecheck(::AbstractGrove)

Returns Bool that tells if Grove is valid

Dendriform.GroveError — Function.

Dendriform.GroveError(::AbstractGrove)

Returns Array with Grove sorting index error

Index Data

Dendriform.treeindex — Function.

treeindex(::AbstractGrove)

Returns tree indices of any PBTree or Grove

Dendriform.groveindex — Function.

groveindex(::AbstractGrove)

Returns the grove index of any Grove

Dendriform.grovebit — Function.

grovebit(::AbstractGrove)

Returns a BitArray of tree indices

Transformations

Dendriform.TreeBase — Function.

Dendriform.TreeBase(::AbstractGrove)

Returns BaseTree objects for any AbstractGrove

Dendriform.TreeInteger — Function.

Dendriform.TreeInteger(::AbstractGrove)

Returns the tree integers of any AbstractGrove

Dendriform.TreeRational — Function.

Dendriform.TreeRational(::AbstractGrove)

Returns the tree rationals of any AbstractGrove

Tools & Options

Dendriform.grovedisplay — Function.

grovedisplay(::Bool)

Toggles the display output of grove index data (disabled by default)

Dendriform.grovesort — Function.

grovesort(::Bool)

Toggles the grovesort algorithm (enabled by default - RECOMMENDED)

Dendriform.treeshift — Function.

treeshift(::Bool)

Toggles the shift for the tree integers / rationals