DependenciesParser.jl

DependenciesParser.jl

DependenciesParser.jl is a package for quickly obtaining the list of dependencies for packages.

DependenciesParser.alldeps — Function.

alldeps(julia::VersionNumber = VersionNumber(1))::Vector{NameRepoVerDeps}

Returns information for all installable packages for the selected julia version

DependenciesParser.dependencies — Function.

dependencies(name::AbstractString,
             direct::Bool = false,
             data::VersionNumber = DependenciesParser.data)::Vector{String}

Returns all dependencies for the package (excludes stdlib)
When direct is true, only direct dependencies are returned
Dependencies.data = alldeps()

DependenciesParser.NameRepoVerDeps — Type.

NameRepoVerDeps

Alias for NamedTuple struct used for name, repository, version, and dependencies.