Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Diana.Parse — Method.

Given a GraphQL source, parses it into a Document. Throws GraphQLError if a syntax error is encountered.

Diana.an_y — Method.

Returns a possibly empty list of parse nodes, determined by the parseFn. This list begins with a lex token of openKind and ends with a lex token of closeKind. Advances the parser to the next lex token after the closing token.

Diana.expect — Method.

If the next token is of the given kind, return that token after advancing the lexer. Otherwise, do not change the parser state and throw an error.

Diana.expectKeyword — Method.

If the next token is a keyword with the given value, return that token after advancing the lexer. Otherwise, do not change the parser state and return false.

Diana.many — Method.

Returns a non-empty list of parse nodes, determined by the parseFn. This list begins with a lex token of openKind and ends with a lex token of closeKind. Advances the parser to the next lex token after the closing token.

Diana.next_token — Method.

If the next token is of the given kind, return true after advancing the lexer. Otherwise, do not change the parser state and return false.

Diana.parseArgument — Method.

Argument : Name : Value

Diana.parseArgumentDefs — Method.

ArgumentsDefinition : ( InputValueDefinition+ )

Diana.parseArguments — Method.

Arguments : ( Argument+ )

Diana.parseDefinition — Method.

Definition : - OperationDefinition - FragmentDefinition - TypeSystemDefinition

Diana.parseDirective — Method.

Directive : @ Name Arguments?

Diana.parseDirectiveDefinition — Method.

DirectiveDefinition : - directive @ Name ArgumentsDefinition? on DirectiveLocations

Diana.parseDirectiveLocations — Method.

DirectiveLocations : - |? Name - DirectiveLocations | Name

Diana.parseDirectives — Method.

Implements the parsing rules in the Directives section.

Directives : Directive+

Diana.parseDocument — Method.

#if length(errors)>0
#	println("Syntax Error GraphQL request (x:x) Unexpected character")
#	errors
	# {"errors":[{"locations":[{"column":5,"line":3}],"message":"Syntax Error GraphQL request (3:5) Unexpected character "".

2: neomatrix{

3:     nombre
   ^

4: linkedin "}]} Document : Definition+

Diana.parseEnumTypeDefinition — Method.

EnumTypeDefinition : enum Name Directives? { EnumValueDefinition+ }

Diana.parseEnumValueDefinition — Method.

EnumValueDefinition : EnumValue Directives?

EnumValue : Name

Diana.parseField — Method.

Field : Alias? Name Arguments? Directives? SelectionSet?

Alias : Name : NAME

Diana.parseFieldDefinition — Method.

FieldDefinition : Name ArgumentsDefinition? : Type Directives?

Diana.parseFragment — Method.

Implements the parsing rules in the Fragments section.

Corresponds to both FragmentSpread and InlineFragment in the spec.

FragmentSpread : ... FragmentName Directives?

InlineFragment : ... TypeCondition? Directives? SelectionSet

Diana.parseFragmentDefinition — Method.

FragmentDefinition : - fragment FragmentName on TypeCondition Directives? SelectionSet

TypeCondition : NamedType

Diana.parseFragmentName — Method.

FragmentName : Name but not on

Diana.parseImplementsInterfaces — Method.

ImplementsInterfaces : implements NamedType+

Diana.parseInputObjectTypeDefinition — Method.

InputObjectTypeDefinition : input Name Directives? { InputValueDefinition+ }

Diana.parseInputValueDef — Method.

InputValueDefinition : Name : Type DefaultValue? Directives?

Diana.parseInterfaceTypeDefinition — Method.

InterfaceTypeDefinition : interface Name Directives? { FieldDefinition+ }

Diana.parseList — Method.

ListValue[Const] : - [ ] - [ Value[?Const]+ ]

Diana.parseName — Method.

Converts a name lex token into a name parse node.

Diana.parseNamedType — Method.

NamedType : Name

Diana.parseObject — Method.

ObjectValue[Const] : - { } - { ObjectField[?Const]+ }

Diana.parseObjectField — Method.

ObjectField[Const] : Name : Value[?Const]

Diana.parseObjectTypeDefinition — Method.

ObjectTypeDefinition : - type Name ImplementsInterfaces? Directives? { FieldDefinition+ }

Diana.parseOperationDefinition — Method.

Implements the parsing rules in the Operations section.

OperationDefinition :

SelectionSet

OperationType Name? VariableDefinitions? Directives? SelectionSet

Diana.parseOperationType — Method.

Implements the parsing rules in the Document section.

OperationType : one of query mutation subscription

Diana.parseScalarTypeDefinition — Method.

ScalarTypeDefinition : scalar Name Directives?

Diana.parseSchemaDefinition — Method.

SchemaDefinition : schema Directives? { OperationTypeDefinition+ }

OperationTypeDefinition : OperationType : NamedType

Diana.parseSelection — Method.

Selection : - Field - FragmentSpread - InlineFragment

Diana.parseSelectionSet — Method.

SelectionSet : { Selection+ }

Diana.parseType — Method.

Given a string containing a GraphQL Type (ex. [Int!]), parse the AST for that type. Throws GraphQLError if a syntax error is encountered.

This is useful within tools that operate upon GraphQL Types directly and in isolation of complete GraphQL documents.

Consider providing the results to the utility function: typeFromAST().

Diana.parseTypeExtensionDefinition — Method.

TypeExtensionDefinition : extend ObjectTypeDefinition

Diana.parseTypeReference — Method.

Implements the parsing rules in the Types section.

Type : - NamedType - ListType - NonNullType

Diana.parseTypeSystemDefinition — Method.

Implements the parsing rules in the Type Definition section.

TypeSystemDefinition : - SchemaDefinition - TypeDefinition - TypeExtensionDefinition - DirectiveDefinition

TypeDefinition : - ScalarTypeDefinition - ObjectTypeDefinition - InterfaceTypeDefinition - UnionTypeDefinition - EnumTypeDefinition - InputObjectTypeDefinition

Diana.parseUnionMembers — Method.

UnionMembers :
- `|`? NamedType
- UnionMembers | NamedType

Diana.parseUnionTypeDefinition — Method.

UnionTypeDefinition : union Name Directives? = UnionMembers

Diana.parseValue — Method.

Given a string containing a GraphQL value (ex. [42]), parse the AST for that value. Throws GraphQLError if a syntax error is encountered.

This is useful within tools that operate upon GraphQL Values directly and in isolation of complete GraphQL documents.

Consider providing the results to the utility function: valueFromAST().

Diana.parseValueLiteral — Method.

Implements the parsing rules in the Values section.

Value[Const] : - [~Const] Variable - IntValue - FloatValue - StringValue - BooleanValue - NullValue - EnumValue - ListValue[?Const] - ObjectValue[?Const]

BooleanValue : one of true false

NullValue : null

EnumValue : Name but not true, false or null

Diana.parseVariable — Method.

Variable : $ Name

Diana.parseVariableDefinition — Method.

VariableDefinition : Variable : Type DefaultValue?

Diana.parseVariableDefinitions — Method.

VariableDefinitions : ( VariableDefinition+ )

Diana.peek — Method.

Determines if the next token is of a given kind

Diana.unexpected — Method.

Helper function for creating an error when an unexpected lexed token is encountered.