Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DiffEqBase.AffineDiffEqOperator — Type.

AffineDiffEqOperator{T} <: AbstractDiffEqOperator{T}

Ex: (A₁(t) + ... + Aₙ(t))*u + B₁(t) + ... + Bₙ(t)

AffineDiffEqOperator{T}(As,Bs,u_cache=nothing)

Takes in two tuples for split Affine DiffEqs

update_coefficients! works by updating the coefficients of the component operators.

Function calls L(u,p,t) and L(u,p,t,du) are fallbacks interpretted in this form. This will allow them to work directly in the nonlinear ODE solvers without modification.

f(u,p,t,du) is only allowed if a u_cache is given

B(t) can be Union{Number,AbstractArray}, in which case they are constants. Otherwise they are interpreted they are functions v=B(t) and B(v,t)

Solvers will see this operator from integrator.f and can interpret it by checking the internals of As and Bs. For example, it can check is_constant(As[1]) etc.

DiffEqBase.check_error — Method.

check_error(integrator)

Check state of integrator and return one of the Return Codes

DiffEqBase.reeval_internals_due_to_modification! — Method.

reeval_internals_due_to_modification!(integrator::DDEIntegrator)

Recalculate interpolation data and update ODE integrator after changes by callbacks.

DiffEqBase.remake — Method.

remake(thing; <keyword arguments>)

Re-construct thing with new field values specified by the keyword arguments.

DiffEqBase.set_t! — Method.

set_t!(integrator::DEIntegrator, t::Real)

Set current time point of the integrator to t.

DiffEqBase.set_u! — Method.

set_u!(integrator::DEIntegrator, u)

Set current state of the integrator to u.

DiffEqBase.step! — Method.

step!(integ::DEIntegrator [, dt [, stop_at_tdt]])

Perform one (successful) step on the integrator.

Alternative, if a dt is given, then step! the integrator until there is a temporal difference ≥ dt in integ.t. When true is passed to the optional third argument, the integrator advances exactly dt.

DiffEqBase.u_modified! — Method.

savevalues!(integrator::DEIntegrator,
  force_save=false) -> Tuple{Bool, Bool}

Try to save the state and time variables at the current time point, or the saveat point by using interpolation when appropriate. It returns a tuple that is (saved, savedexactly). If savevalues! saved value, then saved is true, and if savevalues! saved at the current time point, then savedexactly is true.

The saving priority/order is as follows:

save_on

saveat

force_save

save_everystep/timeseries_steps

DiffEqBase.ExplicitRKTableau — Type.

ExplicitRKTableau

Holds a tableau which defines an explicit Runge-Kutta method.

DiffEqBase.ImplicitRKTableau — Type.

ImplicitRKTableau

Holds a tableau which defines an implicit Runge-Kutta method.

DiffEqBase.check_error! — Method.

check_error!(integrator)

Same as check_error but also set solution's return code (integrator.sol.retcode) and run postamble!.

DiffEqBase.copy_fields — Method.

copy_fields(arr:AbstractArray, template::DEDataArray)

Create DEDataArray that wraps arr with all other fields set to a deep copy of the value in template.

copy_fields!(dest::T, src::T2) where {T<:DEDataArray,T2<:DEDataArray}

Replace all fields of dest except of its wrapped array with a copy of the value in src. Arrays are recursively copied.

DiffEqBase.interpolant! — Method.

Constant Interpolation

DiffEqBase.interpolant! — Method.

Constant Interpolation

DiffEqBase.interpolant! — Method.

Hairer Norsett Wanner Solving Ordinary Differential Euations I - Nonstiff Problems Page 190

Hermite Interpolation

DiffEqBase.interpolant! — Method.

Hermite Interpolation

DiffEqBase.interpolant! — Method.

Hermite Interpolation

DiffEqBase.interpolant! — Method.

Hermite Interpolation

DiffEqBase.interpolant! — Method.

Linear Interpolation

DiffEqBase.interpolant! — Method.

Linear Interpolation

DiffEqBase.interpolant — Method.

Constant Interpolation

DiffEqBase.interpolant — Method.

Hairer Norsett Wanner Solving Ordinary Differential Equations I - Nonstiff Problems Page 190

Hermite Interpolation

DiffEqBase.interpolant — Method.

Hermite Interpolation

DiffEqBase.interpolant — Method.

Hermite Interpolation

DiffEqBase.interpolant — Method.

Hermite Interpolation

DiffEqBase.interpolant — Method.

Linear Interpolation

DiffEqBase.interpolation — Function.

interpolation(tval::Number,t,u,ks)

Get the value at tval where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.interpolation! — Function.

interpolation(tvals,t,u,ks)

Get the value at tvals where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.interpolation! — Function.

interpolation!(out,tval::Number,t,u,ks)

Get the value at tval where the solution is known at the times t (sorted), with values u and derivatives ks

DiffEqBase.numargs — Method.

numparameters(f)

Returns the number of parameters of f for the method which has the most parameters.

DiffEqBase.set_ut! — Method.

set_ut!(integrator::DEIntegrator, u, t)

Set current state of the integrator to u and t

DiffEqBase.undefined_exports — Method.

undefined_exports(mod)

List symbols export'ed but not actually defined.

RecursiveArrayTools.recursivecopy! — Method.

recursivecopy!(dest::T, src::T) where {T<:DEDataArray}

Recursively copy fields of src to dest.