Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DiffEqCallbacks.SavedValues — Type.

SavedValues{tType<:Real, savevalType}

A struct used to save values of the time in t::Vector{tType} and additional values in saveval::Vector{savevalType}.

DiffEqCallbacks.SavedValues — Method.

SavedValues(tType::DataType, savevalType::DataType)

Return SavedValues{tType, savevalType} with empty storage vectors.

DiffEqCallbacks.FunctionCallingCallback — Method.

FunctionCallingCallback(func;
                funcat=Vector{Float64}(),
                func_everystep=isempty(funcat),
                func_start = true,
                tdir=1)

A DiscreteCallback applied after every step to call func(u,t,integrator) If func_everystep, every step of the integrator gives a func call. If funcat is specified, the function is called at the given times, using interpolation if necessary. If the time tdir direction is not positive, i.e. tspan[1] > tspan[2], tdir = -1 has to be specified.

DiffEqCallbacks.SavingCallback — Method.

SavingCallback(save_func, saved_values::SavedValues;
                saveat=Vector{eltype(saved_values.t)}(),
                save_everystep=isempty(saveat),
                save_start = true,
                tdir=1)

A DiscreteCallback applied after every step, saving the time t and the value of save_func(u, t, integrator) in saved_values.

If save_everystep, every step of the integrator is saved. If saveat is specified, the values are saved at the given times, using interpolation if necessary. If the time tdir direction is not positive, i.e. tspan[1] > tspan[2], tdir = -1 has to be specified.

DiffEqCallbacks.affect! — Method.

affect!(integrator, f::AbstractDomainAffect)

Apply domain callback f to integrator.

DiffEqCallbacks.isaccepted — Method.

isaccepted(u, abstol, f::AbstractDomainAffect, args...)

Return whether u is an acceptable state vector at the next time point given absolute tolerance abstol, callback f, and other optional arguments.

DiffEqCallbacks.modify_u! — Method.

modify_u!(integrator, f::AbstractDomainAffect)

Modify current state vector u of integrator if required, and return whether it actually was modified.

DiffEqCallbacks.setup — Method.

setup(f::AbstractDomainAffect, integrator)

Setup callback f and return an arbitrary tuple whose elements are used as additional arguments in checking whether time step is accepted.