Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Discretizers.DiscretizeUniformWidth — Type.

DiscretizeUniformWidth(alg::Symbol) If nbins is a symbol, automatically determine the number of bins to use

Discretizers.LinearDiscretizer — Type.

LinearDiscretizer

A discretizer which encodes (typically) continuous values into discrete bins. A univariate domain is divided into discrete by edges. A value V will be encoded into bin B if V ∈ [Bₗ Bᵣ) (or V ∈ [Bₗ Bᵣ] if B is the rightmost bin)

If forceoutliersto_closest is set to true, then values outside of binedges will be shunted into the nearest bin. Otherwise an error is thrown.

TODO(tim): handle NA handle Nullable handle NaN ability to specify decoding behavior

Discretizers.datalineardiscretizer — Method.

a hybrid discretizers that maps a special missing value indicator (typically Inf or NaN) to a discrete bin, but otherwise is a linear discretizer

Discretizers.decode — Method.

There are several methods for decoding a LinearDiscretizer. The default is SampleUniform, which uniformly samples from the bin. SampleBinCenter returns the bin's center value.