Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DocSeeker.alldocs — Method.

alldocs() -> Vector{DocObj}

Find all docstrings in all currently loaded Modules.

DocSeeker.baseURL — Method.

baseURL(links::Vector{Markdown.Link}) -> String

Find the most common host and return the first URL in links with that host.

DocSeeker.createdocsdb — Method.

createdocsdb()

Asynchronously create a "database" of all local docstrings in Pkg.dir(). This is done by loading all packages and using introspection to retrieve the docstrings – the obvious limitation is that only packages that actually load without errors are considered.

DocSeeker.docsdir — Method.

docsdir(pkg) -> String

Find the directory conataining the documenatation for package pkg. Will fall back to returning a package's README.md. Returns an empty String if no indication of documentation is found.

DocSeeker.docsurl — Method.

docsurl(pkg) -> String

Return the most likely candidate for a package's online documentation or an empty string.

DocSeeker.finddocsURL — Method.

finddocsURL(pkg) -> Vector{Markdown.Link}

Search pkgs readme for links to documentation.

DocSeeker.loaddocsdb — Method.

loaddocsdb() -> Vector{DocObj}

Retrieve the docstrings from the "database" created by createdocsdb(). Will return an empty vector if the database is locked by createdocsdb().

DocSeeker.score — Function.

score(needle, s::Docs.DocObj) -> Float

Scores s against the search query needle. Returns a Float between 0 and 1.