Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DocumenterTools.generate — Method.

DocumenterTools.generate(path::String; name = nothing, format = :html)

Create a documentation stub in path, which is usually a sub folder in the package root. The name of the package is determined automatically, but can be given with the name keyword argument.

generate creates the following files in path:

.gitignore
src/index.md
make.jl
mkdocs.yml
Project.toml

Arguments

path file path to the documentation directory.

Keywords Arguments

name is the name of the package (without .jl). If name is not given generate tries to detect it automatically.

format can be either :html (default), :markdown or :pdf corresponding to the format keyword to Documenter's makedocs function, see Documenter's manual.

Examples

julia> using DocumenterTools

julia> Documenter.generate("path/to/MyPackage/docs")
[ ... output ... ]

DocumenterTools.generate — Method.

DocumenterTools.generate(pkg::Module; dir = "docs", format = :html)

Same as generate(path::String) but the path and name is determined automatically from the module.

Note

The package must be in development mode. Make sure you run pkg> develop pkg from the Pkg REPL, or Pkg.develop("pkg") before generating docs.

Examples

julia> using DocumenterTools

julia> using MyPackage

julia> DocumenterTools.generate(MyPackage)
[ ... output ... ]

DocumenterTools.package_devpath — Method.

package_devpath(pkg)

Returns the path to the top level directory of a devved out package source tree. The package is identified by its top level module pkg.