Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DomainSets.AbstractMap — Type.

A map is any transformation of the form y = f(x), where x has type S and y has type T.

DomainSets.AffineMap — Type.

AffineMap represents y = a*x + b, i.e. it combines a LinearMap and a Translation.

DomainSets.CartToPolarMap — Type.

A Cartesion to Polar map. First dimension is interpreted as radial distance, second as an angle. The unit circle is mapped to a square [-1,1]x[-1,1]

DomainSets.ChebyshevInterval — Type.

The closed interval [-1,1].

DomainSets.CompositeMap — Type.

The composition of several maps.

DomainSets.ConstantMap — Type.

The constant map f(x) = c.

DomainSets.DerivedDomain — Type.

DerivedDomain is an abstract supertype for domains that are implemented in terms of another domain using composition. DerivedDomain transfers the interface of Domain to this superdomain.

A concrete subtype that inherits from DerivedDomain and stores a superdomain is functionally equivalent to that superdomain. Any properties of the superdomain can be modified by overriding a suitable function. For example, in of the concrete domain may be implemented in terms of the in of the superdomain.

DomainSets.EuclideanDomain — Type.

A EuclideanDomain is any domain whose eltype is SVector{N,T}.

DomainSets.HalfLine — Type.

The half-open positive halfline [0,∞).

DomainSets.IdentityMap — Type.

The identity map.

DomainSets.IntersectionDomain — Type.

An IntersectionDomain represents the intersection of a set of domains.

DomainSets.LinearMap — Type.

A LinearMap is an affine map that represents y = a*x, where a can have any type such that a*x maps type S to type T.

DomainSets.NegativeHalfLine — Type.

The open negative halfline (-∞,0).

DomainSets.Point — Type.

Point(x)

represents a single point at x.

DomainSets.PolarToCartMap — Type.

A Polar to Cartesian map. The angle is mapped to the second dimension, radius to the first. A square [-1,1]x[-1,1] is mapped to the unit circle

DomainSets.ProductDomain — Type.

A ProductDomain represents the cartesian product of other domains.

A product domain has two eltypes, an internal type S and an external type T. The internal type S is a tuple containing the eltypes of the elements of the product domain. The external eltype T is a type whose associated space is isomorphic to that of S, but which has been simplified. (See also simplify_product_eltype).

For example, if S is Tuple{Float64,Float64}, then T is SVector{2,Float64}.

DomainSets.ProductMap — Type.

A product map is diagonal and acts on each of the components of x separately: y = f(x) becomes y_i = f_i(x_i).

DomainSets.Translation — Type.

Translation represents y = x + v, where v is a vector in the same space as x and y.

DomainSets.UnionDomain — Type.

A UnionDomain represents the union of a set of domains.

DomainSets.UnitHyperBall — Type.

The unit ball (of radius 1) in N dimensions.

DomainSets.UnitHyperSphere — Type.

The unit sphere (of radius 1) in N dimensions.

DomainSets.UnitInterval — Type.

The closed unit interval [0,1].

DomainSets.ℂ — Type.

The complex plane with Float64 real and imaginar parts (ℂ = \BbbC).

DomainSets.ℚ — Type.

The set of rational numbers of type Rational{Int} (ℚ = \BbbQ).

DomainSets.ℝ — Type.

The set of reals of type Float64 (ℝ = \BbbR).

DomainSets.ℝ1 — Type.

The space ℝ^1

DomainSets.ℝ2 — Type.

The space ℝ^2

DomainSets.ℝ3 — Type.

The space ℝ^3

DomainSets.ℝ4 — Type.

The space ℝ^4

DomainSets.ℤ — Type.

The set of integers of type Int (ℤ = \BbbZ).

DomainSets.:↪ — Function.

The symbol ↪ (\hookrightarrow) is a synonym for embedded.

DomainSets.:≅ — Function.

The symbol ≅ (\cong) is a synonym for isomorphic.

DomainSets.apply_inverse — Method.

Apply the inverse of the given map. By default, this constructs the inverse of m. However, some maps know its own inverse and may provide a more efficient implementation of apply_inverse.

DomainSets.cartesianproduct — Method.

Create a cartesian product of the supplied arguments.

The function cartesianproduct applies some simplifications and does not necessarily return a Product type.

A cartesianproduct(a) with just a single element returns a.

For integer n, cartesianproduct(a, n) becomes cartesianproduct(a, a, ..., a). A type-safe variant is cartesianproduct(a, Val{N}).

DomainSets.ellipse — Method.

Create an ellipse curve with semi-axes lengths a and b respectively.

DomainSets.ellipse_shape — Method.

Create an ellipse-shaped domain with semi-axes lengths a and b respectively.

DomainSets.embedded — Method.

Returns true if A is embedded in B.

DomainSets.interval_map — Method.

Map the interval [a,b] to the interval [c,d].

DomainSets.isomorphic — Method.

Returns true if A is isomorphic to B, and false otherwise.

DomainSets.issubspace — Method.

A space A is a subspace of space B if B is a supertype of A.

DomainSets.left_inverse — Method.

Return a left inverse of the given map. This left inverse mli is not unique, but in any case it is such that (mli ∘ m) * x = x for each x in the domain of m.

DomainSets.linear_map — Method.

Make the linear map y = a*x + b.

DomainSets.origin — Method.

The origin of a space is its zero element.

DomainSets.right_inverse — Method.

Return a right inverse of the given map. This right inverse mri is not unique, but in any case it is such that (m ∘ mri) * y = y for each y in the range of m.

DomainSets.scaling_map — Method.

Scale the variables by a, b, c and d.

DomainSets.scaling_map — Method.

Scale the variables by a, b and c.

DomainSets.scaling_map — Method.

Scale the variables by a and b.

DomainSets.scaling_map — Method.

Scale all variables by a.

DomainSets.spaceof — Method.

Return the geometric space of all elements with the same type as x.

DomainSets.spacetype — Method.

Return the geometric space type with eltype T.

DomainSets.subeltype — Method.

If the type T is a container type, the elements of T may have a different subeltype. If T is not a container, subeltype simply evaluates to T.

DomainSets.subeltype — Method.

Some geometric spaces have an eltype that is composed in terms of an underlying eltype. For example, Complex{Float64} and SVector{2,Float64} are based on Float64, which is their subeltype.

DomainSets.superspace — Method.

Return the superspace of the given geometric space A. The superspace of A should be larger than A, and A should be embedded in it.

Superspaces are used to automatically discover embeddings and promotion rules. Its role is analogous to that of supertype in the Julia type system. Indeed, AnySpace is a superspace of all spaces, much like Any is a supertype of all types.

DomainSets.AbstractAffineMap — Type.

An affine map has the general form y = a*x + b, with types for a, b, x and y such that the expression is valid.

We use matrix and vector to denote a and b respectively.

DomainSets.AbstractConstantMap — Type.

The supertype of several maps that map to a constant value.

DomainSets.AngleMap — Type.

AngleMap is a left inverse of UnitCircleMap. A 2D vector x is projected onto the intersection point with the unit circle of the line connecting x to the origin. The angle of this point, scaled to the interval [0,1), is the result.

DomainSets.AnySpace — Type.

AnySpace is the superset of all possible geometric spaces.

DomainSets.ArraySpace — Type.

A general Array space, with arrays of dimension N and element type T. Note that the arrays can have any size. Thus, ArraySpace{2,Float64} contains all possible matrices of size m × n for any combination of m,n ∈ ℕ.

DomainSets.BidirectionalMappedDomain — Type.

A BidirectionalMappedDomain stores the source domain and both the map f and a left inverse.

DomainSets.ComplexSpace — Type.

A geometric space with complex type.

DomainSets.EmbeddingMap — Type.

A map between embedded spaces.

DomainSets.FixedInterval — Type.

The abstract type FixedInterval is the supertype of intervals with endpoints determined by the type, rather than field values. Examples include UnitInterval and ChebyshevInterval.

DomainSets.ForwardMappedDomain — Type.

A ForwardMappedDomain stores the source domain and the forward map f, which maps source to target.

DomainSets.GeometricSpace — Type.

A domain is always a subset of a geometric space. A geometric space is completely characterized by the type T of its elements. There are no restrictions on concrete values: each instance of type T belongs to GeometricSpace{T}. Conversely, each numeric type T can give rise to only one geometric space.

The goal of GeometricSpace is to provide a lightweight framework for converting elements between spaces. Two features are embeddings and isomorphisms. It is lightweight because the space is characterized by the type and has no state, hence it can be inferred from any variable based only on its type. A restriction that results from this choice is that there can only be one type of embedding from one space to another. More general embeddings could be implemented using domains or specific maps.

Embeddings: A space A is embedded in a space B if each element of A corresponds to an element of B. In that case an element of A can be promoted (using promote_space) to an element of type B. This is not necessarily the same as the promotion of the type of A to the type of B. For example, a scalar can be embedded into a two-dimensional space (say with the second component equal to zero). It would be undesirable to implement that using the standard Julia promotion system, since it would apply to all Julia code used concurrently.

Isomorphisms: If A is embedded in B and B is embedded in A than they A and B are isomorphic. In that case, an element of A can be converted to an element of B and vice-versa. The conversion is invertible. One example is the isomorphism between ℝ^2 and ℂ.

New spaces can be defined by defining a new numeric type. Embeddings and isomorphisms are defined by defining new embedding_reduction and isomorphism_reduction rules, along with conversiong using convert_space and restrict_space. See the documentation of these functions for information on how to use them.

DomainSets.IntegerSpace — Type.

A geometric space with integer type.

DomainSets.InverseMappedDomain — Type.

An InverseMappedDomain stores the source and the inverse of the map f.

DomainSets.IsomorphismMap — Type.

A map between isomorphic spaces.

DomainSets.MappedDomain — Type.

There are three objects involved in the mapping of a domain:

the original domain (denoted source)

the map (denoted f)

the resulting domain (target)

If f maps a variable of type S to a variable of type T, then the source domain has eltype S and the target domain has eltype T.

The characteristic function of the resulting domain is defined in terms of the inverse of the map f, i.e.:

in(x, target) = in(inv(f)*x, source)

Concrete mapped domains can be implemented in various ways, e.g. by storing source and f, or by storing source and inv(f), ...

DomainSets.ProductSpace — Type.

By convention a geometric space with a tuple type T represents a product space of the individual entries of T.

DomainSets.RationalSpace — Type.

A geometric space with rational types.

DomainSets.RealSpace — Type.

A geometric space with floating point type.

DomainSets.RestrictionMap — Type.

A restriction map from a space to an embedded space.

DomainSets.TypeFactory — Type.

A TypeFactory{T} is a convenience type to simplify construction of a type.

Having t = TypeFactory{T} overrides getindex such that t[a] invokes T(a).

For example:

v = TypeFactory{SVector}
v[0.1,0.2]

makes an SVector{2,Float64}.

DomainSets.UnitCircleMap — Type.

The map [cos(2πt), sin(2πt)] from [0,1) to the unit circle in ℝ^2.

DomainSets.UnityMap — Type.

The unity map f(x) = 1.

DomainSets.VectorSpace — Type.

A Euclidean space with static vector types. It is the space of all vectors of fixed length N, with entries of type T.

DomainSets.ZeroMap — Type.

The zero map f(x) = 0.

DomainSets._product_eltype — Method.

Create an eltype that is suitable for a product domain. The result is a tuple type, where each of the elements is the eltype of the corresponding element of the product domain.

DomainSets.affine_composition — Method.

Compute the affine map that represents map2 after map1, that is: y = a2(a1x+b1)+b2 = a2a1x + a2*b1 + b2.

DomainSets.convert_space — Method.

Convert the variable x to an element of the space B. This is possible if the space of x is embedded in B.

DomainSets.default_tolerance — Method.

Return a suitable tolerance to use for verifying whether a point is close to a domain. Typically, the tolerance is close to the precision limit of the numeric type associated with the domain.

DomainSets.demote — Method.

The function demote(S, y::T) is a left inverse of convert(T, x::S), for the case where type S promotes to type T (i.e. promote_type(S,T) == T).

This means that demote(S, convert(T, x::S)) == x, while demote(S, y) may be anything for y::T not in the range of convert(T, x::S).

Note that Julia's convert function is its own left inverse, in the sense that convert(S, convert(T,x)) == x usually holds if x::S. However, the convert function generally throws an InexactError() for elements of type T that are not in (or close to) the range of convert(T, x::S). In those cases, demote does not throw an error, but there may be an arbitrarily large difference between convert(T, demote(S, y)) and y itself. In other words, demote may differ wildly from a right inverse of convert.

DomainSets.element — Method.

Return the i-th element of a composite structure.

See also: elements.

DomainSets.elements — Method.

Some types have composite structure, e.g. product domains, a union of domains. These types contain a list of domains.

It is often undesirable to use getindex to access the elements of the composite type. For this reason we introduce the elements functions. Composite types can implement elements and provide a generic way to access their components.

elements(t): returns the elements making up the composite type t

element(t, i): return the i-th element of the composite type t

numelements(t): return the number of elements of the composite type t

DomainSets.embedding — Method.

The function embedding describes in the type domain whether a space with type T is embedded into a space with type B. If so, it returns True, otherwise it returns False.

Embeddings are the result of the following rules:

A{T} is embedded in A{S} if T promotes to S in the Julia type system

If a rule embedding_reduction(A,B) = (C,D) has been defined, then A is embedded in B if C is embedded in D.

Say A and B are isomorphic if C and D are. This automatically results in embedding reduction rules from A to B and from B to A.

A is embedded in B if the superspace of A is embedded in B.

DomainSets.isomorphism — Method.

Return True if A and B are isomorphic. This implies that any point in A can be converted to a point in B, and vice-versa, with an exact inverse.

The functionality is implemented recursively in terms of the isomorphism_reduction function.

DomainSets.isomorphism_reduction — Method.

Indicate a reduction rule for isomorphism. The function either returns () or a 2-tuple. If isomorphism_reduction(A,B) returns (C,D), this means that A and B are isomorphic if C and D are. Here, C and D are the subeltypes of A and B respectively.

For example, VectorSpace{2,T} and ComplexSpace{S} are isomorphic if GeometricSpace{T} and GeometricSpace{S} are isomorphic.

The isomorphism only has to be declared in one direction, and is automatically valid in both directions. For space promotions, in case of ambiguity the first space A here is chosen as the preferred promotion space.

DomainSets.numelements — Method.

Return the number of elements of a composite structure.

See also: elements.

DomainSets.promote_space — Method.

Promote the geometric spaces of the arguments to a joined space, to which all arguments can be converted using embeddings. If no such concrete space exists, the joined space is AnySpace and the arguments remain unaltered.

DomainSets.promotes_to — Method.

Return True if S promotes to T, i.e., if promote_type(S,T) == T.

DomainSets.restrict_space — Method.

Restrict the variable x to an element of the space B. This is possible if B is embedded in the space of x.

Mathematically, restrict_space(A, y) for y in space B is a left inverse of convert_space(B, x) for x in space A. This means that restrict_space(A, convert_space(B, x)) == x for any x in space A. However, restrict_space(A, y) for a y not in the range of convert_space(B, x) could take any value.

DomainSets.return_type — Method.

return_type(map, U) is a generic function that computes the return type when the given map is applied to a variable of type U.

For any AbstractMap{S,T}, we have that return_type(map, S) = T. We also return T if U can be promoted to T using Julia's promotion system.

DomainSets.similar_interval — Method.

Return an interval that is similar to the given interval, but with endpoints a and b instead.

DomainSets.similar_space — Method.

Create a space that is similar to the given space, but with a different subeltype.

For example, similar_space(ComplexSpace{Float64}, BigFloat) yields a ComplexSpace{BigFloat}.

DomainSets.simplify_product_eltype — Method.

Try to simplify the type of a product domain to a type to which it is isomorphic. The goal is to automatically embed the product domain in ℝ^N if possible.

Examples of simplifications: Tuple{Float64,Float64} -> SVector{2,Float64} Tuple{Tuple{Float64,Float64},Float64} -> SVector{3,Float64}

DomainSets.supereltype — Method.

Return the eltype of the superdomain.