Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DrawSimpleGraphs.KnightTourDrawing — Function.

KnightTourDrawing(r,c) illustrates a Knight's tour of an r-by-c chessboard returning true if the tour exists and false if not.

SimpleDrawing.draw — Method.

draw(G::SimpleGraph) draws G in its current embedding. (If G does not have an embedding, then it is given a circular embedding.)