High-level API

Sampling and accessors

Most users would use this function, which initializes and tunes the parameters of the algorithm, then samples. Parameters can be set manually for difficult posteriors.

DynamicHMC.NUTS_init_tune_mcmc — Function.

NUTS_init_tune_mcmc(rng, ℓ, N; args...)

Init, tune, and then draw N samples from ℓ using the NUTS algorithm.

Return the sample (a vector of NUTS_transitions) and the tuned sampler.

rng is the random number generator.

args are passed on to various methods, see NUTS_init and bracketed_doubling_tuner.

Most users would use this function, unless they are doing something that requires manual tuning.

NUTS_init_tune_mcmc(ℓ, N; args...)

Same as the other method, but with random number generator Random.GLOBAL_RNG.

Important

The NUTS sampler saves a random number generator and uses it for random draws. When running in parallel, you should initialize NUTS_init_tune_mcmc with a random number generator as its first argument explicitly, making sure that each thread has its own one.

These functions can be used use to perform the steps above manually.

NUTS_init
tune
mcmc

The resulting sample is a vector of NUTS_Transition objects, for which the following accessors exist:

DynamicHMC.NUTS_Transition — Type.

Single transition by the No-U-turn sampler. Contains new position and diagnostic information.

DynamicHMC.get_position — Function.

Position after transition.

DynamicHMC.get_neg_energy — Function.

Negative energy of the Hamiltonian at the position.

DynamicHMC.get_depth — Function.

Tree depth.

DynamicHMC.get_termination — Function.

Reason for termination, see Termination.

DynamicHMC.get_acceptance_rate — Function.

get_acceptance_rate(x)

Return average Metropolis acceptance rate.

Average acceptance rate over trajectory.

DynamicHMC.get_steps — Function.

Number of integrator steps.

DynamicHMC.get_position_matrix — Function.

variable_matrix(posterior)

Return the samples of the parameter vector as rows of a matrix.

Diagnostics

These are NUTS-specific diagnostics and statistics (except for sample_cov, which is a convenience function). It is also prudent to use generic MCMC convergence diagnostics, as suggested in the introduction.

DynamicHMC.NUTS_statistics — Function.

NUTS_statistics(sample)

Return statistics about the sample (ie not the variables). Mostly useful for NUTS diagnostics.

DynamicHMC.sample_cov — Function.

sample_cov(sample)

Covariance matrix of the sample.

DynamicHMC.EBFMI — Function.

EBFMI(sample)

Energy Bayesian fraction of missing information. Useful for diagnosing poorly chosen kinetic energies.

Low values (≤ 0.3) are considered problematic. See Betancourt (2016).

Fine-grained control

Kinetic energy

DynamicHMC.KineticEnergy — Type.

abstract type KineticEnergy

Kinetic energy specifications.

For all subtypes, it is implicitly assumed that kinetic energy is symmetric in the momentum p, ie.

neg_energy(::KineticEnergy, p, q) == neg_energy(::KineticEnergy, -p, q)

When the above is violated, the consequences are undefined.

DynamicHMC.EuclideanKE — Type.

abstract type EuclideanKE <: KineticEnergy

Euclidean kinetic energies (position independent).

DynamicHMC.GaussianKE — Type.

struct GaussianKE{T<:(AbstractArray{T,2} where T), S<:(AbstractArray{T,2} where T)} <: EuclideanKE

Gaussian kinetic energy.

\[p ∣ q ∼ N(0, M)\]

independently of $q$.

The inverse covariance $M⁻¹$ is stored.

NUTS parameters and tuning

NUTS
StepsizeTuner
StepsizeCovTuner
TunerSequence
mcmc_adapting_ϵ
bracketed_doubling_tuner