Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

DynamicalSystemsBase.ContinuousDynamicalSystem — Type.

ContinuousDynamicalSystem(eom, state, p [, jacobian [, J]]; t0 = 0.0)

A DynamicalSystem restricted to continuous-time systems (also called ODEs).

DynamicalSystemsBase.DiscreteDynamicalSystem — Type.

DiscreteDynamicalSystem(eom, state, p [, jacobian [, J]]; t0::Int = 0)

A DynamicalSystem restricted to discrete-time systems (also called maps).

DynamicalSystemsBase.DynamicalSystem — Type.

DynamicalSystem

The central structure of DynamicalSystems.jl. All functions of the suite that can use known equations of motion expect an instance of this type.

Constructing a DynamicalSystem

DiscreteDynamicalSystem(eom, state, p [, jacobian [, J0]]; t0::Int = 0)
ContinuousDynamicalSystem(eom, state, p [, jacobian [, J0]]; t0 = 0.0)

with eom the equations of motion function (see below). p is a parameter container, which we highly suggest to use a mutable object like Array, LMArray or a dictionary. Pass nothing in the place of p if your system does not have parameters.

t0, J0 allow you to choose the initial time and provide an initialized Jacobian matrix. See CDS_KWARGS for the default options used to evolve continuous systems (through OrdinaryDiffEq).

Equations of motion

The are two "versions" for DynamicalSystem, depending on whether the equations of motion (eom) are in-place (iip) or out-of-place (oop). Here is how to define them:

oop : The eom must be in the form eom(x, p, t) -> SVector which means that given a state x::SVector and some parameter container p it returns an SVector (from the StaticArrays module) containing the next state.

iip : The eom must be in the form eom!(xnew, x, p, t) which means that given a state x::Vector and some parameter container p, it writes in-place the new state in xnew.

t stands for time (integer for discrete systems). iip is suggested for big systems, whereas oop is suggested for small systems. The break-even point at around 100 dimensions, and for using functions that use the tangent space (like e.g. lyapunovs or gali), the break-even point is at around 10 dimensions.

The constructor deduces automatically whether eom is iip or oop. It is not possible however to deduce whether the system is continuous or discrete just from the equations of motion, hence the 2 constructors.

Jacobian

The optional argument jacobian for the constructors is a function and (if given) must also be of the same form as the eom, jacobian(x, p, n) -> SMatrix for the out-of-place version and jacobian!(xnew, x, p, n) for the in-place version.

If jacobian is not given, it is constructed automatically using the module ForwardDiff. It is heavily advised to provide a Jacobian function, as it gives multiple orders of magnitude speedup.

Interface to DifferentialEquations.jl

Continuous systems are solved using DifferentialEquations.jl. The following two interfaces are provided:

ContinuousDynamicalSystem(prob::ODEProblem [, jacobian [, J0]])
ODEProblem(continuous_dynamical_system, tspan, args...)

where in the second case args stands for the standard extra arguments of ODEProblem: callback, mass_matrix.

If you want to use callbacks with tangent_integrator or parallel_integrator, then invoke them with extra arguments as shown in the Advanced Documentation.

Relevant Functions

trajectory, set_parameter!.

DynamicalSystemsBase.dimension — Method.

dimension(thing) -> D

Return the dimension of the thing, in the sense of state-space dimensionality.

DynamicalSystemsBase.get_deviations — Function.

get_deviations(tang_integ)

Return the deviation vectors of the tangent_integrator in a form of a matrix with columns the vectors.

DynamicalSystemsBase.get_state — Method.

get_state(ds::DynamicalSystem)

Return the state of ds.

get_state(integ [, i::Int = 1])

Return the state of the integrator, in the sense of the state of the dynamical system.

If the integrator is a parallel_integrator, passing i will return the i-th state. The function also correctly returns the true state of the system for tangent integrators.

DynamicalSystemsBase.integrator — Function.

integrator(ds::DynamicalSystem [, u0]; diffeq...) -> integ

Return an integrator object that can be used to evolve a system interactively using step!(integ [, Δt]). Optionally specify an initial state u0.

The state of this integrator is a vector.

diffeq... are keyword arguments propagated into init of DifferentialEquations.jl. See trajectory for examples. Only valid for continuous systems.

DynamicalSystemsBase.jacobian — Function.

jacobian(ds::DynamicalSystem, u = ds.u0, t = ds.t0)

Return the jacobian of the system at u, at t.

DynamicalSystemsBase.parallel_integrator — Function.

parallel_integrator(ds::DynamicalSystem, states; kwargs...)

Return an integrator object that can be used to evolve many states of a system in parallel at the exact same times, using step!(integ [, Δt]).

states are expected as vectors of vectors.

Keyword Arguments

diffeq... : Keyword arguments propagated into init of DifferentialEquations.jl. See trajectory for examples. Only valid for continuous systems. These keywords can also include callback for event handling.

It is heavily advised to use the functions get_state and set_state! to manipulate the integrator. Provide i as a second argument to change the i-th state.

DynamicalSystemsBase.set_deviations! — Function.

set_deviations!(tang_integ, Q)

Set the deviation vectors of the tangent_integrator to Q, which must be a matrix with each column being a deviation vector.

Automatically does u_modified!(tang_integ, true).

DynamicalSystemsBase.set_parameter! — Method.

set_parameter!(ds::DynamicalSystem, index, value)
set_parameter!(ds::DynamicalSystem, values)

Change one or many parameters of the system by setting p[index] = value in the first case and p .= values in the second.

The same function also works for any integrator.

DynamicalSystemsBase.set_state! — Method.

set_state!(integ, u [, i::Int = 1])

Set the state of the integrator to u, in the sense of the state of the dynamical system. Works for any integrator (normal, tangent, parallel).

For parallel integrator, you can choose which state to set (using i).

Automatically does u_modified!(integ, true).

DynamicalSystemsBase.tangent_integrator — Function.

tangent_integrator(ds::DynamicalSystem, Q0 | k::Int; kwargs...)

Return an integrator object that evolves in parallel both the system as well as deviation vectors living on the tangent space.

Q0 is a matrix whose columns are initial values for deviation vectors. If instead of a matrix Q0 an integer k is given, then k random orthonormal vectors are choosen as initial conditions.

Keyword Arguments

u0 : Optional different initial state.

diffeq... : Keyword arguments propagated into init of DifferentialEquations.jl. See trajectory for examples. Only valid for continuous systems. These keywords can also include callback for event handling.

It is heavily advised to use the functions get_state, get_deviations, set_state!, set_deviations! to manipulate the integrator.

Description

If $J$ is the jacobian of the system then the tangent dynamics are the equations that evolve in parallel the system as well as a deviation vector (or matrix) $w$:

\[\begin{aligned} \dot{u} &= f(u, p, t) \\ \dot{w} &= J(u, p, t) \times w \end{aligned}\]

with $f$ being the equations of motion and $u$ the system state. Similar equations hold for the discrete case.

DynamicalSystemsBase.trajectory — Function.

trajectory(ds::DynamicalSystem, T [, u]; kwargs...) -> dataset

Return a dataset that will contain the trajectory of the system, after evolving it for total time T, optionally starting from state u. See Dataset for info on how to use this object.

A W×D dataset is returned, with W = length(t0:dt:T) with t0:dt:T representing the time vector (not returned) and D the system dimension. For discrete systems both T and dt must be integers.

Keyword Arguments

dt : Time step of value output during the solving of the continuous system. For discrete systems it must be an integer. Defaults to 0.01 for continuous and 1 for discrete.

Ttr : Transient time to evolve the initial state before starting saving states.

diffeq... : Keyword arguments propagated into init of DifferentialEquations.jl. For example abstol = 1e-9. Only valid for continuous systems. If you want to specify a solver, do so by using the name alg, e.g.: alg = Tsit5(), maxiters = 1000. This requires you to have been first using OrdinaryDiffEq to access the solvers. See DynamicalSystemsBase.CDS_KWARGS for default values. These keywords can also include callback for event handling. Using a SavingCallback with trajectory will lead to unexpected behavior!