Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

EDFPlus.Annotation — Type.

mutable struct Annotation

These are text strings denoting a time, optionally duration, and a list of notes about the signal at that particular time in the recording. The first onset time of the annotation channel gives a fractional second offset adjustment of the start time of that record, which is specified in whole seconds in the header.

EDFPlus.BEDFPlus — Type.

mutable struct BEDFPlus

Data struct for EDF, EDF+, BDF, and BDF+ EEG type signal files.

EDFPlus.ChannelParam — Type.

mutable struct ChannelParam

Parameters for each channel in the EEG record.

EDFPlus.DataFormat — Type.

DataFormat

enum for types this package handles. Current format for a potential translation is also /same/.

source

EDFPlus.FileStatus — Type.

FileStatus

enum for type or state of file: type of data detected, whether any errors

source

EDFPlus.addannotation! — Method.

addannotation!(edfh, onset, duration, description)

Add an annotation at the given onset timepoint IF there is room Note the description arg is a text string, not an array argument here

EDFPlus.annotation_epoch_iterator — Method.

annotation_epoch_iterator(edfh, epochsecs; startsec, endsec)

Return an iterator for a group of annotations for a given epoch as in epoch_iterator

EDFPlus.channeltimesegment — Method.

channeltimesegment(edfh, channel, startsec, endsec, physical)

get the channel's data between the time points

EDFPlus.closefile! — Method.

closefile!(edfh)

Close the file opened by loadfile and loaded to the BEDFPlus struct Releases memory from read data in edfh

EDFPlus.digitalchanneldata — Method.

digitalchanneldata(edfh, channelnumber)

Get a single digital channel of data in its entirety.

Arguments:

edfh the BEDFPlus struct

channelnumber the channel number in the records

EDFPlus.epoch_iterator — Method.

epoch_iterator(edfh, epochsecs; channels, startsec, endsec, physical)

Make an iterator for EEG epochs of a given duration between start and stop times.

Required arguments

edfh BEDFPlus struct

epochsecs second duration of each epoch

Optional arguments

channels List of channel numbers for data, defaults to all signal channels

startsec Starting position from 0 at start of file, defaults to file start

endsec Ending position in seconds from start of file, defaults to file end

physical Whether to return data as translated to the physical units, defaults to true

EDFPlus.highpassfilter — Function.

highpassfilter(signals, fs, cutoff=1.0, order=4)

Apply high pass filter to signals, return filtered data

EDFPlus.loadfile — Function.

loadfile(path::String, read_annotations=true)

Load a BDF+ or EDF+ type file. Takes a pathname. Will ignore annotations if second argument is set false. Returns a BEDFPlus structure including header and data.

EDFPlus.lowpassfilter — Function.

lowpassfilter(signals, fs, cutoff=25.0, order=4)

Apply low pass filter to signals, return filtered data

EDFPlus.multichanneltimesegment — Method.

multichanneltimesegment(edfh, chanlist, startsec, endsec, physical)

Get an multichannel array of lists of datapoints over time segment NB: best if all datapoint signal rates are the same

EDFPlus.notchfilter — Function.

notchfilter(signals, fs, notchfreq=60, q = 35)

Notch filter signals in array signals, return filtered signals

EDFPlus.physicalchanneldata — Method.

physicalchanneldata(edfh, channelnumber)

Get a single data channel in its entirely, in the physical units stated in the header

Arguments

edfh the BEDFPlus struct

channelnumber the channel number in the records– a channel in the mapped_signals list

EDFPlus.samplerate — Method.

samplerate(edfh, channel)

Get sample (sampling) rate (fs) on the channel in sec^-1 units

EDFPlus.trim — Method.

trim(str)

trim whitespace fore and aft, as in trim in java etc

EDFPlus.version — Method.

static function to state version of module

EDFPlus.writefile! — Method.

writefile!(edfh, newpath; acquire=dummyacquire, sigformat=same)

Write to data in the edfh struct to the file indicated by newpath Returns the file handle of the file written, opened for reading

NOTE: The header needs to be completely specified at function start except for the final number of records, which will be updated after all data records are written. For a system that is recording the data as it is written, the acquire(edfh) function should write the data according the the header parameters. NB: iff the function converts from BDF to EDF or EDF to BDF, the edfh struct is changed.

EDFPlus.Int24 — Type.

type Int24

24-bit integer routines for BDF format signal data.

BDF and BDF+ files use 24 bits per data signal point. The module caches these after reading as Int32 to fit LLVM CPU registers.

EDFPlus.annotationtoTAL — Method.

annotationtoTAL(ann)

Create a TAL (timestamped annotation list) text entry out of an annotation

EDFPlus.bytesperdatapoint — Method.

bytesperdatapoint(edfh)

Return how many bytes used per data point entry: 2 for EDF (16-bit), 3 for BDF (24-bit) data.

EDFPlus.checkfile! — Method.

checkfile!(edfh)

Helper function for loadfile and writefile!

EDFPlus.dash2space — Method.

replace underslashes in string with spaces

EDFPlus.datapointinterval — Function.

datapointinterval(edfh, channel)

Time interval in fractions of a second between individual signal data points

EDFPlus.dummyacquire — Method.

dummyacquire(edfh)

Dummy function for call in writefile! for optional acquire function If using package for data acquisition will need to custom write the acquire function for your calls to writefile!

EDFPlus.epochmarkers — Method.

epochmarkers(edfh, secs)

Get a set of (start, stop) positional markers for epochs (sequential windows) given an epoch duration in seconds

EDFPlus.latintoascii — Method.

latintoascii(str)

Helper function for writefile! related functions

EDFPlus.readannotations! — Method.

readannotations!(edfh)

Helper function for loadfile

EDFPlus.readdata! — Method.

readdata!(edfh)

Helper function for loadfile, reads signal data into the BEDFPlus struct

EDFPlus.recordindexat — Method.

recordindexat(edfh, secondsafterstart)

Index of the record point at or closest just before a given time from recording start Translates a values in seconds to a position in the signal data matrix, returns that record's position

EDFPlus.recordslice — Method.

recordslice(edfh, startpos, endpos)

Get a slice of the data in the recording from one data entry position to another

EDFPlus.signalat — Function.

signalat(edfh, secondsafter, channel)

Get the position in the signal data of the data point at or closest after a given time from recording start. Translates a value in seconds to a position in the signal channel matrix, returns that signal data point's 2D position as list

EDFPlus.signaldata — Method.

signaldata(edfh)

Return which BEDFPlus variable holds the signal data

EDFPlus.signalindices — Method.

signalindices(edfh, channelnumber)

Get a pair of indices of a channel's bytes within each of the data records

EDFPlus.translate16to24bits! — Method.

translate16to24bits!(edfh)

Translate 16 bit data to 32-bit width, for change to 24-bit data for writefile!

EDFPlus.translate24to16bits! — Method.

translate24to16bits!(edfh)

Translate data in 24-bit BDF to 16-bit EDF format Helper function for writefile!

EDFPlus.trimrightzeros — Method.

trimrightzeros compact number string by trimming nonsignificant decimal places/point when not zero

EDFPlus.writeBDFrecords! — Method.

writeBDFrecords!(edfh, fh)

Write an BEDFPlus format file Helper file for writefile!

EDFPlus.writeBDFsignalchannel — Method.

writeBDFsignalchannel(edfh,fh, record, channel)

Helper function for writefile! write a BDF record's worth of a signal channel at given record and channel number

EDFPlus.writeEDFrecords! — Method.

writeEDFrecords!(edfh, fh)

Helper function for writefile! Write a record's worth of all channels of a given record

EDFPlus.writeEDFsignalchannel — Method.

writeEDFsignalchannel(edfh, fh, record, channel)

Helper function for writefile! write a record's worth of a signal channel at given record and channel number

EDFPlus.writeheader — Method.

writeheader(edfh, fh)

Helper function for writefile!

EDFPlus.writeleftjust — Function.

writeleftjust(fh, x, len, fillchar=' ')

Write a stringified object to a file in the leftmost portion of chars written, filling with fillchar to len length as needed, truncate if too long for field