Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

EchoviewEvr.regions — Method.

regions(filename::AbstractString)

regions returns an iterator over regions in the given Echoview EVR file.

EchoviewEvr.regions — Method.

regions(filenames::Vector{String})

regions returns an iterator over regions in the given Echoview EVR files.

EchoviewEvr.Region — Type.

An Echoview region.

See documentation.

EchoviewEvr.load — Method.

load(filename::AbstractString)

Loads an Echoview region file.