Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Electron.Application — Method.

function Application()

Start a new Electron application. This will start a new process for that Electron app and return an instance of Application that can be used in the construction of Electron windows.

Electron.Window — Type.

function Window([app::Application,] options::Dict)

Open a new Window in the application app. Pass the content of options to the Electron BrowserWindow constructor.

If app is not specified, use the default Electron application, starting one if needed.

Electron.Window — Method.

function Window([app::Application,] uri::URI)

Open a new Window in the application app. Show the content as a text/html file with utf-8 encoding.

If app is not specified, use the default Electron application, starting one if needed.

Electron.Window — Method.

function Window([app::Application,] uri::URI)

Open a new Window in the application app. Show the content that uri points to in that new window.

If app is not specified, use the default Electron application, starting one if needed.

Electron.load — Method.

load(win::Window, html::AbstractString)

Load html in the Electron window win.

Electron.load — Method.

load(win::Window, uri::URI)

Load uri in the Electron window win.

Base.Filesystem.pwd — Method.

pwd(URI)

Return pwd() as a URI resource.

Base.close — Method.

close(app::Application)

Terminates the Electron application referenced by app.

Base.close — Method.

close(win::Window)

Close the windows referenced by win.

Base.run — Method.

run(app::Application, code::AbstractString)

Run the JavaScript code that is passed in code in the main application thread of the app Electron process. Returns the value that the JavaScript expression returns.

Base.run — Method.

run(win::Window, code::AbstractString)

Run the JavaScript code that is passed in code in the render thread of the win Electron windows. Returns the value that the JavaScript expression returns.

Electron.URI_file — Method.

URI_file(base, filespec)

Construct an absolute URI to filespec relative to base.