Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Elly.HDFSClient — Type.

HDFSClient

A client to the namenode in a HDFS cluster and holds a connection to it. It also stores the folder context for using relative paths in APIs that use the client.

Elly.HDFSFile — Type.

HDFSFile

Holds a HDFSClient and path pointing to a file on HDFS

Elly.HDFSFileInfo — Type.

HDFSFileInfo

Stat structure for file/foledr in HDFS

Elly.YarnApp — Type.

YarnApp represents one instance of application running on the yarn cluster

Elly.YarnAppAttemptStatus — Type.

YarnAppAttemptStatus wraps the protobuf type for ease of use

Elly.YarnAppMaster — Type.

YarnAppMaster is a skeleton application master. It provides the generic scafolding methods which can be used to create specific application masters for different purposes.

Elly.YarnAppStatus — Type.

YarnAppStatus wraps the protobuf type for ease of use

Elly.YarnClient — Type.

YarnClient holds a connection to the Yarn Resource Manager and provides APIs for application clients to interact with Yarn.

Elly.YarnNode — Type.

YarnNode represents a node manager in the yarn cluster and its communication address, resource state and run state.

Elly.hdfs_renewlease — Method.

Applications that write infrequently/slowly must call renewlease periodically to prevent the namenode from assuming the client from having abandoned the file or some other client from recovering the lease.

Elly.HDFSBlockReader — Type.

HDFSBlockReader reads one block of data from a datanode.

Elly.HDFSBlockWriter — Type.

HDFSBlockWriter writes one block worth of data to a datanode.

The write method returns number of bytes accepted for writing into this block, which may be less than what was requested if block is full.

Packets are sent and acks received as and when data being written exceed packet size limit.

The close operation flushes remaining data in the block and waits for all pending acks.

Elly.HDFSException — Type.

HDFSException

Thrown by HDFS APIs.

Elly.HDFSFileReader — Type.

HDFSFileReader

Provides Julia IO APIs for reading HDFS files. Communicates with namenode for file metadata (through HDFSClient) and to datanodes for file data (through HDFSBlockReader)

Elly.HDFSFileWriter — Type.

HDFSFileWriter

Provides Julia IO APIs for writing HDFS files.

Steps:

get datanodes to write onto from namenode

break bytes into packets

write packets into first datanode's data channel (datanode will mirror packets to downstream datanodes)

when complete, confirm namenode with a blockReceived

On close call NameNode.complete to:

remove lease lease is a write lock for file modification. no leases are required for reading files. leases are managed by namenode lease is added on file create or append dfs client should start thread to renew leases periodically

change file from under construction to complete

Elly.HDFSProtocol — Type.

HDFSProtocol: Hadoop RPC client for namenode protocol.

Elly.HadoopDataChannel — Type.

HadoopDataChannel is the connection to a datanode. Also associated with it is a buffer to help in message preparation.

Elly.HadoopDataChannelPool — Type.

HadoopDataChannelPool is a connection pool that holds connections to datanodes. Connections are deemed stale after keepalivesecs.

Elly.HadoopRpcException — Type.

HadoopRpcException is thrown on Rpc interaction errors either with namenode or datanode. Field status contains error code (enum) if received from the connected entity or just ERROR (1) to indicate failure. Though HadoopRpcException is used while communicating with both namenodes and datanodes, SUCCESS and ERROR are coded with the same values in both cases. Other specific error codes need knowledge of the context to interpret.

Elly.HadoopRpcProtocol — Type.

HadoopRpcProtocol binds a channel and controller with a service protocol implementation. Used by actual service implementations.

Elly.PipelinedPacket — Type.

PipelinedPacket holds a packet of data while it is sent to datanodes and acknowledged. The data bytes are reset once acks are received successfully to reduce memory usage.

Elly.RequestPipeline — Type.

RequestPipeline holds entities while they are requested for from yarn resource manager. Application master thread extracts pending items and requests them from RM, whereupon they are moved on to the requested state.

Elly.WriterPipeline — Type.

WriterPipeline holds all packets of data for a block and provides methods to maintain their states. Pipeline failed status is set if any of the acks received is a failure. Field acked_bytes contains the count of bytes successfully sent till an error is encountered.

Elly.YarnAMNMProtocol — Type.

YarnAMNMProtocol: Hadoop RPC client for Yarn application master to node manager protocol.

Elly.YarnAMRMProtocol — Type.

YarnAMRMProtocol: Hadoop RPC client for Yarn application master to resource manager protocol.

Elly.YarnClientProtocol — Type.

YarnClientProtocol: Hadoop RPC client for application client to Yarn resource manager protocol.

Elly.YarnContainers — Type.

YarnContainers holds all containers related to the application. It also holds the allocation and release pipelines that are used by application master for requesting actions from resource manager. Also schedules callbacks as tasks when containers are allocated or terminated.

Elly.YarnException — Type.

YarnException

Thrown by Yarn APIs.

Elly.YarnNodes — Type.

YarnNodes holds node information as visible to the application master. It also caches connection to node masters. Connection are reused if they are required before a set keepalivesecs time.

Elly.read_packet! — Method.

Read one packet into inbuff starting from offset. If inbuff has insufficient space, returns the minimum additional space required in inbuff to read the packet as a negative number. Otherwise, returns the number of bytes available in inbuff after reading the packet.