Entropic Vector

Entropic Vector

Entropic Index

The index of an entropic vector of $n$ variables is a subset of the set $[n] = \{1, 2, \ldots, n\}$. This set is used to index the entropic vector. A set can be created using the set function.

EntropicCone.set — Function.

set(i::Integer)

Return the set of digits of the integer i. Note that this cannot create a set with elements other than 1, 2, 3, 4, 5, 6, 7, 8 and 9. For instance, the set ${2, 4, 10}$ cannot be constructed with this method, use set([2, 4, 10]) instead.

Examples

To create the set $\{2, 4\}$, use set(24) or set(42).

set(I::AbstractArray{<:Integer})

Return the set of elements of I.

Examples

To create the set $\{2, 4\}$, use set([2, 4]) or set([4, 2]).