Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ExcelReaders.ExcelErrorCell — Type.

ExcelErrorCell

An Excel cell that has an Excel error.

You cannot create $ExcelErrorCell$ objects, they are returned if a cell in an Excel file has an Excel error.

ExcelReaders.ExcelFile — Type.

ExcelFile

A handle to an open Excel file.

You can create an instance of an $ExcelFile$ by calling $openxl$.

ExcelReaders.openxl — Method.

openxl(filename)

Open the Excel file $filename$ and return an $ExcelFile$ handle.

The returned $ExcelFile$ handle can later be passed as the first argument to $readxl$ or $readxslsheet$ to read from that file. If you will call either of those functions more than once, performance will be better if you open the file only once with $openxl$.

Example

f = openxl("filename.xlsx")
data = readxl(f, "Sheet1!A1:C4")