Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Exercism.create_submission — Method.

Open the notebook for the excersice 'slug' in the current directory and create a submittable 'slug.jl' file containing the user solution.

Exercism.parse_notebook — Method.

Parses the notebook in 'path' and returns the code of all cells marked with # submit