Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ExpectationStubs.all_expectations_used — Method.

all_expectations_used(stub::Stub)

Checks that every expectation setup for the stub was actually used. It is good to have this as a sanity check at the end of your test script using the stub.

ExpectationStubs.DoNotCare — Type.

DoNotCare{T}

A type that is equal to all things that are of type <:T. For internal use. Will interact weirdly with hash based dicts

ExpectationStubs.ExpectationValueMismatchError — Type.

ExpectationValueMismatchError

Similar to a julia MethodError This is thrown if a call was made on a stub, but the values/types did not match those that were expected.

ExpectationStubs.get_all_matches — Method.

get_all_matches(vd::VectorDict, key)

Returns all values matching the key

ExpectationStubs.name_argvals_and_sig — Method.

name_argvals_and_sig(defn)

Does the extraction of these as is needed for @used and @usedcounts

ExpectationStubs.onlyesc — Method.

onlyesc(v)

Like esc, except it ignores things that can not be escaped. Eg literals.

Internal use

ExpectationStubs.split_vals_and_sig — Method.

split_vals_and_sig(argsexpr)

Takes an expression from a function definition's args eg :([a::Int, ::Integer]) and breaks it down into the values, and the types. When value is not given it subsitutes DoNotCare(), when type is not given it subsitutes typeof(val) Returns a tuple of each, with all things in expression/symbol form.

Internal use

ExpectationStubs.usecount — Method.

usecount(stub:Stub, arg)

Returns how many times the stub has been called with a matching argument