Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FCSFiles.grab_word — Method.

grab_word(iter, state, delimiter) -> word, state

Grabs the next word from the iterator iter. Takes care to handle escaped delimiters specified by delimiter. Returns a string containing the word and the state of the iterator.

FCSFiles.verify_text — Method.

verify_text(text_mappings) -> Void

Checks that all required keywords are present in the textmapping dictionary returned by `parsetext`