Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FastGroupBy.fastby — Method.

group by for DataFrame API

FastGroupBy.sumby — Function.

Perform sum by group

sumby(df::Union{AbstractDataFrame,NDSparse}, by::Symbol, val::Symbol)
sumby(by::AbstractVector  val::AbstractVector)

Arguments

df : an AbstractDataFrame/NDSparse from which to extract the by and val columns

by : data table column to group by

val: data table column to sum

Returns

::Dict : A Dict that maps unqiues values of by to sum of val

Examples

using FastGroupBy
using DataFrames, IndexedTables, Compat, BenchmarkTools
import DataFrames.DataFrame

const N = 10_000_000; const K = 100

# sumby is faster for DataFrame without missings
srand(1);
@time df = DataFrame(id = rand(1:Int(round(N/K)), N), val = rand(round.(rand(K)*100,4), N));
@belapsed DataFrames.aggregate(df, :id, sum)
@belapsed sumby(df, :id, :val)

FastGroupBy.sumby_contiguous — Method.

sumby assuming that the elements are organised contiguously; it does not perform a check

FastGroupBy.sumby_dict — Method.

sumby using Dict - can be quite slow due to slow hash table operations

FastGroupBy.sumby_radixgroup! — Method.

sumby by using radix and counting sort to group by; it's only a partial sort. It's faster for large by

FastGroupBy.sumby_radixsort! — Method.

sumby by sorting the by column using radixsort

FastGroupBy.sumby_sortperm — Method.

This is faster for smaller by and also doesn't change the input

IndexedTables.column — Method.

column(df, :colname)

Extract a column from an AbstractDataFrame

FastGroupBy._contiguousby — Method.

Apply by-operation assuming that the vector is grouped i.e. elements that belong to the same group by stored contiguously

FastGroupBy._contiguousby_vec — Method.

Apply by-operation assuming that the vector is grouped i.e. elements that belong to the same group by stored contiguously and return a vector

FastGroupBy._fastby! — Method.

Internal: single-function fastby, one by, one val

FastGroupBy.genca — Method.

Generate CategoricalArrays

FastGroupBy.select — Method.

select(:col)

Return a funciton that obtains a column with the named symbol from an AbstractDataFrame or NDSparse