Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FilePathsBase.Mode — Type.

The following file contains simple abstractions to make working with posix permissions easier in julia.

For now we're focused on user permissions.

A lot of the low level permissions code below and the corresponding constants have been translated from cpython's Lib/stat.py file.

https://github.com/python/cpython/blob/master/Lib/stat.py

Base.Filesystem.chown — Method.

chown(path::AbstractPath, user::AbstractString, group::AbstractString; recursive=false)

Change the user and group of the path.

FilePathsBase.Path — Method.

Path()
Path(path::AbstractPath)
Path(path::Tuple)
Path(path::AbstractString)

Responsible for creating the appropriate platform specific path (e.g., PosixPath and WindowsPath for Unix and Windows systems respectively)

FilePathsBase.created — Method.

created(path::AbstractPath) -> DateTime

Returns the creation date for the path.

Example

julia> created(p"src/FilePathsBase.jl")
2017-06-20T04:01:09

FilePathsBase.exists — Method.

exists(path::AbstractPath) -> Bool

Returns whether the path actually exists on the system.

FilePathsBase.extension — Method.

extension(path::AbstractPath) -> AbstractString

Extracts the last extension from a filename if there any, otherwise it returns an empty string.

Example

julia> extension(p"~/repos/FilePathsBase.jl/src/FilePathsBase.jl")
"jl"

FilePathsBase.extensions — Method.

extensions(path::AbstractPath) -> AbstractString

Extracts all extensions from a filename if there any, otherwise it returns an empty string.

Example

julia> extensions(p"~/repos/FilePathsBase.jl/src/FilePathsBase.jl.bak")
2-element Array{SubString{String},1}:
 "jl"
 "bak"

FilePathsBase.filename — Method.

filename(path::AbstractPath) -> AbstractString

Extracts the basename without any extensions.

Example

julia> filename(p"~/repos/FilePathsBase.jl/src/FilePathsBase.jl")
"FilePathsBase"

FilePathsBase.hasparent — Method.

hasparent(path::AbstractPath) -> Bool

Returns whether there is a parent directory component to the supplied path.

FilePathsBase.isexecutable — Method.

isexecutable(path::AbstractPath) -> Bool

Returns whether the path is executable for the current user.

FilePathsBase.mode — Method.

mode(path::AbstractPath) -> Mode

Returns the Mode for the specified path.

Example

julia> mode(p"src/FilePathsBase.jl")
-rw-r--r--

FilePathsBase.modified — Method.

modified(path::AbstractPath) -> DateTime

Returns the last modified date for the path.

Example

julia> modified(p"src/FilePathsBase.jl")
2017-06-20T04:01:09

FilePathsBase.parents — Method.

parents{T<:AbstractPath}(path::T) -> Array{T}

Example

julia> parents(p"~/.julia/v0.6/REQUIRE") 3-element Array{FilePathsBase.PosixPath,1}: p"~" p"~/.julia" p"~/.julia/v0.6"

Throws

ErrorException: if path doesn't have a parent

FilePathsBase.relative — Method.

relative{T<:AbstractPath}(path::T, start::T=T("."))

Creates a relative path from either the current directory or an arbitrary start directory.

Base.Filesystem.chmod — Method.

chmod(path::AbstractPath, mode::Mode; recursive=false)
chmod(path::AbstractPath, mode::Integer; recursive=false)
chmod(path::AbstractPath, user::UIn8=0o0, group::UInt8=0o0, other::UInt8=0o0; recursive=false)
chmod(path::AbstractPath, symbolic_mode::AbstractString; recursive=false)

Provides various methods for changing the mode of a path.

Examples

julia> touch(p"newfile")
Base.Filesystem.File(false, RawFD(-1))

julia> mode(p"newfile")
-rw-r--r--

julia> chmod(p"newfile", 0o755)

julia> mode(p"newfile")
-rwxr-xr-x

julia> chmod(p"newfile", "-x")

julia> mode(p"newfile")
-rw-r--r--

julia> chmod(p"newfile", user=(READ+WRITE+EXEC), group=(READ+EXEC), other=READ)

julia> mode(p"newfile")
-rwxr-xr--

julia> chmod(p"newfile", mode(p"src/FilePathsBase.jl"))

julia> mode(p"newfile")
-rw-r--r--

Base.Filesystem.isblockdev — Method.

Return True if mode is from a block special device file.

Base.Filesystem.ischardev — Method.

Return True if mode is from a character special device file.

Base.Filesystem.isdir — Method.

Return True if mode is from a directory.

Base.Filesystem.isfifo — Method.

Return True if mode is from a FIFO (named pipe).

Base.Filesystem.isfile — Method.

Return True if mode is from a regular file.

Base.Filesystem.islink — Method.

Return True if mode is from a symbolic link.

Base.Filesystem.issocket — Method.

Return True if mode is from a socket.

Base.abs — Method.

abs(path::AbstractPath) -> AbstractPath

Creates an absolute path by adding the current working directory if necessary.

Base.isempty — Method.

isempty(path::AbstractPath) -> Bool

Returns whether or not a path is empty.

NOTE: Empty paths are usually only created by Path(), as p"" and Path("") will default to using the current directory (or p".").

Base.isreadable — Method.

isreadable(path::AbstractPath) -> Bool

Returns whether the path is readable for the current user.

Base.iswritable — Method.

iswritable(path::AbstractPath) -> Bool

Returns whether the path is writable for the current user.

Base.join — Method.

join(root::AbstractPath, pieces::Union{AbstractPath, AbstractString}...) -> AbstractPath

Joins path components into a full path.

Example

julia> join(p"~/.julia/v0.6", "REQUIRE")
p"~/.julia/v0.6/REQUIRE"

Base.parent — Method.

parent{T<:AbstractPath}(path::T) -> T

Returns the parent of the supplied path.

Example

julia> parent(p"~/.julia/v0.6/REQUIRE")
p"~/.julia/v0.6"

Throws

ErrorException: if path doesn't have a parent

Base.real — Method.

real(path::AbstractPath) -> AbstractPath

Canonicalizes a path by expanding symlinks and removing "." and ".." entries.

Base.string — Method.

Convert a file's mode to a string of the form '-rwxrwxrwx'.

FilePathsBase._meta — Method.

Return the portion of the file's mode that describes the file type.

FilePathsBase._mode — Method.

Return the portion of the file's mode that can be set by os.chmod().

LinearAlgebra.norm — Method.

norm(path::AbstractPath) -> AbstractPath

Normalizes a path by removing "." and ".." entries.