Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Filetimes.datetime — Method.

datetime(f::Integer)

Returns a DateTime object given a filetime f, being the number of 100 nanosecond intervals since January 1, 1601.

N.B. Possible loss of precision.

Filetimes.filetime — Method.

filetime(s::AbstractString)

Returns the number of 100 nanosecond intervals since January 1, 1601, for the given ISO8601 string.

filetime("2014-09-02T08:20:07")

Filetimes.filetime — Method.

filetime(d::Dates.DateTime)

Returns the number of 100 nanosecond intervals since January 1, 1601 for the given DateTime.

Filetimes.filetime — Method.

filetime()

Returns the current time as the number of 100 nanosecond intervals since January 1, 1601.