Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FixedEffectModels.partial_out — Method.

Partial out variables in a Dataframe

Arguments

df::AbstractDataFrame

model::Model: A Model created using @model. See @model.

add_mean::Bool: Should the initial mean added to the returned variable?

method::Symbol: A symbol for the method. Default is :lsmr (akin to conjugate gradient descent). Other choices are :qr and :cholesky (factorization methods)

maxiter::Integer: Maximum number of iterations

tol::Real: Tolerance

Returns

::DataFrame: a dataframe with as many columns as there are dependent variables and as many rows as the original dataframe.

::Vector{Int}: a vector of iterations for each column

::Vector{Bool}: a vector of success for each column

Details

partial_out returns the residuals of a set of variables after regressing them on a set of regressors. The syntax is similar to reg - but it accepts multiple dependent variables. It returns a dataframe with as many columns as there are dependent variables and as many rows as the original dataframe. The regression model is estimated on only the rows where none of the dependent variables is missing. With the option add_mean = true, the mean of the initial variable is added to the residuals.

Examples

using  RDatasets, DataFrames, FixedEffectModels, Gadfly
df = dataset("datasets", "iris")
df[:SpeciesC] =  categorical(df[:Species])
result = partial_out(df, @model(SepalWidth + SepalLength ~ 1, fe = SpeciesC), add_mean = true)
plot(layer(result[1], x="SepalWidth", y="SepalLength", Stat.binmean(n=10), Geom.point),
   layer(result[1], x="SepalWidth", y="SepalLength", Geom.smooth(method=:lm)))

FixedEffectModels.reg — Method.

Estimate a linear model with high dimensional categorical variables / instrumental variables

Arguments

df::AbstractDataFrame

model::Model: A model created using @model

save::Union{Bool, Symbol} = false: Should residuals and eventual estimated fixed effects saved in a dataframe? Use save = :residuals to only save residuals. Use save = :fe to only save fixed effects.

method::Symbol = :lsmr: Method to deman regressors. :lsmr is akin to conjugate gradient descent. With parallel use :lsmr_parallel. To use multi threaded use lsmr_threads. Other choices are :qr and :cholesky (factorization methods)

contrasts::Dict = Dict() An optional Dict of contrast codings for each categorical variable in the formula. Any unspecified variables will have DummyCoding.

maxiter::Integer = 10000: Maximum number of iterations

tol::Real =1e-8: Tolerance

Details

Models with instruments variables are estimated using 2SLS. reg tests for weak instruments by computing the Kleibergen-Paap rk Wald F statistic, a generalization of the Cragg-Donald Wald F statistic for non i.i.d. errors. The statistic is similar to the one returned by the Stata command ivreg2.

Examples

using DataFrames, RDatasets, FixedEffectModels
df = dataset("plm", "Cigar")
df[:StateC] =  categorical(df[:State])
df[:YearC] =  categorical(df[:Year])
reg(df, @model(Sales ~ Price, fe = StateC + YearC))
reg(df, @model(Sales ~ NDI, fe = StateC + StateC&Year))
reg(df, @model(Sales ~ NDI, fe = StateC*Year))
reg(df, @model(Sales ~ (Price ~ Pimin)))
reg(df, @model(Sales ~ Price, weights = Pop))
reg(df, @model(Sales ~ NDI, subset = State .< 30))
reg(df, @model(Sales ~ NDI, vcov = robust))
reg(df, @model(Sales ~ NDI, vcov = cluster(StateC)))
reg(df, @model(Sales ~ NDI, vcov = cluster(StateC + YearC)))
reg(df, @model(Sales ~ YearC), contrasts = Dict(:YearC => DummyCoding(base = 80)))