Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FixedEffects.solve_coefficients! — Function.

Solve a least square problem for a set of FixedEffects

solve_coefficients!(y, fes, weights; method = :lsmr, maxiter = 10000, tol = 1e-8)

Arguments

y : A AbstractVector{Float64}

fes: A Vector{<:FixedEffect}

weights: A AbstractWeights

method : A Symbol for the method. Choices are :lsmr, :lsmrthreads, :lsmrparallel, :qr and :cholesky

maxiter : Maximum number of iterations

tol : Tolerance

Returns

b : Solution of the least square problem

iterations: Number of iterations

converged: Did the algorithm converge?

Examples

using  FixedEffects
p1 = repeat(1:5, inner = 2)
p2 = repeat(1:5, outer = 2)
x = rand(10)
solve_coefficients!(rand(10), [FixedEffect(p1), FixedEffect(p2)])

FixedEffects.solve_residuals! — Function.

Solve a least square problem for a set of FixedEffects

solve_residuals!(y, fes, weights; method = :lsmr, maxiter = 10000, tol = 1e-8)

Arguments

y : A AbstractVector{Float64} or a AbstractMatrix{Float64}

fes: A Vector{<:FixedEffect}

weights: A AbstractWeights

method : A Symbol for the method. Choices are :lsmr, :lsmrthreads, :lsmrparallel, :qr and :cholesky

maxiter : Maximum number of iterations

tol : Tolerance

Returns

res : Residual of the least square problem

iterations: Number of iterations

converged: Did the algorithm converge?

Examples

using  FixedEffects
p1 = repeat(1:5, inner = 2)
p2 = repeat(1:5, outer = 2)
solve_residuals!(rand(10), [FixedEffect(p1), FixedEffect(p2)])
solve_residuals!(rand(10, 5), [FixedEffect(p1), FixedEffect(p2)])