Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FixedPointAcceleration.execute_function_safely — Method.

execute_function_safely(Func::Function, x::Array{Float64,1})

This function creates a function that executes the function for which a fixed point is sought. It is a helper function that is not exported.

Takes

Func - The function input to fixed_point

x - The point at which to evaluate the function.

Returns

A FunctionEvaluationResult containing the following fields:

Input_ - The input

Output_ - The output of the FunctionExecutor. May be missing if function could not complete without error.

Error_ - A enum of type FP_FunctionEvaluationError representing what error occured.

Examples

Func(x) = sqrt(x) executefunctionsafely(Func, [-1.0,0.0,1.0]) executefunctionsafely(Func,[Missing(),0.0,1.0]) executefunctionsafely(Func,[7.0,0.0,1.0]) executefunctionsafely(Func,[NaN,0.0,1.0]) executefunctionsafely(Func,[Inf,0.0,1.0]) executefunctionsafely(Func,-1.0) executefunctionsafely(Func,Missing()) executefunctionsafely(Func,1.0) executefunctionsafely(Func,NaN) executefunctionsafely(Func,Inf)

FixedPointAcceleration.fixed_point — Method.

fixed_point(func::Function, previous_FixedPointResults::FixedPointResults;
            Algorithm::FixedPointAccelerationAlgorithm = Anderson,  ConvergenceMetric::Function  = supnorm(Resids::Array{Float64, 1}) = maximum(abs.(Resids)),
            ConvergenceMetricThreshold::Float64 = 1e-10, MaxIter::Int = 1000, MaxM::Int = 10, ExtrapolationPeriod::Int = 7, Dampening::Float64 = 1.0,
            PrintReports::Bool = false, ReportingSigFig::Int = 10, ReplaceInvalids::InvalidReplacement = NoAction, ConditionNumberThreshold::Float64 = 1e3)
fixed_point(func::Function, Inputs::Array{Float64, 1};
            Algorithm::FixedPointAccelerationAlgorithm = Anderson,  ConvergenceMetric::Function  = supnorm(Resids::Array{Float64, 1}) = maximum(abs.(Resids)),
            ConvergenceMetricThreshold::Float64 = 1e-10, MaxIter::Int = 1000, MaxM::Int = 10, ExtrapolationPeriod::Int = 7, Dampening::Float64 = 1.0,
            PrintReports::Bool = false, ReportingSigFig::Int = 10, ReplaceInvalids::InvalidReplacement = NoAction, ConditionNumberThreshold::Float64 = 1e3)
fixed_point(func::Function, Inputs::Float64;
            Algorithm::FixedPointAccelerationAlgorithm = Anderson,  ConvergenceMetric::Function  = supnorm(Resids::Array{Float64, 1}) = maximum(abs.(Resids)),
            ConvergenceMetricThreshold::Float64 = 1e-10, MaxIter::Int = 1000, MaxM::Int = 10, ExtrapolationPeriod::Int = 7, Dampening::Float64 = 1.0,
            PrintReports::Bool = false, ReportingSigFig::Int = 10, ReplaceInvalids::InvalidReplacement = NoAction, ConditionNumberThreshold::Float64 = 1e3)
fixed_point(func::Function, Inputs::Array{Float64, 2}; Outputs::Array{Float64,2} = Array{Float64,2}(undef,size(Inputs)[1],0),
            Algorithm::FixedPointAccelerationAlgorithm = Anderson,  ConvergenceMetric::Function  = supnorm(Resids::Array{Float64, 1}) = maximum(abs.(Resids)),
            ConvergenceMetricThreshold::Float64 = 1e-10, MaxIter::Int = 1000, MaxM::Int = 10, ExtrapolationPeriod::Int = 7, Dampening::Float64 = 1.0,
            PrintReports::Bool = false, ReportingSigFig::Int = 10, ReplaceInvalids::InvalidReplacement = NoAction, ConditionNumberThreshold::Float64 = 1e3)

A function for finding the fixed point of another function

Takes

f - This is the function for which a fixed point is sought. This function must take and return a vector of the same size dimension.

Inputs - This can be either a 1D-vector of values that is an initial guess for a fixed point or it can be an N x A matrix of previous inputs for which corresponding outputs are available. In this case N is the dimensionality of the fixed point vector you are seeking (Hence each column is a matrix that is input to f) and A is the number of previous Inputs/Outputs that are being provided to the fixed point. Where a matrix is input, a corresponding outputs must be provided or the last column of the outputs matrix is taken as a startpoint guess and the rest of the inputs and output matrices are discarded.

Outputs - This is a matrix of the Function values for each column of the input. It must be provided so that column k of the outputs matrix is equal to Function(Column k of inputs matrix).

Algorithm - This is the fixed point Algorithm to be used. It can be Anderson, Simple, Aitken, Newton, MPE, RRE, VEA or SEA. See vignette and references to see explanations of these Algorithms.

ConvergenceMetric This is a function that takes in a vector of inputs and a table of outputs and returns a scalar. This scalar should be low when convergence is close to being achieved. By default this is the maximum residual by absolute value (the sup norm in the space of residuals).

ConvergenceMetricThreshold This is the threshold for terminating the algorithm. The algorithm will terminate when the scalar that ConvergenceMetric returns is less than ConvergenceMetricThreshold. This can be set to a negative number in which case the algorithm will run until MaxIter is hit or an error occurs (Note that an error is likely in trying to use any Algorithm other than "Simple" when a fixed point is already found).

MaxIter - This is the maximum number of iterates that will be undertaken.

MaxM - This is the maximum number of saved iterates that are used in the Anderson algorithm. It has no effect if another Algorithm is chosen. Note that the number of previous iterates that will actually be used is the minimum of MaxIter, the dimensionality of the f's vector and the number of inputs that have been tried to previously (the width of the Outputs matrix at each given stage of the algorithm). If PrintReports = TRUE, the number of previous iterates actually used is reported as the algorithm is running.

ExtrapolationPeriod - This is the number of simple iterates to perform before extrapolating. This is used for the MPE, RRE, VEA and SEA Algorithms and has no effect if another Algorithm is chosen Where an epsilon algorithm is used this should be a multiple of 2, ie (4,6,8,etc).

Dampening - This is the dampening parameter. By default it is 1 which means no dampening takes place. It can also be less than 1 (indicating dampening) or more than 1 (indicating extrapolation).

PrintReports - This is a boolean describing whether to print ongoing ConvergenceMetric values for each iterate.

ReportingSigFig - This is the number of significant figures that will be used in printing the convergence values to the console (only if PrintReports is TRUE).

ReplaceInvalids -Sometimes an acceleration algorithm proposed a vector with an invalid coordinate (NaN, Inf or missing). This parameter can be set to ReplaceInvalids (to replace invalid coordinates by the simple iterate values), ReplaceVector (to replace entire vector with a simple iterate) or NoAction (where an imminent error will occur).

ConditionNumberThreshold - This is a threshold for what condition number is acceptable for solving the least squares problem for the Anderson Algorithm. If the condition number is larger than this threshold then fewer previous iterates will be used in solving the problem. This has no effect unless the "Anderson" Algorithm is used.

Returns

A list containing the fixed_point, the Inputs and corresponding Outputs, and convergence values (which are computed under the "ConvergenceMetric"). The list will also include a "Finish" statement describing why it has finished. This is often going to be due to either MaxIter or ConvergenceMetricThreshold being reached. It may also terminate due to an error in generating a new input guess or using the function with that guess. If this occurs the function will terminate early and the "Finish" statement will describe the issue. In this event there will also be additional objects returned in the list "NewInputVector" and possibly "NewOutputVector" that are useful in debugging the issue.

Examples

#' # For the simplest possible example we can seek the fixed point of the cos function with a scalar. #' Inputs = 0.3 #' Func(x) = cos(x) #' A = fixedpoint(Func, Inputs; Algorithm = Aitken, Dampening = 0.5) #' B = fixedpoint(Func, Inputs; Algorithm = Anderson, Dampening = 1.0) #' #' # For this next one the ConvergenceMetricThreshold is negative so the algorithm #' # will keep running until MaxIter is met. #' C = fixedpoint(Func, Inputs; Algorithm = Simple, MaxIter = 4, ConvergenceMetricThreshold = -1.0) #' # But we can continue solving for this fixed point but now switching to the Newton Algorithm. #' D = fixedpoint(Func, C[:Inputs], C[:Outputs]; Algorithm = Newton) #' #' # We can also find a 4 dimensional fixed point vector of this function. #' Inputs = [0.3, 98, 0, pi] #' E = fixedpoint(Func, Inputs; Algorithm = Anderson) #' F = fixedpoint(Func, Inputs; Algorithm = Anderson, MaxM = 4, ReportingSigFig = 13)

FixedPointAcceleration.fixed_point_new_input — Function.

This function takes the previous inputs and outputs from the fixed_point function and determines what vector to try next in seeking a fixed point.

Takes

Inputs - This is an N x A matrix of previous inputs for which corresponding outputs are available. In this case N is the dimensionality of the fixed point vector that is being sought (Hence each column is a matrix that is input to the "Function") and A is the number of previous Inputs/Outputs that are being provided to the fixed point.

Outputs - This is a matrix of Function values for the each column of the "Inputs" matrix.

Algorithm - This is the fixed point Algorithm to be used. It can be "Anderson", "Simple", "Aitken", "Newton", "MPE", "RRE", "VEA", "SEA".

MaxM - This is the number of saved iterates that are used in the Anderson algorithm. This has no role if another Algorithm is used.

SimpleStartIndex - This is the index for what column in the input/output matrices did the algorithm start doing simple iterates without jumps. This is used for all Algorithms except the simple and Anderson Algorithms where it has no effect.

ExtrapolationPeriod - This is the number of simple iterates to perform before extrapolating. This is used for the MPE, RRE, VEA and SEA Algorithms and has no effect if another Algorithm is chosen.

Dampening - This is the dampening parameter. By default it is 1 which means no dampening takes place. It can also be less than 1 (indicating dampening) or more than 1 (indicating extrapolation).

ConditionNumberThreshold - This is what threshold should be chosen to drop previous iterates if the matrix is ill conditioned. Only used in Anderson acceleration.

PrintReports - This is a boolean describing whether to print ongoing ConvergenceMetric values for each iterate.

Returns

A nicely formatted string version of the input number for printing to the console.

Examples

FPFunction = function(x){c(0.5sqrt(abs(x[1] + x[2])), 1.5x[1] + 0.5*x[2])} A = fixedpoint( Function = FPFunction, Inputs = [0.3,900], MaxIter = 6, Algorithm = Simple) NewGuessAnderson = fixedpointnewinput(A[:Inputs], A[:Outputs], Algorithm = Anderson) NewGuessVEA = fixedpointnewinput(A[:Inputs], A[:Outputs], Algorithm = VEA) NewGuessMPE = fixedpointnewinput(A[:Inputs], A[:Outputs], Algorithm = MPE) NewGuessAitken = fixedpointnew_input(A[:Inputs], A[:Outputs], Algorithm = Aitken)

FixedPointAcceleration.put_together_without_jumps — Function.

This function takes the previous inputs and outputs and assembles a matrix with both excluding jumps.

Takes

Inputs - This is an N x A matrix of previous inputs for which corresponding outputs are available. In this case N is the dimensionality of the fixed point vector that is being sought (and each column is a matrix that is input to the "Function") and A is the number of previous Inputs/Outputs that are being provided to the fixed point.

Outputs This is a matrix of "Function" values for each column of the "Inputs" matrix.

AgreementThreshold A parameter for determining when a column in Inputs and a column in Outputs match. They are deemed to match if the sum of the absolute values of the difference in the columns is less than AgreementThreshold.

Returns

A matrix of inputs and outputs excluding jumps.

FixedPointAcceleration.EpsilonExtrapolation — Method.

This is a helper function for EpsilonExtrapolation

Takes

Iterates - A matrix representing different iterates with one iterate per column. Can be pieced together from Inputs and Outputs matrices of the fixedpoint function using the puttogetherwithoutjumps function

Algorithm - Algorithm for epsilon extrapolation. Should be either "VEA" for the vector extrapolation algorithm or "SEA" for the scalar epsilon algorithm.

Returns

A vector with the extrapolated vector.

FixedPointAcceleration.EpsilonExtrapolationVectorOfInverses — Method.

This is a helper function for EpsilonExtrapolation

Takes

DifferenceMatrix - The matrix of the differences in elements to be inverted.

Algorithm - SEA or VEA.

Returns

A vector of the result of inverting each (column) vector in a mmatrix.

FixedPointAcceleration.PolynomialExtrapolation — Method.

This function performs Minimal Polynomial extrapolation (MPE) or Reduced Rank Extrapolation (RRE) given a matrix of previous iterates of the function.

Takes

Iterates - A matrix of inputs and outputs excluding jumps. Can be pieced together from Inputs and Outputs matrices of the fixedpoint function using the puttogetherwithoutjumps function

Algorithm The Algorithm for polynomial extrapolation. Should be either "MPE" for minimal polynomial extrapolation or "RRE" for reduced rank extrapolation.

Returns

A vector containing the extrapolated vector.