Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Base.isapprox — Method.

isapprox(x::FixedPoint, y::FixedPoint; rtol=0, atol=max(eps(x), eps(y)))

For FixedPoint numbers, the default criterion is that x and y differ by no more than eps, the separation between adjacent fixed-point numbers.