Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FlashWeave.graph — Method.

graph(result::FWResult{T}) -> SimpleWeightedGraph{Int, Float64}

Extract the underlying weighted graph from network results.

FlashWeave.learn_network — Function.

learn_network(data_path::AbstractString, meta_data_path::AbstractString) -> FWResult{Int}

Works like learn_network(data::AbstractArray{ElType}), but takes file paths an OTU table and optionally a meta data table as an input (instead of a data matrix).

data_path - path to a file storing an OTU count table (and in the case of JLD2 possibly meta data)

meta_data_path - optional path to a file with meta data

*_key - HDF5 keys to access data sets with OTU counts, Meta variables and variable names in a JLD2 file, if a data item is absent the corresponding key should be 'nothing'. More help under '?load_data'

verbose - print progress information

transposed - if true, rows of data are variables and columns are samples

kwargs... - additional keyword arguments passed to learn_network(data::AbstractArray{ElType})

FlashWeave.learn_network — Method.

learn_network(data::AbstractArray{<:Real}) -> FWResult{Int}

Learn an interaction network from a data table (including OTUs and optionally meta variables).

data - data table with information on OTU counts and (optionally) meta variables

header - names of variable columns in data

meta_mask - true/false mask indicating which variables are meta variables

Algorithmic parameters:

heterogeneous - enable heterogeneous mode for multi-habitat or -protocol data with at least thousands of samples (FlashWeaveHE)

sensitive - enable fine-grained associations (FlashWeave-S, FlashWeaveHE-S), sensitive=false results in the fast modes FlashWeave-F or FlashWeaveHE-F

max_k - maximum size of conditioning sets, high values can strongly increase runtime. max_k=0 results in no conditioning (univariate mode)

alpha - threshold used to determine statistical significance

conv - convergence threshold, i.e. if conv=0.01 assume convergence if the number of edges increased by only 1% after 100% more runtime (checked in intervals)

feed_forward - enable feed-forward heuristic

max_tests - maximum number of conditional tests that should be performed on a variable pair before association is assumed

hps - reliability criterion for statistical tests when sensitive=false

FDR - perform False Discovery Rate correction (Benjamini-Hochberg method) on pairwise associations

n_obs_min - don't compute associations between variables having less reliable samples (i.e. non-zero if heterogeneous=true) than this number. -1: automatically choose a threshold.

time_limit - if feed-forward heuristic is active, determines the interval (seconds) at which neighborhood information is updated

General parameters:

normalize - automatically choose and perform data normalization (based on sensitive and heterogeneous)

track_rejections - store for each discarded edge, which variable set lead to its exclusion (can be memory intense for large networks)

verbose - print progress information

transposed - if true, rows of data are variables and columns are samples

prec - precision in bits to use for calculations (16, 32, 64 or 128)

make_sparse - use a sparse data representation (should be left at true in almost all cases)

update_interval - if verbose=true, determines the interval (seconds) at which network stat updates are printed

FlashWeave.load_data — Function.

load_data(data_path::AbstractString, meta_path::AbstractString) -> (AbstractMatrix{<:Real}, Vector{String}, AbstractMatrix{<:Real}, Vector{String})

Load tables with OTU count and optionally meta data from disc. Available formats are '.tsv', '.csv', '.biom' and '.jld2'.

data_path - path to a file storing an OTU count table

meta_data_path - optional path to a file with meta variable information

*_key - HDF5 keys to access data sets with OTU counts, Meta variables and variable names in a JLD2 file, if a data item is absent the corresponding key should be 'nothing'

transposed - if true, rows of data are variables and columns are samples

FlashWeave.load_network — Method.

load_network(net_path::AbstractString) -> FWResult{Int}

Load network results from disk. Available formats are '.tsv', '.csv', '.gml' and '.jld2'. For GML, only files with structure identical to save_network('network.gml') output can currently be loaded. Run parameters are only available when loading from JLD2.

net_path - path from which to load the network results

FlashWeave.normalize_data — Method.

normalize_data(data::AbstractMatrix{<:Real}) -> AbstractMatrix OR (AbstractMatrix{<:Real}, Vector{String})

Normalize data using various forms of clr transform and discretization. This should only be used manually when experimenting with different normalization techniques.

data - data table with information on OTU counts and (optionally) meta variables

header - names of variable-column s in data

meta_mask - true/false mask indicating which variables are meta variables

test_name - name of a FlashWeave-specific statistical test mode, the appropriate normalization method will be chosen automatically

norm_mode - identifier of a valid normalization mode ('clr-adapt', 'clr-nonzero', 'clr-nonzero-binned', 'pres-abs', 'tss', 'tss-nonzero-binned')

filter_data - whether to remove samples and variables without information from data

verbose - print progress information

prec - precision in bits to use for calculations (16, 32, 64 or 128)

FlashWeave.save_network — Method.

save_network(net_path::AbstractString, net_result::FWResult) -> Void

Save network results to disk. Available formats are '.tsv', '.csv', '.gml' and '.jld2'.

net_path - output path for the network

net_result - network results object that should be saved

detailed - output additional information, such as discarding sets, if available

FlashWeave.StackChannel — Type.

StackChannel{T}(sz::Int)

Constructs a Channel with an internal buffer that can hold a maximum of sz objects of type T. put! calls on a full StackChannel block until an object is removed with take!. Other constructors:

Channel(Inf): equivalent to Channel{Any}(typemax(Int))

Channel(sz): equivalent to Channel{Any}(sz)

FlashWeave.StackChannel — Method.

StackChannel(func::Function; ctype=Any, csize=0, taskref=nothing)

Create a new task from func, bind it to a new StackChannel of type ctype and size csize, and schedule the task, all in a single call. func must accept the bound StackChannel as its only argument. If you need a reference to the created task, pass a Ref{Task} object via keyword argument taskref. Return a Channel.

Examples

julia> chnl = StackChannel(c->foreach(i->put!(c,i), 1:4));
julia> typeof(chnl)
Channel{Any}
julia> for i in chnl
           @show i
       end;
i = 1
i = 2
i = 3
i = 4

Referencing the created task:

julia> taskref = Ref{Task}();
julia> chnl = StackChannel(c->(@show take!(c)); taskref=taskref);
julia> istaskdone(taskref[])
false
julia> put!(chnl, "Hello");
take!(c) = "Hello"
julia> istaskdone(taskref[])
true

Base.close — Method.

close(c::StackChannel)

Close a StackChannel. An exception is thrown by:

put! on a closed StackChannel.

take! and fetch on an empty, closed StackChannel.

Base.fetch — Method.

fetch(c::StackChannel)

Wait for and get the first available item from the StackChannel. Does not remove the item.

Base.names — Method.

variable_ids(result::FWResult{T}) -> Vector{T}

Extract the IDs/names of all variables (nodes) in the network.

Base.put! — Method.

put!(c::StackChannel, v)

Append an item v to the StackChannel c. Blocks if the StackChannel is full.

Base.take! — Method.

take!(c::StackChannel)

Remove and return a value from a Channel. Blocks until data is available.

FlashWeave.benjamini_hochberg! — Method.

Accelerated version of that found in MultipleTesting.jl

FlashWeave.bind — Method.

bind(chnl::StackChannel, task::Task)

Associate the lifetime of chnl with a task. Channel chnl is automatically closed when the task terminates. Any uncaught exception in the task is propagated to all waiters on chnl. The chnl object can be explicitly closed independent of task termination. Terminating tasks have no effect on already closed Channel objects. When a StackChannel is bound to multiple tasks, the first task to terminate will close the StackChannel. When multiple StackChannels are bound to the same task, termination of the task will close all of the bound StackChannels.

Examples

julia> c = StackChannel(0);
julia> task = @async foreach(i->put!(c, i), 1:4);
julia> bind(c,task);
julia> for i in c
           @show i
       end;
i = 1
i = 2
i = 3
i = 4
julia> isopen(c)
false

julia> c = StackChannel(0);
julia> task = @async (put!(c,1);error("foo"));
julia> bind(c,task);
julia> take!(c)
1
julia> put!(c,1);
ERROR: foo
Stacktrace:
[...]

FlashWeave.channeled_tasks — Method.

channeled_tasks(n::Int, funcs...; ctypes=fill(Any,n), csizes=fill(0,n))

A convenience method to create n StackChannels and bind them to tasks started from the provided functions in a single call. Each func must accept n arguments which are the created StackChannels. StackChannel types and sizes may be specified via keyword arguments ctypes and csizes respectively. If unspecified, all StackChannels are of type Channel{Any}(0). Returns a tuple, (Array{StackChannel}, Array{Task}), of the created StackChannels and tasks.

FlashWeave.isready — Method.

isready(c::StackChannel)

Determine whether a Channel has a value stored to it. Returns immediately, does not block.

FlashWeave.mutual_information — Method.

IMPORTANT NOTE: returns mutual information * number of observations! (avoids repeated calculation later)

FlashWeave.parameters — Method.

parameters(result::FWResult{T}) -> Dict{Symbol, Any}

Extract the used parameters from network results.