Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Flatten.flatten — Method.

Flattening. Flattens a nested type to a Tuple or Vector

Flatten.metaflatten — Method.

Tag flattening. Flattens data attached to a field by methods of a passed in function

Flatten.reconstruct — Method.

Reconstruct an object from partial Tuple or Vector data and another object

Flatten.retype — Method.

Retype an object from partial Tuple or Vector data and another object

Flatten.update! — Method.

Update a mutable object with partial Tuple or Vector data