Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FlexLinearAlgebra.FlexMatrix — Method.

FlexMatrix{T}(rows,cols) creates a new FlexMatrix with rows indexed by rows, columns indexed by cols and all zero entries of type T (which is Number if omitted).

FlexMatrix(v::FlexVector) converts v into a one-column FlexMatrix whose sole column index is Int(1)

FlexLinearAlgebra.FlexVector — Method.

FlexVector{T}(idx) creates a new FlexVector with entries indexed by idx filled with zeros of type T (which defaults to Number)

FlexLinearAlgebra.FlexConvert — Method.

FlexConvert(vec) converts the vector vec into a FlexVector.

FlexLinearAlgebra.FlexOnes — Method.

FlexOnes(T,dom) creates an all 1s vector indexed by dom. If T is missing, values default to Float64.

FlexLinearAlgebra.col_keys — Method.

col_keys(M::FlexMatrix) returns a list of the keys to the columns of M.

FlexLinearAlgebra.delete_col! — Method.

delete_col!(A,c) deletes row c from the FlexMatrix A.

FlexLinearAlgebra.delete_entry! — Method.

delete_entry!(v,x) deletes the entry indexed by x in the FlexVector x.

FlexLinearAlgebra.delete_row! — Method.

delete_row!(A,r) deletes row r from the FlexMatrix A.

FlexLinearAlgebra.row_keys — Method.

row_keys(M::FlexMatrix) returns a list of the keys to the rows of M.