Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FormattedTables.write — Function.

FormattedTables.write(
    file::Union{IO,String},
    table;
    delim = ' ',
    column_names=Tables.schema(table).names,
    formatters=Dict(name=>"" for name in column_names),
    header_fmt=Dict(name=>"" for name in column_names),
)

Write a Tables.jl interface input to a stream or a named file using the format specifiers provided via the formatters keyword argument. The formatters argument must be a dictionary, named tuple, or similar collection that can be indexed using the column names of the table.

Keyword arguments include:

delim: A delimiter to print between list items.

column_names: An iterable collection containing, in order, the names (as Symbols) of the columns of the table to be written.

formatters: A dictionary, named tuple, or similar collection that maps the the names (Symbols) of the columns to be written to a Format.FormatSpec used to format the column.

header_fmt: A dictionary, named tuple, or similar collection that maps the the names (Symbols) of the columns to be written to a Format.FormatSpec used to format the column headers. Set to nothing to skip writing column headers.