Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ForneyLab.Addition — Type.

Description:

An addition constraint factor node

f(out,in1,in2) = δ(in1 + in2 - out)

Interfaces:

1. out
2. in1
3. in2

Construction:

Addition(out, in1, in2, id=:some_id)

ForneyLab.Bernoulli — Type.

Description: Bernoulli factor node

out ∈ {0, 1}
p ∈ [0, 1]

f(out, p) = Ber(out|p) = p^out (1 - p)^{1 - out}

Interfaces: 1. out 2. p

Construction: Bernoulli(id=:some_id)

ForneyLab.Beta — Type.

Description: Beta factor node

Real scalars
a > 0
b > 0

f(out, a, b) = Beta(out|a, b) = Γ(a + b)/(Γ(a) Γ(b)) out^{a - 1} (1 - out)^{b - 1}

Interfaces: 1. out 2. a 3. b

Construction: Beta(id=:some_id)

ForneyLab.Categorical — Type.

Description: Categorical factor node

The categorical node defines a one-dimensional probability
distribution over the normal basis vectors of dimension d

out ∈ {0, 1}^d where Σ_k out_k = 1
p ∈ [0, 1]^d, where Σ_k p_k = 1

f(out, p) = Cat(out | p)
          = Π_i p_i^{out_i}

Interfaces: 1. out 2. p

Construction: Categorical(id=:some_id)

ForneyLab.Clamp — Type.

Description:

A factor that clamps a variable to a constant value.

f(out) = δ(out - value)

Interfaces:

1. out

Construction:

Clamp(out, value, id=:some_id)
Clamp(value, id=:some_id)

ForneyLab.Contingency — Type.

Description: Contingency factor node

The contingency distribution is a multivariate generalization of
the categorical distribution. As a bivariate distribution, the
contingency distribution defines the joint probability
over two unit vectors. The parameter p encodes a contingency matrix
that specifies the probability of co-occurrence.

out1 ∈ {0, 1}^d1 where Σ_j out1_j = 1
out2 ∈ {0, 1}^d2 where Σ_k out2_k = 1
p ∈ [0, 1]^{d1 × d2}, where Σ_jk p_jk = 1

f(out1, out2, p) = Con(out1, out2 | p)
                 = Π_jk p_jk^{out1_j * out2_k}

A Contingency distribution over more than two variables requires
higher-order tensors as parameters; these are not implemented in ForneyLab.

Interfaces: 1. out1 2. out2 3. p

Construction: Contingency(id=:some_id)

ForneyLab.Dirichlet — Type.

Description: Dirichlet factor node

Real vector
a .> 0

Multivariate:
f(out, a) = Dir(out|a)
          = Γ(Σ_i a_i)/(Π_i Γ(a_i)) Π_i out_i^{a_i}

Matrix variate with independent rows:
f(out, a) = Π_j Dir(out|a_j.)
          = Π_j Γ(Σ_k a_jk)/(Π_k Γ(a_jk)) Π_k out_jk^{a_jk}

Interfaces: 1. out 2. a

Construction: Dirichlet(id=:some_id)

ForneyLab.DotProduct — Type.

Description:

out = in1'*in2

in1: d-dimensional vector
in2: d-dimensional vector
out: scalar

       in2
       |
  in1  V   out
----->[⋅]----->

f(out, in1, in2) =  δ(out - in1'*in2)

Interfaces:

1 i[:out], 2 i[:in1], 3 i[:in2]

Construction:

DotProduct(out, in1, in2, id=:my_node)

ForneyLab.Edge — Type.

An Edge joins two interfaces (half-edges) a and b.

ForneyLab.Equality — Type.

Description:

An equality constraint factor node

f([1],[2],[3]) = δ([1] - [2]) δ([1] - [3])

Interfaces:

1, 2, 3

Construction:

Equality(id=:some_id)

The interfaces of an Equality node have to be connected manually.

ForneyLab.Exponential — Type.

Description:

Maps a location to a scale parameter by exponentiation

f(out,in1) = δ(out - exp(in1))

Interfaces:

1. out
2. in1

Construction:

Exponential(out, in1, id=:some_id)

ForneyLab.FactorGraph — Type.

A factor graph consisting of factor nodes and edges.

ForneyLab.Gamma — Type.

Description:

A gamma node with shape-rate parameterization:

f(out,a,b) = Gam(out|a,b) = 1/Γ(a) b^a out^{a - 1} exp(-b out)

Interfaces:

1. out
2. a (shape)
3. b (rate)

Construction:

Gamma(out, a, b, id=:some_id)

ForneyLab.GaussianMeanPrecision — Type.

Description:

A Gaussian with mean-precision parameterization:

f(out,m,w) = 𝒩(out|m,w) = (2π)^{-D/2} |w|^{1/2} exp(-1/2 (out - m)' w (out - m))

Interfaces:

1. out
2. m (mean)
3. w (precision)

Construction:

GaussianMeanPrecision(out, m, w, id=:some_id)

ForneyLab.GaussianMeanVariance — Type.

Description:

A Gaussian with mean-variance parameterization:

f(out,m,v) = 𝒩(out|m,v) = (2π)^{-D/2} |v|^{-1/2} exp(-1/2 (out - m)' v^{-1} (out - m))

Interfaces:

1. out
2. m (mean)
3. v (covariance)

Construction:

GaussianMeanVariance(out, m, v, id=:some_id)

ForneyLab.GaussianMixture — Type.

Description:

A Gaussian mixture with mean-precision parameterization:

f(out, z, m1, w1, m2, w2, ...) = 𝒩(out|m1, w1)^z_1 * 𝒩(out|m2, w2)^z_2 * ...

Interfaces:

1. out
2. z (switch)
3. m1 (mean)
4. w1 (precision)
5. m2 (mean)
6. w2 (precision)
...

Construction:

GaussianMixture(out, z, m1, w1, m2, w2, ..., id=:some_id)

ForneyLab.Interface — Type.

An Interface belongs to a FactorNode and represents a half-edge. An Interface has at most one partner interface, with wich it forms an edge.

ForneyLab.LogNormal — Type.

Description:

A log-normal node with location-scale parameterization:

f(out,m,s) = logN(out|m, s) = 1/out (2π s)^{-1/2} exp(-1/(2s) (log(out) - m)^2))

Interfaces:

1. out
2. m (location)
3. s (squared scale)

Construction:

LogNormal(out, m, s, id=:some_id)

ForneyLab.MarginalRule — Type.

MarginalRule{factortype} specifies a joint marginal update rule with respect to a node of type `factortype`.

ForneyLab.MarginalScheduleEntry — Type.

A MarginalScheduleEntry defines a marginal computation. The marginal_update_rule <: MarginalUpdateRule defines the rule that is used to calculate the (joint) marginal over target.

ForneyLab.MarginalScheduleEntry — Method.

Construct a MarginalScheduleEntry for computing the marginal over cluster through a node-specific joint marginal update rule.

ForneyLab.MarginalScheduleEntry — Method.

Construct a MarginalScheduleEntry for computing the marginal over variable through multiplication of colliding messages.

ForneyLab.MarginalUpdateRule — Type.

A MarginalUpdateRule specifies how a (joint) marginal is calculated from incoming messages (and a node function).

ForneyLab.Message — Type.

Encodes a message, which is a probability distribution with a scaling factor

ForneyLab.MessageUpdateRule — Type.

A MessageUpdateRule specifies how a Message is calculated from the node function and the incoming messages. Use subtypes(MessageUpdateRule) to list the available rules.

ForneyLab.Multiplication — Type.

Description:

For continuous random variables, the multiplication node acts
as a (matrix) multiplication constraint, with node function

f(out, in1, a) = δ(out - a*in1)

Interfaces:

1. out
2. in1
3. a

Construction:

Multiplication(out, in1, a, id=:some_id)

ForneyLab.Nonlinear — Type.

Description:

Nonlinear node modeling a nonlinear relation. Updates for
the nonlinear node are computed through local linearization.

f(out, in1) = δ(out - g(in1))

Interfaces:

1. out
2. in1

Construction:

Nonlinear(out, in1, g::Function, J_g::Function, id=:my_node)

ForneyLab.PointMass — Type.

PointMass is an abstract type used to describe point mass distributions. It never occurs in a FactorGraph, but it is used as a probability distribution type.

ForneyLab.ProbabilityDistribution — Type.

Encodes a probability distribution as a FactorNode of type family with fixed interfaces

ForneyLab.RecognitionFactor — Type.

A RecognitionFactor specifies the subset of variables that comprise a joint factor in the recognition factorization.

ForneyLab.RecognitionFactorization — Type.

A RecognitionFactorization holds a collection of (non-overlapping) recognition factors that specify the recognition factorization over a factor graph that is used for variational inference.

ForneyLab.RecognitionFactorization — Method.

Construct a RecognitionFactorization consisting of one RecognitionFactor for each argument

ForneyLab.ScheduleEntry — Type.

A ScheduleEntry defines a message computation. The msg_update_rule <: MessageUpdateRule defines the rule that is used to calculate the message coming out of interface.

ForneyLab.Sigmoid — Type.

Description: Constrains a continuous, real-valued variable with a binary (boolean) variable.

f(bin, real) = σ(bin⋅real)

Interfaces: 1. bin 2. real

Construction: Sigmoid(id=:some_id)

ForneyLab.Terminal — Type.

Description:

Terminal is a special type of node that is only used in the internal
graph of a CompositeNode. A Terminal is used to terminate an Edge in the
internal graph that is linked to an interface of the CompositeNode.

A Terminal is linked to an interface of the
CompositeNode containing the Terminal.

Interfaces:

1. out

Construction:

Terminal(id=:some_id)

ForneyLab.Transition — Type.

Description:

The transition node models a transition between discrete
random variables, with node function

f(out, in1, a) = Cat(out | a*in1)

Interfaces:

1. out
2. in1
3. a

Construction:

Transition(out, in1, a, id=:some_id)

ForneyLab.Variable — Type.

A Variable encompasses one or more edges in a FactorGraph.

ForneyLab.Wishart — Type.

Description:

A Wishart node:

f(out,v,nu) = W(out|v, nu) = B(v, nu) |out|^{(nu - D - 1)/2} exp(-1/2 tr(v^{-1} out))

Interfaces:

1. out
2. v (scale matrix)
3. nu (degrees of freedom)

Construction:

Wishart(out, v, nu, id=:some_id)

ForneyLab.conditionalDifferentialEntropy — Method.

Compute conditional differential entropy: H(Y|X) = H(Y, X) - H(X)

ForneyLab.constant — Method.

constant creates a Variable which is linked to a new Clamp, and returns this variable.

y = constant(3.0, id=:y)

ForneyLab.currentGraph — Method.

Return currently active FactorGraph. Create one if there is none.

ForneyLab.currentRecognitionFactorization — Method.

Return currently active RecognitionFactorization. Create one if there is none.

ForneyLab.edges — Method.

Collect all edges corresponding with variable(s)

ForneyLab.expectationPropagationAlgorithm — Method.

Create a sum-product algorithm to infer marginals over variables, and compile it to Julia code

ForneyLab.expectationPropagationSchedule — Method.

expectationPropagationSchedule() generates a expectation propagation message passing schedule.

ForneyLab.handle — Method.

Return interface handle name

ForneyLab.marginalSchedule — Method.

marginalSchedule() generates a marginal schedule that computes the marginals for each target entry

ForneyLab.mat — Method.

Helper function to construct 1x1 Matrix

ForneyLab.placeholder — Method.

placeholder(...) creates a Clamp node and registers this node as a data placeholder with the current graph.

# Link variable y to buffer with id :y,
# indicate that Clamp will hold Float64 values.
placeholder(y, :y, datatype=Float64)

# Link variable y to index 3 of buffer with id :y.
# Specify the data type by passing a default value for the Clamp.
placeholder(y, :y, index=3, default=0.0)

# Indicate that the Clamp will hold an array of size `dims`,
# with Float64 elements.
placeholder(X, :X, datatype=Float64, dims=(3,2))

ForneyLab.sumProductAlgorithm — Method.

Create a sum-product algorithm to infer marginals over variables, and compile it to Julia code

ForneyLab.sumProductSchedule — Method.

sumProductSchedule() generates a sum-product message passing schedule that computes the marginals for each of the argument variables.

ForneyLab.ultimatePartner — Method.

ultimatePartner(interface) finds the 'ultimate partner' of interface. If interface.partner does not belong to a Terminal, it simply returns interface.partner. In case of a Terminal node, it finds the first non-Terminal partner on a higher level factor graph.

ForneyLab.variationalAlgorithm — Method.

Create a variational algorithm to infer marginals over a recognition distribution, and compile it to Julia code

ForneyLab.variationalExpectationPropagationAlgorithm — Method.

Create a variational EP algorithm to infer marginals over a recognition distribution, and compile it to Julia code

ForneyLab.variationalExpectationPropagationSchedule — Method.

variationalExpectationPropagationSchedule() generates an expectation propagation message passing schedule that is limited to the recognition_factor. Updates on EP sites are computed with an ExpectationPropagationRule.

ForneyLab.variationalSchedule — Method.

variationalSchedule() generates a variational message passing schedule that computes the marginals for each of the recognition distributions in the recognition factor.

ForneyLab.Cluster — Type.

A Cluster specifies a collection of edges adjacent to node that belong to the same RecognitionFactor. A joint marginal can be computed over a cluster.

ForneyLab.DependencyGraph — Type.

A DependencyGraph is a directed graph in which an edge v -> w represents a dependency of vertex v on vertex w. Dependency graphs are used for example by message scheduling algorithms.

ForneyLab.addNode! — Method.

Add a FactorNode to a FactorGraph

ForneyLab.addVariable! — Method.

Add a Variable to a FactorGraph

ForneyLab.approximate — Method.

Find a local linear approximation to the nonlinear vector function g at x_hat

ForneyLab.associate! — Method.

associate!(interface, variable) associates interface with variable by connecting interface to an Edge belonging to variable.

ForneyLab.breakerTypes — Method.

Constructs breaker types dictionary for breaker sites

ForneyLab.clusterId — Method.

Return the id of the cluster/variable that the node-edge combination belongs to

ForneyLab.collectEPSites — Method.

Find default EP sites present in node_set

ForneyLab.collectInboundTypes — Method.

Find the inbound types that are required to compute a joint marginal over target. Returns a vector with inbound types that correspond with required interfaces.

ForneyLab.collectInboundTypes — Method.

Find the inbound types that are required to compute the message for entry. Returns a vector with inbound types that correspond with required interfaces.

ForneyLab.collectInboundTypes — Method.

Find the inbound types that are required to compute the message for entry. Returns a vector with inbound types that correspond with required interfaces.

ForneyLab.collectInbounds — Method.

Construct the inbound code that computes the marginal for entry. Allows for overloading and for a user the define custom node-specific inbounds collection. Returns a vector with inbounds that correspond with required interfaces.

ForneyLab.collectInbounds — Method.

Find the inbound types that are required to compute the message for entry. Returns a vector with inbound types that correspond with required interfaces.

ForneyLab.collectInbounds — Method.

Construct argument code for naive VB updates

ForneyLab.collectInbounds — Method.

Construct argument code for structured VB updates

ForneyLab.collectInbounds — Method.

Collect and construct SP update code for each inbound.

ForneyLab.collectNaiveVariationalNodeInbounds — Method.

Construct the inbound code that computes the message for entry. Allows for overloading and for a user the define custom node-specific inbounds collection. Returns a vector with inbounds that correspond with required interfaces.

ForneyLab.collectStructuredVariationalNodeInbounds — Method.

Construct the inbound code that computes the message for entry. Allows for overloading and for a user the define custom node-specific inbounds collection. Returns a vector with inbounds that correspond with required interfaces.

ForneyLab.collectSumProductNodeInbounds — Method.

Construct the inbound code that computes the message for entry. Allows for overloading and for a user the define custom node-specific inbounds collection. Returns a vector with inbounds that correspond with required interfaces.

ForneyLab.condense — Method.

Contruct a condensed schedule.

ForneyLab.connect! — Method.

Connect loose end of edge to interface b.

ForneyLab.disconnect! — Method.

Disconnect edge from interface.

ForneyLab.ensureMatrix — Method.

ensureMatrix: cast input to a Matrix if necessary

ForneyLab.extend — Method.

Find the smallest legal subgraph that includes the argument edges. Default setting terminates the search at soft factors and does not constrain the search to a limiting set (as specified by an empty limit_set argument).

ForneyLab.find_vertex_indexes — Method.

children(vertices, graph; allowcycles=false, breakersites=[], restrict_to=[])

Return a vector consisting of vertices and all their children in graph. v is a child of u iff there exists a path from u to v. The resulting array is sorted in reverse topological order, i.e. for each directed edge u -> v, v (child of u) appears before u.

Optional keyword arguments:

allow_cycles: set to true to accept cycles.

breaker_sites: a Set of vertices on which the search will terminate.

restrict_to: a Set of vertices to restrict the search to.

This function can be used to generate message passing schedules if graph is a dependency graph.

ForneyLab.flatten — Method.

Flatten a schedule by inlining all internal message passing schedules. This yields a simple, linear schedule.

ForneyLab.generateId — Method.

Automatically generate a unique id based on the current counter value for the element type.

ForneyLab.hasCollider — Method.

Return whether the subgraph contains a collider. If a collider is found, this will lead to conditional dependencies in the recognition distribution (posterior).

ForneyLab.hasCollider — Method.

Return whether connectedcluster contains a collider. This function assumes the graph for connectedcluster is a connected tree.

ForneyLab.hasNode — Method.

hasNode(graph, node) checks if node is part of graph.

ForneyLab.hasVariable — Method.

hasVariable(graph, var) checks if var is part of graph.

ForneyLab.inferMarginalRule — Method.

Infer the rule that computes the joint marginal over cluster

ForneyLab.inferUpdateRule! — Method.

Infer the update rule that computes the message for entry, as dependent on the inbound types

ForneyLab.inferUpdateRule! — Method.

Infer the update rule that computes the message for entry, as dependent on the inbound types

ForneyLab.inferUpdateRules! — Method.

inferUpdateRules!(schedule) infers specific message update rules for all schedule entries.

ForneyLab.interfaceToScheduleEntryIdx — Method.

Generate a mapping from interface to schedule entry index. Multiple interfaces can map to the same schedule entry if the graph contains composite nodes.

ForneyLab.internalSumProductSchedule — Method.

internalSumProductSchedule() generates a sum-product message passing schedule for the inner graph of a CompositeNode. This schedule produces the sum-product message out of the specified outbound_interface.

ForneyLab.isApproxEqual — Method.

isApproxEqual: check approximate equality

ForneyLab.isRoundedPosDef — Method.

isRoundedPosDef: is input matrix positive definite? Round to prevent fp precision problems that isposdef() suffers from.

ForneyLab.isValid — Method.

isValid: return true if the parameter field exists and (the first element of) the parameter is not NaN

ForneyLab.laplace — Method.

Gamma approximation to the log-normal distribution using Laplace's method

ForneyLab.leaftypes — Method.

leaftypes(datatype) returns all subtypes of datatype that are leafs in the type tree.

ForneyLab.localClusterIds — Method.

Return the ids of the clusters/variables to which edges connected to node belong

ForneyLab.localRecognitionFactorIds — Method.

Return the ids of the recognition factors to which edges connected to node belong

ForneyLab.localRecognitionFactorization — Method.

Return a dictionary from recognition factor-id to variable/cluster-ids local to node

ForneyLab.mapToBernoulliParameterRange — Method.

Map m to range of Bernoulli parameter p ∈ [0, 1]

ForneyLab.marginalString — Method.

Depending on the origin of the Clamp node message, contruct the marginal code.

ForneyLab.matches — Method.

Special inheritance rules for parametric Message types

ForneyLab.messageString — Method.

Depending on the origin of the Clamp node message, contruct the outbound message code.

ForneyLab.nodesConnectedToExternalEdges — Method.

Find the nodes in recognition_factor that are connected to external edges

ForneyLab.nodesConnectedToExternalEdges — Method.

Find the nodes connected to internal_edges that are also connected to external edges

ForneyLab.pack — Method.

This function ensures the argument expression is evaluated at runtime, allowing access to local variables

ForneyLab.recognitionFactorId — Method.

Return the id of the RecognitionFactor that edge belongs to

ForneyLab.summaryDependencyGraph — Method.

summaryDependencyGraph(edgeset)

Returns a DependencyGraph (directed graph) that encodes the dependencies among summary messages (such as sum-product messages) in edgeset. All Interfaces in edgeset are vertices in the dependency graph. The dependency graph can be used for loop detection, scheduling, etc.

ForneyLab.summaryPropagationSchedule — Method.

summaryPropagationSchedule(variables) builds a generic summary propagation Schedule for calculating the marginal distributions of every variable in variables. The message update rule in each schedule entry is set to Nothing.

ForneyLab.trigammaInverse — Method.

trigammaInverse(x): solve trigamma(y) = x for y.

Uses Newton's method on the convex function 1/trigramma(y). Iterations converge monotonically. Based on trigammaInverse implementation in R package "limma" by Gordon Smyth: https://github.com/Bioconductor-mirror/limma/blob/master/R/fitFDist.R

ForneyLab.variateType — Method.

Extract VariateType from dist