Functions

Functions

Functions exported from FourierFlows:

FourierFlows.TimeStepper — Function.

TimeStepper(stepper, eq, dt=nothing)

Generalized timestepper constructor. If stepper is explicit, dt is not used.

FourierFlows.cxtype — Method.

cxtype(T)

Returns T when T is Complex, or Complex{T} when T is Real.

FourierFlows.dealias! — Method.

dealias!(a, g, kalias)

Dealias a on the grid g with aliased x-wavenumbers kalias.

FourierFlows.fltype — Method.

fltype(T)

Returns T when T<:AbstractFloat or Tf when T<:Complex{Tf}.

FourierFlows.gridpoints — Method.

gridpoints(g)

Returns the collocation points of the grid g in 2D arrays X, Y.

FourierFlows.increment! — Method.

increment!(diag)

Increment the Diagnostic diag, or an array of Diagnostics diags.

FourierFlows.innereltype — Method.

innereltype(x)

Recursively determine the 'innermost' type in by the collection x (which may be, for example, a collection of a collection).

FourierFlows.savediagnostic — Method.

savediagnostic(diag, diagname)

Save diagnostics in diag to file, labeled by diagname.

FourierFlows.saveoutput — Method.

saveoutput(out)

Save the fields in out.fields to out.path.

FourierFlows.saveproblem — Method.

saveproblem(prob, filename)

Save certain aspects of a problem.

FourierFlows.stepforward! — Method.

stepforward!(prob, diags, nsteps)

Step forward prob for nsteps, incrementing diags along the way. diags may be a single Diagnostic or a Vector of Diagnostics.

FourierFlows.stepforward! — Method.

stepforward!(prob, nsteps)

Step forward prob for nsteps.

FourierFlows.stepforward! — Method.

stepforward!(prob)

Step forward prob one time step.

FourierFlows.superzeros — Method.

superzeros(T, A)

Returns an array like A, but full of zeros. If innereltype(A) can be promoted to T, then the innermost elements of the array will have type T.

FourierFlows.update! — Method.

update!(diag)

Update diag.

Functions exported from KuramotoSivashinsky:

Modules = [FourierFlows.KuramotoSivashinsky]
Private = false
Order = [:function]

Functions exported from TracerAdvDiff:

Modules = [FourierFlows.TracerAdvDiff]
Private = false
Order = [:function]