Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FredData.Fred — Type.

A connection to the Fred API.

Constructors

Fred(): Key detected automatically. First, looks for the environment variable FREDAPIKEY, then looks for the file ~/.freddatarc.

Fred(key::AbstractString): User specifies key directly

Arguments

key: Registration key provided by the Fred.

Notes

Set the API url with set_api_url!(f::Fred, url::AbstractString)

FredData.FredSeries — Type.

FredSeries(...)

Represent a single data series, and all associated metadata, as queried from FRED.

The following fields are available:

id

title

units_short

units

seas_adj_short

seas_adj

freq_short

freq

realtime_start

realtime_end

last_updated

notes

trans_short

data

FredData.get_data — Method.

get_data(f::Fred, series::AbstractString; kwargs...)

Request one series using the FRED API.

Arguments

f: Fred connection object

series: series mnemonic

Optional Arguments

kwargs...: key-value pairs to be appended to the FRED request. Accepted keys include:

realtime_start: the start of the real-time period as YYYY-MM-DD string

realtime_end: the end of the real-time period as YYYY-MM-DD string

limit: maximum number of results to return

offset: non-negative integer

sort_order: "asc", "desc"

observation_start: the start of the observation period as YYYY-MM-DD string

observation_end: the end of the observation period as YYYY-MM-DD string

units: one of "lin", "chg", "ch1", "pch", "pc1", "pca", "cch", "cca", "log"

frequency: one of "d", "w", "bw", "m", "q", "sa", "a", "wef", "weth", "wew", "wetu", "wem", "wesu", "wesa", "bwew", "bwem"

aggregation_method: one of "avg", "sum", "eop"

output_type: one of 1 (obsevations by real-time period), 2 (observations by vintage date, all observations), 3 (observations by vintage date, new and revised observations only), 4 (observations, initial release only)

vintage_dates: vintage dates as comma-separated YYYY-MM-DD strings