Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FreeTypeAbstraction.renderstring! — Method.

renderstring!(img::AbstractMatrix, str::String, face, pixelsize, y0, x0;
fcolor=one_or_typemax(T), bcolor=zero(T), halign=:hleft, valign=:vbaseline) -> Matrix

Render str into img using the font face of size pixelsize at coordinates y0,x0.

Arguments

y0,x0: origin is in upper left with positive y going down

fcolor: foreground color; typemax(T) for T<:Integer, otherwise one(T)

bcolor: background color; set to nothing for transparent

halign: :hleft, :hcenter, or :hright

valign: :vtop, :vcenter, :vbaseline, or :vbottom