Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

FunctionalTables.FunctionalTable — Type.

FunctionalTable(itr)
FunctionalTable(itr, sortspecs)
FunctionalTable(itr, sortspecs, sortingpolicy; cfg)

Create a FunctionalTable from an iterable that returns NamedTuples.

Returned values need to have the same names (but not necessarily types).

sorting specifies sorting, and is a tuple of :key or :key => reverse elements.

cfg determines sink configuration for collecting elements of the columns, see SinkConfig.

FunctionalTables.RLEVector — Type.

struct RLEVector{C, T, S}

An RLE encoded vector, using negative lengths for missing values.

When an elemenet in counts is positive, it encodes that many of the corresponding element in data.

Negative counts encode missing values of type S (has to be a concrete singleton). In this case there is no corresponding value in data, ie data may have fewer elements than counts.

An RLEVector can also act as a column.

FunctionalTables.SinkConfig — Type.

struct SinkConfig{M}

FunctionalTables.SinkConfig — Method.

SinkConfig(; useRLE, missingvalue)

Make a sink configuration, using defaults.

FunctionalTables.TrustSorting — Type.

Accept the specified sorting to hold without any checks (except for verifying that column names are valid).

Note

This can lead to incorrect results, use cautiously. VerifySorting is recommended instead as it has little overhead.

FunctionalTables.TrySorting — Type.

Try the specified sorting, then gracefully degrade to a subset of it that holds.

FunctionalTables.VerifySorting — Type.

Verify that the specified sorting holds. This is the default sorting policy.

FunctionalTables.by — Method.

by(indexkeys, ft; cfg)

An iterator that groups rows of tables by the columns indexkeys, returning (index::NamedTupe, table::FunctionalTable) for each contiguous block of the index keys.

cfg is used for collecting table.

FunctionalTables.collect_column — Method.

collect_column(cfg, itr)

Collect results from itr into a sink (using config cfg), then finalize and return the column.

FunctionalTables.collect_columns — Method.

columns, sorting =

collect_columns(cfg, itr, sorting, sortingpolicy)

Collect results from itr, which are supposed to be NamedTuples with the same names, into sinks (using config cfg), then finalize and return the NamedTuple of the columns and the sorting (which is ≡ sorting unless sortingpolicy ≡ SortingPolicy(:prefix)).

sorting is ignored when sortingpolicy ≡ SortingPolicy(:accept), otherwise used according to the value of the latter.

FunctionalTables.columns — Method.

columns(ft; vector, mutable)

Return the columns in a NamedTuple.

When mutable, all columns will be mutable <: AbstractVector, and not share (shallow) structure.

When vector, all columns will be <: AbstractVector, but may be immutable or share structure.

FunctionalTables.select — Method.

select(ft, keep)

select(ft, keep...) select(ft; drop)

Select a subset of columns from the table.

select(ft, keep) and select(ft, keep...) returns the table with the given columns.

select(ft; drop = keys) is a convenience form for keeping all but the given columns.

FunctionalTables.ColumnSort — Type.

ColumnSort(key)
ColumnSort(key, rev)

Process an individual sort specification, called by column_sorting.

FunctionalTables.ColumnSort — Type.

struct ColumnSort{K, R}

Sort specification for a column. K::Symbol is a key for sorting, R::Bool reverses sorting for this key.

This type is not part of the interface, for internal representation.

FunctionalTables.ColumnSorting — Type.

Type for sorting, used internally.

FunctionalTables.Keys — Type.

Type for keys, used internally.

FunctionalTables.RepeatedValue — Type.

RepeatedValue(value, len)

Equivalent to a vector containing len instances of value. Used internally.

FunctionalTables.SortingPolicy — Type.

struct SortingPolicy{K}

Policy for dealing with specified sortings.

See VerifySorting, TrustSorting, and TrySorting.

FunctionalTables.SplitTable — Type.

struct SplitTable{K, T<:FunctionalTable, C<:SinkConfig}

Implements by.

Iterator state is a tuple, with

sinks and firstkey, created from the element with a non-matching key,

itrstate, the iteration state for itr.

Base.map — Method.

map(f, st; cfg)

Map a table split with by using f.

Specifically, f(indexkeys, table) receives the index keys (a NamedTuple) and a FunctionalTable.

It is supposed to return an iterable that returns rows (can be a FunctionalTable). These will be prepended with the corresponding index, and collected into a FunctionalTable with cfg.

When f returns just a single row (eg aggregation), wrap by Ref to create a single-element iterable.

Base.map — Method.

map(f, ft; cfg)

Base.merge — Method.

merge(a, b; replace)

Merge two FunctionalTables.

When replace == true, columns in the first one are replaced by second one, otherwise an error is thrown if column names overlap.

Base.merge — Method.

merge(ft, f; cfg, replace)

Map ft using f by rows, then merge the two. See map(::Callable,::FunctionalTable).

cfg is passed to map, replace governs replacement of overlapping columns in merge.

FunctionalTables.append1 — Method.

append1(v, elt)

Append elt to v, allocating a new vector and copying the contents.

Type of new collection is calculated using promote_type.

FunctionalTables.cancontain — Method.

cancontain(T, elt)

Test if a collection of element type T can contain a new element elt without any loss of precision.

FunctionalTables.checkvalidkeys — Method.

checkvalidkeys(argkeys, validkeys)

Check that argkeys are a subset of the valid keys.

When that does not hold, throw and error with an informative message.

FunctionalTables.cmp_sorting — Method.

cmp_sorting(cs, a, b)

Compare rows a and b, which support the getproperty interface, with the given column sorting.

Internal.

FunctionalTables.column_sorting — Function.

column_sorting(sortspecs)
column_sorting(sortspecs, colkeys)

Process sorting specifications for columns (an iterable or possibly a ColumnSorting), return a value of type ColumnSorting.

Verify that sort keys are unique. When colkeys is given, verify that the sort keys are a subset of it.

Accepted syntax:

:key, for sorting a column in ascending order

:key => reverse, for sorting a column in descending order

All functions which accept sort specs should use this, but the function itself is not part of the API.

FunctionalTables.dropkeys — Method.

dropkeys(ftkeys, drop)

Check that drop ⊆ ftkeys, then return ftkeys ∖ drop.

FunctionalTables.finalize_sink — Method.

finalize_sink(?, sink)

Convert sink to a column.

sink may share structure with the result and is not supposed to be used for saving any more elements.

FunctionalTables.finalize_sinks — Method.

finalize_sinks(cfg, sinks)

Finalize a (named) tuple of sinks.

FunctionalTables.fuse — Method.

fuse(f, index, tables)

Prepend the index as repeated columns to f(index, tables...).

FunctionalTables.make_sink — Method.

make_sink(cfg, elt)

Create and return a sink using configuration cfg that stores elt.

FunctionalTables.make_sinks — Method.

make_sinks(cfg, elts)

Make sinks for a (named) tuple pf elements.

FunctionalTables.merge_repeated — Method.

merge_repeated(index, ft)

Make a functional table from index, repeating each value for a column to match the length of ft, then merge the two.

FunctionalTables.merge_sorting — Method.

merge_sorting(cs, otherkeys)

Calculate sorting when a table with sorting is merged with a table containing otherkeys, which may replace columns.

FunctionalTables.narrow — Method.

narrow(x)

Convert the argument to a narrower type if possible without losing precision.

Note

This function is not type stable, use only when new container types are determined.

FunctionalTables.retained_sorting — Method.

retained_sorting(cs, a, b)

Return the part of sorting under which a ≤ b.

FunctionalTables.select_sorting — Method.

select_sorting(cs, keep)

Calculate sorting when only keep keys are kept.

keep may contain keys not in the sorting, ie those of a FunctionalTable.

FunctionalTables.sortkey — Method.

sortkey(cs)

Accessor for sort key. Internal.

FunctionalTables.split_namedtuple — Method.

split_namedtuple(splitter, nt)

Splits a named tuple in two, based on the names in splitter.

Returns two NamedTuples; the first one is ordered as splitter, the second one with the remaining values as in the original argument.

julia> split_namedtuple(NamedTuple{(:a, :c)}, (c = 1, b = 2, a = 3, d = 4))
((a = 3, c = 1), (b = 2, d = 4))

FunctionalTables.store!_or_reallocate — Method.

store!_or_reallocate(?, sink, elt)

Either store elt in sink (in which case the returned value is ≡ sink), or allocate a new sink that can do this, copy the contents, save elt and return that (then the returned value is ≢ sink).