Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GLFW.Window — Method.

Function to create a pure GLFW OpenGL window

GLFW.CreateWindowSurface — Function.

CreateWindowSurface(instance, window, allocator=C_NULL)

Create a Vulkan surface for the specified window.

GLFW.GetInstanceProcAddress — Method.

GetInstanceProcAddress(instance, procname) -> funcptr

Return the address of the specified Vulkan core or extension function for the specified instance. funcptr can be used directly as the first argument of ccall: ccall(funcptr, ...).

GLFW.GetPhysicalDevicePresentationSupport — Method.

GetPhysicalDevicePresentationSupport(instance, device, queuefamily)

Return whether the specified queue family of the specified physical device supports presentation to the platform GLFW was built for.

GLFW.GetRequiredInstanceExtensions — Method.

GetRequiredInstanceExtensions()

Return a vector of names of Vulkan instance extensions.

GLFW.SetWindowIcon — Function.

GLFW.SetWindowIcon(window::Window, image::Matrix{NTuple{4, UInt8}})

GLFW.SetWindowIcon(window::Window, images::Vector{<:AbstractMatrix{NTuple{4,UInt8}}})

Set the window icon, where a single image may be passed or a vector of images with different icon sizes. The images must be of RGBA format. Before calling this function it might be necessary to reinterpret the image as a matrix of element type NTuple{4, UInt8} if the icons are loaded with type RGBA{N0f8}

Examples

using FileIO
icons = load.(["icon-16.png", "icon-32.png", "icon-128.png"])
buffs = reinterpret.(NTuple{4, UInt8}, icons)
GLFW.SetWindowIcon(win, buffs)
GLFW.PollEvents() # needs a poll events to become active

GLFW.VulkanSupported — Method.

VulkanSupported()

Return whether the Vulkan loader has been found.

GLFW.primarymonitorresolution — Method.

Returns the monitor resolution of the primary monitor.

GLFW.set_visibility! — Method.

Sets visibility of OpenGL window. Will still render if not visible. Only applies to the root screen holding the opengl context.

GLFW.standard_context_hints — Method.

Tries to create sensible context hints! Taken from lessons learned at: GLFW

GLFW.standard_screen_resolution — Method.

Takes half the resolution of the primary monitor. This should make for sensible defaults!

GLFW.standard_window_hints — Method.

Standard window hints for creating a plain context without any multisampling or extra buffers beside the color buffer