Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GLPK.get_status — Method.

get_status(model::Optimizer)

Get the status from GLPK depending on which method was used to solve the model.

GLPK.AbstractCallbackData — Type.

AbstractCallbackData

An abstract type to prevent recursive type definition of Optimizer and CallbackData, each of which need the other type in a field.

GLPK.CallbackData — Type.

CallbackData

GLPK.CallbackFunction — Type.

CallbackFunction

The attribute to set the callback function in GLPK. The function takes a single argument of type CallbackData.

GLPK.__internal_callback__ — Method.

__internal_callback__(tree::Ptr{Cvoid}, info::Ptr{Cvoid})

Dummy callback function for internal use only. Responsible for updating the objective bound.

GLPK._certificates_potentially_available — Method.

_certificates_potentially_available(model::Optimizer)

Return true if an infeasiblity certificate or an unbounded ray is potentially available (i.e., the model has been solved using either the Simplex or Exact methods).

GLPK._throw_invalid_method — Method.

_throw_invalid_method(instance::Optimizer)

A helper function to throw an error when the method is set incorrectly. Mainly used to enforce type-stability in functions that have a run-time switch on the method.

GLPK.add_lazy_constraint! — Method.

add_lazy_constraint!(cb_data::GLPK.CallbackData, func::LQOI.Linear, set::S) where S <: Union{LQOI.LE, LQOI.GE, LQOI.EQ}

Add a lazy constraint to the model cb_data.model. This can only be called in a callback from GLPK.IROWGEN.

GLPK.add_row! — Method.

add_row!(problem::GLPK.Prob, columns::Vector{Int},
         coefficients::Vector{Float64}, sense::Cchar, rhs::Real)

Helper function to add a row to the problem. Sense must be one of 'E' (ax == b), 'G' (ax >= b), 'L' (ax <= b) , or 'R' (b <= ax).

If the sense is 'R' the rhs should be the lower bound, and the bounds should be set in a new API call to enforce the upper bound.

GLPK.change_row_sense! — Method.

change_row_sense!(model::Optimizer, row, sense)

Convert a linear constraint into another type of linear constraint by changing the comparison sense.

Constraint types supported are 'E' (equality), 'L' (less-than), and 'G' (greater-than).

For example, ax <= b can become ax >= b or ax == b.

GLPK.copy_function_result! — Method.

copy_function_result!(dest::Vector, foo, model::GLPK.Prob)

A helper function that loops through the indices in dest and stores the result of foo(model, i) for the ith index.

GLPK.get_col_bound_type — Method.

get_col_bound_type(lower::Float64, upper::Float64)

Return the GLPK type of the variable bound given a lower bound of lower and an upper bound of upper.

GLPK.get_infeasibility_ray — Method.

get_infeasibility_ray(model::Optimizer, ray::Vector{Float64})

Get the Farkas certificate of primal infeasiblity.

Can only be called when GLPK.simplex is used as the solver.

GLPK.get_unbounded_ray — Method.

get_unbounded_ray(model::Optimizer, ray::Vector{Float64})

Get the certificate of primal unboundedness.

Can only be called when GLPK.simplex is used as the solver.

GLPK.get_variable_types — Method.

get_variable_types(model::Optimizer)

Return a vector of symbols (one element for each variable) of the variable type. The symbols are given by the key-value mapping in GLPK.VARIABLE_TYPE_MAP.

GLPK.load_variable_primal! — Method.

load_variable_primal!(cb_data::CallbackData)

Load the variable primal solution in a callback.

This can only be called in a callback from GLPK.IROWGEN. After it is called, you can access the VariablePrimal attribute as usual.

GLPK.round_bounds_to_integer — Method.

round_bounds_to_integer(model)::Tuple{Bool, Vector{Float64}, Vector{Float64}}

GLPK does not allow integer variables with fractional bounds. Therefore, we round the bounds of binary and integer variables to integer values prior to solving.

Returns a tuple of the original bounds, along with a Boolean flag indicating if they need to be reset after solve.

GLPK.set_parameter — Method.

set_parameter(param_store, key::Symbol, value)

Set the field name key in a param_store type (that is one of InteriorParam, IntoptParam, or SimplexParam) to value.

GLPK.set_variable_bound — Method.

set_variable_bound(model::Optimizer, column::Int, lower::Float64, upper::Float64)

Set the bounds of the variable in column column to [lower, upper].