Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GaloisFields.GaloisField — Function.

F = GaloisField(p)
F,α = GaloisField(p, :β => [1, 0, 1])
F,α = GaloisField(p, n, :β)

Return a type representing a finite field.

The single-argument signature returns the finite field $ℤ/pℤ$.

The two-arguments signature returns an algebraic extension of that field, with minimum polynomial given by the second argument: a dense representation of the univariate, monic polynomial, with ascending degree.

The three-arguments signature returns an algebraic extension of that field, with minimum polynomial equal to the Conway polynomial for $(p,n)$. The GaloisFields package ships with a database of Conway polynomials and will raise an error if it does not contain an entry for $(p,n)$.

Note that in the latter two cases, the variable name (e.g. β above) is part of the type. This lets you define identifications between isomorphic (sub)fields. For example, with the following definition

F = @GaloisField! 𝔽₂ β^2 + β + 1
G = @GaloisField! 𝔽₂ γ^2 + γ + 1

the fields $F$ and $G$ are isomorphic, but not canonically. We might define

@GaloisFields.identify β => γ + 1
@GaloisFields.identify γ => β + 1

to allow for conversions like

G(β)
convert(F, γ + 1)

In the Conway case, you do not have to define your own identifications, as the Conway polynomials satisfy compatibility relations that allow us to use certain distinguished inclusions between them.

GaloisFields.char — Method.

p = char(GaloisField(3)) # returns 3

Return the characteristic of a finite field, or 0 for <:Integer or <:Rational{<Integer}.

GaloisFields.AbstractExtensionField — Type.

abstract type AbstractExtensionField <: AbstractGaloisField end

A type representing a finite extension of an underlying finite field.

GaloisFields.AbstractGaloisField — Type.

abstract type AbstractGaloisField <: Number end

A type representing finite fields.

GaloisFields.ExtensionField — Type.

F = ExtensionField{F <: AbstractGaloisField, N, α, MinPoly}

Algebraic extension of a finite field $F$ of degree $N$.

GaloisFields.PrimeField — Type.

PrimeField{I<:Integer, p}

A type representing an element in ℤ/pℤ.

GaloisFields.Reduced — Type.

Reduced()

A helper singleton used for asserting that an input value has already been reduced mod p.

GaloisFields.defaultshow — Method.

defaultshow(io, t)

Overloading display of types can be a bit hairy; I've seen a declaration like

show(..., ::Type{<:Val{B}}) where B

being called for non-concrete types. (I haven't dug deep enough to find a nice minimal example.)

That's why all show overloads for types do

!isconcretetype(t) && return defaultshow(io, t)