Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GaussianDistributions.Gaussian — Type.

Gaussian(μ, Σ) -> P

Gaussian distribution with mean μ and covariance Σ. Defines rand(P) and (log-)pdf(P, x). Designed to work with Numbers, UniformScalings, StaticArrays and PSD-matrices.

Implementation details: On Σ the functions logdet, whiten and unwhiten (or chol as fallback for the latter two) are called.

GaussianDistributions.PSD — Type.

PSD{T}

Simple wrapper for the lower triangular Cholesky root of a positive (semi-)definite element σ.

GaussianDistributions.BiNormal — Type.

Bivariate standard normal distribution with correlation ρ

GaussianDistributions.conditional — Method.

conditional(P::Gaussian, A, B, xB)

Conditional distribution of X[i for i in A] given X[i for i in B] == xB if $X ~ P$.

GaussianDistributions.logpdfnormal — Method.

logpdfnormal(x, Σ)

Logarithm of the probability density function of centered Gaussian with covariance Σ

GaussianDistributions.sumlogdiag — Method.

Sum of the log of the diagonal elements. Second argument d is used to handle UniformScaling and other linear operators whose dimensions are determined by the dimension of the argument they work on.