Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GaussianMixtureTest.asymptoticdistribution — Method.

asymptoticdistribution(x, wi, mu, sigmas);

Simulate the asymptotic distribution of test statistic for kstest. nrep is the number of random values to generate. debuginfo is whether to show the debug information. When number of components of null distribution is greater than 1, the test statistic has no closed form asymptotic distribution. When the null distribution is just normal, the asymptotic distribution is just Chisq(2).

GaussianMixtureTest.gmm — Function.

gmm(x, ncomponent)

Estimate parameters of ncomponent gaussian mixture on the data x. Optional arguments of gmm:

ncomponent: the number of components

wi_init mu_init and sigmas_init: the initial values

maxiteration: the number of iterations

tol: the tolerance of convergence criteria

an and sn: the penalty weight and variance term

taufixed: for kstest, whether fix the tau value

whichtosplit and tau: for kstest, which component to split and the split proportion

mu_lb and mu_ub for kstest, the lower and upper limits of components means

pl: wheter the penalty on sigmas be included in the log likelihood in the final two EM steps. Note that the starting value with largest penalized log likelihood is picked, but the penalty term should not be included in the likelihood ratio

ptau: whether to add the penalty on tau be included in likelihood. Better to be true since the more tau values we try the larger the test statistic

GaussianMixtureTest.gmmrepeat — Method.

gmmrepeat(x, C)

Repeat the gmm for ntrials randomly generated starting values and pick the one with largest penalized likelihood.

wi_init, mu_lb, mu_ub, sigmas_lb, sigmas_ub: specify how to generate the random starting values

taufixed, whichtosplit, tau: whether keep the ratio between wi[whichtosplit] and wi[whichtosplit]+wi[whichtosplit+1] fixed at a constant tau.

sn and an: the penalty

debuginfo and tol: whether print the debug information and the convergence critera

pl: wheter the penalty on sigmas be included in the log likelihood in the final two EM steps. Note that the starting value with largest penalized log likelihood is picked, but the penalty term should not be included in the likelihood ratio

ptau: whether to add the penalty on tau be included in likelihood. Better to be true since the more tau values we try the larger the test statistic

GaussianMixtureTest.kstest — Method.

kstest(x, C0)

Do the EM test under null Hypothesis of C0 components. If rejected, then it suggest the true number of components is greater than C0. Optional arguments for kstest

vtau: the finite set of tau value, default to try 0.5 only

ntrials: the number of initial values to try, default to be 25

debuginfo: whether show the debug information