Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GaussianMixtures.CSstats — Type.

CSstats a type holding centered and scaled zeroth and first order GMM statistics

GaussianMixtures.Cstats — Type.

Cstats, a type holding centered zeroth, first and second order GMM statistics

GaussianMixtures.Data — Type.

Data is a type for holding an array of feature vectors (i.e., matrices), or references to files on disk. The data is automatically loaded when needed, e.g., by indexing.

GaussianMixtures.GMM — Type.

GMM is the type that stores information of a Guassian Mixture Model. Currently two main covariance types are supported: full covarariance and diagonal covariance.

GaussianMixtures.GMM — Method.

GMM(n::Int, d::Int, kind::Symbol=:diag) initializes a GMM with means 0 and Indentity covariances

GaussianMixtures.GMMprior — Type.

GMMprior is a type that holds the prior for training GMMs using Variational Bayes.

GaussianMixtures.History — Type.

History, a type to record the history of how a GMM is built.

GaussianMixtures.VGMM — Type.

VGMM is the type that is used to store a GMM in the Variational Bayes training.

GaussianMixtures.covars — Method.

covars(::GMM) returns the covariance matrices Σ of the Gaussians in the mixture.

GaussianMixtures.kind — Method.

kind(::GMM) returns the kind of GMM, either :diag or :full

GaussianMixtures.means — Method.

means(::GMM) returns the means μ of the Gaussians in the mixture

GaussianMixtures.nparams — Method.

nparams(::GMM) returns the number of free parameters in the GMM

GaussianMixtures.weights — Method.

weights(::GMM) returns the weights w, or priors, of the Gaussians in the mixture

GaussianMixtures.GaussianMixture — Type.

GaussianMixture, an abstract type for a mixture of full-covariance or diagonal-covariance Gaussian distributions

GaussianMixtures.addhist! — Method.

addhist!(::GMM, s) adds a comment s to the GMMM

GaussianMixtures.covar — Method.

covar(GMM.Σ) extracts the covariances Σ (which may be encoded as chol(inv(Σ))

GaussianMixtures.density — Method.

density(gmm::GMM, X) returns log(P(X|μ, Σ))

GaussianMixtures.full — Method.

full(::GMM) turns a diagonal covariance GMM into a full-covariance GMMM

LinearAlgebra.diag — Method.

diag(::GMM) turns a full-covariance GMM into a diagonal-covariance GMM, by ignoring off-diagonal elements

ScikitLearnBase.score_samples — Method.

score_samples(gmm::GMM, X::Matrix)

Return the per-sample likelihood of the data under the model.