Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GenericSVD.PackedUVt — Type.

PackedUVt <: Factorization{T}

Packed storage of bidiagonalizing QR reflectors U and V'.

GenericSVD.bidiagonalize_tall! — Method.

bidiagonalize_tall!{T}(A::Matrix{T},B::Bidiagonal)

Bidiagonalize a tall matrix A into B. Both arguments are overwritten.

GenericSVD.is_offdiag_approx_zero! — Method.

is_offdiag_approx_zero!(B::Bidiagonal,i,ɛ)

Tests if the element B[i-1,i] is approximately zero, using the criteria

\[ |B_{i-1,i}| ≤ ɛ*(|B_{i-1,i-1}| + |B_{i,i}|)\]

If true, sets the element to exact zero.

GenericSVD.svd_bidiag! — Method.

svd_bidiag!(B::Bidiagonal [, U, Vt [, ϵ]])

Compute the SVD of a bidiagonal matrix B, via an implicit QR algorithm with shift (known as a Golub-Kahan iterations).

Optional arguments:

U and Vt: orthogonal matrices which pre- and post-multiply B, for computing the SVD of a full matrix X = U*B*Vt from its bidiagonalized form.

ϵ: the tolerance for testing zeros of the offdiagonal elements of B (see below).

Algorithm:

This proceeds by iteratively finding the lowest strictly-bidiagonal submatrix, i.e. n₁, n₂ such that

     [ d ?           ]
     [   d 0         ]
  n₁ [     d e       ]
     [       d e     ]
  n₂ [         d 0   ]
     [           d 0 ]

then applying a Golub-Kahan QR iteration.

GenericSVD.svd_gk! — Method.

svd_gk!(B::Bidiagonal{T},U,Vt,n₁,n₂,shift) where T <: Real

Applies a Golub-Kahan SVD step (an implicit QR with shift) to the submatrix B[n₁:n₂,n₁:n₂], applying the inverse transformations to U and Vt (preserving U*B*Vt).

A Givens rotation is applied to the top 2x2 matrix, and the resulting "bulge" is "chased" down the diagonal to the bottom of the matrix.

GenericSVD.svd_zerodiag_col! — Method.

svd_zerodiag_col!(B::Bidiagonal,Vt,n₁,n₂)

Sets B[n₂,n₂] to zero, then zeros out column n₂ by applying sequential column (right) Givens rotations up to n₁, and the corresponding inverse rotations to Vt (preserving B*Vt).

GenericSVD.svd_zerodiag_row! — Method.

svd_zerodiag_row!(U,B,n₁,n₂)

Sets B[n₁,n₁] to zero, then zeros out row n₁ by applying sequential row (left) Givens rotations up to n₂, and the corresponding inverse rotations to U (preserveing U*B.

GenericSVD.svdvals2x2 — Method.

svdvals2x2(f, h, g)

The singular values of the matrix

B = [ f g ;
      0 h ]

(i.e. the sqrt-eigenvalues of B'*B).

This is a direct translation of LAPACK DLAS2.