Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GenericSchur.eigvalscond — Method.

eigvalscond(S::Schur,nsub::Integer)

Estimate the reciprocal of the condition number of the nsub leading eigenvalues of S. (Use ordschur to move a subspace of interest to the front of S.)

See the LAPACK User's Guide for details of interpretation.

GenericSchur.subspacesep — Method.

subspacesep(S::Schur,nsub::Integer)

Estimate the reciprocal condition of the separation angle for the invariant subspace corresponding to the leading block of size nsub of a Schur decomposition. (Use ordschur to move a subspace of interest to the front of S.)

See the LAPACK User's Guide for details of interpretation.

GenericSchur.triangularize — Method.

triangularize(S::Schur{T}) => Schur{complex{T})

convert a (standard-form) quasi-triangular real Schur factorization into a triangular complex Schur factorization.

GenericSchur._trexchange! — Method.

reorder T by a unitary similarity transformation so that T[iold,iold] is moved to T[inew,inew]. Also right-multiply Z by the same transformation.

GenericSchur.adjtrsylvester! — Method.

adjtrsylvesterh!(A,B,C) => X, σ

solve the Sylvester equation $Aᴴ X - X Bᴴ = σ C$, for upper triangular A and B, overwriting C and setting σ to avoid overflow.

GenericSchur.gschur! — Method.

gschur!(A::StridedMatrix) -> F::Schur

Destructive version of gschur (q.v.).

GenericSchur.gschur — Method.

gschur(A::StridedMatrix) -> F::Schur

Computes the Schur factorization of matrix A using a generic implementation. See LinearAlgebra.schur for usage.

GenericSchur.norm1est! — Method.

norm1est!(applyA!,applyAH!,y::Vector) => γ

Estimate the 1-norm of a linear operator A expressed as functions which apply A and adjoint(A) to a vector such as y.

cf. N.J. Higham, SIAM J. Sci. Stat. Comp. 11, 804 (1990)

GenericSchur.safemin — Method.

a "safe" version of floatmin, such that 1/sfmin does not overflow.

GenericSchur.safescale! — Method.

multiply an array A by a real number cto/cfrom expressed as a ratio, computing without over/underflow where possible.

cfrom must not be zero.

GenericSchur.trsylvester! — Method.

trsylvester!(A,B,C) => X, σ

solve the Sylvester equation $A X - X B = σ C$ for upper triangular and square A and B, overwriting C and setting σ to avoid overflow.

LinearAlgebra.eigvecs — Method.

eigvecs(S::Schur{<:Complex}; left=false)

Compute right or left eigenvectors from a Schur decomposition.