Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GenomicAnnotations.Locus — Type.

Struct for storing information on genomic locations. strand can be '+', '-', or '.' when the strand is irrelevant.

GenomicAnnotations.Locus — Method.

Locus()
Locus(position::UnitRange{Int})
Locus(position::UnitRange{Int}, strand::Char)

BioCore.sequence — Method.

sequence(gene::AbstractGene)

Return genomic sequence for gene.

GenomicAnnotations.addgene! — Method.

addgene!(chr::Chromosome, feature, locus; kw...)

Add gene to chr. locus can be a Locus, a UnitRange, or a StepRange.

GenomicAnnotations.iscomplement — Method.

iscomplement(gene::Abstract)

Return true if gene.locus.compliment == '-', otherwise return false.

GenomicAnnotations.pushproperty! — Method.

pushproperty!(gene::AbstractGene, name::Symbol, x::T)

Add a property to gene, similarly to Base.setproperty!(::gene), but if the property is not missing in gene, it will be transformed to store a vector instead of overwriting existing data.

julia> eltype(chr.genedata[:EC_number])
Union{Missing,String}

julia> chr.genes[1].EC_number = "EC:1.2.3.4"
"EC:1.2.3.4"

julia> pushproperty!(chr.genes[1], :EC_number, "EC:4.3.2.1"); chr.genes[1].EC_number
2-element Array{String,1}:
 "EC:1.2.3.4"
 "EC:4.3.2.1"

julia> eltype(chr.genedata[:EC_number])
Union{Missing, Array{String,1}}

GenomicAnnotations.readgbk — Method.

readgbk(filename)

Parse GenBank-formatted file filename, returning a Chromosome.

Base.delete! — Method.

delete!(genes::AbstractArray{Gene, 1})

Delete all genes in genes from genes[1].parent.

Base.delete! — Method.

delete!(gene::Gene)

Delete gene from gene.parent. Warning: does not work when broadcasted! Use delete!(::AbstractVector{Gene}) instead.

Base.get — Method.

get(g::Gene, key, default)

Retrieve key from g. If key is missing, return default.

GenomicAnnotations.filterseq — Method.

Parse footer (sequence) portion of a GenBank file, returning a String. When BioSequences is available for 0.7 this will be changed to a DNASequence.

GenomicAnnotations.parsechromosome — Method.

parsechromosome(lines)

Parse and return one chromosome entry, and the line number that it ends at.

GenomicAnnotations.parseposition — Method.

Parse lines encoding genomic position, returning the feature as a String, and an instance of Locus.

GenomicAnnotations.vectorise — Method.

vectorise(A::AbstractArray{Union{Missing, T}, 1}) where T

Convert an array of type Vector{Union{Missing, T}} to Vector{Union{Missing, Vector{T}}}.