Empirical variograms

Empirical variograms

An empirical variogram has the form:

\[\newcommand{\x}{\boldsymbol{x}} \hat{\gamma}(h) = \frac{1}{2|N(h)|} \sum_{(i,j) \in N(h)} (z_i - z_j)^2\]

where $N(h) = \left\{(i,j) \mid ||\x_i - \x_j|| = h\right\}$ is the set of pairs of locations at a distance $h$ and $|N(h)|$ is the cardinality of the set. Empirical variograms can be estimated using general distance functions. These can be used in order to for example:

Model anisotropy (e.g. ellipsoid distance)

Perform geostatistical simulation on spherical coordinate systems (e.g. haversine distance)

Please see Distances.jl for a complete list of options.

Omnidirectional

Variography.EmpiricalVariogram — Type.

EmpiricalVariogram(X, z₁, z₂=z₁; [optional parameters])

Computes the empirical (a.k.a. experimental) omnidirectional (cross-)variogram from data locations X and values z₁ and z₂.

EmpiricalVariogram(spatialdata, var₁, var₂=var₁; [optional parameters])

Alternatively, compute the (cross-)variogram for the variables var₁ and var₂ stored in a spatialdata object.

EmpiricalVariogram(partition, var₁, var₂=var₁; [optional parameters])

Alternatively, compute the (cross-)variogram on a partition of the data.

Parameters

nlags - number of lags (default to 20)

maxlag - maximum lag (default to half of maximum lag of data)

distance - custom distance function (default to Euclidean distance)

See also: DirectionalVariogram

Directional

Variography.DirectionalVariogram — Function.

DirectionalVariogram(spatialdata, direction, var₁, var₂=var₁; [optional parameters])

Computes the empirical (cross-)variogram for the variables var₁ and var₂ stored in spatialdata along a given direction.

Optional parameters include the parameters for EmpiricalVariogram and the parameters for DirectionPartitioner.

Notes

A DirectionalVariogram is just a function that first partitions the spatialdata using a DirectionPartitioner and then passes the result to the corresponding EmpiricalVariogram constructor.

See also: EmpiricalVariogram, DirectionPartitioner

Variogram plane plots are also available, please see Plotting.