Fitting variograms

Fitting variograms

Fitting theoretical variograms to empirical observations is an important preprocessing step to ensure valid mathematical models of spatial continuity for posterior estimation/simulation with variogram-based methods such as Kriging and sequential Gaussian simulation.

Given an empirical variogram, the fit function can be used to perform the fit:

Variography.fit — Function.

fit(V, γ, [algo])

Fit theoretical variogram type V to empirical variogram γ using algorithm algo. Default algorithm is WeightedLeastSquares.

fit(Variogram, γ, [algo])

Fit all stationary variogram types to empirical variogram γ, which are subtypes of Variogram, and return the one with minimum error as defined by the algorithm algo.

Currently the following fitting methods are implemented:

Weighted least squares

Variography.WeightedLeastSquares — Type.

WeightedLeastSquares(weightfun)

Fit theoretical variogram using weighted least squares with weighting function weightfun (e.g. h -> 1/h). If not weighting function is provided, bin counts of empirical variogram are normalized and used as weights.