Problems and solvers

Problems and solvers

One of the greatest features of GeoStats.jl is the ability to define geostatistical problems independently of the solution strategy. This design allows researchers and practioners to perform fair comparisons between different solvers. It is perhaps the single most important contribution of this project.

If you are an experienced user of geostatistics or if you do research in the field, you know how hard it is to compare algorithms fairly. Often a new algorithm is proposed in the literature, and yet the task of comparing it with the state of the art is quite demanding. Even when a comparison is made by the author after a great amount of effort, it is inevitably biased.

Part of this issue is attributed to the fact that a general definition of the problem is missing. What is it that we call an "estimation problem" in geostatistics? What about "stochastic simulation"? The answer to these questions is given below in the form of code.

Estimation problem

An estimation problem in geostatitsics is a triplet:

Spatial data (i.e. data with coordinates)

Spatial domain (e.g. regular grid, point collection)

Target variables (or variables to be estimated)

Each of these components is constructed separately, and then grouped (no memory is copied) in an EstimationProblem.

GeoStatsBase.EstimationProblem — Type.

EstimationProblem(spatialdata, domain, targetvars)

A spatial estimation problem on a given domain in which the variables to be estimated are listed in targetvars. The data of the problem is stored in spatialdata.

Examples

Create an estimation problem for rainfall precipitation measurements:

julia> EstimationProblem(spatialdata, domain, :precipitation)

Create an estimation problem for precipitation and CO₂:

julia> EstimationProblem(spatialdata, domain, [:precipitation, :CO₂])

Please check Spatial data and Domains for currently implemented data and domain types.

Simulation problem

Likewise, a stochastic simulation problem in geostatistics is represented with the same triplet. However, the spatial data in this case is optional in order to accomodate the concept of conditional versus unconditional simulation.

GeoStatsBase.SimulationProblem — Type.

SimulationProblem(spatialdata, domain, targetvars, nreals)
SimulationProblem(domain, targetvars, nreals)

A spatial simulation problem on a given domain in which the variables to be simulated are listed in targetvars.

For conditional simulation, the data of the problem is stored in spatialdata.

For unconditional simulation, a dictionary targetvars must be provided mapping variable names to their types.

In both cases, a number nreals of realizations is requested.

Examples

Create a conditional simulation problem for porosity and permeability with 100 realizations:

julia> SimulationProblem(spatialdata, domain, [:porosity,:permeability], 100)

Create an unconditional simulation problem for porosity and facies type with 100 realizations:

julia> SimulationProblem(domain, Dict(:porosity => Float64, :facies => Int), 100)

Notes

To check if a simulation problem has data (i.e. conditional vs. unconditional) use the hasdata method.

GeoStatsBase.hasdata — Function.

hasdata(problem)

Return true if simulation problem has data.

List of solvers

Below is the list of solvers distributed with GeoStats.jl. For more solvers, please check the project page on GitHub where a table is provided with links to accompanying repositories.

Estimation

GeoStats.Kriging — Type.

Kriging(var₁=>param₁, var₂=>param₂, ...)

A polyalgorithm Kriging estimation solver.

Each pair var=>param specifies the KrigingParam param for the Kriging variable var. In order to avoid boilerplate code, the constructor expects pairs of Symbol and NamedTuple instead.

Parameters

variogram - Variogram model (default to GaussianVariogram())

mean - Simple Kriging mean

degree - Universal Kriging degree

drifts - External Drift Kriging drift functions

Latter options override former options. For example, by specifying drifts, the user is telling the algorithm to ignore degree and mean. If no option is specified, Ordinary Kriging is used by default with the variogram only.

maxneighbors - Maximum number of neighbors (default to nothing)

neighborhood - Search neighborhood (default to nothing)

metric - Metric used to find nearest neighbors (default to Euclidean())

The maxneighbors option can be used to perform approximate Kriging with a subset of data points per estimation location. Two neighborhood search methods are available depending on the value of neighborhood:

If a neighborhood is provided, local Kriging is performed by sliding the neighborhood in the domain.

If neighborhood is not provided, the Kriging system is built using maxneighbors nearest neighbors according to a metric.

Examples

Solve the variable :var₁ with Simple Kriging by specifying the mean, and the variable :var₂ with Universal Kriging by specifying the degree and the variogram model.

julia> Kriging(
  :var₁ => (mean=1.,),
  :var₂ => (degree=1, variogram=SphericalVariogram(range=20.))
)

Solve all variables of the problem with the default parameters (i.e. Ordinary Kriging with unit Gaussian variogram):

julia> Kriging()

Simulation

GeoStats.SeqGaussSim — Type.

SeqGaussSim(var₁=>param₁, var₂=>param₂, ...)

A polyalgorithm sequential Gaussian simulation solver.

Each pair var=>param specifies the SeqGaussSimParam param for the simulation variable var. In order to avoid boilerplate code, the constructor expects pairs of Symbol and NamedTuple instead. See Kriging documentation for examples.

Parameters

variogram - Variogram model (default to GaussianVariogram())

mean - Simple Kriging mean

degree - Universal Kriging degree

drifts - External Drift Kriging drift functions

Latter options override former options. For example, by specifying drifts, the user is telling the algorithm to ignore degree and mean. If no option is specified, Ordinary Kriging is used by default with the variogram only.

maxneighbors - Maximum number of neighbors (default to 10)

neighborhood - Search neighborhood (default to nothing)

metric - Metric used to find nearest neighbors (default to Euclidean())

path - Simulation path (default to RandomPath)

For each location in the simulation path, a maximum number of neighbors maxneighbors is used to fit a Gaussian distribution. The nearest neighbors are searched according to a metric or according to a neighborhood when the latter is provided.

GeoStats.CookieCutter — Type.

CookieCutter(master, others)

A cookie-cutter simulation solver.

Parameters

master - Master simulation solver (a.k.a. facies solver)

others - A list of pairs mapping categories to solvers

Examples

Simulate lithology facies with image quilting and fill property with direct Gaussian simulation:

julia> fsolver  = ImgQuilt(:facies => (TI=Strebelle, template=(30,30,1)))
julia> psolver₀ = DirectGaussSim(:property => (variogram=SphericalVariogram(range=10.),))
julia> psolver₁ = DirectGaussSim(:property => (variogram=SphericalVariogram(range=20.),))
julia> solver   = CookieCutter(fsolver, [0 => psolver₀, 1 => psolver₁])