Spatial data

Spatial data

In GeoStats.jl, data and domain are disconnected one from another. This design enables levels of parallelism that would be difficult or even impossible to implement otherwise.

The data is accessed by solvers only if strictly necessary. One of our goals is to be able to handle massive datasets that may not fit in random-access memory (RAM).

GeoDataFrame

For point (or hard) data in spreadsheet format (e.g. CSV, TSV), the GeoDataFrame object is a lightweight wrapper over Julia's DataFrame types.

GeoStatsDevTools.GeoDataFrame — Type.

GeoDataFrame(data, coordnames)

A dataframe object data with additional metadata for tracking the columns coordnames that represent spatial coordinates.

Examples

If the data was already loaded in a normal DataFrame data, and there exists columns named x, y and z, wrap the data and specify the column names:

julia> GeoDataFrame(data, [:x,:y,:z])

Alternatively, load the data directly into a GeoDataFrame object by using the method readgeotable.

Notes

This type is a lightweight wrapper over Julia's DataFrame types. No additional storage is required other than a vector of symbols with the columns names representing spatial coordinates.

GeoStatsDevTools.readgeotable — Function.

readgeotable(args; coordnames=[:x,:y,:z], kwargs)

Read data from disk using CSV.read, optionally specifying the columns coordnames with spatial coordinates.

The arguments args and keyword arguments kwargs are forwarded to the CSV.read function, please check their documentation for more details.

This function returns a GeoDataFrame object.

PointSetData

The PointSetData object is equivalent to GeoDataFrame except that it stores the data in a simple Julia Dict instead of in a DataFrame.

GeoStatsDevTools.PointSetData — Type.

PointSetData(data, coords)

Spatial data georeferenced with coordinate matrix coords. The data argument is a dictionary mapping variable names to Julia arrays with the actual data.

See also: PointSet

RegularGridData

In the case that the data is regularly spaced in a grid, the GeoGridData object provides fast access across multiple overlaid images.

GeoStatsDevTools.RegularGridData — Type.

RegularGridData(data, origin, spacing)

Regularly spaced data georeferenced with origin and spacing. The data argument is a dictionary mapping variable names to Julia arrays with the actual data.

NaN or missing values in the Julia arrays are interpreted as non-valid. They can be used to mask the variables on the grid.

Examples

Given poro and perm two 2-dimensional Julia arrays containing values of porosity and permeability, the following code can be used to georeference the data:

julia> data = Dict(:porosity => poro, :permeability => perm)
julia> RegularGridData(data, [0.,0.,0.], [1.,1.,1.])

Alternatively, one can omit origin and spacing for default values of zeros and ones:

julia> RegularGridData{Float64}(data)

See also: RegularGrid

StructuredGridData

A StructuredGridData is a direct generalization of RegularGridData in which points can be localized in space with indices i,j,k... even though they are not regularly spaced. This format is often found in satellite data, NetCDF, etc.

GeoStatsDevTools.StructuredGridData — Type.

StructuredGridData(data, X, Y, Z, ...)

Data spatially distributed on a structured grid where points are georeferenced by coordinates X, Y, Z, ...

The data argument is a dictionary mapping variable names to Julia arrays with the actual data.

Examples

A very popular structured grid data format is NetCDF. Given 2D arrays LAT and LON with coordinates and arrays with climate data precipitation, temperature, the following code can be used to construct a structured grid:

julia> data = Dict(:precipitation => precipitation, :temperature => temperature)
julia> StructuredGridData(data, LAT, LON)