Theoretical variograms

Theoretical variograms

\[\newcommand{\x}{\boldsymbol{x}} \newcommand{\R}{\mathbb{R}} \newcommand{\1}{\mathbb{1}}\]

In an intrinsic isotropic model, the variogram is only a function of the distance between any two points $\x_1,\x_2 \in \R^m$:

\[\gamma(\x_1,\x_2) = \gamma(||\x_1 - \x_2||) = \gamma(h)\]

Under the additional assumption of 2nd-order stationarity, the well-known covariance is directly related via $\gamma(h) = cov(0) - cov(h)$. Anisotropic models are easily obtained by defining an ellipsoid distance in place of the Euclidean distance. For a list of available distances, please see Distances.jl.

This package implements a few commonly used and other more excentric variogram models. They all share the same default parameters:

sill=1

range=1

nugget=0

distance=Euclidean()

Some of them have extra parameters that can be set with keyword arguments:

GaussianVariogram(nugget=.1) # set nugget effect
MaternVariogram(order=1) # set order of Bessel function

Additionally, a composite (additive) variogram model $\gamma(h) = \gamma_1(h) + \gamma_2(h) + \cdots \gamma_n(h)$ can be constructed from a list of variogram models:

γ = GaussianVariogram() + ExponentialVariogram()

Like the other variogram models, a composite variogram $\gamma$ can be evaluated as an isotropic model $\gamma(h)$ or as a model with a custom distance $\gamma(\x_1,\x_2)$.

Finally, the 2nd-order stationarity property of a variogram can be checked with the isstationary method:

Variography.isstationary — Function.

isstationary(γ)

Check if variogram γ possesses the 2nd-order stationary property.

Gaussian

\[\gamma(h) = (s - n) \left[1 - \exp\left(-3\left(\frac{h}{r}\right)^2\right)\right] + n \cdot \1_{(0,\infty)}(h)\]

Variography.GaussianVariogram — Type.

GaussianVariogram(sill=s, range=r, nugget=n, distance=d)

A Gaussian variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Exponential

\[\gamma(h) = (s - n) \left[1 - \exp\left(-3\left(\frac{h}{r}\right)\right)\right] + n \cdot \1_{(0,\infty)}(h) \]

Variography.ExponentialVariogram — Type.

ExponentialVariogram(sill=s, range=r, nugget=n, distance=d)

An exponential variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Matern

\[\gamma(h) = (s - n) \left[1 - \frac{2^{1-\nu}}{\Gamma(\nu)} \left(\sqrt{2\nu}\frac{h}{r}\right)^\nu K_\nu\left(\sqrt{2\nu}\frac{h}{r}\right)\right] + n \cdot \1_{(0,\infty)}(h)\]

Variography.MaternVariogram — Type.

MaternVariogram(sill=s, range=r, nugget=n, order=ν, distance=d)

A Matérn variogram with sill s, range r and nugget n. The parameter ν is the order of the Bessel function. Optionally, use a custom distance d.

source

Spherical

\[\gamma(h) = (s - n) \left[\left(\frac{3}{2}\left(\frac{h}{r}\right) + \frac{1}{2}\left(\frac{h}{r}\right)^3\right) \cdot \1_{(0,r)}(h) + \1_{[r,\infty)}(h)\right] + n \cdot \1_{(0,\infty)}(h)\]

Variography.SphericalVariogram — Type.

SphericalVariogram(sill=s, range=r, nugget=n, distance=d)

A spherical variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Cubic

\[\gamma(h) = (s - n) \left[\left(7\left(\frac{h}{r}\right)^2 - \frac{35}{4}\left(\frac{h}{r}\right)^3 + \frac{7}{2}\left(\frac{h}{r}\right)^5 - \frac{3}{4}\left(\frac{h}{r}\right)^7\right) \cdot \1_{(0,r)}(h) + \1_{[r,\infty)}(h)\right] + n \cdot \1_{(0,\infty)}(h)\]

Variography.CubicVariogram — Type.

CubicVariogram(sill=s, range=r, nugget=n, distance=d)

A cubic variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Pentaspherical

\[\gamma(h) = (s - n) \left[\left(\frac{15}{8}\left(\frac{h}{r}\right) - \frac{5}{4}\left(\frac{h}{r}\right)^3 + \frac{3}{8}\left(\frac{h}{r}\right)^5\right) \cdot \1_{(0,r)}(h) + \1_{[r,\infty)}(h)\right] + n \cdot \1_{(0,\infty)}(h)\]

Variography.PentasphericalVariogram — Type.

PentasphericalVariogram

A pentaspherical variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Power

\[\gamma(h) = sh^a + n \cdot \1_{(0,\infty)}(h)\]

Variography.PowerVariogram — Type.

PowerVariogram(scaling=s, exponent=a, nugget=n, distance=d)

A power variogram with scaling s, exponent a and nugget n. Optionally, use a custom distance d.

source

Sine hole

\[\gamma(h) = (s - n) \left[1 - \frac{\sin(\pi h / r)}{\pi h / r}\right] + n \cdot \1_{(0,\infty)}(h)\]

Variography.SineHoleVariogram — Type.

SineHoleVariogram(sill=s, range=r, nugget=n, distance=d)

A sine hole variogram with sill s, range r and nugget n. Optionally, use a custom distance d.

source

Composite

Variography.CompositeVariogram — Type.

CompositeVariogram(γ₁, γ₂, ..., γₙ)

A composite (additive) model of variograms γ(h) = γ₁(h) + γ₂(h) + ⋯ + γₙ(h).