Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GeoStatsBase.AbstractDomain — Type.

AbstractDomain{T,N}

A spatial domain with N dimensions in which points are represented with coordinates of type T.

GeoStatsBase.AbstractEstimSolverComparison — Type.

AbstractEstimSolverComparison

A method for comparing estimation solvers.

GeoStatsBase.AbstractEstimationSolver — Type.

AbstractEstimationSolver

A solver for a geostatistical estimation problem.

GeoStatsBase.AbstractMapper — Type.

AbstractMapper

A mapping strategy for mapping spatial data onto a domain.

GeoStatsBase.AbstractProblem — Type.

AbstractProblem

A generic problem in geostatistics.

GeoStatsBase.AbstractSimSolverComparison — Type.

AbstractSimSolverComparison

A method for comparing simulation solvers.

GeoStatsBase.AbstractSimulationSolver — Type.

AbstractSimulationSolver

A solver for a geostatistical simulation problem.

GeoStatsBase.AbstractSolution — Type.

AbstractSolution

A generic solution to a problem in geostatistics.

GeoStatsBase.AbstractSolver — Type.

AbstractSolver

A solver for geostatistical problems.

GeoStatsBase.AbstractSolverComparison — Type.

AbstractSolverComparison

A method for comparing geostatistical solvers.

GeoStatsBase.AbstractSpatialData — Type.

AbstractSpatialData{T,N}

Spatial data distributed in a N-dimensional space using coordinates of type T.

GeoStatsBase.CopyMapper — Type.

CopyMapper

A mapping strategy in which data points are copied directly to the domain at specified locations.

GeoStatsBase.EstimationProblem — Type.

EstimationProblem(spatialdata, domain, targetvars)

A spatial estimation problem on a given domain in which the variables to be estimated are listed in targetvars. The data of the problem is stored in spatialdata.

Examples

Create an estimation problem for rainfall precipitation measurements:

julia> EstimationProblem(spatialdata, domain, :precipitation)

Create an estimation problem for precipitation and CO₂:

julia> EstimationProblem(spatialdata, domain, [:precipitation, :CO₂])

GeoStatsBase.EstimationSolution — Type.

EstimationSolution

A solution to a spatial estimation problem.

GeoStatsBase.SimpleMapper — Type.

SimpleMapper

A mapping strategy in which data points are assigned to their nearest point in the domain.

GeoStatsBase.SimulationProblem — Type.

SimulationProblem(spatialdata, domain, targetvars, nreals)
SimulationProblem(domain, targetvars, nreals)

A spatial simulation problem on a given domain in which the variables to be simulated are listed in targetvars.

For conditional simulation, the data of the problem is stored in spatialdata.

For unconditional simulation, a dictionary targetvars must be provided mapping variable names to their types.

In both cases, a number nreals of realizations is requested.

Examples

Create a conditional simulation problem for porosity and permeability with 100 realizations:

julia> SimulationProblem(spatialdata, domain, [:porosity,:permeability], 100)

Create an unconditional simulation problem for porosity and facies type with 100 realizations:

julia> SimulationProblem(domain, Dict(:porosity => Float64, :facies => Int), 100)

Notes

To check if a simulation problem has data (i.e. conditional vs. unconditional) use the hasdata method.

GeoStatsBase.SimulationSolution — Type.

SimulationSolution

A solution to a spatial simulation problem.

GeoStatsBase.compare — Method.

compare(solvers, problem, method)

Compare solvers on a given problem using a comparison method.

GeoStatsBase.coordinates! — Method.

coordinates!(buff, domain, location)

Non-allocating version of coordinates.

GeoStatsBase.coordinates! — Method.

coordinates!(buff, spatialdata, ind)

Non-allocating version of coordinates.

GeoStatsBase.coordinates! — Method.

coordinates!(buff, domain, locations)

Non-allocating version of coordinates

GeoStatsBase.coordinates! — Method.

coordinates!(buff, spatialdata, indices)

Non-allocating version of coordinates

GeoStatsBase.coordinates — Method.

coordinates(domain)

Return the coordinates of all locations in domain.

GeoStatsBase.coordinates — Method.

coordinates(spatialdata)

Return the coordinates of all indices in spatialdata.

GeoStatsBase.coordinates — Method.

coordinates(domain, locations)

Return the coordinates of locations in the domain.

GeoStatsBase.coordinates — Method.

coordinates(domain, location)

Return the coordinates of the location in the domain.

GeoStatsBase.coordinates — Method.

coordinates(spatialdata, inds)

Return the coordinates of inds in the spatialdata.

GeoStatsBase.coordinates — Method.

coordinates(spatialdata, ind)

Return the coordinates of the ind-th point in spatialdata.

GeoStatsBase.coordnames — Method.

coordnames(spatialdata)

Return the name of the coordinates in spatialdata and their types.

GeoStatsBase.coordtype — Method.

coordtype(domain)

Return the coordinate type of a spatial domain.

GeoStatsBase.coordtype — Method.

coordtype(spatialdata)

Return the coordinate type of a spatial data.

GeoStatsBase.data — Method.

data(problem)

Return the spatial data of the estimation problem.

GeoStatsBase.data — Method.

data(problem)

Return the spatial data of the simulation problem.

GeoStatsBase.datamap — Method.

datamap(problem, targetvar)

Return the mapping from domain locations to data locations for the targetvar of the problem.

GeoStatsBase.datamap — Method.

datamap(problem, targetvar)

Return the mapping from domain locations to data locations for the targetvar of the problem.

GeoStatsBase.digest — Method.

digest(solution)

Convert solution to a dictionary-like format where the keys of the dictionary are the variables of the problem.

GeoStatsBase.digest — Method.

digest(solution)

Convert solution to a dictionary-like format where the keys of the dictionary are the variables of the problem.

GeoStatsBase.domain — Method.

domain(problem)

Return the spatial domain of the estimation problem.

GeoStatsBase.domain — Method.

domain(solution)

Return the domain of the estimation solution.

GeoStatsBase.domain — Method.

domain(problem)

Return the spatial domain of the simulation problem.

GeoStatsBase.domain — Method.

domain(solution)

Return the domain of a simulation solution.

GeoStatsBase.hasdata — Method.

hasdata(problem)

Return true if simulation problem has data.

GeoStatsBase.mapper — Method.

mapper(problem)

Return the mapper of the estimation problem.

GeoStatsBase.mapper — Method.

mapper(problem)

Return the mapper of the simulation problem.

GeoStatsBase.nearestlocation — Method.

nearestlocation(domain, coords)

Return the nearest location of coords in the domain.

GeoStatsBase.npoints — Method.

npoints(domain)

Return the number of points of a spatial domain.

GeoStatsBase.npoints — Method.

npoints(spatialdata)

Return the number of points in spatialdata.

GeoStatsBase.nreals — Method.

nreals(problem)

Return the number of realizations of the simulation problem.

GeoStatsBase.preprocess — Method.

preprocess(problem, solver)

Preprocess the simulation problem once before generating each realization with simulation solver.

Notes

The output of the function is defined by the solver developer. Default implementation returns nothing.

GeoStatsBase.solve — Method.

solve(problem, solver)

Solve the problem with the solver.

GeoStatsBase.solve — Method.

solve(problem, solver)

Solve the simulation problem with the simulation solver.

Notes

Default implementation calls solve_single in parallel.

GeoStatsBase.solve_single — Method.

solve_single(problem, var, solver, preproc)

Solve a single realization of var in the simulation problem with the simulation solver, optionally using preprocessed data in preproc.

Notes

By implementing this function instead of solve, the developer is informing the framework that realizations generated with his/her solver are indenpendent one from another. GeoStats.jl will trigger the algorithm in parallel (if enough processes are available).

GeoStatsBase.valid — Method.

valid(spatialdata, var)

Return all points in spatialdata with a valid value for var. The output is a tuple with the matrix of coordinates as the first item and the vector of values as the second item.

GeoStatsBase.value — Method.

value(spatialdata, ind, var)

Return the value of var for the ind-th point in spatialdata.

GeoStatsBase.valuetype — Method.

valuetype(spatialdata, var)

Return the value type of var in spatialdata.

GeoStatsBase.variables — Method.

variables(spatialdata)

Return the variable names in spatialdata and their types.

GeoStatsBase.variables — Method.

variables(problem)

Return the variable names of the estimation problem and their types.

GeoStatsBase.variables — Method.

variables(problem)

Return the target variables of the simulation problem and their types.

Base.isvalid — Method.

isvalid(spatialdata, ind, var)

Return true if the ind-th point in spatialdata has a valid value for var.

Base.map — Method.

map(spatialdata, domain, targetvars, mapper)

Map the targetvars in spatialdata to domain with mapper.

Base.ndims — Method.

ndims(domain)

Return the number of dimensions of a spatial domain.

Base.ndims — Method.

ndims(spatialdata)

Return the number of dimensions of domain underlying spatialdata.

Base.view — Method.

view(spatialdata, inds)

Return a view of spatialdata with all points in inds.