Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Geodesy.Datum — Type.

Abstract type for geodetic datums

A datum is a set of reference objects and assigned coordinates, relative to which other objects may be positioned. We model these in code with subtypes of Datum. Each geodetic datum has an associated ellipsoid model of the Earth which is required when transforming between coordinate systems. The ellipsoid can be accessed with the ellipsoid() function.

Geodesy.ECEF — Type.

ECEF(x, y, z)

Earth-Centered-Earth-Fixed (ECEF) coordinates. A global Cartesian coordinate system rotating with the Earth.

Geodesy.ECEFfromENU — Type.

ECEFfromENU(origin, datum)
ECEFfromENU(origin::UTM, zone, isnorth, datum)
ECEFfromENU(origin::ECEF, lat, lon)

Construct a Transformation object to convert from local ENU coordinates centred at origin to global ECEF coodinates. This object pre-caches both the ECEF coordinates and latitude and longitude of the origin for maximal efficiency.

Geodesy.ECEFfromLLA — Type.

ECEFfromLLA(ellipsoid_or_datum)

Construct a Transformation object to convert from LLA coordinates to ECEF coordinates.

Geodesy.ENU — Type.

ENU(e, n, u = 0.0)

East-North-Up (ENU) coordinates. A local Cartesian coordinate system, linearized about a reference point.

Geodesy.ENUfromECEF — Type.

ENUfromECEF(origin, datum)
ENUfromECEF(origin::UTM, zone, isnorth, datum)
ENUfromECEF(origin::ECEF, lat, lon)

Construct a Transformation object to convert from global ECEF coordinates to local ENU coordinates centered at the origin. This object pre-caches both the ECEF coordinates and latitude and longitude of the origin for maximal efficiency.

Geodesy.Ellipsoid — Type.

An ellipsoidal representation of the Earth, for converting between LLA and other co-ordinate systems such as ECEF.

Geodesy.GDA94 — Type.

GDA94 - Geocentric Datum of Australia, 1994

Geodesy.ITRF — Type.

ITRF{Year}([epoch])

Construct an object representing the International Terrestrial Reference Frame for the given Year of realization. ITRF is the standard high accuracy terrestrial reference frame for worldwide use. An optional epoch parameter defines the time of interest - typically a date at which coordinates were measured, using, eg a GPS device. Without the epoch parameter, the resulting ITRF{Year} object represents the full dynamic datum.

A realization is created every few years by computing the position of a large set of ground control stations from satellite and celestial measurements. The Year parameter represents the last year from which data was used in the frame processing regression problem. A list of recent realizations is available at http://itrf.ensg.ign.fr/ITRF_solutions; as of 2016-07 valid Years were 2014 2008 2005 2000 1997 1996 1994 1993 1992 1991 1990 1989 1988.

See http://itrf.ensg.ign.fr/general.php for a technical overview. Useful technical papers:

"IERS Conventions (2010)", Petit and Luzum (eds.), IERS Technical note No.36, Chapter 4, https://www.iers.org/IERS/EN/Publications/TechnicalNotes/tn36.html

"ITRF2008: an improved solution of the international terrestrial reference frame", Altamimi et al., J. Geodesy (2011) 85: 457, http://dx.doi.org/10.1007/s00190-011-0444-4

"IGS08: the IGS realization of ITRF2008", Rebischung et al., GPS Solutions (2012) 16: 483, http://dx.doi.org/10.1007/s10291-011-0248-2, ftp://igs.org/pub/resource/pubs/IGS08TheIGSRealizationof_ITRF2008.pdf

"The IGS contribution to ITRF2014", Rebischung et al., J. Geodesy (2016) 90: 611, http://dx.doi.org/10.1007/s00190-016-0897-6

Geodesy.LLA — Type.

LLA(lat, lon, alt = 0.0)
LLA(lat = ϕ, lon = Θ, alt = h)

Latitude, longitude, and alititude co-ordinates. Note: assumes degrees not radians

Geodesy.LLAfromECEF — Method.

LLAfromECEF(ellipsoid)

Construct a Transformation object to convert from ECEF coordinates to LLA coordinates. Pre-caches ellipsoidal parameters for efficiency.

Geodesy.LLAfromUTM — Type.

LLAfromUTM(zone, isnorth::Bool, datum)

Construct a Transformation object to convert from UTM coordinates in the specified zone and hemisphere (isnorth = true or false) to global LLA coordinates. Pre-caches ellipsoidal parameters for efficiency and performs Charles Karney's accurate 6th-order series expansion algorithm.

(See also LLAfromUTMZ)

Geodesy.LLAfromUTMZ — Type.

LLAfromUTMZ(datum)

Construct a Transformation object to convert from global UTMZ coordinates to global LLA coordinates. Pre-caches ellipsoidal parameters for efficiency and performs Charles Karney's accurate 6th-order series expansion algorithm.

(See also LLAfromUTM)

Geodesy.LatLon — Type.

LatLon(lat, lon)
LatLon(lat = ϕ, lon = Θ)

Latitude and longitude co-ordinates. Note: assumes degrees not radians

Geodesy.NAD27 — Type.

NAD27 - North American Datum of 1927

Geodesy.NAD83 — Type.

NAD83 - North American Datum of 1983

For technical details, see "NAD83 (NSRS2007) National Readjustment Final Report" http://www.ngs.noaa.gov/PUBS_LIB/NSRS2007/NOAATRNOSNGS60.pdf

Geodesy.OSGB36 — Type.

OSGB36 - Datum for Ordinance Survey of Great Britain, 1936

Geodesy.UTM — Type.

UTM(x, y, z = 0.0)

Universal transverse Mercator (UTM) coordinates. Common projection type for world points. Zone not included in coordinates - it is a parameterized in the relavant transformations UTMfromLLA and LLAfromUTM (see also the UTMZ type).

This type may be used to parameterize UPS coordinates (Universal Polar Stereographic) to accurately represent the polar regions, in zone "0".

Geodesy.UTMZ — Type.

UTMZ(x, y, z = 0.0, zone::Integer, hemisphere::Bool)

Universal transverse Mercator (UTM) coordinates with zone number. Common projection type for world points. The UTM zone is included in coordinates (see also the UTM type).

This type may be used to parameterize UPS coordinates (Universal Polar Stereographic) to accurately represent the polar regions, in zone "0".

Geodesy.UTMZfromLLA — Type.

UTMZfromLLA(datum)

Construct a Transformation object to convert from global LLA coordinates to global UTMZ coordinates. The zone and hemisphere is automatically calculated following the standard definitions (including exceptions in Norway). Pre-caches ellipsoidal parameters for efficiency and performs Charles Karney's accurate 6th-order series expansion algorithm.

(See also UTMfromLLA)

Geodesy.UTMZfromUTM — Type.

UTMZfromUTM(zone, isnorth, datum)

Transformation to append the UTM/UPS zone and hemisphere to a raw UTM data point.

Geodesy.UTMfromLLA — Type.

UTMfromLLA(zone, isnorth::Bool, datum)

Construct a Transformation object to convert from global LLA coordinates to UTM coordinates in the specified zone and hemisphere (isnorth = true or false). Pre-caches ellipsoidal parameters for efficiency and performs Charles Karney's accurate 6th-order series expansion algorithm.

(See also UTMZfromLLA)

Geodesy.UTMfromUTMZ — Type.

UTMfromUTMZ(zone, isnorth, datum)

Transformation to remove the zone and hemisphere from UTMZ data point, and automatically convert the data to the specified zone if necessary.

Geodesy.WGS84 — Type.

WGS84()

Construct an object representing the World Geodetic System of 1984, as a dynamic datum - see below for gory details about specific WGS84 frames).

If you're getting positions from a consumer GPS device, you're probably going to have WGS84 by default because it's the datum in which GPS satellites broadcast their position ("broadcast ephemerides"). Note however that many devices can also provide position in a national datum, so you should check your device settings to be sure.

As a special case for low accuracy work (worse than a meter or so), WGS84 will assume that coordinates supplied without a capture time are in the latest frame realization known to Geodesy.jl, WGS84 (G1762). Note that this may not be correct if you're processing historical data, or Geodesy.jl itself is out of date. For higher accuracy, you should supply a date of capture - either with each coordinate, or explicitly using the GpsWeek parameter to the WGS84 type:

WGS84{GpsWeek}()

Construct an object representing the static WGS84 datum computed using data gathered prior to the given GpsWeek. WGS84 is maintained and updated by the US National Geospatial-Intelligence-Agency (NGA) at irregular intervals to align with the ITRF to within 0.1m (see [1]); if you care about accuracy at that level, you probably want to be solving for position in a different datum, for example, ITRF. As of 2016, GpsWeek should be one out of [0, 730, 873, 1150, 1674, 1762].

Note that the dates of implementation of these frames as broadcast by the satellites are not the same as the associated GPS week - see Ref. [1], table 2.1. (TODO: Perhaps Geodesy should have a table to figure out which frame to use at a given date? Does anybody care?)

"World Geodetic System 1984", NGA standard NGA.STND.00361.0.0WGS84, 2014-07-08, http://earth-info.nga.mil/GandG/publications/NGASTND0036100WGS84/NGA.STND.00361.0.0WGS84.pdf

Geodesy.ECEFfromUTM — Method.

ECEFfromUTM(zone, isnorth, datum)

Creates composite transformation ECEFfromLLA(datum) ∘ LLAfromUTM(zone, isnorth, datum).

Geodesy.ECEFfromUTMZ — Method.

ECEFfromUTMZ(datum)

Creates composite transformation ECEFfromLLA(datum) ∘ LLAfromUTMZ(datum).

Geodesy.ENUfromLLA — Method.

ENUfromLLA(origin, datum)

Creates composite transformation ENUfromECEF(origin, datum) ∘ ECEFfromLLA(datum).

Geodesy.ENUfromUTM — Method.

ENUfromUTM(origin, zone, isnorth, datum)

Creates composite transformation UTMfromLLA(zone, isnorth, datum) ∘ LLAfromENU(origin, datum). If origin is a UTM point, then it is assumed it is in the given specified zone and hemisphere.

Geodesy.ENUfromUTMZ — Method.

ENUfromUTMZ(origin, datum)

Creates composite transformation ENUfromLLA(origin, datum) ∘ LLAfromUTMZ(datum).

Geodesy.LLAfromENU — Method.

LLAfromENU(origin, datum)

Creates composite transformation LLAfromECEF(datum) ∘ ECEFfromENU(origin, datum).

Geodesy.UTMZfromECEF — Method.

UTMZfromECEF(datum)

Creates composite transformation UTMZfromLLA(datum) ∘ LLAfromECEF(datum).

Geodesy.UTMZfromENU — Method.

UTMZfromENU(origin, datum)

Creates composite transformation UTMZfromLLA(datum) ∘ LLAfromENU(origin, datum).

Geodesy.UTMfromECEF — Method.

UTMfromECEF(zone, isnorth, datum)

Creates composite transformation UTMfromLLA(zone, isnorth, datum) ∘ LLAfromECEF(datum).

Geodesy.UTMfromENU — Method.

UTMfromENU(origin, zone, isnorth, datum)

Creates composite transformation UTMfromLLA(zone, isnorth, datum) ∘ LLAfromENU(origin, datum). If origin is a UTM point, then it is assumed it is in the given specified zone and hemisphere.

Geodesy.datum_shift_ECEF — Method.

datum_shift_ECEF(dest_datum, source_datum)

Return a transformation object which transforms ECEF points from source_datum to dest_datum. The function should attempt to supply the current publicly accepted best estimate.

Note that the best known version of a datum transformation will inherently improve with time (see below), so we cannot simultaneously guarantee that:

We return the best publicly accepted version of a datum shift.

We return the same thing in future versions of Geodesy.jl.

This function will attempt to satisfy condition 1. rather than 2.; if you want a stable version of a transformation you should use one of the lower level functions, for example, GDA94_from_ITRF_Dawson2010().

Important note about accuracy

If you care about accuracy, you should always store long term archival data in the source datum where possible, along with metadata defining the full datum and coordinate system in use. The time to do a datum shift is when you want to compare information from two different datums.

Why all this bother about inaccuracy? The errors are twofold: First, the parameters of a datum shift come from a physical measurement process. Typically this involves measuring the coordinates of some physical locations in both datums, and a physical measurement procedure is always subject to some inaccuracy. Second, these tie points are used to infer a compact representation of the datum shift, with as few numerical parameters as possible. A small number of parameters will result in an overly smooth representation; this modelling error is a second source of inaccuracy. Both these errors can be reduced if you choose to measure more tie points or improve the complexity of the numerical model at a future date.

Geodesy.distance — Function.

distance(utm1, utm2, zone, isnorth, [datum = wgs84])
distance(a, utm2, zone, isnorth, [datum = wgs84])
distance(utm1, b, zone, isnorth, [datum = wgs84])

If one or both points are UTM, we need the zone (and particularly the hemisphere, isnorth = true/false) to determine the Cartesian distance.

Geodesy.distance — Method.

distance(a, b, [datum = wgs84])

The Cartesian distance between points a and b. Uses datum to perform transformations as necessary, and unlike other Geodesy functions, this defaults to use the common WGS-84 datum for convenience.

Geodesy.utm_zone — Method.

(zone, isnorth) = utm_zone(lat, lon)
(zone, isnorth) = utm_zone(ll::LatLon)
(zone, isnorth) = utm_zone(lla::LLA)
(zone, isnorth) = utm_zone(ecef::ECEF, datum)

Find the UTM zone and hemisphere (isnorth = true or false) for the given latitude and longitude (or world point), including the special rules for Norway and Svalbard. Zone 0 corresponds to the poles, using the UPS regions.

Geodesy.TransverseMercator — Type.

immutable TransverseMercator{MaxPow}
TransverseMercator(datum)
TransverseMercator(ellipsoid)
TransverseMercator(a, f, [::Type{Val{MaxPow}} = Val{6}])

Cache of ellipsoidal calculations necessary for transverse-Mercator and polar-stereographic transformations. Series expansion coefficients up to order MaxPow (between 4 and 8, default 6) are calculated and stored for fast transverse-Mercator and UTM calculations.

Geodesy.AngNormalize — Method.

%AngNormalize: Reduce angle to range [-180, 180) % % x = AngNormalize(x) reduces angles to the range [-180, 180).

Geodesy.GDA94_from_ITRF_Dawson2010 — Method.

GDA94_from_ITRF_Dawson2010(ITRF_year, epoch)

Return a Transformation converting ECEF points from a given ITRF to GDA94. Datum shift parameters are taken from [1], supporting ITRF_year 2008, 2005, 2000, 1997, 1996. epoch is the Date (or DateTime) of interest at which the input ECEF coordinates to the transformation were measured in ITRF.

[1] J. Dawson and A. Woods, "ITRF to GDA94 coordinate transforms", Journal of Applied Geodesy, 4, p. 189 (2010).

TODO: We don't yet support epoch varying per input point, but there should be a Transformation object for this at some stage.

Geodesy.LatFix — Method.

LatFix(x) returns x is it is in the range [-90, 90]; otherwise it returns NaN.

Geodesy.eatanhe — Method.

eatanhe(x, e2) returns eatanh(ex) where e = sqrt(e2) e2 is a scalar; x can be any shape.

Geodesy.polarst_fwd — Method.

(x, y, gamma, k) = polarst_fwd(northpole::Bool, k0::Float64, tm::TransverseMercator, lat, lon)

Perform polar-stereographic projection of lat and lon with respect to north or south pole northpole and horizontal scaling k0 (= 0.994 for UPS). γ and k are the local convergence and scaling factors, respectively.

Geodesy.polarst_inv — Method.

(lat, lon, gamma, k) = polarst_inv(northp::Bool, k0::Float64, tm::TransverseMercator, x, y)

Invert polar-stereographic projection of x and y with respect to north or south pole northpole and horizontal scaling k0 (= 0.994 for UPS). γ and k are the local convergence and scaling factors, respectively.

Geodesy.polyval — Method.

Evaluate polynomial with coefficients p[i] at x

Geodesy.tauf — Method.

%TAUF tan(phi) % % TAUF(taup, e2) returns tangent of phi in terms of taup the tangent of % chi. e2, the square of the eccentricity, is a scalar; taup can be any % shape.

Geodesy.taupf — Method.

taupf(tau, e2) returns tangent of chi in terms of tau the tangent of phi. e2, the square of the eccentricity, is a scalar; tau can be any shape.

Geodesy.transverse_mercator_forward — Method.

(x, y, γ, k) = transverse_mercator_forward(lon0, lat, lon, k0, tm::TransverseMercator)

Perform transverse-Mercator projection of lat and lon with respect to reference meridian lat0 and horizontal scaling k0 (= 0.9996 for UTM) using a series expansion approach (see TransverseMercator). γ and k are the local convergence and scaling factors, respectively.

Geodesy.transverse_mercator_reverse — Method.

(lat, lon, γ, k) = transverse_mercator_reverse(lon0, x, y, k0, tm::TransverseMercator)

Invert transverse-Mercator projection of x and y with respect to reference meridian lat0 and horizontal scaling k0 (= 0.9996 for UTM) using a series expansion approach (see TransverseMercator). γ and k are the local convergence and scaling factors, respectively.

Geodesy.utm_meridian — Method.

utm_meridian(zone)

Central meridian of the given UTM zone (note - does not include the conventiional 500km false easting offset).