Functions

Functions

Functions exported from GeophysicalFlows:

Functions exported from TwoDTurb:

GeophysicalFlows.TwoDTurb.Problem — Method.

Problem(; parameters...)

Construct a 2D turbulence problem.

GeophysicalFlows.TwoDTurb.dissipation — Method.

dissipation(prob)

Returns the domain-averaged dissipation rate. nnu must be >= 1.

GeophysicalFlows.TwoDTurb.drag — Method.

drag(prob)

Returns the extraction of domain-averaged energy by drag/hypodrag mu.

GeophysicalFlows.TwoDTurb.energy — Method.

energy(prob)

Returns the domain-averaged kinetic energy in the Fourier-transformed vorticity solution sol.

GeophysicalFlows.TwoDTurb.enstrophy — Method.

enstrophy(prob)

Returns the domain-averaged enstrophy in the Fourier-transformed vorticity solution sol.

GeophysicalFlows.TwoDTurb.set_q! — Method.

set_q!(prob, q)

Set the solution sol as the transform of q and update variables v on the grid g.

GeophysicalFlows.TwoDTurb.updatevars! — Method.

updatevars!(prob)

Update the vars in v on the grid g with the solution in sol.

GeophysicalFlows.TwoDTurb.work — Method.

work(prob)
work(sol, v, g)

Returns the domain-averaged rate of work of energy by the forcing Fh.

Functions exported from BarotropicQG:

GeophysicalFlows.BarotropicQG.dissipation — Method.

dissipation(prob)
dissipation(s, v, p, g)

Returns the domain-averaged dissipation rate. nnu must be >= 1.

GeophysicalFlows.BarotropicQG.drag — Method.

drag(prob)
drag(s, v, p, g)

Returns the extraction of domain-averaged energy by drag mu.

GeophysicalFlows.BarotropicQG.energy — Method.

Calculate the domain-averaged kinetic energy.

GeophysicalFlows.BarotropicQG.enstrophy — Method.

Returns the domain-averaged enstrophy.

GeophysicalFlows.BarotropicQG.meanenergy — Method.

Returns the energy of the domain-averaged U.

GeophysicalFlows.BarotropicQG.meanenstrophy — Method.

Returns the enstrophy of the domain-averaged U.

GeophysicalFlows.BarotropicQG.set_zeta! — Method.

set_zeta!(prob, zeta)
set_zeta!(s, v, g, zeta)

Set the solution s.sol as the transform of zeta and update variables v on the grid g.

GeophysicalFlows.BarotropicQG.updatevars! — Method.

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in s.sol.

GeophysicalFlows.BarotropicQG.work — Method.

work(prob)
work(s, v, p, g)

Returns the domain-averaged rate of work of energy by the forcing Fqh.