Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Git.attached — Method.

Git.attached(; dir="")

Determine whether HEAD is attached to a commit in the given respository.

Git.branch — Method.

Git.branch(; dir="")

Return the name of the current active branch in the given repository.

Git.cmd — Method.

Git.cmd(args; dir="")

Return a Git command from the given arguments, acting on the repository given in dir.

Git.different — Method.

Git.different(verA, verB, path; dir="")

Determine whether two trees are different with respect to the given path.

Git.dir — Method.

Git.dir(d)

Return the path to the default .git for the given repository directory, or the path to use in place of the default .git.

Git.dirty — Method.

Git.dirty([paths]; dir="")

Determine whether the paths in the given repository are dirty, i.e. contain modified but uncommitted tracked files.

Git.git — Method.

Git.git([d])

Return a Git command that refers to the work tree and directory given by d, or the current work tree and directory if d is not specified.

Git.head — Method.

Git.head(; dir="")

Return the commit to which HEAD currently refers.

Git.is_ancestor_of — Method.

Git.is_ancestor_of(a, b; dir="")

Determine whether the commit a is an ancestor of the commit b in the given repository.

Git.iscommit — Method.

Git.iscommit(name; dir="")

Determine whether name refers to a commit in the repository dir. name can be a single SHA1 or a vector of SHA1s.

Git.modules — Method.

Git.modules(args; dir="")

Apply the Git command with the given arguments on the given repository to the configuration file .gitmodules and read the result as a string.

Git.normalize_url — Method.

Git.normalize_url(url)

Normalize the given URL to a valid GitHub repository URL.

Git.readchomp — Method.

Git.readchomp(args; dir="")

Read the result of the Git command using the given arguments on the given repository as a string, removing a single trailing newline if present.

Git.readstring — Method.

Git.readstring(args; dir="")

Read the result of the Git command using the given arguments on the given repository as a string.

Git.restore — Method.

Git.restore(s::State; dir="")

Restore the given repository to the state s.

Git.run — Method.

Git.run(args; dir="", out=stdout)

Execute the Git command from the given arguments args on the repository dir, writing the results to the output stream out.

Git.set_remote_url — Method.

Git.set_remote_url(url; remote="origin", dir="")

Add a remote remote to the given repository from the URL url.

Git.snapshot — Method.

Git.snapshot(; dir="")

Return a State object that captures a snapshot of the given repository.

Git.staged — Method.

Git.staged([paths]; dir="")

Determine whether the paths in the given repository contain staged files.

Git.success — Method.

Git.success(args; dir="")

Determine whether the Git command using the given arguments on the given repository executed successfully.

Git.transact — Method.

Git.transact(f; dir="")

Attempt to execute the function f. If this fails, the repository is restored to its state prior to execution.

Git.unstaged — Method.

Git.unstaged([paths]; dir="")

Determine whether the paths in the given repository contain unstaged files.

Git.version — Method.

Git.version()

Return the version of Git being used by the package.