Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Base.size — Method.

size(wv)

Return the word vector length and the number of words as a tuple.

Glowe.analogy — Method.

analogy(wv, pos, neg, n=5)

Compute the analogy similarity between two lists of words. The positions and the similarity values of the top n similar words will be returned. For example, king - man + woman = queen will be pos=["king", "woman"], neg=["man"].

Glowe.analogy_words — Function.

analogy_words(wv, pos, neg, n=5)

Return the top n words computed by analogy similarity between positive words pos and negaive words neg. from the WordVectors wv.

Glowe.cooccur — Method.

cooccur(corpus, vocab, cooccurrences; verbose=2, symmetric=0, window_size=15, memory=4.0, max_product=nothing, overflow_length=nothing, overflow_file="overflow", distance_weighting=1)

Calculates word-word cooccurrence statistics.

Arguments

corpus::AbstractString input corpus file path

vocab::AbstractString input vocabulary file path (the vocabulary contains truncated unigram counts, produced by vocab_count)

cooccurrences::AbstractString output cooccurrences file path

Keyword arguments

verbose::Int set verbosity: 0, 1, 2 (default) or 3

symmetric::Int if <int> = 0, only use left context; if <int> = 1 (default), use left and right

window_size::Int number of context words to the left (and to the right, if symmetric = 1); default 15

memory::Float64 soft limit for memory consumption, in GB – based on simple heuristic, so not extremely accurate; default 4.0

max_product::Union{Nothing, Int} limit the size of dense cooccurrence array by specifying the max product <int> of the frequency counts of the two cooccurring words. This value overrides that which is automatically produced by memory. Typically only needs adjustment for use with very large corpora

overflow_length::Union{Nothing, Int} limit to length <int> the sparse overflow array, which buffers cooccurrence data that does not fit in the dense array, before writing to disk. This value overrides that which is automatically produced by memory. Typically only needs adjustment for use with very large corpora

overflow_file::String filename, excluding extension, for temporary files; default "overflow"

distance_weighting::Int if <int> = 0, do not weight cooccurrence count by distance between words; if <int> = 1 (default), weight the cooccurrence count by inverse of distance between words"

Examples

# It is assumed that vocab.txt exists and has been created by `vocab_count`
julia> cooccur("corpus.txt", "vocab.txt", "cooccurrences.bin", verbose=2, symmetric=0, window_size=10, memory=8.0, overflow_file="tempoverflow")

Glowe.cosine — Function.

cosine(wv, word, n=10)

Return the position of n (by default n = 10) neighbors of word and their cosine similarities.

Glowe.cosine_similar_words — Function.

cosine_similar_words(wv, word, n=10)

Return the top n (by default n = 10) most similar words to word from the WordVectors wv.

Glowe.get_vector — Method.

get_vector(wv, word [; oov=false, oov_key="<unk>")

Returns the vector representation of word from the WordVectors wv. If oov is true, the vector corresponding to the key oov_key is returned for out-of-vocabulary words.

Glowe.in_vocabulary — Method.

in_vocabulary(wv, word)

Return true if word is part of the vocabulary of the WordVector wv and false otherwise.

Glowe.index — Method.

index(wv, word)

Return the index of word from the WordVectors wv.

Glowe.shuffle — Method.

shuffle(cooccurences, shuffled; verbose=2, memory=4.0, array_size=nothing, temp_file="temp_shuffle")

Shuffles entries of word-word cooccurrence files.

Arguments

cooccurences::AbstractString input cooccurrences file path

shuffled::AbstractString output shuffled cooccurences file path

Keyword arguments

verbose::Int set verbosity: 0, 1, or 2 (default)

memory::Float64 soft limit for memory consumption, in GB; default 4.0

array_size::Union{Nothing, Int} limit to length <int> the buffer which stores chunks of data to shuffle before writing to disk. This value overrides that which is automatically produced by memory

temp_file::String filename, excluding extension, for temporary files; default "temp_shuffle"

Examples

# It is assumed that cooccurences.bin exists and has been created by `cooccur`
julia> shuffle("cooccurences.bin", "shuffled.bin", verbose=2, memory=8.0)

Glowe.similarity — Method.

similarity(wv, word1, word2)

Return the cosine similarity value between two words word1 and word2.

Glowe.vocab_count — Method.

vocab_count(corpus, vocab; max_vocab=100_000, min_count=10, verbose=2)

Extracts unigram counts.

Arguments

corpus::AbstractString input corpus file path

vocab::AbstractString output vocabulary file path

Keyword arguments

verbose::Int set verbosity: 0, 1 or 2 (default)

max_vocab::Int upper bound on vocabulary size, i.e. keep the <int> most frequent words. The minimum frequency words are randomly sampled so as to obtain an even distribution over the alphabet.

min_count::Int lower limit such that words which occur fewer than <int> times are discarded.

Examples

julia> vocab_count("corpus.txt", "vocab.txt", verbose=2, max_vocab=100_000, min_count=10)

Glowe.vocabulary — Method.

vocabulary(wv)

Return the vocabulary as a vector of words of the WordVectors wv.

Glowe.wordvectors — Method.

wordvectors(filename [,type=Float64][; kind=:text, header=false, normalize=true, vocabulary=nothing])

Generate a WordVectors type object from a file.

Arguments

filename::AbstractString the embeddings file name

type::Type type of the embedding vector elements; default Float64

Keyword arguments

kind::Symbol specifies whether the embeddings file is textual (:text)

or binary (:binary); default :text

header::Union{Nothing, Bool} in text embeddings files specifies

whether the file contains a header i.e. number of lines, columns or not. If the header is nothing, the loader will attempt to autodetect the presence of a header; default nothing

normalize:Bool specifies whether to normalize the embedding vectors

i.e. return unit vectors; default true

vocabulary::Union{Nothing, AbstractString} path to the vocabulary

file generated with vocab_count (needed for binary embeddings); default nothing

load_bias::Bool specifies whether to load the bias term or not

if using binary embedding files; default false

Glowe.autodetect_header — Method.

autodetect_header(filename)

Function that attempts at autodetecting the presence of a header in a GloVe embeddings file in a text format. If the function fails to detect a header, a false value is returned.

Note: The function explicitly expects a text format for the embeddings. The behaviour is undetermined for binary formats and no attempt at detecting the file format is done.