Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Gnuplot.exec — Method.

exec(sid::Symbol, s::Vector{String})

Directly execute commands on the underlying Gnuplot process, and return the result(s).

Examples:

exec("print GPVAL_TERM")
exec("plot sin(x)")

Gnuplot.gnuplot — Method.

gnuplot(sid::Symbol)

gnuplot(sid::Symbol, cmd::AbstractString)

Initialize a new session and (optionally) the associated Gnuplot process

Arguments:

sid: the session name (a Julia symbol);

cmd: a string specifying the complete file path to a gnuplot executable. If not given a dry session will be created, i.e. a session without underlying gnuplot process.

Gnuplot.quit — Method.

quit()

Quit the session and the associated gnuplot process (if any).

Gnuplot.quitall — Method.

quitall()

Quit all the sessions and the associated gnuplot processes.

Gnuplot.save — Method.

save(...)

Save the data and commands in the current session to either:

the gnuplot process (i.e. produce a plot): save(term="", output="");

an IO stream: save(stream::IO; term="", output="");

a file: save(file::AbstractStrings; term="", output="").

To save the data and command from a specific session pass the ID as first argument, i.e.:

save(sid::Symbol, term="", output="");

save(sid::Symbol, stream::IO; term="", output="");

save(sid::Symbol, file::AbstractStrings; term="", output="").

In all cases the term keyword allows to specify a gnuplot terminal, and the output keyword allows to specify an output file.

Gnuplot.setverbose — Method.

setverbose(b::Bool)

Set verbose flag to true or false (default: false).

Base.println — Method.

println

Send a string to gnuplot's STDIN.

The commands sent through send are not stored in the current session (use newcmd to save commands in the current session).

Arguments:

gp: a Session object;

str::String: command to be sent;

capture=false: set to true to capture and return the output.

Gnuplot.CheckGnuplotVersion — Method.

CheckGnuplotVersion

Check whether gnuplot is runnable with the command given in cmd. Also check that gnuplot version is >= 4.7 (required to use data blocks).

Gnuplot.newcmd — Method.

newcmd

Send a command to gnuplot process and store it in the current session.