Google JSON APIs

Google JSON APIs

GoogleCloud.api — Module.

General framework for representing Google JSON APIs.

GoogleCloud.api.APIMethod — Type.

APIMethod(verb, path, description)

Maps a method in the API to an HTTP verb and path.

GoogleCloud.api.APIResource — Type.

APIResource(path, methods)

Represents a resource in the API, typically rooted at a specific point in the REST hierarchy.

GoogleCloud.api.APIRoot — Type.

APIRoot(...)

Represent a Google JSON API containing resources, accessible via scopes.

GoogleCloud.api.get_session — Method.

get_session(api)

Get the default session for a specific API.

GoogleCloud.api.iserror — Method.

Check if response is/contains an error

GoogleCloud.api.set_session! — Method.

set_session!(api, session)

Set the default session for a specific API. Set session to nothing to forget session.

GoogleCloud.api.execute — Method.

execute(session::GoogleSession, resource::APIResource, method::APIMethod, 
        path_args::AbstractString...[; ...])

Execute a method against the provided path arguments.

Optionally provide parameters and data (with optional MIME content-type).

GoogleCloud.api.path_replace — Method.

path_replace(path, values)

Replace path tokens in path with values.

Assumes values are provided in the same order in which tokens appear in the path.

path_replace("/{foo}/{bar}/{baz}", ["this", "is", "it"])
# output
"/this/is/it"

GoogleCloud.api.path_tokens — Method.

path_tokens(path)

Extract tokens from a path, e.g.

path_tokens("/{foo}/{bar}/x/{baz}")
# output
3-element Array{SubString{String},1}:
 "{foo}"
 "{bar}"
 "{baz}"

Implemented APIs

GoogleCloud.api._storage — Module.

Google Cloud Storage API

GoogleCloud.api._storage.storage — Constant.

Google Cloud Storage API root.

GoogleCloud.api._compute — Module.

Google Compute Engine API

GoogleCloud.api._compute.compute — Constant.

Google Compute Engine API root.