Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GoogleCloudObjectStores._create! — Method.

Create bucket. If successful return nothing, else return an error message as a String.

GoogleCloudObjectStores._create! — Method.

Create object. If successful return nothing, else return an error message as a String.

v is either:

NamedTuple: (mimetype="application/json", value="some value")

An arbitrary value

GoogleCloudObjectStores._delete! — Method.

Delete bucket. If successful return nothing, else return an error message as a String.

GoogleCloudObjectStores._delete! — Method.

Delete object. If successful return nothing, else return an error message as a String.

GoogleCloudObjectStores._read — Method.

List the contents of a specific bucket. If successful return (true, value), else return (false, errormessage::String).

GoogleCloudObjectStores._read — Method.

Read object. If successful return (true, value), else return (false, errormessage::String).