Public API

Public API documentation

Documentation for Granular.jl's public interface.

See Package-internal documentation for internal package docs.

Index

Granular.Granular

Granular.addAtmosphereDrag!

Granular.addBodyForce!

Granular.addGrain!

Granular.addGrainCylindrical!

Granular.addOceanDrag!

Granular.addWall!

Granular.addWallLinearFrictionless!

Granular.applyAtmosphereDragToGrain!

Granular.applyAtmosphereVorticityToGrain!

Granular.applyOceanDragToGrain!

Granular.applyOceanVorticityToGrain!

Granular.areaOfQuadrilateral

Granular.areaOfTriangle

Granular.checkAndAddContact!

Granular.checkForContacts

Granular.checkTimeParameters

Granular.compareAtmospheres

Granular.compareGrains

Granular.compareOceans

Granular.compareSimulations

Granular.conformalQuadrilateralCoordinates

Granular.convertGrainDataToArrays

Granular.createEmptyAtmosphere

Granular.createEmptyOcean

Granular.createRegularAtmosphereGrid

Granular.createRegularOceanGrid

Granular.createSimulation

Granular.disableAtmosphereDrag!

Granular.disableGrain!

Granular.disableOceanDrag!

Granular.disableOutputFiles!

Granular.enableAtmosphereDrag!

Granular.enableOceanDrag!

Granular.findCellContainingPoint

Granular.findContacts!

Granular.findContactsAllToAll!

Granular.findContactsInGrid!

Granular.findEmptyPositionInGridCell

Granular.findLargestGrainStiffness

Granular.findSmallestGrainMass

Granular.fitGridToGrains!

Granular.getCellCenterCoordinates

Granular.getCellCornerCoordinates

Granular.getGridCornerCoordinates

Granular.getNonDimensionalCellCoordinates

Granular.getWallNormalStress

Granular.getWallSurfaceArea

Granular.grainCircumreference

Granular.grainHorizontalSurfaceArea

Granular.grainKineticRotationalEnergy

Granular.grainKineticTranslationalEnergy

Granular.grainMass

Granular.grainMomentOfInertia

Granular.grainSideSurfaceArea

Granular.grainThermalEnergy

Granular.grainVolume

Granular.harmonicMean

Granular.incrementCurrentTime!

Granular.interGrainPositionVector

Granular.interact!

Granular.interactGrains!

Granular.interpolateAtmosphereState

Granular.interpolateOceanState

Granular.interpolateOceanVelocitiesToCorners

Granular.irregularPacking!

Granular.isPointInCell

Granular.isPointInGrid

Granular.plotGrainSizeDistribution

Granular.plotGrains

Granular.printGrainInfo

Granular.printMemoryUsage

Granular.randpower

Granular.readOceanGridNetCDF

Granular.readOceanNetCDF

Granular.readOceanStateNetCDF

Granular.readSimulation

Granular.readSimulation

Granular.readSimulationStatus

Granular.reflectGrainsFromImpermeableBoundaries!

Granular.regularPacking!

Granular.removeSimulationFiles

Granular.render

Granular.reportGridBoundaryConditions

Granular.reportSimulationTimeToStdout

Granular.resetTime!

Granular.run!

Granular.setBodyForce!

Granular.setCurrentTime!

Granular.setGridBoundaryConditions!

Granular.setMaximumNumberOfContactsPerGrain!

Granular.setOutputFileInterval!

Granular.setTimeStep!

Granular.setTotalTime!

Granular.sortGrainsInGrid!

Granular.status

Granular.totalGrainKineticRotationalEnergy

Granular.totalGrainKineticTranslationalEnergy

Granular.totalGrainThermalEnergy

Granular.updateGrainKinematics!

Granular.updateGrainKinematicsThreeTermTaylor!

Granular.updateGrainKinematicsTwoTermTaylor!

Granular.updateWallKinematics!

Granular.updateWallKinematicsThreeTermTaylor!

Granular.updateWallKinematicsTwoTermTaylor!

Granular.vecTo3d

Granular.writeGrainInteractionVTK

Granular.writeGrainVTK

Granular.writeParaviewPythonScript

Granular.writeSimulation

Granular.writeSimulationStatus

Granular.writeVTK

Granular.zeroForcesAndTorques!

Granular.zeroKinematics!

Public Interface

Granular.Granular — Module.

Granular.jl

Offline granular dynamics simulator module.

Granular.addAtmosphereDrag! — Method.

Add drag from linear and angular velocity difference between atmosphere and all grains.

Granular.addBodyForce! — Method.

setBodyForce!(grain, force)

Add to the value of the external body force on a grain.

Arguments

grain::GrainCylindrical: the grain to set the body force for.

force::Vector{Float64}: a vector of force [N]

Granular.addGrain! — Function.

addGrain!(simulation::Simulation,
          grain::GrainCylindrical,
          verbose::Bool = false)

Add an grain to the simulation object. If verbose is true, a short confirmation message will be printed to stdout.

Granular.addGrainCylindrical! — Method.

function addGrainCylindrical!(simulation, lin_pos, contact_radius,
                                thickness[, areal_radius, lin_vel, lin_acc,
                                force, ang_pos, ang_vel, ang_acc, torque,
                                density, contact_stiffness_normal,
                                contact_stiffness_tangential,
                                contact_viscosity_normal,
                                contact_viscosity_tangential,
                                contact_static_friction,
                                contact_dynamic_friction,
                                youngs_modulus, poissons_ratio,
                                tensile_strength, shear_strength,
                                strength_heal_rate,
                                fracture_toughness,
                                ocean_drag_coeff_vert,
                                ocean_drag_coeff_horiz,
                                atmosphere_drag_coeff_vert,
                                atmosphere_drag_coeff_horiz,
                                pressure, fixed,
                                allow_x_acc, allow_y_acc, allow_z_acc,
                                rotating, enabled, verbose,
                                ocean_grid_pos, atmosphere_grid_pos,
                                n_contact, granular_stress, ocean_stress,
                                atmosphere_stress,
                                thermal_energy,
                                color])

Creates and adds a cylindrical grain to a simulation. Most of the arguments are optional, and come with default values. The only required arguments are simulation, lin_pos, contact_radius, and thickness.

Arguments

simulation::Simulation: the simulation object where the grain should be added to.

lin_pos::Vector{Float64}: linear position of grain center [m]. If a two-component vector is used, the values will be mapped to x and y, and the z component will be set to zero.

contact_radius::Float64: grain radius for granular interaction [m].

thickness::Float64: grain thickness [m].

areal_radius = false: grain radius for determining sea-ice concentration [m].

lin_vel::Vector{Float64} = [0., 0., 0.]: linear velocity [m/s]. If a two-component vector is used, the values will be mapped to x and y, and the z component will be set to zero.

lin_acc::Vector{Float64} = [0., 0., 0.]: linear acceleration [m/s^2]. If a two-component vector is used, the values will be mapped to x and y, and the z component will be set to zero.

force::Vector{Float64} = [0., 0., 0.]: linear force balance [N]. If a two-component vector is used, the values will be mapped to x and y, and the z component will be set to zero.

ang_pos::Float64 = [0., 0., 0.]: angular position around its center vertical axis [rad]. If a scalar is used, the value will be mapped to z, and the x and y components will be set to zero.

ang_vel::Float64 = [0., 0., 0.]: angular velocity around its center vertical axis [rad/s]. If a scalar is used, the value will be mapped to z, and the x and y components will be set to zero.

ang_acc::Float64 = [0., 0., 0.]: angular acceleration around its center vertical axis [rad/s^2]. If a scalar is used, the value will be mapped to z, and the x and y components will be set to zero.

torque::Float64 = [0., 0., 0.]: torque around its center vertical axis [Nm]. If a scalar is used, the value will be mapped to *z, and the x and y components will be set to zero.

density::Float64 = 934.: grain mean density [kg/m^3].

contact_stiffness_normal::Float64 = 1e7: contact-normal stiffness [N/m]; overridden if youngs_modulus is set to a positive value.

contact_stiffness_tangential::Float64 = 0.: contact-tangential stiffness [N/m]; overridden if youngs_modulus is set to a positive value.

contact_viscosity_normal::Float64 = 0.: contact-normal viscosity [N/m/s].

contact_viscosity_tangential::Float64 = 0.: contact-tangential viscosity [N/m/s].

contact_static_friction::Float64 = 0.4: contact static Coulomb frictional coefficient [-].

contact_dynamic_friction::Float64 = 0.4: contact dynamic Coulomb frictional coefficient [-].

youngs_modulus::Float64 = 2e7: elastic modulus [Pa]; overrides any value set for contact_stiffness_normal.

poissons_ratio::Float64 = 0.185: Poisson's ratio, used to determine the contact-tangential stiffness from youngs_modulus [-].

tensile_strength::Float64 = 0.: contact-tensile (cohesive) bond strength [Pa].

shear_strength::Float64 = 0.: shear strength of bonded contacts [Pa].

strength_heal_rate::Float64 = 0.: rate at which contact bond strength is obtained [Pa/s].

fracture_toughness::Float64 = 0.: fracture toughness which influences the maximum compressive strength on granular contact [m^{1/2}Pa]. A value of 1.285e3 m^{1/2}Pa is used for sea ice by Hopkins 2004.

ocean_drag_coeff_vert::Float64 = 0.85: vertical drag coefficient for ocean against grain sides [-].

ocean_drag_coeff_horiz::Float64 = 5e-4: horizontal drag coefficient for ocean against grain bottom [-].

atmosphere_drag_coeff_vert::Float64 = 0.4: vertical drag coefficient for atmosphere against grain sides [-].

atmosphere_drag_coeff_horiz::Float64 = 2.5e-4: horizontal drag coefficient for atmosphere against grain bottom [-].

pressure::Float64 = 0.: current compressive stress on grain [Pa].

fixed::Bool = false: grain is fixed to a constant velocity (e.g. zero).

allow_x_acc::Bool = false: override fixed along x.

allow_y_acc::Bool = false: override fixed along y.

allow_z_acc::Bool = false: override fixed along z.

rotating::Bool = true: grain is allowed to rotate.

enabled::Bool = true: grain interacts with other grains.

verbose::Bool = true: display diagnostic information during the function call.

ocean_grid_pos::Array{Int, 1} = [0, 0]: position of grain in the ocean grid.

atmosphere_grid_pos::Array{Int, 1} = [0, 0]: position of grain in the atmosphere grid.

n_contacts::Int = 0: number of contacts with other grains.

granular_stress::Vector{Float64} = [0., 0., 0.]: resultant stress on grain from granular interactions [Pa].

ocean_stress::Vector{Float64} = [0., 0., 0.]: resultant stress on grain from ocean drag [Pa].

atmosphere_stress::Vector{Float64} = [0., 0., 0.]: resultant stress on grain from atmosphere drag [Pa].

thermal_energy::Float64 = 0.0: thermal energy of grain [J].

color::Int=0: type number, usually used for associating a color to the grain during visualization.

Examples

The most basic example adds a new grain to the simulation sim, with a center at [1., 2., 0.], a radius of 1. meter, and a thickness of 0.5 meter:

Granular.addGrainCylindrical!(sim, [1., 2.], 1., .5)

Note that the z component is set to zero if a two-component vector is passed.

The following example will create a grain with tensile and shear strength, and a velocity of 0.5 m/s towards -x:

Granular.addGrainCylindrical!(sim, [4., 2.], 1., .5,
                              tensile_strength = 200e3,
                              shear_strength = 100e3,
                              lin_vel = [-.5, 0.])

Fixed grains are useful for creating walls or coasts, and loops are useful for creating regular arrangements:

for i=1:5
    Granular.addGrainCylindrical!(sim, [i*2., 0., 3.], 1., .5, fixed=true)
end

Granular.addOceanDrag! — Method.

Add drag from linear and angular velocity difference between ocean and all ice floes.

Granular.addWall! — Function.

addWall!(simulation::Simulation,
         wall::WallLinearFrictionless,
         verbose::Bool = false)

Add an wall to the simulation object. If verbose is true, a short confirmation message will be printed to stdout.

Granular.addWallLinearFrictionless! — Method.

function addWallLinear!(simulation, normal, pos[, bc, mass, thickness, 
                        normal_stress, vel, acc, force,
                        contact_viscosity_normal, verbose])

Creates and adds a linear (flat) and frictionless dynamic wall to a grain to a simulation. Most of the arguments are optional, and come with default values. The only required arguments are simulation, normal, pos, and bc.

Arguments

simulation::Simulation: the simulation object where the wall should be added to.

normal::Vector{Float64}: 3d vector denoting the normal to the wall [m]. The wall will only interact in the opposite direction of this vector, so the normal vector should point in the direction of the grains. If a 2d vector is passed, the third (z) component is set to zero.

pos::Float64: position along axis parallel to the normal vector [m].

bc::String="fixed": boundary condition, possible values are "fixed" (default), "normal stress", or "velocity".

mass::Float64=NaN: wall mass, which is used if wall boundary conditions differs from bc="fixed". If the parameter is left to its default value, the wall mass is set to be equal the total mass of grains in the simulation. Units: [kg]

thickness::Float64=NaN: wall thickness, which is used for determining wall surface area. If the parameter is left to its default value, the wall thickness is set to be equal to the thickest grain in the simulation. Units: [m].

normal_stress::Float64=0.: the wall normal stress when bc == "normal stress" [Pa].

vel::Float64=0.: the wall velocity along the normal vector. If the wall boundary condition is bc = "velocity" the wall will move according to this constant value. If bc = "normal stress" the velocity will be a free parameter. Units: [m/s]

force::Float64=0.: sum of normal forces on the wall from interaction with grains [N].

contact_viscosity_normal::Float64=0.: viscosity to apply in parallel to elasticity in interactions between wall and particles [N/(m/s)]. When this term is larger than zero, the wall-grain interaction acts like a sink of kinetic energy.

verbose::Bool=true: show verbose information during function call.

Examples

The most basic example adds a new fixed wall to the simulation sim, with a wall-face normal of [1., 0.] (wall along y and normal to x), a position of 1.5 meter:

Granular.addWallLinearFrictionless!(sim, [1., 0., 0.], 1.5)

The following example creates a wall with a velocity of 0.5 m/s towards -y:

Granular.addWallLinearFrictionless!(sim, [0., 1., 0.], 1.5,
                                    bc="velocity",
                                    vel=-0.5)

To create a wall parallel to the y axis pushing downwards with a constant normal stress of 100 kPa, starting at a position of y = 3.5 m:

Granular.addWallLinearFrictionless!(sim, [0., 1., 0.], 3.5,
                                    bc="normal stress",
                                    normal_stress=100e3)

Granular.applyAtmosphereDragToGrain! — Method.

Add Stokes-type drag from velocity difference between atmosphere and a single grain.

Granular.applyAtmosphereVorticityToGrain! — Method.

Add Stokes-type torque from angular velocity difference between atmosphere and a single grain. See Eq. 9.28 in "Introduction to Fluid Mechanics" by Nakayama and Boucher, 1999.

Granular.applyOceanDragToGrain! — Method.

Add Stokes-type drag from velocity difference between ocean and a single ice floe.

Granular.applyOceanVorticityToGrain! — Method.

Add Stokes-type torque from angular velocity difference between ocean and a single grain. See Eq. 9.28 in "Introduction to Fluid Mechanics" by Nakayama and Boucher, 1999.

Granular.areaOfQuadrilateral — Method.

Returns the area of a quadrilateral with corner coordinates a, b, c, and d. Corners a and c should be opposite of each other, the same must be true for b and d. This is true if the four corners are passed as arguments in a "clockwise" or "counter-clockwise" manner.

Granular.areaOfTriangle — Method.

Returns the area of an triangle with corner coordinates a, b, and c.

Granular.checkAndAddContact! — Function.

checkAndAddContact!(simulation, i, j)

Check for contact between two grains and register the interaction in the simulation object. The indexes of the two grains is stored in simulation.contact_pairs as [i, j]. The overlap vector is parallel to a straight line connecting the grain centers, points away from grain i and towards j, and is stored in simulation.overlaps. A zero-length vector is written to simulation.contact_parallel_displacement.

Arguments

simulation::Simulation: the simulation object containing the grains.

i::Int: index of the first grain.

j::Int: index of the second grain.

distance_Modifier::Vector{Float64}: vector modifying percieved inter-particle distance, which is used for contact search across periodic boundaries.

Granular.checkForContacts — Method.

checkForContacts(grid, position, radius)

Perform an O(n*log(n)) cell-based contact search between a candidate grain with position position and radius, against all grains registered in the grid. Returns the number of contacts that were found as an Integer value, unless return_when_overlap_found is true.

Arguments

simulation::Simulation: Simulation object containing grain positions.

grid::Any: Ocean or Atmosphere grid containing sorted particles.

x_candidate::Vector{Float64}: Candidate center position to probe for contacts with existing grains [m].

r_candidate::Float64: Candidate radius [m].

return_when_overlap_found::Bool (default: false): Return true if no contacts are found, or return false as soon as a contact is found.

Granular.checkTimeParameters — Method.

Checks if simulation temporal parameters are of reasonable values.

Granular.compareAtmospheres — Method.

compareAtmospheres(atmosphere1::atmosphere, atmosphere2::atmosphere)

Compare values of two atmosphere objects using the Base.Test framework.

Granular.compareGrains — Method.

compareGrains(if1::GrainCylindrical, if2::GrainCylindrical)

Compare values of two grain objects using the Base.Test framework.

Granular.compareOceans — Method.

compareOceans(ocean1::Ocean, ocean2::Ocean)

Compare values of two Ocean objects using the Base.Test framework.

Granular.compareSimulations — Method.

compareSimulations(sim1::Simulation, sim2::Simulation)

Compare values of two Simulation objects using the Base.Test framework.

Granular.conformalQuadrilateralCoordinates — Method.

Returns the non-dimensional coordinates [x_tilde, y_tilde] of a point p within a quadrilateral with corner coordinates A, B, C, and D. Points must be ordered in counter-clockwise order, starting from south-west corner.

Granular.convertGrainDataToArrays — Method.

Gathers all grain parameters from the GrainCylindrical type in continuous arrays in an GrainArrays structure.

Granular.createEmptyAtmosphere — Method.

Returns empty ocean type for initialization purposes.

Granular.createEmptyOcean — Method.

Returns empty ocean type for initialization purposes.

Granular.createRegularAtmosphereGrid — Method.

createRegularAtmosphereGrid(n, L[, origo, time, name,
                            bc_west, bc_south, bc_east, bc_north])

Initialize and return a regular, Cartesian Atmosphere grid with n[1] by n[2] cells in the horizontal dimension, and n[3] vertical cells. The cell corner and center coordinates will be set according to the grid spatial dimensions L[1], L[2], and L[3]. The grid u, v, h, and e fields will contain one 4-th dimension matrix per time step. Sea surface will be at z=0. with the atmosphere spanning z<0.. Vertical indexing starts with k=0 at the sea surface, and increases downwards.

Arguments

n::Vector{Int}: number of cells along each dimension [-].

L::Vector{Float64}: domain length along each dimension [m].

origo::Vector{Float64}: domain offset in each dimension m.

time::Vector{Float64}: vector of time stamps for the grid [s].

name::String: grid name (default = "unnamed").

bc_west::Integer: grid boundary condition for the grains.

bc_south::Integer: grid boundary condition for the grains.

bc_east::Integer: grid boundary condition for the grains.

bc_north::Integer: grid boundary condition for the grains.

Granular.createRegularOceanGrid — Method.

createRegularOceanGrid(n, L[, origo, time, name,
                       bc_west, bc_south, bc_east, bc_north])

Initialize and return a regular, Cartesian Ocean grid with n[1] by n[2] cells in the horizontal dimension, and n[3] vertical cells. The cell corner and center coordinates will be set according to the grid spatial dimensions L[1], L[2], and L[3]. The grid u, v, h, and e fields will contain one 4-th dimension matrix per time step. Sea surface will be at z=0. with the ocean spanning z<0.. Vertical indexing starts with k=0 at the sea surface, and increases downwards.

Arguments

n::Vector{Int}: number of cells along each dimension [-].

L::Vector{Float64}: domain length along each dimension [m].

origo::Vector{Float64}: domain offset in each dimension m.

time::Vector{Float64}: vector of time stamps for the grid [s].

name::String: grid name (default = "unnamed").

bc_west::Integer: grid boundary condition for the grains.

bc_south::Integer: grid boundary condition for the grains.

bc_east::Integer: grid boundary condition for the grains.

bc_north::Integer: grid boundary condition for the grains.

Granular.createSimulation — Method.

createSimulation([id])

Create a simulation object to contain all relevant variables such as temporal parameters, fluid grids, grains, and contacts. The parameter id is used to uniquely identify the simulation when it is written to disk.

The function returns a Simulation object, which you can add grains to, e.g. with addGrainCylindrical!.

Optional argument

id::String="unnamed": simulation identifying string.

Granular.disableAtmosphereDrag! — Method.

disableAtmosphereDrag!(grain)

Disable atmosphere-caused drag on the grain.

Granular.disableGrain! — Method.

Disable grain with index i in the simulation object.

Granular.disableOceanDrag! — Method.

disableOceanDrag!(grain)

Disable ocean-caused drag on the grain.

Granular.disableOutputFiles! — Method.

Disables the write of output files to disk during a simulation.

Granular.enableAtmosphereDrag! — Method.

enableAtmosphereDrag!(grain)

Enable atmosphere-caused drag on the grain, with values by Hunke and Comeau (2011).

Granular.enableOceanDrag! — Method.

enableOceanDrag!(grain)

Enable ocean-caused drag on the grain, with values by Hunke and Comeau (2011).

Granular.findCellContainingPoint — Function.

findCellContainingPoint(grid, point[, method])

Returns the i, j index of the grid cell containing the point. The function uses either an area-based approach (method = "Area"), or a conformal mapping approach (method = "Conformal"). The area-based approach is more robust. This function returns the coordinates of the cell. If no match is found the function returns (0,0).

Arguments

grid::Any: grid object containing ocean or atmosphere data.

point::Vector{Float64}: two-dimensional vector of point to check.

method::String: approach to use for determining if point is inside cell or not, can be "Conformal" (default) or "Area".

Granular.findContacts! — Method.

findContacts!(simulation[, method])

Top-level function to perform an inter-grain contact search, based on grain linear positions and contact radii.

The simplest contact search algorithm (method="all to all") is the most computationally expensive (O(n^2)). The method "ocean grid" bins the grains into their corresponding cells on the ocean grid and searches for contacts only within the vicinity. When this method is applied, it is assumed that the contact_radius values of the grains are smaller than half the cell size.

Arguments

simulation::Simulation: the simulation object containing the grains.

method::String: the contact-search method to apply. Valid options are "all to all" and "ocean grid".

Granular.findContactsAllToAll! — Method.

findContactsAllToAll!(simulation)

Perform an O(n^2) all-to-all contact search between all grains in the simulation object.

Granular.findContactsInGrid! — Method.

findContactsInGrid!(simulation)

Perform an O(n*log(n)) cell-based contact search between all grains in the simulation object.

Granular.findEmptyPositionInGridCell — Method.

findEmptyPositionInGridCell(simulation, grid, i, j, r[, n_iter, seed,
                            verbose])

Attempt locate an empty spot for an grain with radius r with center coordinates in a specified grid cell (i, j) without overlapping any other grains in that cell or the neighboring cells. This function will stop attempting after n_iter iterations, each with randomly generated positions.

This function assumes that existing grains have been binned according to the grid (e.g., using sortGrainsInGrid()).

If the function sucessfully finds a position it will be returned as a two-component Vector{Float64}. If a position is not found, the function will return false.

Arguments

simulation::Simulation: the simulation object to add grains to.

grid::Any: the grid to use for position search.

i::Int: the grid-cell index along x.

j::Int: the grid-cell index along y.

r::Float64: the desired grain radius to fit into the cell.

n_iter::Int = 30: the number of attempts for finding an empty spot.

seed::Int = 1: seed for the pseudo-random number generator.

verbose::Bool = false: print diagnostic information.

Granular.findLargestGrainStiffness — Method.

Finds the largest elastic stiffness of all grains in a simulation. Used to determine the optimal time step length.

Granular.findSmallestGrainMass — Method.

Finds the smallest mass of all grains in a simulation. Used to determine the optimal time step length.

Granular.fitGridToGrains! — Method.

fitGridToGrains!(simulation, grid[, padding])

Fit the ocean or atmosphere grid for a simulation to the current grains and their positions.

Arguments

simulation::Simulation: simulation object to manipulate.

grid::Any: Ocean or Atmosphere grid to manipulate.

padding::Real: optional padding around edges [m].

verbose::Bool: show grid information when function completes.

Granular.getCellCenterCoordinates — Method.

getCellCenterCoordinates(grid, i, j)

Returns grid center coordinates (h-point).

Arguments

xh::Array{Float64, 2}: nominal longitude of h-points [degrees_E]

yh::Array{Float64, 2}: nominal latitude of h-points [degrees_N]

i::Int: x-index of cell.

j::Int: y-index of cell.

Granular.getCellCornerCoordinates — Method.

getCellCornerCoordinates(xq, yq, i, j)

Returns grid-cell corner coordinates in the following order (south-west corner, south-east corner, north-east corner, north-west corner).

Arguments

xq::Array{Float64, 2}: nominal longitude of q-points [degrees_E]

yq::Array{Float64, 2}: nominal latitude of q-points [degrees_N]

i::Int: x-index of cell.

j::Int: y-index of cell.

Granular.getGridCornerCoordinates — Method.

getGridCornerCoordinates(xq, yq)

Returns grid corner coordinates in the following order (south-west corner, south-east corner, north-east corner, north-west corner).

Arguments

xq::Array{Float64, 2}: nominal longitude of q-points [degrees_E]

yq::Array{Float64, 2}: nominal latitude of q-points [degrees_N]

Granular.getNonDimensionalCellCoordinates — Method.

Returns the non-dimensional conformal mapped coordinates for point point in cell i,j, based off the coordinates in the grid.

This function is a wrapper for getCellCornerCoordinates() and conformalQuadrilateralCoordinates().

Granular.getWallNormalStress — Method.

getWallNormalStress(simulation[, wall_index, stress_type])

Returns the current "effective" or "defined" normal stress on the wall with index wall_index inside the simulation object. The returned value is given in Pascal.

Arguments

simulation::Simulation: the simulation object containing the wall.

wall_index::Integer=1: the wall number in the simulation object.

stress_type::String="effective": the normal-stress type to return. The defined value corresponds to the normal stress that the wall is asked to uphold. The effective value is the actual current normal stress. Usually, the magnitude of the effective normal stress fluctuates around the defined normal stress.

Granular.getWallSurfaceArea — Method.

getWallSurfaceArea(simulation, wall_index)

Returns the surface area of the wall given the grid size and its index.

Arguments

simulation::Simulation: the simulation object containing the wall.

wall_index::Integer=1: the wall number in the simulation object.

Granular.grainCircumreference — Method.

Returns the circumreference of the grain

Granular.grainHorizontalSurfaceArea — Method.

Returns the top or bottom (horizontal) surface area of the grain

Granular.grainKineticRotationalEnergy — Method.

Returns the rotational kinetic energy of the grain

Granular.grainKineticTranslationalEnergy — Method.

Returns the translational kinetic energy of the grain

Granular.grainMass — Method.

Returns the mass of the grain

Granular.grainMomentOfInertia — Method.

Returns the moment of inertia of the grain

Granular.grainSideSurfaceArea — Method.

Returns the surface area of the grain sides

Granular.grainThermalEnergy — Method.

Returns the thermal energy of the grain, produced by Coulomb slip

Granular.grainVolume — Method.

Returns the volume of the grain

Granular.harmonicMean — Method.

harmonicMean(a, b)

Returns the harmonic mean of two numbers a::Number and b::Number.

Granular.incrementCurrentTime! — Method.

incrementCurrentTime!(simulation::Simulation, t::Float64)

Sets the current simulation time of the simulation object to t, with parameter value checks.

Granular.interGrainPositionVector — Method.

interGrainPositionVector(simulation, i, j)

Returns a vector pointing from grain i to grain j in the simulation.

Arguments

simulation::Simulation: the simulation object containing the grains.

i::Int: index of the first grain.

j::Int: index of the second grain.

Granular.interact! — Method.

interact!(simulation::Simulation)

Resolve mechanical interaction between all particle pairs.

Arguments

simulation::Simulation: the simulation object containing the grains.

Granular.interactGrains! — Method.

interactGrains!(simulation::Simulation, i::Int, j::Int, ic::Int)

Resolve an grain-to-grain interaction using a prescibed contact law. This function adds the compressive force of the interaction to the grain pressure field of mean compressive stress on the grain sides.

The function uses the macroscopic contact-stiffness parameterization based on Young's modulus and Poisson's ratio if Young's modulus is a positive value.

Granular.interpolateAtmosphereState — Method.

Atmosphere data is containted in Atmosphere type at discrete times (Atmosphere.time). This function performs linear interpolation between time steps to get the approximate atmosphere state at any point in time. If the Atmosphere data set only contains a single time step, values from that time are returned.

Granular.interpolateOceanState — Method.

Ocean data is containted in Ocean type at discrete times (Ocean.time). This function performs linear interpolation between time steps to get the approximate ocean state at any point in time. If the Ocean data set only contains a single time step, values from that time are returned.

Granular.interpolateOceanVelocitiesToCorners — Method.

Convert gridded data from Arakawa-C type (decomposed velocities at faces) to Arakawa-B type (velocities at corners) through interpolation.

Granular.irregularPacking! — Method.

irregularPacking!(simulation[, radius_max, radius_min, sample_limit,
                  padding_factor, binary_radius_search,
                  binary_sampling_quality, thickness, seed,
                  plot_during_packing, verbose)

Generate a dense disc packing in 2D using Poisson disc sampling with O(N) complexity, as described by Robert Bridson (2007) "Fast Poisson disk sampling in arbitrary dimensions". The simulation can be empty or already contain grains. However, an simulation.ocean or simulation.atmosphere grid is required.

Arguments

simulation::Simulation: simulation object where grains are inserted.

radius_max::Real: largest grain radius to use.

radius_min::Real: smallest grain radius to use.

sample_limit::Integer=30: number of points to sample around each grain before giving up.

padding_factor::Real=0.: if positive and binary_radius_search = false, try to add an occasional grain from the current active grain (radius_max*padding_factor).

binary_radius_search::Bool=false: use a binary radius-sampling procedure to fit the largest possible grains into the packing. This option will create the highest packing density.

binary_sampling_quality::Real=100.: the quality to enforce during the binary radius search when binary_radius_search = true. Larger values create denser packings but take longer to complete.

seed::Integer: seed value to the pseudo-random number generator.

plot_during_packing::Bool=false: produce successive plots as the packing is generated. Requires gnuplot (default).

verbose::Bool=true: show diagnostic information to stdout.

Granular.isPointInCell — Function.

Check if a 2d point is contained inside a cell from the supplied grid. The function uses either an area-based approach (method = "Area"), or a conformal mapping approach (method = "Conformal"). The area-based approach is more robust. This function returns true or false.

Granular.isPointInGrid — Function.

Check if a 2d point is contained inside the grid. The function uses either an area-based approach (method = "Area"), or a conformal mapping approach (method = "Conformal"). The area-based approach is more robust. This function returns true or false.

Granular.plotGrainSizeDistribution — Method.

plotGrainSizeDistribution(simulation, [filename_postfix, nbins,
                            size_type, filetype, gnuplot_terminal,
                            skip_fixed, log_y, verbose)

Plot the grain size distribution as a histogram and save it to the disk. The plot is saved accoring to the simulation id, the optional filename_postfix string, and the filetype, and is written to the current folder.

Arguments

simulation::Simulation: the simulation object containing the grains.

filename_postfix::String: optional string for the output filename.

nbins::Int: number of bins in the histogram (default = 12).

size_type::String: specify whether to use the contact or areal radius for the grain size. The default is contact.

filetype::String: the output file type (default = "png").

gnuplot_terminal::String: the gnuplot output terminal to use (default = "png").

skip_fixed::Bool: ommit grains that are fixed in space from the size distribution (default = true).

log_y::Bool: plot y-axis in log scale.

verbose::String: show output file as info message in stdout (default = true).

Granular.plotGrains — Method.

plotGrains(simulation, [filetype, gnuplot_terminal, verbose])

Plot the grains using Gnuplot and save the figure to disk.

Arguments

simulation::Simulation: the simulation object containing the grains.

filetype::String: the output file type (default = "png").

gnuplot_terminal::String: the gnuplot output terminal to use (default = "png crop size 1200,1200").

plot_interactions::Bool: show grain-grain interactions in the plot.

verbose::String: show output file as info message in stdout (default = true).

Granular.printGrainInfo — Method.

printGrainInfo(grain::GrainCylindrical)

Prints the contents of an grain to stdout in a formatted style.

Granular.printMemoryUsage — Method.

printMemoryUsage(sim::Simulation)

Shows the memory footprint of the simulation object.

Granular.randpower — Function.

randpower([nvals], [distribution_power], [min_val], [max_val])

Returns one or more random numbers from a power-law probability distribution.

Arguments

dims::Any: the dimensions of random values (default = 1)

distribution_power::Number: the distribution power (default = 1.)

min_val::Number: the lower bound of the distribution range (default = 0.)

max_val::Number: the upper bound of the distribution range (default = 1.)

Granular.readOceanGridNetCDF — Method.

Read NetCDF file with ocean supergrid information generated by MOM6 (e.g. ocean_hrid.nc) from disk and return as Ocean data structure. This file is located in the simulation INPUT/ subdirectory.

Returns

xh::Array{Float64, 2}: Longitude for cell centers [deg]

yh::Array{Float64, 2}: Latitude for cell centers [deg]

xq::Array{Float64, 2}: Longitude for cell corners [deg]

yq::Array{Float64, 2}: Latitude for cell corners [deg]

Granular.readOceanNetCDF — Method.

Read ocean NetCDF files generated by MOM6 from disk and return as Ocean data structure.

Arguments

velocity_file::String: path to NetCDF file containing ocean velocities, etc., (e.g. prog__####_###.nc).

grid_file::String: path to NetCDF file containing ocean super-grid information (typically INPUT/ocean_hgrid.nc).

regular_grid::Bool=false: true if the grid is regular (all cells equal and grid is Cartesian) or false (default).

Granular.readOceanStateNetCDF — Method.

Read NetCDF file with ocean state generated by MOM6 (e.g. prog__####_###.nc or ########.ocean_month.nc) from disk and return time stamps, velocity fields, layer thicknesses, interface heights, and vertical coordinates.

Returns

time::Vector{Float64}: Time [s]

u::Array{Float64, 2}: Cell corner zonal velocity [m/s], dimensions correspond to placement in [xq, yq, zl, time]

v::Array{Float64, 2}: Cell corner meridional velocity [m/s], dimensions correspond to placement in [xq, yq, zl, time]

h::Array{Float64, 2}: layer thickness [m], dimensions correspond to placement in [xh, yh, zl, time]

e::Array{Float64, 2}: interface height relative to mean sea level [m], dimensions correspond to placement in [xh, yh, zi, time]

zl::Vector{Float64}: layer target potential density [kg m^-3]

zi::Vector{Float64}: interface target potential density [kg m^-3]

Granular.readSimulation — Method.

readSimulation(simulation::Simulation;
               step::Integer = -1,
               verbose::Bool = true)

Read the simulation state from disk and return as new simulation object.

Arguments

simulation::Simulation: use the simulation.id to determine the file name to read from, and read information from the file into this object.

step::Integer=-1: attempt to read this output file step. At its default value (-1), the function will try to read the latest file, determined by calling readSimulationStatus.

verbose::Bool=true: confirm to console that the file has been read.

Granular.readSimulation — Method.

readSimulation(filename::String="";
               verbose::Bool=true)

Return Simulation content read from disk using the JLD2 format.

Arguments

filename::String: path to file on disk containing the simulation information.

verbose::Bool=true: confirm to console that the file has been read.

Granular.readSimulationStatus — Method.

readSimulationStatus(simulation_id[, folder, verbose])

Read the current simulation status from disk (<sim.id>/<sim.id>.status.txt) and return the last output file number.

Arguments

simulation_id::String: the simulation identifying string.

folder::String=".": the folder in which to search for the status file.

verbose::Bool=true: show simulation status in console.

Granular.reflectGrainsFromImpermeableBoundaries! — Method.

reflectGrainsFromImpermeableBoundaries!(simulation::Simulation)

If the ocean or atmosphere grids are impermeable, reflect grain trajectories by reversing the velocity vectors normal to the boundary. This function is to be called after temporal integration of the grain positions.

Granular.regularPacking! — Method.

regularPacking!(simulation, n, r_min, r_max[, tiling, padding_factor,
                size_distribution, size_distribution_parameter, seed])

Create a grid-based regular packing with grain numbers along each axis specified by the n vector.

Arguments

simulation::Simulation: simulation object where the grains are inserted, preferably not containing prior grains.

n::Vector{Integer}: 2-element vector determining number of grains along the x and y axes.

r_min::Real: minimum desired grain radius.

r_max::Real: maximum desired grain radius.

tiling::String: the packing method to use, valid values are "square" (default) and "triangular" (see Wikipedia).

padding_factor::Real: percentage-wise padding around each grain to allow for random perturbations to grain position (default = 0.0).

origo::Vector{Real}: spatial offset for the packing (default [0.0, 0.0]).

size_distribution::String: grain-size distribution to sample. Valid values are "powerlaw" and "uniform".

size_distribution_parameter::Real: parameter to pass to the grain-size distribution generating function.

seed::Integer: seed value to the pseudo-random number generator.

Granular.removeSimulationFiles — Method.

removeSimulationFiles(simulation[, folder])

Remove all simulation output files from the specified folder.

Granular.render — Method.

render(simulation[, pvpython, images, animation])

Wrapper function which calls writeParaviewPythonScript(...) and executes it from the shell using the supplied pvpython argument.

Arguments

simulation::Simulation: simulation object containing the grain data.

pvpython::String: path to the pvpython executable to use. By default, the script uses the pvpython in the system PATH.

images::Bool: render images to disk (default: true)

gif::Bool: merge images as GIF and save to disk (default: false, requires images=true)

animation::Bool: render animation as movie to disk (default: false). If ffmpeg is available on the system, the .avi file is converted to .mp4.

trim::Bool: trim images in animated sequence (default: true)

reverse::Bool: if images=true additionally render reverse-animated gif (default: false)

Granular.reportGridBoundaryConditions — Method.

reportGridBoundaryConditions(grid)

Report the boundary conditions for the grid to the console.

Granular.reportSimulationTimeToStdout — Method.

Prints the current simulation time and total time to standard out

Granular.resetTime! — Method.

resetTime!(simulation)

Reset the current time to zero, and reset output file counters in order to restart a simulation. This function does not overwrite the time step (Simulation.time_step), the output file interval (Simulation.file_time_step), or the total simulation time (Simulation.time_total).

Arguments

simulation::Simulation: the simulation object for which to reset the temporal parameters.

Granular.run! — Method.

run!(simulation[,
     verbose::Bool = true,
     status_interval = 100.,
     show_file_output = true,
     single_step = false,
     temporal_integration_method = "Three-term Taylor"],
     write_jld2 = false)

Run the simulation through time until simulation.time equals or exceeds simulatim.time_total. This function requires that all grains are added to the simulation and that the length of the computational time step is adjusted accordingly.

The function will search for contacts, determine the force balance on each ice floe, and integrate all kinematic degrees of freedom accordingly. The temporal integration is explicit and of length simulation.time_step. This function will write VTK files to disk in the intervals simulation.file_time_step by the function writeVTK. If this value is negative, no output files will be written to disk.

Arguments

simulation::Simulation: the simulation to run (object is modified)

verbose::Bool=true: show verbose information during the time loop

status_interval::Bool=true: show verbose information during the time loop

show_file_output::Bool=true: report to stdout when output file is written

single_step::Bool=false: run simulation for a single time step only. If this causes simulation.time to exceed simulation.time_total, the latter is increased accordingly.

temporal_integration_method::String="Three-term Taylor": type of integration method to use. See updateGrainKinematics for details.

write_jld2::Bool=false: write simulation state to disk as JLD2 files (see Granular.writeSimulation(...) whenever saving VTK output.

Granular.setBodyForce! — Method.

setBodyForce!(grain, force)

Set the value of the external body force on a grain.

Arguments

grain::GrainCylindrical: the grain to set the body force for.

force::Vector{Float64}: a vector of force [N]

Granular.setCurrentTime! — Method.

setCurrentTime!(simulation::Simulation, t::Float64)

Sets the current simulation time of the simulation object to t, with parameter value checks.

Granular.setGridBoundaryConditions! — Function.

setGridBoundaryConditions!(grid, grid_face, mode)

Set boundary conditions for the granular phase at the edges of Ocean or Atmosphere grids. The target boundary can be selected through the grid_face argument, or the same boundary condition can be applied to all grid boundaries at once.

When the center coordinate of grains crosses an inactive boundary (mode = "inactive"), the grain is disabled (GrainCylindrical.enabled = false). This keeps the grain in memory, but stops it from moving or interacting with other grains. By default, all boundaries are inactive.

If the center coordinate of a grain crosses a periodic boundary (mode = periodic), the grain is repositioned to the opposite side of the model domain. Grains can interact mechanically across the periodic boundary.

Arguments

grid::Any: Ocean or Atmosphere grid to apply the boundary condition to.

grid_face::String: Grid face to apply the boundary condition to. Valid values are any combination and sequence of "west" (-x), "south" (-y), "east" (+x), "north" (+y), or simply any combination of "-x", "+x", "-y", and "+y". The specifiers may be delimited in any way. Also, and by default, all boundaries can be selected with "all" (-x, -y, +x, +y), which overrides any other face selection.

mode::String: Boundary behavior, accepted values are "inactive", "periodic", and "impermeable". You cannot specify more than one mode at a time, so if several modes are desired as boundary conditions for the grid, several calls to this function should be made.

verbose::Bool: Confirm boundary conditions by reporting values to console.

Examples

Set all boundaries for the ocean grid to be periodic:

setGridBoundaryConditions!(ocean, "periodic", "all")

Set the south-north boundaries to be inactive, but the west-east boundaries to be periodic:

setGridBoundaryConditions!(ocean, "inactive", "south north")
setGridBoundaryConditions!(ocean, "periodic", "west east")

or specify the conditions from the coordinate system axes:

setGridBoundaryConditions!(ocean, "inactive", "-y +y")
setGridBoundaryConditions!(ocean, "periodic", "-x +x")

Granular.setMaximumNumberOfContactsPerGrain! — Method.

setMaximumNumberOfContactsPerGrain!(simulation, number_of_contacts)

Change the maximum number of contacts per grain, which changes simulation.Nc_max and reallocates memory for each grain. Larger values require more memory, but allow simulation of wider grain-size distributions. The default value is a maximum of 32 contacts per grain, which is sufficient for most practical purposes.

Arguments

simulation::Simulation: the Simulation object to modify

number_of_contacts::Int: the maximum number of contacts per grain to allow.

Granular.setOutputFileInterval! — Method.

setOutputFileInterval!(simulation::Simulation, t::Float64)

Sets the simulation-time interval between output files are written to disk. If this value is zero or negative, no output files will be written.

Granular.setTimeStep! — Method.

setTimeStep!(simulation[, epsilon, verbose])

Find the computational time step length suitable given the grain radii, contact stiffnesses, and grain density. Uses the scheme by Radjaii et al. 2011.

Arguments

simulation::Simulation: the simulation object to modify.

epsilon::Float64=0.07: safety factor in the time step scheme. Larger values are more likely to cause unstable behavior than smaller values.

verbose::Bool=true: display the resultant time step in the console.

Granular.setTotalTime! — Method.

setTotalTime!(simulation::Simulation, t::Float64)

Sets the total simulation time of the simulation object to t, with parameter value checks.

Granular.sortGrainsInGrid! — Method.

Find grain positions in grid, based on their center positions.

Granular.status — Function.

status(folder[, loop, t_int, colored_output, write_header, render)

Shows the status of all simulations with output files written under the specified folder, which is the current working directory by default.

Arguments

folder::String=".": directory (including subdirectories) to scan for simulation output. loop::Bool=false: continue printing the status every t_int seconds. t_int::Int=10: interval between status updates when loop=true. colored_output::Bool=true: display output with colors. write_header::Bool=true: write header line explaining the data. visualize::Bool=false: render the simulation output. Does not work well whenloop=true`, as the script regenerates (and overwrites) all output graphics on every call.

Granular.totalGrainKineticRotationalEnergy — Method.

totalGrainKineticRotationalEnergy(simulation)

Returns the sum of rotational kinetic energies of all grains in a simulation

Granular.totalGrainKineticTranslationalEnergy — Method.

totalGrainKineticTranslationalEnergy(simulation)

Returns the sum of translational kinetic energies of all grains in a simulation

Granular.totalGrainThermalEnergy — Method.

totalGrainKineticTranslationalEnergy(simulation)

Returns the sum of thermal energy of all grains in a simulation

Granular.updateGrainKinematics! — Method.

updateGrainKinematics!(simulation::Simulation[,
                         method::String = "Three-term Taylor"])

Update the grain kinematic parameters using a temporal integration scheme, the current force and torque balance, and gravitational acceleration. If the simulation contains a grid with periodic boundaries, affected grain positions are adjusted accordingly.

Arguments

simulation::Simulation: update the grain positions in this object according to temporal integration of length simulation.time_step.

method::String = "Three-term Taylor": the integration method to use. Available methods are "Two-term Taylor" and "Three-term Taylor". The three-term Taylor expansion is slightly more computationally expensive than the two-term Taylor expansion, but offers an order-of-magnitude increase in precision of grain positions. The two-term expansion can obtain similar precision if the time step is 1/10 the length.

Granular.updateGrainKinematicsThreeTermTaylor! — Method.

Use a three-term Taylor expansion for integrating the kinematic degrees of freedom for a grain.

Granular.updateGrainKinematicsTwoTermTaylor! — Method.

Use a two-term Taylor expansion for integrating the kinematic degrees of freedom for a grain.

Granular.updateWallKinematics! — Method.

updateWallKinematics!(simulation::Simulation[,
                      method::String = "Three-term Taylor"])

Update the wall kinematic parameters using a temporal integration scheme, the current force and torque balance, and gravitational acceleration. If the simulation contains a grid with periodic boundaries, affected wall positions are adjusted accordingly.

Arguments

simulation::Simulation: update the wall positions in this object according to temporal integration of length simulation.time_step.

method::String = "Three-term Taylor": the integration method to use. Available methods are "Two-term Taylor" and "Three-term Taylor". The three-term Taylor expansion is slightly more computationally expensive than the two-term Taylor expansion, but offers an order-of-magnitude increase in precision of wall positions. The two-term expansion can obtain similar precision if the time step is 1/10 the length.

Granular.updateWallKinematicsThreeTermTaylor! — Method.

updateWallKinematicsThreeTermTaylor!(wall, simulation)

Use a two-term Taylor expansion for integrating the kinematic degrees of freedom for a wall.

Granular.updateWallKinematicsTwoTermTaylor! — Method.

updateWallKinematicsTwoTermTaylor!(wall, simulation)

Use a two-term Taylor expansion for integrating the kinematic degrees of freedom for a wall.

Granular.vecTo3d — Method.

function vecTo3d(input, fill)

Convert a scalar or 2d vector to 3d by filling the missing component with the value fill. The returned 3-component vector is a Vector (or 1d Array) of the same type as the input.

Arguments

input: a scalar or two-component vector.

fill::Real: value to use for third

Granular.writeGrainInteractionVTK — Method.

writeGrainInteractionVTK(simulation::Simulation,
                           filename::String;
                           verbose::Bool=false)

Saves grain interactions to .vtp files for visualization with VTK, for example in Paraview. Convert Cell Data to Point Data and use with Tube filter.

Granular.writeGrainVTK — Method.

Write a VTK file to disk containing all grains in the simulation in an unstructured mesh (file type .vtu). These files can be read by ParaView and can be visualized by applying a Glyph filter. This function is called by writeVTK().

Granular.writeParaviewPythonScript — Method.

function writeParaviewPythonScript(simulation, [filename, folder, vtk_folder, verbose])

Create a ".py" script for visualizing the simulation VTK files in Paraview. The script can be run from the command line with pvpython (bundled with Paraview), or from the interactive Python shell inside Paraview.

Arguments

simulation::Simulation: input simulation file containing the data.

filename::String: output file name for the Python script. At its default (blank) value, the script is named after the simulation id (simulation.id).

folder::String: output directory, current directory the default.

vtk_folder::String: directory containing the VTK output files, by default points to the full system path equivalent to "./<simulation.id>/".

save_animation::Bool: make the generated script immediately save a rendered animation to disk when the ".py" script is called.

verbose::Bool: show diagnostic information during

function call, on by default.

Granular.writeSimulation — Method.

writeSimulation(simulation::Simulation;
                     filename::String="",
                     folder::String=".",
                     verbose::Bool=true)

Write all content from Simulation to disk in JLD2 format. If the filename parameter is not specified, it will be saved to a subdirectory under the current directory named after the simulation identifier simulation.id.

Granular.writeSimulationStatus — Method.

writeSimulationStatus(simulation::Simulation;
                      folder::String=".",
                      verbose::Bool=false)

Write current simulation status to disk in a minimal txt file.

Granular.writeVTK — Method.

Write a VTK file to disk containing all grains in the simulation in an unstructured mesh (file type .vtu). These files can be read by ParaView and can be visualized by applying a Glyph filter.

If the simulation contains an Ocean data structure, it's contents will be written to separate .vtu files. This can be disabled by setting the argument ocean=false. The same is true for the atmosphere.

The VTK files will be saved in a subfolder named after the simulation.

Granular.zeroForcesAndTorques! — Method.

Sets the force and torque values of all grains and dynamic walls to zero.

Granular.zeroKinematics! — Method.

zeroKinematics!(simulation)

Set all grain forces, torques, accelerations, and velocities (linear and rotational) to zero in order to get rid of all kinetic energy.