Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Graph500.get_min_type — Method.

get_min_type(n)

Return the minimum Integer type which can contain n

Graph500.get_next_type — Method.

get_next_type(T)

Return a data with a Integer Datatype twice the size of T

Graph500.kernel_1 — Method.

kernel_1(ij)

Create Graph from edge list ij based on kernel1 benchmark(http://graph500.org/?pageid=12#sec-4) Return LightGraphs.SimpleGraph structure