Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GraphPlot.circular_layout — Method.

This function wrap from NetworkX

Position nodes on a circle.

Parameters

G a graph

Returns

locsx, locsy Locations of the nodes. Can be any units you want, but will be normalized and centered anyway

Examples

julia> g = simple_house_graph()
julia> locs_x, locs_y = circular_layout(g)

GraphPlot.gplot — Method.

Given a graph and two vectors of X and Y coordinates, returns a Compose tree of the graph layout

Arguments

G Graph to draw

layout Optional. Layout algorithm. Currently can be one of [randomlayout, circularlayout, springlayout, shelllayout, stressmajorizelayout, spectrallayout]. Default: spring_layout

locs_x, locs_y Locations of the nodes. Can be any units you want, but will be normalized and centered anyway

NODESIZE Optional. Max size for the nodes. Default: 3.0/sqrt(N)

nodesize Optional. Relative size for the nodes, can be a Vector. Default: 1.0

nodelabel Optional. Labels for the vertices, a Vector or nothing. Default: nothing

nodelabelc Optional. Color for the node labels, can be a Vector. Default: colorant"black"

nodelabeldist Optional. Distances for the node labels from center of nodes. Default: 0.0

nodelabelangleoffset Optional. Angle offset for the node labels. Default: π/4.0

NODELABELSIZE Optional. Largest fontsize for the vertice labels. Default: 4.0

nodelabelsize Optional. Relative fontsize for the vertice labels, can be a Vector. Default: 1.0

nodefillc Optional. Color to fill the nodes with, can be a Vector. Default: colorant"turquoise"

nodestrokec Optional. Color for the nodes stroke, can be a Vector. Default: nothing

nodestrokelw Optional. Line width for the nodes stroke, can be a Vector. Default: 0.0

edgelabel Optional. Labels for the edges, a Vector or nothing. Default: []

edgelabelc Optional. Color for the edge labels, can be a Vector. Default: colorant"black"

edgelabeldistx, edgelabeldisty Optional. Distance for the edge label from center of edge. Default: 0.0

EDGELABELSIZE Optional. Largest fontsize for the edge labels. Default: 4.0

edgelabelsize Optional. Relative fontsize for the edge labels, can be a Vector. Default: 1.0

EDGELINEWIDTH Optional. Max line width for the edges. Default: 0.25/sqrt(N)

edgelinewidth Optional. Relative line width for the edges, can be a Vector. Default: 1.0

edgestrokec Optional. Color for the edge strokes, can be a Vector. Default: colorant"lightgray"

arrowlengthfrac Optional. Fraction of line length to use for arrows. Equal to 0 for undirected graphs. Default: 0.1 for the directed graphs

arrowangleoffset Optional. Angular width in radians for the arrows. Default: π/9 (20 degrees)

GraphPlot.graphfamous — Method.

read some famous graphs

Paramenters

graphname Currently, graphname can be one of ["karate", "football", "dolphins", "netscience", "polbooks", "power", "cond-mat"]

Return a graph

Example julia> g = graphfamous("karate")

GraphPlot.random_layout — Method.

Position nodes uniformly at random in the unit square. For every node, a position is generated by choosing each of dim coordinates uniformly at random on the interval [0.0, 1.0).

Parameters

G graph or list of nodes, A position will be assigned to every node in G

Return

locsx, locsy Locations of the nodes. Can be any units you want, but will be normalized and centered anyway

Examples

julia> g = simple_house_graph()

julia> loc_x, loc_y = random_layout(g)

GraphPlot.readedgelist — Method.

read graph from in edgelist format

GraphPlot.shell_layout — Function.

This function is copy from IainNZ's GraphLayout.jl

Position nodes in concentric circles.

Parameters

G a graph

nlist Vector of Vector, Vector of node Vector for each shell.

Examples

julia> g = graphfamous("karate")
julia> nlist = Array{Vector{Int}}(2)
julia> nlist[1] = [1:5]
julia> nlist[2] = [6:num_vertiecs(g)]
julia> locs_x, locs_y = shell_layout(g, nlist)

GraphPlot.spectral_layout — Function.

This function wrap from NetworkX

Position nodes using the eigenvectors of the graph Laplacian.

Parameters

G a graph

weight array or nothing, optional (default=nothing) The edge attribute that holds the numerical value used for the edge weight. If None, then all edge weights are 1.

Examples

julia> g = graphfamous("karate")
julia> weight = rand(num_edges(g))
julia> locs_x, locs_y = spectral_layout(g, weight)

GraphPlot.spring_layout — Function.

This function is copy from IainNZ's GraphLayout.jl

Use the spring/repulsion model of Fruchterman and Reingold (1991):

Attractive force: f_a(d) = d^2 / k

Repulsive force: f_r(d) = -k^2 / d

where d is distance between two vertices and the optimal distance between vertices k is defined as C * sqrt( area / num_vertices ) where C is a parameter we can adjust

Parameters

G a graph

C Constant to fiddle with density of resulting layout

MAXITER Number of iterations we apply the forces

INITTEMP Initial "temperature", controls movement per iteration

Examples

julia> g = graphfamous("karate")
julia> locs_x, locs_y = spring_layout(g)

GraphPlot.stressmajorize_layout — Function.

Compute graph layout using stress majorization

Inputs:

δ: Matrix of pairwise distances
p: Dimension of embedding (default: 2)
w: Matrix of weights. If not specified, defaults to
       w[i,j] = δ[i,j]^-2 if δ[i,j] is nonzero, or 0 otherwise
X0: Initial guess for the layout. Coordinates are given in rows.
    If not specified, default to random matrix of Gaussians

Additional optional keyword arguments control the convergence of the algorithm and the additional output as requested:

maxiter:   Maximum number of iterations. Default: 400size(X0, 1)^2
abstols:   Absolute tolerance for convergence of stress.
           The iterations terminate if the difference between two
           successive stresses is less than abstol.
           Default: √(eps(eltype(X0))
reltols:   Relative tolerance for convergence of stress.
           The iterations terminate if the difference between two
           successive stresses relative to the current stress is less than
           reltol. Default: √(eps(eltype(X0))
abstolx:   Absolute tolerance for convergence of layout.
           The iterations terminate if the Frobenius norm of two successive
           layouts is less than abstolx. Default: √(eps(eltype(X0))
verbose:   If true, prints convergence information at each iteration.
           Default: false
returnall: If true, returns all iterates and their associated stresses.
           If false (default), returns the last iterate

Output:

The final layout X, with coordinates given in rows, unless returnall=true.

Reference:

The main equation to solve is (8) of:

@incollection{
    author = {Emden R Gansner and Yehuda Koren and Stephen North},
    title = {Graph Drawing by Stress Majorization}
    year={2005},
    isbn={978-3-540-24528-5},
    booktitle={Graph Drawing},
    seriesvolume={3383},
    series={Lecture Notes in Computer Science},
    editor={Pach, J'anos},
    doi={10.1007/978-3-540-31843-9_25},
    publisher={Springer Berlin Heidelberg},
    pages={239--250},
}

GraphPlot.LZ — Method.

Computes L^Z defined in (5) of the Reference

Input: Z: current layout (coordinates) d: Ideal distances (default: all 1) w: weights (default: d.^-2)

GraphPlot.graphline — Method.

Return lines and arrow heads

GraphPlot.stress — Function.

Stress function to majorize

Input: X: A particular layout (coordinates in rows) d: Matrix of pairwise distances w: Weights for each pairwise distance

See (1) of Reference

GraphPlot.weightedlaplacian — Method.

Compute weighted Laplacian given ideal weights w

Lʷ defined in (4) of the Reference