Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GraphRecipes.directed_curve — Method.

This function builds a BezierCurve which leaves point p vertically upwards and arrives point q vertically upwards. It may create a loop if necessary. It assumes the view is [0,1]. That can be modified using the xview and yview keyword arguments (default: 0:1).

GraphRecipes.partialcircle — Function.

get an array of tuples of points on a circle with radius r

GraphRecipes.random_control_point — Method.

Randomly pick a point to be the center control point of a bezier curve, which is both equidistant between the endpoints and normally distributed around the midpoint.