Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Graphics.BoundingBox — Type.

BoundingBox(xmin, xmax, ymin, ymax) -> bb

Create a representation bb of a rectangular region, specifying the coordinates of the horizontal (x) and vertical (y) edges.

Graphics.BoundingBox — Method.

BoundingBox(bb0::BoundingBox, bboxes::BoundingBox...) -> bb

Compute the BoundingBox bb that minimally encloses all of the input boxes.

Graphics.BoundingBox — Method.

BoundingBox(p0::Point, points::Point...) -> bb

Compute the BoundingBox bb that minimally encloses all of the input points.

Graphics.Point — Type.

Point(x, y) -> p

Create a Cartesian representation p of a point in two dimensions.

Graphics.Vec2 — Type.

Vec2(x, y) -> v

Create a Cartesian representation v of a vector (or point) in two dimensions.

Graphics.aspect_ratio — Method.

aspect_ratio(bb::BoundingBox) -> r

Compute the ratio r of the height and width of bb.

Graphics.deform — Method.

deform(bb::BoundingBox, Δl, Δr, Δt, Δb) -> bbnew

Add Δl (left), Δr (right), Δt (top), and Δb (bottom) to the edges of a BoundingBox.

Graphics.inner_canvas — Method.

inner_canvas(c::GraphicsContext, device::BoundingBox, user::BoundingBox)
inner_canvas(c::GraphicsContext, x, y, w, h, l, r, t, b)

Create a rectangular drawing area inside device (represented in device-coordinates), giving it user-coordinates user. Any drawing that occurs outside this box is clipped.

x, y, w, and h are an alternative parametrization of device, and l, r, t, b parametrize user.

See also: set_coordinates.

Graphics.isinside — Method.

isinside(bb::BoundingBox, p::Point) -> tf::Bool
isinside(bb::BoundingBox, x, y) -> tf::Bool

Determine whether the point lies within bb.

Graphics.rotate — Method.

rotate(bb::BoundingBox, angle, o) -> bbnew

Rotate bb around o by angle, returning the BoundingBox that encloses the vertices of the rotated box.

Graphics.rotate — Method.

rotate(p::Vec2, angle::Real, o::Vec2) -> pnew

Rotate p around o by angle.

Graphics.set_coordinates — Method.

set_coordinates(c::GraphicsContext, device::BoundingBox, user::BoundingBox)
set_coordinates(c::GraphicsContext, user::BoundingBox)

Set the device->user coordinate transformation of c so that device, expressed in "device coordinates" (pixels), is equivalent to user as expressed in "user coordinates". If device is omitted, it defaults to the full span of c, BoundingBox(0, width(c), 0, height(c)).

See also get_matrix, set_matrix.

Graphics.shift — Method.

shift(bb::BoundingBox, Δx, Δy) -> bbnew

Shift center by (Δx,Δy), keeping width & height fixed.

Base.:& — Method.

bb1 & bb2 -> bb

Compute the intersection of two BoundingBoxes.

Base.:* — Method.

s*bb -> bbnew
bb*s -> bbnew

Scale width & height of BoundingBox bb by s, keeping center fixed.

Base.:+ — Method.

bb1 + bb2 -> bb

Compute the BoundingBox bb that minimally contains bb1 and bb2