Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GroupedErrors._cumulative — Method.

`_cumulative!(df, xaxis, x) = ecdf(df[x])(xaxis)`

Cumulative density function of x, computed along xaxis

GroupedErrors._density — Method.

`_density(df,xaxis::AbstractRange, x; kwargs...)`

Kernel density of x, computed along xaxis

GroupedErrors._density — Method.

`_density(df, xaxis, x)`

Normalized histogram of x (which is discrete: every value is its own bin)

GroupedErrors._hazard — Method.

`_hazard(df,xaxis, x; kwargs...)`

Hazard rate of x, computed along xaxis. Keyword arguments are passed to the function computing the density

GroupedErrors._locreg — Method.

`_locreg(df, xaxis::AbstractRange, x, y; kwargs...)`

Apply loess regression, training the regressor with x and y and predicting xaxis

GroupedErrors._locreg — Method.

`_locreg(df, xaxis, x, y; estimator = mean)`

In the discrete case, the function computes the estimate of y for a given value of x using the function estimator (default is mean)