Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Hadamard.fwht — Function.

fwht(X, dims=1:ndims(X))

Return the fast Walsh-Hadamard transform (WHT) of the array X along the dimensions dims (an integer, tuple, or array of dimensions, defaulting to all dimensions). Only power-of-two sizes (along the transformed dimensions) are supported. The result is returned in the sequency ordering.

Our WHT is normalized so that the forward transform has a 1/n coefficient (where n is the product of the transformed dimensions) and the inverse WHT has no scaling factor.

Hadamard.fwht_dyadic — Function.

fwht_dyadic(X, dims=1:ndims(X))

Return the fast Walsh-Hadamard transform (WHT) of the array X along the dimensions dims (an integer, tuple, or array of dimensions, defaulting to all dimensions). Only power-of-two sizes (along the transformed dimensions) are supported. The result is returned in the dyadic (Paley, bit-reversed) ordering.

Our WHT is normalized so that the forward transform has a 1/n coefficient (where n is the product of the transformed dimensions) and the inverse WHT has no scaling factor.

Hadamard.fwht_natural — Function.

fwht_natural(X, dims=1:ndims(X))

Return the fast Walsh-Hadamard transform (WHT) of the array X along the dimensions dims (an integer, tuple, or array of dimensions, defaulting to all dimensions). Only power-of-two sizes (along the transformed dimensions) are supported. The result is returned in the natural (Hadamard) ordering.

Our WHT is normalized so that the forward transform has a 1/n coefficient (where n is the product of the transformed dimensions) and the inverse WHT has no scaling factor.

Hadamard.fwht_natural! — Function.

fwht_natural!(X, dims=1:ndims(X))

Similar to fwht_natural, but works in-place on the input array X.

Hadamard.hadamard — Method.

hadamard(n)

Return a Hadamard matrix of order n. The known Hadamard matrices up to size 256 are currently supported (via a lookup table), along with any size that factorizes into products of these known sizes and/or powers of two.

The return value is a matrix of Int8 values. If you want to do further matrix computations with this matrix, you may want to convert to Float64 first via float(hadamard(n)).

You can pretty-print a Hadamard matrix as a table of + and - (indicating the signs of the entries) via Hadamard.printsigns, e.g. Hadamard.printsigns(hadamard(28)) for the 28×28 Hadamard matrix.

Hadamard.ifwht — Function.

ifwht(X, dims=1:ndims(X))

Return the inverse fast Walsh-Hadamard transform (WHT) of the array X along the dimensions dims (an integer, tuple, or array of dimensions, defaulting to all dimensions). Only power-of-two sizes (along the transformed dimensions) are supported. The result is returned in the sequency ordering.

Our WHT is normalized so that the forward transform has a 1/n coefficient (where n is the product of the transformed dimensions) and the inverse WHT has no scaling factor.

Hadamard.ifwht_dyadic — Function.

ifwht_dyadic(X, dims=1:ndims(X))

Return the inverse fast Walsh-Hadamard transform (WHT) of the array X along the dimensions dims (an integer, tuple, or array of dimensions, defaulting to all dimensions). Only power-of-two sizes (along the transformed dimensions) are supported. The result is returned in the dyadic (Paley, bit-reversed) ordering.

Our WHT is normalized so that the forward transform has a 1/n coefficient (where n is the product of the transformed dimensions) and the inverse WHT has no scaling factor.

Hadamard.ifwht_natural — Function.

ifwht_natural(X, dims=1:ndims(X))

Return the inverse fast Walsh-Hadamard transform (WHT) of the array X along the dimensions dims (an integer, tuple, or array of dimensions, defaulting to all dimensions). Only power-of-two sizes (along the transformed dimensions) are supported. The result is returned in the natural (Hadamard) ordering.

Our WHT is normalized so that the forward transform has a 1/n coefficient (where n is the product of the transformed dimensions) and the inverse WHT has no scaling factor.

Hadamard.ifwht_natural! — Function.

ifwht_natural!(X, dims=1:ndims(X))

Similar to ifwht_natural, but works in-place on the input array X.

Hadamard.printsigns — Method.

printsigns([io, ] A)

Print a table of + and - characters indicating the signs of the entries of A (with 0 for zero entries). The output stream io defaults to stdout.