Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

HeaderREPLs.append_keymaps! — Method.

append_keymaps!(keymaps, repl::HeaderREPL{H})

Append Dict{Any,Any} key maps to keymaps in order of highest priority first. Some typically useful keymaps (in conventional order of priority):

trigger_search_keymap

mode_termination_keymap

trigger_prefix_keymap

REPL.LineEdit.history_keymap

REPL.LineEdit.default_keymap

REPL.LineEdit.escape_defaults

HeaderREPLs.clear_io — Method.

clear_io(s, repl)

Erases both the input line and the header.

HeaderREPLs.count_display_lines — Method.

nlines = count_display_lines(io, ds)

Count the number of lines needed to display the contents of io in a terminal of displaysize ds. This handles "line wrap" as well as newlines.

HeaderREPLs.find_prompt — Method.

find_prompt(mi, "julia")
find_prompt(mi, PrefixHistoryPrompt)

Return the selected prompt from mi, searching either for the prompt-string of a Prompt or, for other TextInterfaces, searching by type.

HeaderREPLs.mode_termination_keymap — Method.

keymap_dict = mode_termination_keymap(repl::HeaderREPL, default_prompt::Prompt)

Default back to default_prompt for "^C" and hitting backspace as the first character of the line.

HeaderREPLs.print_header — Method.

print_header(io::IO, header::CustomHeader)

Print header to io. header must be a mutable struct containing a field nlines, and before exiting print_header should set this field to the number of lines occupied by the display of your header.

While you have to define print_header, generally you should not call it directly. If you need to display the header, call refresh_header.

HeaderREPLs.refresh_header — Method.

refresh_header(s, repl; clearheader=true)
refresh_header(repl, s, termbuf, terminal; clearheader=true)

Clear (if clearheader is true) and redraw the header and input line.

HeaderREPLs.trigger_prefix_keymap — Method.

keymap_dict = trigger_prefix_keymap(p::PrefixHistoryPrompt)
keymap_dict = trigger_prefix_keymap(repl::HeaderREPL)

Sets up the arrow keys and "^P" and "^N" to trigger reverse and forward prefix-search, respectively.

HeaderREPLs.trigger_search_keymap — Method.

keymap_dict = trigger_search_keymap(p::HistoryPrompt)
keymap_dict = trigger_search_keymap(repl::HeaderREPL)

Sets up "^R" and "^S" to trigger reverse and forward search, respectively.

HeaderREPLs.activate_header — Method.

activate_header(header, prompt, state, termbuf, term)

Customize this if actions need to be taken to initialize your header when switching from other prompts to your custom prompt. The default is to do nothing.

See also deactivate_header.

HeaderREPLs.clear_header_area — Method.

clear_header_area(terminal, header::AbstractHeader)

Move to the top of the area used for display of header, clearing lines as you go.

In most cases you can probably rely on the fallback implementation, as long as you update header.nlines appropriately.

HeaderREPLs.deactivate_header — Method.

deactivate_header(header, prompt, state, termbuf, term)

Customize this if actions need to be taken to clean up your header when switching from your custom prompt to other prompts. The default is to do nothing.

See also activate_header.

HeaderREPLs.setup_prompt — Method.

prompt, modesym = setup_prompt(repl::HeaderREPL{H}, hascolor::Bool)

Return prompt::LineEdit.Prompt and a mode symbol modesym::Symbol that will appear in the julia history file.