Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Hilbert.hilbert — Method.

hilbert(x)
hilbert(x, n)

Analytic signal, computed using the Hilbert transform. y = hilbert(x) gives the analytic signal y = x + i*xi where xi is the Hilbert transform of vector x. If x is complex, only the real part is used. If x is a matrix, then hilbert operates along columns. If n is provided, the n-point Hilbert transform is computed.

Example

x = randn(10,10)
y = hilbert(x)