Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Homebrew.BrewPkg — Type.

BrewPkg

A simple type to give us some nice ways of representing our packages to the user

It contains important information such as the name of the package, the tap it came from, the version of the package and whether it was translated or not

Base.show — Method.

show(io::IO, b::BrewPkg)

Writes a BrewPkg to io, showing tap, name and version number

Homebrew.__init__ — Method.

__init__()

Initialization function. Calls install_brew() to ensure that everything we need is downloaded/installed, then calls update_env() to set the environment properly so that packages being installed can find their binaries.

Homebrew.add — Method.

add(pkg::Union{AbstractString,BrewPkg}; verbose::Bool=false, keep_translations::Bool=false)

Install package pkg and all dependencies, using bottles only, unlinking any previous versions if necessary, and linking the new ones in place. Will attempt to install non-relocatable bottles from Homebrew/core by translating formulae and forcing cellar :any into the formulae.

Automatically deletes all translated formulae before adding formulae and after, unless keep_translations is set to true.

Homebrew.add_flags — Method.

add_flags(cmd::AbstractString, flags::Dict{String,Bool})

Given a mapping of flags to Bools, return [cmd, flag1, flag2...] if the respective Bools are true. Useful for adding --verbose and --force flags onto the end of commands

Homebrew.brew — Method.

brew(cmd::Cmd; no_stderr=false, no_stdout=false, verbose=false, force=false, quiet=false)

Run command cmd using the configured brew binary, optionally suppressing stdout and stderr, and providing flags such as --verbose to the brew binary.

Homebrew.brewchomp — Method.

brewchomp(cmd::Cmd; no_stderr=false, no_stdout=false, verbose=false, force=false, quiet=false))

Run command cmd using the configured brew binary, optionally suppressing stdout and stderr, and providing flags such as --verbose to the brew binary.

This function uses readchomp(), as opposed to brew() which uses run()

Homebrew.cleanup — Method.

cleanup()

Cleans up old installed versions of formulae, as well as purging all downloaded bottles

Homebrew.clt_installed — Method.

clt_installed()

Checks whether the command-line tools are installed, as reported by xcode-select

Homebrew.delete_all_translated_formulae — Method.

delete_all_translated_formulae(;verbose::Bool=false)

Delete all translated formulae from the staticfloat/juliatranslated tap. This is useful for debugging misbehaving formulae during translation.

Homebrew.delete_translated_formula — Method.

delete_translated_formula(name::AbstractString; verbose::Bool=false)

Delete a translated formula from the staticfloat/juliatranslated tap.

Homebrew.deps_sorted — Method.

deps_sorted(pkg::Union{AbstractString,BrewPkg}; build_deps::Bool)

Return a sorted Vector{BrewPkg} of all dependencies (direct and indirect) such that each entry in the list appears after all of its own dependencies.

If build_deps is true include formula depependencies marked as :build.

Homebrew.deps_tree — Method.

deps_tree(pkg::Union{AbstractString,BrewPkg}; build_deps::Bool=false)

Return a dictionary mapping every dependency (both direct and indirect) of pkg to a Vector{BrewPkg} of all of its dependencies. Used in deps_sorted().

If build_deps is true include formula depependencies marked as :build.

Homebrew.direct_deps — Method.

direct_deps(pkg::Union{AbstractString,BrewPkg}; build_deps::Bool=false)

Return a list of all direct dependencies of pkg as a Vector{BrewPkg} If build_deps is true include formula depependencies marked as :build.

Homebrew.download_and_unpack — Method.

download_and_unpack(url::AbstractString, target_dir::AbstractString)

Download a tarball from url and unpack it into target_dir.

Homebrew.formula_path — Method.

formula_path(pkg::Union{AbstractString,BrewPkg})

Returns the absolute path on-disk of the given package pkg.

Homebrew.formula_tap — Method.

formula_tap(name::AbstractString)

Given a formula name, return the formula name and the tap it is from, replacing "" for "Homebrew/core", as we don't care about that particular prefix.

Homebrew.fullname — Method.

fullname(pkg::BrewPkg)

Return the fully-qualified name for a package, dropping "Homebrew/core"

Homebrew.git_installed — Method.

git_installed()

Checks whether git is truly installed or not, dealing with stubs in /usr/bin Also ensure that the version is new enough (e.g. >= 2.0.0.0) that it will work with git fetch --unshallow on Homebrew.

Homebrew.has_bottle — Method.

has_bottle(name::AbstractString)

Checks if a given formula has a bottle at all

Homebrew.has_relocatable_bottle — Method.

has_relocatable_bottle(name::AbstractString)

Checks to see if a given formula has a bottle that can be installed anywhere

Homebrew.info — Method.

info(name::AbstractString)

Returns information about a particular package name as a BrewPkg. This is our basic String -> BrewPkg converter.

Homebrew.info — Method.

info(names::Vector{String})

For each name in names, returns information about that particular package name as a BrewPkg. This is our batched String -> BrewPkg converter.

Homebrew.insert_after_dependencies — Method.

insert_after_dependencies(tree::Dict, sorted_deps::Vector{BrewPkg}, name::AbstractString)

Given a mapping from names to dependencies in tree, and a list of sorted dependencies in sorted_deps, insert a new dependency name into sorted_deps after all dependencies of name. If a dependency of name is not already in sorted_deps, then recursively add that dependency as well.

Homebrew.install_and_link — Method.

install_and_link(pkg::Union{AbstractString,BrewPkg}; verbose=false)

Installs, and links package pkg. Used by add(). Don't call manually unless you really know what you're doing, as this doesn't deal with dependencies, and so can trigger compilation when you don't want it to.

Homebrew.install_brew — Method.

install_brew()

Ensures that Homebrew is installed as desired, that our basic Taps are available and that we have whatever binary tools we need, such as install_name_tool

Homebrew.installed — Method.

installed(pkg::Union{AbstractString,BrewPkg})

Return true if the given package pkg is a directory in the Cellar, showing that it has been installed (but possibly not linked, see linked())

Homebrew.json — Method.

json(pkg::Union{AbstractString,BrewPkg})

Return the full JSON object for pkg, the result of brew info --json=v1 $pkg. If brew info fails, throws an error. If brew info returns an empty object, (e.g. "[]"), this returns an empty Dict.

Note that running brew info --json=v1 is somewhat expensive, so we cache the results in a global dictionary, and batching larger requests with the vectorized json() function similarly increases performance.

Homebrew.json — Method.

json(names::Vector{AbstractString})

For each package name in names, return the full JSON object for name, the result of brew info --json=v1 $name, stored in a dictionary keyed by the names passed into this function. If brew info fails, throws an error. If brew info returns an empty object "[]", that object is represented by an empty dictionary.

Note that running brew info --json=v1 is somewhat expensive, so we cache the results in a global dictionary, and batching larger requests with this function similarly increases performance.

Homebrew.link — Method.

link(pkg::Union{AbstractString,BrewPkg}; verbose=false, force=true)

Link package name into the global namespace, uses --force if force == true

Homebrew.linked — Method.

linked(pkg::Union{AbstractString,BrewPkg})

Returns true if the given package pkg is linked to LinkedKegs, signifying all files installed by this package have been linked into the global prefix.

Homebrew.list — Method.

list()

Returns a list of all installed packages as a Vector{BrewPkg}

Homebrew.outdated — Method.

outdated()

Returns a list of all installed packages that are out of date as a Vector{BrewPkg}

Homebrew.postinstall — Method.

postinstall(pkg::Union{AbstractString,BrewPkg}; verbose=false)

Runs brew postinstall against package pkg, useful for debugging complicated formulae when a bottle doesn't install right and you want to re-run postinstall.

Homebrew.prefix — Method.

prefix(pkg::Union{AbstractString,BrewPkg})

Returns the prefix for a particular package's latest installed version.

Homebrew.prefix — Method.

prefix()

Returns brew_prefix, the location where all Homebrew files are stored.

Homebrew.read_formula — Method.

read_formula(pkg::Union{AbstractString,BrewPkg})

Returns the string contents of a package's formula.

Homebrew.refresh! — Method.

refresh!(;verbose=false)

Forcibly remove all packages and add them again. This should only be used to fix a broken installation, normal operation should never need to use this.

Homebrew.rm — Method.

rm(pkg::Union{AbstractString,BrewPkg}; verbose::Bool=false, force::Bool=true)

Remove package pkg, use --force if force == true

Homebrew.rm — Method.

rm(pkgs::Vector{String or BrewPkg}; verbose::Bool=false, force::Bool=true)

Remove packages pkgs, use --force if force == true

Homebrew.tap — Method.

tap(tap_name::AbstractString; full::Bool=true, verbose::Bool=false)

Runs brew tap $tap_name if the tap does not already exist. If full is true adds the flag --full to clone a full tap instead of Homebrew's default shallow.

If git is not available, manually tap it with curl and tar. This tap will be unshallowed at the next brew update when git is available.

Homebrew.tap_exists — Method.

tap_exists(tap_name::AbstractString)

Check to see if a tap called tap_name (ex: "staticfloat/juliadeps") exists

Homebrew.translate_formula — Method.

translate_formula(pkg::Union{AbstractString,BrewPkg}; verbose::Bool=false)

Given a formula name, return the fully-qualified name of a translated formula if it is translatable. Translation copies a Homebrew/core formula to staticfloat/juliatranslated, adding appropriate cellar :any and root_url lines to any bottle stanzas. This allows us to transparently install non-cellar-any formulae from Homebrew/core.

This function is fairly strict, bailing out at every possible opportunity, and returning the original name. If a formula is non-translatable, it's possible it needs manual intervention, check out the staticfloat/juliadeps tap for examples.

Homebrew.unlink — Method.

unlink(pkg::Union{AbstractString,BrewPkg}; verbose::Bool=false, quiet::Bool=true)

Unlink package pkg from the global namespace, uses --quiet if quiet == true

Homebrew.update — Method.

update(;verbose::Bool=false)

Runs brew update to update Homebrew itself and all taps. Then runs upgrade() to upgrade all formulae that have fallen out of date.

Homebrew.update_env — Method.

update_env()

Updates environment variables PATH and HOMEBREWCACHE, and modifies DLLOAD_PATH to point to our Homebrew installation, allowing us to use things inside of Homebrew transparently. This causes BinDeps to find the binaries during Pkg.build() time, writing the absolute path into deps/deps.jl. Because the paths are written into deps/deps.jl, packages do not need to load in the entire Homebrew package just to find their dependencies.

HOMEBREW_CACHE stores our bottle download cache in a separate place, separating ourselves from other Homebrew installations so we don't conflict with anyone

Homebrew.update_tag — Method.

update_tag(;verbose::Bool = false)

We maintain our own "stable" tag that overrides Homebrew so that we can update at our own pace along with them. Make sure to call update_env() before calling this, as calling our git doesn't work properly otherwise.

Homebrew.upgrade — Method.

upgrade(;verbose::Bool=false)

Iterate over all packages returned from outdated(), removing the old version and adding a new one. Note that we do not simply call brew upgrade here, as we have special logic inside of add() to install from our tap before trying to install from mainline Homebrew.

Homebrew.write_formula — Method.

write_formula(name::AbstractString, formula::AbstractString)

Write out fully-qualified formula name with contents formula to disk. Note that writing out without a tap name is not allowed; we won't write new formulae out to Homebrew/core, only to taps.