Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

HttpCommon.Cookie — Type.

A Cookie represents an HTTP cookie. It has three fields: name and value are strings, and attrs is dictionary of pairs of strings.

HttpCommon.Headers — Type.

Headers represents the header fields for an HTTP request.

HttpCommon.Request — Type.

A Request represents an HTTP request sent by a client to a server. It has five fields:

method: an HTTP methods string (e.g. "GET")

resource: the resource requested (e.g. "/hello/world")

headers: see Headers above

data: the request data as a vector of bytes

uri: additional details, normally not used

HttpCommon.Response — Type.

A Response represents an HTTP response sent to a client by a server. It has six fields:

status: HTTP status code (see STATUS_CODES) [default: 200]

headers: Headers [default: HttpCommmon.headers()]

cookies: Dictionary of strings => Cookies

data: the request data as a vector of bytes [default: UInt8[]]

finished: true if the Reponse is valid, meaning that it can be converted to an actual HTTP response [default: false]

requests: the history of requests that generated the response. Can be greater than one if a redirect was involved.

Response has many constructors - use methods(Response) for full list.

HttpCommon.escapeHTML — Method.

escapeHTML(i::String)

Returns a string with special HTML characters escaped: &, <, >, ", '

HttpCommon.parsequerystring — Method.

parsequerystring(query::String)

Convert a valid querystring to a Dict:

q = "foo=bar&baz=%3Ca%20href%3D%27http%3A%2F%2Fwww.hackershool.com%27%3Ehello%20world%21%3C%2Fa%3E"
parsequerystring(q)
# Dict{String,String} with 2 entries:
#   "baz" => "<a href='http://www.hackershool.com'>hello world!</a>"
#   "foo" => "bar"