Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Hungarian.hungarian — Method.

hungarian(costMat) -> (assignment, cost)

Find an optimal solution of the rectangular assignment problem represented by the $N x M$ matrix costMat. Return the optimal column indices and corresponding minimal cost.

The costMat[n,m] denotes the cost to assign the nth worker to the mth job. The zero element in the return value assignment means that these workers have no assigned job.

Elements in the matrix can be set to missing. In this case, the corresponding matching cannot be considered by the algorithm.

Examples

julia> A = [ 24     1     8;
              5     7    14;
              6    13    20;
             12    19    21;
             18    25     2];

julia> assignment, cost = hungarian(A)
([2,1,0,0,3],8)

julia> assignment, cost = hungarian(A')
([2,1,5],8)

julia> using Missings

julia> costMat = [ missing  1   1
                      1     0   1
                      1     1   0 ]
3×3 Array{Union{Float64, Missings.Missing},2}:
missing  1.0  1.0
  1.0    0.0  1.0
  1.0    1.0  0.0

julia> hungarian(costMat)
([2, 1, 3], 2)

Hungarian.exit_criteria — Method.

exit_criteria(colCovered, ZsDims) -> stepNum

We adjust Munkres's algorithm in order to deal with rectangular matrices, so only K columns are counted here, where K = min(size(Zs))

Hungarian.munkres! — Method.

munkres!(costMat) -> Zs

Identical to munkres, except that it directly modifies its input matrix costMat instead of allocating a copy. As a result, the value of this matrix in the caller code will be modified and should therefore no more be used! This function should rather be used by advanced users to improve performance of critical code.

Hungarian.munkres — Method.

munkres(costMat) -> Zs

Find an optimal solution of the assignment problem represented by the square matrix costMat. Return an sparse matrix illustrating the optimal matching.

Examples

julia> costMat = ones(3, 3) - eye(3, 3)
3×3 Array{Float64,2}:
 0.0  1.0  1.0
 1.0  0.0  1.0
 1.0  1.0  0.0

julia> matching = Hungarian.munkres(costMat)
3×3 sparse matrix with 3 Int64 nonzero entries:
	[1, 1]  =  2
	[2, 2]  =  2
	[3, 3]  =  2

julia> full(matching)
3×3 Array{Int64,2}:
 2  0  0
 0  2  0
 0  0  2

Hungarian.step1! — Method.

Step 1 of the original Munkres' Assignment Algorithm

Hungarian.step2! — Method.

Step 2 of the original Munkres' Assignment Algorithm

Hungarian.step3! — Method.

Step 3 of the original Munkres' Assignment Algorithm