Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

HyperbolicPlane.HCircle — Type.

HCircle(P::HPoint,r::Real) creates a new hyperbolic circle centered at P with radius r.

HyperbolicPlane.HCircle — Method.

HCircle(A,B,C) creates a circle that includes the three given points.

HyperbolicPlane.HContainer — Type.

HContainer is a device for holding a collection of hyperbolic objects. It is like a set, but we have to do a lot of work before adding a new element because equal hyperbolic objects might differ a tiny amount and that would mess up hashing.

C = HContainer() creates a new container.

C = HContainer(items...) creates a new container with the items.

HyperbolicPlane.HLine — Method.

HLine(P,Q) creates a new line from the given two points.

HyperbolicPlane.HLine — Method.

HLine(S::HSegment) extends the segment S to give a (new) line.

HyperbolicPlane.HPoint — Type.

HPoint(z::Complex) creates a new point in the hyperbolic plane. The argument z must have absolute value less than 1.

HPoint(r,theta) creates a new point with polar coordinates (r,theta). See also: polar.

HyperbolicPlane.HPolygon — Type.

HPolygon() creates a new polygon (with no points).

HPolygon(list) creates a polygon whose points are specified in list.

See: add_point!

HyperbolicPlane.HRay — Type.

HRay(P::HPoint, t::Real) returns a ray with vertex P pointing to exp(im*t).

HyperbolicPlane.HRay — Method.

HRay(A,B) where A and B are points creates the ray with vertex A passing through B.

HyperbolicPlane.HSegment — Type.

HSegment(A,B) creates a new line segment with endpoints A and B.

HyperbolicPlane.HTriangle — Type.

HTriangle(A,B,C) creates a new hyperbolic triangle.

HyperbolicPlane.Horocycle — Type.

Horocycle(P::HPoint, theta::Real) create the horocycle containing the point P and the ideal point at exp(im*theta).

AbstractLattices.dist — Method.

dist(P::HPoint,L::HLine) is the distance from P to the nearest point on L.

AbstractLattices.dist — Method.

dist(P,Q) gives the distance betwen two points in the hyperbolic plane. If Q is omitted, give the distance from P to HPoint(0).

Base.angle — Method.

angle(A,B,C) finds the angle betwen BA and BC.

HyperbolicPlane.RandomHCircle — Method.

RandomHCircle() creates a random circle.

HyperbolicPlane.RandomHLine — Method.

RandomHLine() returns a random line in the hyperbolic plane.

Algorithm: choose two values s,t in [0,2pi) uniformly at random and then make the line from exp(s*im) to exp(t*im).

HyperbolicPlane.RandomHPoint — Method.

RandomHPoint() generates a point at random in the hyperbolic plane.

HyperbolicPlane.RandomHPolygon — Function.

RandomHPolygon(n::Int,simple::Bool=false) create a new HPolygon with n points chosen at random. With simple set to true, return a polygon that does not self-intersect.

HyperbolicPlane.RandomHRay — Method.

RandomHRay() creates a random ray. RandomRay(P::HPoint) creates a random ray with vertex P.

HyperbolicPlane.RandomHTriangle — Method.

RandomHTriangle() creates a random triangle via three calls to RandomHPoint().

HyperbolicPlane.RandomHorocycle — Method.

RandomHorocycle() creates a random horocycle by choosing a point at random by RandomHPoint and a random ideal point (uniformly between 0 and 2π).

HyperbolicPlane.add_object! — Method.

add_object!(C::HContainer, X::HObject) adds X to the container C.

HyperbolicPlane.add_point! — Method.

add_point!(X::HPolygon, P::HPoint) adds the point P as the last point of the polygon X

HyperbolicPlane.angles — Method.

angles(P::HPolygon) returns a list of the angles at the vertices of P.

The results are always in the interval [0,pi].

The order of the angles is the order of the vertices in P.plist.

HyperbolicPlane.angles — Method.

angles(T::HTriangle) returns a sorted triple containing the angles at the three corners of the triangle.

HyperbolicPlane.area — Method.

area(C::HCircle) returns the area of the circle.

HyperbolicPlane.area — Method.

area(X::HPolygon) returns the area of the polygon, but is only reliable if X does not self-intersect. See is_simple.

HyperbolicPlane.area — Method.

area(T::HTriangle) returns the area of the triangle.

HyperbolicPlane.between — Method.

between(a,b,c) determines if the hyperbolic point b lies on the segment from a to c.

HyperbolicPlane.bisector — Method.

bisector(S::HSegment) yields an HLine that's the perpendicular bisector of the segment S. May also be invoked bisector(A,B) where A and B are points.

HyperbolicPlane.circumference — Method.

circumference(C::HCircle) returns the circumference of the circle.

HyperbolicPlane.collinear — Method.

collinear checks if the arguments are collinear. Arguments are:

a,b,c: three points

a,L: point and segment (in either order)

L,LL: two segments

HyperbolicPlane.copy_attr — Method.

copy_attr(A,B) copies the attributes assigned to B into A.

HyperbolicPlane.endpoints — Method.

endpoints(X::HPolygon) returns the list of vertices (in order) of the polygon.

HyperbolicPlane.endpoints — Method.

endpoints(T::HTriangle) returns the corner points of the triangle.

HyperbolicPlane.equality_threshold — Method.

equality_threshold(val) sets the "sloppiness" for equality checking. The default value is 100. In general, two objects are equal if their stored values are within val * eps(1.0).

Calling equality_threshold() with no arguments returns the current value.

HyperbolicPlane.equiangular — Method.

equiangular(n,theta) creates a regular n-gon where the vertex angles equal theta

HyperbolicPlane.equilateral — Method.

equilateral(n,s) creates a regular n-gon with side lengths s centered at the origin. First point is on the positive x-axis.

HyperbolicPlane.get_center — Method.

get_center(C::HCircle) returns the center of the circle.

HyperbolicPlane.get_radius — Method.

get_radius(C::HCircle) returns the radius of the circle.

HyperbolicPlane.get_vertex — Method.

get_vertex(R::HRay) returns the vertex (end point) of the ray.

HyperbolicPlane.getz — Method.

getz(P::HPoint) returns the point (complex number) in the interior of the unit disc that represents P.

HyperbolicPlane.interior_point — Method.

interior_point(T::HTriangle) returns a point in the interior of the triangle. The interior point is the intersection of the triangle's medians (the centroid).

HyperbolicPlane.is_simple — Method.

is_simple(X::HPolygon) determines if the polygon edges do not self-intersect. Be sure the polygon is legit using polygon_check first.

HyperbolicPlane.meet — Method.

meet(L,LL) finds a point on lines L and LL or throws an error if they don't intersect. Also for a line and a segment, or two segments.

See meet_check.

HyperbolicPlane.meet_check — Method.

meet_check(L::HLine,LL::HLine) determines if two lines intersect. Also for any combination of lines, segments, or rays.

HyperbolicPlane.midpoint — Method.

midpoint(p,q) finds the mid point of the line segment from p to q. Also midpoint(L::HSegment).

HyperbolicPlane.move2xplus — Method.

move2xplus(P::HPoint) returns an isometry of H^2 that maps P onto the positive real axis.

HyperbolicPlane.move2xplus — Method.

move2xplus(L::HLine) returns a linear fractional transformation that maps points on L to the positive x-axis but is not an isometry of the hyperbolic plane.

HyperbolicPlane.move2xplus — Method.

move2xplus(A,B) or move2xplus(L::HSegment) gives an isometry f so that f(A) is 0 and f(B) is on the positive real axis.

HyperbolicPlane.move2zero — Method.

move2zero(P::Hpoint) returns a LFT that's an isometry of H^2 that maps P to the origin.

HyperbolicPlane.npoints — Method.

npoints(X::HPolygon) returns the number of points on the polygon.

HyperbolicPlane.perimeter — Method.

perimeter(T::HTriangle) or perimeter(P::HPolygon) returns the perimeter of the figure.

HyperbolicPlane.perpendicular — Method.

perpendicular(L::HLine, P::HPoint) returns a line that is perpendicular to L and contains P.

HyperbolicPlane.perpendicular — Method.

perpendicular(L) returns an arbitrary line that is perpendicular to L.

HyperbolicPlane.point_on_line — Method.

point_on_line(L) returns a point on the hyperbolic line L.

See also: points_on_line.

HyperbolicPlane.point_on_ray — Method.

point_on_ray(R::HRay) returns a point in the interior of the ray.

HyperbolicPlane.points_on_circle — Method.

points_on_circle(C::HCircle) returns a 3-tuple of points that lie on the circle C.

HyperbolicPlane.points_on_line — Function.

points_on_line(L,n) returns a list of n distinct points on the line L.

See also: point_on_line.

HyperbolicPlane.polar — Method.

polar(P::HPoint) gives the polar coordinates of P

HyperbolicPlane.polygon_check — Function.

polygon_check(X::HPolygon,quiet=true) checks that the polygon is nondegenerate. Possible degeneracies are:

Repeated vertices

Fewer than three distinct vertices

Angles that are either 0 degrees or 180 degrees

If quiet is false, then a reason for the failed check is printed.

HyperbolicPlane.reflect_across — Method.

reflect_across(X::HObject,L::HSegment/HLine) returns the object formed by refecting X across the line segment/lineL`.

HyperbolicPlane.rotation — Method.

rotation(theta) is an isometry of H^2 corresponding to a rotation about the origin of the amount theta

HyperbolicPlane.same_side — Method.

same_side(P,Q,L) determines if the points P and Q lie in the same (closed) halfplane as determined by L. If either point is on L then the result is true.

HyperbolicPlane.set_color — Function.

set_color(X,col) sets the color of the hyperbolic object X.

HyperbolicPlane.set_fill_alpha — Function.

set_fill_alpha(X,alpha) sets the alpha value for the object's fill. Only works for HPlane and HCircle. Default value is 1 (not transparent)

HyperbolicPlane.set_fill_color — Function.

set_fill_color(X,col) sets the color used to fill X. Works for HPlane and HCircle. Default is :yellow

HyperbolicPlane.set_line_style — Function.

set_line_style(X,style) sets the line style for drawing X. Default is :solid.

HyperbolicPlane.set_no_fill — Method.

set_no_fill(X) removes fill from the HObject. It now draws as unfilled.

HyperbolicPlane.set_radius — Function.

set_radius(P,rad) sets the radius for an HPoint. The default is 1.

HyperbolicPlane.set_thickness — Function.

set_thickness(X,thk) sets the thickness of the line used to draw the hyperbolic object X. Default is 1.

HyperbolicPlane.sides — Method.

sides(P:::HPolygon/HTriangle) returns a list of the line segments that are the sides of the polygon.

HyperbolicPlane.stab_count — Method.

stab_count(R,targets) is used to count how many items in targets are intersected by R where targets is a list of HLinear objects and R is an HLinear.

HyperbolicPlane.tesselation — Function.

tesselation(n,k,deep): Tesselate the hyperbolic plane by regular n-gons in which each vertex is a corner of k polygons. deep controls how many layers. The center of the first k-gon is placed at the origin.

May also be called tesselation(n,k,deep,true) in which case a vertex of the first k-gon is placed at the origin and the tesselation is seeded by copies of this first polygon around the origin.

HyperbolicPlane.triangulate — Method.

triangulate(X::HPolygon) returns a list of triangles that triangulate X. The polygon should have at least three sides, and have no bad angles, and not self-intersect.

SimpleDrawing.draw — Method.

draw(X) draws the hyperbolic object X in a graphics window.

draw may be applied to a list of hyperbolic objects.

The typical sequence of drawing starts by clearing the screen with the plot() function (from the Plots module), then various to calls to draw and then concludes with a call to finish() (see the help message for that function).

Base.:- — Method.

-X where X is a hyperbolic object is a new object reflected through the origin.

Base.adjoint — Method.

adjoint(X::HObject) (that is, X') returns a new X that is reflected across the x-axis.

Base.delete! — Method.

delete!(C::HContainer, X::HObject) deletes X from the container C returning true if successful (or false if X was not in the container).

Base.in — Method.

in(p::HPoint, X::HPolygon) determines if the point lies on the boundary or, or is interior to, the polygon. Be sure that polygon_check(X) and is_simple(X) both return true.

Base.in — Method.

in(P::HPoint, R::HRay) determine if P lies on the ray R.

Base.in — Method.

in(P::HPoint, T::HTriangle) determines if P is in the triangle T, either in one of its sides or in its interior.

HyperbolicPlane._cycle — Method.

_cycle(A,k) returns a k-step shift of A. We require k to be in the interval [0,n-1] where n=length(A). No checking is done.

HyperbolicPlane._cyclic_equal — Method.

_cyclic_equal(A,B) checks if some cyclic shift of one list equals the other.

HyperbolicPlane._solver — Method.

_solver(f,goal,lo,hi) assumes f is increasing on the interval [lo,hi]

HyperbolicPlane._solver — Method.

_solver(f,goal) assumes that f is increasing and there is a nonnegative x so that f(x)==goal.

HyperbolicPlane.alt_mod — Method.

alt_mod(k,n) is mod(k,n) unless the result is zero, in which case we return n.

HyperbolicPlane.check_diagonal — Method.

check_diagonal(X::HPolygon, k::Int) checks to see if the diagonal between vertices k-1 and k+1 (a) does not intersect the boundary of X and (b) goes through the interior of X.

HyperbolicPlane.expander — Method.

expander(P::HPolygon) returns a list of polygons formed by reflecting P across each of its sides.

HyperbolicPlane.find_ear_diagonal — Method.

find_ear_diagonal(X::HPolygon) returns an index k such that the diagonal from k-1 to k+1 passes check_diagonal. Returns 0 if no ear diagonal is found.

HyperbolicPlane.solve_dist — Method.

solve_dist(d) is the inverse of _dist()