Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Hyperopt.BlueNoiseSampler — Type.

Try to spread out parameters as blue noise (only high frequency). Should sample the space better than random sampling. Use this if you intend to run less than, say, 2000.

source

Hyperopt.RandomSampler — Type.

Sample a value For each parameter uniformly at random from the candidate vectors. Log-uniform sampling available by providing a log-spaced candidate vector.

Hyperopt.TreeSampler — Type.

random_init::Int = 5 # Number of initial function with random sampling before a tree is built
samples_per_leaf::Int = 3   # Average number of samples per leaf
n_tries::Int = 20    # How many random samples to query the tree model with before selecting a sample

source