Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Hyperspecialize.allsubtypes — Method.

allsubtypes(t)

Return an Array containing all subtypes of t at load time.

Examples

julia> Hyperspecialize.allsubtypes(Real)
24-element Array{Type,1}:
 Real
 AbstractFloat
 BigFloat
 Float16
 Float32
 Float64
 AbstractIrrational
 Irrational
 Integer
 Bool
 Signed
 BigInt
 Int128
 Int16
 Int32
 Int64
 Int8
 Unsigned
 UInt128
 UInt16
 UInt32
 UInt64
 UInt8
 Rational

Hyperspecialize.concretesubtypes — Method.

concretesubtypes(t)

Return an Array containing all concrete subtypes of t at load time.

Examples

julia> Hyperspecialize.concretesubtypes(Real)
16-element Array{Any,1}:
 BigFloat
 Float16
 Float32
 Float64
 Bool
 BigInt
 Int128
 Int16
 Int32
 Int64
 Int8
 UInt128
 UInt16
 UInt32
 UInt64
 UInt8