Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ImageDraw.CirclePointRadius — Type.

circle = CirclePointRadius(center, ρ)

A Drawable circle having center center and radius ρ

ImageDraw.CircleThreePoints — Type.

circle = CircleThreePoints(p1, p2, p3)

A Drawable circle passing through points p1, p2 and p3

ImageDraw.Drawable — Type.

Type representing any object drawable on image

ImageDraw.Ellipse — Type.

ellipse = Ellipse(center, ρx, ρy)

A Drawable ellipse with center center and parameters ρx and ρy

ImageDraw.LineNormal — Type.

line = LineNormal(ρ, θ)

A Drawable infinte length line having perpendicular length ρ from origin and angle θ between the perpendicular and x-axis

ImageDraw.LineSegment — Type.

ls = LineSegment(p1, p2)

A Drawable finite length line between p1 and p2

ImageDraw.LineTwoPoints — Type.

line = LineTwoPoints(p1, p2)

A Drawable infinite length line passing through the two points p1 and p2.

ImageDraw.Path — Type.

path = Path([point])

A Drawable sequence of line segments connecting consecutive pairs of points in [point].

Note

This will create a non-closed path. For a closed path, see Polygon

ImageDraw.Point — Type.

p = Point(x,y)
p = Point(c)

A Drawable point on the image

ImageDraw.Polygon — Type.

polygon = Polygon([vertex])

A Drawable polygon i.e. a closed path created by joining the consecutive points in [vertex] along with the first and last point.

Note

This will create a closed path. For a non-closed path, see Path

ImageDraw.RegularPolygon — Type.

rp = RegularPolygon(center, side_count, side_length, θ)

A Drawable regular polygon.

#Arguments

center::Point : the center of the polygon

side_count::Int : number of sides of the polygon

side_length::Real : length of each side

θ::Real : orientation of the polygon w.r.t x-axis (in radians)

ImageDraw.draw! — Method.

img = draw!(img, drawable, color)
img = draw!(img, drawable)

Draws drawable on img using color color which defaults to oneunit(eltype(img))

ImageDraw.draw! — Method.

img = draw!(img, [drawable], [color])
img = draw!(img, [drawable] ,color)
img = draw!(img, [drawable])

Draws all objects in [drawable] in the given order on img using corresponding colors from [color] which defaults to oneunit(eltype(img)) If only a single color color is specified then all objects will be colored with that color.

ImageDraw.draw — Method.

img_new = draw(img, drawable, color)
img_new = draw(img, [drawable], [color])

Draws the drawable object on a copy of image img using color color. Can also draw multiple Drawable objects when passed as a AbstractVector{Drawable} with corresponding colors in [color]