Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ImageInTerminal.colorant2ansi — Method.

colorant2ansi(color::Colorant) -> Int

Converts the given colorant into an integer index that corresponds to the closest 256-colors ANSI code.

julia> colorant2ansi(RGB(1., 1., 0.))
226

This function also tries to make good use of the additional number of available shades of gray (ANSI codes 232 to 255).

julia> colorant2ansi(RGB(.5, .5, .5))
244

julia> colorant2ansi(Gray(.5))
244

ImageInTerminal.imshow — Function.

imshow([stream], img, [depth::TermColorDepth], [maxsize])

Displays the given image img using unicode characters and terminal colors (defaults to 256 colors). img has to be an array of Colorant.

If working in the REPL, the function tries to choose the encoding based on the current display size. The image will also be downsampled to fit into the display (using restrict).

ImageInTerminal.imshow24bit — Method.

imshow24bit([stream], img, [maxsize])

Displays the given image img using unicode characters and the 24 terminal colors that some modern terminals support. img has to be an array of Colorant.

If working in the REPL, the function tries to choose the encoding based on the current display size. The image will also be downsampled to fit into the display (using restrict).

ImageInTerminal.imshow256 — Method.

imshow256([stream], img, [maxsize])

Displays the given image img using unicode characters and the widely supported 256 terminal colors. img has to be an array of Colorant.

If working in the REPL, the function tries to choose the encoding based on the current display size. The image will also be downsampled to fit into the display (using restrict).

ImageInTerminal.encodeimg — Function.

encodeimg(enc::ImageEncoder, colordepth::TermColorDepth, img, [maxheight], [maxwidth])

Transforms the pixel of the given image img, which has to be an array of Colorant, into a string of unicode characters using ansi terminal colors.

The encoder enc specifies which kind of unicode represenation should be used.

The colordepth can either be TermColor256() or TermColor24bit() and specifies which terminal color codes should be used.

maxheight and maxwidth specify the maximum numbers of string characters that should be used for the resulting image. Larger images are downscaled automatically using restrict.

The function returns a tuple with three elements:

A vector of strings containing the encoded image. Each element represent one line. The lines do not contain newline characters.

Number of lines in the vector.

Number of visible characters per line (the remaining are colorcodes).

ImageInTerminal.use_24bit — Method.

use_24bit()

Triggers imshow24bit automatically if an array of colorants is to be displayed in the julia REPL. Call ImageInTerminal.use_256() to restore default behaviour.

ImageInTerminal.use_256 — Method.

use_256()

Triggers imshow256 automatically if an array of colorants is to be displayed in the julia REPL. (This is the default)