Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ImageInpainting.Crimisini — Type.

Crimisini(tilesize)

Examplar-based inpainting based on confidence and isophote maps. tilesize is the patch size as a tuple of integers.

References

Crimisini, A., Pérez, P., Toyama, K., 2004. Region Filling and Object Removal by Examplar-based Image Inpainting.

ImageInpainting.inpaint — Method.

inpaint(img, mask, algo)

Inpaint img on pixels marked as true in mask using algorithm algo.

ImageInpainting.InpaintAlgo — Type.

InpaintAlgo

An inpainting algorithm with given parameters.

ImageInpainting.convdist — Method.

convdist(img, kern; [weights])

Compute distance between all subimages of img and kern. Optionally, activate/deactivate pixels in the kern using weights.

ImageInpainting.imfilter_cpu — Method.

imfilter_cpu(img, kern)

Perform filtering on img with kernel kern using the FFT algorithm.

ImageInpainting.pointgradients — Method.

pointgradients(img, points; [method])

Compute the gradients along all dimensions of N-dimensional img at points using method. Default method is :ando3.

ImageInpainting.selectpatch — Method.

selectpatch(imgs, tilesize, center)

Helper function to extract a patch of size tilesize from imgs centered at Cartesian index center.