Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ImageSegmentation.ImageEdge — Type.

edge = ImageEdge(index1, index2, weight)

Construct an edge in a Region Adjacency Graph. index1 and index2 are the integers corresponding to individual pixels/voxels (in the sense of linear indexing via sub2ind), and weight is the edge weight (measures the dissimilarity between pixels/voxels).

ImageSegmentation.SegmentedImage — Type.

SegmentedImage type contains the index-label mapping, assigned labels, segment mean intensity and pixel count of each segment.

ImageSegmentation.fast_scanning — Method.

seg_img = fast_scanning(img, threshold, [diff_fn])

Segments the N-D image using a fast scanning algorithm and returns a SegmentedImage containing information about the segments.

Arguments:

img : N-D image to be segmented (arbitrary axes are allowed)

threshold : Upper bound of the difference measure (δ) for considering pixel into same segment; an AbstractArray can be passed having same number of dimensions as that of img for adaptive thresholding

diff_fn : (Optional) Function that returns a difference measure (δ) between the mean color of a region and color of a point

Examples:

julia> img = zeros(Float64, (3,3));

julia> img[2,:] .= 0.5;

julia> img[:,2] .= 0.6;

julia> seg = fast_scanning(img, 0.2);

julia> labels_map(seg)
3×3 Array{Int64,2}:
 1  4  5
 4  4  4
 3  4  6

Citation:

Jian-Jiun Ding, Cheng-Jin Kuo, Wen-Chih Hong, "An efficient image segmentation technique by fast scanning and adaptive merging"

ImageSegmentation.felzenszwalb — Function.

segments                = felzenszwalb(img, k, [min_size])
index_map, num_segments = felzenszwalb(edges, num_vertices, k, [min_size])

Segments an image using Felzenszwalb's graph-based algorithm. The function can be used in either of two ways -

segments = felzenszwalb(img, k, [min_size])

Segments an image using Felzenszwalb's segmentation algorithm and returns the result as SegmentedImage. The algorithm uses euclidean distance in color space as edge weights for the region adjacency graph.

Parameters:

img = input image

k = Threshold for region merging step. Larger threshold will result in bigger segments.

min_size = Minimum segment size

index_map, num_segments = felzenszwalb(edges, num_vertices, k, [min_size])

Segments an image represented as Region Adjacency Graph(RAG) using Felzenszwalb's segmentation algorithm. Each pixel/region corresponds to a node in the graph and weights on each edge measure the dissimilarity between pixels. The function returns the number of segments and index mapping from nodes of the RAG to segments.

Parameters:

edges = Array of edges in RAG. Each edge is represented as ImageEdge.

num_vertices = Number of vertices in RAG

k = Threshold for region merging step. Larger threshold will result in bigger segments.

min_size = Minimum segment size

ImageSegmentation.hmin_transform — Method.

out = hmin_transform(img, h)

Suppresses all minima in grayscale image whose depth is less than h.

H-minima transform is defined as the reconstruction by erosion of (img + h) by img. See Morphological image analysis by Soille pg 170-172.

ImageSegmentation.labels_map — Method.

img_labeled = labels_map(seg)

Return an array containing the label assigned to each pixel.

ImageSegmentation.meanshift — Method.

segments                = meanshift(img, spatial_radius, range_radius; iter=50, eps=0.01))

Segments the image using meanshift clustering. Returns a SegmentedImage.

Parameters:

img = input grayscale image

spatialradius/rangeradius = bandwidth parameters of truncated normal kernel. Controlling the size of the kernel determines the resolution of the mode detection.

iter/eps = stopping criterion for meanshift procedure. The algorithm stops after iter iterations or if kernel center moves less than eps distance in an update step, whichever comes first.

ImageSegmentation.prune_segments — Method.

new_seg = prune_segments(seg, rem_labels, diff_fn)

Removes all segments that have labels in rem_labels replacing them with their neighbouring segment having least diff_fn. rem_labels is a Vector of labels.

new_seg = prune_segments(seg, is_rem, diff_fn)

Removes all segments for which is_rem returns true replacing them with their neighbouring segment having least diff_fn.

is_rem(label) -> Bool

Returns true if label label is to be removed otherwise false.

d = diff_fn(rem_label, neigh_label)

A difference measure between label to be removed and its neighbors. isless must be defined for objects of the type of d.

ImageSegmentation.region_adjacency_graph — Method.

G, vert_map = region_adjacency_graph(seg, weight_fn)

Constructs a region adjacency graph (RAG) from the SegmentedImage. It returns the RAG along with a Dict(label=>vertex) map. weight_fn is used to assign weights to the edges.

weight_fn(label1, label2)

Returns a real number corresponding to the weight of the edge between label1 and label2.

ImageSegmentation.region_splitting — Method.

seg = region_splitting(img, homogeneous)

Segments img by recursively splitting it until all the segments are homogeneous.

b = homogeneous(img)

Returns true if img is homogeneous.

Examples

julia> img = 0.1*rand(6, 6);

julia> img[4:end, 4:end] .+= 10;

julia> function homogeneous(img)
           min, max = extrema(img)
           max - min < 0.2
       end
homogeneous (generic function with 1 method)

julia> seg = region_splitting(img, homogeneous);

ImageSegmentation.region_tree — Method.

t = region_tree(img, homogeneous)

Creates a region tree from img by splitting it recursively until all the regions are homogeneous.

b = homogeneous(img)

Returns true if img is homogeneous.

Examples

julia> img = 0.1*rand(6, 6);

julia> img[4:end, 4:end] .+= 10;

julia> function homogeneous(img)
           min, max = extrema(img)
           max - min < 0.2
       end
homogeneous (generic function with 1 method)

julia> t = region_tree(img, homogeneous);

ImageSegmentation.rem_segment! — Method.

rem_segment!(seg, label, diff_fn)

In place removal of the segment having label label, replacing it with the neighboring segment having least diff_fn value.

d = diff_fn(rem_label, neigh_label)

A difference measure between label to be removed and its neighbors. isless must be defined for objects of the type of d.

Examples

    # This removes the label `l` and replaces it with the label of
    # neighbor having maximum pixel count.
    julia> rem_segment!(seg, l, (i,j)->(-seg.segment_pixel_count[j]))

    # This removes the label `l` and replaces it with the label of
    # neighbor having the least value of euclidian metric.
    julia> rem_segment!(seg, l, (i,j)->sum(abs2, seg.segment_means[i]-seg.segment_means[j]))

ImageSegmentation.rem_segment — Method.

new_seg = rem_segment(seg, label, diff_fn)

Removes the segment having label label and returns the new SegmentedImage. For more info, see remove_segment!

ImageSegmentation.seeded_region_growing — Method.

seg_img = seeded_region_growing(img, seeds, [kernel_dim], [diff_fn])
seg_img = seeded_region_growing(img, seeds, [neighbourhood], [diff_fn])

Segments the N-D image img using the seeded region growing algorithm and returns a SegmentedImage containing information about the segments.

Arguments:

img : N-D image to be segmented (arbitrary axes are allowed)

seeds : Vector containing seeds. Each seed is a Tuple of a CartesianIndex{N} and a label. See below note for more information on labels.

kernel_dim : (Optional) Vector{Int} having length N or a NTuple{N,Int} whose ith element is an odd positive integer representing the length of the ith edge of the N-orthotopic neighbourhood

neighbourhood : (Optional) Function taking CartesianIndex{N} as input and returning the neighbourhood of that point.

diff_fn : (Optional) Function that returns a difference measure(δ) between the mean color of a region and color of a point

Note

The labels attached to points must be positive integers, although multiple points can be assigned the same label. The output includes a labelled array that has same indexing as that of input image. Every index is assigned to either one of labels or a special label '0' indicating that the algorithm was unable to assign that index to a unique label.

Examples

julia> img = zeros(Gray{N0f8},4,4);

julia> img[2:4,2:4] .= 1;

julia> seeds = [(CartesianIndex(3,1),1),(CartesianIndex(2,2),2)];

julia> seg = seeded_region_growing(img, seeds);

julia> labels_map(seg)
4×4 Array{Int64,2}:
 1  1  1  1
 1  2  2  2
 1  2  2  2
 1  2  2  2

Citation:

Albert Mehnert, Paul Jackaway (1997), "An improved seeded region growing algorithm", Pattern Recognition Letters 18 (1997), 1065-1071

ImageSegmentation.segment_labels — Method.

labels = segment_labels(seg)

Returns the list of assigned labels

ImageSegmentation.segment_mean — Method.

m = segment_mean(seg, l)

Returns the mean intensity of label l. If no label is supplied, it returns a Dict(label=>mean) of all the labels.

ImageSegmentation.segment_pixel_count — Method.

c = segment_pixel_count(seg, l)

Returns the count of pixels that are assigned label l. If no label is supplied, it returns a Dict(label=>pixel_count) of all the labels.

ImageSegmentation.unseeded_region_growing — Method.

seg_img = unseeded_region_growing(img, threshold, [kernel_dim], [diff_fn])
seg_img = unseeded_region_growing(img, threshold, [neighbourhood], [diff_fn])

Segments the N-D image using automatic (unseeded) region growing algorithm and returns a SegmentedImage containing information about the segments.

Arguments:

img : N-D image to be segmented (arbitrary axes are allowed)

threshold : Upper bound of the difference measure (δ) for considering pixel into same segment

kernel_dim : (Optional) Vector{Int} having length N or a NTuple{N,Int} whose ith element is an odd positive integer representing the length of the ith edge of the N-orthotopic neighbourhood

neighbourhood : (Optional) Function taking CartesianIndex{N} as input and returning the neighbourhood of that point.

diff_fn : (Optional) Function that returns a difference measure (δ) between the mean color of a region and color of a point

Examples

julia> img = zeros(Gray{N0f8},4,4);

julia> img[2:4,2:4] .= 1;

julia> seg = unseeded_region_growing(img, 0.2);

julia> labels_map(seg)
4×4 Array{Int64,2}:
 1  1  1  1
 1  2  2  2
 1  2  2  2
 1  2  2  2

ImageSegmentation.watershed — Method.

segments                = watershed(img, markers)

Segments the image using watershed transform. Each basin formed by watershed transform corresponds to a segment. If you are using image local minimas as markers, consider using hmin_transform to avoid oversegmentation.

Parameters:

img = input grayscale image

markers = An array (same size as img) with each region's marker assigned a index starting from 1. Zero means not a marker. If two markers have the same index, their regions will be merged into a single region. If you have markers as a boolean array, use label_components.