Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ImageTransformations.InvWarpedView — Type.

InvWarpedView(img, tinv, [indices]) -> wv

Create a view of img that lazily transforms any given index I passed to wv[I] to correspond to img[inv(tinv)(I)]. While technically this approach is known as backward mode warping, note that InvWarpedView is created by supplying the forward transformation

The conceptual difference to WarpedView is that InvWarpedView is intended to be used when reasoning about the image is more convenient that reasoning about the indices. Furthermore, InvWarpedView allows simple nesting of transformations, in which case the transformations will be composed into a single one.

The optional parameter indices can be used to specify the domain of the resulting wv. By default the indices are computed in such a way that wv contains all the original pixels in img.

see invwarpedview for more information.

ImageTransformations.WarpedView — Type.

WarpedView(img, tform, [indices]) -> wv

Create a view of img that lazily transforms any given index I passed to wv[I] to correspond to img[tform(I)]. This approach is known as backward mode warping.

The optional parameter indices can be used to specify the domain of the resulting wv. By default the indices are computed in such a way that wv contains all the original pixels in img. To do this inv(tform) has to be computed. If the given transformation tform does not support inv, then the parameter indices has to be specified manually.

see warpedview for more information.

ImageTransformations.imresize — Method.

imresize(img, sz) -> imgr
imresize(img, inds) -> imgr
imresize(img; ratio) -> imgr

Change img to be of size sz (or to have indices inds). If ratio is used, then sz = ceil(Int, size(img).*ratio). This interpolates the values at sub-pixel locations. If you are shrinking the image, you risk aliasing unless you low-pass filter img first.

Examples

julia> img = testimage("lena_gray_256") # 256*256
julia> imresize(img, 128, 128) # 128*128
julia> imresize(img, 1:128, 1:128) # 128*128
julia> imresize(img, (128, 128)) # 128*128
julia> imresize(img, (1:128, 1:128)) # 128*128
julia> imresize(img, (1:128, )) # 128*256
julia> imresize(img, 128) # 128*256
julia> imresize(img, ratio = 0.5) # 128*128

σ = map((o,n)->0.75*o/n, size(img), sz)
kern = KernelFactors.gaussian(σ)   # from ImageFiltering
imgr = imresize(imfilter(img, kern, NA()), sz)

See also restrict.

ImageTransformations.imrotate — Method.

imrotate(img, θ, [indices], [degree = Linear()], [fill = NaN]) -> imgr

Rotate image img by θ∈[0,2π) in a clockwise direction around its center point. To rotate the image counterclockwise, specify a negative value for angle.

By default, rotated image imgr will not be cropped. Bilinear interpolation will be used and values outside the image are filled with NaN if possible, otherwise with 0.

Examples

julia> img = testimage("cameraman")

# rotate with bilinear interpolation but without cropping 
julia> imrotate(img, π/4)

# rotate with bilinear interpolation and with cropping
julia> imrotate(img, π/4, axes(img))

# rotate with nearest interpolation but without cropping
julia> imrotate(img, π/4, Constant())

See also warp.

ImageTransformations.invwarpedview — Method.

invwarpedview(img, tinv, [indices], [degree = Linear()], [fill = NaN]) -> wv

Create a view of img that lazily transforms any given index I passed to wv[I] to correspond to img[inv(tinv)(I)]. While technically this approach is known as backward mode warping, note that InvWarpedView is created by supplying the forward transformation. The given transformation tinv must accept a SVector as input and support inv(tinv). A useful package to create a wide variety of such transformations is CoordinateTransformations.jl.

When invoking wv[I], values for img must be reconstructed at arbitrary locations inv(tinv)(I). InvWarpedView serves as a wrapper around WarpedView which takes care of interpolation and extrapolation. The parameters degree and fill can be used to specify the b-spline degree and the extrapolation scheme respectively.

The optional parameter indices can be used to specify the domain of the resulting wv. By default the indices are computed in such a way that wv contains all the original pixels in img.

ImageTransformations.restrict — Method.

restrict(img[, region]) -> imgr

Reduce the size of img by two-fold along the dimensions listed in region, or all spatial coordinates if region is not specified. It anti-aliases the image as it goes, so is better than a naive summation over 2x2 blocks.

See also imresize.

ImageTransformations.warp — Method.

warp(img, tform, [indices], [degree = Linear()], [fill = NaN]) -> imgw

Transform the coordinates of img, returning a new imgw satisfying imgw[I] = img[tform(I)]. This approach is known as backward mode warping. The transformation tform must accept a SVector as input. A useful package to create a wide variety of such transformations is CoordinateTransformations.jl.

Reconstruction scheme

During warping, values for img must be reconstructed at arbitrary locations tform(I) which do not lie on to the lattice of pixels. How this reconstruction is done depends on the type of img and the optional parameter degree.

When img is a plain array, then on-grid b-spline interpolation will be used. It is possible to configure what degree of b-spline to use with the parameter degree. For example one can use degree = Linear() for linear interpolation, degree = Constant() for nearest neighbor interpolation, or degree = Quadratic(Flat()) for quadratic interpolation.

In the case tform(I) maps to indices outside the original img, those locations are set to a value fill (which defaults to NaN if the element type supports it, and 0 otherwise). The parameter fill also accepts extrapolation schemes, such as Flat(), Periodic() or Reflect().

For more control over the reconstruction scheme –- and how beyond-the-edge points are handled –- pass img as an AbstractInterpolation or AbstractExtrapolation from Interpolations.jl.

The meaning of the coordinates

The output array imgw has indices that would result from applying inv(tform) to the indices of img. This can be very handy for keeping track of how pixels in imgw line up with pixels in img.

If you just want a plain array, you can "strip" the custom indices with parent(imgw).

Examples: a 2d rotation (see JuliaImages documentation for pictures)

julia> using Images, CoordinateTransformations, TestImages, OffsetArrays

julia> img = testimage("lighthouse");

julia> axes(img)
(Base.OneTo(512),Base.OneTo(768))

# Rotate around the center of `img`
julia> tfm = recenter(RotMatrix(-pi/4), center(img))
AffineMap([0.707107 0.707107; -0.707107 0.707107], [-196.755,293.99])

julia> imgw = warp(img, tfm);

julia> axes(imgw)
(-196:709,-68:837)

# Alternatively, specify the origin in the image itself
julia> img0 = OffsetArray(img, -30:481, -384:383);  # origin near top of image

julia> rot = LinearMap(RotMatrix(-pi/4))
LinearMap([0.707107 -0.707107; 0.707107 0.707107])

julia> imgw = warp(img0, rot);

julia> axes(imgw)
(-293:612,-293:611)

julia> imgr = parent(imgw);

julia> axes(imgr)
(Base.OneTo(906),Base.OneTo(905))

ImageTransformations.warpedview — Method.

warpedview(img, tform, [indices], [degree = Linear()], [fill = NaN]) -> wv

Create a view of img that lazily transforms any given index I passed to wv[I] to correspond to img[tform(I)]. This approach is known as backward mode warping. The given transformation tform must accept a SVector as input. A useful package to create a wide variety of such transformations is CoordinateTransformations.jl.

When invoking wv[I], values for img must be reconstructed at arbitrary locations tform(I) which do not lie on to the lattice of pixels. How this reconstruction is done depends on the type of img and the optional parameter degree. When img is a plain array, then on-grid b-spline interpolation will be used, where the pixel of img will serve as the coeficients. It is possible to configure what degree of b-spline to use with the parameter degree. The two possible values are degree = Linear() for linear interpolation, or degree = Constant() for nearest neighbor interpolation.

In the case tform(I) maps to indices outside the domain of img, those locations are set to a value fill (which defaults to NaN if the element type supports it, and 0 otherwise). Additionally, the parameter fill also accepts extrapolation schemes, such as Flat(), Periodic() or Reflect().

The optional parameter indices can be used to specify the domain of the resulting WarpedView. By default the indices are computed in such a way that the resulting WarpedView contains all the original pixels in img. To do this inv(tform) has to be computed. If the given transformation tform does not support inv, then the parameter indices has to be specified manually.

warpedview is essentially a non-coping, lazy version of warp. As such, the two functions share the same interface, with one important difference. warpedview will insist that the resulting WarpedView will be a view of img (i.e. parent(warpedview(img, ...)) === img). Consequently, warpedview restricts the parameter degree to be either Linear() or Constant().