Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ImageView.CLim — Type.

CLim(cmin, cmax)

Specify contrast limits where x <= cmin will be rendered as black, and x >= cmax will be rendered as white.

ImageView.canvasgrid — Function.

grid, frames, canvases = canvasgrid((ny, nx))

Create a grid of ny-by-nx canvases for drawing. grid is a GtkGrid layout, frames is an ny-by-nx array of GtkAspectRatioFrames that contain each canvas, and canvases is an ny-by-nx array of canvases.

ImageView.imlink — Method.

imlink(imgs...; gridsize=(1,length(imgs)), dims=(1,2))

Show multiple images in a single window, linking higher-dimensional axes to shared GUI control(s).

ImageView.imshow! — Method.

imshow!(canvas, img) -> drawsignal
imshow!(canvas, img::Signal, zr::Signal{ZoomRegion}) -> drawsignal
imshow!(frame::Frame, canvas, img::Signal, zr::Signal{ZoomRegion}) -> drawsignal
imshow!(..., annotations=Signal(Dict{UInt,Any}()))

Display the image img, in the specified canvas. Use the version with zr if you have already turned on rubber-banding or other pan/zoom interactivity for canvas. Returns the Reactive drawsignal used for updating the canvas.

If you supply frame, then the pixel aspect ratio will be set to that of pixelspacing(img).

With any of these forms, you may optionally supply annotations.

This only creates the draw method for canvas; mouse- or key-based interactivity can be set up via imshow or, at a lower level, using GtkReactive's tools:

init_zoom_rubberband

init_zoom_scroll

init_pan_scroll

init_pan_drag

Example

```julia using ImageView, GtkReactive, Gtk.ShortNames, TestImages

Create a window with a canvas in it

win = Window() c = canvas(UserUnit) push!(win, c) Gtk.showall(win)

Load images

mri = testimage("mri")

Display the image

imshow!(c, mri[:,:,1])

Update with a different image

imshow!(c, mri[:,:,8])

ImageView.imshow — Method.

imshow(img; axes=(1,2), name="ImageView") -> guidict
imshow(img, clim; axes=(1,2), name="ImageView") -> guidict
imshow(img, clim, zoomregion, slicedata, annotations; axes=(1,2), name="ImageView") -> guidict

Display the image img in a new window titled with name, returning a dictionary guidict containing any Reactive signals or GtkReactive widgets. If the image is 3 or 4 dimensional, GUI controls will be added for slicing along "extra" axes. By default the two-dimensional slice containing axes 1 and 2 are shown, but that can be changed by passing a different setting for axes.

If the image is grayscale, by default contrast is set by a scaleminmax object whose end-points can be modified by right-clicking on the image. If clim == nothing, the image's own native contrast is used (clamp01nan). You may also pass a custom contrast function.

Finally, you may specify GtkReactive.ZoomRegion and SliceData signals. See also roi, as well as any annotations that you wish to apply.

ImageView.imshow — Method.

imshow()

Choose an image to display via a file dialog.

ImageView.imshow — Method.

imshow(canvas, imgsig::Signal) -> guidict
imshow(canvas, imgsig::Signal, zr::Signal{ZoomRegion}) -> guidict
imshow(frame::Frame, canvas, imgsig::Signal, zr::Signal{ZoomRegion}) -> guidict

Display imgsig (a Signal of an image) in canvas, setting up panning and zooming. Optionally include a frame for preserving aspect ratio. imgsig must be two-dimensional (but can be a Signal-view of a higher-dimensional object).

Example

using ImageView, TestImages, Gtk
mri = testimage("mri");
# Create a canvas `c`. There are other approaches, like stealing one from a previous call
# to `imshow`, or using GtkReactive directly.
guidict = imshow_gui((300, 300))
c = guidict["canvas"];
# To see anything you have to call `showall` on the window (once)
Gtk.showall(guidict["window"])
# Create the image Signal
imgsig = Signal(mri[:,:,1]);
# Show it
imshow(c, imgsig)
# Now anytime you want to update, just push! a new image
push!(imgsig, mri[:,:,8])

ImageView.imshow_gui — Function.

guidict = imshow_gui(canvassize; name="ImageView", aspect=:auto)
guidict = imshow_gui(canvassize, slicedata, gridsize=(1,1); name="ImageView", aspect=:auto)

Create an image-viewer GUI. By default creates a single canvas, but with custom gridsize you can create a grid of canvases. canvassize = (szx, szy) describes the desired size of the (or each) canvas. slicedata is an object created by roi that encodes the necessary information for creating player widgets for viewing multidimensional images.

ImageView.roi — Method.

roi(A) -> zr::Signal(ZoomRegion), slicedata::SliceData
roi(A, dims=(1,2)) -> zr::Signal(ZoomRegion), slicedata::SliceData
roi(A, (:namey, :namex)) -> zr::Signal(ZoomRegion), slicedata::SliceData

Create the initial "region of interest" for viewing A. For multidimensional objects, optionally select two dimensions (the first two, by default) for slicing. The outputs zr and slicedata describe the within-view and player-controlled axes, respectively.

See also: slice2d.

ImageView.scalebar — Method.

scalebar(guidict::Dict, length; x = 0.8, y = 0.1, color = RGB(1,1,1))

Add a scale bar annotation to the image display controlled by guidict (returned by imshow). If the pixelspacing of the image is set using Unitful quantities, length should also be expressed in physical units.

x and y control the placement of the scalebar, and color its rendered color.

ImageView.slice2d — Method.

slice2d(A, zr, sd) -> A2

Create a two-dimensional slice A2 using the current ZoomRegion zr and SliceData sd.

ImageView.SliceData — Method.

SliceData{transpose::Bool}(signals::NTuple{N,Signal{Int}}, axes::NTuple{N,Axes})

Specify slice information for a (possibly) multidimensional image. signals hold the currently-selected slices for the selected axes, all of which are effectively "orthogonal" to the plane in the viewer.

ImageView.canvas_size — Method.

canvas_size(win, requested_size) -> (xsz, ysz)
canvas_size(screensize, requested_size) -> (xsz, ysz)

Limit the requested canvas size by the screen size. Both the output and screensize are supplied in Gtk order (x, y).

When supplying a GtkWindow win, the canvas size is limited to 60% of the total screen size.

ImageView.closeall — Method.

closeall()

Closes all windows opened by ImageView2.

ImageView.default_canvas_size — Function.

default_canvas_size(imagesize, pixelaspectratio=1) -> (xsz, ysz)

Compute the canvas size for an image of size imagesize with the defined pixelaspectratio. Note that imagesize is supplied in coordinate order, i.e., (y, x) order, whereas the returned canvas size is in Gtk order, i.e., (x, y) order.

ImageView.imshowlabeled — Method.

imshowlabeled(img, label)

Display img, but showing the pixel's label rather than the color value in the status bar.

ImageView.normalized_lengths — Method.

wnorm, hnorm = normalized_lengths(img, width, height)

Convert absolute width and height into normalized variants. Depends on the pixelspacing of img.

ImageView.sliceinds — Method.

inds = sliceinds(img, roi, slices...)
inds = sliceinds(axs, roi, slices...)

Return an indices-tuple inds that selects the region-of-interest roi at particular slices along orthogonal coordinates. Use AxisArrays.Axis to indicate out-of-order axes, either by dimension number (e.g., Axis{3}) or by name (Axis{:z}). If you're using dimension numbers, slices must list them in increasing order.

Examples:

julia> using AxisArrays: Axis

# Mimic a 4d array
julia> axs = (1:1080, 1:1920, 1:20, 1:1000)
(1:1080, 1:1920, 1:20, 1:1000)

julia> ImageView.sliceinds(axs, (6:10, 1:5), Axis{3}(7), Axis{4}(15))
(6:10, 1:5, 7, 15)

julia> ImageView.sliceinds(axs, (6:10, 1:5), Axis{2}(7), Axis{4}(15))
(6:10, 7, 1:5, 15)

# Mimic a 4d AxisArray
julia> axs = (Axis{:y}(1:1080), Axis{:x}(1:1920), Axis{:z}(1:20), Axis{:t}(1:1000))
(Axis{:y,UnitRange{Int64}}(1:1080), Axis{:x,UnitRange{Int64}}(1:1920), Axis{:z,UnitRange{Int64}}(1:20), Axis{:t,UnitRange{Int64}}(1:1000))

julia> ImageView.sliceinds(axs, (6:10, 1:5), Axis{:t}(15), Axis{:z}(7))   # out-of-order OK
(6:10, 1:5, 7, 15)

julia> ImageView.sliceinds(axs, (6:10, 1:5), Axis{:y}(15), Axis{:z}(7))
(15, 6:10, 7, 1:5)