Context

Impute.Context — Type.

Context

Stores common summary information for all Imputor types.

Fields

num::Int: number of observations

count::Int: number of missing values found

limit::Float64: allowable limit for missing values to impute

missing::Function: returns a Bool if the value counts as missing.

Base.findfirst — Method.

findfirst(ctx::Context, data::AbstractVector) -> Int

Returns the first not missing index in data.

Arguments

ctx::Context: the context to pass into ismissing

data::AbstractVector: the data array to search

Returns

Int: the first index in data that isn't missing

Base.findlast — Method.

findlast(ctx::Context, data::AbstractVector) -> Int

Returns the last not missing index in data.

Arguments

ctx::Context: the context to pass into ismissing

data::AbstractVector: the data array to search

Returns

Int: the last index in data that isn't missing

Base.findnext — Method.

findnext(ctx::Context, data::AbstractVector) -> Int

Returns the next not missing index in data.

Arguments

ctx::Context: the context to pass into ismissing

data::AbstractVector: the data array to search

Returns

Int: the next index in data that isn't missing

Base.ismissing — Method.

ismissing(ctx::Context, x) -> Bool

Uses ctx.missing to determine if x is missing. If x is a data row or an abstract array then ismissing will return true if ctx.missing returns true for any element. The ctx.count is increased whenever whenever we return true and if ctx.count / ctx.num exceeds our ctx.limit we throw an ImputeError

Arguments

ctx::Context: the contextual information about missing information.

x: the value to check (may be an single values, abstract array or row)