Impute

Impute.ImputeError — Type.

ImputeError{T} <: Exception

Is thrown by impute methods when the limit of imputable values has been exceeded.

Fields

msg::T - the message to print.

Base.Iterators.drop — Method.

drop(data::Dataset; limit=1.0)

Utility method for impute(data, :drop; limit=limit)

Impute.chain! — Method.

chain!(data::Dataset, missing::Function, imputors::Imputor...; kwargs...)

Creates a Chain with imputors and calls impute!(imputor, missing, data; kwargs...)

Impute.chain! — Method.

chain!(data::Dataset, imputors::Imputor...; kwargs...)

Creates a Chain with imputors and calls impute!(imputor, data; kwargs...)

Impute.chain — Method.

chain(data::Dataset, args...; kwargs...)

Copies the data before calling chain!(data, args...; kwargs...)

Impute.drop! — Method.

drop!(data::Dataset; limit=1.0)

Utility method for impute!(data, :drop; limit=limit)

Impute.impute! — Method.

impute!(data::Dataset, missing::Function, method::Symbol=:interp, args...; limit::Float64=0.1)

Creates the appropriate Imputor type and Context (using missing function) in order to call impute!(imputor::Imputor, ctx::Context, data::Dataset) with them.

Arguments

data::Dataset: the datset containing missing elements we should impute.

missing::Function: the missing data function to use

method::Symbol: the imputation method to use (options: [:drop, :fill, :interp, :locf, :nocb])

args::Any...: any arguments you should pass to the Imputor constructor.

limit::Float64: missing data ratio limit/threshold (default: 0.1)

Impute.impute! — Method.

impute!(data::Dataset, method::Symbol=:interp, args...; limit::Float64=0.1)

Looks up the Imputor type for the method, creates it and calls impute!(imputor::Imputor, data::Dataset, limit::Float64) with it.

Arguments

data::Dataset: the datset containing missing elements we should impute.

method::Symbol: the imputation method to use (options: [:drop, :fill, :interp, :locf, :nocb])

args::Any...: any arguments you should pass to the Imputor constructor.

limit::Float64: missing data ratio limit/threshold (default: 0.1)

Impute.impute — Method.

impute(data::Dataset, args...; kwargs...)

Copies the data before calling impute!(new_data, args...; kwargs...)

Impute.interp! — Method.

interp!(data::Dataset; limit=1.0)

Utility method for impute!(data, :interp; limit=limit)

Impute.interp — Method.

interp(data::Dataset; limit=1.0)

Utility method for impute(data, :interp; limit=limit)