Imputors

Impute.impute! — Function.

impute!(imp::Imputor, data::Dataset, limit::Float64=0.1)

Creates a Context using information about data. These include

missing data function which defaults to missing

number of elements: *(size(data)...)

Arguments

imp::Imputor: the Imputor method to use

data::Dataset: the data to impute

limit::Float64: missing data ratio limit/threshold (default: 0.1)

Return

Dataset: the input data with values imputed.

Impute.impute! — Method.

impute!(imp::Imputor, ctx::Context, data::AbstractMatrix)

Imputes the data in a matrix by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

ctx::Context: the contextual information for missing data

data::AbstractMatrix: the data to impute

Returns

AbstractMatrix: the input data with values imputed

Impute.impute! — Method.

impute!(imp::Imputor, ctx::Context, data::DataFrame)

Imputes the data in a DataFrame by imputing the values 1 column at a time; if this is not the desired behaviour custom imputor methods should overload this method.

Arguments

imp::Imputor: the Imputor method to use

ctx::Context: the contextual information for missing data

data::DataFrame: the data to impute

Returns

DataFrame: the input data with values imputed

Drop

Impute.impute! — Method.

impute!(imp::Drop, ctx::Context, data::AbstractVector)

Uses filter! to remove missing elements from the array.

Arguments

imp::Drop: this Imputor method

ctx::Context: contextual information for missing data

data::AbstractVector: the data to impute

Returns

AbstractVector: our data array with missing elements removed

Impute.impute! — Method.

impute!(imp::Drop, ctx::Context, data::AbstractMatrix)

Finds the missing rows in the matrix and uses a mask (Vector{Bool}) to return the data with those rows removed. Unfortunately, the mask approach requires copying the matrix.

NOTES (or premature optimizations):

We use view, but this will change the type of the data by returning a SubArray

We might be able to do something clever by:

reshaping the data to a vector

running deleteat! for the appropriate indices and

reshaping the data back to the desired shape.

Arguments

imp::Drop: this Imputor method

ctx::Context: contextual information for missing data

data::AbstractMatrix: the data to impute

Returns

AbstractMatrix: a new matrix with missing rows removed

Impute.impute! — Method.

impute!(imp::Drop, ctx::Context, data::DataFrame)

Finds the missing rows in the DataFrame and deletes them.

NOTE: this isn't quite as fast as dropnull in DataFramess as we're using an arbitrary missing function rather than using the precomputed dt.isnull vector of bools.

Arguments

imp::Drop: this Imputor method

ctx::Context: contextual information for missing data

data::DataFrame: the data to impute

Returns

DataFrame: our data with the missing rows removed.

Impute.Drop — Type.

Drop <: Imputor

Removes missing values from the AbstractArray or DataFrame provided.

Fill

Impute.impute! — Method.

impute!(imp::Fill, ctx::Context, data::AbstractVector)

Computes the fill value if imp.value is a Function (i.e., imp.value(drop(copy(data)))) and replaces all missing values in the data with that value.

Impute.Fill — Type.

Fill <: Imputor

Fills in the missing data with a specific value.

Fields

value::Any: A scalar missing value or a function that returns the a scalar if passed the data with missing data removed (e.g, mean)

Impute.Fill — Method.

Fill() -> Fill

By default Fill() will use the mean of the existing values as the fill value.

Interpolate

Impute.impute! — Method.

impute!(imp::Interpolate, ctx::Context, data::AbstractVector)

Uses linear interpolation between existing elements of a vector to fill in missing data.

WARNING: Missing values at the head or tail of the array cannot be interpolated if there are no existing values on both sides. As a result, this method does not guarantee that all missing values will be imputed.

Impute.Interpolate — Type.

Interpolate <: Imputor

Performs linear interpolation between the nearest values in an vector.

Last Observation Carried Forward (LOCF)

Impute.impute! — Method.

impute!(imp::LOCF, ctx::Context, data::AbstractVector)

Iterates forwards through the data and fills missing data with the last existing observation.

WARNING: missing elements at the head of the array may not be imputed if there is no existing observation to carry forward. As a result, this method does not guarantee that all missing values will be imputed.

Usage

Next Observation Carried Backward (NOCB)

Impute.impute! — Method.

impute!(imp::NOCB, ctx::Context, data::AbstractVector)

Iterates backwards through the data and fills missing data with the next existing observation.

WARNING: missing elements at the tail of the array may not be imputed if there is no existing observation to carry backward. As a result, this method does not guarantee that all missing values will be imputed.

Usage

Impute.NOCB — Type.

NOCB <: Imputor

Fills in missing data using the Next Observation Carried Backward (NOCB) approach.

Chain

Impute.impute! — Method.

impute!(imp::Chain, missing::Function, data::Dataset; limit::Float64=0.1)

Creates a Context and runs the Imputors on the supplied data.

Arguments

imp::Chain: the chain to run

missing::Function: the missing function to use in the Context to pass to the Imputors

data::Dataset: our data to impute

limit::Float64: the missing data ration limit/threshold

Returns

Dataset: our imputed data

Impute.impute! — Method.

impute!(imp::Chain, data::Dataset; limit::Float64=0.1)

Infers the missing data function from the data and passes that to impute!(imp::Chain, missing::Function, data::Dataset; limit::Float64=0.1).

Arguments

imp::Chain: the chain to run

data::Dataset: our data to impute

limit::Float64: the missing data ration limit/threshold

Returns

Dataset: our imputed data

Impute.Chain — Type.

Chain <: Imputor

Runs multiple Imputors on the same data in the order they're provided.

Fields

imputors::Array{Imputor}

Impute.Chain — Method.

Chain(imputors::Imputor...) -> Chain

Creates a Chain using the Imputors provided (ordering matters).