Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

InPlace.inclusiveinplace! — Method.

a = inclusiveinplace!(op, a, b...)

Compute and return op(a, b...). If a is mutable, possibly modify its value in-place.

In the case where a is mutable, it is an implementation detail whether its value is actually modified, and for this reason, one should always also assign the result of this call to a. Moreover, one should use it only on values for which the current stackframe holds the only reference; e.g. by using deepcopy.

InPlace.inplace! — Method.

a = inplace!(op, a, b...)

Compute and return op(b...). If a is mutable, possibly modify its value in-place.

In the case where a is mutable, it is an implementation detail whether its value is actually modified, and for this reason, one should always also assign the result of this call to a. Moreover, one should use it only on values for which the current stackframe holds the only reference; e.g. by using deepcopy.