Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

IncrementalInference.addFactor! — Method.

addFactor!(fgl, Xi, usrfnc; multihypo, ready, api, labels, uid, autoinit, threadmodel)

Add factor with user defined type <: FunctorInferenceType to the factor graph object. Define whether the automatic initialization of variables should be performed. Use order sensitive multihypo keyword argument to define if any variables are related to data association uncertainty.

IncrementalInference.addFactor! — Method.

addFactor!(fgl, xisyms, usrfnc; multihypo, ready, api, labels, uid, autoinit, threadmodel)

Add factor with user defined type <: FunctorInferenceType to the factor graph object. Define whether the automatic initialization of variables should be performed. Use order sensitive multihypo keyword argument to define if any variables are related to data association uncertainty.

IncrementalInference.addNode! — Method.

addNode!(fg, lbl, softtype; N, autoinit, ready, dontmargin, labels, api, uid, smalldata)

Add a node (variable) to a graph. Use this over the other dispatches.

IncrementalInference.addNode! — Method.

addNode!(fg, lbl, softtype; N, autoinit, ready, dontmargin, labels, api, uid, smalldata)

Add a node (variable) to a graph. Use this over the other dispatches.

IncrementalInference.allnums — Method.

allnums(str)

Test if all elements of the string is a number: Ex, "123" is true, "1_2" is false.

IncrementalInference.approxConv — Method.

approxConv(fgl, fct, towards; api, N)

Draw samples from the approximate convolution of towards symbol using factor fct relative to the other variables. In addition the api can be adjusted to recover the data from elsewhere (likely to be replaced/removed in the future).

IncrementalInference.batchSolve! — Method.

batchSolve!(fgl; drawpdf, show, N, recursive, dbg)

Perform multimodal incremental smoothing and mapping (mm-iSAM) computations over given factor graph fgl::FactorGraph on the local computer. A pdf of the Bayes (Junction) tree will be generated in the working folder with drawpdf=true

IncrementalInference.childCliqs — Method.

childCliqs(treel, cliq)

Return a vector of child cliques to cliq.

IncrementalInference.convert2packedfunctionnode — Function.

convert2packedfunctionnode(fgl, fsym)
convert2packedfunctionnode(fgl, fsym, api)

Encode complicated function node type to related 'Packed<type>' format assuming a user supplied convert function .

IncrementalInference.decodefg — Method.

decodefg(fgs; api)

Unpack PackedFunctionNodeData formats back to regular FunctonNodeData.

IncrementalInference.doautoinit! — Method.

doautoinit!(fgl, Xi; api, singles, N)

initialize destination variable nodes based on this factor in factor graph, fg, generally called during addFactor!. Destination factor is first (singletons) or second (dim 2 pairwise) variable vertex in Xi.

IncrementalInference.doautoinit! — Method.

doautoinit!(fgl, xsyms; api, singles, N)

Initialize destination variable nodes based on this factor in factor graph, fg, generally called during addFactor!. Destination factor is first (singletons) or second (dim 2 pairwise) variable vertex in Xi.

IncrementalInference.dwnMsg — Method.

dwnMsg(cliq)

Return the last down message stored in cliq of Bayes (Junction) tree.

IncrementalInference.emptyFactorGraph — Method.

emptyFactorGraph(; reference)

Construct an empty FactorGraph object with the minimum amount of information / memory populated.

IncrementalInference.encodefg — Method.

encodefg(fgl; api)

Make a full memory copy of the graph and encode all composite function node types – assuming that convert methods for 'Packed<type>' formats exist. The same converters are used for database persistence with CloudGraphs.jl.

IncrementalInference.evalFactor2 — Method.

evalFactor2(fgl, fct, solvefor; N, dbg)

Single entry point for evaluating factors from factor graph, using multiple dispatch to locate the correct evalPotentialSpecific function.

IncrementalInference.fifoFreeze! — Method.

fifoFreeze!(fgl)

Freeze nodes that are older than the quasi fixed-lag length defined by fg.qfl, according to fg.fifo ordering.

Future:

Allow different freezing strategies beyond fifo.

IncrementalInference.findRelatedFromPotential — Function.

findRelatedFromPotential(fg, idfct, vertid, N)
findRelatedFromPotential(fg, idfct, vertid, N, dbg)

Compute proposal belief on vertid through idfct representing some constraint in factor graph. Always full dimension variable node – partial constraints will only influence subset of variable dimensions. The remaining dimensions will keep pre-existing variable values.

IncrementalInference.getCurrentWorkspaceFactors — Method.

getCurrentWorkspaceFactors()

Return all factors currently registered in the workspace.

IncrementalInference.getCurrentWorkspaceVariables — Method.

getCurrentWorkspaceVariables()

Return all variables currently registered in the workspace.

IncrementalInference.getKDE — Method.

getKDE(v)

Get KernelDensityEstimate kde estimate stored in variable node.

IncrementalInference.getNumPts — Method.

getNumPts(v)

Get the number of points used for the current marginal belief estimate represtation for a particular variable in the factor graph.

IncrementalInference.getVal — Method.

getVal(v)

Convenience function to get point values sampled i.i.d from marginal of lbl variable in the current factor graph.

IncrementalInference.getVertKDE — Method.

getVertKDE(v)

Get KernelDensityEstimate kde estimate stored in variable node.

IncrementalInference.initializeNode! — Method.

initializeNode!(fgl, sym; N, api)

Initialize the belief of a variable node in the factor graph struct.

IncrementalInference.landmarks — Method.

landmarks(fgl::FactorGraph, vsym::Symbol)

Return Vector{Symbol} of landmarks attached to vertex vsym in fgl.

IncrementalInference.loadjld — Method.

loadjld(; file)

Opposite of savejld(fg, gt=gt, file="tempfg.jl") to load data from file. This function uses the unpacking converters for converting all PackedInferenceType to FunctorInferenceType.

IncrementalInference.localProduct — Method.

localProduct(fgl, sym; N, dbg, api)

Using factor graph object fg, project belief through connected factors (convolution with conditional) to variable sym followed by a approximate functional product.

Return: product belief, full proposals, partial dimension proposals, labels

IncrementalInference.ls — Method.

ls(fgl, lbls; api, ring)

Experimental union of elements version of ls(::FactorGraph, ::Symbol). Not mean't to replace broadcasting ls.(fg, [:x1;:x2])

IncrementalInference.ls — Method.

ls(fgl, lbl; api, ring)

IncrementalInference.ls — Method.

ls(fgl; key1, key2)

List the nodes in a factor graph.

Examples

ls(fg)

IncrementalInference.lsRear — Function.

lsRear(fgl)
lsRear(fgl, n)

Return array of all variable nodes connected to the last n many poses (:x*).

Example:

# Shallow copy the tail end of poses from a factor graph `fg1`
vars = lsRear(fg1, 5)
fg1_r5 = subgraphFromVerts(fg1, vars)

IncrementalInference.lsf — Method.

lsf(fgl, lbl; api)

List factors in a factor graph.

Examples

lsf(fg)

IncrementalInference.manualinit! — Method.

manualinit!(fgl, sym, usefcts)

Workaround function when first-version (factor graph based) auto initialization fails. Usually occurs when using factors that have high connectivity to multiple variables.

IncrementalInference.numericRootGenericRandomizedFnc! — Method.

numericRootGenericRandomizedFnc!(ccwl; perturb, testshuffle)

Solve free variable x by root finding residual function fgr.usrfnc(x, res) randomly shuffle x dimensions if underconstrained by measurement z dimensions small random perturbation used to prevent trivial solver cases, div by 0 etc. result stored in fgr.Y ccw.X must be set to memory ref the param[varidx] being solved, at creation of ccw

IncrementalInference.parentCliq — Method.

parentCliq(treel, cliq)

Return cliq's parent clique.

IncrementalInference.printgraphmax — Method.

printgraphmax(fgl)

Print the maximum point values form all variables approximate marginals in the factor graph. The full marginal can be recovered for example X0 = getVertKDE(fg, :x0).

IncrementalInference.savejld — Method.

savejld(fgl; file, groundtruth)

Save mostly complete Factor Graph type by converting complicated FunctionNodeData types to 'Packed' types using user supplied converters. Ground truth can also be saved and recovered by the associated loadjld(file="tempfg.jld2") method.

IncrementalInference.setDwnMsg! — Method.

setDwnMsg!(cliql, msgs)

Set the downward passing message for Bayes (Junction) tree clique cliql.

IncrementalInference.setUpMsg! — Method.

setUpMsg!(cliql, msgs)

Set the upward passing message for Bayes (Junction) tree clique cliql.

IncrementalInference.shuffleXAltD! — Method.

shuffleXAltD!(ccwl, X)

Shuffle incoming X into random positions in fr.Y. Shuffled fr.Y will be placed back into fr.X[:,fr.gwp.particleidx] upon fr.gwp.usrfnc(x, res).

IncrementalInference.subgraphFromVerts — Method.

subgraphFromVerts(fgl, verts; neighbors)

Explore all shortest paths combinations in verts, add neighbors and reference subgraph using unique index into graph data structure.

IncrementalInference.treeProductUp — Method.

treeProductUp(fg, tree, cliq, sym; N, dbg)

Calculate a fresh–-single step–-approximation to the variable sym in clique cliq as though during the upward message passing. The full inference algorithm may repeatedly calculate successive apprimxations to the variable based on the structure of variables, factors, and incoming messages to this clique. Which clique to be used is defined by frontal variable symbols (cliq in this case) – see whichCliq(...) for more details. The sym symbol indicates which symbol of this clique to be calculated. Note that the sym variable must appear in the clique where cliq is a frontal variable.

IncrementalInference.upMsg — Method.

upMsg(cliq)

Return the last up message stored in cliq of Bayes (Junction) tree.

IncrementalInference.whichCliq — Method.

whichCliq(bt, frt)

Return the Graphs.ExVertex node object that represents a clique in the Bayes (Junction) tree, as defined by one of the frontal variables frt.

IncrementalInference.wipeBuildNewTree! — Method.

wipeBuildNewTree!(fg; ordering, drawpdf, show, filepath, viewerapp)

Build a completely new Bayes (Junction) tree, after first wiping clean all temporary state in fg from a possibly pre-existing tree.

IncrementalInference.writeGraphPdf — Method.

writeGraphPdf(fgl; viewerapp, filepath, engine, show)

Export a dot and pdf file drawn by Graphviz showing the factor graph.

KernelDensityEstimate.kde! — Method.

kde!(em)

Construct a BallTreeDensity KDE object from an IIF.EasyMessage object.

IncrementalInference.addNewVarVertInGraph! — Method.

addNewVarVertInGraph!(fgl, vert, id, lbl, ready, smalldata)

Initialize a new Graphs.ExVertex which will be added to some factor graph.

IncrementalInference.approxConvOnElements! — Method.

approxConvOnElements!(ccwl, elements, ?)

Perform the nonlinear numerical operations to approximate the convolution with a particular user defined likelihood function (conditional), which as been prepared in the frl object. This function uses root finding to enforce a non-linear function constraint.

Notes:

remember this is a deepcopy of original sfidx, since we are generating a proposal distribution and not directly replacing the existing variable belief estimate

Future work:

once Threads.@threads have been optmized JuliaLang/julia#19967, also see area4 branch

improve handling of n and particleidx, especially considering future multithreading support

IncrementalInference.approxConvOnElements! — Method.

approxConvOnElements!(ccwl, elements, ?)

Perform the nonlinear numerical operations to approximate the convolution with a particular user defined likelihood function (conditional), which as been prepared in the frl object. This function uses root finding to enforce a non-linear function constraint.

Notes:

remember this is a deepcopy of original sfidx, since we are generating a proposal distribution and not directly replacing the existing variable belief estimate

Future work:

once Threads.@threads have been optmized JuliaLang/julia#19967, also see area4 branch

improve handling of n and particleidx, especially considering future multithreading support

IncrementalInference.approxConvOnElements! — Method.

approxConvOnElements!(ccwl, elements)

Perform the nonlinear numerical operations to approximate the convolution with a particular user defined likelihood function (conditional), which as been prepared in the frl object. This function uses root finding to enforce a non-linear function constraint.

Notes:

remember this is a deepcopy of original sfidx, since we are generating a proposal distribution and not directly replacing the existing variable belief estimate

Future work:

once Threads.@threads have been optmized JuliaLang/julia#19967, also see area4 branch

improve handling of n and particleidx, especially considering future multithreading support

IncrementalInference.assembleHypothesesElements! — Method.

assembleHypothesesElements!(mh, maxlen, sfidx, lenXi)

This function explicitly codes that marginalization of a discrete categorical selection variable for ambiguous data association situations. Improved implementations should implicitly induce the same behaviour through summation (integration) when marginalizing any number of discrete variables. This function populates allelements with particle indices associated with particular multihypothesis selection while activehypo simultaneously contains the hypothesis index and factor graph variables associated with that hypothesis selection. The return value certainidx are the hypotheses that are not in question.

# `allelements` example BearingRange [:x1, 0.5:l1a, 0.5:l1b]
# sfidx = (1=:x1,2=:l1a,3=:l1b)
if solvefor :x1, then allelem = [mhidx.==:l1a; mhidx.==l1b]
if solvefor :l1a, then allelem = [mhidx.==:l1a] and ARR[solvefor][:,mhidx.==:l1b]=ARR[:l1b][:,mhidx.==:l1b]
if solvefor :l1b, then allelem = [mhidx.==:l1b] and ARR[solvefor][:,mhidx.==:l1a]=ARR[:l1a][:,mhidx.==:l1a]
if solvefor 1, then allelem = [mhidx.==2; mhidx.==3]
if solvefor 2, then allelem = [mhidx.==2] and ARR[solvefor][:,mhidx.==3]=ARR[3][:,mhidx.==3]
if solvefor 3, then allelem = [mhidx.==3] and ARR[solvefor][:,mhidx.==2]=ARR[2][:,mhidx.==2]

# `activehypo` in example mh=[1.0;0.5;0.5]
sfidx=1, mhidx=2:  ah = [1;2]
sfidx=1, mhidx=3:  ah = [1;3]
sfidx=2, mhidx=2:  ah = [1;2]
sfidx=2, mhidx=3:  2 should take a value from 3
sfidx=3, mhidx=2:  3 should take a value from 2
sfidx=3, mhidx=3:  ah = [1;3]

# `activehypo` in example mh=[1.0;0.33;0.33;0.34]
sfidx=1, mhidx=2:  ah = [1;2]
sfidx=1, mhidx=3:  ah = [1;3]
sfidx=1, mhidx=4:  ah = [1;4]

sfidx=2, mhidx=2:  ah = [1;2]
sfidx=2, mhidx=3:  2 should take a value from 3
sfidx=2, mhidx=4:  2 should take a value from 4

sfidx=3, mhidx=2:  3 should take a value from 2
sfidx=3, mhidx=3:  ah = [1;3]
sfidx=3, mhidx=4:  3 should take a value from 4

sfidx=4, mhidx=2:  4 should take a value from 2
sfidx=4, mhidx=3:  4 should take a value from 3
sfidx=4, mhidx=4:  ah = [1;4]

IncrementalInference.assembleNullHypothesis — Method.

assembleNullHypothesis(ccwl, maxlen, spreadfactor)

Prepare data required for null hypothesis cases during convolution.

IncrementalInference.cliqGibbs — Method.

cliqGibbs(fg, cliq, vertid, inmsgs, N, dbg)

Perform one step of the minibatch clique Gibbs operation for solving the Chapman-Kolmogov trasit integral – here involving separate approximate functional convolution and product operations.

IncrementalInference.computeAcrossHypothesis! — Method.

computeAcrossHypothesis!(ccwl, allelements, activehypo, certainidx, sfidx)

Common function to compute across a single user defined multi-hypothesis ambiguity per factor. This function dispatches both FunctorPairwise and FunctorPairwiseMinimize factors.

IncrementalInference.computeAcrossNullHypothesis! — Method.

computeAcrossNullHypothesis!(ccwl, allelements, nhc, ENT)

Do true and null hypothesis computations based on data structures prepared earlier – specific to FunctorPairwiseNH. This function will be merged into a standard case for FunctorPairwise/Minimize in the future.

IncrementalInference.downMsgPassingRecursive — Method.

downMsgPassingRecursive(inp; N, dbg, drawpdf)

Pass NBPMessages back down the tree – pre order tree traversal.

IncrementalInference.evalPotentialSpecific — Method.

evalPotentialSpecific(Xi, ccwl, solvefor; N, dbg)

Multiple dispatch wrapper for evaluating the ccwl::CommonConvWrapper{<: FunctorSingleton} types.

IncrementalInference.evalPotentialSpecific — Method.

evalPotentialSpecific(Xi, ccwl, solvefor; N, spreadfactor, dbg)

Multiple dispatch wrapper for evaluating the ccwl::CommonConvWrapper{<: FunctorSingletonNH} types. Planned changes will fold null hypothesis in as a standard feature and no longer appear as a separate InferenceType.

IncrementalInference.evalPotentialSpecific — Method.

evalPotentialSpecific(Xi, ccwl, solvefor; N, dbg)

Multiple dispatch wrapper for <:FunctorPairwise types, to prepare and execute the general approximate convolution with user defined factor residual functions. This method also supports multihypothesis operations as one mechanism to introduce new modality into the proposal beliefs.

IncrementalInference.fmcmc! — Function.

fmcmc!(fgl, cliq, fmsgs, IDs, N, MCMCIter)
fmcmc!(fgl, cliq, fmsgs, IDs, N, MCMCIter, dbg)

Iterate successive approximations of clique marginal beliefs by means of the stipulated proposal convolutions and products of the functional objects for tree clique cliq.

IncrementalInference.packFromLocalPotentials! — Function.

packFromLocalPotentials!(fgl, dens, wfac, cliq, vertid, N)
packFromLocalPotentials!(fgl, dens, wfac, cliq, vertid, N, dbg)

Add all potentials associated with this clique and vertid to dens.

IncrementalInference.prepareCommonConvWrapper! — Method.

prepareCommonConvWrapper!(ccwl, Xi, solvefor, N)

Prepare a common functor computation object prepareCommonConvWrapper{T} containing the user factor functor along with additional variables and information using during approximate convolution computations.

IncrementalInference.prepareparamsarray! — Method.

prepareparamsarray!(ARR, Xi, N, solvefor)

Prepare the particle arrays ARR to be used for approximate convolution. This function ensures that ARR has te same dimensions among all the parameters. Function returns with ARR[sfidx] pointing at newly allocated deepcopy of the existing values in getVal(Xi[.index==solvefor]). Return values sfidx is the element in ARR where Xi.index==solvefor and maxlen is length of all (possibly resampled) ARR contained particles. Note Xi is order sensitive.

IncrementalInference.prodmultiplefullpartials — Method.

prodmultiplefullpartials(dens, partials, Ndims, N)

Multiply various full and partial dimension constraints.

IncrementalInference.prodmultipleonefullpartials — Method.

prodmultipleonefullpartials(dens, partials, Ndims, N)

Multiply a single full and several partial dimension constraints.

IncrementalInference.productpartials! — Method.

productpartials!(pGM, dummy, partials)

Multiply different dimensions from partial constraints individually.

IncrementalInference.setfreeze! — Method.

setfreeze!(fgl, sym)

Set variable(s) sym of factor graph to be marginalized – i.e. not be updated by inference computation.

IncrementalInference.treeProductDwn — Method.

treeProductDwn(fg, tree, cliq, sym; N, dbg)

Calculate a fresh–-single step–-approximation to the variable sym in clique cliq as though during the downward message passing. The full inference algorithm may repeatedly calculate successive apprimxations to the variable based on the structure of variables, factors, and incoming messages to this clique. Which clique to be used is defined by frontal variable symbols (cliq in this case) – see whichCliq(...) for more details. The sym symbol indicates which symbol of this clique to be calculated. Note that the sym variable must appear in the clique where cliq is a frontal variable.

IncrementalInference.upGibbsCliqueDensity — Function.

upGibbsCliqueDensity(inp)
upGibbsCliqueDensity(inp, N)
upGibbsCliqueDensity(inp, N, dbg)

Perform computations required for the upward message passing during belief propation on the Bayes (Junction) tree. This function is usually called as part via remote_call for multiprocess dispatch.

IncrementalInference.updateFGBT! — Method.

updateFGBT!(fg, bt, cliqID, ddt; dbg, fillcolor)

Update cliq cliqID in Bayes (Juction) tree bt according to contents of ddt – intended use is to update main clique after a downward belief propagation computation has been completed per clique.

IncrementalInference.updateFGBT! — Method.

updateFGBT!(fg, bt, cliqID, urt; dbg, fillcolor)

Update cliq cliqID in Bayes (Juction) tree bt according to contents of urt – intended use is to update main clique after a upward belief propagation computation has been completed per clique.