Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

IndexableBitVectors.RRR — Type.

Compressed indexable bit vector.

RRR compresses a bit vector in an information-theoretically optimal representation. This compression is based on local abundance of bits; if 0s or 1s are clustered in a bit vector, it will be compressed even when the numbers of 0s and 1s are globally equal.

Let n be the length of a RRR, the asymptotic query times are

getindex: O(1)

rank: O(1)

select: O(log n)

See (Raman et al, 2007, doi:10.1145/1290672.1290680) for more details.

IndexableBitVectors.SucVector — Type.

Static indexable bit vector.

SucVector requires 5/4 bits per bit in order to keep bits and the number of bits (i.e. a SucVector object becomes ~25% bigger when converted from a BitVector object). Bits are split into blocks: four chunks of 64 bits are stored in a block and blocks are stored in a vector. Blocks are contiguous in momery so that memory cache works efficiently.

The number of bits is limited up to 2^40 bits. This is 128GiB so would be enough in most cases.

Let n be the length of a SucVector, the asymptotic query times are

getindex: O(1)

rank: O(1)

select: O(log n)

IndexableBitVectors.rank — Function.

rank(x, bv, i)

Count the number of xs within bv[1:i].

IndexableBitVectors.rank0 — Function.

rank0(rb, i)

Count the number of 0s (falses) within bv[1:i].

IndexableBitVectors.rank1 — Function.

rank1(bv, i)

Count the number of 1s (trues) within bv[1:i].

IndexableBitVectors.rsearch — Function.

rsearch(x, bv, i)

Search the position of the previous x in bv starting from i.

IndexableBitVectors.rsearch0 — Function.

rsearch0(bv, i)

Search the position of the previous 0 in bv starting from i.

IndexableBitVectors.rsearch1 — Function.

rsearch1(bv, i)

Search the position of the previous 1 in bv starting from i.

IndexableBitVectors.search — Function.

search(x, bv, i)

Search the position of the next x in bv starting from i.

IndexableBitVectors.search0 — Function.

search0(bv, i)

Search the position of the next 0 in bv starting from i.

IndexableBitVectors.search1 — Function.

search1(bv, i)

Search the position of the next 1 in bv starting from i.

IndexableBitVectors.select — Function.

select(x, bv, j)

Return the position of the j-th occurrence of x in bv.

IndexableBitVectors.select0 — Function.

select0(bv, j)

Return the position of the j-th occurrence of 0 in bv.

IndexableBitVectors.select1 — Function.

select1(bv, j)

Return the position of the j-th occurrence of 1 in bv.