Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Indexing.ViewArray — Type.

ViewArray(parent, indices)

Create an array out which is a lazy view of parent. The indices of out match the indices of indices, and the values of indices are used to index parent. Unlike SubArray, the parent type could be any indexable container (for instance, we can create an AbstractArray view of the values stored in an AbstractDict).

See also ViewDict, view.

Examples

julia> d = Dict(:a => 11, :b => 12, :c => 13)
Dict{Symbol,Int64} with 3 entries:
  :a => 11
  :b => 12
  :c => 13

julia> ViewArray(d, [:a, :c])
2-element ViewArray{Int64,1,Dict{Symbol,Int64},Array{Symbol,1}}:
 11
 13

Indexing.ViewDict — Type.

ViewDict(parent, indices)

Create an array out which is a lazy view of parent. The indices of out match the indices of indices, and the values of indices are used to index parent. The parent must be an indexable type - for instance, an array or dictionary.

See also ViewArray, view.

Examples

julia> d = Dict(:a => "Alice", :b => "Bob", :c => "Charlie")
Dict{Symbol,String} with 3 entries:
  :a => "Alice"
  :b => "Bob"
  :c => "Charlie"

julia> ViewDict(d, Dict(:aa => :a, :cc => :c))
ViewDict{Symbol,String,Dict{Symbol,String},Dict{Symbol,Symbol}} with 2 entries:
  :aa => "Alice"
  :cc => "Charlie"

julia> v = [11, 12, 13]
3-element Array{Int64,1}:
 11
 12
 13

julia> ViewDict(v, Dict(:a =>1 , :c => 3))
ViewDict{Symbol,Int64,Array{Int64,1},Dict{Symbol,Int64}} with 2 entries:
  :a => 11
  :c => 13

Indexing.getindices — Method.

getindices(container, indices)

Return an indexable container with indices keys(indices) and values container[i] for i ∈ values(indices). This generalizes scalar getindex(container, index) for multiple indices, for dictionaries, tuples, named tuples, etc.

Examples

julia> d = Dict(:a => "Alice", :b => "Bob", :c => "Charlie")
Dict{Symbol,String} with 3 entries:
  :a => "Alice"
  :b => "Bob"
  :c => "Charlie"

julia> getindices(d, [:a, :c])
2-element Array{String,1}:
 "Alice"  
 "Charlie"

julia> getindices(d, (:a, :c))
("Alice", "Charlie")

julia> getindices(d, Dict("Wife" => :a, "Husband" => :c))
Dict{String,String} with 2 entries:
  "Wife"    => "Alice"
  "Husband" => "Charlie"

Indexing.setindices! — Method.

setindices!(container, value, indices)

Store the given value at all the indices in values(indices) of container. This generalizes scalar setindex! for dictionaries, etc.

Base.view — Method.

view(d::AbstractDict, inds::AbstractArray)

Create an array view over the dictionary d with the indices stored in inds.

See also ViewArray.

Examples

julia> d = Dict(:a => 11, :b => 12, :c => 13)
Dict{Symbol,Int64} with 3 entries:
  :a => 11
  :b => 12
  :c => 13

julia> view(d, [:a, :c])
2-element ViewArray{Int64,1,Dict{Symbol,Int64},Array{Symbol,1}}:
 11
 13

Base.view — Method.

view(parent, inds::AbstractDict)

Create a dictionary view over the indexable container parent. The indices of the output are the indices of inds, and the values are the corresponding parent[i] for i in values(inds).

See also ViewDict.

Examples

julia> d = Dict(:a => "Alice", :b => "Bob", :c => "Charlie")
Dict{Symbol,String} with 3 entries:
  :a => "Alice"
  :b => "Bob"
  :c => "Charlie"

julia> view(d, Dict(:aa => :a, :cc => :c))
ViewDict{Symbol,String,Dict{Symbol,String},Dict{Symbol,Symbol}} with 2 entries:
  :aa => "Alice"
  :cc => "Charlie"

julia> v = [11, 12, 13]
3-element Array{Int64,1}:
 11
 12
 13

julia> view(v, Dict(:a =>1 , :c => 3))
ViewDict{Symbol,Int64,Array{Int64,1},Dict{Symbol,Int64}} with 2 entries:
  :a => 11
  :c => 13