Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Indicators.adx — Method.

Average directional index

adx(hlc::Array{Float64}; n::Int64=14, wilder=true)::Array{Float64}

Output

Column 1: DI+

Column 2: DI-

Column 3: ADX

Indicators.alma — Method.

Arnaud-Legoux moving average (ALMA)

alma{Float64}(x::Array{Float64}; n::Int64=9, offset::Float64=0.85, sigma::Float64=6.0)::Array{Float64}

Indicators.aroon — Method.

Aroon up/down/oscillator

aroon(hl::Array{Float64,2}; n::Int64=25)::Array{Float64}

Output

Column 1: Aroon Up

Column 2: Aroon Down

Column 3: Aroon Oscillator

Indicators.atr — Method.

Average true range (uses exponential moving average)

atr(hlc::Matrix{Float64}; n::Int64=14)::Array{Float64}

Indicators.bbands — Method.

Bollinger bands (moving average with standard deviation bands)

bbands(x::Array{Float64}; n::Int64=10, sigma::Float64=2.0)::Matrix{Float64}

Output

Column 1: lower band

Column 2: middle band

Column 3: upper band

Indicators.cci — Method.

Commodity channel index

cci(hlc::Array{Float64,2}; n::Int64=20, c::Float64=0.015, ma::Function=sma)::Array{Float64}

Indicators.crossover — Method.

Find where x crosses over y (returns boolean vector where crossover occurs)

Indicators.crossunder — Method.

Find where x crosses under y (returns boolean vector where crossunder occurs)

Indicators.dema — Method.

Double exponential moving average (DEMA)

dema(x::Array{Float64}; n::Int64=10, alpha=2.0/(n+1), wilder::Bool=false)::Array{Float64}

Indicators.diffn — Method.

Lagged differencing

diffn(X::Array{Float64,2}; n::Int64=1)::Array{Float64}

Indicators.diffn — Method.

Lagged differencing

diffn(x::Array{Float64}; n::Int64=1)::Array{Float64}

Indicators.donch — Method.

Donchian channel (if inclusive is set to true, will include current bar in calculations.)

donch(hl::Array{Float64,2}; n::Int64=10, inclusive::Bool=true)::Array{Float64}

Output

Column 1: Lowest low of last n periods

Column 2: Average of highest high and lowest low of last n periods

Column 3: Highest high of last n periods

Indicators.ema — Method.

Exponential moving average (EMA)

ema(x::Array{Float64}; n::Int64=10, alpha::Float64=2.0/(n+1.0), wilder::Bool=false)::Array{Float64}

Indicators.hma — Method.

Hull moving average (HMA)

hma(x::Array{Float64}; n::Int64=20)::Array{Float64}

Indicators.kama — Method.

Kaufman adaptive moving average (KAMA)

Indicators.keltner — Method.

Keltner bands

keltner(hlc::Matrix{Float64}; nema::Int64=20, natr::Int64=10, mult::Int64=2)::Matrix{Float64}

Output Column 1: lower band Column 2: middle band Column 3: upper band

Indicators.kst — Method.

KST (Know Sure Thing) – smoothed and summed rates of change

kst(x::Array{Float64};
    nroc::Array{Int64}=[10,15,20,30], navg::Array{Int64}=[10,10,10,15],
    wgts::Array{Int64}=collect(1:length(nroc)), ma::Function=sma)::Array{Float64}

Indicators.macd — Method.

Moving average convergence-divergence

macd(x::Array{Float64}; nfast::Int64=12, nslow::Int64=26, nsig::Int64=9)::Array{Float64}

Output

Column 1: MACD

Column 2: MACD Signal Line

Column 3: MACD Histogram

Indicators.mama — Method.

MESA adaptive moving average (MAMA)

mama(x::Array{Float64}; fastlimit::Float64=0.5, slowlimit::Float64=0.05)::Matrix{Float64}

Indicators.mlr — Method.

Moving linear regression predictions

Indicators.mlr_bands — Method.

Moving linear regression bands

Output:

Column 1: Lower bound

Column 2: Regression estimate

Column 3: Upper bound

Indicators.mlr_beta — Method.

Moving linear regression intercept (column 1) and slope (column 2)

Indicators.mlr_intercept — Method.

Moving linear regression y-intercept

Indicators.mlr_lb — Method.

Moving linear regression lower bound

Indicators.mlr_rsq — Method.

Moving linear regression R-squared or adjusted R-squared

Indicators.mlr_se — Method.

Moving linear regression standard errors

Indicators.mlr_slope — Method.

Moving linear regression slope

Indicators.mlr_ub — Method.

Moving linear regression upper bound

Indicators.mma — Method.

Modified moving average (MMA)

mma(x::Array{Float64}; n::Int64=10)::Array{Float64}

Indicators.mode — Method.

(Adapted from StatsBase: https://raw.githubusercontent.com/JuliaStats/StatsBase.jl/master/src/scalarstats.jl)

Compute the mode of an arbitrary array::Array{Float64}

Indicators.momentum — Method.

Momentum indicator (price now vs price n periods back)

momentum(x::Array{Float64}; n::Int64=1)::Array{Float64}

Indicators.psar — Method.

Parabolic stop and reverse (SAR)

psar(hl::Array{Float64}; af_min::Float64=0.02, af_max::Float64=0.2, af_inc::Float64=af_min)::Array{Float64}

Arguments

hl: 2D array of high and low prices in first and second columns respectively

af_min: starting/initial value for acceleration factor

af_max: maximum acceleration factor (accel factor capped at this value)

af_inc: increment to the acceleration factor (speed of increase in accel factor)

Indicators.renko — Method.

Renko chart patterns

Methods

Traditional (Constant Box Size): renko(x::Array{Float64}; box_size::Float64=10.0)::Array{Int}

ATR Dynamic Box Size: renko(hlc::Matrix{Float64}; box_size::Float64=10.0, use_atr::Bool=false, n::Int=14)::Array{Int}

Output

Array{Int} object of size Nx1 (where N is the number rows in x) where each element gives the Renko bar number of the corresponding row in x.

Indicators.roc — Method.

Rate of change indicator (percent change between i'th observation and (i-n)'th observation)

roc(x::Array{Float64}; n::Int64=1)::Array{Float64}

Indicators.rsi — Method.

Relative strength index

rsi(x::Array{Float64}; n::Int64=14, ma::Function=ema, args...)::Array{Float64}

Indicators.runacf — Method.

Compute the running/rolling autocorrelation of a vector.

function runacf(x::Array{Float64};
                n::Int = 10,
                maxlag::Int = n-3,
                lags::AbstractArray{Int,1} = 0:maxlag,
                cumulative::Bool = true)::Matrix{Float64}

Indicators.runcor — Method.

Compute the running or rolling correlation of two arrays

runcor(x::Array{Float64}, y::Array{Float64}; n::Int64=10, cumulative::Bool=true)::Array{Float64}

Indicators.runcov — Method.

Compute the running or rolling covariance of two arrays

runcov(x::Array{Float64}, y::Array{Float64}; n::Int64=10, cumulative::Bool=true)::Array{Float64}

Indicators.runmad — Method.

Compute the running or rolling mean absolute deviation of an array

runmad(x::Array{Float64}; n::Int64=10, cumulative::Bool=true, fun::Function=median)::Array{Float64}

Indicators.runmax — Method.

Compute the running or rolling maximum of an array

runmax(x::Array{Float64}; n::Int64=10, cumulative::Bool=true, inclusive::Bool=true)::Array{Float64}

Indicators.runmean — Method.

Compute a running or rolling arithmetic mean of an array.

runmean(x::Array{Float64}; n::Int64=10, cumulative::Bool=true)::Array{Float64}

Indicators.runmin — Method.

Compute the running or rolling minimum of an array

runmin(x::Array{Float64}; n::Int64=10, cumulative::Bool=true, inclusive::Bool=true)::Array{Float64}

Indicators.runquantile — Method.

Compute the running/rolling quantile of an array

runquantile(x::Vector{Float64}; p::Float64=0.05, n::Int=10, cumulative::Bool=true)::Vector{Float64}

Indicators.runsd — Method.

Compute the running or rolling standard deviation of an array

runsd(x::Array{Float64}; n::Int64=10, cumulative::Bool=true)::Array{Float64}

Indicators.runsum — Method.

Compute a running or rolling summation of an array.

runsum(x::Array{Float64}; n::Int64=10, cumulative::Bool=true)::Array{Float64}

Indicators.runvar — Method.

Compute the running or rolling variance of an array

runvar(x::Array{Float64}; n::Int64=10, cumulative=true)::Array{Float64}

Indicators.sma — Method.

Simple moving average (SMA)

sma(x::Array{Float64}; n::Int64=10)::Array{Float64}

Indicators.smi — Method.

SMI (stochastic momentum oscillator)

smi(hlc::Array{Float64,2}; n::Int64=13, nFast::Int64=2, nSlow::Int64=25, nSig::Int64=9,
    maFast::Function=ema, maSlow::Function=ema, maSig::Function=sma)::Matrix{Float64}

Indicators.stoch — Method.

Stochastic oscillator (fast or slow)

stoch(hlc::Array{Float64,2}; nK::Int64=14, nD::Int64=3, kind::Symbol=:fast, ma::Function=sma, args...)::Matrix{Float64}

Indicators.swma — Method.

Sine-weighted moving average

Indicators.tema — Method.

Triple exponential moving average (TEMA)

tema(x::Array{Float64}; n::Int64=10, alpha=2.0/(n+1), wilder::Bool=false)::Array{Float64}

Indicators.tr — Method.

True range

tr(hlc::Matrix{Float64})::Array{Float64}

Indicators.trima — Method.

Triangular moving average (TRIMA)

trima(x::Array{Float64}; n::Int64=10, ma::Function=sma, args...)::Array{Float64}

Indicators.wilder_sum — Method.

Welles Wilder summation of an array

wilder_sum(x::Array{Float64}; n::Int64=10)::Array{Float64}

Indicators.wma — Method.

Weighted moving average (WMA)

wma(x::Array{Float64}; n::Int64=10, wts::Array{Float64}=collect(1:n)/sum(1:n))::Array{Float64}

Indicators.wpr — Method.

Williams %R

wpr(hlc::Array{Float64,2}, n::Int64=14)::Array{Float64}

Indicators.zlema — Method.

Zero-lag exponential moving average (ZLEMA)

zlema(x::Array{Float64}; n::Int=10, ema_args...)::Array{Float64}