Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Inflate.InflateGzipStream — Type.

InflateGzipStream(stream::IO)

Streaming decompression of Gzip compressed stream. For an in-memory counterpart, see inflate_gzip.

gzip_headers = Dict{String, Any}()
InflateGzipStream(stream::IO; headers = gzip_headers)

Also retrieve gzip headers in the provided Dict. The headers are available directly after the object is constructed.

InflateGzipStream(stream::IO; ignore_checksum = true)

Skip computing CRC of the content, as well as the header, for consistency checking.

Reference: RFC 1952

Inflate.InflateStream — Type.

InflateStream(stream::IO)

Streaming decompression of unwrapped deflate compressed stream. For an in-memory counterpart, see inflate.

Reference: RFC 1951

Inflate.InflateZlibStream — Type.

InflateZlibStream(stream::IO)

Streaming decompression of Zlib compressed stream. For an in-memory counterpart, see inflate_zlib.

InflateZlibStream(stream::IO; ignore_checksum = true)

Skip computing Adler checksum for consistency checking.

Reference: RFC 1950

Inflate.inflate — Method.

inflate(source::Vector{UInt8})

Decompress in-memory source, in unwrapped deflate format. The output will also be a Vector{UInt8}. For a streaming counterpart, see InflateStream.

Reference: RFC 1951

Inflate.inflate_gzip — Method.

inflate_gzip(filename::AbstractString)

Convenience wrapper for reading a gzip compressed text file. The result is returned as a string.

Inflate.inflate_gzip — Method.

inflate_gzip(source::Vector{UInt8})

Decompress in-memory source, in Gzip compressed format. The output will also be a Vector{UInt8}. For a streaming counterpart, see InflateGzipStream.

gzip_headers = Dict{String, Any}()
inflate_gzip(source::Vector{UInt8}; headers = gzip_headers)

Also retrieve gzip headers in the provided Dict.

inflate_gzip(source::Vector{UInt8}; ignore_checksum = true)

Skip computing CRC of the content, as well as the header, for consistency checking.

Reference: RFC 1952

Inflate.inflate_zlib — Method.

inflate_zlib(source::Vector{UInt8})

Decompress in-memory source, in Zlib compressed format. The output will also be a Vector{UInt8}. For a streaming counterpart, see InflateZlibStream.

inflate_zlib(source::Vector{UInt8}; ignore_checksum = true)

Skip computing Adler checksum for consistency checking.

Reference: RFC 1950