Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

InformationMeasures.discretize_values — Method.

discretize_values(values_x; <keyword arguments>)

Assign continuous measurements to discrete bins.

Arguments

values_x: the data values.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

InformationMeasures.get_conditional_entropy — Method.

get_conditional_entropy(values_x, values_y; <keyword arguments>)

Estimate conditional entropy of one set of values conditioned on another set of values.

Arguments:

values_x: the data values.

values_y: the data values to be conditioned on.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_conditional_entropy — Method.

get_conditional_entropy(xy; <keyword arguments>)

Arguments:

xy: the 2D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xy is probabilities. If false, xy is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_conditional_mutual_information — Method.

get_conditional_mutual_information(values_x, values_y, values_z; <keyword arguments>)

Estimate the conditional mutual information between two sets of values, conditioned on a third.

Arguments:

values_x: the data values.

values_y: a second set of data values.

values_z: the data values to be conditioned on.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_conditional_mutual_information — Method.

get_conditional_mutual_information(xyz; <keyword arguments>)

Arguments:

xyz: the 3D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xyz is probabilities. If false, xyz is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_cross_entropy — Method.

get_cross_entropy(values_x, values_y; <keyword arguments>)

Estimate the cross-entropy between two sets of values.

Arguments:

values_x: the first set of data values.

values_y: the second set of data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

discretized=false: whether the data values are already discretized.

lambda=nothing: the shrinkage intensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_entropy — Method.

get_entropy(values_x; <keyword arguments>)

Estimate entropy (or joint entropy, if more than one set of data values is given).

Arguments:

values_x: the data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

discretized=false: whether the data values are already discretized.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_interaction_information — Method.

get_interaction_information(values_x, values_y, values_z; <keyword arguments>)

Estimate the interaction information between three sets of values.

Arguments:

values_x: the data values.

values_y: a second set of data values.

values_z: a third set of data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_interaction_information — Method.

get_interaction_information(xyz; <keyword arguments>)

Arguments:

xyz: the 3D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xyz is probabilities. If false, xyz is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_mutual_information — Method.

get_mutual_information(values_x, values_y; <keyword arguments>)

Estimate the mutual information between two sets of values.

Arguments:

values_x: the data values.

values_y: a second set of data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_mutual_information — Method.

get_mutual_information(xy; <keyword arguments>)

Arguments:

xy: the 2D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xy is probabilities. If false, xy is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_partial_information_decomposition — Method.

get_partial_information_decomposition(values_x, values_y, values_z; <keyword arguments>)

Estimate the partial information decomposition between three sets of values.

Performance can be improved by setting all_orientations, include_unique and include_synergy according to the use case.

Arguments:

values_x: the data values.

values_y: a second set of data values.

values_z: a third set of data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equqivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

all_orientations=false: whether to use each set of values as the target. If false, only values_z is the target.

include_unique=true: whether to get the unique information.

include_synergy=true: whether to get the synergy.

InformationMeasures.get_partial_information_decomposition — Method.

get_partial_information_decomposition(xyz; <keyword arguments>)

Arguments:

xyz: the 3D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xyz is probabilities. If false, xyz is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

all_orientations=false: whether to use each set of values as the target. If false, only values_z is the target.

include_unique=true: whether to get the unique information.

include_synergy=true: whether to get the synergy.

InformationMeasures.get_probabilities — Method.

get_probabilities(estimator, frequencies; <keyword arguments>)

Estimate probabilities from a set of discrete values.

Arguments:

estimator: the entropy estimator.

frequencies: the bin frequencies for the discretized data values.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_redundancy — Method.

get_redundancy(values_x, values_y, values_z; <keyword arguments>)

Estimate the redundancy between three sets of values.

Arguments:

values_x: the data values.

values_y: a second set of data values.

values_z: a third set of data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equqivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

all_orientations=false: whether to use each set of values as the target. If false, only values_z is the target.

InformationMeasures.get_redundancy — Method.

get_redundancy(xyz; <keyword arguments>)

Arguments:

xyz: the 3D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xyz is probabilities. If false, xyz is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

all_orientations=false: whether to use each set of values as the target. If false, only values_z is the target.

InformationMeasures.get_total_correlation — Method.

get_total_correlation(values_x, values_y, values_z; <keyword arguments>)

Estimate the total correlation between three sets of values.

Arguments:

values_x: the data values.

values_y: a second set of data values.

values_z: a third set of data values.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

mode="uniform_width": the discretization algorithm.

number_of_bins=0: the number of bins (will be overridden if mode is "bayesian_blocks").

get_number_of_bins=get_root_n: the method for calculating the number of bins (only called if number_of_bins is 0).

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".

InformationMeasures.get_total_correlation — Method.

get_total_correlation(xyz; <keyword arguments>)

Arguments:

xyz: the 3D frequencies or probabilities.

estimator="maximum_likelihood": the entropy estimator.

base=2: the base of the logarithm, equivalent to the units of information.

probabilities=false: whether xyz is probabilities. If false, xyz is frequencies.

lambda=nothing: the shrinkage instensity, only used if estimator is "shrinkage".

prior=1: the Dirichlet prior, only used if estimator is "dirichlet".