Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Base.kill — Method.

kill(search::InteractiveCodeSearch.BackgroundSearch)

Stop search.

InteractiveCodeSearch.bg — Function.

InteractiveCodeSearch.bg(id::Int = 0)

Lookup background search by id. For convenience, id = 0 is the last search and id = -1 is the second last search. More generally, non-positive id is treated as the offset from end.

InteractiveCodeSearch.find_source_file — Method.

find_source_file(file)

Find source file and return its full path. It just calls Base.find_source_file and return its result for normal Julia installation. For nightly Julia build, it tries to guess the right path when Base.find_source_file failed.

InteractiveCodeSearch.get_history_provider — Method.

get_history_provider() :: REPLHistoryProvider

InteractiveCodeSearch.load_history_provider — Method.

load_history_provider() :: REPLHistoryProvider

InteractiveCodeSearch.read_stdout — Method.

read_stdout(input::AbstractString, cmd)
read_stdout(input_provider, cmd)

Julia implementation of "echo {input} | {cmd}".

InteractiveCodeSearch.stop — Method.

InteractiveCodeSearch.stop(search::BackgroundSearch)
InteractiveCodeSearch.stop(id::Int = 0)

Stop search. The first form can be invoked via kill(search) as well (where kill is the function exported from Base). See also bg.