Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

InternedStrings.intern — Method.

Substrings are interned as their parent string type

InternedStrings.intern — Method.

intern(::Type{T}, s)

Intern s as if it were type T, converting it if required. Note that this will lead to unexpected behavour if the type of s, and T, do not have equivalent equality and hash functions (i.e. this is not safe if hash(s) != hash(convert(T, s))).

InternedStrings.intern — Method.

intern(s::T)

Return a reference to a interned instance of s, adding it to the interning pool if it did not already exist.