API

API

IntervalConstraintProgramming.Vol — Type.

N-dimensional Volume with lower and upper bounds

IntervalConstraintProgramming.pave — Method.

pave takes the given working list of boxes and splits them into inner and boundary lists with the given separator

IntervalConstraintProgramming.pave — Method.

pave(S::Separator, domain::IntervalBox, eps)`

Find the subset of domain defined by the constraints specified by the separator S. Returns (sub)pavings inner and boundary, i.e. lists of IntervalBox.

IntervalConstraintProgramming.CombinationSeparator — Type.

CombinationSeparator is a separator that is a combination (union, intersection, or complement) of other separators.

IntervalConstraintProgramming.ConstraintFunction — Type.

A ConstraintFunction contains the created forward and backward code

IntervalConstraintProgramming.ConstraintSeparator — Type.

ConstraintSeparator is a separator that represents a constraint defined directly using @constraint.

IntervalConstraintProgramming.Contractor — Type.

Contractor represents a Contractor from $\mathbb{R}^N$ to $\mathbb{R}^N$. Nout is the output dimension of the forward part.

IntervalConstraintProgramming.GeneratedFunction — Type.

A generated function, with the code that generated it

Base.:∩ — Method.

∩(S1::Separator, S2::Separator)

Separator for the intersection of two sets given by the separators S1 and S2. Takes an iterator of intervals (IntervalBox, tuple, array, etc.), of length equal to the total number of variables in S1 and S2; it returns inner and outer tuples of the same length

IntervalConstraintProgramming.isuniqued — Method.

Check if a symbol like :a has been uniqued to :_a_1_

IntervalConstraintProgramming.make_forward_function — Method.

Generate code for an anonymous function with given input arguments, output arguments, and code block.

IntervalConstraintProgramming.make_symbol — Method.

Return a new, unique symbol like z3

IntervalConstraintProgramming.process_assignment! — Method.

An assigment is of the form a = f(...). The name a is currently retained. TODO: It should later be made unique.

IntervalConstraintProgramming.process_call! — Function.

A call is something like +(x, y). A new variable is introduced for the result; its name can be specified using the new_var optional argument. If none is given, then a new, generated name is used.

IntervalConstraintProgramming.process_iterated_function! — Method.

Processes something of the form (f↑4)(x) (write as \uparrow<TAB>) by rewriting it to the equivalent set of iterated functions

IntervalConstraintProgramming.unify_variables — Method.

Unify the variables of two separators