API

API

IntervalRootFinding.Bisection — Type.

Bisection{F} <: Contractor{F}

Contractor type for the bisection method.

IntervalRootFinding.Krawczyk — Type.

Krawczyk{F, FP} <: Contractor{F}

Contractor type for the Krawczyk method.

# Fields
    - `f::F`: function whose roots are searched
    - `f::FP`: derivative or jacobian of `f`

IntervalRootFinding.Newton — Type.

Newton{F, FP} <: Contractor{F}

Contractor type for the Newton method.

# Fields
    - `f::F`: function whose roots are searched
    - `f::FP`: derivative or jacobian of `f`

IntervalRootFinding.Root — Type.

Object representing a possible root inside a given interval. status is either :unknown or :unique. If status is :unique then we know that there is a unique root of the function in question inside the given interval.

IntervalRootFinding.RootSearch — Type.

RootSearch{R <: Union{Interval,IntervalBox}, S <: Contractor, T <: Real}

Type implementing the Base.Iterator interface to the branch and prune routine. Returns the RootSearchState at each iteration. Note: Each iteration mutates the RootSearchState. Use copy(state::RootSearchState) to create an independent instance if necessary.

IntervalRootFinding.branch_and_prune — Function.

branch_and_prune(X, contract, tol=1e-3)

Generic branch and prune routine for finding isolated roots using the contract function as the contractor.

Inputs:

X: Interval or IntervalBox

contract: function that determines the status of a given box X. It returns the new box and a symbol indicating the status. Current possible values are of type Bisection or Newton

IntervalRootFinding.gauss_elimination_interval! — Method.

Solves the system of linear equations using Gaussian Elimination. Preconditioning is used when the precondition keyword argument is true.

REF: Luc Jaulin et al., Applied Interval Analysis, pg. 72

IntervalRootFinding.gauss_seidel_interval! — Method.

Iteratively solves the system of interval linear equations and returns the solution set. Uses the Gauss-Seidel method (Hansen-Sengupta version) to solve the system. Keyword precondition to turn preconditioning off. Eldon Hansen and G. William Walster : Global Optimization Using Interval Analysis - Chapter 5 - Page 115

IntervalRootFinding.newton1d — Method.

newton1d performs the interval Newton method on the given function f with its derivative f′ and initial interval x. Optional keyword arguments give the tolerances reltol and abstol. reltol is the tolerance on the relative error whereas abstol is the tolerance on |f(X)|, and a debug boolean argument that prints out diagnostic information.

IntervalRootFinding.newton1d — Method.

newton1d performs the interval Newton method on the given function f and initial interval x. Optional keyword arguments give the tolerances reltol and abstol. reltol is the tolerance on the relative error whereas abstol is the tolerance on |f(X)|, and a debug boolean argument that prints out diagnostic information.

IntervalRootFinding.quadratic_roots — Method.

Function to solve a quadratic equation where the coefficients are intervals. Returns an array of intervals of the roots. Arguments a, b and c are interval coefficients of x², x and 1 respectively. The interval case differs from the non-interval case in that there might be three disjoint interval roots. In the third case, one interval root extends to −∞ and another extends to +∞. This algorithm finds the set of points where F.lo(x) ≥ 0 and the set of points where F.hi(x) ≤ 0 and takes the intersection of these two sets. Eldon Hansen and G. William Walster : Global Optimization Using Interval Analysis - Chapter 8

IntervalRootFinding.step! — Method.

step!(state::RootSearchState, contractor, tolerance)

Progress state by treating one of its working regions. Note: state.working is always modified. If a root is found, it is added to state.outputs.

Polynomials.roots — Method.

roots(f, X, contractor, tol=1e-3)
roots(f, deriv, X, contractor, tol=1e-3)

Uses a generic branch and prune routine to find in principle all isolated roots of a function f:R^n → R^n in a box X, or a vector of boxes.

Inputs:

f: function whose roots will be found

X: Interval or IntervalBox

contractor: function that, when applied to the function f, determines the status of a given box X. It returns the new box and a symbol indicating the status. Current possible values are Bisection, Newton and Krawczyk

deriv ; explicit derivative of f for Newton and Krawczyk

IntervalRootFinding.Compl — Type.

Complex numbers as 2-vectors, enough for polynomials.

IntervalRootFinding.Contractor — Type.

Contractor{F}

Abstract type for contractors.

IntervalRootFinding.gauss_elimination_interval1! — Method.

Using Base.`

IntervalRootFinding.guarded_derivative_midpoint — Method.

Returns two intervals, the first being a point within the interval x such that the interval corresponding to the derivative of f there does not contain zero, and the second is the inverse of its derivative

IntervalRootFinding.newton — Method.

newton performs the interval Newton method on the given function f with its optional derivative f_prime and initial interval x. Optional keyword arguments give the tolerance, maxlevel at which to stop subdividing, and a debug boolean argument that prints out diagnostic information.

IntervalRootFinding.newton_refine — Method.

If a root is known to be inside an interval, newton_refine iterates the interval Newton method until that root is found.

IntervalRootFinding.newton_refine — Method.

If a root is known to be inside an interval, newton_refine iterates the interval Newton method until that root is found.

IntervalRootFinding.newtonlike_contract — Method.

newtonlike_contract(op, X, tol)

Contraction operation for contractors using the first derivative of the
function. This contraction use a bisection scheme to refine the intervals
with `:unkown` status.

Currently `Newton` and `Krawczyk` contractors uses this.

IntervalRootFinding.preconditioner — Method.

Preconditions the matrix A and b with the inverse of mid(A)

IntervalRootFinding.quadratic_helper! — Method.

Helper function for quadratic_interval that computes roots of a real quadratic using interval arithmetic to bound rounding errors.

IntervalRootFinding.realify — Method.

Takes a complex (polynomial) function f and returns a function g:R^2 -> R^2 that implements it.

IntervalRootFinding.realify_derivative — Method.

Takes the derivative of a complex function and returns the real jacobian that implements it.

IntervalRootFinding.refine — Function.

Generic refine operation for Krawczyk and Newton. This function assumes that it is already known that X contains a unique root. Call using e.g. op = X -> N(f, f′, X)

IntervalRootFinding.𝒦 — Method.

Multi-variable Krawczyk operator

IntervalRootFinding.𝒦 — Method.

Single-variable Krawczyk operator

IntervalRootFinding.𝒩 — Method.

Multi-variable Newton operator.

IntervalRootFinding.𝒩 — Method.

Single-variable Newton operator