Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

IntervalSets.AbstractInterval — Type.

A subtype of AbstractInterval{T} represents an interval subset of type T, that provides endpoints, closedendpoints.

IntervalSets.ClosedInterval — Type.

A ClosedInterval(left, right) is an interval set that includes both its upper and lower bounds. In mathematical notation, the constructed range is [left, right].

IntervalSets.Interval — Type.

An Interval{L,R}(left, right) where L,R are :open or :closed is an interval set containg x such that

left ≤ x ≤ right if L == R == :closed

left < x ≤ right if L == :open and R == :closed

left ≤ x < right if L == :closed and R == :open, or

left < x < right if L == R == :open

IntervalSets.OpenInterval — Type.

An TypedEndpointsInterval{:open,:open}(left, right) is an interval set that includes both its upper and lower bounds. In mathematical notation, the constructed range is (left, right).

IntervalSets.:.. — Method.

iv = l..r

Construct a ClosedInterval iv spanning the region from l to r.

IntervalSets.:± — Method.

iv = center±halfwidth

Construct a ClosedInterval iv spanning the region from center - halfwidth to center + halfwidth.

IntervalSets.closedendpoints — Method.

A tuple of Bool's encoding whether the left/right endpoints are closed.

IntervalSets.duration — Method.

duration(iv)

calculates the the total number of integers or dates of an integer or date valued interval. For example, duration(0..1) is 2, while width(0..1) is 1.

IntervalSets.endpoints — Method.

A tuple containing the left and right endpoints of the interval.

IntervalSets.isleftclosed — Method.

Is the interval closed at the left endpoint?

IntervalSets.isleftopen — Method.

Is the interval open at the left endpoint?

IntervalSets.isrightclosed — Method.

Is the interval closed at the right endpoint?

IntervalSets.isrightopen — Method.

Is the interval open at the right endpoint?

IntervalSets.leftendpoint — Method.

The left endpoint of the interval.

IntervalSets.rightendpoint — Method.

The right endpoint of the interval.

IntervalSets.width — Method.

w = width(iv)

Calculate the width (max-min) of interval iv. Note that for integers l and r, width(l..r) = length(l:r) - 1.

IntervalSets.Domain — Type.

A subtype of Domain{T} represents a subset of type T, that provides in.

IntervalSets.TypedEndpointsInterval — Type.

A subtype of TypedEndpointsInterval{L,R,T} where L and R are :open or :closed, that represents an interval subset of type T, and provides endpoints.