Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

IterativeRefinement.condInfest — Method.

condInfest(A,F,anorm)

computes an approximation to the condition of matrix A in the infinity-norm, using factorization F and the precomputed infinity norm anorm of A.

IterativeRefinement.equilibrators — Method.

equilibrators(A) -> R,C

compute row- and column-wise scaling vectors R,C for a matrix A such that the absolute value of the largest element in any row or column of Diagonal(R)*A*Diagonal(C) is close to unity. Designed to reduce the condition number of the working matrix.

IterativeRefinement.rfldiv — Method.

rfldiv(A,b,f=lu; kwargs...) -> x,bnorm,bcomp

Compute an accurate solution to a linear system $A x = b$ using extra-precise iterative refinement, with error bounds.

Returns solution x, a normwise relative forward error estimate bnorm, and maximum componentwise relative error estimate bcomp. Specifically, bnorm is an estimate of $‖xtrue - x‖ / ‖x‖$ (max norms). If the problem is so ill-conditioned that a good solution is unrealizable, bnorm and bcomp are set to unity (unless expert).

Arguments

A: a matrix,

b: a vector with the same eltype,

f: a factorization function such as lu.

Keywords

DT: higher-precision type for refinement; defaults to Float64 if the eltype of A and b is Float32, Double64 for Float64, etc.

verbosity: 0(default): quiet, 1: report on iterations, 2: details.

equilibrate::Bool: whether the function should equilibrate A (default true).

maxiter: default 20.

tol: relative tolerance for convergence, in units of eps(T).

expert::Bool: whether to return questionable bounds in extreme cases.

κ: the (max-norm) condition of A (see below).

F: a factorization of A (see below).

If A has already been equilibrated, and a Factorization object F and condition estimate κ have already been computed, they may be provided as keyword arguments; no check for consistency is done here.

Uses the algorithm of Demmel et al. ACM TOMS, 32, 325 (2006).

IterativeRefinement.norm1est! — Method.

norm1est!(applyA!,applyAH!,y::Vector) => γ

Estimate the 1-norm of a linear operator A expressed as functions which apply A and adjoint(A) to a vector such as y.

cf. N.J. Higham, SIAM J. Sci. Stat. Comp. 11, 804 (1990)