Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

JDBC.JDBCRowIterator — Type.

Iterator to get rows of tables as array of tuples.

JDBC.commit — Method.

Commit any pending transaction to the database. Throws a JDBCError if connection is null.

Returns nothing.

JDBC.commit — Method.

commit(connection::JConnection)

Commits the transaction

Args

connection: The connection object

Returns

None

JDBC.connection — Method.

Return the corresponding connection for a given cursor.

JDBC.createStatement — Method.

createStatement(connection::JConnection)

Initializes a Statement

Args

connection: The connection object

Returns

The JStatement object

JDBC.cursor — Method.

Create a new database cursor.

Returns a JDBCCursor instance.

JDBC.execute! — Method.

Run a query on a database.

The results of the query are not returned by this function but are accessible through the cursor.

parameters can be any iterable of positional parameters, or of some T<:Associative for keyword/named parameters.

Throws a JDBCError if query caused an error, cursor is not initialized or connection is null.

Returns nothing.

JDBC.execute — Method.

execute(stmt::JStatement, query::AbstractString)

Executes the auery based on JStatement or any of its sub-types

Args

stmt: The JStatement object or any of its sub-types

query: The query to be executed

Returns

A boolean indicating whether the execution was successful or not

JDBC.execute — Method.

execute(stmt::(Union{JPreparedStatement, JCallableStatement}))

Executes the auery based on the Prepared Statement or Callable Statement

Args

stmt: The Prepared Statement or the Callable Statement object

Returns

A boolean indicating whether the execution was successful or not

JDBC.executeQuery — Method.

executeQuery(stmt::JStatement, query::AbstractString)

Executes the auery and returns the results as a JResultSet object.

Args

stmt: The Statement object

query: The query to be executed

Returns

The result set as a JResultSet object

JDBC.executeQuery — Method.

executeQuery(stmt::Union{JPreparedStatement, JCallableStatement})

Executes the auery based on a JPreparedStatement object or a JCallableStatement object

Args

stmt: The JPreparedStatement object or JCallableStatement object

Returns

The result set as a JResultSet object

JDBC.executeUpdate — Method.

executeUpdate(stmt::JStatement, query::AbstractString)

Executes the update auery and returns the status of the execution of the query

Args

stmt: The Statement object

query: The query to be executed

Returns

An integer representing the status of the execution

JDBC.executeUpdate — Method.

executeUpdate(stmt::Union{JPreparedStatement, JCallableStatement})

Executes the update auery based on a JPreparedStatement object or a JCallableStatement object

Args

stmt: The JPreparedStatement object or JCallableStatement object

Returns

An integer indicating the status of the execution of the query

JDBC.getColumnCount — Method.

getColumnCount(rsmd::JResultSetMetaData)

Returns the number of columns based on the JResultSetMetaData object

Args

rsmd: The JResultSetMetaData object

Returns

The number of columns.

JDBC.getColumnName — Method.

getColumnName(rsmd::JResultSetMetaData, col::Integer)

Returns the column's name based on the JResultSetMetaData object and the column number

Args

rsmd: The JResultSetMetaData object

col: The column number

Returns

The column name

JDBC.getColumnType — Method.

getColumnType(rsmd::JResultSetMetaData, col::Integer)

Returns the column's data type based on the JResultSetMetaData object and the column number

Args

rsmd: The JResultSetMetaData object

col: The column number

Returns

The column type as an integer

JDBC.getDate — Method.

getDate(rs::Union{JResultSet, JCallableStatement}, fld::AbstractString)

Returns the Date object based on the result set or a callable statement. The value is extracted based on the column name.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column name

Returns

The Date object.

JDBC.getDate — Method.

getDate(rs::Union{JResultSet, JCallableStatement}, fld::Integer)

Returns the Date object based on the result set or a callable statement. The value is extracted based on the column number.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column number

Returns

The Date object.

JDBC.getMetaData — Method.

getMetaData(rs::JResultSet)

Returns information about the types and properties of the columns in the ResultSet object

Args

stmt: The JResultSet object

Returns

The JResultSetMetaData object.

JDBC.getResultSet — Method.

getResultSet(stmt::JStatement)

Returns the result set based on the previous execution of the query based on a JStatement

Args

stmt: The JStatement object

Returns

The JResultSet object.

JDBC.getTableMetaData — Method.

Get the metadata (column name and type) for each column of the table in the result set rs.

Returns an array of (column name, column type) tuples.

JDBC.getTime — Method.

getTime(rs::Union{JResultSet, JCallableStatement}, fld::AbstractString)

Returns the Time object based on the result set or a callable statement. The value is extracted based on the column name.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column name

Returns

The Time object.

JDBC.getTime — Method.

getTime(rs::Union{JResultSet, JCallableStatement}, fld::Integer)

Returns the Time object based on the result set or a callable statement. The value is extracted based on the column number.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column number

Returns

The Time object.

JDBC.getTimestamp — Method.

getTimestamp(rs::Union{JResultSet, JCallableStatement}, fld::AbstractString)

Returns the Timestamp object based on the result set or a callable statement. The value is extracted based on the column name.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column name

Returns

The Timestamp object.

JDBC.getTimestamp — Method.

getTimestamp(rs::@Union{JResultSet, JCallableStatement}, fld::Integer)

Returns the Timestamp object based on the result set or a callable statement. The value is extracted based on the column number.

Args

stmt: The JResultSet or JCallableStatement object

fld: The column number

Returns

The Timestamp object.

JDBC.isopen — Method.

Returns a boolean indicating whether connection conn is open.

JDBC.prepareCall — Method.

prepareCall(connection::JConnection, query::AbstractString)

Prepares the Callable Statement for the given query

Args

connection: The connection object

query: The query string

Returns

The JCallableStatement object

JDBC.prepareStatement — Method.

prepareStatement(connection::JConnection, query::AbstractString)

Prepares the Statement for the given query

Args

connection: The connection object

query: The query string

Returns

The JPreparedStatement object

JDBC.rollback — Method.

Roll back to the start of any pending transaction. Throws a JDBCError if connection is null.

Returns nothing.

JDBC.rollback — Method.

rollback(connection::JConnection)

Rolls back the transactions.

Args

connection: The connection object

Returns

None

JDBC.rows — Method.

Create a row iterator.

This method returns an instance of an iterator type which returns one row on each iteration. Each row returnes a Tuple{...}.

Throws a JDBCError if execute! was not called on the cursor or connection is null.

Returns a JDBCRowIterator instance.

JDBC.setAutoCommit — Method.

setAutoCommit(connection::JConnection, x::Bool)

Set the Auto Commit flag to either true or false. If set to false, commit has to be called explicitly

Args

connection: The connection object

Returns

None

JDBC.setFetchSize — Method.

setFetchSize(stmt::Union{JStatement, JPreparedStatement, JCallableStatement }, x::Integer)

Sets the fetch size in a JStatement or a JPreparedStatement object or a JCallableStatement object. The number of records that are returned in subsequent query executions are determined by what is set here.

Args

stmt: The JPreparedStatement object or JCallableStatement object

x: The number of records to be returned

Returns

None

JDBC.Connection — Method.

Open a JDBC Connection to the specified host. The username and password can be optionally passed as a Dictionary props of the form Dict("user" => "username", "passwd" => "password"). The JDBC connector location can be optionally passed as connectorpath, if it is not added to the java class path.

Returns a JDBCConnection instance.

Base.close — Method.

Closes the JDBCConnection conn. Throws a JDBCError if connection is null.

Returns nothing.

Base.close — Method.

Close the JDBCCursor csr. Throws a JDBCError if cursor is not initialized.

Returns nothing.

JDBC.absolute! — Method.

absolute!(rs::JResultSet, row::Int)

Move the cursor to the specified row.

Args

rs: The JResultSet object.

row: The row to move to.

Returns

A boolean indicating success.

JDBC.beforeFirst! — Method.

beforeFirst!(rs::JResultSet)

Move the cursor to the front of the JResultSet, before the first row.

Args

rs: The JResultSet object.

Returns

nothing

JDBC.clearParameters — Method.

clearParameters(stmt::Union{JPreparedStatement, JCallableStatement})

Clears the currently held parameters in a JPreparedStatement object or a JCallableStatement object

Args

stmt: The JPreparedStatement object or JCallableStatement object

Returns

None