Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

JSONSchema.Schema — Type.

Schema(sch::String)

Create a schema for document validation. sch should be a String containing a valid JSON.

Schema(sch::Dict)

Create a schema but with sch being a parsed JSON created with JSON.parse() or JSON.parsefile().

Example

julia> myschema = Schema("
  {
    "properties": {
      "foo": {},
      "bar": {}
    },
    "required": ["foo"]
  }
  ")

julia> sch = JSON.parsefile(filepath)
julia> myschema = Schema(sch)

Base.isvalid — Method.

isvalid(x, s::Schema)

Check that a document x is valid against the Schema s.

Examples

julia> myschema = Schema("
  {
    "properties": {
      "foo": {},
      "bar": {}
    },
    "required": ["foo"]
  }
  ")

julia> isvalid( JSON.parse("{ "foo": true }"), myschema)
true
julia> isvalid( JSON.parse("{ "bar": 12.5 }"), myschema)
false

JSONSchema.diagnose — Method.

diagnose(x, s::Schema)

Check that a document x is valid against the Schema s. If valid return nothing, and if not, return a diagnostic String containing a selection of one or more likely causes of failure.

Examples

julia> diagnose( JSON.parse("{ "foo": true }"), myschema)
nothing
julia> diagnose( JSON.parse("{ "bar": 12.5 }"), myschema)
"in [] : required property 'foo' missing"