Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Jackknife.Reducer — Type.

Reducer(f)

Type that holds a function and enforces at runtime that the function returns either a Real value or missing. This is used internally to sanity check the input for functions which accept a function argument.

Jackknife.bias — Method.

Jackknife.bias(estimator::Function, x)

Compute the jackknife estimate of the bias of estimator, which is given as a function that computes a point estimate when passed a real-valued vector x.

Jackknife.estimate — Method.

Jackknife.estimate(estimator::Function, x)

Compute the bias-corrected jackknife estimate of the parameter estimator(x).

Jackknife.leaveoneout — Method.

Jackknife.leaveoneout(estimator::Function, x::AbstractVector{<:Real})

Estimate the parameter $\hat{\theta}$ using the estimating function estimator for each subsample of x, systematically omitting each index one at a time. The result is a vector of length length(x)-1 of parameter estimates.

Jackknife.variance — Method.

Jackknife.variance(estimator::Function, x)

Compute the jackknife estimate of the variance of estimator, which is given as a function that computes a point estimate when passed a real-valued vector x.