Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

JavaCall.classforname — Method.

classforname(name::String)

Create an instance of Class<name> (same as Class.forName(name) in Java)

Args

name: The name of a class to instantiate

Returns

JavaObject Instance of Class<name>

JavaCall.getname — Method.

getname(cls::JClass)

Returns the fully qualified name of the java class

Args

cls: The JClass object

Returns

The fully qualified name of the java class

JavaCall.getname — Method.

getname(method::JMethod)

Returns the fully qualified name of the java method

Args

cls: The JClass method

Returns

The fully qualified name of the method

JavaCall.getparametertypes — Method.

getparametertypes(method::JMethod)

Returns the parameter types of the java method

Args

method: The java method object

Returns

Vector the parametertypes

JavaCall.getreturntype — Method.

getreturntype(method::JMethod)

Returns the return type of the java method

Args

method: The java method object

Returns

Returns the type of the return object as a JClass

JavaCall.isnull — Method.

isnull(obj::JavaMetaClass)

Checks if the passed JavaMetaClass is null or not

Args

obj: The object of type JavaMetaClass

Returns

true if the passed object is null else false

JavaCall.isnull — Method.

isnull(obj::JavaObject)

Checks if the passed JavaObject is null or not

Args

obj: The object of type JavaObject

Returns

true if the passed object is null else false

JavaCall.listmethods — Method.

listmethods(obj::JavaObject)

Lists the methods that are available on the java object passed

Args

obj: The java object

Returns

List of methods

JavaCall.listmethods — Method.

listmethods(obj::JavaObject, name::AbstractString)

Lists the methods that are available on the java object passed. The methods are filtered based on the name passed

Args

obj: The java object

name: The filter (e.g. method name)

Returns

List of methods available on the java object and matching the name passed

JavaCall.narrow — Method.

Given a JavaObject{T} narrows down T to a real class of the underlying object. For example, JavaObject{:java.lang.Object} pointing to java.lang.String will be narrowed down to JavaObject{:java.lang.String}