Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

JeszenszkiBasis.RestrictedSzbasis — Type.

Basis of occupation vectors with a site occupation restriction.

JeszenszkiBasis.RestrictedSzbasis — Method.

RestrictedSzbasis(K::Int, N::Int, M::Int)

Create a basis for K sites and N particles, with no more than M particles per site.

JeszenszkiBasis.Szbasis — Type.

Basis of occupation vectors.

JeszenszkiBasis.Szbasis — Method.

Szbasis(K::Int, N::Int)

Create a basis for K sites and N particles.

JeszenszkiBasis.num_vectors — Method.

num_vectors(N::Int, K::Int, M::Int)

Compute the number of vectors in a basis with N particles and K sites, and a limit of M particles per site.

JeszenszkiBasis.num_vectors — Method.

num_vectors(N::Int, K::Int)

Compute the number of vectors in a basis with N particles and K sites.

JeszenszkiBasis.num_vectors — Method.

num_vectors(basis::AbstractSzbasis, N::Int, K::Int)

Compute the number of vectors in a basis with N particles and K sites, and a limit of as many particles per site as in basis.

JeszenszkiBasis.serial_num — Method.

serial_num(K::Int, N::Int, v::AbstractVector{Int})

Compute the serial number of occupation vector v in a basis with K sites and N particles.

JeszenszkiBasis.serial_num — Method.

serial_num(K::Int, N::Int, M::Int, v::AbstractVector{Int})

Compute the serial number of occupation vector v in a basis with K sites and N particles, and a limit of M particles per site.

JeszenszkiBasis.serial_num — Method.

serial_num(basis::AbstractSzbasis, v::AbstractVector{Int})

Compute the serial number of occupation vector v in basis.

JeszenszkiBasis.serial_num — Method.

serial_num(basis::AbstractSzbasis, K::Int, N::Int, v::AbstractVector{Int})

Compute the serial number of occupation vector v in a basis with N particles and K sites, and a limit of as many particles per site as in basis.

JeszenszkiBasis.site_max — Method.

site_max(basis::AbstractSzbasis)

Get the maximum number of particles on a site in basis.

JeszenszkiBasis.sub_serial_num — Method.

sub_serial_num(basis::AbstractSzbasis, v::AbstractVector{Int})

Compute the serial number of the reduced occupation vector v, which has only a subset of the sites present in basis.

JeszenszkiBasis.to_str — Method.

to_str(v::Vector{Int})

Convert an occupation vector v to a space-separated string.