Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

JuMP.macro_assign_and_return — Method.

macro_assign_and_return(code, variable, name;
                        register_fun::Union{Nothing, Function}=nothing,
                        model=nothing)

Return runs code in a local scope which returns the value of variable and then assign variable to name. If register_fun is given, register_fun(model, name, variable) is called.

JuMP.macro_return — Method.

macro_return(model, code, variable)

Return a block of code that 1. runs the code block code in a local scope and 2. returns the value of a local variable named variable. variable must reference a variable defined by code.