Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

JuliaWebAPI.APIInvoker — Type.

APIInvoker holds the transport and format used for a remote api call.

JuliaWebAPI.APIInvoker — Type.

APIInvoker holds the transport and format used for a remote api call. This method creates an APIInvoker over ZMQ transport using JSON message format. (provided for backward compatibility)

JuliaWebAPI.APIResponder — Type.

APIResponder holds the transport and format used for data exchange and the endpoint specifications.

JuliaWebAPI.APIResponder — Type.

APIResponder holds the transport and format used for data exchange and the endpoint specifications. This method creates an APIResponder over ZMQ transport using JSON message format. (provided for backward compatibility)

JuliaWebAPI.DictMsgFormat — Type.

Intermediate format that is just a Dict. No serialization. A Dict with cmd (string), args (array), vargs (dict).

JuliaWebAPI.InProcTransport — Type.

Transport layer over in-process Channels

JuliaWebAPI.JSONMsgFormat — Type.

Intermediate format based on JSON. A JSON object with cmd (string), args (array), vargs (dict).

JuliaWebAPI.SerializedMsgFormat — Type.

Intermediate format based on Julia serialization. A dict with cmd (string), args (array), vargs (dict).

JuliaWebAPI.ZMQTransport — Type.

Transport layer over ZMQ sockets

JuliaWebAPI.apicall — Method.

Calls a remote api cmd with args... and data.... The response is formatted as specified by the formatter specified in conn.

JuliaWebAPI.fnresponse — Method.

extract and return the response data as a direct function call would have returned but throw error if the call was not successful.

JuliaWebAPI.httpresponse — Method.

construct an HTTP Response object from the API response

JuliaWebAPI.process — Method.

start processing as a server

JuliaWebAPI.register — Method.

Register a function as API call. TODO: validate method belongs to module?

JuliaWebAPI.call_api — Method.

call the actual API method, and send the return value back as response

JuliaWebAPI.respond — Method.

send a response over the transport in the specified format