Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Juno.clearconsole — Method.

clearconsole()

Clear the console if Juno is used; does nothing otherwise.

Juno.input — Function.

input(prompt = "") -> "..."

Prompt the user to input some text, and return it. Optionally display a prompt.

Juno.selector — Method.

selector([xs...]) -> x

Allow the user to select one of the xs.

xs should be an iterator of strings. Currently there is no fallback in other environments.

Juno.structure — Method.

structure(x)

Display x's underlying representation, rather than using its normal display method.

For example, structure(:(2x+1)) displays the Expr object with its head and args fields instead of printing the expression.

Juno.defaultrepr — Function.

defaultrepr(x, lazy = false)

render fallback for types without any specialized show methods.

If lazy is true, then the type's fields will be loaded lazily when expanding the tree. This is useful when the fields contain big elements that might need to be inspectable.

Can be used by packages to restore Juno's default printing if they have defined a show method that should not be used by Juno:

Juno.render(i::Juno.Inline, x::myType) = Juno.render(i, Juno.defaultrepr(x))

Juno.info — Method.

info(msg)

Show the given message in Juno's console using blue styling, or fall back to Base.info.

In a package, you can use import Juno: info to replace the default version with this one.

Juno.isactive — Method.

isactive()

Will return true when the current Julia process is connected to a running Juno frontend.

Juno.notify — Method.

notify(msg)

Display msg as an OS specific notification.

Useful for signaling the end of a long running computation or similar. This disregards the Notifications setting in julia-client. Falls back to info(msg) in other environments.

Juno.plotsize — Method.

plotsize()

Get the size of Juno's plot pane in px. Returns [100, 100] as a fallback value.

Juno.profiler — Method.

profiler()

Show currently collected profile information as an in-editor flamechart.

Juno.profiletree — Method.

profiletree()

Show currently collected profile information in tree-form. Falls back to Profile.print().

Juno.progress — Method.

progress(f::Function; name = "", msg = "")

Evaluates f with id as its argument and makes sure to destroy the progress bar afterwards. To update the progress bar in f you can call a logging statement like @info or even just @logmsg with _id=id and progress as arguments.

progress can take either of the following values:

0 <= progress < 1: create or update progress bar

progress == nothing || progress = NaN: set progress bar to indeterminate progress

progress > 1 || progress == "done": destroy progress bar

The logging message (e.g. "foo" in @info "foo") will be used as the progress bar's name.

Juno.progress() do id
    for i = 1:10
        sleep(0.5)
        @info "iterating" progress=i/10 _id=id
    end
end

Juno.syntaxcolors — Method.

syntaxcolors(selectors = Atom.SELECTORS)::Dict{String, UInt32}

Get the colors used by the current Atom theme. selectors should be a Dict{String, Vector{String}} which assigns a css selector (e.g. syntax--julia) to a name (e.g. variable).