Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

KrigingEstimators.ExternalDriftKriging — Type.

ExternalDriftKriging(γ, drifts)
ExternalDriftKriging(X, z, γ, drifts)

External Drift Kriging with variogram model γ and external drifts functions.

Optionally, pass the coordinates X and values z to the fit function.

Notes

External drift functions should be smooth

Kriging system with external drift is often unstable

Include a constant drift (e.g. x->1) for unbiased estimation

OrdinaryKriging is recovered for drifts = [x->1]

For polynomial mean, see UniversalKriging

KrigingEstimators.KrigingEstimator — Type.

KrigingEstimator

A Kriging estimator (e.g. Simple Kriging).

KrigingEstimators.OrdinaryKriging — Type.

OrdinaryKriging(γ)
OrdinaryKriging(X, z, γ)

Ordinary Kriging with variogram model γ.

Optionally, pass the coordinates X and values z to the fit function.

KrigingEstimators.SimpleKriging — Type.

SimpleKriging(γ, μ)
SimpleKriging(X, z, γ, μ)

Simple Kriging with variogram model γ and constant mean μ.

Optionally, pass the coordinates X and values z to the fit function.

Notes

Simple Kriging requires stationary variograms

KrigingEstimators.UniversalKriging — Type.

UniversalKriging(γ, degree, dim)
UniversalKriging(X, z, γ, degree)

Universal Kriging with variogram model γ and polynomial degree on a spatial domain of dimension dim.

Optionally, pass the coordinates X and values z to the fit function.

Notes

OrdinaryKriging is recovered for 0th degree polynomial

For non-polynomial mean, see ExternalDriftKriging

KrigingEstimators.fit — Method.

fit(estimator, X, z)

Build Kriging system from coordinates X and values z and return a fitted estimator.

KrigingEstimators.predict — Method.

predict(estimator, xₒ)

Compute mean and variance for the estimator at coordinates xₒ.

KrigingEstimators.status — Method.

status(fittedkrig)

Return the status of the fittedkrig object, meaning the factorization of the Kriging system was successful.

KrigingEstimators.weights — Method.

weights(estimator, xₒ)

Compute the weights λ (and Lagrange multipliers ν) for the estimator at coordinates xₒ.

KrigingEstimators.FittedKriging — Type.

FittedKriging(estimator, state)

An object that can be used for making predictions using the parameters in estimator and the current Kriging state.

KrigingEstimators.KrigingState — Type.

KrigingState(X, z, LHS, RHS)

A Kriging state stores information needed to perform estimation at any given location.

KrigingEstimators.KrigingWeights — Type.

KrigingWeights(λ, ν)

An object storing Kriging weights λ and Lagrange multipliers ν.

KrigingEstimators.combine — Method.

combine(estimator, weights, z)

Combine weights with values z to produce mean and variance.

KrigingEstimators.factorize — Method.

factorize(estimator, LHS)

Factorize LHS of Kriging system with appropriate factorization method.

KrigingEstimators.lhs — Method.

lhs(estimator, X)

Return LHS of Kriging system using spatial configuration X.

KrigingEstimators.nconstraints — Method.

nconstraints(estimator)

Return number of constraints for estimator.

KrigingEstimators.set_constraints_lhs! — Method.

set_constraints_lhs!(estimator, LHS, X)

Set constraints in LHS of Kriging system.

KrigingEstimators.set_constraints_rhs! — Method.

set_constraints_rhs!(estimator, xₒ)

Set constraints in RHS of Kriging system.

KrigingEstimators.set_rhs! — Method.

set_rhs!(estimator, xₒ)

Set RHS of Kriging system at coodinates xₒ.