Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LBFGSB.setulb — Method.

setulb(n, m, x, l, u, nbd, f, g, factr, pgtol, wa, iwa, task, iprint, csave, lsave, isave, dsave)

This function wraps L-BFGS-B's fortran subroutine setulb which partitions the working arrays wa and iwa, and then uses the limited memory BFGS method to solve the bound constrained optimization problem by calling mainlb.

Note, nmax is the dimension of the largest problem to be solved, mmax is the maximum number of limited memory corrections.

Arguments

n::Ref{Cint}: the dimension of the problem.

m::Ref{Cint}: the maximum number of variable metric corrections used to define the limited memory matrix.

x::Vector{Cdouble}: an approximation to the solution, of length n.

l::Vector{Cdouble}: the lower bound on x, of length n.

u::Vector{Cdouble}: the upper bound on x, of length n.

nbd::Vector{Cint}: represents the type of bounds imposed on the variables, and must be specified as follows:

nbd[i]=0 if x[i] is unbounded;

nbd[i]=1 if x(i) has only a lower bound;

nbd[i]=2 if x[i] has both lower and upper bounds;

nbd[i]=3 if x[i] has only an upper bound.

f::Ref{Cdouble}: the value of the function at x(on first entry f is unspecified).

g::Vector{Cdouble}: the value of the gradient at x(on first entry g is unspecified).

factr::Ref{Cdouble}: the iteration will stop when (f^k - f^{k+1})/max{|f^k|,|f^{k+1}|,1} <= factr*epsmch, where epsmch is the machine precision, which is automatically generated by the code. Typical values for factr:

1.e12 for low accuracy;

1.e7 for moderate accuracy;

1.e1 for extremely high accuracy.

pgtol::Ref{Cdouble}: the iteration will stop when max{|proj gi | i = 1, ..., n} <= pgtol where pgi is the ith component of the projected gradient.

wa::Vector{Cdouble}: the working array of length (2mmax + 5)nmax + 12mmax^2 + 12mmax.

iwa::Vector{Cint}: the integer working array of length 3nmax.

task::Vector{Cuchar}: a working string of characters of length 60 indicating the current job when entering and quitting this function.

iprint::Ref{Cint}: it controls the frequency and type of output generated:

iprint<0 no output is generated;

iprint=0 print only one line at the last iteration;

0<iprint<99 print also f and |proj g| every iprint iterations;

iprint=99 print details of every iteration except n-vectors;

iprint=100 print also the changes of active set and final x;

iprint>100 print details of every iteration including x and g.

when iprint > 0, the file iterate.dat will be created to summarize the iteration.

csave::Vector{Cuchar}: a working string of characters of length 60.

lsave::Vector{Bool}: a logical working array of dimension 4. if task = NEW_X, the following information is available:

if lsave[1] == true then the initial X has been replaced by its projection in the feasible set;

if lsave[2] == true then the problem is constrained;

if lsave[3] == true then each variable has upper and lower bounds.

isave::Vector{Cint}: an integer working array of dimension 44. if task = NEW_X, the following information is available:

isave[22] = the total number of intervals explored in the search of Cauchy points;

isave[26] = the total number of skipped BFGS updates before the current iteration;

isave[30] = the number of current iteration;

isave[31] = the total number of BFGS updates prior the current iteration;

isave[33] = the number of intervals explored in the search of Cauchy point in the current iteration;

isave[34] = the total number of function and gradient evaluations;

isave[36] = the number of function value or gradient evaluations in the current iteration;

if isave[37] = 0 then the subspace argmin is within the box;

if isave[37] = 1 then the subspace argmin is beyond the box;

isave[38] = the number of free variables in the current iteration;

isave[39] = the number of active constraints in the current iteration;

n + 1 - isave[40] = the number of variables leaving the set of active constraints in the current iteration;

isave[41] = the number of variables entering the set of active constraints in the current iteration.

dsave::Vector{Cdouble}: a double precision working array of dimension 29. if task = NEW_X, the following information is available:

dsave[1] = current 'theta' in the BFGS matrix;

dsave[2] = f(x) in the previous iteration;

dsave[3] = factr*epsmch;

dsave[4] = 2-norm of the line search direction vector;

dsave[5] = the machine precision epsmch generated by the code;

dsave[7] = the accumulated time spent on searching for Cauchy points;

dsave[8] = the accumulated time spent on subspace minimization;

dsave[9] = the accumulated time spent on line search;

dsave[11] = the slope of the line search function at the current point of line search;

dsave[12] = the maximum relative step length imposed in line search;

dsave[13] = the infinity norm of the projected gradient;

dsave[14] = the relative step length in the line search;

dsave[15] = the slope of the line search function at the starting point of the line search;

dsave[16] = the square of the 2-norm of the line search direction vector.

LBFGSB.timer — Method.

timer(x)

The double precision cpu timing subroutine in the L-BFGS-B code.