Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LPVSpectral.SpectralExt — Type.

LPV Spectral estimation result type.

See ls_spectral_lpv for additional help.

An object of this type can be plotted if Plots.jl is installed. Use regular Plots-syntax, with the additional attributes

normalization= :none / :sum / :max
normdim = :freq / :v # Only applies if normalization= :sum or :max
dims = 2 or 3 (default = 2)

Fields:

Y::AbstractVector
X::AbstractVector
V::AbstractVector
w
Nv
λ
coulomb::Bool
normalize::Bool
x                   # The estimated parameters
Σ                   # Covariance of the estimated parameters

LPVSpectral.basis_activation_func — Method.

basisactivationfunc(V,Nv,normalize,coulomb)

Returns a func v->ϕ(v) ∈ ℜ(Nv) that calculates the activation of Nv basis functions spread out to cover V nicely. If coulomb is true, then we get twice the number of basis functions, 2Nv

LPVSpectral.detrend — Method.

detrend(x, order=0, t = 1:length(x)) Removes the trend of order order, i.e, mean and, if order=1, the slope of the signal x If order = 1, then the sampling points of x can be supplied as t (default: t = 1:length(x))

LPVSpectral.ls_cohere — Method.

ls_cohere(y,u,t,freqs; nw = 10, noverlap = -1, estimator=ls_spectral, kwargs...)

Perform spectral coherence estimation using the least-squares method.

estimator is the spectral estimatio function to use, default is ls_spectral. For sparse estimation, try estimator = ls_sparse_spectral See ls_sparse_spectral for more help. See also ls_windowcsd and ls_spectral for additional help.

LPVSpectral.ls_sparse_spectral — Function.

ls_sparse_spectral(y,t,f=(0:((length(y)-1)/2))/length(y), [window]; λ=1, proxg = ProximalOperators.NormL1(λ), kwargs...)

perform spectral estimation using the least-squares method with (default) a L1-norm penalty on the Fourier coefficients, change kwarg proxg to e.g. NormL0(λ) for a different behavior or proxg = IndBallL0(4) if the number of frequencies is known in advance. Promotes a sparse spectrum. See ?ADMM for keyword arguments to control the solver.

y is the signal to be analyzed t is the sampling points f is a vector of frequencies

LPVSpectral.ls_sparse_spectral_lpv — Method.

`lssparsespectral_lpv(Y,X,V,w,Nv::Int; λ = 1, coulomb = false, normalize=true, coulomb = false, normalize = true)

Perform LPV spectral estimation using the method presented in Bagge Carlson et al. "Linear Parameter-Varying Spectral Decomposition." modified to include a sparsity-promoting L1 group-lasso penalty on the coefficients. The groups are based on frequency, meaning a solution in which either all parameters For a particular frequency are zero, or all are non-zero. This is useful in the identification of frequency components among a large set of possible frequencies. See the paper or README For additional details.

Y output

X sample locations

V scheduling signal

w frequency vector

Nv number of basis functions

λ Regularization parameter

coulomb Assume discontinuity at v=0 (useful For signals where, e.g., Coulomb friction might cause issues.)

normalize Use normalized basis functions (See paper For details).

See ?ADMM for keyword arguments to control the solver.

See also psd, ls_spectral_lpv and ls_windowpsd_lpv

LPVSpectral.ls_spectral — Function.

ls_spectral(y,t,f=(0:((length(y)-1)/2))/length(y); λ=0)

perform spectral estimation using the least-squares method y is the signal to be analyzed t is the sampling points f is a vector of frequencies

See also ls_sparse_spectral tls_spectral

LPVSpectral.ls_spectral — Method.

ls_spectral(y,t,f,W::AbstractVector) W is a vector of weights, for weighted least-squares

LPVSpectral.ls_spectral_lpv — Method.

ls_spectral_lpv(Y,X,V,w,Nv::Int; λ = 1e-8, coulomb = false, normalize=true)

Perform LPV spectral estimation using the method presented in Bagge Carlson et al. "Linear Parameter-Varying Spectral Decomposition." See the paper For additional details.

Y output

X sample locations

V scheduling signal

w frequency vector

Nv number of basis functions

λ Regularization parameter

coulomb Assume discontinuity at v=0 (useful for signals where, e.g., Coulomb friction might cause issues.)

normalize Use normalized basis functions (See paper for details).

The method will issue a warning if less than 90% of the variance in Y is described by the estimated model. If this is the case, try increasing either the number of frequencies or the number of basis functions per frequency. Alternatively, try lowering the regularization parameter λ.

See also psd, ls_sparse_spectral_lpv and ls_windowpsd_lpv

LPVSpectral.ls_windowcsd — Method.

ls_windowcsd(y,u,t,freqs; nw = 10, noverlap = -1, window_func=rect, estimator=ls_spectral, kwargs...)

Perform windowed cross spectral density estimation using the least-squares method.

y and u are the two signals to be analyzed and t::AbstractVector are their sampling points window_func defaults to Windows.rect

estimator is the spectral estimatio function to use, default is ls_spectral. For sparse estimation, try estimator = ls_sparse_spectral See ls_sparse_spectral for more help.

See ls_spectral for additional help.

LPVSpectral.ls_windowpsd — Method.

ls_windowpsd(y,t,freqs; nw = 10, noverlap = -1, window_func=rect, estimator=ls_spectral, kwargs...)

perform widowed spectral estimation using the least-squares method. window_func defaults to Windows.rect

estimator is the spectral estimatio function to use, default is ls_spectral. For sparse estimation, try estimator = ls_sparse_spectral See ls_sparse_spectral for more help. kwargs are passed to estimator.

See ls_spectral for additional help.

LPVSpectral.ls_windowpsd_lpv — Function.

lswindowpsdlpv(Y::AbstractVector,X::AbstractVector,V::AbstractVector,w,Nv::Integer, nw::Int=10, noverlap=0; kwargs...)

Perform windowed psd estimation using the LPV method. A rectangular window is always used.

See ?ls_spectral_lpv for additional help.

LPVSpectral.pdf — Method.

f(cn::ComplexNormal, z)

Probability density Function f(z) for a complex normal distribution. This can probably be more efficiently implemented

LPVSpectral.psd — Method.

psd(se::SpectralExt) Compute the power spectral density For a SpectralExt object

See also ls_windowpsd_lpv

LPVSpectral.tls_spectral — Function.

tls_spectral(y,t,f=(0:((length(y)-1)/2))/length(y)) Perform total least-squares spectral estimation using the SVD-method. See ls_spectral for additional help

LPVSpectral.Windows2 — Type.

Windows2(y,t,nw,noverlap,window_func) noverlap = -1 sets the noverlap to dpw/2

LPVSpectral.Windows3 — Type.

Windows3(y,t,v,nw::Integer,noverlap::Integer,window_func::Function) noverlap = -1 sets the noverlap to dpw/2

LPVSpectral.ADMM — Method.

ADMM(x,proxf,proxg;
iters      = 10000,   # ADMM maximum number of iterations
tol        = 1e-5,    # ADMM tolerance
printerval = 100,     # Print this often
cb(x,z)    = nothing, # Callback function
μ          = 0.05`)   # ADMM tuning parameter. If results oscillate, lower this value.

LPVSpectral.real_complex_bs — Function.

real_complex_bs(A,b, λ=0) Replaces the backslash operator For complex arguments. Expands the A-matrix into [real(A) imag(A)] and performs the computation using real arithmetics. Optionally accepts λ to solve the ridge regression problem using the formulation [A;λI]\[b;0]. λ should be given with the same dimension as the columns of A, i.e. if λ represents a standard deviation, then λ = σ, not λ = σ²

LPVSpectral.reshape_params — Method.

Returns params as a [nω × N] matrix

LPVSpectral.ridgereg — Method.

ridgereg(A,b,λ) Accepts λ to solve the ridge regression problem using the formulation [A;λI]\[b;0]. λ should be given with the same dimension as the columns of A, i.e. if λ represents an inverse standard deviation, then 1/λ = σ, not 1/λ = σ²