Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ColorTypes.RGB — Method.

The RGB color associated with this point

LasIO.LasPoint — Type.

Abstract type for ASPRS LAS point data record formats 0 - 3

LasIO.LasPoint0 — Type.

ASPRS LAS point data record format 0

source

LasIO.LasPoint1 — Type.

ASPRS LAS point data record format 1

source

LasIO.LasPoint2 — Type.

ASPRS LAS point data record format 2

source

LasIO.LasPoint3 — Type.

ASPRS LAS point data record format 3

source

LasIO.LasVariableLengthRecord — Type.

A LAS "variable length record" - the generic way to store extra user or organization defined binary metadata in LAS files.

LasIO.LasVariableLengthRecord — Method.

Construct a projection VLR based on an EPSG code

LasIO.PointVector — Type.

Custom PointVector struct for memory mapped LasPoints. Inspiration taken from UnalignedVector.jl and extended it with custom indexing and packing.

ColorTypes.blue — Method.

The blue image channel value associated with this point

ColorTypes.green — Method.

The green image channel value associated with this point

ColorTypes.red — Method.

The red image channel value associated with this point

LasIO.classification — Method.

Classification value as defined in the ASPRS classification table.

LasIO.edge_of_flight_line — Method.

If true, it is the last point before the scanner changes direction.

LasIO.flag_byte — Method.

Flag byte, contains return number, number of returns, scan direction flag and edge of flight line

LasIO.flag_byte — Method.

Flag byte, as represented in the point data, built up from components

LasIO.gps_time — Method.

Convert DateTime to GPS Float64, as represented in the point data, assumes time is recorded in Adjusted Standard GPS Time; see is_standard_gps

LasIO.gps_time — Method.

Time tag value at which the point was aquired, a Float64; see is_standard_gps for what it represents

LasIO.intensity — Method.

Integer representation of the pulse return magnitude.

LasIO.key_point — Method.

If true, this point is considered to be a model key-point.

LasIO.number_of_returns — Method.

The total number of returns for a given pulse.

LasIO.pt_src_id — Method.

This value indicates the file from which this point originated.

LasIO.raw_classification — Method.

Raw classification byte, contains classification, synthetic, key point and withheld

LasIO.raw_classification — Method.

Raw classification byte, as represented in the point data, built up from components

LasIO.return_number — Method.

The pulse return number for a given output pulse, starting at one.

LasIO.scan_angle — Method.

Angle at which the laser point was output, including the roll of the aircraft.

LasIO.scan_direction — Method.

If true, the scanner mirror was traveling from left to right at the time of the output pulse.

LasIO.synthetic — Method.

If true, the point was not created from lidar collection

LasIO.update! — Method.

Update the header bounding box and count based on point data

LasIO.user_data — Method.

This field may be used at the user’s discretion.

LasIO.withheld — Method.

If true, this point should not be included in processing

LasIO.xcoord — Method.

X coordinate (Float64), apply scale and offset according to the header

LasIO.xcoord — Method.

X value (Int32), as represented in the point data, reversing the offset and scale from the header

LasIO.ycoord — Method.

Y coordinate (Float64), apply scale and offset according to the header

LasIO.ycoord — Method.

Y value (Int32), as represented in the point data, reversing the offset and scale from the header

LasIO.zcoord — Method.

Z coordinate (Float64), apply scale and offset according to the header

LasIO.zcoord — Method.

Z value (Int32), as represented in the point data, reversing the offset and scale from the header

Dates.DateTime — Method.

Get the DateTime that the point was collected. Assumes time is recorded in Adjusted Standard GPS Time; see is_standard_gps. Note that DateTime has millisecond precision, any higher precision is lost.

LasIO.GeoKeys — Method.

Create GeoKeys from EPSG code. Assumes CRS is projected and in meters.

LasIO.epsg_code! — Method.

Set the projection in the header, without altering the points

LasIO.epsg_code — Method.

Get the EPSG code of the projection in the header

LasIO.generate_read — Method.

Generate read (unpack) method for structs.

LasIO.generate_write — Method.

Generate write (pack) method for structs.

LasIO.is_srs — Method.

test whether a vlr is a GeoKeyDirectoryTag, GeoDoubleParamsTag or GeoAsciiParamsTag

LasIO.is_standard_gps — Method.

If true, GPS Time is standard GPS Time (satellite GPS Time) minus 1e9. If false, GPS Time is GPS Week Time.

Note that not all software sets this encoding correctly.

LasIO.is_wkt — Method.

Check if the projection information is in WKT format (true) or GeoTIFF (false)