Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

GLM.linkfun — Method.

link of underlying GLM

Lasso.distfun — Method.

distribution of underlying GLM

StatsBase.coef — Method.

coef(path::RegularizationPath; select=:AICc)

Returns a p by nλ coefficient array where p is the number of coefficients (including any intercept) and nλ is the number of path segments, or a selected segment's coefficients.

If model was only initialized but not fit, returns a p vector of zeros. Consistent with StatsBase.coef, if the model has an intercept it is included.

Example:

  m = fit(LassoPath,X,y)
  coef(m; select=:CVmin)

Keywords

select=:all returns a p by nλ matrix of coefficients

select=:AIC selects the AIC minimizing segment

select=:AICc selects the corrected AIC minimizing segment

select=:BIC selects the BIC minimizing segment

select=:CVmin selects the segment that minimizing out-of-sample mean squared error from cross-validation with nCVfolds random folds.

select=:CV1se selects the segment whose average OOS deviance is no more than 1 standard error away from the one minimizing out-of-sample mean squared error from cross-validation with nCVfolds random folds.

nCVfolds=10 number of cross-validation folds

kwargs... are passed to minAICc(::RegularizationPath) or to cross_validate_path(::RegularizationPath)

StatsBase.deviance — Method.

deviance at each segment of the path for the fitted model and data

StatsBase.deviance — Method.

deviance at each segement of the path for (potentially new) data X and y select=:all or :AICc like in coef()

StatsBase.deviance — Method.

deviance at each segment of the path for (potentially new) y and predicted values μ

StatsBase.dof — Method.

dof(path::RegularizationPath)

Approximates the degrees-of-freedom in each segment of the path as the number of non zero coefficients plus a dispersion parameter when appropriate. Note that for GammaLassoPath this may be a crude approximation, as gamlr does this differently.

Base.size — Method.

size(path) returns (p,nλ) where p is the number of coefficients (including any intercept) and nλ is the number of path segments. If model was only initialized but not fit, returns (p,1).

Lasso.computeω! — Method.

Compute coefficient specific weights vector ωj^t based on previous iteration coefficients β times penaltyfactor