Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

RecipesBase.plot — Method.

plot(path::RegularizationPath, args...; <keyword arguments>)

Plots a RegularizationPath fitted with the Lasso package.

The minimum AICc segment is represented by a solid vertical line and the CVmin and CV1se cross-validation selected segments by dashed vertical lines.

By default it shows nonzero coefficients at the AICc in color and the rest grayed out.

LassoPlot uses Plots.jl, so you can set any supported backend before plotting, and add any features of the plot afterwards.

Example:

    using Lasso, LassoPath, Plots
    path = fit(LassoPath, X, y, dist, link)
    plot(path)

Arguments

args... additional arguments passed to Plots.plot()

Keywords

x=:segment one of (:segment, :λ, :logλ)

varnames=nothing specify variable names

select=:AICc Selection criteria in (:AIC, :CVmin, :CV1se) for which coefficients will be shown in color. The rest are grayed out.

showselectors=[:AICc,:CVmin,:CV1se] shown vertical lines

selectedvars=[] Subset of the variables to present, or empty vector for all

kwargs... additional keyword arguments passed along to fit(GammaLassoPath,...)