Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LatticeSites.HilbertSpace — Type.

HilbertSpace{L, S, M}

M dimension Hilbert space of Site type L in with shape S.

LatticeSites.down — Function.

down(site) -> site
down(site_type) -> site

down tag for this label. e.g. 0 for Bit, -0.5 for Half.

LatticeSites.flip — Function.

flip(site)

Defines how to flip this type of site.

LatticeSites.flip! — Function.

flip!(S, I...)

Flips given configuration S at index I....

LatticeSites.randflip! — Function.

randflip!(config) -> (proposed_index, config)
randflip!(::FlipStyle, config) -> (proposed_index, config)

Flip the lattice configuration randomly using given flip style, default flip style is UniformFlip{1}, which choose one site in the configuration uniformly and flip it.

One should always be able to use config[proposed_index] to get the current value of this lattice configuration. Whether one can change the site, depends on whether the configuration is stored in a mutable type.

LatticeSites.up — Function.

up(site) -> site
up(site_type) -> site

up tag for this label. e.g. 1 for Bit, 0.5 for Half.

LatticeSites.FlipStyle — Type.

FlipStyle

Abstract type for flip styles.

LatticeSites.UniformFlip — Type.

UniformFlip{N}

Choose N sites in the configuration uniformly and flip it