Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LearnBase.AbstractBatchView — Type.

abstract AbstractBatchView{TElem, TData} <: DataView{TElem, TData}

Baseclass for all vector-like views of some data structure, that views it as some form or vector of equally sized batches.

see MLDataPattern.BatchView for a concrete example.

LearnBase.AbstractObsView — Type.

abstract AbstractObsView{TElem, TData} <: DataView{TElem, TData}

Baseclass for all vector-like views of some data structure, that views it as some form or vector of observations.

see MLDataPattern.ObsView for a concrete example.

LearnBase.BatchIterator — Type.

abstract BatchIterator{TElem,TData} <: DataIterator{TElem,TData}

Baseclass for all types that iterate over of some data source one batch at a time.

@assert typeof(RandomBatches(X, size=10)) <: BatchIterator

for x in RandomBatches(X, size=10)
    @assert nobs(x) == 10
    # ...
end

see MLDataPattern.RandomBatches

LearnBase.Cost — Type.

Baseclass for any kind of cost. Notable examples for costs are Loss and Penalty.

LearnBase.DataIterator — Type.

abstract DataIterator{TElem,TData}

Baseclass for all types that iterate over a data source in some manner. The total number of observations may or may not be known or defined and in general there is no contract that getobs or nobs has to be supported by the type of data. Furthermore, length should be used to query how many elements the iterator can provide, while nobs may return the underlying true amount of observations available (if known).

see MLDataPattern.RandomObs, MLDataPattern.RandomBatches

LearnBase.DataView — Type.

abstract DataView{TElem, TData} <: AbstractVector{TElem}

Baseclass for all vector-like views of some data structure. This allow for example to see some design matrix as a vector of individual observation-vectors instead of one matrix.

see MLDataPattern.ObsView and MLDataPattern.BatchView for examples.

LearnBase.DiscreteSet — Type.

Set of discrete items

LearnBase.DistanceLoss — Type.

A supervised loss that can be simplified to L(targets, outputs) = L(targets - outputs) is considered distance-based.

LearnBase.IntervalSet — Type.

A continuous range (inclusive) between a lo and a hi

LearnBase.Loss — Type.

Baseclass for all losses. A loss is some (possibly simplified) function L(features, targets, outputs), where outputs are the result of some function f(features).

LearnBase.MarginLoss — Type.

A supervised loss, where the targets are in {-1, 1}, and which can be simplified to L(targets, outputs) = L(targets * outputs) is considered margin-based.

LearnBase.Minimizable — Type.

Baseclass for any prediction model that can be minimized. This means that an object of a subclass contains all the information needed to compute its own current loss.

LearnBase.ObsIterator — Type.

abstract ObsIterator{TElem,TData} <: DataIterator{TElem,TData}

Baseclass for all types that iterate over some data source one observation at a time.

using MLDataPattern
@assert typeof(RandomObs(X)) <: ObsIterator

for x in RandomObs(X)
    # ...
end

see MLDataPattern.RandomObs

LearnBase.SupervisedLoss — Type.

A loss is considered supervised, if all the information needed to compute L(features, targets, outputs) are contained in targets and outputs, and thus allows for the simplification L(targets, outputs).

LearnBase.Transformation — Type.

Anything that takes an input and performs some kind of function to produce an output. For example a linear prediction function.

LearnBase.TupleSet — Type.

Groups several heterogenous sets. Used mainly for proper dispatch.

LearnBase.UnsupervisedLoss — Type.

A loss is considered unsupervised, if all the information needed to compute L(features, targets, outputs) are contained in features and outputs, and thus allows for the simplification L(features, outputs).

LearnBase.grad — Function.

Return the gradient of the learnable parameters w.r.t. some objective

LearnBase.grad! — Function.

Do a backward pass, updating the gradients of learnable parameters and/or inputs

LearnBase.inputdomain — Function.

Returns an AbstractSet representing valid input values

LearnBase.targetdomain — Function.

Returns an AbstractSet representing valid output/target values

LearnBase.transform! — Function.

Do a forward pass, and return the output

LearnBase.ObsDimension — Type.

see ?ObsDim

LearnBase.default_obsdim — Function.

default_obsdim(data)

The specify the default obsdim for a specific type of data. Defaults to ObsDim.Undefined()