Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LearningStrategies.Breaker — Type.

Breaker(f)

Stop learning when f(model, i) returns true.

LearningStrategies.Converged — Type.

Converged(f; tol = 1e-6, every = 1)

Stop learning when norm(f(model) - lastf) ≦ tol. f must return a Vector{Float64}.

LearningStrategies.ConvergedTo — Type.

ConvergedTo(f, goal; tol=1e-6, every=1)

Stop learning when ‖f(model) - goal‖ ≦ tol.

LearningStrategies.IterFunction — Type.

IterFunction(f)
IterFunction(f, b)
IterFunction(b, f)

Call f(model, i) every b iterations at hook call. Default value of b is 1.

LearningStrategies.LearningStrategy — Type.

A LearningStrategy should implement at least one of the following methods:

setup!(strat, model, data)

cleanup!(strat, model)

hook(strat, model, i)

finished(strat, model, data, i)

update!(model, strat, item)

LearningStrategies.MaxIter — Type.

MaxIter(n)

Stop learning after n iterations.

Wrapping MaxIter with Verbose:

julia> learn!(nothing, Verbose(MaxIter(5)), 1:100)
INFO: MaxIter: 1/5
INFO: MaxIter: 2/5
INFO: MaxIter: 3/5
INFO: MaxIter: 4/5
INFO: MaxIter: 5/5
INFO: MaxIter(5) finished

LearningStrategies.MetaStrategy — Type.

MetaStrategy(strats::LearningStrategy...)

A collection of learning strategies in a type-stable way.

LearningStrategies.TimeLimit — Type.

TimeLimit(s)

Stop learning after s seconds.

LearningStrategies.Tracer — Type.

Tracer{T}(::Type{T}, f, b=1)

Store f(model, i) every b iterations.

To extract the data Tracer collected, collect(tracer). Note that this operation will copy.

julia> t = Tracer(Int, (model, i) -> @show(i))
Tracer(1, #5, 0-element Array{Int64,1})

julia> learn!(nothing, strategy(MaxIter(3), t))
i = 1
i = 2
i = 3

julia> collect(t)
3-element Array{Int64,1}:
 1
 2
 3

LearningStrategies.Verbose — Type.

Verbose(s::LearningStrategy)
Verbose(s::LearningStrategy, io::IO)

Allow the LearningStrategy s to print output.

Will automatically print when finished(s, args...) == true.

Other methods should be overloaded to add printout.

For example: update!(model, v::Verbose{MyStrategy}, item) = ...

LearnBase.learn! — Method.

learn!(model, strategy, data) -> model

Learn a model from data using strategy. New models/strategies/data types should overload at least one of the following:

setup!

update!

hook

finished

cleanup!

learn! Implementation:

function learn!(model, s::LearningStrategy, data)
    setup!(s, model[, data])
    for (i, item) in enumerate(data)
        update!(model, s[, i], item)
        hook(s, model[, data], i)
        finished(s, model[, data], i) && break
    end
    cleanup!(s, model)
end

LearningStrategies.strategy — Method.

strategy(s::LearningStrategy...)
strategy(ms::MetaStrategy, s::LearningStrategy...)

Create a MetaStrategy from LearningStrategies or add a LearningStrategy to an existing MetaStrategy.