Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LegacyStrings.utf16 — Function.

utf16(::Union{Ptr{UInt16},Ptr{Int16}} [, length])

Create a string from the address of a NUL-terminated UTF-16 string. A copy is made; the pointer can be safely freed. If length is specified, the string does not have to be NUL-terminated.

LegacyStrings.utf16 — Method.

utf16(s)

Create a UTF-16 string from a byte array, array of UInt16, or any other string type. (Data must be valid UTF-16. Conversions of byte arrays check for a byte-order marker in the first two bytes, and do not include it in the resulting string.)

Note that the resulting UTF16String data is terminated by the NUL codepoint (16-bit zero), which is not treated as a character in the string (so that it is mostly invisible in Julia); this allows the string to be passed directly to external functions requiring NUL-terminated data. This NUL is appended automatically by the utf16(s) conversion function. If you have a UInt16 array A that is already NUL-terminated valid UTF-16 data, then you can instead use UTF16String(A) to construct the string without making a copy of the data and treating the NUL as a terminator rather than as part of the string.

LegacyStrings.utf32 — Function.

utf32(::Union{Ptr{Char},Ptr{UInt32},Ptr{Int32}} [, length])

Create a string from the address of a NUL-terminated UTF-32 string. A copy is made; the pointer can be safely freed. If length is specified, the string does not have to be NUL-terminated.

LegacyStrings.utf32 — Method.

utf32(s)

Create a UTF-32 string from a byte array, array of Char or UInt32, or any other string type. (Conversions of byte arrays check for a byte-order marker in the first four bytes, and do not include it in the resulting string.)

Note that the resulting UTF32String data is terminated by the NUL codepoint (32-bit zero), which is not treated as a character in the string (so that it is mostly invisible in Julia); this allows the string to be passed directly to external functions requiring NUL-terminated data. This NUL is appended automatically by the utf32(s) conversion function. If you have a Char or UInt32 array A that is already NUL-terminated UTF-32 data, then you can instead use UTF32String(A) to construct the string without making a copy of the data and treating the NUL as a terminator rather than as part of the string.

LegacyStrings.checkstring — Function.

Validates and calculates number of characters in a UTF-8,UTF-16 or UTF-32 encoded vector/string

This function checks the bounds of the start and end positions Use unsafe_checkstring to avoid that overhead if the bounds have already been checked

Input Arguments:

dat UTF-8 (Vector{UInt8}), UTF-16 (Vector{UInt16}) or UTF-32 (Vector{UInt32}, AbstractString) encoded string

Optional Input Arguments:

startpos start position (defaults to 1)

endpos end position (defaults to lastindex(dat))

Keyword Arguments:

accept_long_null = true # Modified UTF-8 (\0 represented as b"\xc0\x80")

accept_surrogates = true # CESU-8

accept_long_char = false # Accept arbitrary long encodings

Returns:

(total characters, flags, 4-byte, 3-byte, 2-byte)

Throws:

UnicodeError

LegacyStrings.encode_to_utf16 — Method.

Converts an already validated UTF-32 encoded vector of UInt32 to a UTF16String

Input Arguments:

dat Vector{UInt32} of UTF-32 encoded data

len length of output in 16-bit words

Returns:

UTF16String

LegacyStrings.encode_to_utf8 — Method.

Converts an already validated vector of UInt16 or UInt32 to a UTF8String

Input Arguments:

dat Vector of code units (UInt16 or UInt32), explicit   is not converted

len length of output in bytes

Returns:

UTF8String

LegacyStrings.unsafe_checkstring — Function.

Validates and calculates number of characters in a UTF-8,UTF-16 or UTF-32 encoded vector/string

Warning: this function does not check the bounds of the start or end positions Use checkstring to make sure the bounds are checked

Input Arguments:

dat UTF-8 (Vector{UInt8}), UTF-16 (Vector{UInt16}) or UTF-32 (Vector{UInt32}, AbstractString) encoded string

Optional Input Arguments:

pos start position (defaults to 1)

endpos end position (defaults to lastindex(dat))

Keyword Arguments:

accept_long_null = true # Modified UTF-8 (\0 represented as b"\xc0\x80")

accept_surrogates = true # CESU-8

accept_long_char = false # Accept arbitrary long encodings

Returns:

(total characters, flags, 4-byte, 3-byte, 2-byte)

Throws:

UnicodeError