Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LibSndFile.fmt_to_type — Method.

Take a LibSndFile format code and return a suitable sample type

LibSndFile.inferlen — Method.

inferlen(io)

Try to infer the length of io in bytes

LibSndFile.sf_readf — Function.

Wrappers for the family of sfreadf* functions, which read the given number of frames into the given array. Returns the number of frames read.

LibSndFile.sf_writef — Function.

Wrappers for the family of sfwritef* functions, which write the given number of frames into the given array. Returns the number of frames written.