Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Libtask.TArray — Type.

TArray{T}(dims, ...)

Implementation of data structures that automatically perform copy-on-write after task copying.

If currenttask is an existing key in s, then return `s[currenttask]. Otherwise, returns[currenttask] = s[lasttask]`.

Usage:

TArray(dim)

Example:

ta = TArray(4)              # init
for i in 1:4 ta[i] = i end  # assign
Array(ta)                   # convert to 4-element Array{Int64,1}: [1, 2, 3, 4]

Libtask.tfill — Method.

 tfill(val, dim, ...)

Construct a TArray of a specified value.

tfill(val, dim)

Example:

tz = tfill(9.0, 4)            # construct
Array(tz)                     # convert to 4-element Array{Float64,1}:  [9.0  9.0  9.0  9.0]

Libtask.tzeros — Method.

 tzeros(dims, ...)

Construct a distributed array of zeros. Trailing arguments are the same as those accepted by TArray.

tzeros(dim)

Example:

tz = tzeros(4)              # construct
Array(tz)                   # convert to 4-element Array{Int64,1}: [0, 0, 0, 0]