Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Libz.ZlibDeflateInputStream — Method.

ZlibDeflateInputStream(input[; <keyword arguments>])

Construct a zlib deflate input stream to compress gzip/zlib data.

Arguments

input: a byte vector, IO object, or BufferedInputStream containing data to compress.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=false: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

level::Integer=6: compression level in 1-9.

mem_level::Integer=8: memory to use for compression in 1-9.

strategy=Z_DEFAULT_STRATEGY: compression strategy; see zlib documentation.

Libz.ZlibDeflateOutputStream — Method.

ZlibDeflateOutputStream(output[; <keyword arguments>])

Construct a zlib deflate output stream to compress gzip/zlib data.

Arguments

output: a byte vector, IO object, or BufferedInputStream to which compressed data should be written.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=false: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

level::Integer=6: compression level in 1-9.

mem_level::Integer=8: memory to use for compression in 1-9.

strategy=Z_DEFAULT_STRATEGY: compression strategy; see zlib documentation.

Libz.ZlibInflateInputStream — Method.

ZlibInflateInputStream(input[; <keyword arguments>])

Construct a zlib inflate input stream to decompress gzip/zlib data.

Arguments

input: a byte vector, IO object, or BufferedInputStream containing compressed data to inflate.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=falso: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

reset_on_end::Bool=true: on stream end, try to find the start of another stream.

Libz.ZlibInflateOutputStream — Method.

ZlibInflateOutputStream(output[; <keyword arguments>])

Construct a zlib inflate output stream to decompress gzip/zlib data.

Arguments

output: a byte vector, IO object, or BufferedInputStream to which decompressed data should be written.

bufsize::Integer=8192: input and output buffer size.

raw::Bool=false: if true, data is raw compress data, without zlib metadata

gzip::Bool=true: if true, data is gzip compressed; if false, zlib compressed.

Libz.adler32 — Function.

adler32(data)

Compute the Adler-32 checksum over the data input. data can be BufferedInputStream or Vector{UInt8}.

Libz.crc32 — Function.

crc32(data)

Compute the CRC-32 checksum over the data input. data can be BufferedInputStream or Vector{UInt8}.

Libz.Sink — Type.

The mode type parameter must be either :inflate or :deflate.

Libz.Source — Type.

The mode type parameter must be either :inflate or :deflate.

BufferedStreams.readbytes! — Method.

readbytes!(source, buffer, from, to)

Read bytes from the zlib stream to a buffer. Satisfies the BufferedStreams source interface.

BufferedStreams.writebytes — Method.

writebytes(sink, buffer, n, eof)

Write some bytes from a given buffer. Satisfies the BufferedStreams sink interface.