Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LightQuery.By — Type.

By(it, f)

Marks that it has been pre-sorted by the key f.

LightQuery.LeftJoin — Type.

LeftJoin(left::By, right::By)

For each value in left, look for a value with the same key in right.

julia> using LightQuery

julia> LeftJoin(
            By([1, 2, 5, 6], identity),
            By([1, 3, 4, 6], identity)
       ) |> collect
4-element Array{Pair{Int64,Union{Missing, Int64}},1}:
 1 => 1
 2 => missing
 5 => missing
 6 => 6

LightQuery.Name — Type.

Name(name)

julia> using LightQuery

julia> Name(:a)((a = 1, b = 2.0,))
1

LightQuery.Nameless — Type.

A container for a function and the expression that generated it

LightQuery.Names — Type.

struct Names{T} end

julia> using LightQuery

julia> Names(:a, :b)((a = 1, b = 2.0, c = 3//1))
(a = 1, b = 2.0)

LightQuery.based_on — Method.

based_on(data; assignments...)

Apply the functions in assignments to data, and assign to the corresponding keys.

julia> using LightQuery

julia> based_on((a = 1, b = 2.0), c = @_ _.a + _.b)
(c = 3.0,)

LightQuery.column — Method.

column(it, names::Names)

Lazy find name for each item in it.

julia> using LightQuery

julia> it = [(a = 1, b = 1.0), (a = 2, b = 2.0)];

julia> collect(column(it, :a))
2-element Array{Int64,1}:
 1
 2

LightQuery.gather — Method.

gather(data, new_column, columns...)

Gather all the data in columns into a single new_column.

julia> using LightQuery

julia> gather((a = 1, b = 2.0, c = "c"), :d, :a, :c)
(b = 2.0, d = (a = 1, c = "c"))

LightQuery.group — Method.

group(b::By)

Group consecutive keys in b.

julia> using LightQuery

julia> group(By([1, 3, 2, 4], iseven)) |> collect
2-element Array{Pair{Bool,SubArray{Int64,1,Array{Int64,1},Tuple{UnitRange{Int64}},true}},1}:
 false => [1, 3]
  true => [2, 4]

LightQuery.in_common — Method.

in_common(data1, data2)

Find Names in common.

julia> using LightQuery

julia> in_common((a = 1, b = 2.0), (a = 1, c = "3"))
Names{(:a,)}()

LightQuery.key — Method.

key(pair)

The first item

LightQuery.name — Method.

name(data, names...)

Rename data

julia> using LightQuery

julia> name((a = 1, b = 2.0), :c, :d)
(c = 1, d = 2.0)

LightQuery.named — Method.

named(data)

Convert to a named tuple

julia> using LightQuery

julia> named((a = 1, b = 2.0))
(a = 1, b = 2.0)

julia> struct Triple{T1, T2, T3}
            first::T1
            second::T2
            third::T3
        end;

julia> Base.propertynames(t::Triple) = (:first, :second, :third);

julia> named(Triple(1, 1.0, "a"))
(first = 1, second = 1.0, third = "a")

LightQuery.remove — Method.

remove(data, columns...)

Remove columns.

julia> using LightQuery

julia> remove((a = 1, b = 2.0), :b)
(a = 1,)

LightQuery.rename — Method.

rename(data; renames...)

For type stability, use Name.

julia> using LightQuery

julia> rename((a = 1, b = 2.0), c = Name(:a))
(b = 2.0, c = 1)

LightQuery.rows — Method.

rows(n::NamedTuple)

Iterator over rows of a NamedTuple of columns.

julia> using LightQuery

julia> rows((a = [1, 2], b = [2, 1])) |> first
(a = 1, b = 2)

LightQuery.select — Method.

select(data, columns...)

Select columns

julia> using LightQuery

julia> select((a = 1, b = 2.0), :a)
(a = 1,)

LightQuery.separate — Method.

separate(it, into_names...)

Column-wise storage.

julia> using LightQuery

julia> it = [(a = 1, b = 1.0), (a = 2, b = 2.0)];

julia> result = separate(it, :a, :b);

julia> first(result)
(a = 1, b = 1.0)

LightQuery.spread — Method.

spread(data, column::Name)

Unnest nested data in column

julia> using LightQuery

julia> spread((b = 2.0, d = (a = 1, c = "c")), :d)
(b = 2.0, a = 1, c = "c")

LightQuery.transform — Method.

transform(data; assignments...)

Same as based_on, but merge back in the original.

julia> using LightQuery

julia> transform((a = 1, b = 2.0), c = @_ _.a + _.b)
(a = 1, b = 2.0, c = 3.0)

LightQuery.unname — Method.

unname

Remove names. Explicitly define public propertynames for arbitrary structs.

julia> using LightQuery

julia> unname((a = 1, b = 2.0))
(1, 2.0)

julia> unname((1, 2.0))
(1, 2.0)

julia> struct Triple{T1, T2, T3}
            first::T1
            second::T2
            third::T3
        end;

julia> Base.propertynames(t::Triple) = (:first, :second, :third);

julia> unname(Triple(1, 1.0, "a"))
(1, 1.0, "a")

LightQuery.unzip — Method.

unzip(it, n)

Unzip an iterator it which returns tuples of length n.

julia> using LightQuery

julia> f(x) = (x, x + 1.0);

julia> unzip(over([1], f), 2)
([1], [2.0])

julia> unzip(over([1, missing], f), 2);

julia> unzip(zip([1], [1.0]), 2)
([1], [1.0])

julia> unzip([(1, 1.0)], 2)
([1], [1.0])

LightQuery.value — Method.

value(pair)

The second item