Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LimitedLDLFactorizations.lldl — Method.

lldl(A)

Compute the limited-memory LDLᵀ factorization of A without pivoting.

Arguments

A::SparseMatrixCSC{Tv,Ti}: matrix to factorize (its strict lower triangle and diagonal will be extracted)

Keyword arguments

memory::Int=0: extra amount of memory to allocate for the incomplete factor L. The total memory allocated is nnz(T) + n * memory, where T is the strict lower triangle of A and n is the size of A.

α::Tv=Tv(0): initial value of the shift in case the incomplete LDLᵀ factorization of A is found to not exist. The shift will be gradually increased from this initial value until success.

droptol::Tv=Tv(0): to further sparsify L, all elements with magnitude smaller than droptol are dropped.

LimitedLDLFactorizations.abspermute! — Method.

Permute the elements of keys in place so that

abs(x[keys[i]]) ≤ abs(x[keys[k]])  for i = 1, ..., k
abs(x[keys[k]]) ≤ abs(x[keys[i]])  for i = k, ..., n,

where n is the length of keys. The length of x should be at least n. Only keys is modified.

From the MINPACK2 function dsel2 by Kastak, Lin and Moré.