Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LinQuadOptInterface.CSRMatrix — Type.

CSRMatrix{T}

Matrix given in compressed sparse row (CSR) format.

CSRMatrix is analgous to the structure in Julia's SparseMatrixCSC but with the rows and columns flipped. It contains three vectors:

row_pointers is a vector pointing to the start of each row in columns and coefficients;

columns is a vector of column numbers; and

coefficients is a vector of corresponding nonzero values.

The length of row_pointers is the number of rows in the matrix.

This struct is not a subtype of AbstractSparseMatrix as it is intended to be a collection of the three vectors as they are required by solvers such as Gurobi. It is not intended to be used for general computation.

LinQuadOptInterface.LinearQuadraticModel — Function.

LinearQuadraticModel(M::Type{<:LinQuadOptimizer}, env)

Initializes a model given a model type M and an env that might be a nothing for some solvers.

LinQuadOptInterface.__assert_supported_constraint__ — Method.

__assert_supported_constraint__(model::LinQuadOptimizer, ::Type{F}, ::Type{S})

Throw an UnsupportedConstraint{F, S} error if the model cannot add constraints of type F-in-S.

LinQuadOptInterface.__assert_valid__ — Method.

__assert_valid__(model::LinQuadOptimizer, index::MOI.Index)

Throw an MOI.InvalidIndex error if MOI.is_valid(model, index) == false.

LinQuadOptInterface._matching_sparsity_pattern — Method.

_matching_sparsity_pattern(f1::T, f2::T) where {T <: Union{Linear, VecLin}}

Internal function, not intended for external use.

Determines whether functions f1 and f2 have the same sparsity pattern w.r.t. their constraint rows and columns. Assumes both functions are already in canonical form.

LinQuadOptInterface._replace_with_different_sparsity! — Method.

_replace_with_different_sparsity!(model::LinQuadOptimizer,
    previous::Linear, replacement::Linear, row)

Internal function, not intended for external use.

Change the linear constraint function at index row in model from previous to replacement. This function assumes that previous and replacement may have different sparsity patterns.

This function (and _replace_with_matching_sparsity! above) are necessary because the LQOI API currently only allows linear constraint modification through the change_matrix_coefficient! and change_matrix_coefficients! functions. In order to fully replace a linear constraint, we have to zero out the current matrix coefficients and then set the new matrix coefficients. When the sparsity patterns match, the zeroing-out step can be skipped.

LinQuadOptInterface._replace_with_different_sparsity! — Method.

_replace_with_different_sparsity!(model::LinQuadOptimizer,
    previous::VecLin, replacement::VecLin, row)

Internal function, not intended for external use.

Change the linear constraint function at index row in model from previous to replacement. This function assumes that previous and replacement may have different sparsity patterns.

This function (and _replace_with_matching_sparsity! above) are necessary because the LQOI API currently only allows linear constraint modification through the change_matrix_coefficient! and change_matrix_coefficients! functions. In order to fully replace a linear constraint, we have to zero out the current matrix coefficients and then set the new matrix coefficients. When the sparsity patterns match, the zeroing-out step can be skipped.

LinQuadOptInterface._replace_with_matching_sparsity! — Method.

_replace_with_matching_sparsity!(model::LinQuadOptimizer,
    previous::Linear, replacement::Linear, row)

Internal function, not intended for external use.

Change the linear constraint function at index row in model from previous to replacement. This function assumes that previous and replacement have exactly the same sparsity pattern w.r.t. which variables they include and that both constraint functions are in canonical form (as returned by MOIU.canonical(). Neither assumption is checked within the body of this function.

LinQuadOptInterface._replace_with_matching_sparsity! — Method.

_replace_with_matching_sparsity!(model::LinQuadOptimizer,
    previous::VecLin, replacement::VecLin, row)

Internal function, not intended for external use.

Change the linear constraint function at index row in model from previous to replacement. This function assumes that previous and replacement have exactly the same sparsity pattern w.r.t. which variables they include and that both constraint functions are in canonical form (as returned by MOIU.canonical(). Neither assumption is checked within the body of this function.

LinQuadOptInterface.add_linear_constraint — Method.

add_linear_constraint(model::LinQuadOptimizer, func::Linear, set::S)

Add a constraint of type func-in-set to model.

LinQuadOptInterface.add_linear_constraints! — Function.

add_linear_constraints!(m, A::CSRMatrix{Float64},
    sense::Vector{Cchar}, rhs::Vector{Float64})::Nothing

Adds linear constraints of the form Ax (sense) rhs to the model m.

sense and rhs contain one element for each row in A. The sense is given by backend_type(m, set).

Ranged constraints (set=MOI.Interval) should be added via add_ranged_constraint! instead.

See also: LinQuadOptInterface.CSRMatrix.

LinQuadOptInterface.add_mip_starts! — Function.

add_mip_starts!(model::M, cols::Vector{Int}, x::Vector{Float64})::Nothing

Add a primal start x for the variables in the columns cols of model.

Note that if this method is implemented, solvers of type M must also declare that they support VariablePrimalStarts by overloading the following method:

function MOI.supports(model::M,
                      ::MOI.VariablePrimalStart,
                      ::Type{MOI.VariableIndex})
    return true
end

LinQuadOptInterface.add_quadratic_constraint! — Function.

add_quadratic_constraint!(m, cols::Vector{Int}, coefs::Vector{Float64}, rhs::Float64,
    sense, I::Vector{Int}, J::Vector{Int}, V::Vector{Float64})::Nothing

Add a quadratic constraint a'x + 0.5 x' Q x. See add_linear_constraints! for information of linear component. Arguments (I,J,V) given in triplet form for the Q matrix in 0.5 x' Q x.

LinQuadOptInterface.add_ranged_constraints! — Function.

add_ranged_constraints!(m, A::CSRMatrix{Float64},
    lowerbound::Vector{Float64}, upperbound::Vector{Float64})

Adds linear constraints of the form lowerbound <= Ax <= upperbound to the model m.

This is a special case compared to standard add_linear_constraints! since it is often implemented via multiple API calls.

LinQuadOptInterface.add_sos_constraint! — Function.

add_sos_constraint!(m, cols::Vector{Int}, vals::Vector{Float64}, typ::Symbol)::Nothing

Add the SOS constraint to the model m. typ is either :SOS1 or :SOS2.

LinQuadOptInterface.add_variables! — Function.

add_variables!(m, n::Int)::Nothing

Add n new variables to the model m.

LinQuadOptInterface.backend_type — Method.

backend_type(m, ::MOI.AbstractSet)::CChar

An overloadable type for dispatching the appropriate types to the backends.

For example, GLPK.jl uses 'E' for a'x=b constraints, where as Gurobi.jl uses ==.

Three are special cases:

Val{:Continuous}: for the type of a continuous variable

Val{:Upperbound}: for the upper bound of a variable

Val{:Lowerbound}: for the lower bound of a variable

Defaults

MOI.GreaterThan  - 'G'
MOI.LessThan     - 'L'
MOI.EqualTo      - 'E'

MOI.Zeros        - 'E'
MOI.Nonpositives - 'L'
MOI.Nonnegatives - 'G'

MOI.ZeroOne        - 'B'
MOI.Integer        - 'I'
MOI.Semicontinuous - 'S'
MOI.Semiinteger    - 'N'

MOI.SOS1 - :SOS1  # '1'
MOI.SOS2 - :SOS2  # '2'

Val{:Continuous} - 'C'
Val{:Upperbound} - 'U'
Val{:Lowerbound} - 'L'

LinQuadOptInterface.canonical_reduction — Method.

canonical_reduction(model::LinQuadOptimizer, func::Quad)

Reduce a ScalarQuadraticFunction into five arrays, returned in the following order:

a vector of affine column indices

a vector of affine coefficients

a vector of quadratic row indices

a vector of quadratic column indices

a vector of quadratic coefficients

LinQuadOptInterface.change_both_variable_bounds! — Method.

change_both_variable_bounds!(model::LinQuadOptimizer, columns::Vector{Int},
    lower_bounds::Vector{Float64}, upper_bounds::Vector{Float64})

Set the lower bound of column columns[i] to lower_bounds[i] and the upper bound to upper_bounds[i]. Alternatively, the lower or upper bound can be left blank by passing an array of length 0 instead.

Examples: changebothvariablebounds!(model, [1, 2], [-0.5, 0.0], [1.0, 2.0]) changebothvariablebounds!(model, [1, 2], [-0.5, 0.0], Float64[]) changebothvariable_bounds!(model, [1, 2], Float64[], [1.0, 2.0])

LinQuadOptInterface.change_linear_constraint_sense! — Function.

change_linear_constraint_sense!(m, rows::Vector{Int}, sense::Vector{Symbol})::Nothing

Change the sense of the linear constraints in rows to sense.

sense is the output of backend_type(m, set), where set is the corresponding set for the row rows[i].

Interval constraints require a call to change_range_value!.

LinQuadOptInterface.change_matrix_coefficient! — Function.

change_matrix_coefficient!(m, row, col, coef)

Set the linear coefficient of the variable in column col, constraint row to coef.

LinQuadOptInterface.change_matrix_coefficients! — Method.

change_matrix_coefficients!(m, rows, cols, coefs)

Change multiple linear coefficients simultaneously. Sets the linear coefficient of the variable at column cols[i] in constraint rows[i] to coefs[i] for i in 1:length(rows). Requires that length(rows) == length(cols) == length(coefs).

By default, this method just calls change_matrix_coefficient!(m, rows[i], cols[i], coefs[i]) repeatedly, but some solver interfaces may offer more efficient implementations.

LinQuadOptInterface.change_objective_coefficient! — Function.

changeobjectivecoefficient!(m, col, coef)

Set the linear coefficient of the variable in column col to coef in the objective function.

LinQuadOptInterface.change_objective_sense! — Function.

change_objective_sense!(m, sense::Symbol)::Nothing

Change the optimization sense of the model m to sense. sense must be :min or :max.

LinQuadOptInterface.change_rhs_coefficient! — Function.

changerhscoefficient!(m, row, coef)

Set the rhs of the constraint in row row to coef.

LinQuadOptInterface.change_variable_bounds! — Function.

change_variable_bounds!(m, cols::Vector{Int}, values::Vector{Float64}, senses::Vector)

Change the bounds of the variable. The sense of the upperbound is given by backend_type(m, Val{:Upperbound}()). The sense of the lowerbound is given by backend_type(m, Val{:Lowerbound}())

LinQuadOptInterface.change_variable_types! — Function.

change_variable_types(m, cols::Vector{Int}, types):Nothing

Change the variable types. Type is the output of one of:

backend_type(m, ::ZeroOne), for binary variables;

backend_type(m, ::Integer), for integer variables; and

backend_type(m, Val{:Continuous}()), for continuous variables.

LinQuadOptInterface.delete_from_dictionary_by_value — Method.

delete_from_dictionary_by_value(dict::Dict{K, V}, index::V)

Delete the entries of a dictionary where the value of a key=>value pair is equal to index.

This is used to remove bounds when deleting a variable.

LinQuadOptInterface.delete_linear_constraints! — Function.

delete_linear_constraints!(m, start_row::Int, end_row::Int)::Nothing

Delete the linear constraints start_row, start_row+1, ..., end_row from the model m.

LinQuadOptInterface.delete_quadratic_constraints! — Function.

delete_quadratic_constraints!(m, start_row::Int, end_row::Int)::Nothing

Delete the quadratic constraints start_row, start_row+1, ..., end_row from the model m.

LinQuadOptInterface.delete_sos! — Function.

delete_sos!(m, start_idx::Int, end_idx::Int)::Nothing

Delete the SOS constraints start_idx, start_idx+1, ..., end_idx from the model m.

LinQuadOptInterface.delete_variables! — Function.

delete_variables!(m, start_col::Int, end_col::Int)::Nothing

Delete the columns start_col, start_col+1, ..., end_col from the model m.

LinQuadOptInterface.fakesolve — Method.

fakesolve(instance::MockLinQuadOptimizer)

Set solutions upon solve calls. Data held in MockLinQuadOptimizer stored data is passed to the low-level emulator MockLinQuadModel.

LinQuadOptInterface.functions_to_CSRMatrix — Method.

functions_to_CSRMatrix(model::LinQuadOptimizer, functions::Vector{Linear}, num_non_zeros::Int)

Convert a vector of ScalarAffineFunctions into a CSRMatrix with num_non_zeros non-zero coefficients.

LinQuadOptInterface.get_barrier_iterations — Function.

get_barrier_iterations(m)

Get the number of barrier iterations performed during the most recent optimization of the model m.

LinQuadOptInterface.get_column — Method.

get_column(model::LinQuadOptimizer, variable::MOI.SingleVariable)

Return the column of the variable inside a MOI.SingleVariable function in model.

LinQuadOptInterface.get_column — Method.

get_column(model::LinQuadOptimizer, index::MOI.VariableIndex)

Return the column of the variable index in model.

LinQuadOptInterface.get_constant_objective — Method.

get_constant_objective(m)::Float64

Return the constant (i.e. offset) component of the objective.

LinQuadOptInterface.get_dual_status — Function.

get_dual_status(m)

Get the dual status of the model m.

LinQuadOptInterface.get_farkas_dual! — Function.

get_farkas_dual!(m, x::Vector{Float64})

Get the farkas dual (certificate of primal infeasiblility) for the linear constraints in the model m, and store in x. xmust have one element for each linear constraint.

LinQuadOptInterface.get_farkas_dual_bounds! — Method.

get_farkas_dual_bounds!(model, dest::Vector{Float64})

Get the farkas dual (certificate of primal infeasibility) for the variable bounds in the model model, and store in dest. dest must have one element for each variable.

Since most solvers do not have this feature, this function has a default fallback that does nothing.

LinQuadOptInterface.get_iteration_count — Function.

get_iteration_count(m)

Get the number of simplex iterations performed during the most recent optimization of the model m.

LinQuadOptInterface.get_linear_constraint — Function.

get_linear_constraint(m, row::Int)::Tuple{Vector{Int}, Vector{Float64}}

Get the linear component of the constraint in the 1-indexed row row in the model m. Returns a tuple of (cols, vals).

LinQuadOptInterface.get_linear_dual_solution! — Function.

get_linear_dual_solution!(m, x::Vector{Float64})

Get the dual solution for the linear constraints in the model m, and store in x. xmust have one element for each linear constraint.

LinQuadOptInterface.get_linear_objective! — Function.

get_linear_objective!(m, x::Vector{Float64})

Get the linear coefficients of the objective and store in x.

LinQuadOptInterface.get_linear_primal_solution! — Function.

get_linear_primal_solution!(m, x::Vector{Float64})

Given a set of linear constraints l <= a'x <= b in the model m, get the constraint primal a'x for each constraint, and store in x. x must have one element for each linear constraint.

LinQuadOptInterface.get_node_count — Function.

get_node_count(m)

Get the number of branch-and-cut nodes expolored during the most recent optimization of the model m.

LinQuadOptInterface.get_number_linear_constraints — Function.

get_number_linear_constraints(m)::Int

Get the number of linear constraints in the model m.

LinQuadOptInterface.get_number_quadratic_constraints — Function.

get_number_quadratic_constraints(m)::Int

Get the number of quadratic constraints in the model m.

LinQuadOptInterface.get_number_variables — Function.

get_number_variables(m)::Int

Get the number of variables in the model m.

LinQuadOptInterface.get_objective_bound — Function.

get_objective_bound(m)

Get the objective bound of the model m.

LinQuadOptInterface.get_objective_value — Function.

get_objective_value(m)

Get the objective value of the solved model m.

LinQuadOptInterface.get_objectivesense — Function.

get_objectivesense(m)::MOI.OptimizationSense

Get the optimization sense of the model m.

LinQuadOptInterface.get_primal_status — Function.

get_primal_status(m)

Get the primal status of the model m.

LinQuadOptInterface.get_quadratic_constraint — Function.

get_quadratic_constraint(m, row::Int)::Tuple{Vector{Int}, Vector{Float64}, SparseMatrixCSC{Float64,Int64}}

Get the linear and quadratic components of the constraint in the 1-indexed row row in the model m. Returns a tuple of (lin_cols, lin_vals, Q). Where Q represents the matrix in CSC format.

LinQuadOptInterface.get_quadratic_dual_solution! — Function.

get_quadratic_dual_solution!(m, x::Vector{Float64})

Get the dual solution for the quadratic constraints in the model m, and store in x. xmust have one element for each quadratic constraint.

LinQuadOptInterface.get_quadratic_primal_solution! — Function.

get_quadratic_primal_solution!(m, x::Vector{Float64})

Given a set of quadratic constraints l <= a'x + x'Qx <= b in the model m, get the constraint primal a'x + x'Qx for each constraint, and store in x. x must have one element for each quadratic constraint.

LinQuadOptInterface.get_quadratic_rhs — Function.

get_quadratic_rhs(m, row::Int)::Float64

Get the right hand-side term of quadratic constraint in row row in model m.

LinQuadOptInterface.get_quadratic_terms_objective — Function.

get_quadratic_terms_objective(m)::SparseMatrixCSC{Float64,Int64}

Get quadratic terms of the objective function returned in sparse CSC format.

LinQuadOptInterface.get_range — Function.

get_range(m, row::Int)::Tuple{Float64,Float64}

Get the range which the constraint row belongs to. The output of the function is the tuple lowerbound, upperbound of bounds: lowerbound <= a'x < = upperbound

LinQuadOptInterface.get_relative_mip_gap — Function.

get_relative_mip_gap(m)

Get the relative MIP gap of the solved model m.

LinQuadOptInterface.get_rhs — Function.

get_rhs(m, row::Int)::Float64

Get the right-hand side of the linear constraint in the 1-indexed row row in the model m.

LinQuadOptInterface.get_sos_constraint — Function.

get_sos_constraint(m, idx::Int)::Tuple{Vector{Int}, Vector{Float64}, Symbol}

Get the SOS constraint idx from the model m. Returns the triplet (cols, vals, typ).

LinQuadOptInterface.get_termination_status — Function.

get_termination_status(m)

Get the termination status of the model m.

LinQuadOptInterface.get_unbounded_ray! — Function.

get_unbounded_ray!(m, x::Vector{Float64})

Get the unbounded ray (certificate of dual infeasiblility) for the linear constraints in the model m, and store in x. xmust have one element for each variable.

LinQuadOptInterface.get_variable_dual_solution! — Function.

get_variable_dual_solution!(m, x::Vector{Float64})

Get the dual solution (reduced-costs) for the variables in the model m, and store in x. xmust have one element for each variable.

LinQuadOptInterface.get_variable_lowerbound — Function.

get_variable_lowerbound(m, col::Int)::Float64

Get the lower bound of the variable in 1-indexed column col of the model m.

LinQuadOptInterface.get_variable_primal_solution! — Function.

get_variable_primal_solution!(m, x::Vector{Float64})

Get the primal solution for the variables in the model m, and store in x. xmust have one element for each variable.

LinQuadOptInterface.get_variable_upperbound — Function.

get_variable_upperbound(m, col::Int)::Float64

Get the upper bound of the variable in 1-indexed column col of the model m.

LinQuadOptInterface.has_integer — Method.

has_integer(model::LinQuadOptimizer)::Bool

A helper function to determine if model has any integer components (i.e. binary, integer, special ordered sets, semicontinuous, or semi-integer variables).

LinQuadOptInterface.has_value — Method.

has_value(dict::Dict{K, V}, value::V) where {K, V}

Return true if dict has a key=>value pair with value.

LinQuadOptInterface.is_binding — Method.

is_binding(set, value::Float64)

Return true if value is an extreme point of the set set.

LinQuadOptInterface.make_problem_type_continuous — Method.

make_problem_type_continuous(m)::Nothing

If an explicit call is needed to change the problem type continuous (e.g., CPLEX).

LinQuadOptInterface.make_problem_type_integer — Method.

make_problem_type_integer(m)::Nothing

If an explicit call is needed to change the problem type integer (e.g., CPLEX).

LinQuadOptInterface.modify_ranged_constraints! — Function.

modify_ranged_constraints!(m, rows::Vector{Int}, lowerbound::Vector{Float64}, upperbound::Vector{Float64})

Modify the lower and upperbounds of a ranged constraint in the model m.

This is a special case compared to standard the change_rhs_coefficient! since it is often implemented via multiple API calls.

LinQuadOptInterface.print_low_level_solution — Method.

Solution builder to create solutions for the mock solver. Should be called inside MOI.optimize!(model::LinQuadOptimizer) right before changing signs. This means that these solutions replicate the low-level solver output. The idea is to run the test with a solver like Xpress, Gurobi, GLPK or CPLEX to obtain the detailed solution.

LinQuadOptInterface.reduce_duplicates! — Method.

reduce_duplicates!(rows::Vector{T}, cols::Vector{T}, vals::Vector{S})

Given a matrix specified by row indices in rows, column indices in cols and coefficients in vals, return new rows, cols, and vals vectors with duplicate elements summed and any coefficients in the lower triangle moved to the upper triangle.

This function swaps element i in rows and cols if rows[i]>cols[i].

Examples

julia> reduce_duplicates!(
    [1,   2, 2, 2],  # rows
    [1,   1, 2, 2],  # cols
    [1, 0.5, 1, 1]   # vals
    )
([1, 1, 2], [1, 2, 2], [1.0, 0.5, 2.0])

LinQuadOptInterface.set_constant_objective! — Method.

set_constant_objective!(m, value)::Nothing

Set the constant (i.e. offset) component of the objective function to the given value.

Solver interfaces that overload this behavior (e.g. to pass that constant objective to the solver itself) must also overload get_constant_objective(m).

LinQuadOptInterface.set_linear_objective! — Function.

set_linear_objective!(m, cols::Vector{Int}, coefs::Vector{Float64})::Nothing

Set the linear component of the objective.

LinQuadOptInterface.set_quadratic_objective! — Function.

set_quadratic_objective!(m, I::Vector{Int}, J::Vector{Int}, V::Vector{Float64})::Nothing

Set the quadratic component of the objective. Arguments given in triplet form for the Q matrix in 0.5 x' Q x.

LinQuadOptInterface.shift_references_after_delete_affine! — Method.

shift_references_after_delete_affine!(m, row)

This function updates all of the references in m after we have deleted row row in the affine constraint matrix.

LinQuadOptInterface.shift_references_after_delete_quadratic! — Method.

shift_references_after_delete_quadratic!(m, row)

This function updates all of the references in m after we have deleted row row in the quadratic constraint matrix.

LinQuadOptInterface.solve_linear_problem! — Function.

solve_linear_problem!(m)::Nothing

Solve a linear program m.

LinQuadOptInterface.solve_mip_problem! — Function.

solve_mip_problem!(m)::Nothing

Solve a mixed-integer model m.

LinQuadOptInterface.solve_quadratic_problem! — Function.

solve_quadratic_problem!(m)::Nothing

Solve a model m with quadratic components.

LinQuadOptInterface.supported_constraints — Function.

supported_constraints(m)::Vector{
    Tuple{MOI.AbstractFunction, MOI.AbstractSet}
}

Get a list of supported constraint types in the model m.

For example, [(LQOI.Linear, LQOI.EQ)]

LinQuadOptInterface.supported_objectives — Function.

supported_objectives(m)::Vector{MOI.AbstractScalarFunction}

Get a list of supported objective types in the model m.

For example, [LQOI.Linear, LQOI.Quad]

LinQuadOptInterface.unload — Function.

unload(from::Vector, to, warn = true)

Helper function to remove the first element of a vector and return it. If the vector is empty data in default is returned instead. Used in fakesolve.

LinQuadOptInterface.unsafe_set! — Method.

unsafe_set!(m, ::MOI.ObjectiveFunction{F}, objective::Linear) where F

Sets a linear objective function without cannonicalizing objective.