Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LinearTimeVaryingModelsBase.AbstractCost — Type.

Cost interface, implement the following functions

see also AbstractModel, ModelAndCost

function calculate_cost(::Type{AbstractCost}, x::AbstractVector, u)::Number

function calculate_cost(::Type{AbstractCost}, x::AbstractMatrix, u)::AbstractVector

function calculate_final_cost(::Type{AbstractCost}, x::AbstractVector)::Number

function dc(::Type{AbstractCost}, x, u)
    return cx,cu,cxx,cuu,cxu
end

LinearTimeVaryingModelsBase.AbstractModel — Type.

Model interface, implement the following functions

see also AbstractCost, ModelAndCost

fit_model(::Type{AbstractModel}, batch::Batch)::AbstractModel

predict(model::AbstractModel, x, u)

function df(model::AbstractModel, x, u, I::UnitRange)
    return fx,fu,fxx,fxu,fuu
end

LinearTimeVaryingModelsBase.ModelAndCost — Type.

Define types that implement the interfaces AbstractModel and AbstractCost.

Create object modelcost = ModelAndCost(model, cost)

Run macro @definemodelcostfunctions(modelcost). This macro defines the following functions

f(x, u, i)  = f(modelcost, x, u, i)
fT(x)       = fT(modelcost, x)
df(x, u, I) = df(modelcost, x, u, I)

see also AbstractModel, AbstractCost

LinearTimeVaryingModelsBase.df — Method.

fx,fu,fxx,fxu,fuu = df(model::AbstractModel, x, u)

Get the linearized dynamics at x,u

LinearTimeVaryingModelsBase.df — Method.

fx,fu,fxx,fxu,fuu,cx,cu,cxx,cxu,cuu = df(modelcost::ModelAndCost, x, u)

Get the linearized dynamics and cost at x,u

LinearTimeVaryingModelsBase.predict — Method.

xnew = predict(model::AbstractModel, t::Trajectory [, i])
xnew = predict(model::AbstractModel, x, u [, i])

Predict the next state given the current state and action

LinearTimeVaryingModelsBase.simulate — Method.

xnew = simulate(model::AbstractModel, t::Trajectory)
xnew = simulate(model::AbstractModel, x0, u)

Simulate system forward in time given the initial state and actions

LinearTimeVaryingModelsBase.fit_model! — Method.

fit_model!(model::AbstractModel, x,u)::AbstractModel

Refits a model to new data

LinearTimeVaryingModelsBase.fit_model — Method.

model = fit_model(::Type{AbstractModel}, x,u)::AbstractModel

Fits a model to data