Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LocalCoverage.clean_coverage — Method.

clean_coverage(pkg)

Clean up after generate_coverage.

LocalCoverage.generate_coverage — Method.

generate_coverage(pkg; genhtml)

Generate a coverage report for package pkg.

When genhtml, the corresponding external command will be called to generate a HTML report. This can be found in eg the package lcov on Debian/Ubuntu.

*.cov files are near the source files as generated by Julia, everything else is placed in Pkg.dir(pkg, "coverage"). The summary is in Pkg.dir(pkg, "coverage", "lcov.info").

Use clean_coverage for cleaning.

LocalCoverage.open_coverage — Method.

open_coverage(pkg)

Open the HTML coverage results in a browser for pkg if they exist.

See generate_coverage.