Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LocalDiskObjectStores._create! — Method.

Create bucket. If successful return nothing, else return an error message as a String.

LocalDiskObjectStores._create! — Method.

Create object. If successful return nothing, else return an error message as a String.

LocalDiskObjectStores._delete! — Method.

Delete bucket. If successful return nothing, else return an error message as a String.

LocalDiskObjectStores._delete! — Method.

Delete object. If successful return nothing, else return an error message as a String.

LocalDiskObjectStores._read — Method.

Read bucket. If successful return (true, value), else return (false, errormessage::String).

LocalDiskObjectStores._read — Method.

Read object. If successful return (true, value), else return (false, errormessage::String).