Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LocalFunctionApproximation.compute_value — Function.

compute_value(lfa::LocalFunctionApproximator, v::AbstractVector)

Return the value of the function at some query point v, based on the local function approximator

compute_value(lfa::LocalFunctionApproximator, v_list::AbstractVector{V}) where V <: AbstractVector{Float64}

Return the value of the function for a list of query points, based on the local function approximator

LocalFunctionApproximation.get_all_interpolating_points — Function.

get_all_interpolating_points(lfa::LocalFunctionApproximator)

Return the vector of points (in a specific order) that are used to interpolate

LocalFunctionApproximation.get_all_interpolating_values — Function.

get_all_interpolating_values(lfa::LocalFunctionApproximator)

Return the vector of all interpolating values (in the same order as the interpolating points)

LocalFunctionApproximation.get_interpolating_nbrs_idxs_wts — Function.

get_interpolating_nbrs_idxs_wts(lfa::LocalFunctionApproximator, v::AbstractVector)

Return a tuple of (indices, weights) for the interpolants for a specific query v

LocalFunctionApproximation.n_interpolating_points — Function.

n_interpolating_points(lfa::LocalFunctionApproximator)

Return the number of interpolanting points that the approximator is using

LocalFunctionApproximation.set_all_interpolating_values — Function.

set_all_interpolating_values(lfa::LocalFunctionApproximator, vals::AbstractVector)

Set the values of all interpolating points to the input vector