Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Loess.loess — Method.

loess(xs, ys, normalize=true, span=0.75, degreee=2)

Fit a loess model.

Args: xs: A n by m matrix with n observations from m independent predictors ys: A length n response vector. normalize: Normalize the scale of each predicitor. (default true when m > 1) span: The degree of smoothing, typically in [0,1]. Smaller values result in smaller local context in fitting. degree: Polynomial degree.

Returns: A fit LoessModel.

Loess.KDTree — Method.

KDTree(xs, leaf_size_factor, leaf_diameter factor)

Construct a kd-tree

Args:

xs: an n x m matrix containing n, m-dimensional observations.

leaf_size_factor: Stop spliting if a node contains fewer than leaf_size_factor * n elements.

leaf_diameter_factor: Stop spliting if a node's bounding hypercube has diameter less that leaf_diameter_factor times the diameter of the root node's bounding hypercube.

Returns: A KDTree object

Loess.bounds_verts — Method.

bounds_verts(bounds)

Given a bounding hypecube bounds, return its verticies

Loess.build_kdtree — Method.

build_kdtree(xs, perm, bounds, leaf_size_cutoff, leaf_diameter_cutoff, verts)

Recursively build a kd-tree

Args:

xs: Data being orginized.

perm: Permutation of the data, used to avoid directly sorting or modifying xs.

bounds: Bounding hypercube of the node.

leaf_size_cutoff: stop spliting on nodes with more than this many values.

leaf_diameter_cutoff: stop splitting on nodes with less than this diameter.

verts: current set of vertexes

Modifies: perm, verts

Returns: Either a KDLeafNode or a KDInternalNode

Loess.diameter — Method.

diameter(bounds)

Compute the diamater of a hypercube (i.e. the maximum distance between any 2 points in the hypercube) defined by bounds.

Args:

bounds: A 2 by n matrix where bounds[1,i] gives the lower bound in the ith dimension and bounds[2,j] the upper bound.

Returns: Computed diameter

Loess.evalpoly — Method.

evalpoly(xs,bs)

Evaluate a multivariate polynomial with coefficients bs at xs. bs should be of length 1+length(xs)*d where d is the degree of the polynomial.

bs[1] + xs[1]*bs[2] + xs[1]^2*bs[3] + ... + xs[end]^d*bs[end]

Loess.tnormalize! — Method.

tnormalize!(x,q)

Default normalization procedure for predictors.

This simply normalizes by the mean of everything between the 10th an 90th percentiles.

Args: xs: a matrix of predictors q: cut the ends of at quantiles q and 1-q

Modifies: xs

Loess.traverse — Method.

traverse(kdtree, x)

Traverse the tree kdtree to the bottom and return the verticies of the bounding hypercube of the leaf node containing the point x.

Loess.tricubic — Method.

tricubic(u)

Tricubic weight function.

Args: u: Distance between 0 and 1

Returns: A weighting of the distance u