Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

LoggingExtras.DemuxLogger — Method.

DemuxLogger(loggers...; include_current=true)

Sets the globallogger to demux, so that messages are sent to all the loggers. If `includecurrent_global` is true, then messages are also sent to the global logger (or rather to what was the global logger when this was constructed). Normally this would be the ConsoleLogger in the REPL etc.

LoggingExtras.FilteredLogger — Type.

FilteredLogger(filter, logger)

Wraps logger in a filter. While loggers intrinsictally have in built filtering mechanisms. Wrapping it in a FilterLogger allows for extract control, at the cost of a bit of overhead.

The filter should be a function that returns a boolean. true if the message should be logged and false if not. It should take as inputs: (level, message, _module, group, id, file, line; kwargs...)

These argument come from the logging macro (@info,@warn` etc).

level::LogLevel Warn, Info, etc,

message::String the message to be logged

_module::Module can be used to specify a different originating module from the source location of the message.

group::Symbol can be used to override the message group (this is normally derived from the base name of the source file).

id::Symbol can be used to override the automatically generated unique message identifier. This is useful if you need to very closely associate messages generated on different source lines.

file::String and line::Int can be used to override the apparent source location of a log message.

kwargs... any keyword or positional arguments to the logging macro.