Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MCMCDiagnostics.effective_sample_size — Function.

effective_sample_size(x, v = var(x))

Effective sample size of vector x.

Estimated from autocorrelations. See Gelman et al (2013), section 11.4.

When the variance v is supplied, it saves some calculation time.

MCMCDiagnostics.ess_factor_estimate — Function.

ess_factor_estimate(x, v = var(x))

Estimate for effective sample size factor.

Return τ, K where τ is estimated effective sample size / sample size, and K is the last lag used for autocorrelation estimation.

Notes

See Gelman et al (2013), section 11.4.

τ is capped at 1, this is relevant when the sample has large negative autocorrelation (happens with HMC/NUTS).

Some implementations (eg Stan) use FFT for autocorrelations, which yields the whole spectrum. In practice, a <50-100 lags are usually sufficient for reasonable samplers, so the “naive” version may be more efficient.

MCMCDiagnostics.potential_scale_reduction — Method.

potential_scale_reduction(chains...)

Potential scale reduction factor (for possibly ragged chains).

Also known as R̂. Always ≥ 1 by construction, but values much larger than 1 (say 1.05) indicate poor mixing.

Uses formula from Stan Development Team (2017), section 28.3.

MCMCDiagnostics.autocorrelation — Function.

autocorrelation(x, k, v = var(x))

Estimate of lag-k autocorrelation of x from a variogram. v is the variance of x, used when supplied.

See Gelman et al (2013), section 11.4.