Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MIDI.MIDIEvent — Type.

MIDIEvent <: TrackEvent

See TrackEvent.

MIDI.MIDIFile — Type.

MIDIFile <: Any

Type representing a file of MIDI data.

Fields

format::UInt16 : The format of the file. Can be 0, 1 or 2.

tpq::Int16 : The time division of the track, ticks-per-quarter-note.

tracks::Array{MIDITrack, 1} : The array of contained tracks.

MIDI.MIDITrack — Type.

MIDITrack <: Any

MIDITrack is simply a container for TrackEvents, since its only field is events::Vector{TrackEvent}.

Track chunks begin with four bytes spelling out "MTrk", followed by the length (in bytes) of the track (see readvariablelength), followed by a sequence of events.

MIDI.MetaEvent — Type.

MetaEvent <: TrackEvent

See TrackEvent.

MIDI.Note — Type.

Note <: AbstractNote

Data structure describing a "music note".

Fields:

pitch::UInt8 : Pitch, starting from C0 = 0, adding one per semitone (middle-C is 60).

velocity::UInt8 : Dynamic intensity. Cannot be higher than 127 (0x7F).

position::UInt : Position in absolute time (since beginning of track), in ticks.

duration::UInt : Duration in ticks.

channel::UInt8 = 0 : Channel of the track that the note is played on. Cannot be higher than 127 (0x7F).

If the channel of the note is 0 (default) it is not printed with show.

MIDI.Notes — Type.

Notes{N<:AbstractNote}

Data structure describing a collection of "music notes", bundled with a ticks per quarter note measure.

Fields:

notes::Vector{N}

tpq::Int16 : Ticks per quarter note. Defines the fundamental unit of measurement of a note's position and duration, as well as the length of one quarter note. Takes values from 1 to 960.

Notes is iterated and accessed as if iterating or accessing its field notes.

MIDI.SysexEvent — Type.

SysexEvent <: TrackEvent

See TrackEvent.

MIDI.TrackEvent — Type.

TrackEvent <: Any

Abstract supertype for all MIDI events.

All track events begin with a variable length time value (see readvariablelength) and have a field named dT which contains it. This number notes after how many ticks since the last event does the current even takes place.

MIDIEvents then resume with a MIDI channel message defined in constants.jl. They're followed by 1 or 2 bytes, depending on the channel message (see MIDI.EVENTTYPETOLENGTH). If no valid channel message is identified, the previous seen channel message is used. After that the MIDI command is encoded.

MetaEvents and SysexEvents both resume with a specific byte (see constants.jl).

MIDI.BPM — Method.

BPM(midi)

Return the BPM where the given MIDIFile was exported at.

MIDI.addevent! — Method.

addevent!(track::MIDITrack, time::Int, event::TrackEvent)

Add an event to the track at given time. The time is in absolute time, not relative.

If you want to add multiple events in one go, you should use the addevents! function instead.

MIDI.addevents! — Method.

addevents!(track::MIDITrack, times, events)

Add given events to given track at given times, internally doing all translations from absolute time to relative time.

Using this function is more efficient than a loop over single addevent! calls.

MIDI.addnote! — Method.

addnote!(track::MIDITrack, note::AbstractNote)

Add given note to given track, internally doing the translation from absolute time to relative time.

MIDI.addnotes! — Method.

addnotes!(track::MIDITrack, notes)

Add given notes to given track, internally doing all translations from absolute time to relative time.

MIDI.addtrackname! — Method.

addtrackname!(track::MIDI.MIDITrack, name::String)

Add a name to the given track by attaching the "track name" MetaEvent to the start of the track.

MIDI.findtextevents — Method.

findtextevents(eventtype, track)

Find all "text" events specifield by eventtype in the track. The eventtype can be :text, :lyric, :marker, which will find the appropriate MetaEvents.

For convenience, this function does not return the events themselves. Instead, it returns three vectors: the first is the strings of the events, the second is the indices of the events in the track and the third is the absolute position of the events (since start of track).

Notice - common music score editors like e.g. MuseScore, GuitarPro, etc., do not export the lyrics and text information when exporting midi files.

MIDI.getnotes — Function.

getnotes(midi::MIDIFile, trackno = 2)

Find all NOTEON and NOTEOFF midi events in the trackno track of a midi, that correspond to the same note value (pitch) and convert them into the Note datatype. There are special cases where NOTEOFF is actually encoded as NOTEON with 0 velocity. getnotes takes care of this.

Notice that the first track of a midi doesn't have any notes.

getnotes(track::MIDITrack, tpq = 960)

Find the notes from track directly, passing also the ticks per quarter note.

Returns: Notes{Note}, setting the ticks per quarter note as tpq. You can find the originally exported ticks per quarter note from the original MIDIFile through midi.tpq.

MIDI.ms_per_tick — Function.

ms_per_tick(tpq, bpm)
ms_per_tick(midi::MIDIFile)

Return how many miliseconds is one tick, based on the beats per minute bpm and ticks per quarter note tpq.

MIDI.name_to_pitch — Method.

name_to_pitch(p::String) -> Int

Return the pitch value of the given note name p, which can be of the form capital_letter*sharp*octave. E.g. name_to_pitch(D#6) gives 75.

capital_letter : from "A" to "G".

sharp : one of "#" "♯" or "".

octave : any integer (as a string), the octave number (an octave is 12 pitches). If not given it is assumed "5".

Reminder: middle C has pitch 60 and is displayed as C5.

MIDI.pitch_to_name — Method.

pitch_to_name(pitch) -> string

Return the name of the pitch, e.g. F5, A♯3 etc. in modern notation given the value in integer.

Reminder: middle C has pitch 60 and is displayed as C5.

MIDI.readMIDIFile — Method.

readMIDIFile(filename::AbstractString)

Read a file into a MIDIFile data type.

MIDI.readvariablelength — Method.

readvariablelength(f::IO)

Variable length numbers in MIDI files are represented as a sequence of bytes. If the first bit is 0, we're looking at the last byte in the sequence. The remaining 7 bits indicate the number.

MIDI.testmidi — Method.

testmidi()

Return the path to a test MIDI file.

MIDI.textevent — Method.

textevent(eventtype, text)

Create an event using the string text. The eventtype can be :text, :lyric, :marker, which will create the appropriate type of MetaEvent.

The returned event can be added to a MIDITrack via either addevent! or addevents! for multiple events.

Notice - Cubase can read the marker events and MuseScore can read the lyrics events. We haven't seen any editor that can read the text events, so far.

MIDI.trackname — Method.

trackname(track::MIDI.MIDITrack)

Return the name of the given track as a string, by finding the "track name" MetaEvent.

MIDI.writeMIDIFile — Method.

writeMIDIFile(filename::AbstractString, data::MIDIFile)

Write a MIDIFile as a ".mid" file to the given filename.

writeMIDIFile(filename::AbstractString, notes::Notes)

Create a MIDIFile directly from notes, using format 1.

MIDI.writevariablelength — Method.

writevariablelength(f::IO, number::Int)

Write on f the given number, firstly converting it to the "variable length" format. See the documentation for more.

MIDI.addevent_hint! — Method.

addevent_hint!(track::MIDITrack, time::Int, event::TrackEvent,
                eventindex::UInt, eventtime::UInt)

Add an event to the track at given time. The time is in absolute time, not relative. eventindex and eventtime have to be the index and the absolute time of a known event in track which lays BEFORE the position where event shall be added. This shortens the search for the correct position for event by skipping all TrackEvents before the specified one.

Returns the index and absolute time of the added event.

MIDI.get_abs_pos — Method.

get_abs_pos(track::MIDITrack, idxs)

Return the absolute positions (since track start) of the events given by the indices idxs of track.events.

MIDI.programchange — Method.

programchange(track::MIDITrack, time::Integer, channel::UInt8, program::UInt8)

Change the program (instrument) on the given channel. Time is absolute, not relative to the last event.

The program must be specified in the range 1-128, not in 0-127!