Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MMTF.decodedata — Method.

Decodes the MsgPack unpacked dict of the MMTF file.

MMTF.encodedata — Method.

Encodes the MMTF Dict.

MMTF.fetchmmtf — Method.

Fetches a file from RCSB server and decodes it into a MMTFDict. pdbid: ID of the PDB file to be fetched and decoded.

MMTF.parsemmtf — Method.

Parses the MMTF file into a MMTFDict. filepath: Path of the input file.

MMTF.writemmtf — Method.

Writes the MMTFDict back to a file. MMTFDict: Dict containing the decoded MMTF data to be written to file. filepath: Path to write the file

MMTF.addheader — Method.

Add the header to the appropriate array

MMTF.bytes — Method.

Convert an integer array into a byte arrays.

MMTF.chars — Method.

Convert integers to chars.

MMTF.decodearray — Method.

Parse the header of an input byte array and then decode using the input array, the codec and the appropirate parameter. input_array: the array to be decoded

MMTF.decodechainlist — Method.

Convert a list of bytes to a list of strings. Each string is of length mmtf.CHAIN_LEN

MMTF.deltadecode — Method.

A function to delta decode an int array. in_array: the input array of integers

MMTF.deltaencode — Method.

A function to delta decode an int array. in_array: the input array to be delta encoded

MMTF.encodearray — Method.

Encode the array using the method and then add the header to this array. input_array: the array to be encoded codec: the integer index of the codec to use param: the integer parameter to use in the function

MMTF.encodechainlist — Method.

Convert a list of strings to a list of byte arrays.

MMTF.floats — Method.

Convert integers to floats by division. in_ints: the integer array divider: the divider

MMTF.ints — Method.

Convert an array of chars to an array of ints.

MMTF.ints — Method.

Convert floating points to integers using a multiplier. in_floats: the input floats multiplier: the multiplier to be used for conversion. Corresponds to the precisison.

MMTF.ints — Method.

Convert a byte array into an integer array. The number of bytes forming an integer is defined by num in_bytes: the input bytes num: the number of bytes per int

MMTF.parseheader — Method.

Parse the header and return it along with the input array minus the header. Returns the codec, the length of the decoded array, the parameter and the remainder of the array

MMTF.recursiveindexdecode — Method.

Unpack an array of integers using recursive indexing.

MMTF.recursiveindexencode — Method.

Pack an integer array using recursive indexing.

MMTF.runlengthdecode — Method.

A function to run length decode an int array. in_array: the input array of integers

MMTF.runlengthencode — Method.

A function to run length decode an int array. in_array: the inptut array of integers