Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MacroTools.inexpr — Method.

inexpr(expr, x)

Simple expression match; will return true if the expression x can be found inside expr.

inexpr(:(2+2), 2) == true

MacroTools.isdef — Method.

Test for function definition expressions.

MacroTools.isexpr — Method.

isexpr(x, ts...)

Convenient way to test the type of a Julia expression. Expression heads and types are supported, so for example you can call

isexpr(expr, String, :string)

to pick up on all string-like expressions.

MacroTools.namify — Method.

An easy way to get pull the (function/type) name out of expressions like foo{T} or Bar{T} <: Vector{T}.

MacroTools.prettify — Method.

prettify(ex)

Makes generated code generaly nicer to look at.

MacroTools.rmlines — Method.

rmlines(x)

Remove the line nodes from a block or array of expressions.

Compare quote end vs rmlines(quote end)

MacroTools.splitarg — Method.

splitarg(arg)

Match function arguments (whether from a definition or a function call) such as x::Int=2 and return (arg_name, arg_type, is_splat, default). arg_name and default are nothing when they are absent. For example:

> map(splitarg, (:(f(a=2, x::Int=nothing, y, args...))).args[2:end])
4-element Array{Tuple{Symbol,Symbol,Bool,Any},1}:
 (:a, :Any, false, 2)
 (:x, :Int, false, :nothing)
 (:y, :Any, false, nothing)
 (:args, :Any, true, nothing)

MacroTools.splitdef — Method.

splitdef(fdef)

Match any function definition

function name{params}(args; kwargs)::rtype where {whereparams}
   body
end

and return Dict(:name=>..., :args=>..., etc.). The definition can be rebuilt by calling MacroTools.combinedef(dict), or explicitly with

rtype = get(dict, :rtype, :Any)
all_params = [get(dict, :params, [])..., get(dict, :whereparams, [])...]
:(function $(dict[:name]){$(all_params...)}($(dict[:args]...);
                                            $(dict[:kwargs]...))::$rtype
      $(dict[:body])
  end)

MacroTools.unblock — Method.

unblock(expr)

Remove outer begin blocks from an expression, if the block is redundant (i.e. contains only a single expression).

MacroTools.alias_gensyms — Method.

alias_gensyms(expr)

Replaces gensyms with animal names This makes gensym'd code far easier to follow.

MacroTools.assoc! — Method.

assoc!(d, k, v)

is the same as d[k] = v but returns d rather than v.

MacroTools.combinearg — Method.

combinearg(arg_name, arg_type, is_splat, default)

combinearg is the inverse of splitarg.

MacroTools.combinedef — Method.

combinedef(dict::Dict)

combinedef is the inverse of splitdef. It takes a splitdef-like Dict and returns a function definition.

MacroTools.gatherwheres — Method.

gatherwheres(:(f(x::T, y::U) where T where U)) => (:(f(x::T, y::U)), (:U, :T))

MacroTools.gensym_ids — Method.

gensym_ids(expr)

Replaces gensyms with unique ids (deterministically)

MacroTools.postwalk — Method.

postwalk(f, expr)

Applies f to each node in the given expression tree, returning the result. f sees expressions after they have been transformed by the walk. See also prewalk.

MacroTools.prewalk — Method.

prewalk(f, expr)

Applies f to each node in the given expression tree, returning the result. f sees expressions before they have been transformed by the walk, and the walk will be applied to whatever f returns.

This makes prewalk somewhat prone to infinite loops; you probably want to try postwalk first.