Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MappedArrays.mappedarray — Method.

M = mappedarray(f, A)
M = mappedarray(f, A, B, C...)

Create a view M of the array A that applies f to every element of A; M == map(f, A), with the difference that no storage is allocated for M. The view is read-only (you can get values but not set them).

When multiple input arrays are supplied, M[i] = f(A[i], B[i], C[i]...).

MappedArrays.mappedarray — Method.

M = mappedarray(f, finv, A)
M = mappedarray(f, finv, A, B, C...)

creates a view of the array A that applies f to every element of A. The inverse function, finv, allows one to also set values of the view and, correspondingly, the values in A.

When multiple input arrays are supplied, M[i] = f(A[i], B[i], C[i]...).

MappedArrays.of_eltype — Method.

M = of_eltype(T, A)
M = of_eltype(val::T, A)

creates a view of A that lazily-converts the element type to T.