Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MarriageMarkets.SearchMatch — Type.

SearchMatch(λ, δ, r, σ, θ_m, θ_f, γ_m, γ_f, ψ_m, ψ_f, ℓ_m, ℓ_f, h; β=0.5, verbose=false, step=0.2)

Construct a Shimer & Smith (2000) marriage market model and solve for the equilibrium. When match-specific shocks are included, the divorce process is endogenized as in Goussé (2014).

The equilibrium is the solution to a nested fixed-point mapping. The inner fixed-point solves for a matching equilibrium: consistent strategies given fixed singles densities. The outer fixed-point solves for a market equilibrium: singles densities that are consistent with strategies.

Model selection depends which arguments are provided:

Match-specific additive shocks $z ~ N(0, σ)$ when $σ > 0$

* Note: some randomness is required for the fixed point iteration to converge with sex asymmetry

Closed system or inflow/outflow:

* `ℓ_m, ℓ_f` exogenous: population circulates between singlehood and marriage, no birth/death
* Death rates `ψ_m, ψ_f`, inflows `γ_m, γ_f`: population distributions `ℓ_m, ℓ_f` endogenous

MarriageMarkets.SearchMatch — Method.

Outer constructor that constructs the production array from the types and production function.

MarriageMarkets.StaticMatch — Type.

StaticMatch(mtypes, ftypes, mdist, fdist, surplus)

Construct a Choo & Siow (2006) marriage market model and solve for the equilibrium.

MarriageMarkets.StaticMatch — Method.

Outer constructor that takes the production function to build the surplus array

MarriageMarkets.StaticMatch — Method.

Outer constructor for one dimensional females case

MarriageMarkets.StaticMatch — Method.

Outer constructor for one dimensional males case

MarriageMarkets.StaticMatch — Method.

Outer constructor for one dimensional case

MarriageMarkets.SearchClosed — Method.

SearchClosed(λ, δ, r, σ, θ_m, θ_f, ℓ_m, ℓ_f, g; β=0.5, verbose=false, step=0.2)

Constructs marriage market equilibrium of closed-system model with match-specific productivity shocks and production function $g(x,y)$.

MarriageMarkets.SearchClosed — Method.

SearchClosed(λ, δ, r, σ, θ_m, θ_f, ℓ_m, ℓ_f, g; β=0.5, verbose=false, step=0.2)

Method for one-dimensional types.

MarriageMarkets.SearchInflow — Method.

SearchInflow(λ, δ, r, σ, θ_m, θ_f, γ_m, γ_f, ψ_m, ψ_f, g; β=0.5, verbose=false, step=0.2)

Constructs marriage market equilibrium of inflow and death model with match-specific productivity shocks and production function $g(x,y)$.

MarriageMarkets.SearchInflow — Method.

SearchInflow(λ, δ, r, σ, θ_m, θ_f, γ_m, γ_f, ψ_m, ψ_f, g; β=0.5, verbose=false, step=0.2)

Method for one-dimensional types.

MarriageMarkets.estimate_static_surplus — Method.

Estimate the marital surplus from observed matches and singles.

MarriageMarkets.demand — Method.

Quasi-demand functions

Assuming that surplus is generated from equal gains on both sides. This is w.l.o.g. as gains and transfers cannot be separately identified.

MarriageMarkets.equilibrium — Method.

Compute equilibrium shares of singles and marriages.

Solves the matching equilibrium by searching for the transfers (analogous to prices) which clear the market.

MarriageMarkets.generate_surplus — Method.

Construct production array from function.

MarriageMarkets.prod_array — Method.

Construct production array from function.

MarriageMarkets.sex_solve — Method.

Wrapper for nlsolve that handles the concatenation and splitting of sex-specific arrays.

MarriageMarkets.sex_split — Method.

Split vector v into male/female pieces, where idx is number of male types.

MarriageMarkets.surplusdiv — Method.

Wife's consumption out of surplus (aggregate share).