Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MathOptInterfaceMosek.MosekModel — Type.

MosekModel <: MathOptInterface.AbstractModel

Linear variables and constraint can be deleted. For some reason MOSEK does not support deleting PSD variables.

Note also that adding variables and constraints will permanently add some (currently between 1 and 3) Int64s that a delete! will not remove. This ensures that Indices (Variable and constraint) that are deleted are thereafter invalid.

MathOptInterfaceMosek.checkconsistency — Method.

Check consistency of the internal structures.

MathOptInterfaceMosek.deleteblock — Method.

Move a block to the free list.

MathOptInterfaceMosek.ensurefree — Method.

ensurefree(s::LinkedInts, N :: Int)

Ensure that there are at least N elements free, and allocate as necessary.

MathOptInterfaceMosek.getfreeindexes — Method.

Get a list if the currently free elements.

MathOptInterfaceMosek.getindexes — Method.

getindexes(s::LinkedInts, id :: Int)

MathOptInterfaceMosek.getusedindexes — Method.

Get a list if the currently used elements.

MathOptInterfaceMosek.newblock — Method.

newblock(s::LinkedInts, N :: Int)

Add a new block in list idx