Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MatrixMarket.mmread — Function.

mmread(filename, infoonly::Bool=false, retcoord::Bool=false)

Read the contents of the Matrix Market file 'filename' into a matrix, which will be either sparse or dense, depending on the Matrix Market format indicated by 'coordinate' (coordinate sparse storage), or 'array' (dense array storage).

If infoonly is true (default: false), only information on the size and structure is returned from reading the header. The actual data for the matrix elements are not parsed.

If retcoord is true (default: false), the rows, column and value vectors are returned, if it is a sparse matrix, along with the header information.

MatrixMarket.mmwrite — Method.

mmwrite(filename, matrix::SparseMatrixCSC)

Write a sparse matrix to file 'filename'.