Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MaximumLikelihoodPower.KSstatistic — Method.

KSstatistic(data::AbstractVector, alpha)

Return the Колмогоров-Смирнов (Kolmogorov-Smirnov) statistic comparing data to a power law with power alpha. The elements of data are assumed to be unique.

MaximumLikelihoodPower.mle — Method.

(estimate, stderr) = mle(data::AbstractVector)

Return the maximum likelihood estimate and standard error of the exponent of a power law applied to the sorted vector data.

MaximumLikelihoodPower.mleKS — Method.

mleKS{T<:AbstractFloat}(data::AbstractVector{T})

Return the maximum likelihood estimate and standard error of the exponent of a power law applied to the sorted vector data. Also return the Kolmogorov-Smirnov statistic. Results are returned in an instance of type MLEKS.

MaximumLikelihoodPower.scanKS — Method.

scanKS(data, powers)

Compute the Kolmogorov Smirnov statistic for several values of α in the iterator powers. Return the value of α that minimizes the KS statistic and the two neighboring values.

MaximumLikelihoodPower.scanmle — Method.

scanmle(data; ntrials=100, stderrcutoff=0.1, useKS=false)

Perform mle approximately ntrials times on data, increasing xmin. Stop trials if the standard error of the estimate alpha is greater than stderrcutoff. If useKS is true, then the application of mle giving the smallest KS statistic is returned. Return an object containing statistics of the scan.

scanmle is intended to analayze the power-law behavior of the tail of data.

MaximumLikelihoodPower.MLEKS — Type.

MLEKS{T}

Container for storing results of MLE estimate and Kolmogorov-Smirnov statistic of the exponent of a power law.

MaximumLikelihoodPower.MLEScan — Type.

MLEScan{T <: AbstractFloat}

Record best estimate of alpha and associated parameters.

MaximumLikelihoodPower._scanmle — Method.

_scanmle{T<:AbstractFloat, V <: Integer}(data::AbstractVector{T}, range::AbstractVector{V},stderrcutoff)

Inner function for scanning power-law mle for power alpha over xmin. range specifies which xmin to try. stderrcutoff specifies a standard error in alpha at which we stop trials. range should be increasing.

MaximumLikelihoodPower.comparescan — Method.

comparescan(mle::MLEKS, i, data, mlescan::MLEScan)

Compare the results of MLE estimation mle to record results in mlescan and update mlescan.