Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MbedTLS.Cipher — Method.

Cipher(info::Union{CipherID, CipherKind}) -> Cipher

Construct a cipher object and set it to use the specified cipher algorithm.

The algorithm can either be specific (ie, CIPHER_AES_256_CBC), or general (ie, CIPHER_AES). In the latter case, a default choice of specific cipher will be used. See ?CipherInfo for more details.

MbedTLS.decrypt — Function.

decrypt(cipher, key, msg, [iv]) -> Vector{UInt8}

Decrypt a message using the given cipher. The cipher can be specified as

a generic cipher (like CIPHER_AES)

a specific cipher (like CIPHERAES256_CBC)

a Cipher object

key is the symmetric key used for cryptography, given as either a String or a Vector{UInt8}. It must be the right length for the chosen cipher; for example, CIPHERAES256_CBC requires a 32-byte (256-bit) key.

msg is the message to be encoded. It should either be convertible to a String or be a Vector{UInt8}.

iv is the initialization vector, whose size must match the block size of the cipher (eg, 16 bytes for AES) and correspond to the iv used by the encryptor. By default, it will be set to all zeros.

MbedTLS.digest — Function.

digest(kind::MDKind, msg::Vector{UInt8}, [key::Vector{UInt8}]) -> Vector{UInt8}

Perform a digest of the given type on the given message (a byte array), return a byte array with the digest.

If an optional key is given, perform an HMAC digest.

MbedTLS.digest! — Function.

digest!(kind::MDKind, msg::Vector{UInt8}, [key::Vector{UInt8}, ], buffer::Vector{UInt8})

In-place version of digest that stores the digest to buffer.

It is the user's responsibility to ensure that buffer is long enough to contain the digest. get_size(kind::MDKind) returns the appropriate size.

MbedTLS.encrypt — Function.

encrypt(cipher, key, msg, [iv]) -> Vector{UInt8}

Encrypt a message using the given cipher. The cipher can be specified as

a generic cipher (like CIPHER_AES)

a specific cipher (like CIPHERAES256_CBC)

a Cipher object

key is the symmetric key used for cryptography, given as either a String or a Vector{UInt8}. It must be the right length for the chosen cipher; for example, CIPHERAES256_CBC requires a 32-byte (256-bit) key.

msg is the message to be encoded. It should either be convertible to a String or be a Vector{UInt8}.

iv is the initialization vector, whose size must match the block size of the cipher (eg, 16 bytes for AES). By default, it will be set to all zeros, which is not secure. For security reasons, it should be set to a different value for each encryption operation.

MbedTLS.CipherInfo — Method.

CipherInfo(id::CipherID, key_bitlength, mode::CipherMode) -> CipherInfo

Construct a custom cipher info object.

id: A CipherID, such as CIPHERAES of CIPHERBLOWFIHS

key_bitlength: The bit length of the secret key. Available options depend

on the specific cipher id.

mode: Either Encrypt or Decrypt to indicate which cipher operation will

be performed with this cipher info.

MbedTLS.CipherInfo — Method.

CipherInfo(id::CipherID) -> CipherInfo

Build a CipherInfo for the given cipher id using the strongest available key size and the CBC block mode.

Warning: In CBC block mode, it is imperative that you use a unique IV (initial value) for each encryption operation to maintain security.

Base.close — Method.

Send a TLS close_notify message to the peer.

Base.eof — Method.

True if not isreadable and there are no more bytesavailable to read.

Base.isopen — Method.

Same as iswritable(ctx).

"...a closed stream may still have data to read in its buffer, use eof to check for the ability to read data." [?Base.isopen]

Base.isreadable — Method.

True unless:

TLS close_notify was received, or

the peer closed the connection (and the TLS buffer is empty), or

an un-handled exception occurred while reading.

Base.iswritable — Method.

True unless:

close(::SSLContext) is called, or

the peer closed the connection.

Base.readavailable — Method.

Read available decrypted data from ctx, but don't wait for more data to arrive.

The amount of decrypted data that can be read at once is limited by MBEDTLS_SSL_MAX_CONTENT_LEN.

Base.readbytes! — Function.

Copy at most nbytes of decrypted data from ctx into buf. If all=true: wait for sufficient decrypted data to be available. Less than nbytes may be copied if the peer sends TLS close_notify or closes the connection. Returns number of bytes copied into buf (<= nbytes).

Base.unsafe_read — Method.

Copy nbytes of decrypted data from ctx into buf. Wait for sufficient decrypted data to be available. Throw EOFError if the peer sends TLS close_notify or closes the connection before nbytes have been copied.

MbedTLS.f_recv — Method.

Copy at most nbytes of encrypted data to buf from the bio connection. If no encrypted bytes are available return:

MBEDTLS_ERR_SSL_WANT_READ if the connection is still open, or

MBEDTLS_ERR_NET_CONN_RESET if it is closed.

MbedTLS.f_send — Method.

Copy nbytes of encrypted data from buf to the underlying bio connection.

MbedTLS.get_size — Method.

get_size(kind::MDKind) -> Int

Returns the size of the digest in bytes that the given digest type requires.

For example,

get_size(MD_SHA256) == 32

MbedTLS.set_bio! — Method.

Connect f_send and f_recv callback functions to SSLContext.

MbedTLS.ssl_abandon — Method.

The documentation for ssl_read, ssl_write and ssl_close_notify all say:

If this function returns something other than 0 or MBEDTLSERRSSLWANTREAD/WRITE, you must stop using the SSL context for reading or writing, and either free it or call

This function ensures that the SSLContext is won't be used again.

MbedTLS.ssl_check_pending — Method.

ssl_check_pending(::SSLContext)::Bool

Check if there is data already read from the underlying transport but not yet processed.

If the SSL/TLS module successfully returns from an operation - e.g. a handshake or an application record read - and you're awaiting incoming data next, you must not immediately idle on the underlying transport to have data ready, but you need to check the value of this function first. The reason is that the desired data might already be read but not yet processed. If, in contrast, a previous call to the SSL/TLS module returned MBEDTLSERRSSLWANTREAD, it is not necessary to call this function, as the latter error code entails that all internal data has been processed.

https://tls.mbed.org/api/ssl_8h.html#a4075f7de9877fd667bcfa2e819e33426

MbedTLS.ssl_close_notify — Method.

ssl_close_notify(::SSLContext)

Notify the peer that the connection is being closed. 0 if successful, or a specific SSL error code.

If this function returns something other than 0 or MBEDTLSERRSSLWANTREAD/WRITE, you must stop using the SSL context for reading or writing, and either free it or call mbedtlssslsession_reset() on it before re-using it for a new connection; the current connection must be closed.

https://tls.mbed.org/api/ssl_8h.html#ac2c1b17128ead2df3082e27b603deb4c

MbedTLS.ssl_conf_renegotiation! — Method.

Enable / Disable renegotiation support for connection when initiated by peer (MBEDTLSSSLRENEGOTIATIONENABLED or MBEDTLSSSLRENEGOTIATIONDISABLED). See: https://tls.mbed.org/api/ssl_8h.html#aad4f50fc1c0a018fd5eb18fd9621d0d3

MbedTLS.ssl_get_bytes_avail — Method.

Return the number of application data bytes remaining to be read from the current record.

https://tls.mbed.org/api/ssl_8h.html#ad43142085f3182e9b0dc967ec582032b:

MbedTLS.ssl_read — Method.

ssl_read(::SSLContext, ptr, n)

Returns One of the following: 0 if the read end of the underlying transport was closed, the (positive) number of bytes read, or a negative error code on failure.

If MBEDTLSERRSSLWANTREAD is returned, no application data is available from the underlying transport. In this case, the function needs to be called again at some later stage.

If this function returns something other than a positive value or MBEDTLSERRSSLWANTREAD/WRITE or MBEDTLSERRSSLCLIENTRECONNECT, you must stop using the SSL context for reading or writing, and either free it or call mbedtlssslsession_reset() on it before re-using it for a new connection; the current connection must be closed.

https://tls.mbed.org/api/ssl_8h.html#aa2c29eeb1deaf5ad9f01a7515006ede5

MbedTLS.ssl_session_reset — Method.

Reset an already initialized SSL context for re-use while retaining application-set variables, function pointers and data.

MbedTLS.ssl_unsafe_read — Method.

ssl_unsafe_read(::SSLContext, buf, nbytes)

Copies at most nbytes of decrypted data into buf. Never blocks to wait for more data to arrive. Returns number of bytes copied into buf (<= nbytes). Updates ctx.bytesavailable with the number of decrypted bytes remaining in the TLS buffer.

Stops when:

nbytes have been copied, or

there are no more decrypted bytes available in the TLS buffer, or

a TLS close_notify message is received.

When TLS close_notify is received:

isreadable is set to false and bytesavailable is set to zero. [RFC5246 7.2.1]: "Any data received after a closure alert is ignored."

the number of bytes read before the close_notify is returned as usual.

Throws a MbedException if ssl_read returns an unhandled error code.

When an unhandled exception occurs isreadable is set to false.

MbedTLS.ssl_unsafe_write — Method.

Copy nbytes from buf to ctx.

The TLS library function ssl_write is called as many times as needed to send all the data. The TLS library encrypts the data and passes it to the f_send function which sends it to the underlying connection (ctx.bio). See f_send and set_bio! below.

MbedTLS.ssl_write — Method.

ssl_write(::SSLContext, ptr, n)

When this function returns MBEDTLSERRSSLWANTWRITE/READ, it must be called later with the same arguments, until it returns a value greater that or equal to 0. When the function returns MBEDTLSERRSSLWANTWRITE there may be some partial data in the output buffer, however this is not yet sent.

If this function returns something other than 0, a positive value or MBEDTLSERRSSLWANTREAD/WRITE, you must stop using the SSL context for reading or writing, and either free it or call mbedtlssslsession_reset() on it before re-using it for a new connection; the current connection must be closed.

https://tls.mbed.org/api/ssl_8h.html#a5bbda87d484de82df730758b475f32e5

MbedTLS.update! — Method.

update!(cipher::Cipher, buf_in, buf_out::Vector{UInt8}) -> Int

Run the given cipher on buf_in (a String or Vector{UInt8}) and store the result of the cipher in buf_out (a Vector{UInt8}).

It is your responsibility to ensure that buf_out is at least as large as necessary to hold the result. It should be at least as big as the size of buf_in plus the block size associated with cipher.

MbedTLS.wait_for_decrypted_data — Method.

While there are no decrypted bytes available but the connection is readable:

If the TLS buffer has no pending (unprocessed) data wait for more encrypted data to arrive on the underlying connection.

Run a zero-byte read to allow the library to process its internal buffer, and/or read from the underlying connection.

ssl_unsafe_read updates the isreadable and bytesavailable state.