Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Media.getdisplay — Method.

getdisplay(T)

Find out what output device T will display on.

Media.media — Method.

media(T) gives the media type of the type T. The default is Textual.

media(Gadfly.Plot) == Media.Plot

Media.setdisplay — Method.

setdisplay([input], T, output)

Display T objects using output when produced by input.

T is an object type or media type, e.g. Gadfly.Plot or Media.Graphical.

display(Editor(), Image, Console())