Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MeshCat.Visualizer — Type.

vis = Visualizer()

Construct a new MeshCat visualizer instance.

Useful methods:

vis[:group1] # get a new visualizer representing a sub-tree of the scene
setobject!(vis, geometry) # set the object shown by this visualizer's sub-tree of the scene
settransform!(vis], tform) # set the transformation of this visualizer's sub-tree of the scene

Base.delete! — Method.

delete!(vis)

Delete the geometry at this visualizer's path and all of its descendants.

MeshCat.setobject! — Method.

setobject!(vis, obj)

Set the object at this visualizer's path. This replaces whatever geometry was presently at that path. To draw multiple geometries, place them at different paths by using the slicing notation:

setobject!(vis[:group1][:box1], geometry1)
setobject!(vis[:group1][:box2], geometry2)

MeshCat.setprop! — Method.

setprop!(vis, property, value)

Set a single property for the object at the given path.

(this is named setprop! instead of setproperty! to avoid confusion with the Base.setproperty! function introduced in Julia v0.7)

MeshCat.settransform! — Method.

settransform!(vis, tform)

Set the transform of this visualizer's path. This can be done before or after adding an object at that path. The overall transform of an object is the composition of the transforms of all of its parents, so setting the transform of vis[:group1] affects the poses of the objects at vis[:group1][:box1] and vis[:group1][:box2].

Base.wait — Method.

wait(v)

Wait until at least one browser has connected to the visualizer's server. This is useful in scripts to delay execution until the browser window has opened.

MeshCat.intrinsic_transform — Method.

intrinsic_transform(g)

Different tools disagree about what various geometric primitives mean. For example, GeometryTypes.jl considers the "origin" of a cube to be its bottom-left corner, where DrakeVisualizer and MeshCat consider its origin to be the center. The intrinsic_transform(g) returns the transform from the GeometryTypes origin to the MeshCat origin.

MeshCat.lower — Function.

Convert a geometry, material, object, or transform into the appropriate plain data structures expected by three.js. Most objects are lowered into Dicts matching the JSON structure used by three.js.