Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MeshCatMechanisms.animate — Method.

animate(vis::MechanismVisualizer,
        times::Vector{Float64},
        configurations::Vector{Vector{Float64}};
        fps::Float64=60, realtimerate::Float64=1.)

Animate the given mechanism passing through a time-coded series of configurations by linearly interpolating the configuration vectors.

MeshCatMechanisms.setelement! — Function.

setelement!(mvis::MechanismVisualizer, point::Point3D, radius::Real=0.05, name::AbstractString="<element>")

Add a HyperSphere geometry with the given radius to the visualizer at the given point

MeshCatMechanisms.setelement! — Function.

setelement!(mvis::MechanismVisualizer, element::VisualElement, name::AbstractString="<element>")

Attach the given visual element to the visualizer. The element's frame will determine how its geometry is attached to the scene tree, so that any other geometries attached to the same body will all move together.

MeshCatMechanisms.setelement! — Function.

setelement!(mvis::MechanismVisualizer, frame::CartesianFrame3D, object::AbstractObject, name::AbstractString="<element>")

Attach the given geometric object (geometry + material) to the visualizer at the given frame

MeshCatMechanisms.setelement! — Function.

setelement!(mvis::MechanismVisualizer, frame::CartesianFrame3D, geometry::GeometryLike, name::AbstractString="<element>")

Attach the given geometry to the visualizer at the given frame, using its default material.

MeshCatMechanisms.setelement! — Function.

setelement!(mvis::MechanismVisualizer, frame::CartesianFrame3D, geometry::GeometryLike, material::AbstractMaterial, name::AbstractString="<element>")

Construct an object with the given geometry and material and attach it to the visualizer

MeshCatMechanisms.setelement! — Function.

setelement!(mvis::MechanismVisualizer, frame::CartesianFrame3D, scale::Real=0.5, name::AbstractString="<element>")

Add a Triad geometry with the given scale to the visualizer at the specified frame

RigidBodyDynamics.set_configuration! — Method.

set_configuration!(mvis::MechanismVisualizer, args...)

Set the configuration of the mechanism visualizer and re-render it.

Examples

julia> set_configuration!(vis, [1., 2., 3.])

julia> set_configuration!(vis, findjoint(robot, "shoulder"), 1.0)