Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Missings.allowmissing — Method.

allowmissing(x::AbstractArray)

Return an array equal to x allowing for missing values, i.e. with an element type equal to Union{eltype(x), Missing}.

When possible, the result will share memory with x (as with convert).

See also: disallowmissing

Missings.disallowmissing — Method.

disallowmissing(x::AbstractArray)

Return an array equal to x not allowing for missing values, i.e. with an element type equal to Missings.T(eltype(x)).

When possible, the result will share memory with x (as with convert). If x contains missing values, a MethodError is thrown.

See also: allowmissing

Missings.levels — Method.

levels(x)

Return a vector of unique values which occur or could occur in collection x, omitting missing even if present. Values are returned in the preferred order for the collection, with the result of sort as a default.

Contrary to unique, this function may return values which do not actually occur in the data, and does not preserve their order of appearance in x.

Missings.passmissing — Method.

passmissing(f)

Return a function that returns missing if any of its positional arguments are missing (even if their number or type is not consistent with any of the methods defined for f) and otherwise applies f to these arguments.

passmissing does not support passing keyword arguments to the f function.

Examples

```jldoctest julia> passmissing(sqrt)(4) 2.0

julia> passmissing(sqrt)(missing) missing

julia> passmissing(sqrt).([missing, 4]) 2-element Array{Union{Missing, Float64},1}: missing 2.0

julia> passmissing((x,y)->"x y")(1, 2) "1 2"

julia> passmissing((x,y)->"x y")(missing) missing

Missings.fail — Method.

Missings.fail(itr)

Return an iterator wrapping iterable itr which will throw a MissingException if a missing value is found.

Use collect to obtain an Array containing the resulting values. If itr is an array, the resulting array will have the same dimensions.

See also: skipmissing, Missings.replace

Examples

julia> collect(Missings.fail([1 2; 3 4]))
2×2 Array{Int64,2}:
 1  2
 3  4

julia> collect(Missings.fail([1, missing, 2]))
ERROR: MissingException: missing value encountered by Missings.fail
[...]

Missings.replace — Method.

Missings.replace(itr, replacement)

Return an iterator wrapping iterable itr which replaces missing values with replacement. When applicable, the size of itr is preserved. If the type of replacement differs from the element type of itr, it will be converted.

See also: skipmissing, Missings.fail

Examples

julia> collect(Missings.replace([1, missing, 2], 0))
3-element Array{Int64,1}:
 1
 0
 2

julia> collect(Missings.replace([1 missing; 2 missing], 0))
2×2 Array{Int64,2}:
 1  0
 2  0