Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Mocking.Bindings — Method.

Bindings(expr)

Takes a function signature expression and extracts all of the bindings into either internal (defined by the signature) or external (defined outside of the signature).

julia> Bindings(:(f{T<:Integer}(a::T, b::Base.Real=a, c=z)))
Mocking.Bindings(Set(Any[:c,:T,:a,:b,:f]),Set(Any[:z,:(Base.Real),:Integer]))

Mocking.binding_expr — Function.

binding_expr(x) -> Expr

Converts a Module, Type, or Function into an expression which includes the entire module hierarchy.

julia> binding_expr(Int8)
:(Core.Int8)

julia> binding_expr(Dates.Hour)
:(Base.Dates.Hour)

Mocking.compiled_modules_enabled — Method.

compiled_modules_enabled() -> Bool

Determine if the julia command line flag --compiled-modules has been set to "yes".

Mocking.extract_kwargs — Method.

extract_kwargs(expr::Expr) -> Vector{Expr}

Extract the :parameters and :kw value into an array of :kw expressions we don't evaluate any expressions for values yet though.

Mocking.set_compiled_modules — Method.

set_compiled_modules(state::Bool) -> Void

Override the julia command line flag --compiled-modules with a runtime setting. Code run before this the value is modified will use the original setting. Not meant for general purpose usage.