Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MortalityTables.qx — Method.

qx(table::XTbMLTable, age)

Given a mortality table and an age returns the qx.

If the table is a select and ultimate table, it will return the select mortality during the ulitmate period because duration was not given.

If the table does not have select rates, will just return the rate for the given age.

MortalityTables.qx — Method.

qx(table::XTbMLTable, age)

Given a mortality table and an age returns the appropriate ultimate qx.

MortalityTables.qx — Method.

qx(table::XTbMLTable, issueAge, duration)

Given a mortality table, an issue age, and a duration, returns the appropriate select or ultimate qx.