Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MotifSequenceGenerator.all_possible_sums — Method.

all_possible_sums(summands, n)

Compute all possible sums from combining n elements from summands (with repetition), only using unique combinations.

Return a vector of tuples: the first entry of each tuple is the sum, while the second is the indices of summands used to compute the sum.

MotifSequenceGenerator.random_sequence — Method.

random_sequence(motifs::Vector{M}, q, limits, translate, δq = 0; kwargs...)

Create a random sequence of motifs of type M, under the constraint that the sequence has "length" ℓ exactly within q - δq ≤ ℓ ≤ q + δq. Return the sequence itself as well as the sequence of indices of motifs used to create it.

"length" here means an abstracted length defined by the struct M, based on the limits and translate functions. It does not refer to the amount of elements!

M can be anything, given the two functions

limits(motif) : Some function that given the motif it returns the (start, fine) of the the motif in the same units as q. This function establishes a measure of length, which simply is fine - start.

translate(motif, t) : Some function that given the motif it returns a new motif which is translated by t (either negative or positive), with respect to the same units as q.

Keywords

Please see the source code (use @which) for a full description of the algorithm.

tries = 5 : Up to how many initial random sequences are accepted.

taulcut = 2 : Up to how times an element is dropped from the initial guess.

summands = 3 : Up to how many motifs may be combined as a sum to complete a sequence.

MotifSequenceGenerator._motifs_at_origin — Method.

_motifs_at_origin(motifs, limits, translate) -> (motifs0, motiflens)

Bring all motifs to the origin and compute the motif lengths.

MotifSequenceGenerator._random_sequence_try — Method.

_random_sequence_try(motiflens, q) -> seq, seq_length

Return a random sequence of motif indices so that the total sequence is guaranteed to have total length of q - δq ≤ ℓ ≤ q - δq + maximum(motiflens).