Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MuladdMacro.args — Method.

args(ex)

Return arguments of function call in ex.

MuladdMacro.iscall — Method.

iscall(ex, op)

Determine whether ex is a call of operation op with at least two arguments.

MuladdMacro.isdotcall — Method.

isdotcall(ex, op)

Determine whether ex is a dot call of operation op with at least two arguments.

MuladdMacro.isdotcall — Method.

isdotcall(ex)

Determine whether ex is a dot call.

MuladdMacro.ismul — Method.

ismul(ex, dot::Bool)

Determine whether expression ex is a multiplication that is dotted if dot is true and not dotted otherwise.

MuladdMacro.issub — Method.

issub(ex)

Determine whether ex is a subtraction.

MuladdMacro.issum — Method.

issum(ex)

Determine whether ex is a sum.

MuladdMacro.newargs — Method.

newargs(ex, args)

Create new expression of function call ex with arguments args.

Unary function calls are not considered, i.e. if only one function argument is provided it is returned.

MuladdMacro.newmuladd — Method.

newmuladd(x, y, z, dot::Bool)

Return expression (muladd).(x, y, z) if dot is true and muladd(x, y, z) otherwise.

MuladdMacro.splitargs — Method.

splitargs(ex)

Split arguments of function call ex before last argument and combine first arguments to one expression if possible.

MuladdMacro.sub_to_muladd — Method.

sub_to_muladd(ex)

Replace subtraction ex by muladd if possible.

MuladdMacro.sum_to_muladd — Method.

sum_to_muladd(ex)

Replace sum ex by sequence of muladd if possible. Hereby the order of summation might be changed.

MuladdMacro.to_muladd — Method.

to_muladd(ex)

Convert every combination of addition/subtraction and multiplication in expression ex to a call of muladd.

If both of the involved operators are dotted, muladd is applied as a dot call The order of summation might be changed.