Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Multisets.Multiset — Type.

A Multiset is an unordered collection of things with repetition permitted. A new Multiset container is created with Multiset{T}() where T is the type of the objects held in the multiset. If T is omitted, it defaults to Any.

A Multiset can be created from a collection list (such as a Vector or Set) with Multiset(list). If an element is repeated in list it has the appropriate multiplicity.

Multisets.set_braces_show — Method.

Set braces display mode for multisets, like this:

{1,2,2,3,3,3,3}

See also set_short_show and set_julia_show.

Multisets.set_julia_show — Method.

Set Julia style display mode for multisets, like this:

Multiset(Int64[1,2,2,3,3,3,3])

See also set_short_show and set_braces_show.

Multisets.set_short_show — Method.

Set show display mode for multisets, like this:

Multiset{Int64} with 7 elements

See also set_braces_show and set_julia_show.

Base.:& — Method.

A&B for multisets is intersect(A,B).

Base.:* — Method.

n*A is the scalar multiple of a multiset in which the multiplicity of x is n*A[x]. Of course, we require n >= 0.

Base.:* — Method.

A*B for the Cartesian product of multisets A and B.

Base.:+ — Method.

A+B for multisets is the disjoint union, i.e., a new multiset in which the multiplicity of x is A[x]+B[x]. This can be abbreviated A|B.

Base.:- — Method.

A-B for multisets is the multiset difference, i.e., a new multiset in which the multiplicity of x is A[x]-B[x] unless this goes below 0, in which case the multiplicity is 0.

Base.:| — Method.

For multisets, A|B is union(A,B). See also + which behaves differently.

Base.getindex — Method.

For a M[t] where M is a Multiset returns the multiplicity of t in M. A value of 0 means that t is not a member of M.

Base.intersect — Method.

intersect(A,B) for multisets creates a new multiset in which the multiplicity of x is min(A[x],B[x]). This may be abbreviated A&B.

Base.push! — Method.

push!(M,x,incr) increases the multiplicity of x in M by incr (which defaults to 1). incr can be negative, but it is not possible to decrease the multiplicty below 0.

Base.union — Method.

union(A,B) for multisets creates a new multiset in which the multiplicity of x is max(A[x],B[x]).

Multisets.clean! — Method.

clean!(M) removes elements of multiplicy 0 from the underlying data structure supporting M.