Atoms extraction

Atoms extration

Moment matrix

MultivariateMoments.MatMeasure — Type.

mutable struct MatMeasure{T, MT <: AbstractMonomial, MVT <: AbstractVector{MT}} <: AbstractMeasureLike{T}
    Q::SymMatrix{T}
    x::MVT
    support::Union{Nothing, AlgebraicSet}
end

Measure $\nu$ represented by the moments of the monomial matrix $x x^\top$ in the symmetric matrix Q. The set of points that are zeros of all the polynomials $p$ such that $\mathbb{E}_{\nu}[p] = 0$ is stored in the field support when it is computed.

MultivariateMoments.matmeasure — Function.

matmeasure(μ::Measure, x)

Creates a matrix the moment matrix for the moment matrix $x x^\top$ using the moments of μ.

Atomic measure

MultivariateMoments.WeightedDiracMeasure — Type.

struct WeightedDiracMeasure{T}
    center::Vector{T}
    weight::T
end

Represents a weighted dirac measure with centered at center and with weight weight.

MultivariateMoments.AtomicMeasure — Type.

struct AtomicMeasure{T, V} <: AbstractMeasureLike{T}
    variables::V                           # Vector/Tuple of variables
    atoms::Vector{WeightedDiracMeasure{T}} # Atoms of the measure
end

An measure is said to be atomic if it is a sum of weighted dirac measures. For instance, $\eta = 2 \delta_{(1, 0)} + 3 \delta_{(1/2, 1/2)}$ is an atomic measure since it is a sum of the diracs centered at $(1, 0)$ and $(1/2, 1/2)$ and weighted respectively by 2 and 3. That is, $\mathbb{E}_{\eta}[p] = 2p(1, 0) + 3p(1/2, 1/2)$.

The AtomicMeasure struct stores an atomic measure where variables contains the variables and atoms contains atoms of the measure.

Atoms extraction

Given a MatMeasure with a positive semidefinite moment matrix, an algebraic system for which the set of solution is a superset of the support of the measure. If the measure is atomic and the MatMeasure contains enough moments, the algebraic system will only have a finite number of solutions which are the centers of the diracs of the measure.

MultivariateMoments.computesupport! — Function.

MultivariateMoments.computesupport!(ν::MatMeasure, ranktol, [dec])

Computes the support field of ν. The ranktol and dec parameters are passed as is to the lowrankchol function.

MultivariateMoments.extractatoms — Function.

extractatoms(ν::MatMeasure, ranktol, [dec])

Return an AtomicMeasure with the atoms of ν if it is atomic or nothing if ν is not atomic. The ranktol and dec parameters are passed as is to the lowrankchol function.

This system is obtained from a low rank cholesky decomposition of the moment matrix. This decomposition can either be obtained by a cholesky or SVD decomposition from which we remove the rows corresponding to the negligeable eigenvalues/singular values.

MultivariateMoments.LowRankChol — Type.

LowRankChol

Method for computing a $r \times n$ matrix U of a $n \times n$ rank $r$ psd matrix Q such that Q = U'U.

MultivariateMoments.ShiftChol — Type.

ShiftChol <: LowRankChol

Shift the matrix by shift times the identity matrix before cholesky.

MultivariateMoments.SVDChol — Type.

SVDChol <: LowRankChol

Use SVD decomposition.

MultivariateMoments.lowrankchol — Function.

MultivariateMoments.lowrankchol(Q::AbstractMatrix, dec::LowRankChol, ranktol)

Returns a $r \times n$ matrix $U$ of a $n \times n$ rank $r$ positive semidefinite matrix $Q$ such that $Q = U^\top U$. The rank of $Q$ is the number of singular values larger than ranktol${} \cdot \sigma_1$ where $\sigma_1$ is the largest singular value.