Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MusicManipulations.MoreVelNote — Type.

MoreVelNote

Enables velocities higher than 127 but other than that identical with Note.

When Notes with MoreVelNote are written into a MIDITrack, any velocity higher than 127 is equaled to 127.

Base.transpose — Method.

transpose(notes, semitones)

Transpose the notes for the given amount of semitones.

MusicManipulations.allowedpitches — Method.

allowedpitches(notes::Notes, allowed) -> newnotes

Only keep the notes that have a pitch specified in allowed (one or many pitches).

MusicManipulations.classify — Method.

classify(notes::Notes, grid)
classify(note::AbstractNote, grid, tpq::Integer)

Classify given notes according to the given grid.

Returns an integer (or vector of integers) that corresponds to the index of the closest grid point to the note position modulo the quarter note. 1 means start of the grid and length(grid) means end of the grid (i.e. next quarter note).

MusicManipulations.firstnotes — Method.

firstnotes(notes, grid)

Return the notes that first appear in each grid point, without quantizing them.

This function does not consider the notes modulo the quarter note! Different quarter notes have different grid points.

MusicManipulations.getnotnotes — Method.

getnotnotes(track::MIDI.MIDITrack) -> (abs_pos, events)

Find all events in track that are not NOTEON or NOTEOFF. Return the found events and their positions in absolute time (in ticks).

Each event can be added to another MIDITrack using

for ev in zip(abs_pos, events)
    MIDI.addevent!(newtrack, ev...)
end

See also replace_notes.

MusicManipulations.quantize — Function.

quantize(notes::Notes, grid, duration = true)
quantize(note::AbstractNote, grid, tpq::Integer)

Return a quantized copy of the given notes on the given grid, which can be any sorted iterable that starts on 0 and ends on 1.

Each note is quantized (relocated) to its closest point of the grid, by first identifying that point using classify. It is assumed that the grid is the same for all quarter notes of the track.

If duration is true, the function also quantizes the duration of the notes on the same grid.

This function respects the notes' absolute position and quantizes in absolute position, not relative.

MusicManipulations.quantize_duration! — Method.

quantize_duration!(notes::Notes, grid)
quantize_duration!(note::Note, grid, tpq)

Quantize the duration of given notes on the grid.

MusicManipulations.random_notes_sequence — Method.

random_notes_sequence(motifs::Vector{Notes{N}}, q, δq = 0)

Create a random sequence from a pool of notes (motifs) such that it has total length ℓ exactly q - δq ≤ ℓ ≤ q + δq. Notice that q is measured in ticks.

Return the result as a single Notes container, and also return the sequence of motifs used.

This function uses random_sequence from the module MotifSequenceGenerator, adapted to the Notes struct.

MusicManipulations.randomnotes — Function.

randomnotes(n::Int, tpq = 960)

Generate some random notes that start sequentially.

MusicManipulations.replace_notes — Method.

replace_notes(oldtrack::MIDITrack, notes::Notes) -> newtrack

Create a new midi track that copies all MIDIEvents that are not NOTEON or NOTEOFF from the oldtrack while it also adds all notes from notes.

MusicManipulations.separatepitches — Method.

separatepitches(notes::Notes [, allowed])

Get a dictionary "pitch"=>"notes of that pitch". Optionally only keep pitches that are contained in allowed.

MusicManipulations.subdivision — Method.

subdivision(n::Int, tpq)

Return how many ticks is the duration of the subdivision of a 4/4-bar into n equal parts, assuming the ticks per quarter note are tpq.

For example, for sixteenth notes you would do subdivision(16, tpq), for eigth-note triplets subdivision(12, tpq) and so on.

MusicManipulations.timeseries — Function.

timeseries(notes::Notes, property, f [, grid]) -> tvec, ts

Produce a timeseries of the property of the given notes, optionally first quantizing on the given grid.

After quantization, it is often the case that many notes are in the same bin of the grid. The function f denotes which value of the vector of the property of the notes to keep. Typical values are minimum, maximum, mean, etc (for type stability the returned timeseries are always Float64).

The property can be :velocity, :pitch, or :duration. Grid bins without any notes are given the value 0. This can be problematic if you request for :pitch and your notes also include notes which actually have pitch 0, i.e. C0.

MusicManipulations.translate — Method.

translate(notes, ticks)

Translate the notes for the given amount of ticks.