Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Mustache.render — Method.

render([io], tokens, view)

Render a set of tokens with a view, using optional io object to print or store.

Arguments

io::IO: Optional IO object.

tokens: Either Mustache tokens, or a string to parse into tokens

view: A view provides a context to look up unresolved symbols demarcated by mustache braces. A view may be specified by a dictionary, a module, a composite type, a vector, a named tuple, a data frame, a Tables object, or keyword arguments.

Mustache.render_from_file — Method.

Renders a template from filepath and view. If it has seen the file before then it finds the compiled MustacheTokens in TEMPLATES rather than calling parse a second time.