Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

MySQL.Query — Method.

MySQL.Query(conn, sql, sink=Data.Table; append::Bool=false) => MySQL.Query

execute an sql statement and return a MySQL.Query object. Result rows can be iterated as NamedTuples via Data.rows(query) where query is the MySQL.Query object. Results can also be streamed to any valid Data.Sink via Data.stream!(query, sink).

MySQL.Stmt — Type.

MySQL.Stmt(conn, sql) => MySQL.Stmt

Prepare an SQL statement that may contain ? parameter placeholders.

A MySQL.Stmt may then be executed by calling MySQL.execute!(stmt, params) where params are the values to be bound to the ? placeholders in the original SQL statement. Params must be provided for every ? and will be matched in the same order they appeared in the original SQL statement.

Bulk statement execution can be accomplished by "streaming" a param source like:

Data.stream!(source, stmt)

where source is any valid Data.Source (from DataStreams.jl). As with MySQL.execute!, the source must provide enough params and will be matched in the same order.

MySQL.connect — Method.

MySQL.connect(host::String, user::String, passwd::String; db::String = "", port = "3306", socket::String = MySQL.API.MYSQL_DEFAULT_SOCKET, opts = Dict())

Connect to a MySQL database.

MySQL.disconnect — Method.

MySQL.disconnect(conn::MySQL.Connection)

Close a handle to a MySQL database opened by MySQL.connect.

MySQL.escape — Method.

MySQL.escape(conn::MySQL.Connection, str::String) -> String

Escapes a string using mysql_real_escape_string(), returns the escaped string.

MySQL.execute! — Method.

MySQL.execute!(conn, sql) => Nothing

execute an sql statement without returning results (useful for DDL statements, update, delete, etc.)

MySQL.query — Function.

MySQL.query(conn, sql, sink=Data.Table, args...; append::Bool=false) => sink

execute an sql statement and return the results in sink, which can be any valid Data.Sink (interface from DataStreams.jl), and args... are any necessary arguments to the sink. By default, a NamedTuple of Vectors are returned.

Passing append=true as a keyword argument will cause the resultset to be appended to the sink instead of replacing.

To get the results as a DataFrame, you can just do MySQL.query(conn, sql, DataFrame).

See list of DataStreams implementations here