Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NBInclude.nbinclude — Method.

nbinclude(m::Module, path; ...)

Like @nbinclude(path; ...) but allows you to specify a module to evaluate in, similar to include(m, path).

NBInclude.my_include_string — Method.

my_include_string(m::Module, s::AbstractString, path::AbstractString, prev, softscope)

Like includestring (or softscopeincludestring, depending on the softscope flag), but also change the current source path just as include(filename) would do. We are hacking undocumented internals of Julia here (see `base/loading.jl:includerelative), but it hasn't changed from Julia 0.2 to Julia 0.7 so it's not too crazy.prevshould be the previous path returned byBase.source_path`.