Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NFFT.NFFTPlan — Type.

    NFFTPlan(x, N, ...) -> plan

Compute D dimensional NFFT plan for sampling locations x (a vector or a D-by-M matrix) that can be applied on arrays of size N (a tuple of length D).

The optional arguments control the accuracy.

It takes as optional keywords all the keywords supported by plan_fft function (like flags and timelimit). See documentation of plan_fft for reference.

NFFT.NFFTPlan — Method.

    NFFTPlan(x, d, N, ...) -> plan

Compute directional NFFT plan: A 1D plan that is applied along dimension d of a D dimensional array of size N with sampling locations x (a vector).

It takes as optional keywords all the keywords supported by plan_fft function (like flags and timelimit). See documentation of plan_fft for reference.

NFFT.nfft — Method.

    nfft(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the NFFT of a D dimensional array f of size N. fHat is a vector of length M. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

NFFT.nfft_adjoint — Method.

    nfft_adjoint(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the adjoint NFFT of a length M vector fHat f is a D dimensional array of size N. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

NFFT.nfft! — Method.

    nfft!(p, f, fHat) -> fHat

Calculate the NFFT of f with plan p and store the result in fHat.

Both f and fHat must be complex arrays.

NFFT.nfft_adjoint! — Function.

    nfft_adjoint!(p, fHat, f) -> f

Calculate the adjoint NFFT of fHat and store the result in f.

Both f and fHat must be complex arrays.