Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NLSolversBase.gradient! — Method.

Evaluates the gradient value at x.

Stores the value in obj.DF.

NLSolversBase.gradient — Method.

Evaluates the gradient value at x

This does not update obj.DF or obj.x_df.

NLSolversBase.gradient — Method.

Get the ith element of the most recently evaluated gradient of obj.

NLSolversBase.gradient — Method.

Get the most recently evaluated gradient of obj.

NLSolversBase.hessian — Method.

Get the most recently evaluated Hessian of obj

NLSolversBase.jacobian — Method.

Get the most recently evaluated Jacobian of obj.

NLSolversBase.value!! — Method.

Force (re-)evaluation of the objective value at x.

Returns f(x) and stores the value in obj.F

NLSolversBase.value! — Method.

Evaluates the objective value at x.

Returns f(x) and stores the value in obj.F

NLSolversBase.value — Method.

Evaluates the objective value at x.

Returns f(x), but does not store the value in obj.F

NLSolversBase.value — Method.

Get the most recently evaluated objective value of obj.

NLSolversBase.f!_from_f — Method.

f!_from_f(f, F::Abstractarray)

Return an inplace version of f

NLSolversBase.gradient!! — Method.

Force (re-)evaluation of the gradient value at x.

Stores the value in obj.DF.

NLSolversBase.nconstraints — Method.

nconstraints(bounds) -> nc

The number of linear/nonlinear constraint functions supplied by the user. This does not include bounds-constraints on variables.

See also: nconstraints_x.

NLSolversBase.nconstraints_x — Method.

nconstraints_x(bounds) -> nx

The number of "meaningful" constraints (not ±Inf) on the x coordinates.

See also: nconstraints.

NLSolversBase.parse_constraints — Method.

parse_constraints(T, l, u) -> eq, val, ineq, σ, b

From user-supplied constraints of the form

l_i ≤  v_i  ≤ u_i

(which include both inequality and equality constraints, the latter when l_i == u_i), convert into the following representation:

- `eq`, a vector of the indices for which `l[eq] == u[eq]`
- `val = l[eq] = u[eq]`
- `ineq`, `σ`, and `b` such that the inequality constraints can be written as
         σ[k]*(v[ineq[k]] - b[k]) ≥ 0
   where `σ[k] = ±1`.

Note that since the same v_i might have both lower and upper bounds, ineq might have the same index twice (once with σ=-1 and once with σ=1).

Supplying ±Inf for elements of l and/or u implies that v_i is unbounded in the corresponding direction. In such cases there is no corresponding entry in ineq/σ/b.

T is the element-type of the non-Int outputs