Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NNlib.elu — Function.

elu(x, α = 1) =
  x > 0 ? x : α * (exp(x) - 1)

Exponential Linear Unit activation function. See Fast and Accurate Deep Network Learning by Exponential Linear Units. You can also specify the coefficient explicitly, e.g. elu(x, 1).

NNlib.leakyrelu — Function.

leakyrelu(x) = max(0.01x, x)

Leaky Rectified Linear Unit activation function. You can also specify the coefficient explicitly, e.g. leakyrelu(x, 0.01).

NNlib.logsoftmax — Function.

logsoftmax(xs) = log.(exp.(xs) ./ sum(exp.(xs)))

logsoftmax computes the log of softmax(xs) and it is more numerically stable than softmax function in computing the cross entropy loss.

NNlib.logσ — Method.

logσ(x)

Return log(σ(x)) which is computed in a numerically stable way.

julia> logσ(0.)
-0.6931471805599453
julia> logσ.([-100, -10, 100.])
3-element Array{Float64,1}:
 -100.0
  -10.0
   -0.0

NNlib.relu — Method.

relu(x) = max(0, x)

Rectified Linear Unit activation function.

NNlib.selu — Method.

selu(x) = λ * (x ≥ 0 ? x : α * (exp(x) - 1))

λ ≈ 1.0507
α ≈ 1.6733

Scaled exponential linear units. See Self-Normalizing Neural Networks.

NNlib.softmax — Function.

softmax(xs) = exp.(xs) ./ sum(exp.(xs))

Softmax takes log-probabilities (any real vector) and returns a probability distribution that sums to 1.

If given a matrix it will treat it as a batch of vectors, with each column independent.

julia> softmax([1,2,3.])
3-element Array{Float64,1}:
  0.0900306
  0.244728
  0.665241

NNlib.softplus — Method.

softplus(x) = log(exp(x) + 1)

See Deep Sparse Rectifier Neural Networks.

NNlib.softsign — Method.

softsign(x) = x / (1 + |x|)

See Quadratic Polynomials Learn Better Image Features.

NNlib.swish — Method.

swish(x) = x * σ(x)

Self-gated actvation function. See Swish: a Self-Gated Activation Function.

NNlib.σ — Method.

σ(x) = 1 / (1 + exp(-x))

Classic sigmoid activation function.

NNlib.log_fast — Function.

log_fast(x)

Compute the natural logarithm of x. The inverse of the natural logarithm is the natural expoenential function exp(x)