Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NaNMath.extrema — Method.

NaNMath.extrema(A)

Args:

A: An array of floating point numbers

Returns:

Returns the minimum and maximum of all elements in the array, ignoring NaN's.

Examples:

using NaNMath as nm
nm.extrema([1., 2., NaN]) # result: 1.0, 2.0

NaNMath.max — Method.

NaNMath.max(x, y)

Compute the IEEE 754-2008 compliant maximum of x and y. As of version 0.6 of Julia, Base.max(x, y) will return NaN if x or y is NaN. NaNMath.max favors values over NaN, and will return whichever x or y is not NaN in that case.

Examples

julia> NaNMath.max(NaN, 0.0)
0.0

julia> NaNMath.max(1, 2)
2

NaNMath.maximum — Method.

NaNMath.maximum(A)

Args:

A: An array of floating point numbers

Returns:

Returns the maximum of all elements in the array, ignoring NaN's.

Examples:

using NaNMath as nm
nm.maximum([1., 2., NaN]) # result: 2.0

NaNMath.mean — Method.

NaNMath.mean(A)

Args:

A: An array of floating point numbers

Returns:

Returns the arithmetic mean of all elements in the array, ignoring NaN's.

Examples:

using NaNMath as nm
nm.mean([1., 2., NaN]) # result: 1.5

NaNMath.mean_count — Method.

Returns a tuple of the arithmetic mean of all elements in the array, ignoring NaN's, and the number of non-NaN values in the array.

NaNMath.min — Method.

NaNMath.min(x, y)

Compute the IEEE 754-2008 compliant minimum of x and y. As of version 0.6 of Julia, Base.min(x, y) will return NaN if x or y is NaN. NanMath.min favors values over NaN, and will return whichever x or y is not NaN in that case.

Examples

julia> NanMath.min(NaN, 0.0)
0.0

julia> NaNMath.min(1, 2)
1

NaNMath.minimum — Method.

NaNMath.minimum(A)

Args:

A: An array of floating point numbers

Returns:

Returns the minimum of all elements in the array, ignoring NaN's.

Examples:

using NaNMath as nm
nm.minimum([1., 2., NaN]) # result: 1.0

NaNMath.std — Method.

NaNMath.std(A)

Args:

A: A one dimensional array of floating point numbers

Returns:

Returns the standard deviation of a vector A. The algorithm will return an estimator of the generative distribution's standard deviation under the assumption that each entry of v is an IID drawn from that generative distribution. This computation is equivalent to calculating
sqrt(sum((v - mean(v)).^2) / (length(v) - 1)). NaN values are ignored.

Examples:

using NaNMath as nm
nm.std([1., 2., NaN]) # result: 0.7071067811865476

NaNMath.sum — Method.

NaNMath.sum(A)

Args:

A: An array of floating point numbers

Returns:

Returns the sum of all elements in the array, ignoring NaN's.

Examples:

using NaNMath as nm
nm.sum([1., 2., NaN]) # result: 3.0

NaNMath.var — Method.

NaNMath.var(A)

Args:

A: A one dimensional array of floating point numbers

Returns:

Returns the sample variance of a vector A. The algorithm will return an estimator of the generative distribution's variance under the assumption that each entry of v is an IID drawn from that generative distribution. This computation is equivalent to calculating
sum((v - mean(v)).^2) / (length(v) - 1). NaN values are ignored.

Examples:

using NaNMath as nm
nm.var([1., 2., NaN]) # result: 0.5