Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NaiveBayes.HybridNB — Type.

Initialize a HybridNB model with continuous and/or discrete features

Constructors

HybridNB(labels::AbstractVector, kde_names::AbstractVector, discrete_names::AbstractVector)
HybridNB(labels::AbstractVector, kde_names::AbstractVector)
HybridNB(labels::AbstractVector, num_kde::Int, num_discrete::Int)

Arguments

labels : An AbstractVector{Any} of feature labels

kde_names : An AbstractVector{Any} of the names of continuous features

discrete_names : An AbstractVector{Any} of the names of discrete features

num_kde : Number of continuous features

num_discrete : Number of discrete features

NaiveBayes.HybridNB — Method.

HybridNB(labels::Vector{Int64}) -> model_h

HybridNB(labels::Vector{Int64}, AstractString) -> model_h

A constructor for both types of features

NaiveBayes.MultinomialNB — Method.

Multinomial Naive Bayes classifier

classes : array of objects Class names n_vars : Int64 Number of variables in observations alpha : Number (optional, default 1) Smoothing parameter. E.g. if alpha equals 1, each variable in each class is believed to have 1 observation by default

NaiveBayes.NBModel — Type.

Base type for Naive Bayes models. Inherited classes should have at least following fields: ccounts::Dict{C, Int64} - count of ocurrences of each class nobs::Int64 - total number of observations

NaiveBayes.predict_logprobs — Method.

predict_logprobs(m::HybridNB, features_c::Vector{Vector{Float64}, features_d::Vector{Vector{Int})

Return the log-probabilities for each column of X, where each row is the class

NaiveBayes.predict_logprobs — Method.

Predict log probabilities for all classes

NaiveBayes.predict_logprobs — Method.

Predict log probabilities for all classes

NaiveBayes.predict_proba — Method.

predict_proba{V<:Number}(m::HybridNB, f_c::Vector{Vector{Float64}}, f_d::Vector{Vector{Int64}})

Predict log-probabilities for the input features. Returns tuples of predicted class and its log-probability estimate.

NaiveBayes.predict_proba — Method.

Predict logprobs, return tuples of predicted class and its logprob

NaiveBayes.restructure_matrix — Method.

restructure_matrix(M::Matrix) -> V::Dict{Symbol, Vector}

Restructure a matrix as vector of vectors

NaiveBayes.to_matrix — Method.

to_matrix(D::Dict{Symbol, Vector}}) -> M::Matrix

convert a dictionary of vectors into a matrix

NaiveBayes.train — Method.

train(HybridNB, continuous, discrete, labels) -> model2

StatsBase.fit — Method.

fit(m::HybridNB, f_c::Vector{Vector{Float64}}, f_d::Vector{Vector{Int64}}, labels::Vector{Int64})

Train NB model with discrete and continuous features by estimating P(x⃗|c)

StatsBase.fit — Method.

fit(m::HybridNB, f_c::Matrix{Float64}, labels::Vector{Int64})

Train NB model with continuous features only

StatsBase.predict — Method.

predict(m::HybridNB, f_c::Vector{Vector{Float64}}, f_d::Vector{Vector{Int64}}) -> labels

Predict hybrid naive bayes for continuos featuers only

StatsBase.predict — Method.

Predict kde naive bayes for continuos featuers only

NaiveBayes.ePDF — Type.

a wrapper around key value pairs for a discrete probability distribution

NaiveBayes.ePDF — Method.

Constructor of ePDF

NaiveBayes.logprob_x_given_c — Method.

Calculate log P(x|C)

NaiveBayes.logprob_x_given_c — Method.

Calculate log P(x|C)

NaiveBayes.logprob_x_given_c — Method.

Calculate log P(x|C)

NaiveBayes.logprob_x_given_c — Method.

Calculate log P(x|C)

NaiveBayes.num_samples — Method.

compute the number of samples

NaiveBayes.probability — Method.

query the ePDF to get the probability of n

NaiveBayes.sum_log_x_given_c! — Method.

computes log[P(x⃗ⁿ|c)] ≈ ∑ᵢ log[p(xⁿᵢ|c)]