Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NamedTupleTools.delete — Method.

delete(namedtuple, symbol(s)|Tuple) delete(ntprototype, symbol(s)|Tuple)

Generate a namedtuple [ntprototype] from the first arg omitting fields present in the second arg.

see: merge

NamedTupleTools.fieldvalues — Method.

fieldvalues

obtain values assigned to fields of a struct type (in field order)

NamedTupleTools.isprototype — Method.

isprototype( ntprototype )
isprototype( namedtuple  )

Predicate that identifies NamedTuple prototypes.

see: namedtuple

NamedTupleTools.namedtuple — Method.

namedtuple(  name1, name2, ..  )
namedtuple( (name1, name2, ..) )
namedtuple(  namedtuple )

Generate a NamedTuple prototype by specifying or obtaining the fieldnames. The prototype is applied to fieldvalues, giving a completed NamedTuple.

Example

julia> ntprototype = namedtuple( :a, :b, :c )

NamedTuple{(:a, :b, :c),T} where T<:Tuple

julia> nt123 = ntprototype(1, 2, 3)

(a = 1, b = 2, c = 3)

julia> ntAb3 = ntprototype("A", "b", 3)

(a = "A", b = "b", c = 3)

see: isprototype

NamedTupleTools.valtypes — Method.

valtypes( namedtuple )
valtypes( typeof(namedtuple) )

Retrieve the values' types as a tuple.

see: valtype

Base.merge — Method.

merge(namedtuple1, namedtuple2)
merge(nt1, nt2, nt3, ..)

Generate a namedtuple with all fieldnames and values of namedtuple2 and every fieldname of namedtuple1 that does not occur in namedtuple2 along with their values.

see: delete!

Base.valtype — Method.

valtype( namedtuple )

Retrieve the values' types as a typeof(tuple).

see: valtypes