Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NearestNeighborDescent.DescentGraph — Type.

DescentGraph(data::AbstractMatrix, n_neighbors::Integer[, metric::SemiMetric = Euclidean()]; <keyword arguments>)

NearestNeighborDescent.DescentGraph — Method.

DescentGraph(data::Vector{V}, n_neighbors::Integer[, metric::SemiMetric = Euclidean()]; <keyword arguments>)

Build an approximate kNN graph of data using nearest neighbor descent.

Arguments

max_iters::Integer = 10: Limits the number of iterations to refine candidate

nearest neighbors. Higher values trade off speed for accuracy. Note that graph construction may terminate early if little progress is being made.

sample_rate::AbstractFloat = 1.: The sample rate for calculating local joins

around each point. Lower values trade off accuracy for speed.

precision::AbstractFloat = .001: The threshold for early termination,

where precision is "roughly the fraction of true kNN allowed to be missed due to early termination". Lower values take longer but return more accurate results.

NearestNeighborDescent.search — Function.

search(graph::DescentGraph, queries::AbstractMatrix, n_neighbors::Integer, queue_size::Real = 1.)

NearestNeighborDescent.search — Method.

search(graph::DescentGraph, queries::Vector{V}, n_neighbors::Integer, queue_size=1.) -> indices, distances

Search the kNN graph for the nearest neighbors of the points in queries. queue_size controls how large the candidate queue should be as a multiple of n_neighbors. Larger values increase accuracy at the cost of speed.

NearestNeighborDescent._check_tuple — Method.

Check if a tuple exists in a heap at index i, and optionally update its dist. Returns (exists::Bool, updated::Bool)

NearestNeighborDescent._heappush! — Function.

_heappush!(heap::BinaryHeap, tup::NNTuple, max_candidates)

Try to push a neighbor tup to heap. This will fail (return 0) if tup is already in heap, if tup.dist > top(heap).dist. Otherwise return 1. If length(heap) > max_candidates after pushing, pop the largest candidate.

NearestNeighborDescent._neighbors — Function.

Get the neighbors of each point in a KNN graph, graph, as an array of ids.

NearestNeighborDescent._unchecked_heappush! — Function.

Push tup onto heap without checkout if it already exists in the heap.

NearestNeighborDescent.brute_knn — Method.

Brute-force (𝒪(n²)) kNN search algorithm. Returns an KxN array of tuples (index, distance) of the k nearest neighbors for each point in data.

NearestNeighborDescent.brute_search — Method.

brute_search(data, queries, k, metric) -> idxs, dists

Search for the nearest k neighbors of q in queries in data.

NearestNeighborDescent.build_graph — Method.

Return a kNN graph for the input data according to the given metric.

NearestNeighborDescent.deheap_knns — Method.

Remove the k nearest neighbors from each heap in knn_heaps. Return two k x length(knn_heaps) arrays for the indices and distances to each point's kNN.

NearestNeighborDescent.sample_neighbors — Method.

Sample n_neighbors elements from a set of ints 1:npoints. The ints in exclude won't be sampled.