Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NearestNeighbors.BallTree — Method.

BallTree(data [, metric = Euclidean(), leafsize = 10]) -> balltree

Creates a BallTree from the data using the given metric and leafsize.

NearestNeighbors.BruteTree — Method.

BruteTree(data [, metric = Euclidean()) -> brutetree

Creates a BruteTree from the data using the given metric.

NearestNeighbors.DataFreeTree — Method.

DataFreeTree(treetype, data[, reorderbufffer = similar(data), kwargs...]) -> datafreetree

Creates a DataFreeTree which wraps a KDTree or BallTree. Keywords arguments are passed to their respective constructors.

The KDTree or BallTree will be stored without a reference to the underlaying data. injectdata has to be used to re-link them to a data array before use.

NearestNeighbors.KDTree — Method.

KDTree(data [, metric = Euclidean(), leafsize = 10]) -> kdtree

Creates a KDTree from the data using the given metric and leafsize. The metric must be a MinkowskiMetric.

NearestNeighbors.injectdata — Method.

injectdata(datafreetree, data) -> tree

Returns the KDTree/BallTree wrapped by datafreetree, set up to use data for the points data.

NearestNeighbors.inrange — Method.

inrange(tree::NNTree, points, radius [, sortres=false]) -> indices

Find all the points in the tree which is closer than radius to points. If sortres = true the resulting indices are sorted.

NearestNeighbors.knn — Method.

knn(tree::NNTree, points, k [, sortres=false]) -> indices, distances

Performs a lookup of the k nearest neigbours to the points from the data in the tree. If sortres = true the result is sorted such that the results are in the order of increasing distance to the point. skip is an optional predicate to determine if a point that would be returned should be skipped.