Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NetworkLearning.BayesRN — Type.

Naive-Bayes relational neighbour learner (trainable). Calculates neighbourhood likelihoods (i.e. given a vertex's class, the class distribution in its neighbourhood) and uses the resulting information to compute class estimates for each vertex using a Bayesian approach.

NetworkLearning.ClassDistributionRN — Type.

Class-distribution relational neighbour (trainable). Claculates a reference vector (RV) for each class (using the vertex neighbourhood information) and compares vertices to the reference vectors corresponding to each class using a similarity measure.

NetworkLearning.ComputableAdjacency — Type.

Adjacency type where the adjacency information is represented by a . function f and data on which the function can be applied. The result of f(data) has to be either a matrix or a graph.

NetworkLearning.EmptyAdjacency — Type.

Adjacency type indicating the lack of any adjacency information.

NetworkLearning.GibbsSamplingInferer — Type.

Gibbs sapmpling object. Stores the parameters necessary for the algorithm.

NetworkLearning.GraphAdjacency — Type.

Adjacency type where the adjacency information is represented by an AbstractGraph.

NetworkLearning.IterativeClassificationInferer — Type.

Iterative classification object. Stores the parameters necessary for the algorithm.

NetworkLearning.MatrixAdjacency — Type.

Adjacency type where the adjacency information is represented by an AbstractMatrix.

NetworkLearning.NetworkLearnerEnt — Type.

Entity-based network learning model type.

NetworkLearning.NetworkLearnerObs — Type.

Observation-based network learning model type.

NetworkLearning.PartialAdjacency — Type.

Adjacency type where the adjacency information is represented by a function f which can be used to calculate an adjacency matrix or graph, given proper data."

NetworkLearning.RelaxationLabelingInferer — Type.

Relaxation labeling object. Stores the parameters necessary for the algorithm.

NetworkLearning.SimpleRN — Type.

Simple relational neighbour learner. Counts for each vertex how many neighbours from each class are in its neighbourhood.

NetworkLearning.WeightedRN — Type.

Weighted relational neighbour learner. For each vertex, it sums up the estimates from neighboring vertices.

NetworkLearning.add_adjacency! — Method.

add_adjacency!(model, Av)

Function that adds or, creates and then adds, adjacency objects contained in a vector Av to a network learning model. The method is used in 'ouf-of-graph' learning i.e. the training data is not used in the predictions for new samples. In such circumstances, the adjacency structures - graphs, matrices etc. pertinent to the new observations have to be created either using the same functions used in training or, separately supplied, so that relational variables can be created.

NetworkLearning.adjacency — Method.

adjacency([a,data])

Constructs an adjacency object. If a is a AbstractMatrix, AbstractGraph or Tuple, it will return usable adjacencies. If a is a Function or PartialAdjacency, data has to be present. If both arguments are missing, the function returns an EmptyAdjacency.

Examples

julia> using NetworkLearning, LightGraphs

julia> A = [0 1 0; 1 0 0; 0 0 0];

julia> Am = adjacency(A)
Matrix adjacency, 3 obs

julia> Ag = adjacency(Graph(A))
Graph adjacency, 3 obs

julia> Ac = adjacency(x->x,A)
Computable adjacency

NetworkLearning.adjacency_graph — Method.

adjacency_graph(a)

Returns an adjacency graph computed from the adjacency information of a.

NetworkLearning.adjacency_matrix — Method.

adjacency_matrix(a)

Returns an adjacency matrix computed from the adjacency information of a.

NetworkLearning.adjacency_obs — Method.

adjacency_obs(A, r, obsdim)

Selects a range r::UnitRange of observations from the matrix A, along the dimension obsdim::LearnBase.ObsDimension.

NetworkLearning.fit — Method.

fit(::Type{NetworkLearnerEnt}, X, update, Adj, fl_train, fl_exec, fr_train, fr_exec [;kwargs])

Training method for the entity-based network learning framework.

Arguments

Xo::AbstractMatrix initial estimates for the entities

update::BitVector mask that indicates wether estimates can be updated (true value) or not (false value); false values

generally can be associated with estimates of training samples

Adj::Vector{AbstractAdjacency} a vector containing the entity relational structures (adjacency objects)

fr_train relational model training function; can be anything that suports the call fr_train((Xr,y)) where y = f_targets(Xo)

fr_exec relational model prediction function; can be anything that suports the call fr_exec(Mr,Xr) where Mr = fr_train((Xr,y))

and Xr is a dataset of relational variables generated by the relational learner using the estimates Xo and the adjacency structures.

Keyword arguments

priors::Vector{Float64} class priors (if applicable)

learner::Symbol relational learner (i.e. variable generator); available options :rn, :wrn, :bayesrn and :cdrn (default :wrn)

inference::Symbol collective inference method; available options :rl, :ic and :gs (default :rl)

normalize::Bool whether to normalize the relational variables per-entity to the L1 norm (default true)

f_targets::Function function that extracts targets from estimates generated by the local/relational models

(default f_targets = x->MLDataPattern.targets(indmax,x))

obsdim::Int observation dimension (default 2)

tol::Float64 maximum admissible mean estimate error for collective inference convergence (default 1e-6)

κ::Float64 relaxation labeling starting constant, used if learner == :rl (default 1.0)

α::Float64 relaxation labeling decay constant, used if learner == :rl (default 0.99)

maxiter::Int maximum number of iterations for collective inference (default 100)

bratio::Float64 percentage of iterations i.e. maxiter used for Gibbs sampling burn-in (default 0.1)

NetworkLearning.fit — Method.

fit(::Type{NetworkLearnerObs}, X, y, Adj, fl_train, fl_exec, fr_train, fr_exec [;kwargs])

Training method for the observation-based network learning framework.

Arguments

X::AbstractMatrix training data (used by fl_train, fl_exec; if use_local_data==true, it is also used by fr_train)

Y::AbstractAray data targets (used by fl_train, fr_train)

Adj::Vector{AbstractAdjacency} a vector containing the observation relational structures (adjacency objects)

fl_train local model training 'function'; can be anything that supports the call fl_train((X,y))

fl_exec local model prediction 'function'; can be anything that supports the call fl_exec(Ml,X) where Ml = fl_train((X,y))

fr_train relational model training function; can be anything that suports the call fr_train((Xr,y))

fr_exec relational model prediction function; can be anything that suports the call fr_exec(Mr,Xr) where Mr = fr_train((Xr,y))

and Xr is a dataset of relational variables generated by the relational learner using the results of the local model prediction function and adjacency structures.

Keyword arguments

priors::Vector{Float64} class priors (if applicable)

learner::Symbol relational learner (i.e. variable generator); available options :rn, :wrn, :bayesrn and :cdrn (default :wrn)

inference::Symbol collective inference method; available options :rl, :ic and :gs (default :rl)

normalize::Bool whether to normalize the relational variables per-observation to the L1 norm (default true)

use_local_data::Bool whether the relational model should use the local data provided (i.e. in X) (default true)

f_targets::Function function that extracts targets from estimates generated by the local/relational models

(default f_targets = x->MLDataPattern.targets(indmax,x))

obsdim::Int observation dimension (default 2)

tol::Float64 maximum admissible mean estimate error for collective inference convergence (default 1e-6)

κ::Float64 relaxation labeling starting constant, used if learner == :rl (default 1.0)

α::Float64 relaxation labeling decay constant, used if learner == :rl (default 0.99)

maxiter::Int maximum number of iterations for collective inference (default 100)

bratio::Float64 percentage of iterations i.e. maxiter used for Gibbs sampling burn-in (default 0.1)

NetworkLearning.fit — Method.

Training method for the network learning framework. This method should not be called directly.

NetworkLearning.fit — Method.

Training method for the network learning framework. This method should not be called directly.

NetworkLearning.infer! — Method.

Function that calls collective inference using the information in contained in the entity-based network learner

NetworkLearning.intdim — Method.

intdim(::LearnBase.ObsDimension)

Returns the integer associated to a dimension object i.e. intdim(ObsDim.Constant{3}) returns 3. The function is designed to work on matices so intdim(::ObsDim.Last) will return 2.

NetworkLearning.matrix_prealloc — Method.

matrix_prealloc(no, nv, obsdim, val)

Returns a Matrix{T} filled with values equal to val::T, having the size no (number of observations) on dimension obsdim and nv (number of variables) in the other dimension.

NetworkLearning.oppdim — Method.

oppdim(::LearnBase.ObsDimension)

Returns the other dimension for a matrix i.e. if provided ObsDim.Constant{1} returns ObsDim.Constant{2}.

NetworkLearning.predict! — Method.

In-place prediction method for the network learning framework.

NetworkLearning.predict — Method.

Prediction method for the network learning framework.

NetworkLearning.strip_adjacency — Method.

strip_adjacency(a)

Function that removes adjancency information (i.e. matrix, graph or other data) from adjacency objects and returns a PartialAdjacency that can be used in conjunction with the adjacency function to build a new adjacency object.

Examples

julia> using NetworkLearning, LightGraphs

julia> A = [0 1 0; 1 0 0; 0 0 0];

julia> Am = adjacency(A)
Matrix adjacency, 3 obs

julia> Ac = adjacency(x->x,A)
Computable adjacency

julia> Sm = strip_adjacency(Am)
Partial adjacency, not computable

julia> adjacency(Sm, A) # Sm has to be used with a matrix
Matrix adjacency, 3 obs

julia> Sc = strip_adjacency(Ac)
Partial adjacency, not computable

julia> adjacency(Sc, A) # Sc can be use with any data; calls f(A) where f = x->x
Matrix adjacency, 3 obs

NetworkLearning.update_adjacency! — Method.

update_adjacency!(a,f_update)

Function that updates the data of an adjacency object. a has to be a MatrixAdjacency or GraphAdjacency while f_update has to of the form f_update = x->update_function!(x).

Examples

julia> using NetworkLearning, LightGraphs

julia> A = [0 1 0; 1 0 0; 0 0 0];

julia> Am = adjacency(A) Matrix adjacency, 3 obs

julia> updatefunction!(X,x,y) = begin X[x,y] += 1 X[y,x] += 1 return X end updatefunction! (generic function with 2 methods)

julia> fupdate(x,y) = X->updatefunction!(X,x,y) f_update (generic function with 1 method)

julia> for i in 1:3 updateadjacency!(Am, fupdate(1,3)) # call function three times end

julia> adjacency_matrix(Am) 3×3 Array{Int64,2}: 0 1 3 1 0 0 3 0 0

NetworkLearning.NetworkLearnerState — Type.

Entity-based network learning model state. It consists of an Array with estimates and a an update mask in the form of a BitVector indicating which observation estimates are to be updated (the ones that are not updated are considered training/stable observations).

NetworkLearning.nvars — Method.

Returns the number of variables given a data object which supports the nobs function. The data object must ideally present two dimensions i.e. matrix.