Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NormalizeQuantiles.normalizeQuantiles — Method.

qnmatrix::Array{Float64} function normalizeQuantiles(matrix::AbstractArray)

Calculate the quantile normalized data for the input matrix

Parameter: matrix::AbstractArray The input data as an array of values interpreted as matrix(rows,columns)

Return value: qnmatrix::Array{Float64} The quantile normalized data as Array{Float64}

Example:

using NormalizeQuantiles

array = [ 3.0 2.0 1.0 ; 4.0 5.0 6.0 ; 9.0 7.0 8.0 ; 5.0 2.0 8.0 ]
qn = normalizeQuantiles(array)

row = 2
column = 2
array=convert(Array{Any},array)
array[row,column] = missing
qn = normalizeQuantiles(array)

NormalizeQuantiles.sampleRanks — Method.

(Array{Union{Missing,Int}},Dict{Int,Array{Int}}) sampleRanks(array::AbstractArray;tiesMethod::qnTiesMethods=tmMin,naIncreasesRank=false,resultMatrix=false)

Calculate ranks of the values of a given vector.

Parameters:

array: the input array
tiesMethod: the method how ties (equal values) are treated
   possible values: tmMin tmMax tmOrder tmReverse tmRandom tmAverage
   default is tmMin
naIncreasesRank: if true than any NA increases the following ranks by 1
resultMatrix: if true than return a dictionary of rank keys and array of indices values

If the input array can be altered use sampleRanks!(matrix::AbstractArray) to be more memory efficient.

Example:

using NormalizeQuantiles

a = [ 5.0 2.0 4.0 3.0 1.0 ]

(r,m)=sampleRanks(a)

(r,m)=sampleRanks(a,tiesMethod=tmMin,naIncreasesRank=false,resultMatrix=true)

r is the vector of ranks.

m is a dictionary with rank as keys and as value the indices of all values of this rank.