Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

NumericalIntegration.cumul_integrate — Method.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::Trapezoidal)

Use Trapezoidal rule.

NumericalIntegration.cumul_integrate — Method.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEvenFast)

Use Trapezoidal rule, assuming evenly spaced vector x. Unsafe method: no bound checking.

NumericalIntegration.cumul_integrate — Method.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEven)

Use Trapezoidal rule, assuming evenly spaced vector x.

NumericalIntegration.cumul_integrate — Method.

cumul_integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalFast)

Use Trapezoidal rule. Unsafe method: no bound checking. This is the default when no method is supplied.

NumericalIntegration.cumul_integrate — Method.

cumul_integrate(x::AbstractVector, y::AbstractMatrix, method; dims=2)

When y is an array, compute integral along each dimension specified by dims (default 2: columns)

NumericalIntegration.cumul_integrate — Method.

cumul_integrate(x,y...)

Compute cumulative numerical integral of y(x) from x=x[1] to x=x[end]. Return a vector with elements of the same type as the input. If not method is supplied, use TrapezdoialFast.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractVector, ::Trapezoidal)

Use Trapezoidal rule.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractVector, ::SimpsonEven)

Use Simpson's rule, assuming evenly spaced vector x. Unsafe method: no bound checking.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractVector, ::SimpsonEven)

Use Simpson's rule, assuming evenly spaced vector x.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEvenFast)

Use Trapezoidal rule, assuming evenly spaced vector x. Unsafe method: no bound checking.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalEven)

Use Trapezoidal rule, assuming evenly spaced vector x.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractVector, ::TrapezoidalFast)

Use Trapezoidal rule. Unsafe method: no bound checking. This is the default when no method is supplied.

NumericalIntegration.integrate — Method.

integrate(x::AbstractVector, y::AbstractMatrix, method; dims=2)

When y is an array, compute integral along dimension specified by dims (default 2: columns).

NumericalIntegration.integrate — Method.

integrate(x,y...)

Compute numerical integral of y(x) from x=x[1] to x=x[end]. Return a scalar of the same type as the input. If not method is supplied, use TrapezdoialFast.