Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

OBOParse.Ontology — Type.

The collection of all ontology terms and their relations.

OBOParse.Term — Type.

Ontology term.

The Term object is a node in the direct acyclic ontology graph. Its outgoing and incoming edges represent the relations with the other nodes and could be retrieved by

relationship(term, sym)

and

rev_relationship(term, sym)

respectively, where sym is the relationship annotation (e.g. :part_of, :is_a, :regulates).

OBOParse.Stanza — Type.

Represents one entry in the OBO file, e.g.

[Term]
id: GO:0000002
namespace: biological_process
def: BBB
name: two

is stored as Stanza with Typ = "Term", id = "GO:0000002" and tagvalues = Dict("id" => "GO:0000002", "namespace" => ["biological_process"], "def" => ["BBB"], "name" => "two").

OBOParse.Typedef — Type.