Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

OMETIFF.ImageSlice — Type.

An ImageSlice struct contains all the information from a TiffData element in the OME-XML.

OMETIFF.build_axes — Method.

build_axes(omexml::EzXML.Node)

Returns an array of ints with dimension sizes and an array of AxisArrays.Axis objects both in XYZCT order given the Pixels node of the OME-XML document

OMETIFF.check_bswap — Method.

check_bswap(io::Union{Stream, IOStream})

Check endianness of TIFF file to see if we need to swap bytes

OMETIFF.extract_filename — Method.

Extract the name of the file backing a stream

OMETIFF.load_master_xml — Method.

load_master_xml(file::TiffFile)

Loads the master OME-XML file from file or from a linked file.

OMETIFF.loadxml — Method.

loadxml(file::TiffFile)

Extract the OME-XML embedded in the TiffFile file.

OMETIFF.myendian — Method.

myendian()

Returns the TIFF endian byte order expected if its endianness matches the host's

OMETIFF.next — Method.

next(file::TiffFile)

Loads the next IFD in file as determined by file.loc and returns a list of the strip offsets to load the data stored in this IFD. If the IFD has not been loaded yet then this function loads it from disk, otherwise it returns the IFD location from memory.

OMETIFF.read_tiffdata — Method.

read_tiffdata(tiffdata::EzXML.Node, files::Dict{String, TiffFile}, orig_file::TiffFile)

Reads a Tiff Data entry and generates an ImageSlice object from it. Keeps track of all relevant files.

OMETIFF.reset — Method.

reset(file)

Resets the IFD pointer to the beginning of the file. This is helpful for when loading the next position from files that may already be fully loaded.

OMETIFF.to_symbol — Method.

to_symbol(input::String)

Cleans up input string and converts it into a symbol