Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

PlotlyBase.savefig — Method.

savefig(p::Plot, fn::AbstractString; format=nothing, scale=nothing,
width=nothing, height=nothing)

Save a plot p to a file named fn. If format is given and is one of (png, jpeg, webp, svg, pdf, eps), it will be the format of the file. By default the format is guessed from the extension of fn. scale sets the image scale. width and height set the dimensions, in pixels. Defaults are taken from p.layout, or supplied by plotly