Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ObjectStores.ObjectStore — Type.

Required fields:

id::String; # From Authorization.AbstractClient

id2permission::Dict{String, Permission}; # Resource ID => Permission (from Authorization.AbstractClient)

idpattern2permission::Dict{Regex, Permission}; # Resource ID pattern => Permission (from Authorization.AbstractClient)

type2permission::Dict{DataType, Permission}; # Resource type => Permission (from Authorization.AbstractClient)

rootbucketID::String # ID of root bucket # Specific to ObjectStores

Base.delete! — Method.

Delete object. If successful return nothing, else return an error message as a String.

Base.getindex — Method.

Read object. If successful return the value, else @warn the error message and return nothing.

Base.setindex! — Method.

Create/update object. If successful return nothing, else return an error message as a String.

ObjectStores.createbucket! — Function.

Create bucket. If successful return nothing, else return an error message as a String.

ObjectStores.deletebucket! — Function.

Delete bucket. If successful return nothing, else return an error message as a String.

ObjectStores.isbucket — Method.

Returns true if name refers to a bucket.

ObjectStores.isobject — Method.

Returns true if name refers to an object.

ObjectStores.listcontents — Function.

List the contents of the bucket. If successful return the value, else @warn the error message and return nothing.