Internals

Package Internals

Documentation of the OpenEphysLoader.jl internals.

OpenEphysLoader.BlockBuffer — Type.

Type to buffer continuous file contents

OpenEphysLoader.BlockHeader — Type.

Represents the header of each data block

OpenEphysLoader.DataBlock — Type.

Represents the entirety of a data block

OpenEphysLoader.MATLABdata — Type.

Abstract class for representing matlab code fragments

OpenEphysLoader.MATfloat — Type.

type for representing Matlab floatingpoint numbers

OpenEphysLoader.MATint — Type.

Type for representing Matlab integers

OpenEphysLoader.MATstr — Type.

Type for representing Matlab strings

OpenEphysLoader.OESignalTree — Type.

OESignalTree{T<:OEProcessor}(chain_e::LightXML.XMLElement, [recording_names::Set])

Signal tree for recording processors. Since OpenEphysLoader currently on works on .continuous file types, this will search for the first OERhythmProcessor and make a signal tree up to that point.

Construct with a XML signalchain element, and a set of processor names that are valid recording nodes.

See Tree for field information.

OpenEphysLoader.SignalNode — Type.

SignalNode{T<:OEProcessor}

Node type for OEProcessor signal chain, subtype of TreeNode.

See TreeNode for information on fields.

OpenEphysLoader.Tree — Type.

Tree{T}

Abstract type for tree structure, with type T content.

Contains a group of TreeNode in the single required field:

Required Fields

nodes Indexable list of TreeNode elements.

OpenEphysLoader.TreeNode — Type.

TreeNode{T}

Abstract node type for tree structure, with type T content.

Subtypes must have the following fields:

Required Fields

content T content of node.

parent Int ID of parent node

children Vector{Int} IDs of children node

OpenEphysLoader.add_continuous_meta! — Method.

Add data from Continuous_Data.openephys to OESettings from settings.xml

OpenEphysLoader.channel_arr — Method.

Parse XML Element PROCESSOR and recover channel metadata.

OpenEphysLoader.check_filesize — Method.

Check that file could be comprised of header and complete data blocks

OpenEphysLoader.convert_block! — Method.

Convert the wacky data format in OpenEphys continuous files

OpenEphysLoader.find_matching_proc — Method.

Find id of processor in OESignalTree that matches id of XML processor element

OpenEphysLoader.matread — Method.

read a Matlab source line

OpenEphysLoader.parseline — Function.

Parse a line of Matlab source code

OpenEphysLoader.parseto — Function.

Convert a string to the desired type

OpenEphysLoader.prepare_block! — Method.

Load data block if necessary

OpenEphysLoader.read_into! — Function.

Read file data block into data block buffer

OpenEphysLoader.read_into! — Method.

Read block header into header buffer

OpenEphysLoader.seek_to_block — Method.

Move io to data block

OpenEphysLoader.verify_tail! — Method.

Verify end of block marker