Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

OpenIDConnect.OIDCCtx — Type.

Holds an OpenID Connect context that can be used in subsequent OpenID request flows. The context holds request states, and configuration options.

OpenIDConnect.flow_get_authorization_code — Method.

Given the params from the redirected response from the authentication request, extract the authorization code. See sections 3.1.2.5 and 3.1.2.6 of https://openid.net/specs/openid-connect-core-1_0.html.

Returns the authorization code on success. Returns one of APIError or AuthServerError on failure.

OpenIDConnect.flow_get_token — Method.

Token Request. Given the authorization code obtained, invoke the token end point and obtain an idtoken, accesstoken, refreshtoken. See section 3.1.3.1 of https://openid.net/specs/openid-connect-core-10.html.

Returns a JSON object containing tokens on success. Returns a AuthServerError or APIError object on failure.

OpenIDConnect.flow_refresh_token — Method.

Token Refresh. Given the refresh code obtained, invoke the token end point and obtain new tokens. See section 12 of https://openid.net/specs/openid-connect-core-1_0.html.

Returns a JSON object containing tokens on success. Returns a AuthServerError or APIError object on failure.

OpenIDConnect.flow_request_authorization_code — Method.

Authentication request. Uses the authorization code flow. Acceptable optional args as listed in section 3.1.2.1 of specifications (https://openid.net/specs/openid-connect-core-1_0.html)

Returns a String with the redirect URL. Caller must perform the redirection.

OpenIDConnect.flow_validate_id_token — Method.

Validate an OIDC token. Validates both the structure and signature. See section 3.1.3.7 of https://openid.net/specs/openid-connect-core-1_0.html

OpenIDConnect.APIError — Type.

API calling error detected by this library

OpenIDConnect.AuthServerError — Type.

Error returned from OpenID server See section 3.1.2.6 of https://openid.net/specs/openid-connect-core-1_0.html