Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

OptimizationProblems.AMPGO02 — Method.

Univariate multimodal minimization problem AMPGO02

OptimizationProblems.AMPGO03 — Method.

Univariate multimodal minimization problem AMPGO03

OptimizationProblems.AMPGO04 — Method.

Univariate multimodal minimization problem AMPGO04

OptimizationProblems.AMPGO05 — Method.

Univariate multimodal minimization problem AMPGO05

OptimizationProblems.AMPGO06 — Method.

Univariate multimodal minimization problem AMPGO06

OptimizationProblems.AMPGO07 — Method.

Univariate multimodal minimization problem AMPGO07

OptimizationProblems.AMPGO08 — Method.

Univariate multimodal minimization problem AMPGO08

OptimizationProblems.AMPGO09 — Method.

Univariate multimodal minimization problem AMPGO09

OptimizationProblems.AMPGO10 — Method.

Univariate multimodal minimization problem AMPGO10

OptimizationProblems.AMPGO11 — Method.

Univariate multimodal minimization problem AMPGO11

OptimizationProblems.AMPGO12 — Method.

Univariate multimodal minimization problem AMPGO12

OptimizationProblems.AMPGO13 — Method.

Univariate multimodal minimization problem AMPGO13

OptimizationProblems.AMPGO14 — Method.

Univariate multimodal minimization problem AMPGO14

OptimizationProblems.AMPGO15 — Method.

Univariate multimodal minimization problem AMPGO15

OptimizationProblems.AMPGO18 — Method.

Univariate multimodal minimization problem AMPGO18

OptimizationProblems.AMPGO20 — Method.

Univariate multimodal minimization problem AMPGO20

OptimizationProblems.AMPGO21 — Method.

Univariate multimodal minimization problem AMPGO21

OptimizationProblems.AMPGO22 — Method.

Univariate multimodal minimization problem AMPGO22

OptimizationProblems.Dus2_1 — Method.

Univariate unimodal minimization problem Dus2_1

OptimizationProblems.Dus2_3 — Method.

Univariate unimodal minimization problem Dus2_3

OptimizationProblems.Dus2_9 — Method.

Univariate multimodal minimization problem Dus2_9

OptimizationProblems.Duscube — Method.

Univariate multimodal minimization problem Duscube

OptimizationProblems.Shpak1 — Method.

Univariate multimodal minimization problem Shpak1

OptimizationProblems.Shpak2 — Method.

Univariate multimodal minimization problem Shpak2

OptimizationProblems.Shpak3 — Method.

Univariate multimodal minimization problem Shpak3

OptimizationProblems.Shpak4 — Method.

Univariate multimodal minimization problem Shpak4

OptimizationProblems.Shpak5 — Method.

Univariate multimodal minimization problem Shpak5

OptimizationProblems.Shpak6 — Method.

Univariate multimodal minimization problem Shpak6

OptimizationProblems.arglina — Function.

Linear function with n parameters and m observations - full rank

OptimizationProblems.arglinb — Function.

Linear function with n parameters and m observations - rank 1

OptimizationProblems.arglinc — Function.

Linear function with n parameters and m observations - rank 1, zero columns and rows

OptimizationProblems.arwhead — Function.

Arrow head model in size n

OptimizationProblems.bdqrtic — Function.

Banded quartic model in size n

OptimizationProblems.beale — Method.

Beale Model

OptimizationProblems.brownbs — Method.

Brownbs Model

OptimizationProblems.brownden — Function.

Brown and Dennis function

OptimizationProblems.broydn7d — Function.

Broyden 7-diagonal model in size n

OptimizationProblems.brybnd — Function.

Broyden banded model in size n

OptimizationProblems.bt1 — Method.

BT1 Model

OptimizationProblems.chainwoo — Function.

The chained Woods function in size n, a variant on the Woods function

OptimizationProblems.cliff — Method.

The 'cliff problem' in 2 variables

OptimizationProblems.clplatea — Function.

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.clplateb — Function.

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.clplatec — Function.

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.cosine — Function.

The cosine function in size n

OptimizationProblems.cragglvy — Function.

The extented Cragg and Levy function in size n

OptimizationProblems.curly — Function.

Curly function in size n with semi-bandwidth b

OptimizationProblems.curly10 — Function.

Curly function in size n with semi-bandwidth 10

OptimizationProblems.curly20 — Function.

Curly function in size n with semi-bandwidth 20

OptimizationProblems.curly30 — Function.

Curly function in size n with semi-bandwidth 30

OptimizationProblems.dixmaane — Function.

Dixon-Maany function in size n (version E by default)

OptimizationProblems.dixmaanf — Function.

Dixon-Maany function in size n (version F)

OptimizationProblems.dixmaang — Function.

Dixon-Maany function in size n (version G)

OptimizationProblems.dixmaanh — Function.

Dixon-Maany function in size n (version H)

OptimizationProblems.dixmaani — Function.

Dixon-Maany function in size n (version I by default)

OptimizationProblems.dixmaanj — Function.

Dixon-Maany function in size n (version J)

OptimizationProblems.dixmaank — Function.

Dixon-Maany function in size n (version K)

OptimizationProblems.dixmaanl — Function.

Dixon-Maany function in size n (version L)

OptimizationProblems.dixmaanm — Function.

Dixon-Maany function in size n (version M by default)

OptimizationProblems.dixmaann — Function.

Dixon-Maany function in size n (version N)

OptimizationProblems.dixmaano — Function.

Dixon-Maany function in size n (version O)

OptimizationProblems.dixmaanp — Function.

Dixon-Maany function in size n (version P)

OptimizationProblems.dixon3dq — Function.

Dixon's tridiagonal quadratic.

OptimizationProblems.dqdrtic — Function.

Diagonal quadratic problem

OptimizationProblems.dqrtic — Function.

Diagonal quartic model in size n

OptimizationProblems.edensch — Function.

Extended Dennis-Schnabel model in size n

OptimizationProblems.eg2 — Function.

model in size n

OptimizationProblems.engval1 — Function.

The Engval1 model in size n

OptimizationProblems.genrose — Function.

Generalized Rosenbrock model in size n

OptimizationProblems.genrose_nash — Function.

Nash's variant of genrose() in size n

OptimizationProblems.hs1 — Method.

HS1 model

OptimizationProblems.hs10 — Method.

HS10 model

OptimizationProblems.hs100 — Method.

HS100 model

OptimizationProblems.hs101 — Method.

HS101 model

OptimizationProblems.hs102 — Method.

HS102 model

OptimizationProblems.hs103 — Method.

HS103 model

OptimizationProblems.hs104 — Method.

HS104 model

OptimizationProblems.hs105 — Method.

HS105 model

OptimizationProblems.hs106 — Method.

HS106 model

OptimizationProblems.hs107 — Method.

HS107 model

OptimizationProblems.hs108 — Method.

HS108 model

OptimizationProblems.hs109 — Method.

HS109 model

OptimizationProblems.hs11 — Method.

HS11 model

OptimizationProblems.hs110 — Method.

HS110 model

OptimizationProblems.hs111 — Method.

HS111 model

OptimizationProblems.hs112 — Method.

HS112 model

OptimizationProblems.hs113 — Method.

HS113 model

OptimizationProblems.hs114 — Method.

HS114 model

OptimizationProblems.hs116 — Method.

HS116 model

OptimizationProblems.hs117 — Method.

HS117 model

OptimizationProblems.hs118 — Method.

HS118 model

OptimizationProblems.hs119 — Method.

HS119 model

OptimizationProblems.hs12 — Method.

HS12 model

OptimizationProblems.hs13 — Method.

HS13 model

OptimizationProblems.hs14 — Method.

HS14 model

OptimizationProblems.hs15 — Method.

HS15 model

OptimizationProblems.hs16 — Method.

HS16 model

OptimizationProblems.hs17 — Method.

HS17 model

OptimizationProblems.hs18 — Method.

HS18 model

OptimizationProblems.hs19 — Method.

HS19 model

OptimizationProblems.hs2 — Method.

HS2 model

OptimizationProblems.hs20 — Method.

HS20 model

OptimizationProblems.hs21 — Method.

HS21 model

OptimizationProblems.hs22 — Method.

HS22 model

OptimizationProblems.hs23 — Method.

HS23 model

OptimizationProblems.hs24 — Method.

HS24 model

OptimizationProblems.hs25 — Method.

HS25 model

OptimizationProblems.hs26 — Method.

HS26 Model

OptimizationProblems.hs27 — Method.

HS27 Model

OptimizationProblems.hs28 — Method.

HS28 model

OptimizationProblems.hs29 — Method.

HS29 model

OptimizationProblems.hs3 — Method.

HS3 model

OptimizationProblems.hs30 — Method.

HS30 model

OptimizationProblems.hs31 — Method.

HS31 model

OptimizationProblems.hs32 — Method.

HS32 model

OptimizationProblems.hs33 — Method.

HS33 model

OptimizationProblems.hs34 — Method.

HS34 model

OptimizationProblems.hs35 — Method.

HS35 model

OptimizationProblems.hs36 — Method.

HS36 model

OptimizationProblems.hs37 — Method.

HS37 model

OptimizationProblems.hs38 — Method.

HS38 model

OptimizationProblems.hs39 — Method.

HS39 model

OptimizationProblems.hs4 — Method.

HS4 model

OptimizationProblems.hs40 — Method.

HS40 model

OptimizationProblems.hs41 — Method.

HS41 model

OptimizationProblems.hs42 — Method.

HS42 model

OptimizationProblems.hs43 — Method.

HS43 model

OptimizationProblems.hs44 — Method.

HS44 model

OptimizationProblems.hs45 — Method.

HS45 model

OptimizationProblems.hs46 — Method.

HS46 model

OptimizationProblems.hs47 — Method.

HS47 model

OptimizationProblems.hs48 — Method.

HS48 model

OptimizationProblems.hs49 — Method.

HS49 model

OptimizationProblems.hs5 — Method.

HS5 model

OptimizationProblems.hs50 — Method.

HS50 model

OptimizationProblems.hs51 — Method.

HS51 model

OptimizationProblems.hs52 — Method.

HS52 model

OptimizationProblems.hs53 — Method.

HS53 model

OptimizationProblems.hs54 — Method.

HS54 model

OptimizationProblems.hs55 — Method.

HS55 model

OptimizationProblems.hs56 — Method.

HS56 model

OptimizationProblems.hs57 — Method.

HS57 model

OptimizationProblems.hs59 — Method.

HS59 model

OptimizationProblems.hs6 — Method.

HS6 model

OptimizationProblems.hs60 — Method.

HS60 model

OptimizationProblems.hs61 — Method.

HS61 model

OptimizationProblems.hs62 — Method.

HS62 model

OptimizationProblems.hs63 — Method.

HS63 model

OptimizationProblems.hs64 — Method.

HS64 model

OptimizationProblems.hs65 — Method.

HS65 model

OptimizationProblems.hs66 — Method.

HS66 model

OptimizationProblems.hs7 — Method.

HS7 model

OptimizationProblems.hs70 — Method.

HS70 model

OptimizationProblems.hs71 — Method.

HS71 model

OptimizationProblems.hs72 — Method.

HS72 model

OptimizationProblems.hs73 — Method.

HS73 model

OptimizationProblems.hs74 — Method.

HS74 model

OptimizationProblems.hs75 — Method.

HS75 model

OptimizationProblems.hs76 — Method.

HS76 model

OptimizationProblems.hs77 — Method.

HS77 model

OptimizationProblems.hs78 — Method.

HS78 model

OptimizationProblems.hs79 — Method.

HS79 model

OptimizationProblems.hs8 — Method.

HS8 model

OptimizationProblems.hs80 — Method.

HS80 model

OptimizationProblems.hs81 — Method.

HS81 model

OptimizationProblems.hs83 — Method.

HS83 model

OptimizationProblems.hs84 — Method.

HS84 model

OptimizationProblems.hs86 — Method.

HS86 model

OptimizationProblems.hs9 — Method.

HS9 Model

OptimizationProblems.hs93 — Method.

HS93 model

OptimizationProblems.hs95 — Method.

HS95 model

OptimizationProblems.hs96 — Method.

HS96 model

OptimizationProblems.hs97 — Method.

HS97 model

OptimizationProblems.hs98 — Method.

HS98 model

OptimizationProblems.meyer3 — Method.

Meyer function

OptimizationProblems.nasty — Method.

Nasty problem

OptimizationProblems.palmer1c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer1d — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer2c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer3c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer4c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer5c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer5d — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer6c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer7c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.palmer8c — Method.

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.penalty2 — Function.

A penalty problem by Gill, Murray and Pitfield in size 'n'

OptimizationProblems.penalty3 — Function.

A penalty problem by Gill, Murray and Pitfield in size 'n'

OptimizationProblems.powellsg — Function.

The extended Powell singular problem in size 'n'

OptimizationProblems.power — Function.

The Power problem by Oren.

OptimizationProblems.quartc — Function.

A simple quartic function.

OptimizationProblems.rosenbrock — Method.

Classic Rosenbrock problem

OptimizationProblems.sbrybnd — Function.

Broyden banded system of nonlinear equations in size 'n'

OptimizationProblems.schmvett — Function.

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.scosine — Function.

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.sinquad — Function.

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.sparsine — Function.

A sparse problem involving sine functions in size 'n'

OptimizationProblems.sparsqur — Function.

A sparse quartic problem in size 'n'

OptimizationProblems.srosenbr — Function.

The separable extension of Rosenbrock's function 'n'

OptimizationProblems.tointgss — Function.

Toint's Gaussian problem in size 'n'

OptimizationProblems.tquartic — Function.

A quartic function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.tridia — Function.

Shanno's TRIDIA quadratic tridiagonal problem.

OptimizationProblems.vardim — Function.

Variable dimension problem.

OptimizationProblems.woods — Function.

The extended Woods problem n

OptimizationProblems.woods — Method.

The Woods function.

OptimizationProblems.zangwil3 — Method.

Zangwil3 Model