Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Base.get — Method.

get(opt::OptData)

Get data from opt. Throw an exception if no data has been pushed to opt before.

Base.push! — Method.

push!(opt::OptData, ::Type{T}, args...) where T

Construct an object of type T from args and push it to opt.

Base.push! — Method.

push!(opt::OptData{T}, data::T) where T

Push data of type T to opt.

OptionalData.isavailable — Method.

isavailable(opt::OptData)

Check whether data has been pushed to opt.