Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Oracle.CommonCreateParams — Type.

Safe version of OraCommonCreateParams

Oracle.ConnCreateParams — Type.

Safe version of OraConnCreateParams

Oracle.Connection — Type.

Connection handles are used to represent connections to the database. These can be standalone connections created by calling the function dpiConncreate() or acquired from a session pool by calling the function dpiPoolacquireConnection(). They can be closed by calling the function dpiConnclose() or releasing the last reference to the connection by calling the function dpiConnrelease(). Connection handles are used to create all handles other than session pools and context handles.

Oracle.EncodingInfo — Type.

Mirrors ODPI-C's OraEncodingInfo struct, but using Julia types.

Oracle.NativeValue — Type.

High-level type as an aggregation of dpiNativeTypeNum and dpiData.

Oracle.OraCommonCreateParams — Method.

This function may be unsafe: safe_params must outlive the OraCommonCreateParams generated by this function.

Oracle.OraConnCreateParams — Type.

This structure is used for creating connections to the database, whether standalone or acquired from a session pool. All members are initialized to default values using the dpiContext_initConnCreateParams() function. Care should be taken to ensure a copy of this structure exists only as long as needed to create the connection since it can contain a clear text copy of credentials used for connecting to the database.

Oracle.OraConnCreateParams — Method.

This function may be unsafe: safe_params must outlive the ConnCreateParams generated by this function.

Oracle.OraData — Type.

This structure is used for passing data to and from the database for variables and for manipulating object attributes and collection values.

Oracle.OraPoolCreateParams — Type.

This structure is used for creating session pools, which can in turn be used to create connections that are acquired from that session pool.

All members are initialized to default values using the dpiContext_initPoolCreateParams() function.

Oracle.StmtInfo — Type.

High-level version for OraStmtInfo using Bool Julia type.

Oracle.execute! — Method.

execute!(stmt::Stmt; exec_mode::dpiExecMode=ORA_MODE_EXEC_DEFAULT) :: UInt32

Returns the number of columns which are being queried. If the statement does not refer to a query, the value is set to 0.

Oracle.fetch! — Method.

fetch!(stmt::Stmt)

Fetches a single row from the statement.

Oracle.row_count — Method.

Number of affected rows in a DML statement.