Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

OrdinaryDiffEq.NLFunctional — Method.

(S::NLFunctional)(integrator) -> (z, η, iter, fail_convergence)

Perform functional iteration that is used by implicit methods, where z is the solution, η is used to measure the iteration error (see [HW96]), iter is the number of iteration, and fail_convergence reports whether the algorithm succeed. It solves

\[G(z) = dt⋅f(tmp + γ⋅z, p, t+c⋅h) z = G(z)\]

by iterating

\[zᵏ⁺¹ = G(zᵏ).\]

[HW96]
Ernst Hairer and Gerhard Wanner, "Solving Ordinary Differential

Equations II, Springer Series in Computational Mathematics. ISBN 978-3-642-05221-7. Section IV.8. doi:10.1007/978-3-642-05221-7

OrdinaryDiffEq.NLNewton — Method.

(S::NLNewton)(integrator) -> (z, η, iter, fail_convergence)

Perform numerically stable modified NLNewton iteration that is specialized for implicit methods (see [HS96] and [HW96]), where z is the solution, η is used to measure the iteration error (see [HW96]), iter is the number of iteration, and fail_convergence reports whether the algorithm succeed. It solves

\[G(z) = dt⋅f(tmp + γ⋅z, p, t+c⋅h) - z = 0⃗\]

by iterating

\[W Δᵏ = dt⋅f(tmp + γ⋅zᵏ, p, t+c⋅h) - zᵏ zᵏ⁺¹ = zᵏ + Δᵏ\]

where W=M-dt⋅γJ, M is the mass matrix, dt is the step size, γ is a constant, and J is the Jacobian matrix.

[HS96]
M.E.Hoseaa and L.F.Shampine, "Analysis and implementation of TR-BDF2",

Applied Numerical Mathematics, Volume 20, Issues 1–2, February 1996, Pages 21-37. doi:10.1016/0168-9274(95)00115-8

[HW96]
Ernst Hairer and Gerhard Wanner, "Solving Ordinary Differential

Equations II, Springer Series in Computational Mathematics. ISBN 978-3-642-05221-7. Section IV.8. doi:10.1007/978-3-642-05221-7

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Called to add the extra k9, k10, k11 steps for the Order 5 interpolation when needed

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

By default, Hermite interpolant so update the derivative at the two ends

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

DiffEqBase.addsteps! — Function.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Called to add the extra k9, k10, k11 steps for the Order 5 interpolation when needed

DiffEqBase.addsteps! — Function.

OrdinaryDiffEq.constructDormandPrince — Function.

constructDormandPrince()

Constructs the tableau object for the Dormand-Prince Order 4/5 method.

OrdinaryDiffEq.ode_interpolant — Method.

odeinterpolant and odeinterpolant! dispatch

OrdinaryDiffEq.Anas5ConstantCache — Type.

An Optimized Runge-Kutta method for the solution of Orbital Problems by Z.A. Anastassi and T.E. Simos Journal of Computational and Applied Mathematics, Volume 175, Issue 1, 1 March 2005, Pages 1 to 9.

OrdinaryDiffEq.BS3ConstantCache — Method.

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

OrdinaryDiffEq.BS3ConstantCache — Method.

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

OrdinaryDiffEq.BS5ConstantCache — Method.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

OrdinaryDiffEq.BS5ConstantCache — Method.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

OrdinaryDiffEq.Feagin10ConstantCache — Method.

constructFeagin10

OrdinaryDiffEq.Feagin10ConstantCache — Method.

constructFeagin10

OrdinaryDiffEq.Feagin12ConstantCache — Method.

constructFeagin12

OrdinaryDiffEq.Feagin12ConstantCache — Method.

constructFeagin12

OrdinaryDiffEq.Feagin14ConstantCache — Method.

constructFeagin14

OrdinaryDiffEq.Feagin14ConstantCache — Method.

constructFeagin14

OrdinaryDiffEq.ODEIntegrator — Type.

ODEIntegrator

Fundamental struct allowing interactively stepping through the numerical solving of a differential equation. The full documentation is hosted here: http://docs.juliadiffeq.org/latest/basics/integrator.html. This docstring describes basic functionality only!

Initialize using integrator = init(prob::ODEProblem, alg; kwargs...). The keyword args which are accepted are the same Common Solver Options used by solve.

For reference, relevant fields of the ODEIntegrator are:

t - time of the proposed step

u - value at the proposed step

opts - common solver options

alg - the algorithm associated with the solution

f - the function being solved

sol - the current state of the solution

tprev - the last timepoint

uprev - the value at the last timepoint

opts holds all of the common solver options, and can be mutated to change the solver characteristics. For example, to modify the absolute tolerance for the future timesteps, one can do:

integrator.opts.abstol = 1e-9

For more info see the linked documentation page.

OrdinaryDiffEq.TanYam7ConstantCache — Method.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

OrdinaryDiffEq.TanYam7ConstantCache — Method.

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

OrdinaryDiffEq.TsitPap8ConstantCache — Method.

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

OrdinaryDiffEq.TsitPap8ConstantCache — Method.

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

OrdinaryDiffEq.Vern6ConstantCache — Method.

From Verner's Website

OrdinaryDiffEq.Vern6ConstantCache — Method.

From Verner's Website

OrdinaryDiffEq.Vern9ConstantCache — Method.

From Verner's Webiste

OrdinaryDiffEq.Vern9ConstantCache — Method.

From Verner's Webiste

OrdinaryDiffEq.WOperator — Type.

WOperator(mass_matrix,gamma,J[;transform=false])

A linear operator that represents the W matrix of an ODEProblem, defined as

\[W = MM - \gamma J\]

or, if transform=true:

\[W = \frac{1}{\gamma}MM - J\]

where MM is the mass matrix (a regular AbstractMatrix or a UniformScaling), γ is a real number proportional to the time step, and J is the Jacobian operator (must be a AbstractDiffEqLinearOperator). A WOperator can also be constructed using a *DEFunction directly as

WOperator(f,gamma[;transform=false])

f needs to have a jacobian and jac_prototype, but the prototype does not need to be a diffeq operator –- it will automatically be converted to one.

WOperator supports lazy * and mul! operations, the latter utilizing an internal cache (can be specified in the constructor; default to regular Vector). It supports all of AbstractDiffEqLinearOperator's interface.

OrdinaryDiffEq.BS5Interp — Method.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Used in the lazy construction of the dense output

k9, k10, k11 are not computed until called in the dense routine

OrdinaryDiffEq.BS5Interp — Method.

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

Used in the lazy construction of the dense output

k9, k10, k11 are not computed until called in the dense routine

OrdinaryDiffEq.BS5Interp_polyweights — Method.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These coefficients are taken from RKSuite

Note that RKSuite has an error: r081 should be 0 and r011 should be 1. This is pretty easy to spot since the first order interpolation is linear from y₀.

OrdinaryDiffEq.BS5Interp_polyweights — Method.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These coefficients are taken from RKSuite

Note that RKSuite has an error: r081 should be 0 and r011 should be 1. This is pretty easy to spot since the first order interpolation is linear from y₀.

OrdinaryDiffEq.Tsit5Interp — Method.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These are the coefficients of the expanded form of the polynomials from

Runge–Kutta pairs of order 5(4) satisfying only the first column simplifying assumption

Ch. Tsitouras

OrdinaryDiffEq.Tsit5Interp — Method.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

These are the coefficients of the expanded form of the polynomials from

Runge–Kutta pairs of order 5(4) satisfying only the first column simplifying assumption

Ch. Tsitouras

OrdinaryDiffEq.Vern6Interp_polyweights — Method.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

OrdinaryDiffEq.Vern6Interp_polyweights — Method.

Coefficients for the polynomial bᵢΘ = ri1Θ + ri2Θ^2 + ri3*Θ^3 + ...

OrdinaryDiffEq._ode_interpolant — Method.

OrdinaryDiffEq._ode_interpolant — Method.

From MATLAB ODE Suite by Shampine

OrdinaryDiffEq.alg_cache_expRK — Method.

alg_cache_expRK(alg,u,uEltypeNoUnits,uprev,f,t,dt,p,du1,tmp,dz,plist)

Construct the non-standard caches (not uType or rateType) for ExpRK integrators.

plist is a list of integers each corresponding to the order of a phiv(!) call in perform_step!.

OrdinaryDiffEq.alg_cache_exprb — Method.

alg_cache_exprb(alg,uEltypeNoUnits,uprev,f,t,p,du1,tmp,dz,plist)

Construct the non-standard caches (not uType or rateType) for Exprb integrators.

plist is a list of integers each corresponding to the order of a phiv(!) call in perform_step!.

OrdinaryDiffEq.calc_J! — Method.

calc_J!(integrator,cache,is_compos)

Interface for calculating the jacobian.

For constant caches, a new jacobian object is returned whereas for mutable caches cache.J is updated. In both cases, if integrator.f has a custom jacobian update function, then it will be called for the update. Otherwise, either ForwardDiff or finite difference will be used depending on the jac_config of the cache.

OrdinaryDiffEq.calc_J — Method.

calc_J(integrator,cache,is_compos)

Interface for calculating the jacobian.

For constant caches, a new jacobian object is returned whereas for mutable caches cache.J is updated. In both cases, if integrator.f has a custom jacobian update function, then it will be called for the update. Otherwise, either ForwardDiff or finite difference will be used depending on the jac_config of the cache.

OrdinaryDiffEq.calculate_residuals! — Method.

calculate_residuals!(out, u₀, u₁, α, ρ)

Save element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

in out.

OrdinaryDiffEq.calculate_residuals! — Method.

calculate_residuals!(out, ũ, u₀, u₁, α, ρ)

Save element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

in out.

OrdinaryDiffEq.calculate_residuals — Method.

calculate_residuals(u₀, u₁, α, ρ)

Calculate element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

OrdinaryDiffEq.calculate_residuals — Method.

calculate_residuals(ũ, u₀, u₁, α, ρ)

Calculate element-wise residuals

\[\frac{ũ}{α+\max{|u₀|,|u₁|}*ρ}\]

OrdinaryDiffEq.choosedeg! — Method.

choosedeg!(cache) -> nothing

Calculate ms[mdeg] (the degree of the Chebyshev polynomial) and cache.recind (the index of recurrence parameters for that degree), where recf[recind:(recind+ms[mdeg]-2)] are the μ,κ pairs for the mdeg degree method.

OrdinaryDiffEq.expRK_operators — Method.

expRK_operators(alg,dt,A) -> ops

Compute operator(s) for an ExpRK algorithm. dt is the time step and A is the matrix form of the linear operator (from either a linear problem or a SplitODEProblem). All ExpRK methods that use caching operators should implement this method.

OrdinaryDiffEq.hermite_interpolant — Method.

Hairer Norsett Wanner Solving Ordinary Differential Euations I - Nonstiff Problems Page 190

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

OrdinaryDiffEq.hermite_interpolant — Method.

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

OrdinaryDiffEq.hermite_interpolant — Method.

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

OrdinaryDiffEq.hermite_interpolant — Method.

Herimte Interpolation, chosen if no other dispatch for ode_interpolant

OrdinaryDiffEq.linear_interpolant — Method.

Linear Interpolation

OrdinaryDiffEq.ode_interpolation — Function.

ode_interpolation(tval::Number,ts,timeseries,ks)

Get the value at tval where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

OrdinaryDiffEq.ode_interpolation — Function.

ode_interpolation(tvals,ts,timeseries,ks)

Get the value at tvals where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

OrdinaryDiffEq.ode_interpolation! — Function.

ode_interpolation(tvals,ts,timeseries,ks)

Get the value at tvals where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

OrdinaryDiffEq.ode_interpolation! — Function.

ode_interpolation!(out,tval::Number,ts,timeseries,ks)

Get the value at tval where the solution is known at the times ts (sorted), with values timeseries and derivatives ks

OrdinaryDiffEq.reset_alg_dependent_opts! — Method.

If no user default, then this will change the default to the defaults for the second algorithm.

OrdinaryDiffEq.ssp_coefficient — Method.

ssp_coefficient(alg)

Return the SSP coefficient of the ODE algorithm alg. If one time step of size dt with alg can be written as a convex combination of explicit Euler steps with step sizes cᵢ * dt, the SSP coefficient is the minimal value of 1/cᵢ.

Examples

julia> ssp_coefficient(SSPRK104())
6