Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

PCRE2.compile — Method.

Wrapper for compile() function, throw error on error return with error info

PCRE2.info — Method.

PCRE2 pattern_info call wrapper

PCRE2.jit_compile — Method.

Wrapper for jit_compile() function, throw error on error return with error info