Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

PEG.parse_next — Method.

parse_next(rule, input; whole=false)

Parse a prefix of the input string as one instance of the given rule. This differs from just calling the rule itself on the input in a few important ways:

When parsing succeeds, parse_next returns only the parsed value, while rule returns a Tuple of the parsed value and the remaining unparsed input.

When parsing fails, parse_next throws an exception, with information on what failed to match starting at the latest point in the string that any parsing expression matched up to. rule just returns nothing.

When whole=true, parsing only succeeds if the whole input is consumed, i.e. the second value in the tuple that rule returns is "".

PEG.parse_whole — Method.

parse_whole(rule, input) = parse_next(rule, input; whole=true)

PEG.squash_ast — Method.

squash_ast(x)

Given an ast returned by a grammar with no semantics, return a simplified version, such that there are no nothings, and no Any[]s with exactly zero or one element.

PEG.Cache — Type.

A Cache maps from (rule name or gensym, input string length remaining) to either nothing or (parsed value, remaining input substring) (which is what rule functions return).

PEG.Failure — Type.

Singleton value that semantics functions may return to cause the parsing expression they were called from to fail, e.g. return Failure().

PEG.flatten_op — Method.

Given e.g. x=:(((a op b) op c) op d), return [a,b,c,d].

PEG.remove_re_flag — Method.

remove_re_flag(flags, flag)

Return (a copy of) regex flags with flag removed, and a boolean indicating whether the flag was there in the first place.

PEG.setdebug! — Function.

setdebug!(val=true)

Enable or disable printing of debugging information.