Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

PackageCompiler.build_executable — Function.

build_executable(
    julia_program, output_name = nothing, c_program = nothing;
    snoopfile = nothing, builddir = nothing, verbose = false, quiet = false,
    copy_julialibs = true, copy_files = nothing, release = false, Release = false,
    sysimage = nothing, home = nothing, startup_file = nothing, handle_signals = nothing,
    sysimage_native_code = nothing, compiled_modules = nothing,
    depwarn = nothing, warn_overwrite = nothing,
    compile = nothing, cpu_target = nothing, optimize = nothing, debug = nothing,
    inline = nothing, check_bounds = nothing, math_mode = nothing,
    cc = nothing, cc_flags = nothing
)
`julia_program` needs to be a Julia script containing a `julia_main` function, e.g. like `examples/hello.jl`
`snoopfile` is optional and can be a Julia script which calls functions that you want to make sure to have precompiled
`builddir` is where the compiled artifacts will end up

PackageCompiler.build_shared_lib — Function.

build_shared_lib(
    julia_program, output_name = nothing;
    snoopfile = nothing, builddir = nothing, verbose = false, quiet = false,
    init_shared = false, copy_julialibs = true, copy_files = nothing, release = false, Release = false,
    sysimage = nothing, home = nothing, startup_file = nothing, handle_signals = nothing,
    sysimage_native_code = nothing, compiled_modules = nothing,
    depwarn = nothing, warn_overwrite = nothing,
    compile = nothing, cpu_target = nothing, optimize = nothing, debug = nothing,
    inline = nothing, check_bounds = nothing, math_mode = nothing,
    cc = nothing, cc_flags = nothing
)
`julia_program` needs to be a Julia script containing a `julia_main` function, e.g. like `examples/hello.jl`
`snoopfile` is optional and can be a Julia script which calls functions that you want to make sure to have precompiled
`builddir` is where the compiled artifacts will end up

PackageCompiler.compile_package — Method.

compile_package(packages...; kw_args...)

with packages being either a string naming a package, or a tuple (package_name, precompile_file). If no precompile file is given, it will use the packages runtests.jl, which is a good canditate for figuring out what functions to compile!

PackageCompiler.compile_package — Method.

compile_package(packages::Tuple{String, String}...; force = false, reuse = false, debug = false, cpu_target = nothing)

Compile a list of packages. Each package comes as a tuple of (package_name, precompile_file) where the precompile file should contain all function calls, that should get compiled into the system image. Usually the runtests.jl file is a good candidate, since it should run all important functions of a package.

PackageCompiler.force_native_image! — Function.

force_native_image!()

Builds a clean system image, similar to a fresh Julia install. Can also be used to build a native system image for a downloaded cross compiled julia binary.

PackageCompiler.revert — Function.

Reverts a forced compilation of the system image. This will restore any previously backed up system image files, or build a new, clean system image.

PackageCompiler.static_julia — Method.

static_julia(juliaprog::String; kw_args...)

compiles the Julia file at path juliaprog with keyword arguments:

cprog                            C program to compile (required only when building an executable, if not provided a minimal driver program is used)
verbose                          increase verbosity
quiet                            suppress non-error messages
builddir                         build directory
outname                          output files basename
snoopfile                        specify script calling functions to precompile
clean                            remove build directory
autodeps                         automatically build required dependencies
object                           build object file
shared                           build shared library
init_shared                      add `init_jl_runtime` and `exit_jl_runtime` to shared library for runtime initialization
executable                       build executable file
rmtemp                           remove temporary build files
copy_julialibs                   copy Julia libraries to build directory
copy_files                       copy user-specified files to build directory (either `nothing` or a string array)
release                          build in release mode, implies `-O3 -g0` unless otherwise specified
Release                          perform a fully automated release build, equivalent to `-atjr`
sysimage <file>                  start up with the given system image file
home <dir>                       set location of `julia` executable
startup_file {yes|no}            load `~/.julia/config/startup.jl`
handle_signals {yes|no}          enable or disable Julia's default signal handlers
sysimage_native_code {yes|no}    use native code from system image if available
compiled_modules {yes|no}        enable or disable incremental precompilation of modules
depwarn {yes|no|error}           enable or disable syntax and method deprecation warnings
warn_overwrite {yes|no}          enable or disable method overwrite warnings
compile {yes|no|all|min}         enable or disable JIT compiler, or request exhaustive compilation
cpu_target <target>              limit usage of CPU features up to <target> (implies default `--sysimage_native_code=no`)
optimize {0,1,2,3}               set the optimization level
debug <level>                    enable / set the level of debug info generation
inline {yes|no}                  control whether inlining is permitted
check_bounds {yes|no}            emit bounds checks always or never
math_mode {ieee,fast}            disallow or enable unsafe floating point optimizations
cc                               system C compiler
cc_flags <flags>                 pass custom flags to the system C compiler when building a shared library or executable (either `nothing` or a string array)

PackageCompiler.build_sysimg — Function.

build_sysimg(sysimg_path=default_sysimg_path(), cpu_target="native", userimg_path=nothing; force=false)

Rebuild the system image. Store it in sysimg_path, which defaults to a file named sys.ji that sits in the same folder as libjulia.{so,dylib}, except on Windows where it defaults to Sys.BINDIR/../lib/julia/sys.ji. Use the cpu instruction set given by cpu_target. Valid CPU targets are the same as for the -C option to julia, or the -march option to gcc. Defaults to native, which means to use all CPU instructions available on the current processor. Include the user image file given by userimg_path, which should contain directives such as using MyPackage to include that package in the new system image. New system image will not replace an older image unless force is set to true.

PackageCompiler.get_backup! — Function.

Returns the system image file stored in the backup folder. If there is no backup, this function will automatically generate a system image in the backup folder.

PackageCompiler.snoop_userimg — Method.

snoop_userimg(userimg, packages::Tuple{String, String}...)

Traces all function calls in packages and writes out `precompile` statements into the file `userimg`