Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

PaddedViews.PaddedView — Type.

datapadded = PaddedView(fillvalue, data, padded_indices)
datapadded = PaddedView(fillvalue, data, padded_indices, data_indices)
datapadded = PaddedView(fillvalue, data, sz)
datapadded = PaddedView(fillvalue, data, sz, first_datum)

Create a padded version of the array data, where any elements within the span of padded_indices not assigned in data will have value fillvalue. If a second set of indices data_indices is not supplied it is assumed the array data spans the indices from the first up until size(data), otherwise data spans the specified data_indices. Alternately, the padded array size sz can be specified along with the location of the first element of data, first_datum. If first_datum is omitted, it is assumed to be the first element of the padded array.

Example

julia> a = reshape(1:9, 3, 3)
3×3 Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}:
 1  4  7
 2  5  8
 3  6  9

julia> PaddedView(-1, a, (4, 5))
4×5 PaddedViews.PaddedView{Int64,2,Tuple{Base.OneTo{Int64},Base.OneTo{Int64}},Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}:
  1   4   7  -1  -1
  2   5   8  -1  -1
  3   6   9  -1  -1
 -1  -1  -1  -1  -1

 julia> PaddedView(-1, a, (1:5,1:5), (2:4,2:4))
 PaddedViews.PaddedView{Int64,2,Tuple{UnitRange{Int64},UnitRange{Int64}},OffsetArrays.OffsetArray{Int64,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}} with indices 1:5×1:5:
  -1  -1  -1  -1  -1
  -1   1   4   7  -1
  -1   2   5   8  -1
  -1   3   6   9  -1
  -1  -1  -1  -1  -1

julia> PaddedView(-1, a, (5,5), (2,2))
5×5 PaddedViews.PaddedView{Int64,2,Tuple{Base.OneTo{Int64},Base.OneTo{Int64}},OffsetArrays.OffsetArray{Int64,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}}}}:
-1  -1  -1  -1  -1
-1   1   4   7  -1
-1   2   5   8  -1
-1   3   6   9  -1
-1  -1  -1  -1  -1

PaddedViews.paddedviews — Method.

Aspad = paddedviews(fillvalue, A1, A2, ....)

Pad the arrays A1, A2, ..., to a common size or set of axes, chosen as the span of axes enclosing all of the input arrays.

Example:

julia> a1 = reshape([1,2], 2, 1)
2×1 Array{Int64,2}:
 1
 2

julia> a2 = [1.0,2.0]'
1×2 Array{Float64,2}:
 1.0  2.0

julia> a1p, a2p = paddedviews(0, a1, a2);

julia> a1p
2×2 PaddedViews.PaddedView{Int64,2,Tuple{Base.OneTo{Int64},Base.OneTo{Int64}},Array{Int64,2}}:
 1  0
 2  0

julia> a2p
2×2 PaddedViews.PaddedView{Float64,2,Tuple{Base.OneTo{Int64},Base.OneTo{Int64}},Array{Float64,2}}:
 1.0  2.0
 0.0  0.0