Docstrings

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Pages.add_library — Method.

Add a JS library to the current page from a url.

Pages.block — Method.

Block Julia control flow until until callback"notify" is called.

Pages.broadcast — Method.

Broadcast a message to all connected web pages to be interpetted by WebSocket listener. For example, in JavaScript:

var sock = new WebSocket('ws://'+window.location.host); sock.onmessage = function( message ){ var msg = JSON.parse(message.data); console.log(msg); }

Pages.message — Method.

Send a message to the specified connection to be interpetted by WebSocket listener. For example, in JavaScript:

var sock = new WebSocket('ws://'+window.location.host); sock.onmessage = function( message ){ var msg = JSON.parse(message.data); console.log(msg); }