Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Pandas.pandas_wrap — Method.

pandas_wrap(pyo::PyObject)

Wrap an instance of a Pandas python class in the Julia type which corresponds to that class.

Pandas.pyattr_set — Method.

pyattr_set(types, methods...)

For each Julia type T<:PandasWrapped in types and each method m in methods, define a new function m(t::T, args...) that delegates to the underlying pyobject wrapped by t.