Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Parsers.Delimited — Type.

Parsers.Delimited(next, delims::Union{Char, String}...=',')

A custom `Parsers.Layer` used to support parsing delimited values in `IO` streams. `delims` can be any number of `Char` or `String` arguments that should collectively be used as "delimiters".

Parsing on a `Parsers.Delimited` will first call `Parsers.parse!(d.next, io, result; kwargs...)`, then expect the next bytes to be one of the expected `delims` arguments.
If one of `delims` is not found, the result is `Parsers.INVALID`, but parsing will continue until a valid `delims` is found. An `eof(io)` is _always_ considered a valid termination state in place of a delimiter.

Parsers.Layer — Type.

An interface to define custom "parsing support" layers to be used in `Parsers.parse!`. Examples implementations include:
  * Parsers.Delimited
  * Parsers.Quoted
  * Parsers.Strip
  * Parsers.Sentinel

The interface to implement for a `Layer` is:
  * `Parsers.parse!(x::MyLayer, io::IO, r::Result{T}; kwargs...) where {T}`: implementation of your custom `MyLayer` type, including calling down to the next layer
  * Include a parameterized `next::I` field, to allow parsing the next layer

Parsers.Quoted — Type.

Parsers.Quoted(next, quotechar='"', escapechar='\', ignore_quoted_whitespace=false)
Parsers.Quoted(next, openquote, closequote, escapechar, ignore_quoted_whitespace)

A custom `Parsers.Layer` used to support parsing potentially "quoted" values. Parsing with a `Parsers.Quoted` does not _require_ the value to be quoted, but will always check for an initial quote and, if found, will then expect (and continue parsing until) a corresponding close quote is found.
A single `quotechar` can be given, indicating the quoted field will start and end with the same character.
Both `quotechar` and `escapechar` arguments are limited to ASCII characters.

Parsers.Result — Type.

Parsers.Result{T}(x::Union{T, Missing}, ::Parsers.ReturnCode, b::UInt8)
Parsers.Result(::Type{T}, code, byte)

A result type used by `Parsers.parse!` to signal the result of trying to parse a certain type. Fields include:
    * `result::Union{T, Missing}`: holds the parsed result of type `T` or `missing` if unable to parse or a valid `Parsers.Sentinel` value was found
    * `code::Parsers.ReturnCode`: a value signaling whether parsing succeeded (`Parsers.OK`) or not (`Parsers.INVALID`); see `?Parsers.ReturnCode` for all possible codes
    * `pos::Int64`: the byte position when a parsing operation started

Parsers.ReturnCode — Type.

Each `Parsers.Result` has a `r.code` field which has type `Parsers.ReturnCode` and is a set of bit flags for various parsing states.
The top bit is used to indicate "SUCCESS" (0) and "INVALID" (1), so all failed parsing attempts will have a code < 0, while successful parsings will be > 0.
Use `Parsers.ok(code)` to check if a `Parsers.Result` is successful or not.
`Parsers.codes(code)` and `Parsers.text(code)` can also be used to get text representations of which bit flags are set in a return code.
Various bit flags include:
  * `OK`: the innermost type parser succeeded in parsing a valid value for the type
  * `SENTINEL`: a sentinel value was matched; mutually exclusive with `OK` (sentinels are only checked for when the underlying type parser fails)
  * `QUOTED`: an opening quote character was detected while parsing (note this does not indicate a _correctly_ quoted value)
  * `DELIMITED`: a valid delimiter was found while parsing; note that EOF is always a valid delimiter
  * `NEWLINE`: a newline character was matched as a delimiter (useful for applications where newlines have special delimiting purposes)
  * `EOF`: parsing encountered the end-of-file while parsing
  * `INVALID_QUOTED_FIELD`: a corresponding closing quote character was not found for a `QUOTED` field
  * `INVALID_DELIMITER`: a delimiter was found, but not at the expected position while parsing
  * `OVERFLOW`: a numeric type overflowed while parsing its value

Parsers.Sentinel — Type.

Parses.Sentinel(next, sentinels::Union{String, Vector{String}})

A custom `Parsers.Layer` to support sentinel value parsing for any type. A single string or vector of strings can be provided which, if encountered during parsing, will result in `missing` being returned with a `Parsers.ReturnCode` of `Parsers.OK`.

One special case of sentinel parsing is that of the "empty" sentinel, i.e. `Parsers.Sentinel(io, "")`. In this case, sentinel parsing will "succeed" only when the underlying type parsing failed to consume any bytes (i.e it immediately encountered invalid characters).

Parsers.Strip — Type.

Parsers.Strip(next, wh1=' ', wh2='	')

A custom `Parsers.Layer` used to remove leading and trailing whitespace.
By default, only `' '` (space) and `'	'` (tab) characters are skipped over.
Any two valid ASCII characters may be used to skip.

Parsers.Trie — Type.

Trie(value::String, value_to_return::T) where {T}
Trie(values::Vector{String}, value_to_return::T) where {T}
Trie(values::Vector{Pair{String, T}}) where {T}

A basic [trie](https://en.wikipedia.org/wiki/Trie) structure for use in parsing sentinel and other special values.
The various constructors take either a single or set of strings to act as "sentinels" (i.e. special values to be parsed), plus an optional `value_to_return` argument, which will be the value returned if the sentinel is found while parsing.
Note the last constructor `Trie(values::Vector{Pair{String, T}})` allows specifying different return values for different sentinel values. Bool parsing uses this like:
```
const BOOLS = Trie(["true"=>true, "false"=>false])
```
The only restriction is that each individual value must be of the same type (i.e. a single `Trie` can only ever return one type of value).

See `?Parsers.match!` for more information on how a `Trie` can be used for special-value parsing.

Parsers.defaultparser — Function.

function used by Parsers to dispatch to default type parser implementations

Parsers.fastseek! — Function.

Parsers.fastseek!(io::IO, n::Integer)

Without valididty checks, seek an `IO` to desired byte position `n`.

Parsers.match! — Function.

Parsers.match!(t::Parsers.Trie, io::IO, r::Parsers.Result, setvalue::Bool=true, ignorecase::Bool=false)

Function that takes an `io::IO` argument, a prebuilt `r::Parsers.Result` argument, and a `t::Parsers.Trie` argument, and attempts to match/detect special values in `t` with the next bytes consumed from `io`.
If special values are found, `r.result` will be set to the value that was associated with `t` when it was constructed.
The return value of `Parsers.match!` is if a special value was indeed detected in `io` (`true` or `false`).
Optionally, if the `setvalue` is `false`, `r.result` will be unaffected (i.e. not set) even if a special value is found.
The optional argument `ignorecase` can be used if case-insensitive matching is desired.

Note that `io` is reset to its original position if no special value is found.

Parsers.parse — Function.

Attempt to parse a value of type T from string str. Throws Parsers.Error on parser failures and invalid values.

Parsers.parse! — Function.

Parsers.parse!(l::Parsers.Layer, io::IO, r::Result{T}; kwargs...)::Parsers.Result{T}

Internal parsing function that returns a full `Parsers.Result` type to indicate the success of parsing a `T` from `io`.

A custom parsing function `f` can be passed, as the innermost "layer", which should have the form `f(io::IO, ::Type{T}, r::Result{T}, args...)::Result{T}`, i.e. it takes an `IO` stream, attemps to parse type `T`, takes a pre-allocated `Result{T}` and should return it after parsing.

Parsers.peekbyte — Function.

Parsers.peekbyte(io::IO)::UInt8

Return, but do not consume, the next byte from an `IO` without checking `eof(io)`.

Parsers.readbyte — Function.

Parsers.readbyte(io::IO)::UInt8

Consume a single byte from an `IO` without checking `eof(io)`.

Parsers.tryparse — Function.

Attempt to parse a value of type T from IO io. Returns nothing on parser failures and invalid values.