Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Poltergeist.MarkovMap — Type.

MarkovMap(fs::Vector, ds::Vector, ran = coveringinterval(ds); dir=Forward, diff=autodiff(...)))

Generate a MarkovMap with branches given by elements of fsdefined on subdomains given byds, onto a vectorran`.

The keyword argument dir stipulates whether the elements of fs are the branches (Forward) or the branches' inverses (Reverse).

The keyword argument diff provides the derivatives of the fs. By default it is the automatic derivatives of fs.

Poltergeist.MarkovMap — Type.

MarkovMap(branches::Vector, domain, rangedomain)

Generate a computer representation of a full-branch uniformly-expanding interval map domain → rangedomain using a vector describing the branches of the map.

Poltergeist.AR_process — Method.

AR_process(f,A::Fun)

Calculate coefficients of Gaussian AR process with the same autocorrelation function as A under the map f.

See Appendix A.2 of Wormell, C.L. & Gottwald, G.A. 'On the Validity of Linear Response Theory in High-Dimensional Deterministic Dynamical Systems' (2018).

Poltergeist.MA_process — Method.

MA_process(f,A::Fun)

Calculate coefficients of Gaussian MA process with the same autocorrelation function as A under the map f.

See Appendix A.2 of Wormell, C.L. & Gottwald, G.A. 'On the Validity of Linear Response Theory in High-Dimensional Deterministic Dynamical Systems' (2018).

Poltergeist.Transfer — Method.

Transfer(m::AbstractMarkovMap)

Create a CachedOperator of a ConcreteTransfer operator encoding the transfer operator of m.

Caching is used for speed, as entries of the transfer operator are most efficiently calculated whole columns at a time.

Poltergeist.acim — Method.

acim(L)

Output a Fun object giving the acim of the associated map

Poltergeist.birkhoffcov — Method.

birkhoffcov(L, A::Fun, B::Fun)

Output covariance of CLT-normalised Birkhoff sums of A and B under the map

Poltergeist.birkhoffvar — Method.

birkhoffvar(L, A::Fun)

Output diffusion coefficient of A under the map using Green-Kubo formula

Poltergeist.correlationsum — Function.

correlationsum(L, A::Fun, ρ=acim(L))

Output resolvent of transfer operator applied to ρ*(A - ∫A dρ)

Poltergeist.covariancefunction — Method.

cfA, cfB = covariancefunction(L, A, B, n)

Compute the lag-covariance function against transfer operator L between observables A and B for n time steps in each direction. Specifically, the (k+1)th entry of cfA is the expectation of B A∘f^(k) and vice versa for cfB.

Poltergeist.covariancefunction — Method.

cfA, cfB = covariancefunction(L, A, B; tol=eps(A*B))

Chooses n so that the covariance declines to a given tolerance.

Poltergeist.covariancefunction — Method.

cfA = covariancefunction(L, A, n)

Compute the lag-covariance function against transfer operator L for observable A for n time steps in each direction. Specifically, the (k+1)th entry is the expectation of A A∘f^(k).

Poltergeist.covariancefunction — Method.

cf = covariancefunction(L, A; tol = eps(A))

Choose n so that the covariance declines to a given tolerance

Poltergeist.doubling — Function.

doubling(d = Interval(0,1.))

Returns the full-branch interval map on domain d with 2 equally-sized branches.

See also: tupling, lanford

Poltergeist.hofbauerextension — Function.

hofbauerextension(m,basedomains;maxdepth=100,forcereturn=trues(length(basedomains)))

Generates a Hofbauer extension of m starting from basedomains, which may be a single domain.

The keyword maxdepth says how deep the Hofbauer extension should go, and forcereturn sets whether a return to given members of basedomains should forced if possible.

For example, in the case of f(x) = mod(2x,1), if basedomains is set to Interval(0.,0.3) then Interval(0.,0.5) might map to Interval(0.,1.)for a given map ifforcereturn=falsebut would map to Interval(0,0.3) and Interval(0.3,1) if forcereturn=true.

Poltergeist.lanford — Function.

lanford(T=Float64)

Return the Lanford map, with type encoding T.

See also: tupling, doubling

Poltergeist.linearresponse — Method.

linearresponse(L, X::Fun)

Output a Fun object giving the first-order change in the acim of the map under perturbation X

Poltergeist.lyapunov — Function.

lyapunov(f, r=acim(f), sp=Space(rangedomain(f)))

Calculate Lyapunov exponent associated with f

Poltergeist.modulomap — Function.

modulomap(f, D, R=dom; diff= autodiff(f,dom))

Output MarkovMap or CircleMap m: D → R such that m(x) = f(x) mod R.

Poltergeist.nbranches — Method.

nbranches(m)

Return the number of branches of m.

Poltergeist.perturb — Method.

perturb(d, X, ϵ)

Construct a self-map on domain d: x ↦ x + ϵ X(x)

Poltergeist.perturb — Method.

perturb(m::AbstractMarkovMap, X, ϵ)

Output perturbation of m: x ↦ m(x) + ϵ X(m(x))

Poltergeist.tupling — Function.

tupling(k::Int, d = Interval(0,1.))

Returns the full-branch interval map on domain d with k equally-sized branches.

See also: doubling, lanford

Poltergeist.HofbauerDomain — Type.

HofbauerDomain

A Domain object annotated with a depth index for HofbauerExtensions. It is not a Domain subtype itself as it should only be used in the context of Hofbauer extensions.

Poltergeist.HofbauerPoint — Type.

HofbauerPoint(x, graphind)

Constructs a HofbauerPoint, representing point x on the graphindth domain of a Hofbauer extension.

LinearAlgebra.eigvals — Method.

eigvals(m::AbstractMarkovMap, n)

Output eigenvalues of Transfer(m) using n×n Galerkin discretisation.

Calls directly to ApproxFun: you can also call eigvals(Transfer(m), n)

LinearAlgebra.eigvecs — Method.

eigvecs(m::AbstractMarkovMap, n)

Output eigenfunctions of Transfer(m) using n×n Galerkin discretisation.

Calls directly to ApproxFun: you can also call eigvecs(Transfer(m), n)

Poltergeist.branches — Method.

branches(m)

Return the branches of the map m.

Poltergeist.eachbranchindex — Method.

eachbranchindex(m)

Return an iterator giving the indices of the branches of m

Poltergeist.maphb — Method.

maphb(b::AbstractBranch, x)

Maps point x according to b. For NeutralBranches calculates the return map out of domain(b).

Poltergeist.zero_to — Function.

zero_to(A::Fun, ρ::Fun=uniform(space(A)))

Output ρ*(A - ∫A dρ)