Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ProximalAlgorithms.AFBA — Method.

Asymmetric forward-backward-adjoint algorithm

AFBA(x0, y0; kwargs)

Solves convex optimization problems of the form

minimize f(x) + g(x) + (h □ l)(L x),

where f is smooth, g and h are possibly nonsmooth and l is strongly convex. Symbol □ denotes the infimal convolution, and L is a linear mapping. Points x0 and y0 are the initial primal and dual iterates, respectively.

Keyword arguments are as follows:

f: smooth, convex function (default: zero)

g: convex function (possibly nonsmooth, default: zero)

h: convex function (possibly nonsmooth, default: zero)

l: strongly convex function (possibly nonsmooth, default: indicator of {0})

L: linear operator (default: identity)

betaQ: Lipschitz constant of gradient of f (default: zero)

betaR: Lipschitz constant of gradient of l conjugate (default: zero)

gamma1: stepsize corresponding to the primal updates (default: see [1] for each case)

gamma2: stepsize corresponding to the dual updates (default: see [1] for each case)

theta: nonnegative algorithm parameter (default: '1')

mu: algorithm parameter in the range 0,1

tol: primal-dual termination tolerance (default: 1e-5)

maxit: maximum number of iterations (default: 10000)

verbose, verbosity level (default: 1)

verbose_freq, verbosity frequency for verbose = 1 (default: 100)

The iterator implements Algorithm 3 of [1] with constant stepsize (α_n=λ) for several prominant special cases:

θ = 2 ==> Corresponds to the Vu-Condat Algorithm [2,3].

θ = 1, μ=1

θ = 0, μ=1

θ ∈ [0,∞), μ=0

See [1, Figure 1] for other special cases and relation to other algorithms.

[1] Latafat, Patrinos. "Asymmetric forward–backward–adjoint splitting for solving monotone inclusions involving three operators" Computational Optimization and Applications, pages 1–37, 2017. [2] Condat. "A primal–dual splitting method for convex optimization involving Lipschitzian, proximable and linear composite terms" Journal of Optimization Theory and Applications 158.2 (2013): 460-479. [3] Vũ. "A splitting algorithm for dual monotone inclusions involving cocoercive operators"" Advances in Computational Mathematics, 38(3), pp.667-681.

ProximalAlgorithms.ChambollePock — Method.

Chambolle-Pock primal-dual algorithm

ChambollePock(x0, y0; kwargs)

Solves convex optimization problems of the form

minimize g(x) + (h □ l)(L x).

where g and h are possibly nonsmooth and l is strongly convex. Symbol □ denotes the infimal convolution, and L is a linear mapping. Points x0 and y0 are the initial primal and dual iterates, respectively.

See documentation of AFBA for the list of keyword arguments.

ProximalAlgorithms.FBS — Method.

Forward-backward splitting

FBS(x0; kwargs...)

Solves a problem of the form

minimize fs(As*x) + fq(Aq*x) + g(x)

where fs is a smooth function, fq is a quadratic function, g is a proximable function and As, Aq are linear operators. Parameter x0 is the initial point. Keyword arguments specify the problem and additional options as follows:

fs, smooth function (default: identically zero function)

fq, quadratic function (default: identically zero function)

g, proximable function (default: identically zero function)

As, linear operator (default: identity)

Aq, linear operator (default: identity)

gamma, stepsize (default: unspecified, determine automatically)

maxit, maximum number of iteration (default: 10000)

tol, halting tolerance on the fixed-point residual (default: 1e-4)

adaptive, adaptively adjust gamma (default: false if gamma is provided)

fast, enables accelerated method (default: false)

verbose, verbosity level (default: 1)

verbose_freq, verbosity frequency for verbose = 1 (default: 100)

ProximalAlgorithms.VuCondat — Method.

Vũ-Condat primal-dual algorithm

VuCondat(x0, y0; kwargs)

Solves convex optimization problems of the form

minimize f(x) + g(x) + (h □ l)(L x).

where f is smooth, g and h are possibly nonsmooth and l is strongly convex.

Symbol □ denotes the infimal convolution, and L is a linear mapping. Points x0 and y0 are the initial primal and dual iterates, respectively.

See documentation of AFBA for the list of keyword arguments.