Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

PyCall.PyArray — Type.

PyArray(o::PyObject)

This converts an ndarray object o to a PyArray.

This implements a nocopy wrapper to a NumPy array (currently of only numeric types only).

If you are using pycall and the function returns an ndarray, you can use PyArray as the return type to directly receive a PyArray.

PyCall.PyDict — Type.

PyDict(o::PyObject)
PyDict(d::Dict{K,V})

This returns a PyDict, which is a no-copy wrapper around a Python dictionary.

Alternatively, you can specify the return type of a pycall as PyDict.

PyCall.PyObject — Type.

PyObject(juliavar)

This converts a julia variable to a PyObject, which is a reference to a Python object. You can convert back to native julia types using convert(T, o::PyObject), or using PyAny(o).

Given o::PyObject, o[:attribute] is equivalent to o.attribute in Python, with automatic type conversion.

Given o::PyObject, get(o, key) is equivalent to o[key] in Python, with automatic type conversion.

PyCall.PyVector — Type.

PyVector(o::PyObject)

This returns a PyVector object, which is a wrapper around an arbitrary Python list or sequence object.

Alternatively, PyVector can be used as the return type for a pycall that returns a sequence object (including tuples).

PyCall.PyNULL — Method.

PyNULL()

Return a PyObject that has a NULL underlying pointer, i.e. it doesn't actually refer to any Python object.

This is useful for initializing PyObject global variables and array elements before an actual Python object is available. For example, you might do const myglobal = PyNULL() and later on (e.g. in a module __init__ function), reassign myglobal to point to an actual object with copy!(myglobal, someobject). This procedure will properly handle Python's reference counting (so that the Python object will not be freed until you are done with myglobal).

PyCall.PyReverseDims — Method.

PyReverseDims(array)

Passes a Julia array to Python as a NumPy row-major array (rather than Julia's native column-major order) with the dimensions reversed (e.g. a 2×3×4 Julia array is passed as a 4×3×2 NumPy row-major array). This is useful for Python libraries that expect row-major data.

PyCall.PyTextIO — Method.

PyTextIO(io::IO)
PyObject(io::IO)

Julia IO streams are converted into Python objects implementing the RawIOBase interface, so they can be used for binary I/O in Python

PyCall.ispynull — Method.

ispynull(o::PyObject)

Test where o contains a NULL pointer to a Python object, i.e. whether it is equivalent to a PyNULL() object.

PyCall.pybuiltin — Method.

pybuiltin(s::AbstractString)

Look up a string or symbol s among the global Python builtins. If s is a string it returns a PyObject, while if s is a symbol it returns the builtin converted to PyAny.

PyCall.pybytes — Method.

pybytes(b::Union{String,DenseVector{UInt8}})

Convert b to a Python bytes object. This differs from the default PyObject(b) conversion of String to a Python string (which may fail if b does not contain valid Unicode), or from the default conversion of a Vector{UInt8} to a bytearray object (which is mutable, unlike bytes).

PyCall.pycall! — Method.

pycall!(ret::PyObject, o::Union{PyObject,PyPtr}, returntype::Type, args...; kwargs...)

Set ret to the result of pycall(o, returntype, args...; kwargs) and return convert(returntype, ret). Avoids allocating an extra PyObject for ret. See pycall for other details.

PyCall.pycall — Method.

pycall(o::Union{PyObject,PyPtr}, returntype::TypeTuple, args...; kwargs...)

Call the given Python function (typically looked up from a module) with the given args... (of standard Julia types which are converted automatically to the corresponding Python types if possible), converting the return value to returntype (use a returntype of PyObject to return the unconverted Python object reference, or of PyAny to request an automated conversion)

PyCall.pyeval — Function.

pyeval(s::AbstractString, returntype::TypeTuple=PyAny, locals=PyDict{AbstractString, PyObject}(),
                            input_type=Py_eval_input; kwargs...)

This evaluates s as a Python string and returns the result converted to rtype (which defaults to PyAny). The remaining arguments are keywords that define local variables to be used in the expression.

For example, pyeval("x + y", x=1, y=2) returns 3.

PyCall.pyfunction — Method.

pyfunction(f, argtypes...; kwtypes...)

Create a Python object that wraps around the Julia function (or callable object) f. Unlike PyObject(f), this allows you to specify the argument types that the Julia function expects — giving you more control and potentially better performance.

kwtypes... should be a set of somekeyword=SomeType arguments giving the desired Julia type for each keyword somekeyword. Unspecified keywords are converted to PyAny (i.e. auto-detected) by default.

The return value ret = f(...) is still converted back to a Python object by PyObject(ret). If you want a different return-type conversion than the default of PyObject(ret), you can instead call pyfunctionret.

PyCall.pyfunctionret — Method.

pyfunctionret(f, returntype, argtypes...; kwtypes...)

Like pyfunction, but also lets you specify the returntype for conversion back to Python. In particular, if ret = f(...) is the return value of f, then it is converted to Python via PyObject(returntype(ret)).

If returntype is Any, then ret is not converted to a "native" Python type at all, and is instead converted to a "wrapped" Julia object in Python. If returntype is nothing, then the return value is discarded and nothing is returned to Python.

PyCall.pygui — Method.

pygui()

Return the current GUI toolkit as a symbol.

PyCall.pygui_start — Function.

pygui_start(gui::Symbol = pygui())

Start the event loop of a certain toolkit.

The argument gui defaults to the current default GUI, but it could be :wx, :gtk, :gtk3, :tk, or :qt.

PyCall.pygui_stop — Function.

pygui_stop(gui::Symbol = pygui())

Stop any running event loop for gui. The gui argument defaults to current default GUI.

PyCall.pyimport — Method.

pyimport(s::AbstractString)

Import the Python module s (a string or symbol) and return a pointer to it (a PyObject). Functions or other symbols in the module may then be looked up by s[name] where name is a string (for the raw PyObject) or symbol (for automatic type-conversion). Unlike the @pyimport macro, this does not define a Julia module and members cannot be accessed with s.name

PyCall.pyimport_conda — Function.

pyimport_conda(modulename, condapkg, [channel])

Returns the result of pyimport(modulename) if possible. If the module is not found, and PyCall is configured to use the Conda Python distro (via the Julia Conda package), then automatically install condapkg via Conda.add(condapkg) and then re-try the pyimport. Other Anaconda-based Python installations are also supported as long as their conda program is functioning.

If PyCall is not using Conda and the pyimport fails, throws an exception with an error message telling the user how to configure PyCall to use Conda for automated installation of the module.

The third argument, channel is an optional Anaconda "channel" to use for installing the package; this is useful for packages that are not included in the default Anaconda package listing.

PyCall.pyimport_e — Method.

pyimport_e(s::AbstractString)

Like pyimport(s), but returns an ispynull(o) == true object if the module cannot be imported rather than throwing an error.

PyCall.pyrepr — Method.

pyrepr(o::PyObject)

Return a string representation of o corresponding to repr(o) in Python.

PyCall.pystr — Method.

pystr(o::PyObject)

Return a string representation of o corresponding to str(o) in Python.

PyCall.pystring — Method.

pystring(o::PyObject)

Return a string representation of o. Normally, this corresponds to repr(o) in Python, but unlike repr it should never fail (falling back to str or to printing the raw pointer if necessary).

PyCall.pytype_mapping — Method.

pytype_mapping(pytype, jltype)

Given a Python type object pytype, tell PyCall to convert it to jltype in PyAny(object) conversions.

PyCall.pytype_query — Function.

pytype_query(o::PyObject, default=PyObject)

Given a Python object o, return the corresponding native Julia type (defaulting to default) that we convert o to in PyAny(o) conversions.

PyCall.pywrap — Function.

pywrap(o::PyObject)

This returns a wrapper w that is an anonymous module which provides (read) access to converted versions of o's members as w.member.

For example, @pyimport module as name is equivalent to const name = pywrap(pyimport("module"))

If the Python module contains identifiers that are reserved words in Julia (e.g. function), they cannot be accessed as w.member; one must instead use w.pymember(:member) (for the PyAny conversion) or w.pymember("member") (for the raw PyObject).

PyCall.__pycall! — Method.

Lowest level version of pycall!(ret, o, ...), assumes pyargsptr and kw have all their args set to Python values, so we can just call the function o. Sets ret.o to the result of the call, and returns ret::PyObject.

PyCall._pycall! — Function.

Low-level version of pycall!(ret, o, ...) for when kw is already in python friendly format but you don't have the python tuple to hold the arguments (pyargsptr). Sets ret.o to the result of the call, and returns ret::PyObject.

PyCall._pycall! — Method.

Low-level version of pycall!(ret, o, ...; kwargs...) Sets ret.o to the result of the call, and returns ret::PyObject

PyCall.anaconda_conda — Method.

anaconda_conda()

Return the path of the conda program if PyCall is configured to use an Anaconda install (but not the Conda.jl Python), and the empty string otherwise.

PyCall.def_py_class — Method.

def_py_class(type_name::AbstractString, methods::Vector;
             base_classes=[], getsets::Vector=[])

def_py_class creates a Python class whose methods are implemented in Julia. @pydef macros expand into a call to this function.

Arguments

methods: a vector of tuples (py_name::String, jl_fun::Function) py_name will be a method of the Python class, which will call jl_fun

base_classes: the Python base classes to inherit from.

Return value: the created class (::PyTypeObject)

PyCall.pyreturn — Method.

pyreturn(x) :: PyPtr

Prepare PyPtr from x for passing it to Python. If x is already a PyObject, the refcount is incremented. Otherwise a PyObject wrapping/converted from x is created.

PyCall.pystealref! — Method.

"Steal" a reference from a PyObject: return the raw PyPtr, while setting the corresponding o.o field to NULL so that no decref will be performed when o is garbage collected. (This means that you can no longer use o.) Used for passing objects to Python.