Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Base.length — Method.

    length(l::Line) -> Float64

The length of the line segment.

Rectangle.avg_min_dist — Method.

    avg_min_dist(r1::Rect, r2::Rect) -> dx::Float64, dy::Float64

Rectangles are essentially point sets. Hence, one can perceive existence of a minimum distance of one point in r1 from r2. Similar, distance would also exist for every point in r2 from r1. While, technically Euclidean distance metric can exist, the computation is fairly cumborsome. Here, we use the city block distance or L1-metric.

dx: The distance in the x-direction dy: The distance in the y-direction

The minimum distance will be zero when the rectangles are intersecting. The distance also will be lower in a specific direction if there is an overlap of the rectangles in that direction

Rectangle.intersects — Method.

    intersects(l1::Line{T}, l2::Line{T}) where {T <: Real} -> Bool

If l1 and l2 intersect each other.

Rectangle.isHorizontal — Method.

    isHorizontal(l::Line) -> Bool
    isVertcal(l::Line) -> Bool

If the Line is horizontal or vertical.

Rectangle.merge_axis_aligned — Method.

    merge_axis_aligned(alines::Vector{Line{T}}, 
                       axis::Int=1, 
                       order::Symbol=:increasing,
                       tol::T=pcTol(T)) -> Vector{Line{T}}

Given an array of axis aligned lines, if the line ends touch the lines are merged into a larger segment. Lines which are not touching the other lines are left intact.

order parameter can be in :increasing or :decreasing order in the direction of the axis.

axis parameter can be 1 for horizontal lines and 2 for vertical lines.

Rectangle.min_dist — Method.

    min_dist(r1::Rect, r2::Rect) -> dx, dy

Minimum distance or gap between two rectangles.

dx: The distance in the x-direction dy: The distance in the y-direction

The minimum distance will be zero when the rectangles are overlapping in a direction.

Rectangle.parallelogram_area — Method.

    parallelogram_area(m::Matrix) -> Float64

Area of the parallelogram. The matrix is a 2x3 matrix.

Rectangle.projectX — Method.

    projectX(r1::Rect, r2::Rect) -> (left, overlap, right)

Projects the rectangles along the X-axis and returns three parts of rectangles.

left: The left segment of the projection overlap: If there is any overlap between the rectangles right: The right segment of the projection

Each portion is returned as a tuple.

If the rectangle is part of the first rectangle, it's returned as the first element of the tuple. nothing is returned for a part when a portion is not available.

Rectangle.projectY — Method.

    projectY(r1::Rect, r2::Rect) -> (bottom, overlap, top)

Projects the rectangles along the Y-axis and returns three parts of rectangles.

bottom: The bottom segment of the projection overlap: If there is any overlap between the rectangles top: The top segment of the projection

Each portion is returned as a tuple.

If the rectangle is part of the first rectangle, it's returned as the first element of the tuple. nothing is returned for a part when a portion is not available.

Rectangle.ratio — Method.

    ratio(l1::Line{T}, p::Vector{T}) -> r::Real

If p is on l1 it divides the line at ratio r:(1-r) else nothing.

Rectangle.visibleX — Method.

    visibleX(r1::Rect, r2::Rect) -> Rect
    visibleY(r1::Rect, r2::Rect) -> Rect

Projects the rectangles along the X-axis (Y-axis) and returns a rectangle area which is completely visible from both rectangles.

nothing is returned when there is no overlap along the X-axis.