Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

RingLists.RingList — Type.

RingList{T}() creates a new, empty RingList holding elements of type T. Example: RingList{Int}().

RingList(list) creates a new RingList from the elements in the one-dimensional array list. Example: RingList([1,2,3]).

RingList(x...) creates a new RingList from the arguments. Example: RingList(1,2,3).

RingLists.insertafter! — Method.

insertafter!(a::RingList,x,y) inserts x into a after y.

RingLists.previous — Method.

previous(a,x) returns the element y so that a[y]==x.

Base.delete! — Method.

delete!(a,x) removes x from the RingList.

Base.first — Function.

first(a::RingList, true_first::Bool = false) returns an element of a. If true_first is true, then try to return the smallest element held in a.

Base.insert! — Method.

insert!(a,x) adds the element x into a. No guarantee where it will go. See also insertafter!

Base.reverse — Method.

reverse(a::RingList) returns a new RingList containing the same elements as a but in reverse order.