Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

KernelDensityEstimatePlotting.plotKDE — Method.

plotKDE(fgl, sym; axis, dims, c, levels, title)

A peneric KDE plotting function that allows marginals of higher dimensional beliefs and various keyword options.

Example:

p = kde!(randn(3,100))

plotKDE(p)
plotKDE(p, dims=[1;2], levels=3)
plotKDE(p, dims=[1])

q = kde!(5*randn(3,100))
plotKDE([p;q])
plotKDE([p;q], dims=[1;2], levels=3)
plotKDE([p;q], dims=[1])

RoMEPlotting.getColorsByLength — Function.

getColorsByLength()
getColorsByLength(len)

Standardize the length colors used by RoMEPlotting.

RoMEPlotting.plotKDEofnc — Method.

plotKDEofnc(fgl, fsym; marg, N)

plot absolute values of variables and measurement surrounding fsym factor.

RoMEPlotting.plotKDEresiduals — Method.

plotKDEresiduals(fgl, fsym; N, levels, dims, fill, api)

Trye plot relative values of variables and measurement surrounding fsym factor.

RoMEPlotting.plotLocalProduct — Method.

plotLocalProduct(fgl, lbl; N, dims, api, levels, show, dirpath, mimetype, sidelength)

Plot the proposal belief from neighboring factors to lbl in the factor graph (ignoring Bayes tree representation), and show with new product approximation for reference.

RoMEPlotting.plotLocalProduct — Method.

plotLocalProduct{T <: AbstractString}(fgl::FactorGraph, lbl::T; N::Int=100, dims::Vector{Int}=Int[])

Plot the proposal belief from neighboring factors to lbl in the factor graph (ignoring Bayes tree representation), and show with new product approximation for reference. String version is obsolete and will be deprecated.

RoMEPlotting.plotPose — Method.

plotPose(fgl, syms; levels, c, show, filepath, app)

Example: pl = plotPose(fg, [:x1; :x2; :x3])

RoMEPlotting.plotPriorsAtCliq — Method.

plotPriorsAtCliq(treel, lb, cllb)

Plot the product of priors and incoming upward messages for variable in clique.

plotPriorsAtCliq(tree, :x2, :x1[, marg=[1;2]])

RoMEPlotting.plotUpMsgsAtCliq — Method.

plotUpMsgsAtCliq(treel, lb, cllb)

Draw the up pass belief of lb from clique cllb.

plotUpMsgsAtCliq(tree, :x2, :x1)