Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

SOFA.iauASTROM — Type.

Star-independent astrometry parameters

Note: - Vectors eb, eh, em and v are all with respect to BCRS axes.

SOFA.iauLDBODY — Type.

Body parameters for light deflection

SOFA.iauA2af — Method.

Decompose radians into degrees, arcminutes, arcseconds, fraction. This function is part of the International Astronomical Union's SOFA (Standards Of Fundamental Astronomy) software collection.

Status: vector/matrix support function.

Given: ndp int resolution (Note 1) angle double angle in radians

Returned: sign char '+' or '-' idmsf int[4] degrees, arcminutes, arcseconds, fraction

Called: iauD2tf decompose days to hms

Notes: 1. The argument ndp is interpreted as follows: ndp resolution : ...0000 00 00 -7 1000 00 00 -6 100 00 00 -5 10 00 00 -4 1 00 00 -3 0 10 00 -2 0 01 00 -1 0 00 10 0 0 00 01 1 0 00 00.1 2 0 00 00.01 3 0 00 00.001 : 0 00 00.000...

2. The largest positive useful value for ndp is determined by the

size of angle, the format of doubles on the target platform, and the risk of overflowing idmsf[3]. On a typical platform, for angle up to 2pi, the available floating-point precision might correspond to ndp=12. However, the practical limit is typically ndp=9, set by the capacity of a 32-bit int, or ndp=4 if int is only 16 bits.

3. The absolute value of angle may exceed 2pi.  In cases where it

does not, it is up to the caller to test for and handle the case where angle is very nearly 2pi and rounds up to 360 degrees, by testing for idmsf[0]=360 and setting idmsf[0-3] to zero.

This revision: 2013 June 18

SOFA release 2018-01-30

SOFA.iauA2tf — Method.

Decompose radians into hours, minutes, seconds, fraction.

This function is part of the International Astronomical Union's SOFA (Standards Of Fundamental Astronomy) software collection.

Status: vector/matrix support function.

Given: ndp int resolution (Note 1) angle double angle in radians

Returned: sign char '+' or '-' ihmsf int[4] hours, minutes, seconds, fraction

Called: iauD2tf decompose days to hms

Notes:

1. The argument ndp is interpreted as follows:
    ndp         resolution
     :      ...0000 00 00
    -7         1000 00 00
    -6          100 00 00
    -5           10 00 00
    -4            1 00 00
    -3            0 10 00
    -2            0 01 00
    -1            0 00 10
     0            0 00 01
     1            0 00 00.1
     2            0 00 00.01
     3            0 00 00.001
     :            0 00 00.000...

2. The largest positive useful value for ndp is determined by the
size of angle, the format of doubles on the target platform, and
the risk of overflowing ihmsf[3].  On a typical platform, for
angle up to 2pi, the available floating-point precision might
correspond to ndp=12.  However, the practical limit is typically
ndp=9, set by the capacity of a 32-bit int, or ndp=4 if int is
only 16 bits.

3. The absolute value of angle may exceed 2pi.  In cases where it
does not, it is up to the caller to test for and handle the
case where angle is very nearly 2pi and rounds up to 24 hours,
by testing for ihmsf[0]=24 and setting ihmsf[0-3] to zero.

This revision: 2013 July 31

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAb — Method.

Apply aberration to transform natural direction into proper direction.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: pnat double[3] natural direction to the source (unit vector) v double[3] observer barycentric velocity in units of c s double distance between the Sun and the observer (au) bm1 double sqrt(1-|v|^2): reciprocal of Lorenz factor

Returned: ppr double[3] proper direction to source (unit vector)

Notes:

The algorithm is based on Expr. (7.40) in the Explanatory

Supplement (Urban & Seidelmann 2013), but with the following changes: - Rigorous rather than approximate normalization is applied. -The gravitational potential term from Expr. (7) in Klioner (2003) is added, taking into account only the Sun's contribution. This has a maximum effect of about 0.4 microarcsecond.

In almost all cases, the maximum accuracy will be limited by the

supplied velocity. For example, if the SOFA iauEpv00 function is used, errors of up to 5 microarcseconds could occur.

References:

Urban, S. & Seidelmann, P. K. (eds), Explanatory Supplement to

the Astronomical Almanac, 3rd ed., University Science Books (2013).

Klioner, Sergei A., "A practical relativistic model for micro-

arcsecond astrometry in space", Astr. J. 125, 1580-1597 (2003).

Called: iauPdp scalar product of two p-vectors

This revision: 2013 October 9

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAe2hd — Method.

Horizon to equatorial coordinates: transform azimuth and altitude to hour angle and declination.

Given: az double azimuth el double altitude (informally, elevation) phi double site latitude

Returned: ha double hour angle (local) dec double declination

Notes:

All the arguments are angles in radians.

The sign convention for azimuth is north zero, east +pi/2.

HA is returned in the range +/-pi. Declination is returned in

the range +/-pi/2.

The latitude phi is pi/2 minus the angle between the Earth's

rotation axis and the adopted zenith. In many applications it will be sufficient to use the published geodetic latitude of the site. In very precise (sub-arcsecond) applications, phi can be corrected for polar motion.

The azimuth az must be with respect to the rotational north pole,

as opposed to the ITRS pole, and an azimuth with respect to north on a map of the Earth's surface will need to be adjusted for polar motion if sub-arcsecond accuracy is required.

Should the user wish to work with respect to the astronomical

zenith rather than the geodetic zenith, phi will need to be adjusted for deflection of the vertical (often tens of arcseconds), and the zero point of ha will also be affected.

The transformation is the same as Ve = Ry(phi-pi/2)Rz(pi)Vh,

where Ve and Vh are lefthanded unit vectors in the (ha,dec) and (az,el) systems respectively and Rz and Ry are rotations about first the z-axis and then the y-axis. (n.b. Rz(pi) simply reverses the signs of the x and y components.) For efficiency, the algorithm is written out rather than calling other utility functions. For applications that require even greater efficiency, additional savings are possible if constant terms such as functions of latitude are computed once and for all.

Again for efficiency, no range checking of arguments is carried

out.

Last revision: 2017 September 12

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAf2a — Method.

Convert degrees, arcminutes, arcseconds to radians.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: s char sign: '-' = negative, otherwise positive ideg int degrees iamin int arcminutes asec double arcseconds

Returned: rad double angle in radians

Returned (function value): int status: 0 = OK 1 = ideg outside range 0-359 2 = iamin outside range 0-59 3 = asec outside range 0-59.999... Notes: 1. The result is computed even if any of the range checks fail.

2. Negative ideg, iamin and/or asec produce a warning status, but
the absolute value is used in the conversion.

3. If there are multiple errors, the status value reflects only the
first, the smallest taking precedence.

This revision:  2013 June 18

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAnp — Method.

Normalize angle into the range 0 <= a < 2pi.

This function is part of the International Astronomical Union's SOFA (Standards Of Fundamental Astronomy) software collection.

Status: vector/matrix support function.

Given: a double angle (radians)

Returned (function value): double angle in range 0-2pi

This revision: 2013 June 18

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAnpm — Method.

Normalize angle into the range -pi <= a < +pi.

This function is part of the International Astronomical Union's SOFA (Standards Of Fundamental Astronomy) software collection.

Status: vector/matrix support function.

Given: a double angle (radians)

Returned (function value): double angle in range +/-pi

This revision: 2013 June 18

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauApcg — Method.

For a geocentric observer, prepare star-independent astrometry parameters for transformations between ICRS and GCRS coordinates. The Earth ephemeris is supplied by the caller.

The parameters produced by this function are required in the parallax, light deflection and aberration parts of the astrometric transformation chain.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: date1 double TDB as a 2-part... date2 double ...Julian Date (Note 1) ebpv double[2][3] Earth barycentric pos/vel (au, au/day) ehp double[3] Earth heliocentric position (au)

Returned: astrom iauASTROM* star-independent astrometry parameters: pmt double PM time interval (SSB, Julian years) eb double[3] SSB to observer (vector, au) eh double[3] Sun to observer (unit vector) em double distance from Sun to observer (au) v double[3] barycentric observer velocity (vector, c) bm1 double sqrt(1-|v|^2): reciprocal of Lorenz factor bpn double[3][3] bias-precession-nutation matrix along double unchanged xpl double unchanged ypl double unchanged sphi double unchanged cphi double unchanged diurab double unchanged eral double unchanged refa double unchanged refb double unchanged

Notes: 1. The TDB date date1+date2 is a Julian Date, apportioned in any convenient way between the two arguments. For example, JD(TDB)=2450123.7 could be expressed in any of these ways, among others:

        date1          date2

     2450123.7           0.0       (JD method)
     2451545.0       -1421.3       (J2000 method)
     2400000.5       50123.2       (MJD method)
     2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases
where the loss of several decimal digits of resolution is
acceptable.  The J2000 method is best matched to the way the
argument is handled internally and will deliver the optimum
resolution.  The MJD method and the date & time methods are both
good compromises between resolution and convenience.  For most
applications of this function the choice will not be at all
critical.

TT can be used instead of TDB without any significant impact on
accuracy.

2. All the vectors are with respect to BCRS axes.

3. This is one of several functions that inserts into the astrom
structure star-independent parameters needed for the chain of
astrometric transformations ICRS <-> GCRS <-> CIRS <-> observed.

The various functions support different classes of observer and
portions of the transformation chain:

      functions         observer        transformation

   iauApcg iauApcg13    geocentric      ICRS <-> GCRS
   iauApci iauApci13    terrestrial     ICRS <-> CIRS
   iauApco iauApco13    terrestrial     ICRS <-> observed
   iauApcs iauApcs13    space           ICRS <-> GCRS
   iauAper iauAper13    terrestrial     update Earth rotation
   iauApio iauApio13    terrestrial     CIRS <-> observed

Those with names ending in "13" use contemporary SOFA models to
compute the various ephemerides.  The others accept ephemerides
supplied by the caller.

The transformation from ICRS to GCRS covers space motion,
parallax, light deflection, and aberration.  From GCRS to CIRS
comprises frame bias and precession-nutation.  From CIRS to
observed takes account of Earth rotation, polar motion, diurnal
aberration and parallax (unless subsumed into the ICRS <-> GCRS
transformation), and atmospheric refraction.

4. The context structure astrom produced by this function is used by
iauAtciq* and iauAticq*.

Called: iauApcs astrometry parameters, ICRS-GCRS, space observer

This revision: 2013 October 9

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauApcg13 — Method.

For a geocentric observer, prepare star-independent astrometry parameters for transformations between ICRS and GCRS coordinates. The caller supplies the date, and SOFA models are used to predict the Earth ephemeris.

The parameters produced by this function are required in the parallax, light deflection and aberration parts of the astrometric transformation chain.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: date1 double TDB as a 2-part... date2 double ...Julian Date (Note 1)

Returned: astrom iauASTROM* star-independent astrometry parameters: pmt double PM time interval (SSB, Julian years) eb double[3] SSB to observer (vector, au) eh double[3] Sun to observer (unit vector) em double distance from Sun to observer (au) v double[3] barycentric observer velocity (vector, c) bm1 double sqrt(1-|v|^2): reciprocal of Lorenz factor bpn double[3][3] bias-precession-nutation matrix along double unchanged xpl double unchanged ypl double unchanged sphi double unchanged cphi double unchanged diurab double unchanged eral double unchanged refa double unchanged refb double unchanged

Notes:

The TDB date date1+date2 is a Julian Date, apportioned in any

convenient way between the two arguments. For example, JD(TDB)=2450123.7 could be expressed in any of these ways, among others:

        date1          date2

     2450123.7           0.0       (JD method)
     2451545.0       -1421.3       (J2000 method)
     2400000.5       50123.2       (MJD method)
     2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 method is best matched to the way the argument is handled internally and will deliver the optimum resolution. The MJD method and the date & time methods are both good compromises between resolution and convenience. For most applications of this function the choice will not be at all critical.

TT can be used instead of TDB without any significant impact on accuracy.

All the vectors are with respect to BCRS axes.

In cases where the caller wishes to supply his own Earth

ephemeris, the function iauApcg can be used instead of the present function.

This is one of several functions that inserts into the astrom

structure star-independent parameters needed for the chain of astrometric transformations ICRS <-> GCRS <-> CIRS <-> observed.

The various functions support different classes of observer and portions of the transformation chain:

      functions         observer        transformation

   iauApcg iauApcg13    geocentric      ICRS <-> GCRS
   iauApci iauApci13    terrestrial     ICRS <-> CIRS
   iauApco iauApco13    terrestrial     ICRS <-> observed
   iauApcs iauApcs13    space           ICRS <-> GCRS
   iauAper iauAper13    terrestrial     update Earth rotation
   iauApio iauApio13    terrestrial     CIRS <-> observed

Those with names ending in "13" use contemporary SOFA models to compute the various ephemerides. The others accept ephemerides supplied by the caller.

The transformation from ICRS to GCRS covers space motion, parallax, light deflection, and aberration. From GCRS to CIRS comprises frame bias and precession-nutation. From CIRS to observed takes account of Earth rotation, polar motion, diurnal aberration and parallax (unless subsumed into the ICRS <-> GCRS transformation), and atmospheric refraction.

The context structure astrom produced by this function is used by

iauAtciq* and iauAticq*.

Called: iauEpv00 Earth position and velocity iauApcg astrometry parameters, ICRS-GCRS, geocenter

This revision: 2013 October 9

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauApci — Method.

For a terrestrial observer, prepare star-independent astrometry parameters for transformations between ICRS and geocentric CIRS coordinates. The Earth ephemeris and CIP/CIO are supplied by the caller.

The parameters produced by this function are required in the parallax, light deflection, aberration, and bias-precession-nutation parts of the astrometric transformation chain.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: date1 double TDB as a 2-part... date2 double ...Julian Date (Note 1) ebpv double[2][3] Earth barycentric position/velocity (au, au/day) ehp double[3] Earth heliocentric position (au) x,y double CIP X,Y (components of unit vector) s double the CIO locator s (radians)

Returned: astrom iauASTROM* star-independent astrometry parameters: pmt double PM time interval (SSB, Julian years) eb double[3] SSB to observer (vector, au) eh double[3] Sun to observer (unit vector) em double distance from Sun to observer (au) v double[3] barycentric observer velocity (vector, c) bm1 double sqrt(1-|v|^2): reciprocal of Lorenz factor bpn double[3][3] bias-precession-nutation matrix along double unchanged xpl double unchanged ypl double unchanged sphi double unchanged cphi double unchanged diurab double unchanged eral double unchanged refa double unchanged refb double unchanged

Notes:

1. The TDB date date1+date2 is a Julian Date, apportioned in any
convenient way between the two arguments.  For example,
JD(TDB)=2450123.7 could be expressed in any of these ways, among
others:

    date1          date2

    2450123.7           0.0       (JD method)
    2451545.0       -1421.3       (J2000 method)
    2400000.5       50123.2       (MJD method)
    2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases
where the loss of several decimal digits of resolution is
acceptable.  The J2000 method is best matched to the way the
argument is handled internally and will deliver the optimum
resolution.  The MJD method and the date & time methods are both
good compromises between resolution and convenience.  For most
applications of this function the choice will not be at all
critical.

TT can be used instead of TDB without any significant impact on
accuracy.

2. All the vectors are with respect to BCRS axes.

3. In cases where the caller does not wish to provide the Earth
ephemeris and CIP/CIO, the function iauApci13 can be used instead
of the present function.  This computes the required quantities
using other SOFA functions.

4. This is one of several functions that inserts into the astrom
structure star-independent parameters needed for the chain of
astrometric transformations ICRS <-> GCRS <-> CIRS <-> observed.

The various functions support different classes of observer and
portions of the transformation chain:

    functions         observer        transformation

iauApcg iauApcg13    geocentric      ICRS <-> GCRS
iauApci iauApci13    terrestrial     ICRS <-> CIRS
iauApco iauApco13    terrestrial     ICRS <-> observed
iauApcs iauApcs13    space           ICRS <-> GCRS
iauAper iauAper13    terrestrial     update Earth rotation
iauApio iauApio13    terrestrial     CIRS <-> observed

Those with names ending in "13" use contemporary SOFA models to
compute the various ephemerides.  The others accept ephemerides
supplied by the caller.

The transformation from ICRS to GCRS covers space motion,
parallax, light deflection, and aberration.  From GCRS to CIRS
comprises frame bias and precession-nutation.  From CIRS to
observed takes account of Earth rotation, polar motion, diurnal
aberration and parallax (unless subsumed into the ICRS <-> GCRS
transformation), and atmospheric refraction.

5. The context structure astrom produced by this function is used by
iauAtciq* and iauAticq*.

Called: iauApcg astrometry parameters, ICRS-GCRS, geocenter iauC2ixys celestial-to-intermediate matrix, given X,Y and s

This revision: 2013 September 25

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauApci13 — Method.

For a terrestrial observer, prepare star-independent astrometry parameters for transformations between ICRS and geocentric CIRS coordinates. The caller supplies the date, and SOFA models are used to predict the Earth ephemeris and CIP/CIO.

The parameters produced by this function are required in the parallax, light deflection, aberration, and bias-precession-nutation parts of the astrometric transformation chain.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: date1 double TDB as a 2-part... date2 double ...Julian Date (Note 1)

Returned: astrom iauASTROM* star-independent astrometry parameters: pmt double PM time interval (SSB, Julian years) eb double[3] SSB to observer (vector, au) eh double[3] Sun to observer (unit vector) em double distance from Sun to observer (au) v double[3] barycentric observer velocity (vector, c) bm1 double sqrt(1-|v|^2): reciprocal of Lorenz factor bpn double[3][3] bias-precession-nutation matrix along double unchanged xpl double unchanged ypl double unchanged sphi double unchanged cphi double unchanged diurab double unchanged eral double unchanged refa double unchanged refb double unchanged eo double* equation of the origins (ERA-GST)

Notes:

The TDB date date1+date2 is a Julian Date, apportioned in any

convenient way between the two arguments. For example, JD(TDB)=2450123.7 could be expressed in any of these ways, among others:

     date1          date2

  2450123.7           0.0       (JD method)
  2451545.0       -1421.3       (J2000 method)
  2400000.5       50123.2       (MJD method)
  2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 method is best matched to the way the argument is handled internally and will deliver the optimum resolution. The MJD method and the date & time methods are both good compromises between resolution and convenience. For most applications of this function the choice will not be at all critical.

TT can be used instead of TDB without any significant impact on accuracy.

All the vectors are with respect to BCRS axes.

In cases where the caller wishes to supply his own Earth

ephemeris and CIP/CIO, the function iauApci can be used instead of the present function.

This is one of several functions that inserts into the astrom

structure star-independent parameters needed for the chain of astrometric transformations ICRS <-> GCRS <-> CIRS <-> observed.

The various functions support different classes of observer and portions of the transformation chain:

     functions         observer        transformation

  iauApcg iauApcg13    geocentric      ICRS <-> GCRS
  iauApci iauApci13    terrestrial     ICRS <-> CIRS
  iauApco iauApco13    terrestrial     ICRS <-> observed
  iauApcs iauApcs13    space           ICRS <-> GCRS
  iauAper iauAper13    terrestrial     update Earth rotation
  iauApio iauApio13    terrestrial     CIRS <-> observed

Those with names ending in "13" use contemporary SOFA models to compute the various ephemerides. The others accept ephemerides supplied by the caller.

The transformation from ICRS to GCRS covers space motion, parallax, light deflection, and aberration. From GCRS to CIRS comprises frame bias and precession-nutation. From CIRS to observed takes account of Earth rotation, polar motion, diurnal aberration and parallax (unless subsumed into the ICRS <-> GCRS transformation), and atmospheric refraction.

The context structure astrom produced by this function is used by

iauAtciq* and iauAticq*.

Called: iauEpv00 Earth position and velocity iauPnm06a classical NPB matrix, IAU 2006/2000A iauBpn2xy extract CIP X,Y coordinates from NPB matrix iauS06 the CIO locator s, given X,Y, IAU 2006 iauApci astrometry parameters, ICRS-CIRS iauEors equation of the origins, given NPB matrix and s

This revision: 2013 October 9

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAper — Method.

In the star-independent astrometry parameters, update only the Earth rotation angle, supplied by the caller explicitly.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: theta double Earth rotation angle (radians, Note 2) astrom iauASTROM* star-independent astrometry parameters: pmt double not used eb double[3] not used eh double[3] not used em double not used v double[3] not used bm1 double not used bpn double[3][3] not used along double longitude + s' (radians) xpl double not used ypl double not used sphi double not used cphi double not used diurab double not used eral double not used refa double not used refb double not used

Returned: astrom iauASTROM* star-independent astrometry parameters: pmt double unchanged eb double[3] unchanged eh double[3] unchanged em double unchanged v double[3] unchanged bm1 double unchanged bpn double[3][3] unchanged along double unchanged xpl double unchanged ypl double unchanged sphi double unchanged cphi double unchanged diurab double unchanged eral double "local" Earth rotation angle (radians) refa double unchanged refb double unchanged

Notes:

This function exists to enable sidereal-tracking applications to

avoid wasteful recomputation of the bulk of the astrometry parameters: only the Earth rotation is updated.

For targets expressed as equinox based positions, such as

classical geocentric apparent (RA,Dec), the supplied theta can be Greenwich apparent sidereal time rather than Earth rotation angle.

The function iauAper13 can be used instead of the present

function, and starts from UT1 rather than ERA itself.

This is one of several functions that inserts into the astrom

structure star-independent parameters needed for the chain of astrometric transformations ICRS <-> GCRS <-> CIRS <-> observed.

The various functions support different classes of observer and portions of the transformation chain:

     functions         observer        transformation

  iauApcg iauApcg13    geocentric      ICRS <-> GCRS
  iauApci iauApci13    terrestrial     ICRS <-> CIRS
  iauApco iauApco13    terrestrial     ICRS <-> observed
  iauApcs iauApcs13    space           ICRS <-> GCRS
  iauAper iauAper13    terrestrial     update Earth rotation
  iauApio iauApio13    terrestrial     CIRS <-> observed

Those with names ending in "13" use contemporary SOFA models to compute the various ephemerides. The others accept ephemerides supplied by the caller.

The transformation from ICRS to GCRS covers space motion, parallax, light deflection, and aberration. From GCRS to CIRS comprises frame bias and precession-nutation. From CIRS to observed takes account of Earth rotation, polar motion, diurnal aberration and parallax (unless subsumed into the ICRS <-> GCRS transformation), and atmospheric refraction.

This revision: 2013 September 25

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAper13 — Method.

In the star-independent astrometry parameters, update only the Earth rotation angle. The caller provides UT1, (n.b. not UTC).

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: ut11 double UT1 as a 2-part... ut12 double ...Julian Date (Note 1) astrom iauASTROM* star-independent astrometry parameters: pmt double not used eb double[3] not used eh double[3] not used em double not used v double[3] not used bm1 double not used bpn double[3][3] not used along double longitude + s' (radians) xpl double not used ypl double not used sphi double not used cphi double not used diurab double not used eral double not used refa double not used refb double not used

Returned: astrom iauASTROM* star-independent astrometry parameters: pmt double unchanged eb double[3] unchanged eh double[3] unchanged em double unchanged v double[3] unchanged bm1 double unchanged bpn double[3][3] unchanged along double unchanged xpl double unchanged ypl double unchanged sphi double unchanged cphi double unchanged diurab double unchanged eral double "local" Earth rotation angle (radians) refa double unchanged refb double unchanged

Notes:

The UT1 date (n.b. not UTC) ut11+ut12 is a Julian Date, apportioned in any convenient way between the arguments ut11 and ut12. For example, JD(UT1)=2450123.7 could be expressed in any of these ways, among others:

  ut11           ut12

2450123.7 0.0 (JD method) 2451545.0 -1421.3 (J2000 method) 2400000.5 50123.2 (MJD method) 2450123.5 0.2 (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 and MJD methods are good compromises between resolution and convenience. The date & time method is best matched to the algorithm used: maximum precision is delivered when the ut11 argument is for 0hrs UT1 on the day in question and the ut12 argument lies in the range 0 to 1, or vice versa.

If the caller wishes to provide the Earth rotation angle itself, the function iauAper can be used instead. One use of this technique is to substitute Greenwich apparent sidereal time and thereby to support equinox based transformations directly.

This is one of several functions that inserts into the astrom structure star-independent parameters needed for the chain of astrometric transformations ICRS <-> GCRS <-> CIRS <-> observed.

The various functions support different classes of observer and portions of the transformation chain:

  functions         observer        transformation

iauApcg iauApcg13 geocentric ICRS <-> GCRS iauApci iauApci13 terrestrial ICRS <-> CIRS iauApco iauApco13 terrestrial ICRS <-> observed iauApcs iauApcs13 space ICRS <-> GCRS iauAper iauAper13 terrestrial update Earth rotation iauApio iauApio13 terrestrial CIRS <-> observed

Those with names ending in "13" use contemporary SOFA models to compute the various ephemerides. The others accept ephemerides supplied by the caller.

The transformation from ICRS to GCRS covers space motion, parallax, light deflection, and aberration. From GCRS to CIRS comprises frame bias and precession-nutation. From CIRS to observed takes account of Earth rotation, polar motion, diurnal aberration and parallax (unless subsumed into the ICRS <-> GCRS transformation), and atmospheric refraction.

Called: iauAper astrometry parameters: update ERA iauEra00 Earth rotation angle, IAU 2000

This revision: 2013 September 25

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.

SOFA.iauAtciqz — Method.

Quick ICRS to CIRS transformation, given precomputed star- independent astrometry parameters, and assuming zero parallax and proper motion.

Use of this function is appropriate when efficiency is important and where many star positions are to be transformed for one date. The star-independent parameters can be obtained by calling one of the functions iauApci[13], iauApcg[13], iauApco[13] or iauApcs[13].

The corresponding function for the case of non-zero parallax and proper motion is iauAtciq.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Given: rc,dc double ICRS astrometric RA,Dec (radians) astrom iauASTROM* star-independent astrometry parameters: pmt double PM time interval (SSB, Julian years) eb double[3] SSB to observer (vector, au) eh double[3] Sun to observer (unit vector) em double distance from Sun to observer (au) v double[3] barycentric observer velocity (vector, c) bm1 double sqrt(1-|v|^2): reciprocal of Lorenz factor bpn double[3][3] bias-precession-nutation matrix along double longitude + s' (radians) xpl double polar motion xp wrt local meridian (radians) ypl double polar motion yp wrt local meridian (radians) sphi double sine of geodetic latitude cphi double cosine of geodetic latitude diurab double magnitude of diurnal aberration vector eral double "local" Earth rotation angle (radians) refa double refraction constant A (radians) refb double refraction constant B (radians)

Returned: ri,di double CIRS RA,Dec (radians)

Note:

All the vectors are with respect to BCRS axes.

References:

Urban, S. & Seidelmann, P. K. (eds), Explanatory Supplement to the Astronomical Almanac, 3rd ed., University Science Books (2013).

Klioner, Sergei A., "A practical relativistic model for micro- arcsecond astrometry in space", Astr. J. 125, 1580-1597 (2003).

Called: iauS2c spherical coordinates to unit vector iauLdsun light deflection due to Sun iauAb stellar aberration iauRxp product of r-matrix and p-vector iauC2s p-vector to spherical iauAnp normalize angle into range +/- pi

This revision: 2013 October 9

SOFA release 2018-01-30

Copyright (C) 2018 IAU SOFA Board. See notes at end.