Sched.jl

Sched.jl

A Julia event scheduler inspired by Python sched.

Sched.Sched — Module.

A generally useful event scheduler class. Each instance of this class manages its own queue. No multi-threading is implied; you are supposed to hack that yourself, or use a single instance per application. Each instance is parametrized with two functions, one that is supposed to return the current time, one that is supposed to implement a delay. You can implement real-time scheduling by substituting time and sleep from built-in module time, or you can implement simulated time by writing your own functions. This can also be used to integrate scheduling with STDWIN events; the delay function is allowed to modify the queue. Time can be expressed as integers or floating point numbers, as long as it is consistent. Events are specified by tuples (time, priority, action, argument, kwargs). As in UNIX, lower priority numbers mean higher priority; in this way the queue can be maintained as a priority queue. Execution of the event means calling the action function, passing it the argument sequence in "argument" (remember that in Python, multiple function arguments are be packed in a sequence) and keyword parameters in "kwargs". The action function may be an instance method so it has another way to reference private data (besides global variables).

Install

Sched is a registered package. To add it to your Julia packages, simply do the following in REPL:

Pkg.add("Sched")

Usage

Download example

Contents

Sched.jl

Install

Usage

Contents

Syntax

Package Internals

See also

Syntax

Sched.Scheduler — Type.

Scheduler(; timefunc=_time, delayfunc=sleep)

Initialize a new Scheduler instance, passing optionaly the time and delay functions

The scheduler struct defines a generic interface to scheduling events. It needs two functions to actually deal with the “outside world”

The timefunc should be callable without arguments, and return a number (the “time”, in any units whatsoever). timefunc default is UTCDateTimeFunc.

The delayfunc function should be callable with one argument, compatible with the output of timefunc, and should delay that many time units. delayfunc will also be called with the argument 0 after each event is run to allow other threads an opportunity to run in multi-threaded applications.

Sched.enterabs — Function.

enterabs(sched, time_, priority, action, args...; kwargs...)

Enter a new event in the queue at an absolute time. Returns an ID for the event which can be used to remove it, if necessary.

Sched.enter — Function.

enter(sched, delay, priority, action, args...; kwargs...)

Enter a new event in the queue at a relative time. A variant of enterabs that specifies the time as a relative time. This is actually the more commonly used interface.

Sched.cancel — Function.

cancel(sched, event)

Remove an event from the queue. This must be presented the ID as returned by enter(). If the event is not in the queue, this raises ValueError.

Base.isempty — Function.

isempty(collection) -> Bool

Determine whether a collection is empty (has no elements).

Examples

julia> isempty([])
true

julia> isempty([1 2 3])
false

source

isempty(cb)

Test whether the buffer is empty.

isempty(hs::HalfSpace)::Bool

Return if a half-space is empty or not.

Input

hs – half-space

Output

false.

isempty(P::AbstractPolytope)::Bool

Determine whether a polytope is empty.

Input

P – abstract polytope

Output

true if the given polytope contains no vertices, and false otherwise.

Algorithm

This algorithm checks whether the vertices_list of the given polytope is empty or not.

isempty(S::AbstractCentrallySymmetric)::Bool

Return if a centrally symmetric set is empty or not.

Input

S – centrally symmetric set

Output

false.

isempty(P::AbstractCentrallySymmetricPolytope)::Bool

Return if a centrally symmetric, polytopic set is empty or not.

Input

P – centrally symmetric, polytopic set

Output

false.

isempty(∅::EmptySet)::Bool

Return if the empty set is empty or not.

Input

∅ – empty set

Output

true.

isempty(P::HPoly{N}; [solver]=GLPKSolverLP())::Bool where {N<:Real}

Determine whether a polyhedron is empty.

Input

P – polyhedron

backend – (optional, default: default_polyhedra_backend(P, N)) the polyhedral computations backend

solver – (optional, default: GLPKSolverLP()) LP solver backend

Output

true if and only if the constraints are inconsistent.

Algorithm

This function uses Polyhedra.isempty which evaluates the feasibility of the LP whose feasible set is determined by the set of constraints and whose objective function is zero.

Notes

This implementation uses GLPKSolverLP as linear programming backend by default.

isempty(hp::Hyperplane)::Bool

Return if a hyperplane is empty or not.

Input

hp – hyperplane

Output

false.

isempty(L::Line)::Bool

Return if a line is empty or not.

Input

L – line

Output

false.

isempty(cp::CartesianProduct)::Bool

Return if a Cartesian product is empty or not.

Input

cp – Cartesian product

Output

true iff any of the sub-blocks is empty.

isempty(cpa::CartesianProductArray)::Bool

Return if a Cartesian product is empty or not.

Input

cp – Cartesian product

Output

true iff any of the sub-blocks is empty.

isempty(ch::ConvexHull)::Bool

Return if a convex hull of two convex sets is empty or not.

Input

ch – convex hull

Output

true iff both wrapped sets are empty.

isempty(cha::ConvexHullArray)::Bool

Return if a convex hull array is empty or not.

Input

cha – convex hull array

Output

true iff all wrapped sets are empty.

isempty(em::ExponentialMap)::Bool

Return if an exponential map is empty or not.

Input

em – exponential map

Output

true iff the wrapped set is empty.

isempty(eprojmap::ExponentialProjectionMap)::Bool

Return if an exponential projection map is empty or not.

Input

eprojmap – exponential projection map

Output

true iff the wrapped set is empty.

isempty(cap::Intersection)::Bool

Return if the intersection is empty or not.

Input

cap – intersection of two convex sets

Output

true iff the intersection is empty.

Notes

The result will be cached, so a second query will be fast.

isempty(lm::LinearMap)::Bool

Return if a linear map is empty or not.

Input

lm – linear map

Output

true iff the wrapped set is empty.

isempty(ms::MinkowskiSum)::Bool

Return if a Minkowski sum is empty or not.

Input

ms – Minkowski sum

Output

true iff any of the wrapped sets are empty.

isempty(msa::MinkowskiSumArray)::Bool

Return if a Minkowski sum array is empty or not.

Input

msa – Minkowski sum array

Output

true iff any of the wrapped sets are empty.

isempty(cms::CacheMinkowskiSum)::Bool

Return if a caching Minkowski sum array is empty or not.

Input

cms – caching Minkowski sum

Output

true iff any of the wrapped sets are empty.

Notes

Forgotten sets cannot be checked anymore. Usually they have been empty because otherwise the support vector query should have crashed before. In that case, the caching Minkowski sum should not be used further.

isempty(p::Rep, solver::MathProgBase.AbstractMathProgSolver=Polyhedra.solver(p))

Check whether the polyhedron p is empty by using the solver solver.

isempty(sched) -> Bool

Check whether the queue is empty.

Base.run — Function.

run(command, args...; wait::Bool = true)

Run a command object, constructed with backticks. Throws an error if anything goes wrong, including the process exiting with a non-zero status (when wait is true).

If wait is false, the process runs asynchronously. You can later wait for it and check its exit status by calling success on the returned process object.

When wait is false, the process' I/O streams are directed to devnull. When wait is true, I/O streams are shared with the parent process. Use pipeline to control I/O redirection.

source

run(sched; blocking=true)

Execute events until the queue is empty. If blocking is False executes the scheduled events due to expire soonest (if any) and then return the deadline of the next scheduled call in the scheduler. When there is a positive delay until the first event, the delay function is called and the event is left in the queue; otherwise, the event is removed from the queue and executed (its action function is called, passing it the argument). If the delay function returns prematurely, it is simply restarted. It is legal for both the delay function and the action function to modify the queue or to raise an exception; exceptions are not caught but the scheduler's state remains well-defined so run() may be called again. A questionable hack is added to allow other threads to run: just after an event is executed, a delay of 0 is executed, to avoid monopolizing the CPU when other threads are also runnable.

Sched.queue — Function.

queue(sched)

Return an ordered list of upcoming events.

Package Internals

Sched.Event — Type.

Event(time_, priority, action, args...; kwargs...)

Event structure

time_: Numeric type compatible with the return value of the timefunc function passed to the constructor.'

priority: Events scheduled for the same time will be executed in the order of their priority.

action: Executing the event means executing action(args...; kwargs...)

args: args is a sequence holding the positional arguments for the action.

kwargs: kwargs is a dictionary holding the keyword arguments for the action.

Sched.Priority — Type.

Priority(time_, priority)

Priority of events

Sched.TimeFunc — Type.

Abstract type for struct that returns real-time or simulated time when called (functor)

Sched.UTCDateTimeFuncStruct — Type.

UTCDateTimeFuncStruct()

Functor that return real-time as DateTime (UTC) when called

Sched.FloatTimeFuncStruct — Type.

FloatTimeFuncStruct()

Functor that return real-time as Float when called

See also

ExtensibleScheduler.jl a more advanced and extensible Julia events scheduler

https://discourse.julialang.org/t/julia-cron-like-event-scheduler/6899