Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

SemidefiniteOptInterface.coefficienttype — Function.

coefficienttype(optimizer::AbstractSDOptimizer)

Returns the coefficienttype that should be used for optimizer.

SemidefiniteOptInterface.getX — Function.

getX(optimizer::AbstractSDOptimizer)

Returns the solution X as a block matrix.

SemidefiniteOptInterface.getZ — Function.

getZ(optimizer::AbstractSDOptimizer)

Returns the solution Z.

SemidefiniteOptInterface.getblockdimension — Function.

getblockdimension(optimizer::AbstractSDOptimizer, blk)

Returns the dimension of the block blk.

SemidefiniteOptInterface.getconstraintcoefficients — Function.

getconstraintcoefficients(optimizer::AbstractSDOptimizer, constr::Integer)

Return the list of entries blk, i, j of the matrix of the constraint constr.

SemidefiniteOptInterface.getconstraintconstant — Function.

getconstraintconstant(optimizer::AbstractSDOptimizer, constr::Integer)

Sets the entry constr of b to val.

SemidefiniteOptInterface.getdualobjectivevalue — Function.

getdualobjectivevalue(optimizer::AbstractSDOptimizer)

Returns the dual objective value.

SemidefiniteOptInterface.getnumberofblocks — Function.

getnumberofblocks(optimizer::AbstractSDOptimizer)

Returns the number of blocks of the block matrix.

SemidefiniteOptInterface.getnumberofconstraints — Function.

getnumberofconstraints(optimizer::AbstractSDOptimizer)

Returns the number of constraints of the model.

SemidefiniteOptInterface.getobjectivecoefficients — Function.

getobjectivecoefficients(optimizer::AbstractSDOptimizer)

Return the list of entries blk, i, j of the objective matrix.

SemidefiniteOptInterface.getprimalobjectivevalue — Function.

getprimalobjectivevalue(optimizer::AbstractSDOptimizer)

Returns the primal objective value.

SemidefiniteOptInterface.gety — Function.

gety(optimizer::AbstractSDOptimizer)

Returns the solution y.

SemidefiniteOptInterface.init! — Function.

init!(optimizer::AbstractSDOptimizer, blkdims::Vector{Int}, nconstrs::Integer)

Initialize the optimizer with nconstrs constraints and blkdims blocks.

SemidefiniteOptInterface.setconstraintcoefficient! — Function.

setconstraintcoefficient!(optimizer::AbstractSDOptimizer, val, constr::Integer, blk::Integer, i::Integer, j::Integer)

Set the entry i, j of the block blk of the matrix of the constraint constr to val.

SemidefiniteOptInterface.setconstraintconstant! — Function.

setconstraintconstant!(optimizer::AbstractSDOptimizer, val, constr::Integer)

Get the entry constr of b.

SemidefiniteOptInterface.setobjectivecoefficient! — Function.

setobjectivecoefficient!(optimizer::AbstractSDOptimizer, val, blk::Integer, i::Integer, j::Integer)

Set the entry i, j of the block blk of the objective matrix to val.