Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

SimpleGraphAlgorithms.ad — Method.

ad(G) is the average degree of a vertex in G.

SimpleGraphAlgorithms.chrome_poly — Method.

chrome_poly(G) computes the chromatic polynomial of the graph G. If G is a directe graph, this returns the chromatic polynomial of G's underlying simple graph.

This function builds a datastructure to prevent computing the chromatic polynomial of the same graph twice. To do this, it uses frequent isomorphism checks.

See size_cpm and reset_cpm.

SimpleGraphAlgorithms.color — Method.

color(G,k): Return a k-coloring of G (or error if none exists).

SimpleGraphAlgorithms.degdeg — Method.

degdeg(G) creates an n-by-n matrix where the nonzero entries in a row are the degrees of the neighbors of that vertex. The rows are lexicographically sorted.

SimpleGraphAlgorithms.edge_chromatic_number — Method.

edge_chromatic_number(G) returns the edge chromatic number of the graph G.

SimpleGraphAlgorithms.edge_color — Method.

edge_color(G,k) returns a proper k-edge coloring of G or throws an error if one does not exist.

SimpleGraphAlgorithms.fast_iso_test — Method.

fast_iso_test(G,H) applies various heuristics to see if the graphs might be isomorphic. A false return certifies the graphs are not isomorphic; a true result indicates the graphs might be (indeed, likely are) isomorphic.

SimpleGraphAlgorithms.info_map — Method.

info_map(G): We create a dictionary mapping the vertices of G to 128-bit integer values in such a way that twin vertices will have the same value but, we hope, nontwin vertices will have different values. (By twin we mean a pair of vertices such that there is an automorphism of the graph mapping one to the other.)

SimpleGraphAlgorithms.iso — Method.

iso(G,H) finds an isomorphism from G to H (or throws an error if the graphs are not isomorphic). Returns a dictionary mapping the vertices of G to H.

SimpleGraphAlgorithms.iso2 — Method.

iso2(G,H) is a variant of iso(G,H) that first applies various heuristics that should speed up processing if the graphs are "far" from vertex transitive.

SimpleGraphAlgorithms.iso_check — Method.

iso_check(G,H,d) checks if d is an isomorphism from G to H.

SimpleGraphAlgorithms.iso_matrix — Method.

iso_matrix(G,H) returns a permutation matrix P such that A*P==P*B where A is the adjacency matrix of G and B is the adjacency matrix of H. If the graphs are not isomorphic, an error is raised.

SimpleGraphAlgorithms.mad — Method.

mad(G) computes the maximum average degree of G. See also mad_core.

SimpleGraphAlgorithms.mad_core — Method.

mad_core(G) returns a subgraph H of G for which ad(H)==mad(G).

SimpleGraphAlgorithms.max_clique — Method.

max_clique(G) returns a maximum size clique of a SimpleGraph.

SimpleGraphAlgorithms.max_indep_set — Method.

max_indep_set(G) returns a maximum size independent set of a SimpleGraph.

SimpleGraphAlgorithms.max_matching — Method.

max_matching(G) returns a maximum matching of a SimpleGraph.

SimpleGraphAlgorithms.min_dom_set — Method.

min_dom_set(G) returns a smallest dominating set of a SimpleGraph. That is, a smallest set S with the property that every vertex of G either is in S or is adjacent to a vertex of S.

SimpleGraphAlgorithms.min_edge_cover — Method.

min_edge_cover(G) finds a smallest subset F of the edges such that every vertex of G is the end point of at least one edge in F. An error is raised if G has a vertex of degree 0.

SimpleGraphAlgorithms.min_vertex_cover — Method.

min_vertex_cover(G) returns a smallest vertex cover S of G. This is a smallest possible set of vertices such that every edge of G has one or both end points in S.

min_vertex_cover(G,k) returns the smallest set of vertices S such that at least k edges are indicent with a vertex in S.

SimpleGraphAlgorithms.reset_cpm — Method.

reset_cpm() clears the datastructure of previously computed chromatic polynomials.

SimpleGraphAlgorithms.set_solver — Method.

set_solver(function) selects which optimization solver to use. For example, set_solver(CbcSolver). To use Gurobi, so this:

using Gurobi
set_solver(GurobiSolver)

Choice remains in effect until there's a subsequent call to set_solver.

SimpleGraphAlgorithms.size_cpm — Method.

size_cpm() reports the number of graphs held in the datastructure of previously computed chromatic polynomials.

SimpleGraphAlgorithms.uhash — Method.

uhash(G) creates an UInt64 hash of the graph such that isomorphic graphs will produce the same value. We hope that nonisomorphic graphs will create different values, but, alas, that need not be the case.

SimpleGraphAlgorithms.fast_iso_test_basic — Method.

fast_iso_test_basic(G,H) is a very quick test to rule out graphs being isomorphic by considering the number of vertices, number of edges, and degree sequences. Returns false if the graphs fail this very basic test of isomorphism. A true result does not imply the graphs are isomorphic.

SimpleGraphAlgorithms.mad_model — Method.

mad_model(G) returns the solved linear program whose optimum value is mad(G).

SimpleGraphAlgorithms.recall — Method.

recall(CPM,G) look up in CPM to see if we have already computed the chromatic polynomial of this graph (or something isomorphic to it).

Short form: P = CPM[G].

SimpleGraphAlgorithms.remember! — Method.

remember!(CPM,G,P) save the polynomial P associated with G in this data structure (if it wasn't in there already).

Short form: CPM[G] = P.