Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

SingularSpectrumAnalysis.elementary — Method.

Ui = elementary(USV,I)

Computes the sum UᵢSᵢVᵢ' for all i in I If I is 1:L, the returned matrix is identical to USV'

SingularSpectrumAnalysis.hankel — Method.

X = hankel(x,window_size)

Form the trajectory matrix X of size KxL, K = N-L+1 x can be a vector or a matrix for multivariate SSA

SingularSpectrumAnalysis.hankelize — Method.

means = hankelize(X)

Computes a timeseries from an approximate Hankel matrix by diagoal averaging (Hankelization). Note that the returned value is a vector and not a hankel matrix. A hankel matrix is easily obtainable by hankel(hankelize(X), size(X,2))

SingularSpectrumAnalysis.hsvd — Method.

USV = hsvd(y,L)

Form a trajectory hankel matrix from data y and compute svd on this Hankelmatrix

SingularSpectrumAnalysis.pairplot — Method.

pairplot(USV, groupings) Usage:

USV = hsvd(data,L)
seasonal_groupings = [1:2, 4:5]
pairplot(USV,seasonal_groupings)

SingularSpectrumAnalysis.reconstruct — Method.

yrt,yrs = reconstruct(USV::SVD, trends, seasonal::AbstractArray)
yr      = reconstruct(USV::SVD, groupings::AbstractArray)

Compute a reconstruction of the time-series based on an SVD object obtained from hsvd and user selected groupings. See also ?SingularSpectrumAnalysis