Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

StrAPI.CharSetASCIICompat — Type.

Characters 0-0x7f same as ASCII

StrAPI.CharSetBMP — Type.

Characters 0-0xd7ff, 0xe000-0xffff follow Unicode BMP

StrAPI.CharSetBinary — Type.

8-bit Binary string, not text

StrAPI.CharSetISOCompat — Type.

Characters 0-0x9f follows ISO 8859

StrAPI.CharSetOther — Type.

Codepoints are not in Unicode compatible order

StrAPI.CharSetStyle — Type.

CharSetStyle(A)
CharSetStyle(typeof(A))

CharSetStyle specifies the information about the character set used by the string or characters. When you define a new AbstractString or AbstractChar type, you can choose to implement it with

StrAPI.CharSetStyle(::Type{<:MyString}) = StrAPI.CharSetISOCompat()

The default is CharSetUnicode()

StrAPI.CharSetUnicode — Type.

Full Unicode character set, no additions

StrAPI.CharSetUnicodePlus — Type.

Unicode character set, plus encodings of invalid characters

StrAPI.CharSetUnknown — Type.

Raw bytes, words, or character string, unknown encoding/character set

StrAPI.CompareStyle — Type.

CompareStyle(Union{A, typeof(A)}, Union{B, typeof(B)})

CompareStyle specifies how to compare two strings with character set encodings A and B

StrAPI.CompareStyle(::Type{<:MyString}, ::Type{String}) = StrAPI.ByteCompare()

The default is CodePointCompare

StrAPI.EncodingStyle — Type.

EncodingStyle(A)
EncodingStyle(typeof(A))

EncodingStyle specifies the whether a character set encoding uses one or multiple codeunits to encode a single codepoint. When you define a new AbstractString type, you can choose to implement it with either single or multi-codeunit indexing.

StrAPI.EncodingStyle(::Type{MyCharSetEncoding}) = StrAPI.MultiCodeUnitEncoding()

The default is SingleCodeUnitEncoding()

StrAPI.ValidatedStyle — Type.

ValidatedStyle(A)
ValidatedStyle(typeof(A))

ValidatedStyle specifies the whether a string or character type is always valid or not When you define a new AbstractString or AbstractChar type, you can choose to implement it as always validated, or validation state unknown.

StrAPI.ValidatedStyle(::Type{<:MyCharType}) = StrAPI.AlwaysValid()

The default is UnknownValidity

StrAPI.basetype — Function.

Get the base type (of CodeUnitTypes) of a character or aligned/swapped type

StrAPI.charset — Function.

Get the character set used by a string or character type

StrAPI.codepoints — Method.

codepoints(str)

An iterator that generates the code points of a string

Examples

julia> a = str("abc🖖")

julia> collect(a)

julia> collect(codepoints(a))

StrAPI.cse — Function.

Get the character set / encoding used by a string type

StrAPI.encoding — Function.

Get the character set used by a string type

StrAPI.is_multi — Function.

Determine if a string has multiple codeunit encoding