Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

StrTables.AbstractEntityTable — Type.

Abstract type for Entity tables: Supports lookupname, matchchar, matches, longestmatches, completions

StrTables.PackedTable — Type.

Compact table Designed to save memory compared to a Vector{Vector{S}} Allows for fast lookup of ranges when input was sorted Can be saved/loaded to/from a file quickly

StrTables.PackedTable — Method.

Make a single table of a vector of elements of type T

StrTables.StrTable — Type.

Compact string table Designed to save memory compared to a Vector{String} Allows for fast lookup of ranges when input was sorted Can be saved/loaded to/from a file quickly

StrTables.completions — Function.

Given a string, return all of the entity names that start with that string, if any

StrTables.longestmatches — Function.

Given a string, return all of the longest matches to the beginning of the string as a vector

StrTables.lookupname — Function.

Given an entity name, return the string it represents, or an empty string if not found

StrTables.matchchar — Function.

Given a character, return all exact matches to the character as a vector

StrTables.matches — Function.

Given a string, return all exact matches to the string as a vector

StrTables.load — Method.

Return a Vector{Any} with data tables loaded from file

StrTables.load — Method.

Return a Vector{Any} with data tables loaded from IO

StrTables.matchfirst — Method.

Return a vector of values that whose beginning matches the string

StrTables.matchfirstrng — Method.

Return the range of indices of values that whose beginning matches the string

StrTables.read_value — Method.

Read a single StrTable, String, UInt, or Vector of UInt value from the io device

StrTables.save — Method.

Save a collection of values (StrTable, String, Float, UInt, Int*) into a file

StrTables.save — Method.

Save a collection of values (StrTable, String, Float, UInt, Int*) into an IO object