Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

SymmetricTensors.SymmetricTensor — Type.

Type constructor

frame - stores ArrayNArrays{T,N} bls - Int, size of ordinary block bln - Int, number of blocks datasize - Int, size of data stored (in each direction the same) sqr - Bool, is the last block size a same as ordinary's block size

SymmetricTensors.SymmetricTensor — Method.

SymmetricTensor(data::Array{N}, bls::Int)

Returns: data in SymmetricTensor form.

julia> a = reshape(collect(1.:16.), 4, 4);

julia> SymmetricTensor(a*a')
SymmetricTensors.SymmetricTensor{Float64,2}(Union{Array{Float64,2}, Void}[[276.0 304.0; 304.0 336.0][332.0 360.0; 368.0 400.0]; nothing [404.0 440.0; 440.0 480.0]], 2, 2, 4, true)

Base.rand — Method.

rand(SymmetricTensor{T, N}, n::Int, b::Int = 2)

Returns N-dimensional random SymmetricTensor with elements of type T drawn from uniform distribution on [0,1), n denotes data size and b denotes block size.

Base.setindex! — Method.

setindex!(st::SymmetricTensor{T,N}, x::T,  mulind::Int...)

Change a SymmetricTensors value at the given multi-index

SymmetricTensors.diag — Method.

diag(st::SymmetricTensor{N})

Return vector of floats, the super-diag of st

SymmetricTensors.unfold — Method.

unfold(ar::Array{N}, mode::Int)

Returns an matrix being an unfold of N dims array in given mode.

julia> A = reshape(collect(1.:8.), 2, 2, 2);

julia> unfold(A, 1)
2×4 Array{Float64,2}:
 1.0  3.0  5.0  7.0
 2.0  4.0  6.0  8.0

 julia> unfold(A, 2)
 2×4 Array{Float64,2}:
  1.0  2.0  5.0  6.0
  3.0  4.0  7.0  8.0

  julia> unfold(A, 3)
  2×4 Array{Float64,2}:
   1.0  2.0  3.0  4.0
   5.0  6.0  7.0  8.0

Core.Array — Method.

Array(st::SymmetricTensor{N})

Return N dims array converted from SymmetricTensor type

Base.getindex — Method.

getindex(st::SymmetricTensor, mulind::Tuple)

Returns a Symmetric Tensor element for a given multi-index

SymmetricTensors.fixpointperms — Method.

fixpointperms(j::NTuple{N, Int}) where N

Returns Vactor{Vector}, a fix point permutation of given multiindex

TODO: this is a naive implementation

SymmetricTensors.frtest — Method.

frtest(data::ArrayNArrays{T,N})

Returns assertion error if: all sizes of nullable array not equal, at least some undergiagonal block not null, some blocks (not last) not squared, some diagonal blocks not symmetric.

SymmetricTensors.getblock — Method.

getblock(st::SymmetricTensor, i::Tuple)

Returns a block from Symmetric Tensor, works for all multi-indices also not sorted

SymmetricTensors.getblockunsafe — Method.

getblockunsafe(st::SymmetricTensor, i::Tuple)

Returns a block from Symmetric Tensor, unsafe works only if multi-index is sorted

SymmetricTensors.ind2range — Method.

ind2range(i::Int, bls::Int, dats::Int)

Returns a range given index i, size of a block and size of data

julia> ind2range(2,3,5)
4:5

SymmetricTensors.issymetric — Method.

issymetric(ar::Array{N}, atol::Float64)

Returns: Assertion Error if not symmetric given a tolerance.

julia> A = reshape(collect(1.:8.), 2, 2, 2);

julia> julia> issymetric(A)
ERROR: AssertionError: not symmetric

SymmetricTensors.pyramidindices — Method.

pyramidindices(dims::Int, tensize::Int)

julia> pyramidindices(2,3)
6-element Array{Tuple{Int64,Int64},1}:
 (1,1)
 (1,2)
 (1,3)
 (2,2)
 (2,3)
 (3,3)

SymmetricTensors.randblock — Method.

randblock(::Type{T}, dims::NTuple{N, Int}, j::NTuple{N, Int})

Returns a block of size dims and position j by a uniformly distributed random number of type T

SymmetricTensors.randsymarray — Function.

randsymarray(T, dim::Int, N::Int = 4)

Returns N-dimmensional random super-symmetric array with Float64 elements drawn from uniform distribution on [0,1), dim denotes data size.

SymmetricTensors.randsymarray — Method.

randsymarray(T, dim ::Int, N::Int = 4)

Returns N-dimmensional random super-symmetric array with elements of type T drawn from uniform distribution on [0,1), dim denotes data size.

SymmetricTensors.setindexunsafe! — Method.

setindexunsafe!(st::SymmetricTensor{T,N}, x::T,  mulind::Int...)

Unsafe change a SymmetricTensors value at the given multi-index

SymmetricTensors.sizetest — Method.

sizetest(dats::Int, bls::Int)

Returns: DimensionMismatch if blocks size is grater than data size.

julia> SymmetricTensors.sizetest(2,3)
ERROR: DimensionMismatch("bad block size 3 > 2")