Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

TimeToLive.TTL — Type.

TTL(ttl::Period; refresh_on_access::Bool=true) -> TTL{Any, Any}
TTL{K, V}(ttl::Period; refresh_on_access::Bool=true) -> TTL{K, V}

A TTL cache. If refresh_on_access is set, expiries are reset whenever they are accessed.

Base.Filesystem.touch — Method.

touch(t::TTL, k)

Reset the expiry time for the value at t[k].