Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

TropicalSemiring.Max — Type.

Max

Indicating that we are in the tropical semiring $(ℝ ∪ {-∞}, ⊕, ⊙)$ with the max convention.

TropicalSemiring.Min — Type.

Min

Indicating that we are in the tropical semiring $(ℝ ∪ {∞}, ⊕, ⊙)$ with the min convention.

TropicalSemiring.Trop — Type.

Trop{MM<:MinMax, T<:Real}

A Trop number is an element of a the tropical semi-ring. These come in two flavours Max and Min. In the Max case this is the semi-ring $(ℝ ∪ {-∞}, ⊕, ⊙)$ where $⊕$ is the usual multiplication and $⊙$ is the usual maximum. In the Min case this is the semi-ring $(ℝ ∪ {∞}, ⊕, ⊙)$ where $⊕$ is the usual multiplication and $⊙$ is the usual minimum.

TropicalSemiring.inf — Method.

inf(M::Type{<:Union{Min, Max}} = Max)

Constructs $-∞$ in the case of Max and $∞$ in the case of Min.