Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

ValueOrientedRiskManagementInsurance.Alloc — Type.

Allocation of investments within an InvestGroup object ig

ValueOrientedRiskManagementInsurance.Asset — Type.

Abstract type for assets

ValueOrientedRiskManagementInsurance.BuInsurance — Type.

Business unit: Line of business

ValueOrientedRiskManagementInsurance.BuInvestments — Type.

Business unit: Investments

ValueOrientedRiskManagementInsurance.BusinessUnit — Type.

A business unit is a line of business or investments

ValueOrientedRiskManagementInsurance.CapMkt — Type.

A collection of Process objects which represents a capital market. Currently we only have two deterministic processes, stock::Stock and rfr::RiskFreeRate.

ValueOrientedRiskManagementInsurance.Debt — Type.

Debt issued by the company

ValueOrientedRiskManagementInsurance.DetermProcess — Type.

Deterministic process, currently the only type of process implemented

ValueOrientedRiskManagementInsurance.Dynamic — Type.

Dynamic projection parameters for a Projection object

ValueOrientedRiskManagementInsurance.GaussCopula — Type.

Gaussian copula

ValueOrientedRiskManagementInsurance.GaussCopula — Method.

Creation of a Gaussian copula with marginal distributions

ValueOrientedRiskManagementInsurance.IGCash — Type.

Invest Group for Invest object InvestCash

ValueOrientedRiskManagementInsurance.IGCost — Type.

Investment costs for an InvestGroup object ig

ValueOrientedRiskManagementInsurance.IGStock — Type.

Invest Group for Invest object InvestStock

ValueOrientedRiskManagementInsurance.InvPort — Type.

Investment portfolio

ValueOrientedRiskManagementInsurance.Invest — Type.

Represents an individual investments and its development in time.

ValueOrientedRiskManagementInsurance.InvestCash — Type.

Invest object associated with a RiskFreeRate process

ValueOrientedRiskManagementInsurance.InvestGroup — Type.

A group of Invest objects of the same sub-type

ValueOrientedRiskManagementInsurance.InvestStock — Type.

Invest object associated with a Stock process

ValueOrientedRiskManagementInsurance.LiabIns — Type.

Liability portfolio

ValueOrientedRiskManagementInsurance.LiabOther — Type.

Portfolio of other liabilities

ValueOrientedRiskManagementInsurance.Liabilities — Type.

Liability portfolio consisting of a single insurance contract

ValueOrientedRiskManagementInsurance.ModelPoint — Type.

Model point for the liability portfolio representing a number of insured with the same contract parameters and the same biometric parameters

Time model:

  project time τ:               0
  real time t:     t_start      t_0
  product time s:  0            s_0
                   |------------|-------------------|-----------
                                |-------------------|
                                         dur
                   |--------------------------------|
                              product.dur

ValueOrientedRiskManagementInsurance.NLLob — Type.

Line of Business (Non-Life)

ValueOrientedRiskManagementInsurance.NLLob — Method.

Constructs line of business & calculates statistical parameters

ValueOrientedRiskManagementInsurance.PLInsurance — Type.

Profit loss account for the insurance result from a line of business (lob)

ValueOrientedRiskManagementInsurance.PLInvestments — Type.

Profit loss account for the investment result result

ValueOrientedRiskManagementInsurance.PLTotal — Type.

Profit loss account for the aggregated insurance company

ValueOrientedRiskManagementInsurance.Product — Type.

Life insurance product / tariff

ValueOrientedRiskManagementInsurance.ProfitLoss — Type.

Profit and loss account of an insurance company component

ValueOrientedRiskManagementInsurance.ProjParam — Type.

Projection parameters for S2-calculations without S2 calibration parameters

ValueOrientedRiskManagementInsurance.Projection — Type.

Projection of a life insurer

ValueOrientedRiskManagementInsurance.RiskFactor — Type.

Risk factor

ValueOrientedRiskManagementInsurance.RiskFreeRate — Type.

A deterministic process representing a short rate

ValueOrientedRiskManagementInsurance.S2 — Type.

Solvency 2 (total)

ValueOrientedRiskManagementInsurance.S2Mkt — Type.

Solvency 2 market risk

ValueOrientedRiskManagementInsurance.S2MktEq — Type.

Solvency 2 market risk: equity risk

ValueOrientedRiskManagementInsurance.S2MktInt — Type.

Solvency 2 market risk: interest rate risk

ValueOrientedRiskManagementInsurance.SSTCapMkt — Type.

Capital market

ValueOrientedRiskManagementInsurance.Stock — Type.

A deterministic process representing a stock index. Additional objects

ValueOrientedRiskManagementInsurance.StockIndex — Type.

Investment in stock

ValueOrientedRiskManagementInsurance.Stress — Type.

Stress scenario

ValueOrientedRiskManagementInsurance.Total — Type.

Total company result

ValueOrientedRiskManagementInsurance.ZeroBond — Type.

Zero bond (an instance of an asset)

Base.rand — Method.

rand(gc::GaussCopula, n::Int)

n random samples from a Gaussian copula gc

ValueOrientedRiskManagementInsurance.delta — Method.

delta(assets::Vector{Asset}, liabs::Liabilities, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the δ-vector

ValueOrientedRiskManagementInsurance.evaluate! — Method.

evaluate!(pl::ProfitLoss, α::Real)

Calculate expected profit, economic capital (expected shortfall at safety level α), and RORAC for pl

ValueOrientedRiskManagementInsurance.forw2spot — Method.

forw2spot(f::Vector{Float64})

calculates the spot rate s rate from the forward rate f:

(1+f[1])(1+f[2])...(1+f[n]) = (1+s[n])^n

ValueOrientedRiskManagementInsurance.gammamatrix — Method.

gammamatrix(assets::Vector{Asset}, liabs::Liabilities, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the Γ-matrix

ValueOrientedRiskManagementInsurance.initialize — Method.

initialize(insurance_input::DataFrame, invest_input::DataFrame, tau_kendall::Matrix{Real}, n_scen::Int)

Set up an insurance company

ValueOrientedRiskManagementInsurance.premrestotalsd — Method.

premrestotalsd(nllobs::Vector{NLLob}, corr_lob::Matrix{Float64})

Calculates the standard deviation of the total premium reserve

ValueOrientedRiskManagementInsurance.profit! — Method.

profit!(pl::PLInsurance, r_distr::Vector{Float64}, s::Real)

Updates pl.profit, where r_distr is the loss distribution and s is the risk free interest rate

ValueOrientedRiskManagementInsurance.profit! — Method.

profit!(pl::PLInvestments, r_distr::Vector{Float64}, s::Real)

Updates pl.profit, where r_distr is the investment result and s is the risk free interest rate. Only the investment return above s counts as profit.

ValueOrientedRiskManagementInsurance.profit! — Method.

profit!(pl::PLTotal, pl_bu::Array{ProfitLoss}, cap_init::Real, costs_fixed::Real, s::Real)

Updates pl.profit, where cap_init is the initial capital, pl_bu are the profit loss accounts of the business units, costs_fixed are the fixed costs, and s is the risk free interest rate.

ValueOrientedRiskManagementInsurance.project — Method.

project(ins_input::DataFrame, inv_input::DataFrame, tau_kendall::Matrix{Real}, n_scen::Int, α::Real, s::Real costs_fixed::Real)

Set up an insurance company and project its results

ValueOrientedRiskManagementInsurance.rtk — Method.

rtk(t::Int, assets::Vector{Asset}, liabs::Liabilities, x::Vector{Float64}, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the risk bearing capital (after sensitivity x)

ValueOrientedRiskManagementInsurance.spot2forw — Method.

spot2forw(s::Vector{Float64})

calculates the forward rate f rate from the spot rate s:

(1 + s[n-1])^(n-1) * (1+f[n]) = (1+s[n])^n

ValueOrientedRiskManagementInsurance.srtk — Method.

srtk(shift::Int, assets::Vector{Asset}, liabs::Liabilities, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates (linear) sensitivities for rtk

ValueOrientedRiskManagementInsurance.srtk — Method.

srtk(shift_1::Int, shift_2::Int, assets::Vector{Asset}, liabs::Liabilities, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates quadratic sensitivities for rtk

ValueOrientedRiskManagementInsurance.value — Method.

value(t::Int, assts::Vector{Asset}, x::Vector{Float64}, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the value of the asset portfolio (after sensitivity x)

ValueOrientedRiskManagementInsurance.value — Method.

value(t::Int, liabs::Liabilities, x::Vector{Float64}, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the value of the liability portfolio (after sensitivity x)

ValueOrientedRiskManagementInsurance.value — Method.

value(t::Int, si::StockIndex,x::Vector{Float64}, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the value of a stock (after sensitivity x)

ValueOrientedRiskManagementInsurance.value — Method.

value(t::Int, zb::ZeroBond, x::Vector{Float64}, rf::RiskFactor, cap_mkt::SSTCapMkt)

Calculates the value of a zero bond (after sensitivity x)

ValueOrientedRiskManagementInsurance.Δrtk — Method.

Δrtk(Δx::Vector{Float64}, δ::Vector{Float64}, Γ::Matrix{Float64})

Calculates the shocked rtk based on shocks Δx

ValueOrientedRiskManagementInsurance.Process — Type.

A stochastic or deterministc process representing a part of the Capital market. It always has an object x representing the development in time

ValueOrientedRiskManagementInsurance.S2Def — Type.

Solvency 2 default risk

ValueOrientedRiskManagementInsurance.S2Def1 — Type.

Solvency 2 default risk of type 1

ValueOrientedRiskManagementInsurance.S2Def2 — Type.

S2 default (type 2) risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2Health — Type.

S2 health risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2Life — Type.

Solvency 2 life risk

ValueOrientedRiskManagementInsurance.S2LifeBio — Type.

Solvency 2 life risk: biometric risks: mortality, longevity, surrender

ValueOrientedRiskManagementInsurance.S2LifeCost — Type.

Solvency 2 life risk: expense risk

ValueOrientedRiskManagementInsurance.S2LifeMorb — Type.

S2 life morbidity risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2LifeRevision — Type.

S2 life revision risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2MktConc — Type.

S2 market concentration risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2MktFx — Type.

S2 market currency risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2MktProp — Type.

S2 market property risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2MktSpread — Type.

S2 market spread risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2Module — Type.

Abstract type for an S2 module

ValueOrientedRiskManagementInsurance.S2NonLife — Type.

S2 non-life risk module, not yet implemented

ValueOrientedRiskManagementInsurance.S2NotImplemented — Type.

Dummy type for those S2 modules that have not been implemented

ValueOrientedRiskManagementInsurance.S2Op — Type.

Solvency 2 operational risk

ValueOrientedRiskManagementInsurance.alloc! — Method.

alloc!(τ, cap_mkt::CapMkt, invs::InvPort)

Dynamically re-allocate investments invs::InvPort at the beginning of year τ

Changed: invs

ValueOrientedRiskManagementInsurance.bioshock! — Method.

bioshock!(l_ins::LiabIns, bio::S2LifeBio, shock_symb::Symbol)

Shock for biometric risk

Changed: l_ins::LiabIns (l_ins.mps)

ValueOrientedRiskManagementInsurance.bioshock! — Method.

bioshock!(mp::ModelPoint, bio::S2LifeBio, symb::Symbol)

Helper function: shock for biometric risk

Changed: mp::ModelPoint

ValueOrientedRiskManagementInsurance.biquotient — Method.

biquotient(τ, yield_eoy, cap_mkt, invs, mp, dyn)

Indicator for bonus rate expectation at the end of year τ

ValueOrientedRiskManagementInsurance.bof — Method.

bof(bal::DataFrame)

Helper function for the calculation of basic own funds

ValueOrientedRiskManagementInsurance.bof — Method.

bof(md::S2Module, scen::Symbol)

Basic own funds for module md and scenario scen

ValueOrientedRiskManagementInsurance.bonus! — Method.

bonus!(τ, invs::InvPort, liabs::LiabIns, dyn::Dynamic, proj, surp_pre_profit_tax_bonus)

Bonus at the end of year τ

Changed: proj::Projection

ValueOrientedRiskManagementInsurance.bonusrate — Method.

bonusrate(yield_eoy, rfr_price, bonus_factor)

Helper function for dynamic bonus rate declaration

ValueOrientedRiskManagementInsurance.bonusrate — Method.

bonusrate(τ, yield_eoy, mp::ModelPoint, dyn)

Dynamic bonus rate declaration for a model point at the end of year τ

ValueOrientedRiskManagementInsurance.catshock! — Method.

catshock!(mp::ModelPoint, bio::S2LifeBio, symb::Symbol)

Shock for mortality catastrophe risk

Changed: mp::ModelPoint

ValueOrientedRiskManagementInsurance.costshock! — Method.

costshock!(invs::InvPort, l_ins::LiabIns, cost::S2LifeCost)

Shock for expense risk

Changed: invs::InvPort, l_ins::LiabIns

ValueOrientedRiskManagementInsurance.dynstate — Method.

dynstate(τ, cap_mkt::CapMkt)

indicator for the state of the economy at the end of year τ

ValueOrientedRiskManagementInsurance.dynstateavg — Method.

dynstateavg(τ, cap_mkt::CapMkt)

Two year average of the indicator for the state of the economy at the end of year τ.

ValueOrientedRiskManagementInsurance.fdb — Method.

fdb(md::S2Module, scen::Symbol)

Future discretionary benefits for module md and scenario scen

ValueOrientedRiskManagementInsurance.freesurp — Method.

freesurp(dyn, invest_pre, liab)

Helper function for the free surplus calculation

ValueOrientedRiskManagementInsurance.freesurp — Method.

freesurp(τ, proj::Projection, dyn)

Free surplus for the dynamic dividend declaration

ValueOrientedRiskManagementInsurance.getloan — Method.

getloan(τ::Int, debt::Debt)

Get the nominal of debt at time τ, if it has been taken out at this point in time, otherwise return 0.0

ValueOrientedRiskManagementInsurance.getloan — Method.

getloan(τ::Int, l_other::LiabOther)

Get the total nominal of all debts within the portfolio of other liabilities l_other, which are taken out at the beginning of year τ

ValueOrientedRiskManagementInsurance.getyield — Method.

getyield(τ, rfr::RiskFreeRate)

calculates the yield of rfr during year τ

ValueOrientedRiskManagementInsurance.getyield — Method.

getyield(τ, stock::Stock)

calculates the yield of stock during year τ

ValueOrientedRiskManagementInsurance.goingconcern! — Method.

goingconcern!(l_other::LiabOther, Δgc::Vector{Float64})

Transforms a portfolio of other liabilities l_other::LiabOther according to the going concern assumption. We do not input the factors gc directly but their year on year differences Δgc.

Changed: l_other

ValueOrientedRiskManagementInsurance.goingconcern — Method.

goingconcern(debts::Vector{Debt}, Δgc::Vector{Float64})

Calculates a vector of debts from an existing vector of debts according to the going concern assumption. The total initial debt is the same, but the new debt vectors mature earlier so that the total nominal decreases according to the going concern factors. We do not input the factors gc directly but their year on year differences Δgc.

ValueOrientedRiskManagementInsurance.investpredivid — Method.

investpredivid(τ, invs::InvPort, proj::Projection)

Market value of assets before payment of dividends

ValueOrientedRiskManagementInsurance.mkt_val0_adj! — Method.

mkt_val0_adj!(proj::Projection, invs::InvPort, mkt_eq, eq_type::Symbol)

Adjust initial market value for S2 balance sheet in order to reflect that the initial fall of the market value of equity investments is reflected in the shockedsolvency balance sheet Must be called after the projection

Changed: proj::Projection

ValueOrientedRiskManagementInsurance.mkteqshock! — Method.

mkteqshock!(invs::InvPort, mkt_eq, eq_type::Symbol)

Shock for equity market risk

Changed: invs::InvPort

ValueOrientedRiskManagementInsurance.mktintshock! — Method.

mktintshock!(cap_mkt::CapMkt, s2_mkt_int, int_type::Symbol)

Shock for interest rate market risk

Changed: cap_mkt::CapMkt

ValueOrientedRiskManagementInsurance.paycoupon — Method.

paycoupon(τ::Int, debt::Debt)

Coupon payment for debt at the end of year τ

ValueOrientedRiskManagementInsurance.paycoupon — Method.

paycoupon(τ::Int, l_other::LiabOther)

Coupon payment for all debts within the portfolio of other liabilities l_other at the end of year τ

ValueOrientedRiskManagementInsurance.payprincipal — Method.

payprincipal(τ::Int, debt::Debt)

Payment of the principal of debt at the end of year τ, if the debt matures at this point in time

ValueOrientedRiskManagementInsurance.payprincipal — Method.

payprincipal(τ::Int, l_other::LiabOther)

Payment of the total principal of all debts within the portfolio of other liabilities l_other, which mature at the end of year τ

ValueOrientedRiskManagementInsurance.premium — Method.

premium(ins_sum, rfr, prob, β, λ)

Calculate the premium of a product

ValueOrientedRiskManagementInsurance.project! — Method.

project!(τ, cap_mkt::CapMkt, invs::InvPort, dyn::Dynamic, proj::Projection)

Project one year, investment results from the year τ

Changed: proj::Projection, invs::InvPort

ValueOrientedRiskManagementInsurance.project! — Method.

project!(τ, cap_mkt::CapMkt, invs::InvPort, liabs::LiabIns, liab_other::LiabOther, dyn::Dynamic, proj::Projection)

Project one year

Changed: proj::Projection, invs::InvPort, dyn::Dynamic

ValueOrientedRiskManagementInsurance.project! — Method.

project!(τ::Int, mv_boy::Float64, invs::InvPort)

Project invs::InvPort one year for a given initial market value.

Changed: invs

ValueOrientedRiskManagementInsurance.project! — Method.

project!(τ::Int, mv_bop_total::Float64, ig::InvestGroup)

project ig::InvestGroup one year for a given initial market value.

Changed: ig

ValueOrientedRiskManagementInsurance.project! — Method.

project!(τ::Int, mv_boy::Float64, invest::Invest)

project invest::Invest one year for a given initial market value.

Changed: invest

ValueOrientedRiskManagementInsurance.projectboy! — Method.

projectboy!(τ, proj::Projection, liabs::LiabIns)

Project one year, update values at the beginning of the year τ

Changed: proj::Projection, liabs::LiabIns (liabs.mp)

ValueOrientedRiskManagementInsurance.projecteoy! — Method.

projecteoy!(τ, cap_mkt::CapMkt, invs::InvPort, liabs::LiabIns, dyn::Dynamic, proj::Projection)

Project one year, update values at the end of the year τ

Changed: proj::Projection, liabs::LiabIns (liabs.mp)

ValueOrientedRiskManagementInsurance.pv — Method.

pv(τ::Int, cap_mkt::CapMkt, debt::Debt)

Present value of debt at the end of year τ. Any servicing of this debt during year τ has occured beforehand

ValueOrientedRiskManagementInsurance.pv — Method.

pv(τ::Int, cap_mkt::CapMkt, l_other::LiabOther)

present value of other liabilities l_other at the end of year τ. Any servicing of debt within this portfolio during year τ has occured beforehand

ValueOrientedRiskManagementInsurance.pvprev — Method.

pvprev(rfr, cf, pv) = (cf + pv) / (1 + rfr)

Recursive step in generic present value calculation

ValueOrientedRiskManagementInsurance.pvvec — Method.

pvvec(rfr::Vector{Float64}, cf)`

Generic present value calculation, where rfr denotes the risk free rate for each year and cf. The length of the vector rfr may not be smaller than the length of the cashflow. Each payment occurs at the end of the corresponding year.

Output: A vector of the same length as cf contaning the present value at the end of each year. The last component is zero.

ValueOrientedRiskManagementInsurance.qxpxshock! — Method.

qxpxshock!(mp::ModelPoint, bio::S2LifeBio, symb::Symbol)

Helper function: shock for mortality risk

Changed: mp::ModelPoint

ValueOrientedRiskManagementInsurance.rfrshock — Method.

rfrshock(rfr::Vector{Float64}, s2_mkt_int, int_type)

Helper function: shock for the risk free interest rate

ValueOrientedRiskManagementInsurance.s2bal — Method.

s2bal(p::ProjParam, md::S2Module, shock!::Any, scen::Symbol)

Calculation of the shocked S2 balance sheet for module md

ValueOrientedRiskManagementInsurance.s2bal — Method.

s2bal(p::ProjParam)

Calculation of the unshocked S2 balance sheet

ValueOrientedRiskManagementInsurance.scr! — Method.

scr!(s2::S2, tax_credit_0::Float64)

Total SCR

Changed: s2

ValueOrientedRiskManagementInsurance.scr! — Method.

scr!(mkt_int::S2MktInt)

SCR for interest rate risk, sub-module mkt_int

Changed: mkt_int::S2MktInt

ValueOrientedRiskManagementInsurance.scr! — Method.

scr!(def::S2Def1)

SCR for interest rate risk, sub-module def::S2Def1

Changed: def::S2Def1

ValueOrientedRiskManagementInsurance.scr! — Method.

scr!(mdl::S2Module)

Scenario based SCR calculation for mdl::S2Module

Changed: mdl::S2Module

ValueOrientedRiskManagementInsurance.scr! — Method.

scr!(op::S2Op, bscr)

SCR for interest rate risk, sub-module op

Changed: op

ValueOrientedRiskManagementInsurance.scr — Method.

scr(md::S2Module, corr::Matrix{Float64})

Aggregation of SCRs of sub-modules

ValueOrientedRiskManagementInsurance.selectmort! — Method.

selectmort!(p::ProjParam, bio::S2LifeBio)

Identify those model points that are subject to mortality risk. This function does not properly take into account second order effects due to the effect of boni. However, for realistic portfolios second order effects are unlikely to change the set of identified model points.

Changed: bio::S2LifeBio (bio.mp_select)

ValueOrientedRiskManagementInsurance.sxshock! — Method.

sxshock!(mp::ModelPoint, bio::S2LifeBio, symb::Symbol)

Helper function: shock for surrender risk

Changed: mp::ModelPoint

ValueOrientedRiskManagementInsurance.tpg — Method.

tpg(τ, rfr, prob, β, λ)

Technical provisions of guaranteed benefits at the end of year τ

ValueOrientedRiskManagementInsurance.tpg — Method.

tpg(τ, rfr, mp)

Best estimate guaranteed technical provisions of a model point mp at the end of year τ

ValueOrientedRiskManagementInsurance.tpgfixed — Method.

tpgfixed(τ, rfr, fixed_cost_gc::Vector)

Technical provisions at the end of year τ for future fixed (going concern) costs

ValueOrientedRiskManagementInsurance.tpgrec — Method.

tpgrec(τ, tpg, rfr, prob, β, λ)

One year backward recursion formula for technical provisions of guaranteed benefits

ValueOrientedRiskManagementInsurance.update! — Method.

update!(τ, proj::Projection, dyn::Dynamic)

Update dynamic parameters

ValueOrientedRiskManagementInsurance.val0! — Method.

val0!(cap_mkt::CapMkt, invs::InvPort, liabs::LiabIns, l_other::LiabOther, proj::Projection)

Valuation at time t_0

Changed: proj::Projection

ValueOrientedRiskManagementInsurance.valbonus! — Method.

valbonus!(rfr::Vector{Float64}, proj::Projection)

Provisions for future bonus payments (at each time τ). Needs to be called after the projection is completed

Changed: proj::Projection

ValueOrientedRiskManagementInsurance.valcostprov! — Method.

valcostprov!(rfr::Vector{Float64}, invs::InvPort, proj::Projection)

Provisions for future costs (at each time τ). Absolute costs (including absolute investment costs from all investments) and relative investment costs for provisions are considered. It is assumed that provisions are backed by cash investments Needs to be called after the projection is completed

Changed: proj::Projection

ValueOrientedRiskManagementInsurance.δsx — Method.

δsx(τ, cap_mkt, invs, mp, dyn)

Dynamic lapse probability factor to adjust the initial estimate