Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

WebIO.Scope — Type.

Scope(id; kwargs...)

An object which can send and receive messages.

Fields:

id::String: A unique ID

imports: An array of js/html/css assets to load before rendering the contents of a scope.

WebIO.setup_provider — Method.

setup_provider(s::Union{Symbol, AbstractString})

Perform any side-effects necessary to set up the given provider. For example, in IJulia, this causes the frontend to load the webio javascript bundle.

WebIO.ConnectionPool — Type.

ConnectionPool(outbox::Channel, connections::Set{AbstractConnection}=Set())

Manages the distribution of messages from the outbox channel to a set of connections. The ConnectionPool asynchronously takes messages from its outbox and sends each message to every connection in the pool. Any closed connections are automatically removed from the pool.

WebIO.SyncCallback — Type.

A callable which updates the frontend

WebIO.encode_scripts — Method.

Any </script> tags in the js/html node representation can cause problems, because if they are printed inside a <script> tag, even if they are in quotes in a javascript string, the html parser will still read them as a closing script tag, and thus end the script content prematurely, causing untold woe.

WebIO.ensure_connection — Method.

Ensure that the pool has at least one connection, potentially blocking the current task until that is the case.

WebIO.re_register_renderables — Method.

Called after a provider is setup

WebIO.register_renderable — Method.

`WebIO.register_renderable(MyType::Type)`

Registers that a WebIO.render method is available for instances of MyType. Allows WebIO to hook into the display machinery of backends such as Atom and IJulia to display the WebIO rendered version of the type as appropriate.

Also defines a Base.show(io::IO, m::MIME"text/html", x::MyType) as Base.show(io, m, WebIO.render(x))

WebIO.render — Function.

render(x::MyType)

Generic function that defines how a Julia object is rendered. Should return a Node object.

WebIO.render — Method.

WebIO.render(obs::Observable)

Returns a WebIO Node whose contents are the richest version of the observable's value, and which updates to display the observable's current value

WebIO.set_nosync — Method.

Set observable without synchronizing with the counterpart on the browser