Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

Widgets.node — Method.

node(w::Widget)

Return primary node for widget w

Widgets.scope! — Method.

scope!(w::Widget, sc)

sets up a primary scope sc for widget w

Widgets.scope — Method.

scope(w::Widget)

Return primary scope for widget w if it exists, nothing otherwise.

Widgets.layout — Method.

layout(f, w::Widget)

Create a new Widget that is a copy of w and whose layout is the layout of w composed with the function f.

Examples

using InteractBase, CSSUtil, Widgets
w = button("OK")
Widgets.layout(w) do t
    hbox("Click here", t)
end

Widgets.layout — Method.

layout(w::Widget)

Return the function that will be used to determine the layout of widget w.