Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

WriteVTK.vtk_write_array — Method.

Write a Julia array to a VTK image data file (.vti).

Useful for general visualisation of arrays. The input can be a 2D or 3D array.

WriteVTK.data_to_xml — Function.

Add numerical data to VTK XML file.

Data is written under the xParent XML node.

Nc corresponds to the number of components of the data.

WriteVTK.data_to_xml_appended — Method.

Add appended raw binary data to VTK XML file.

Data is written to the vtk.buf buffer.

When compression is enabled:

the data array is written in compressed form (obviously);

the header, written before the actual numerical data, is an array of UInt32 values: [num_blocks, blocksize, last_blocksize, compressed_blocksizes] All the sizes are in bytes. The header itself is not compressed, only the data is. For more details, see: http://public.kitware.com/pipermail/paraview/2005-April/001391.html http://mathema.tician.de/what-they-dont-tell-you-about-vtk-xml-binary-formats (This is not really documented in the VTK specification...)

Otherwise, if compression is disabled, the header is just a single UInt32 value containing the size of the data array in bytes.

WriteVTK.data_to_xml_inline — Method.

Add inline, base64-encoded data to VTK XML file.

WriteVTK.datatype_str — Method.

Return the VTK string representation of a numerical data type.

WriteVTK.extent_attribute — Function.

Return the "extent" attribute required for structured (including rectilinear) grids.

WriteVTK.num_cells_structured — Method.

Number of cells in structured grids.

In 3D, all cells are hexahedrons (i.e. VTKHEXAHEDRON), and the number of cells is (Ni-1)(Nj-1)(Nk-1). In 2D, they are quadrilaterals (VTKQUAD), and in 1D they are line segments (VTK_LINE).

WriteVTK.num_components — Method.

Determine number of components of input data.

WriteVTK.save_with_appended_data — Method.

Write VTK XML file containing appended binary data to disk.

In this case, the XML file is written manually instead of using the save_file function of LightXML, which doesn't allow to write raw binary data.

WriteVTK.sizeof_data — Method.

Total size of data in bytes.

WriteVTK.vtk_point_or_cell_data — Method.

Add either point or cell data to VTK file.

Here Nc is the number of components of the data (Nc >= 1).

WriteVTK.write_array — Method.

Write array of numerical data to stream.