Home

Package doesn't contain Documenter docs.

Docs automatically generated by juliadocs.org

YaoBase.AbstractRegister — Type.

AbstractRegister{B, T}

Abstract type for quantum registers. B is the batch size, T is the data type.

YaoBase.QubitMismatchError — Type.

QubitMismatchError <: Exception

Qubit number mismatch error when applying a Block to a Register or concatenating Blocks.

Base.join — Method.

join(::AbstractRegister...) -> register

Merge several registers as one register via tensor product.

source

Base.repeat — Method.

repeat(r::AbstractRegister, n::Int) -> register

Repeat register r for n times on batch dimension.

Example

source

LinearAlgebra.ishermitian — Method.

ishermitian(op) -> Bool

check if this operator is hermitian.

YaoBase.addbit! — Method.

addbit!(register, n::Int) -> register
addbit!(n::Int) -> Function

addbit the register by n bits in state |0>. i.e. |psi> -> |000> ⊗ |psi>, addbit bits have higher indices. If only an integer is provided, then perform lazy evaluation.

source

YaoBase.basis — Method.

basis(register) -> UnitRange

Returns an UnitRange of the all the bits in the Hilbert space of given register.

source

YaoBase.datatype — Method.

datatype(register) -> Int

Returns the numerical data type used by register.

note!!!

`datatype` is not the same with `eltype`, since `AbstractRegister` family
is not exactly the same with `AbstractArray`, it is an iterator of several
registers.

source

YaoBase.density_matrix — Method.

density_matrix(register)

Returns the density matrix of current active qubits.

source

YaoBase.focus! — Method.

focus!(register, locs::Int...) -> register
focus!(locs::Int...) -> f(register) -> register

Focus the wires on specified location.

source

YaoBase.iscommute — Method.

iscommute(ops...) -> Bool

check if operators are commute.

YaoBase.isreflexive — Method.

isreflexive(op) -> Bool

check if this operator is reflexive.

YaoBase.isunitary — Method.

isunitary(op) -> Bool

check if this operator is a unitary operator.

YaoBase.measure — Function.

measure(register[, ntimes=1]) -> Vector{Int}

Return measurement results of current active qubits (regarding to active qubits, see focus! and relax!).

source

YaoBase.measure! — Method.

measure!(register[, locs])

measure and collapse to result state.

source

YaoBase.measure_remove! — Method.

measure_remove!(::AbstractRegister[, locs])

measure the active qubits of this register and remove them.

source

YaoBase.measure_reset! — Method.

measure_reset!(reg::AbstractRegister[, locs]; [val=0]) -> Int

measure and set the register to specific value.

source

YaoBase.nactive — Method.

nactive(register) -> Int

Returns the number of active qubits.

note!!!

Operators always apply on active qubits.

source

YaoBase.nbatch — Method.

nbatch(register) -> Int

Returns the number of batches.

source

YaoBase.nqubits — Method.

nqubits(register) -> Int

Returns the (total) number of qubits. See nactive, nremain for more details.

source

YaoBase.nremain — Method.

nremain(register) -> Int

Returns the number of non-active qubits.

source

YaoBase.relax! — Method.

relax!(register[, locs]) -> register
relax!(nbits, locs) -> f(register) -> register

Inverse transformation of focus!, where nbit is the number of active bits for target register.

source

YaoBase.select! — Method.

select!(dest::AbstractRegister, src::AbstractRegister, bits::Integer...) -> AbstractRegister
select!(register::AbstractRegister, bits::Integer...) -> register
select!(b::Integer) -> f(register)

select a subspace of given quantum state based on input eigen state bits.

Example

select!(reg, 0b110) will select the subspace with (focused) configuration 110. After selection, the focused qubit space is 0, so you may want call relax! manually.

source

YaoBase.select — Method.

select(register, bits...) -> AbstractRegister

Non-inplace version of select!.

source

YaoBase.state — Method.

state(register) -> AbstractMatrix

Returns the raw state of register. This always returns a matrix which is a batch of quantum states.

source

YaoBase.viewbatch — Method.

viewbatch(register, i::Int) -> AbstractRegister{1}

Returns a view of the i-th slice on batch dimension.

source

YaoBase.ρ — Method.

ρ(register)

Returns the density matrix of current active qubits. This is the same as density_matrix.

source

YaoBase.instruct — Function.

instruct(state, operator[, locs, control_bits, control_vals])

instruction implementation for applying an operator to a quantum state.

This operator will be overloaded for different operator or state with different types.